% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{add_left_mut}
\alias{add_left_mut}
\title{add_left_mut}
\usage{
add_left_mut(left, right)
}
\arguments{
\item{left}{Integer input (declared mut in Rust).}

\item{right}{Integer input.

Note: \code{mut} tests macro handling of mutable parameters; becomes unused after expansion.}
}
\description{
Test macro handling of a mutable left parameter.
}
