% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{add_panic_heap}
\alias{add_panic_heap}
\title{add_panic_heap}
\usage{
add_panic_heap(left, right)
}
\arguments{
\item{left}{Ignored integer input.}

\item{right}{Ignored integer input.}
}
\description{
Test that a panic drops a heap-allocated (Box) MsgOnDrop.
}
