% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{add_r_error}
\alias{add_r_error}
\title{add_r_error}
\usage{
add_r_error(left, right)
}
\arguments{
\item{left}{Ignored integer input.}

\item{right}{Ignored integer input.}
}
\description{
Test that Rf_error triggers an R error with a stack-allocated MsgOnDrop.
}
