% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{add_r_error_heap}
\alias{add_r_error_heap}
\title{add_r_error_heap}
\usage{
add_r_error_heap(left, right)
}
\arguments{
\item{left}{Ignored integer input.}

\item{right}{Ignored integer input.}
}
\description{
Test that Rf_error triggers an R error with a heap-allocated (Box) MsgOnDrop.
}
