% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{add_right_mut}
\alias{add_right_mut}
\title{add_right_mut}
\usage{
add_right_mut(left, right)
}
\arguments{
\item{left}{Integer input.}

\item{right}{Integer input.}
}
\description{
Test macro handling of a mutable right parameter.
}
