% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{add_score}
\alias{add_score}
\title{The \code{add_score()} generic}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{add_score()} is an S7 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{SidecarS7}}
}
}
