% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Myfunctions.R
\name{AddFive}
\alias{AddFive}
\title{Add Five to a number}
\usage{
AddFive(x)
}
\arguments{
\item{x}{A numeric value to be added}
}
\value{
the output after add 5
}
\description{
Takes in any parameter value and add 5 to it
}

