% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{aho_test_count}
\alias{aho_test_count}
\title{aho_test_count}
\usage{
aho_test_count(patterns, haystack)
}
\arguments{
\item{patterns}{Character vector of patterns to search for.}

\item{haystack}{Single string to search within.}
}
\description{
Test counting the number of pattern matches in the haystack.
}
