% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{alias_optional_lists}
\alias{alias_optional_lists}
\title{alias_optional_lists}
\usage{
alias_optional_lists(a, b)
}
\arguments{
\item{a}{First vector or list of vectors.}

\item{b}{Second vector or list of vectors.}
}
\description{
Exercise borrowed-slice alias checks with optional lists inputs.
}
