% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{alias_scalar_lists}
\alias{alias_scalar_lists}
\title{alias_scalar_lists}
\usage{
alias_scalar_lists(a, b)
}
\arguments{
\item{a}{First scalar, vector, or list.}

\item{b}{Second scalar, vector, or list.}
}
\description{
Exercise native borrow metadata through scalar lists conversions.
}
