% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{alias_three_arguments}
\alias{alias_three_arguments}
\title{alias_three_arguments}
\usage{
alias_three_arguments(a, b, c)
}
\arguments{
\item{a}{First list of integer vectors.}

\item{b}{Shared integer vector.}

\item{c}{Second list of integer vectors.}
}
\description{
Exercise batched diagnostics for three conflicting arguments.
}
