% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{alias_wrapped_scalars}
\alias{alias_wrapped_scalars}
\title{alias_wrapped_scalars}
\usage{
alias_wrapped_scalars(a, b)
}
\arguments{
\item{a}{First scalar, vector, or list.}

\item{b}{Second scalar, vector, or list.}
}
\description{
Exercise native borrow metadata through wrapped scalars conversions.
}
