% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{AliasPalette}
\alias{AliasPalette}
\alias{AliasPalette$new}
\alias{AliasPalette$paint}
\alias{$.AliasPalette}
\alias{[[.AliasPalette}
\title{AliasPalette Class}
\usage{
AliasPalette

\method{$}{AliasPalette}(self, name)

\method{[[}{AliasPalette}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Arguments of \code{AliasPalette$new()}:

\describe{
\item{\code{color}}{One of "red", "gray", "blue".}
}

Replace the colour and return it.

\describe{
\item{\code{color}}{One of "red", "gray", "blue".}
}
}
