% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{all}
\alias{all}
\alias{all,fm-method}
\alias{all,fmV-method}
\alias{fm.all}
\title{Are All Values True?}
\usage{
fm.all(x, lazy = FALSE)

\S4method{all}{fm}(x, ..., na.rm = FALSE)

\S4method{all}{fmV}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{a logical FlashR vector.}

\item{lazy}{indicates whether or not to evaluate it lazily.}

\item{...}{zero or more logical vectors.}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\value{
The value is a logical vector of length one.
}
\description{
\code{all} tests whether all of the values in a set of logical vectors are
true? \code{fm.all} tests whether all values in a single vector are true.
\code{fm.all} can evaluate it lazily.
}
\examples{
mat <- fm.runif.matrix(100, 10) > 0.5
all(mat)
fm.all(mat)
}

