% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{altrep_sexp_materialize_strings}
\alias{altrep_sexp_materialize_strings}
\title{altrep_sexp_materialize_strings}
\usage{
altrep_sexp_materialize_strings(x)
}
\arguments{
\item{x}{An ALTREP string vector.}
}
\description{
Try to wrap a SEXP in AltrepSexp and materialize strings.

Uses AltrepSexp parameter -- only accepts ALTREP vectors.
}
