% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{animal_kind}
\alias{animal_kind}
\title{The \code{animal_kind()} generic}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{animal_kind()} is an S7 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{S7Animal}}
}
}
