% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{any}
\alias{any}
\alias{any,fm-method}
\alias{any,fmV-method}
\alias{fm.any}
\title{Are some Values True?}
\usage{
fm.any(x, lazy = FALSE)

\S4method{any}{fm}(x, ..., na.rm = FALSE)

\S4method{any}{fmV}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{a logical FlashR vector.}

\item{lazy}{indicates whether or not to evaluate it lazily.}

\item{...}{zero or more logical vectors.}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\value{
The value is a logical vector of length one.
}
\description{
\code{any} tests whether some of the values in a set of logical vectors are
true? \code{fm.any} tests whether some of the values in a single vector are
true. \code{fm.any} can evaluate it lazily.
}
\examples{
mat <- fm.runif.matrix(100, 10) > 0.5
any(mat)
fm.any(mat)
}

