% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber-database_hooks.R
\name{api_database_hooks}
\alias{api_database_hooks}
\title{API Database Hooks}
\usage{
api_database_hooks()
}
\arguments{
\item{...}{Not used}
}
\value{
list
}
\description{
Utility function to close the database connection when the API is stopped.
}
