% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber-err_handler.R
\name{api_err_handler}
\alias{api_err_handler}
\title{Plumber API Error Handler}
\usage{
api_err_handler(req, res, err, ...)
}
\arguments{
\item{req}{Plumber request object}

\item{res}{Plumber response object}

\item{err}{Error object}

\item{...}{Not used}
}
\value{
function
}
\description{
Utility function to handle errors in the API.
}
