% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{api_ping}
\alias{api_ping}
\title{Ping the Plumber API}
\usage{
api_ping(req, res)
}
\arguments{
\item{req}{Plumber request object}

\item{res}{Plumber response object}
}
\value{
plumber response
}
\description{
Utility function used to perform an API healthcheck.
}
