% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{api_sessioninfo}
\alias{api_sessioninfo}
\title{API Session Info}
\usage{
api_sessioninfo(req, res)
}
\arguments{
\item{req}{Plumber request object}

\item{res}{Plumber response object}
}
\value{
plumber response
}
\description{
Utility function to retrieve the session info of the API.
}
