% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{api_version}
\alias{api_version}
\title{Request the API Package's Version}
\usage{
api_version(req, res)
}
\arguments{
\item{req}{Plumber request object}

\item{res}{Plumber response object}
}
\value{
plumber response
}
\description{
Utility function to retrieve the version of the API.
}
