% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arg_error_body_callless}
\alias{arg_error_body_callless}
\title{arg_error_body_callless}
\usage{
arg_error_body_callless(fill)
}
\arguments{
\item{fill}{A string.}
}
\description{
\code{arg_error_body} with \code{call = none}.
}
\keyword{internal}
