% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arg_error_body_worker}
\alias{arg_error_body_worker}
\title{arg_error_body_worker}
\usage{
arg_error_body_worker(fill)
}
\arguments{
\item{fill}{A string.}
}
\description{
\code{arg_error_body} on the worker thread.
}
\keyword{internal}
