% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arg_error_flag_unchecked}
\alias{arg_error_flag_unchecked}
\title{arg_error_flag_unchecked}
\usage{
arg_error_flag_unchecked(flag)
}
\arguments{
\item{flag}{\code{TRUE} or \code{FALSE}.}
}
\description{
\code{no_preconditions} on a logical scalar.
}
\keyword{internal}
