% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arg_error_pair}
\alias{arg_error_pair}
\title{arg_error_pair}
\usage{
arg_error_pair(pair)
}
\arguments{
\item{pair}{\verb{list(<integer>, <string>)}.}
}
\description{
A tuple argument: an unnamed list of a whole number and a string.
}
\keyword{internal}
