% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arg_error_peak}
\alias{arg_error_peak}
\title{arg_error_peak}
\usage{
arg_error_peak(dv)
}
\arguments{
\item{dv}{Numbers, strings or factor labels.}
}
\description{
Largest value of a vector read like \code{as.numeric()}.
}
\keyword{internal}
