% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arg_error_ratio}
\alias{arg_error_ratio}
\title{arg_error_ratio}
\usage{
arg_error_ratio(num, den)
}
\arguments{
\item{num, den}{A number, string or factor label of length 1.}
}
\description{
Ratio of two numbers read like \code{as.numeric()}.
}
\keyword{internal}
