% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arg_error_string_unchecked}
\alias{arg_error_string_unchecked}
\title{arg_error_string_unchecked}
\usage{
arg_error_string_unchecked(s)
}
\arguments{
\item{s}{A single string.}
}
\description{
\code{no_preconditions} on a string.
}
\keyword{internal}
