% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arg_error_strings_unchecked}
\alias{arg_error_strings_unchecked}
\title{arg_error_strings_unchecked}
\usage{
arg_error_strings_unchecked(xs)
}
\arguments{
\item{xs}{A character vector.}
}
\description{
\code{no_preconditions} on a character vector.
}
\keyword{internal}
