% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{Arithmetic}
\alias{*,ANY,fm-method}
\alias{*,ANY,fmV-method}
\alias{*,fm,ANY-method}
\alias{*,fm,fm-method}
\alias{*,fm,fmV-method}
\alias{*,fm,matrix-method}
\alias{*,fmV,ANY-method}
\alias{*,fmV,fm-method}
\alias{*,fmV,fmV-method}
\alias{*,matrix,fm-method}
\alias{+,ANY,fm-method}
\alias{+,ANY,fmV-method}
\alias{+,fm,ANY-method}
\alias{+,fm,fm-method}
\alias{+,fm,fmV-method}
\alias{+,fm,matrix-method}
\alias{+,fmV,ANY-method}
\alias{+,fmV,fm-method}
\alias{+,fmV,fmV-method}
\alias{+,matrix,fm-method}
\alias{-,ANY,fm-method}
\alias{-,ANY,fmV-method}
\alias{-,fm,ANY-method}
\alias{-,fm,fm-method}
\alias{-,fm,fmV-method}
\alias{-,fm,matrix-method}
\alias{-,fm,missing-method}
\alias{-,fmV,ANY-method}
\alias{-,fmV,fm-method}
\alias{-,fmV,fmV-method}
\alias{-,fmV,missing-method}
\alias{-,matrix,fm-method}
\alias{/,ANY,fm-method}
\alias{/,ANY,fmV-method}
\alias{/,fm,ANY-method}
\alias{/,fm,fm-method}
\alias{/,fm,fmV-method}
\alias{/,fm,matrix-method}
\alias{/,fmV,ANY-method}
\alias{/,fmV,fm-method}
\alias{/,fmV,fmV-method}
\alias{/,matrix,fm-method}
\alias{Arithmetic}
\alias{^,ANY,fmV-method}
\alias{^,fm,ANY-method}
\alias{^,fm,fm-method}
\alias{^,fm,fmV-method}
\alias{^,fm,matrix-method}
\alias{^,fmV,ANY-method}
\alias{^,fmV,fmV-method}
\alias{^,matrix,fm-method}
\title{Arithmetic Operators}
\usage{
\S4method{-}{fm,missing}(e1, e2)

\S4method{-}{fmV,missing}(e1, e2)

\S4method{+}{fm,fm}(e1, e2)

\S4method{+}{fmV,fmV}(e1, e2)

\S4method{+}{fm,fmV}(e1, e2)

\S4method{+}{fmV,fm}(e1, e2)

\S4method{+}{fm,matrix}(e1, e2)

\S4method{+}{matrix,fm}(e1, e2)

\S4method{+}{fm,ANY}(e1, e2)

\S4method{+}{ANY,fm}(e1, e2)

\S4method{+}{fmV,ANY}(e1, e2)

\S4method{+}{ANY,fmV}(e1, e2)

\S4method{-}{fm,fm}(e1, e2)

\S4method{-}{fmV,fmV}(e1, e2)

\S4method{-}{fm,fmV}(e1, e2)

\S4method{-}{fmV,fm}(e1, e2)

\S4method{-}{fm,matrix}(e1, e2)

\S4method{-}{matrix,fm}(e1, e2)

\S4method{-}{fm,ANY}(e1, e2)

\S4method{-}{ANY,fm}(e1, e2)

\S4method{-}{fmV,ANY}(e1, e2)

\S4method{-}{ANY,fmV}(e1, e2)

\S4method{*}{fm,fm}(e1, e2)

\S4method{*}{fmV,fmV}(e1, e2)

\S4method{*}{fm,fmV}(e1, e2)

\S4method{*}{fmV,fm}(e1, e2)

\S4method{*}{fm,matrix}(e1, e2)

\S4method{*}{matrix,fm}(e1, e2)

\S4method{*}{fm,ANY}(e1, e2)

\S4method{*}{ANY,fm}(e1, e2)

\S4method{*}{fmV,ANY}(e1, e2)

\S4method{*}{ANY,fmV}(e1, e2)

\S4method{/}{fm,fm}(e1, e2)

\S4method{/}{fmV,fmV}(e1, e2)

\S4method{/}{fm,fmV}(e1, e2)

\S4method{/}{fmV,fm}(e1, e2)

\S4method{/}{fm,matrix}(e1, e2)

\S4method{/}{matrix,fm}(e1, e2)

\S4method{/}{fm,ANY}(e1, e2)

\S4method{/}{ANY,fm}(e1, e2)

\S4method{/}{fmV,ANY}(e1, e2)

\S4method{/}{ANY,fmV}(e1, e2)

\S4method{^}{fm,fm}(e1, e2)

\S4method{^}{fmV,fmV}(e1, e2)

\S4method{^}{fm,fmV}(e1, e2)

\S4method{^}{fm,matrix}(e1, e2)

\S4method{^}{matrix,fm}(e1, e2)

\S4method{^}{ANY,fmV}(e1, e2)

\S4method{^}{fm,ANY}(e1, e2)

\S4method{^}{fmV,ANY}(e1, e2)
}
\arguments{
\item{e1, e2}{One of the operands need to be a FlashR object. If one operand
is a matrix and the other is a vector, we perform the arithmetic operation
on the vector and every column of the matrix. If one operand is a scalar,
we perform the operation on the scalar with every element in the matrix or
the vector.

There are a few exceptions. For example, the left operand of "^" has to be
a FlashR object and the right operand has to be a scalar.}
}
\description{
Perform unary or binary arithmetic operations on FlashR objects.
}
\examples{
mat1 <- fm.runif.matrix(100, 10)
mat2 <- fm.runif.matrix(100, 10)
mat <- -mat1
mat <- mat1 + mat2
mat <- mat1 - mat2
mat <- mat1 * mat2
mat <- mat1 / mat2
mat <- mat1 ^ 2
}

