% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrayvec_roundtrip_dbl}
\alias{arrayvec_roundtrip_dbl}
\title{arrayvec_roundtrip_dbl}
\usage{
arrayvec_roundtrip_dbl(v)
}
\arguments{
\item{v}{Numeric vector from R (max 4 elements).}
}
\description{
Test \verb{ArrayVec<[f64; 4]>} roundtrip through R.
}
