% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrayvec_roundtrip_int}
\alias{arrayvec_roundtrip_int}
\title{arrayvec_roundtrip_int}
\usage{
arrayvec_roundtrip_int(v)
}
\arguments{
\item{v}{Integer vector from R (max 8 elements).}
}
\description{
Test \verb{ArrayVec<[i32; 8]>} roundtrip through R.
}
