% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_f64_roundtrip}
\alias{arrow_f64_roundtrip}
\alias{arrow_f64_sum}
\alias{arrow_f64_len}
\alias{arrow_f64_null_count}
\alias{arrow_i32_roundtrip}
\alias{arrow_i32_sum}
\alias{arrow_i32_null_count}
\alias{arrow_u8_roundtrip}
\alias{arrow_u8_len}
\alias{arrow_bool_roundtrip}
\alias{arrow_bool_null_count}
\alias{arrow_string_roundtrip}
\alias{arrow_string_null_count}
\alias{arrow_recordbatch_roundtrip}
\alias{arrow_recordbatch_nrow}
\alias{arrow_recordbatch_ncol}
\alias{arrow_recordbatch_column_names}
\alias{arrow_arrayref_roundtrip}
\alias{arrow_arrayref_len}
\alias{arrow_f64_empty_roundtrip}
\alias{arrow_i32_empty_roundtrip}
\alias{arrow_factor_roundtrip}
\alias{arrow_factor_len}
\alias{arrow_date_roundtrip}
\alias{arrow_date_len}
\alias{arrow_posixct_roundtrip}
\alias{arrow_posixct_len}
\alias{arrow_recordbatch_typed_roundtrip}
\alias{arrow_f64_filter_non_null}
\alias{arrow_f64_mean}
\alias{arrow_arrayref_type_name}
\title{arrow_f64_roundtrip}
\usage{
arrow_f64_roundtrip(v)

arrow_f64_sum(v)

arrow_f64_len(v)

arrow_f64_null_count(v)

arrow_i32_roundtrip(v)

arrow_i32_sum(v)

arrow_i32_null_count(v)

arrow_u8_roundtrip(v)

arrow_u8_len(v)

arrow_bool_roundtrip(v)

arrow_bool_null_count(v)

arrow_string_roundtrip(v)

arrow_string_null_count(v)

arrow_recordbatch_roundtrip(rb)

arrow_recordbatch_nrow(rb)

arrow_recordbatch_ncol(rb)

arrow_recordbatch_column_names(rb)

arrow_arrayref_roundtrip(v)

arrow_arrayref_len(v)

arrow_f64_empty_roundtrip(v)

arrow_i32_empty_roundtrip(v)

arrow_factor_roundtrip(v)

arrow_factor_len(v)

arrow_date_roundtrip(v)

arrow_date_len(v)

arrow_posixct_roundtrip(v)

arrow_posixct_len(v)

arrow_recordbatch_typed_roundtrip(rb)

arrow_f64_filter_non_null(v)

arrow_f64_mean(v)

arrow_arrayref_type_name(v)
}
\arguments{
\item{v}{Arrow ArrayRef from R.}

\item{rb}{Arrow RecordBatch with typed columns from R.}
}
\description{
Test Float64Array roundtrip through Arrow.

Test summing elements of a Float64Array.

Test getting the length of a Float64Array.

Test counting null values in a Float64Array.

Test Int32Array roundtrip through Arrow.

Test summing elements of an Int32Array.

Test counting null values in an Int32Array.

Test UInt8Array roundtrip through Arrow.

Test getting the length of a UInt8Array.

Test BooleanArray roundtrip through Arrow.

Test counting null values in a BooleanArray.

Test StringArray roundtrip through Arrow.

Test counting null values in a StringArray.

Test RecordBatch roundtrip through Arrow.

Test getting the number of rows in a RecordBatch.

Test getting the number of columns in a RecordBatch.

Test extracting column names from a RecordBatch.

Test ArrayRef roundtrip through Arrow dynamic dispatch.

Test getting the length of a dynamically-typed ArrayRef.

Test empty Float64Array roundtrip through Arrow.

Test empty Int32Array roundtrip through Arrow.

Test factor roundtrip through Arrow StringDictionaryArray.

Test getting the length of a StringDictionaryArray from an R factor.

Test Date roundtrip through Arrow Date32Array.

Test getting the length of a Date32Array.

Test POSIXct roundtrip through Arrow TimestampSecondArray.

Test getting the length of a POSIXct converted to Arrow timestamp.

Test RecordBatch roundtrip with typed columns.

Filter a Float64Array to keep only non-null values.

Compute the mean of a Float64Array (ignoring nulls).

Get the data type name of an ArrayRef.
}
