% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_arrayref_len}
\alias{arrow_arrayref_len}
\title{arrow_arrayref_len}
\usage{
arrow_arrayref_len(v)
}
\arguments{
\item{v}{Arrow ArrayRef from R.}
}
\description{
Test getting the length of a dynamically-typed ArrayRef.
}
