% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_arrayref_roundtrip}
\alias{arrow_arrayref_roundtrip}
\title{arrow_arrayref_roundtrip}
\usage{
arrow_arrayref_roundtrip(v)
}
\arguments{
\item{v}{Arrow ArrayRef from R.}
}
\description{
Test ArrayRef roundtrip through Arrow dynamic dispatch.
}
