% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_arrayref_type_name}
\alias{arrow_arrayref_type_name}
\title{arrow_arrayref_type_name}
\usage{
arrow_arrayref_type_name(v)
}
\arguments{
\item{v}{Arrow ArrayRef from R.}
}
\description{
Get the data type name of an ArrayRef.
}
