% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_bool_null_count}
\alias{arrow_bool_null_count}
\title{arrow_bool_null_count}
\usage{
arrow_bool_null_count(v)
}
\arguments{
\item{v}{Arrow BooleanArray from R.}
}
\description{
Test counting null values in a BooleanArray.
}
