% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_bool_roundtrip}
\alias{arrow_bool_roundtrip}
\title{arrow_bool_roundtrip}
\usage{
arrow_bool_roundtrip(v)
}
\arguments{
\item{v}{Arrow BooleanArray from R.}
}
\description{
Test BooleanArray roundtrip through Arrow.
}
