% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_date_len}
\alias{arrow_date_len}
\title{arrow_date_len}
\usage{
arrow_date_len(v)
}
\arguments{
\item{v}{Arrow Date32Array from R Date.}
}
\description{
Test getting the length of a Date32Array.
}
