% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_date_roundtrip}
\alias{arrow_date_roundtrip}
\title{arrow_date_roundtrip}
\usage{
arrow_date_roundtrip(v)
}
\arguments{
\item{v}{Arrow Date32Array from R Date.}
}
\description{
Test Date roundtrip through Arrow Date32Array.
}
