% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_f64_len}
\alias{arrow_f64_len}
\title{arrow_f64_len}
\usage{
arrow_f64_len(v)
}
\arguments{
\item{v}{Arrow Float64Array from R.}
}
\description{
Test getting the length of a Float64Array.
}
