% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_f64_mean}
\alias{arrow_f64_mean}
\title{arrow_f64_mean}
\usage{
arrow_f64_mean(v)
}
\arguments{
\item{v}{Arrow Float64Array from R.}
}
\description{
Compute the mean of a Float64Array (ignoring nulls).
}
