% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_f64_roundtrip}
\alias{arrow_f64_roundtrip}
\title{arrow_f64_roundtrip}
\usage{
arrow_f64_roundtrip(v)
}
\arguments{
\item{v}{Arrow Float64Array from R.}
}
\description{
Test Float64Array roundtrip through Arrow.
}
