% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_f64_sum}
\alias{arrow_f64_sum}
\title{arrow_f64_sum}
\usage{
arrow_f64_sum(v)
}
\arguments{
\item{v}{Arrow Float64Array from R.}
}
\description{
Test summing elements of a Float64Array.
}
