% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_factor_len}
\alias{arrow_factor_len}
\title{arrow_factor_len}
\usage{
arrow_factor_len(v)
}
\arguments{
\item{v}{Arrow StringDictionaryArray from R factor.}
}
\description{
Test getting the length of a StringDictionaryArray from an R factor.
}
