% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_factor_roundtrip}
\alias{arrow_factor_roundtrip}
\title{arrow_factor_roundtrip}
\usage{
arrow_factor_roundtrip(v)
}
\arguments{
\item{v}{Arrow StringDictionaryArray from R factor.}
}
\description{
Test factor roundtrip through Arrow StringDictionaryArray.
}
