% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_i32_null_count}
\alias{arrow_i32_null_count}
\title{arrow_i32_null_count}
\usage{
arrow_i32_null_count(v)
}
\arguments{
\item{v}{Arrow Int32Array from R.}
}
\description{
Test counting null values in an Int32Array.
}
