% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_na_bool_null_positions}
\alias{arrow_na_bool_null_positions}
\title{arrow_na_bool_null_positions}
\usage{
arrow_na_bool_null_positions(v)
}
\arguments{
\item{v}{logical vector}
}
\description{
Return null positions for BooleanArray.
}
