% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_na_bool_roundtrip}
\alias{arrow_na_bool_roundtrip}
\title{arrow_na_bool_roundtrip}
\usage{
arrow_na_bool_roundtrip(v)
}
\arguments{
\item{v}{logical vector}
}
\description{
Roundtrip logical through Arrow.
}
