% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_na_f64_all_null_altrep}
\alias{arrow_na_f64_all_null_altrep}
\title{arrow_na_f64_all_null_altrep}
\usage{
arrow_na_f64_all_null_altrep(n)
}
\arguments{
\item{n}{length}
}
\description{
Create an all-null Float64Array as ALTREP.
}
