% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_na_i32_add_ten}
\alias{arrow_na_i32_add_ten}
\title{arrow_na_i32_add_ten}
\usage{
arrow_na_i32_add_ten(v)
}
\arguments{
\item{v}{integer vector}
}
\description{
Compute on i32 Arrow array (add 10), preserving nulls.
}
