% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_na_i32_altrep}
\alias{arrow_na_i32_altrep}
\title{arrow_na_i32_altrep}
\usage{
arrow_na_i32_altrep(v)
}
\arguments{
\item{v}{integer vector}
}
\description{
Create a Rust-computed Int32Array (with NAs) and return as ALTREP.
}
