% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_na_i32_null_count}
\alias{arrow_na_i32_null_count}
\title{arrow_na_i32_null_count}
\usage{
arrow_na_i32_null_count(v)
}
\arguments{
\item{v}{integer vector}
}
\description{
Return the null count for an Int32Array.
}
