% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_na_recordbatch_null_counts}
\alias{arrow_na_recordbatch_null_counts}
\title{arrow_na_recordbatch_null_counts}
\usage{
arrow_na_recordbatch_null_counts(rb)
}
\arguments{
\item{rb}{data.frame}
}
\description{
Return per-column null counts as integer vector.
}
