% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_na_recordbatch_roundtrip}
\alias{arrow_na_recordbatch_roundtrip}
\title{arrow_na_recordbatch_roundtrip}
\usage{
arrow_na_recordbatch_roundtrip(rb)
}
\arguments{
\item{rb}{data.frame}
}
\description{
Roundtrip data.frame with various NA patterns through Arrow.
}
