% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_na_string_altrep}
\alias{arrow_na_string_altrep}
\title{arrow_na_string_altrep}
\usage{
arrow_na_string_altrep(v)
}
\arguments{
\item{v}{character vector}
}
\description{
Create a Rust-computed StringArray (with NAs) and return as ALTREP.
}
