% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_na_string_compact}
\alias{arrow_na_string_compact}
\title{arrow_na_string_compact}
\usage{
arrow_na_string_compact(v)
}
\arguments{
\item{v}{character vector}
}
\description{
Filter out null strings, returning only non-null values.
}
