% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_na_string_null_positions}
\alias{arrow_na_string_null_positions}
\title{arrow_na_string_null_positions}
\usage{
arrow_na_string_null_positions(v)
}
\arguments{
\item{v}{character vector}
}
\description{
Return null positions for StringArray.
}
