% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_posixct_len}
\alias{arrow_posixct_len}
\title{arrow_posixct_len}
\usage{
arrow_posixct_len(v)
}
\arguments{
\item{v}{R POSIXct SEXP to convert to Arrow timestamp.}
}
\description{
Test getting the length of a POSIXct converted to Arrow timestamp.
}
