% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_posixct_roundtrip}
\alias{arrow_posixct_roundtrip}
\title{arrow_posixct_roundtrip}
\usage{
arrow_posixct_roundtrip(v)
}
\arguments{
\item{v}{R POSIXct SEXP to convert to Arrow timestamp.}
}
\description{
Test POSIXct roundtrip through Arrow TimestampSecondArray.
}
