% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_recordbatch_column_names}
\alias{arrow_recordbatch_column_names}
\title{arrow_recordbatch_column_names}
\usage{
arrow_recordbatch_column_names(rb)
}
\arguments{
\item{rb}{Arrow RecordBatch from R.}
}
\description{
Test extracting column names from a RecordBatch.
}
