% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_recordbatch_ncol}
\alias{arrow_recordbatch_ncol}
\title{arrow_recordbatch_ncol}
\usage{
arrow_recordbatch_ncol(rb)
}
\arguments{
\item{rb}{Arrow RecordBatch from R.}
}
\description{
Test getting the number of columns in a RecordBatch.
}
