% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_recordbatch_nrow}
\alias{arrow_recordbatch_nrow}
\title{arrow_recordbatch_nrow}
\usage{
arrow_recordbatch_nrow(rb)
}
\arguments{
\item{rb}{Arrow RecordBatch from R.}
}
\description{
Test getting the number of rows in a RecordBatch.
}
