% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_recordbatch_roundtrip}
\alias{arrow_recordbatch_roundtrip}
\title{arrow_recordbatch_roundtrip}
\usage{
arrow_recordbatch_roundtrip(rb)
}
\arguments{
\item{rb}{Arrow RecordBatch from R.}
}
\description{
Test RecordBatch roundtrip through Arrow.
}
