% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_recordbatch_typed_roundtrip}
\alias{arrow_recordbatch_typed_roundtrip}
\title{arrow_recordbatch_typed_roundtrip}
\usage{
arrow_recordbatch_typed_roundtrip(rb)
}
\arguments{
\item{rb}{Arrow RecordBatch with typed columns from R.}
}
\description{
Test RecordBatch roundtrip with typed columns.
}
