% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_string_roundtrip}
\alias{arrow_string_roundtrip}
\title{arrow_string_roundtrip}
\usage{
arrow_string_roundtrip(v)
}
\arguments{
\item{v}{Arrow StringArray from R.}
}
\description{
Test StringArray roundtrip through Arrow.
}
