% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arrow_u8_roundtrip}
\alias{arrow_u8_roundtrip}
\title{arrow_u8_roundtrip}
\usage{
arrow_u8_roundtrip(v)
}
\arguments{
\item{v}{Arrow UInt8Array from R.}
}
\description{
Test UInt8Array roundtrip through Arrow.
}
