% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-shiny.R
\name{as_csv}
\alias{as_csv}
\title{Export a DGEList as a csv file}
\usage{
as_csv(x, ...)
}
\arguments{
\item{x}{DGEList}
}
\value{
data.frame
}
\description{
This file is suitable for:
* degust.erc.monash.edu
* https://kcvi.shinyapps.io/START/ (https://github.com/jminnier/STARTapp)
}
\details{
no pdata
}
