% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{as_data_frame}
\alias{as_data_frame}
\alias{as_data_frame.AsCoerceErrorTest}
\alias{as_list}
\alias{as_list.AsCoerceErrorTest}
\alias{as_character}
\alias{as_character.AsCoerceErrorTest}
\alias{AsCoerceErrorTest}
\title{S3 generic for \code{as_data_frame}
S3 generic for \code{as_data_frame}}
\format{
An object of class \code{environment} of length 1.
}
\source{
Generated by miniextendr from \code{AsCoerceErrorTest::as_data_frame}

Generated by miniextendr from \code{AsCoerceErrorTest::as_data_frame}

Generated by miniextendr from \code{AsCoerceErrorTest::as_list}

Generated by miniextendr from \code{AsCoerceErrorTest::as_list}

Generated by miniextendr from \code{AsCoerceErrorTest::as_character}

Generated by miniextendr from \code{AsCoerceErrorTest::as_character}
}
\usage{
\method{as_data_frame}{AsCoerceErrorTest}(x, ...)

\method{as_list}{AsCoerceErrorTest}(x, ...)

\method{as_character}{AsCoerceErrorTest}(x, ...)

AsCoerceErrorTest
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments passed to methods}
}
\description{
Convert to data.frame - fails when empty.

Convert to list - always returns "not supported" error.

Convert to character - custom error message.
}
\keyword{datasets}
