% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{AtomicCounter}
\alias{AtomicCounter}
\alias{AtomicCounter$new_atomic}
\alias{$.AtomicCounter}
\alias{[[.AtomicCounter}
\alias{AtomicCounter$SharedCounter$value}
\alias{AtomicCounter$SharedCounter$increment}
\alias{AtomicCounter$SharedCounter$add}
\alias{AtomicCounter$SharedCounter$reset}
\title{AtomicCounter Class}
\usage{
AtomicCounter

\method{$}{AtomicCounter}(self, name)

\method{[[}{AtomicCounter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
AtomicCounter Class
}
