% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{batch_update_thread_files}
\alias{batch_update_thread_files}
\title{Update thread files in batches to minimize API requests}
\usage{
batch_update_thread_files(
  params,
  channel_dir,
  token,
  force = FALSE,
  timezone = "UTC",
  max_threads = 500
)
}
\arguments{
\item{params}{List of parameters for get_channel_threads}

\item{channel_dir}{Directory for channel files}

\item{token}{Authentication token}

\item{force}{Force update even if timestamps indicate no changes}

\item{timezone}{Timezone for timestamps}

\item{max_threads}{Maximum number of threads to process}
}
\value{
List of updated thread file paths
}
\description{
Update thread files in batches to minimize API requests
}
