% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{Bell}
\alias{Bell}
\alias{Bell$new}
\alias{Bell$rings}
\alias{$.Bell}
\alias{[[.Bell}
\alias{Bell$QuietBell$ring}
\alias{Bell$QuietBell$ring_quietly}
\alias{Bell$QuietBell$rings_quietly}
\title{Bell  Class}
\usage{
Bell

\method{$}{Bell}(self, name)

\method{[[}{Bell}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
A bell that has not rung yet.

Rings so far.
}
