% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bigint_add}
\alias{bigint_add}
\title{bigint_add}
\usage{
bigint_add(a, b)
}
\arguments{
\item{a}{First big integer as string.}

\item{b}{Second big integer as string.}
}
\description{
Test BigInt addition of two string-encoded integers.
}
