% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bigint_is_positive}
\alias{bigint_is_positive}
\title{bigint_is_positive}
\usage{
bigint_is_positive(s)
}
\arguments{
\item{s}{String representation of a big integer.}
}
\description{
Test whether a BigInt parsed from string is positive.
}
