% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bigint_mul}
\alias{bigint_mul}
\title{bigint_mul}
\usage{
bigint_mul(a, b)
}
\arguments{
\item{a}{First big integer as string.}

\item{b}{Second big integer as string.}
}
\description{
Test BigInt multiplication of two string-encoded integers.
}
