% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome.R
\name{biome}
\alias{biome}
\title{Creates a data frame with the ecosystem type of the sites}
\usage{
biome(database)
}
\arguments{
\item{database}{A list with the aridec structure}
}
\value{
A data frame with the ecosystem types from the database
}
\description{
Creates a data frame with the ecosystem type of the sites
}
\examples{
biome=biome(database=aridec)
}
