% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitflags_complement}
\alias{bitflags_complement}
\title{bitflags_complement}
\usage{
bitflags_complement(flags)
}
\arguments{
\item{flags}{Permission flags as RFlags wrapper.}
}
\description{
Compute the complement (NOT) of a flag set (within defined bits).
}
