% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitflags_count}
\alias{bitflags_count}
\title{bitflags_count}
\usage{
bitflags_count(flags)
}
\arguments{
\item{flags}{Permission flags as RFlags wrapper.}
}
\description{
Count the number of flags that are set.
}
