% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitflags_display}
\alias{bitflags_display}
\title{bitflags_display}
\usage{
bitflags_display(flags)
}
\arguments{
\item{flags}{Permission flags as RFlags wrapper.}
}
\description{
Display the Debug format of a flag set.
}
