% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitflags_from_truncate}
\alias{bitflags_from_truncate}
\title{bitflags_from_truncate}
\usage{
bitflags_from_truncate(value)
}
\arguments{
\item{value}{Integer bitmask to convert.}
}
\description{
Test truncating conversion from integer to flags, ignoring unknown bits.
}
