% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitflags_has_execute}
\alias{bitflags_has_execute}
\title{bitflags_has_execute}
\usage{
bitflags_has_execute(flags)
}
\arguments{
\item{flags}{Permission flags as RFlags wrapper.}
}
\description{
Test checking the EXECUTE flag on a permission set.
}
