% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitflags_has_read}
\alias{bitflags_has_read}
\title{bitflags_has_read}
\usage{
bitflags_has_read(flags)
}
\arguments{
\item{flags}{Permission flags as RFlags wrapper.}
}
\description{
Test checking the READ flag on a permission set.
}
