% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitflags_has_write}
\alias{bitflags_has_write}
\title{bitflags_has_write}
\usage{
bitflags_has_write(flags)
}
\arguments{
\item{flags}{Permission flags as RFlags wrapper.}
}
\description{
Test checking the WRITE flag on a permission set.
}
