% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitflags_intersect}
\alias{bitflags_intersect}
\title{bitflags_intersect}
\usage{
bitflags_intersect(a, b)
}
\arguments{
\item{a}{First permission flags.}

\item{b}{Second permission flags.}
}
\description{
Test computing the intersection (bitwise AND) of two flag sets.
}
