% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitflags_names}
\alias{bitflags_names}
\title{bitflags_names}
\usage{
bitflags_names(flags)
}
\arguments{
\item{flags}{Permission flags as RFlags wrapper.}
}
\description{
Get the names of all set flags as a character vector.
}
