% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitflags_roundtrip}
\alias{bitflags_roundtrip}
\title{bitflags_roundtrip}
\usage{
bitflags_roundtrip(flags)
}
\arguments{
\item{flags}{Permission flags as RFlags wrapper.}
}
\description{
Test RFlags roundtrip through R.
}
