% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitflags_union}
\alias{bitflags_union}
\title{bitflags_union}
\usage{
bitflags_union(a, b)
}
\arguments{
\item{a}{First permission flags.}

\item{b}{Second permission flags.}
}
\description{
Test computing the union (bitwise OR) of two flag sets.
}
