% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitflags_xor}
\alias{bitflags_xor}
\title{bitflags_xor}
\usage{
bitflags_xor(a, b)
}
\arguments{
\item{a}{First permission flags.}

\item{b}{Second permission flags.}
}
\description{
Compute the symmetric difference (XOR) of two flag sets.
}
