% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitvec_roundtrip}
\alias{bitvec_roundtrip}
\alias{bitvec_ones}
\alias{bitvec_zeros}
\alias{bitvec_from_vec}
\alias{bitvec_to_vec}
\alias{bitvec_len}
\alias{bitvec_empty}
\alias{bitvec_all_ones}
\alias{bitvec_all_zeros}
\alias{bitvec_toggle}
\title{bitvec_roundtrip}
\usage{
bitvec_roundtrip(bits)

bitvec_ones(bits)

bitvec_zeros(bits)

bitvec_from_vec(bools)

bitvec_to_vec(bits)

bitvec_len(bits)

bitvec_empty()

bitvec_all_ones(n)

bitvec_all_zeros(n)

bitvec_toggle(bits)
}
\arguments{
\item{bits}{Bit vector to flip.}

\item{bools}{Logical vector to convert.}

\item{n}{Number of bits.}
}
\description{
Test RBitVec roundtrip through R.

Test counting the number of set bits (ones) in a bit vector.

Test counting the number of unset bits (zeros) in a bit vector.

Test creating a bit vector from a logical vector.

Test converting a bit vector back to a logical vector.

Test getting the length of a bit vector.

Test creating and roundtripping an empty bit vector.

Test creating an all-ones bit vector of given length.

Test creating an all-zeros bit vector of given length.

Test toggling (flipping) all bits in a bit vector.
}
