% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitvec_all_ones}
\alias{bitvec_all_ones}
\title{bitvec_all_ones}
\usage{
bitvec_all_ones(n)
}
\arguments{
\item{n}{Number of bits.}
}
\description{
Test creating an all-ones bit vector of given length.
}
