% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitvec_all_zeros}
\alias{bitvec_all_zeros}
\title{bitvec_all_zeros}
\usage{
bitvec_all_zeros(n)
}
\arguments{
\item{n}{Number of bits.}
}
\description{
Test creating an all-zeros bit vector of given length.
}
