% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitvec_from_vec}
\alias{bitvec_from_vec}
\title{bitvec_from_vec}
\usage{
bitvec_from_vec(bools)
}
\arguments{
\item{bools}{Logical vector to convert.}
}
\description{
Test creating a bit vector from a logical vector.
}
