% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitvec_ones}
\alias{bitvec_ones}
\title{bitvec_ones}
\usage{
bitvec_ones(bits)
}
\arguments{
\item{bits}{Bit vector to count ones in.}
}
\description{
Test counting the number of set bits (ones) in a bit vector.
}
