% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitvec_roundtrip}
\alias{bitvec_roundtrip}
\title{bitvec_roundtrip}
\usage{
bitvec_roundtrip(bits)
}
\arguments{
\item{bits}{Bit vector to roundtrip.}
}
\description{
Test RBitVec roundtrip through R.
}
