% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitvec_to_vec}
\alias{bitvec_to_vec}
\title{bitvec_to_vec}
\usage{
bitvec_to_vec(bits)
}
\arguments{
\item{bits}{Bit vector to convert.}
}
\description{
Test converting a bit vector back to a logical vector.
}
