% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitvec_toggle}
\alias{bitvec_toggle}
\title{bitvec_toggle}
\usage{
bitvec_toggle(bits)
}
\arguments{
\item{bits}{Bit vector to flip.}
}
\description{
Test toggling (flipping) all bits in a bit vector.
}
