% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitvec_zeros}
\alias{bitvec_zeros}
\title{bitvec_zeros}
\usage{
bitvec_zeros(bits)
}
\arguments{
\item{bits}{Bit vector to count zeros in.}
}
\description{
Test counting the number of unset bits (zeros) in a bit vector.
}
