% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{borsh_hashmap_roundtrip}
\alias{borsh_hashmap_roundtrip}
\title{borsh_hashmap_roundtrip}
\usage{
borsh_hashmap_roundtrip()
}
\description{
Test roundtripping a HashMap<String, i32> through Borsh.
}
