% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{borsh_nested_roundtrip}
\alias{borsh_nested_roundtrip}
\title{borsh_nested_roundtrip}
\usage{
borsh_nested_roundtrip()
}
\description{
Test roundtripping a nested structure through Borsh serialization.
}
