% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{borsh_option_roundtrip}
\alias{borsh_option_roundtrip}
\title{borsh_option_roundtrip}
\usage{
borsh_option_roundtrip()
}
\description{
Test roundtripping Option\if{html}{\out{<i32>}} values (Some and None) through Borsh.
}
