% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{borsh_vec_bool_roundtrip}
\alias{borsh_vec_bool_roundtrip}
\title{borsh_vec_bool_roundtrip}
\usage{
borsh_vec_bool_roundtrip()
}
\description{
Test roundtripping a Vec\if{html}{\out{<bool>}} through Borsh.
}
