% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{box_share}
\alias{box_share}
\title{Share of values a range box covers, via S3 dispatch}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Share of values a range box covers, via S3 dispatch
}
