% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{box_slice_bool_roundtrip}
\alias{box_slice_bool_roundtrip}
\title{box_slice_bool_roundtrip}
\usage{
box_slice_bool_roundtrip(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
\verb{Box<[bool]>} roundtrip: R logical → Rust → R logical
}
