% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{box_slice_double}
\alias{box_slice_double}
\title{box_slice_double}
\usage{
box_slice_double(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
Transform: double each element using \verb{Box<[f64]>}
}
