% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{box_slice_f64_roundtrip}
\alias{box_slice_f64_roundtrip}
\title{box_slice_f64_roundtrip}
\usage{
box_slice_f64_roundtrip(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
\verb{Box<[f64]>} roundtrip: R numeric → Rust → R numeric
}
