% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{box_slice_option_string_roundtrip}
\alias{box_slice_option_string_roundtrip}
\title{box_slice_option_string_roundtrip}
\usage{
box_slice_option_string_roundtrip(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
\verb{Box<[Option<String>]>} with NA support
}
