% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{box_slice_raw_roundtrip}
\alias{box_slice_raw_roundtrip}
\title{box_slice_raw_roundtrip}
\usage{
box_slice_raw_roundtrip(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
\verb{Box<[u8]>} roundtrip: R raw → Rust → R raw
}
