% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{boxed_data_altrep}
\alias{boxed_data_altrep}
\title{boxed_data_altrep}
\usage{
boxed_data_altrep(n)
}
\arguments{
\item{n}{Length of the vector.}
}
\description{
Example: Using into_altrep() to store wrapper
}
