% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{bubble_ordering}
\alias{bubble_ordering}
\title{Order bubbles by topological depth}
\usage{
bubble_ordering(g, bubbles_df)
}
\arguments{
\item{g}{An \code{igraph} directed acyclic graph representing a gene splicing graph.}

\item{bubbles_df}{A data frame (or \code{S4Vectors::DataFrame}) of detected bubbles as
returned by \code{detect_bubbles_igraph}, with columns \code{source}, \code{sink},
\code{n}, \code{partitions}, and \code{paths}.}
}
\description{
Order bubbles by topological depth
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_edge_names(g)
bubbles_df <- grase::detect_bubbles_igraph(g)
ordered_df <- grase::bubble_ordering(g, bubbles_df)
}
}
