% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bubble_ordering3}
\alias{bubble_ordering3}
\title{Third bubble ordering method (most sophisticated)}
\usage{
bubble_ordering3(g, bubbles_df)
}
\description{
Third bubble ordering method (most sophisticated)
}
