% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjcnt_functions.R
\name{build_sj_matrix}
\alias{build_sj_matrix}
\title{Build a junction count matrix from a directory of STAR SJ.out.tab files}
\usage{
build_sj_matrix(sj_dir, use_multi = FALSE)
}
\arguments{
\item{sj_dir}{Directory containing *.SJ.out.tab files (searched recursively).}

\item{use_multi}{Logical. If TRUE, add n_multi to n_uniq (default FALSE).}
}
\value{
Named list: mat (junction x sample integer matrix), samples (sample names).
}
\description{
Build a junction count matrix from a directory of STAR SJ.out.tab files
}
