\name{builder_init}
\alias{builder_init}
\title{Initialise a builder configuration file}
\usage{
builder_init(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{builder}}.}
}
\description{
Creates a \code{builder.ini} configuration file in the current
working directory.
}
\value{
Exit code from builder.
}
