% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bytes_concat}
\alias{bytes_concat}
\title{bytes_concat}
\usage{
bytes_concat(a, b)
}
\arguments{
\item{a}{First raw vector.}

\item{b}{Second raw vector.}
}
\description{
Test concatenating two Bytes buffers.
}
