% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bytes_mut_roundtrip}
\alias{bytes_mut_roundtrip}
\title{bytes_mut_roundtrip}
\usage{
bytes_mut_roundtrip(data)
}
\arguments{
\item{data}{Raw vector to roundtrip.}
}
\description{
Test BytesMut roundtrip through R raw vector.
}
