% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bytes_roundtrip}
\alias{bytes_roundtrip}
\title{bytes_roundtrip}
\usage{
bytes_roundtrip(data)
}
\arguments{
\item{data}{Raw vector to roundtrip.}
}
\description{
Test Bytes roundtrip through R raw vector.
}
