% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalStream_Order.R
\name{CalcStreamOrder}
\alias{CalcStreamOrder}
\title{Calculate Stream order}
\usage{
CalcStreamOrder(basinID, dsID, segments)
}
\arguments{
\item{basinID}{a column list of basin IDs (can be obtained from the CalcSubbasins function Summary output column 1 "Basin_ID")}

\item{dsID}{downstream basin IDs, the downstream ID of each basin, should be correpsonding to the basin ID 
(can be obtained from the CalcSubbasins function Summary output Column 6 "Downstream_Basin_ID")}

\item{segments}{(optional) a channel mask with segments assigned with basin IDs 
(can use the CalcSubbasins 'segments' output for this).}
}
\value{
summary - A summary table with a row for every basin with three columns "Basin_ID" ""Downstream_ID" "StreamOrder_number"

order_mask (optional)- only available if segments is an input,this will output a spatial raster with the channel orders
}
\description{
Function to calculate Strahler stream orers using stream segments delineated using the CalcSubbasins Function
}
