% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define_Subbasins.R
\name{CalcSubbasins}
\alias{CalcSubbasins}
\title{Defining Subbasins}
\usage{
CalcSubbasins(
  direction,
  area,
  mask,
  d4 = c(1, 2, 3, 4),
  riv_th = 50,
  printflag = F,
  merge_th = 0
)
}
\arguments{
\item{direction}{a matrix of the flow directions for cells that have been processed - defaults to all zeros}

\item{area}{drainage areas for every cell}

\item{mask}{Mask with ones for cells to be processed and zeros for everything else - defaults to a mask of all 1's}

\item{d4}{directional numbering system: the numbers you want to assigne to down, left, top,right (defaults to 1,2,3,4)}

\item{riv_th}{threshold for the drainage area minimum used desigate cells as river cells, defaults to 50}

\item{printflag}{optional flag to print function progress. this is a test}

\item{merge_th}{after all the subbasins have been defined subbasins with areas <merg_th will be combined with their downstream neighbors (Defaults to 0 which means no merging will take place)}
}
\description{
Function to divide the domain into subbasins with individual stream segments
}
