% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalStream_Order.R
\name{CalStreamOrder}
\alias{CalStreamOrder}
\title{Defining Subbasins}
\usage{
CalStreamOrder(subbasin)
}
\arguments{
\item{subbasin}{subbasins calculated from CalcSubbasins function in PrioirtyFlow}
}
\value{
summary - A summary table with a row for every basin with three columns "Basin_ID" ""Downstream_ID" "StreamOrder_number"

channel_orders - If a stream map is provided this will output a spatial map with the channel orders
}
\description{
Function to calculate Strahler stream orers using stream segments delineated using the CalcSubbasins Function
}
