% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon.R
\name{carbon}
\alias{carbon}
\title{Creates a data frame with the carbon content in soil samples}
\usage{
carbon(database)
}
\arguments{
\item{database}{A list with the sidb}
}
\value{
A data.frame with the carbon content and the units of the soil samples from the database
}
\description{
Creates a data frame with the carbon content in soil samples
}
\examples{
sidb=loadEntries(path="~/sidb/data/")
C=carbon(database=sidb)
}
