% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{Create categorical column from expressions}
\usage{
categorize(.data, col, ..., .default = NA_character_)
}
\arguments{
\item{.data}{A data frame}

\item{col}{Bare column name}

\item{...}{Named expressions evaluating a logical value. The name will
be used for the categorical variable. For}

\item{.default}{The value to use for \code{col} if none of the logical
predicates in \code{...} are true.}
}
\value{
A data frame consisting of \code{.data} with the additional column
  \code{col}.
}
\description{
Create a categorical variable from logical expressions on a data frame.
}

