% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize_}
\alias{categorize_}
\title{Create categorical column from expressions (standard evaluation)}
\usage{
categorize_(data, col, dots, default = NA_character_)
}
\arguments{
\item{data}{A data frame}

\item{col}{A character vector with the column name to hold the new categorical
variable}

\item{dots}{An object coercible into a \code{\link{lazy_dots}} object which
contains the categories and conditions associated with them.}

\item{default}{Value for rows unmatched by any condition in \code{dots}.}

\item{...}{Unused.}
}
\description{
This is a S3 generic.
}

