% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ntsa.R
\name{ccm_test}
\alias{ccm_test}
\title{Causality test with CCM (Convergent Cross Mapping)}
\usage{
ccm_test(
  df,
  ref,
  tar,
  lib = matrix(c(1, nrow(df)), nrow = 1),
  E,
  Tp = 0,
  libMin = E + 2,
  numSamples = 100,
  seed = NULL,
  surr = NULL
)
}
\arguments{
\item{df}{vector, matrix, data.frame or tibble,
which contains time series to be analyzed}

\item{ref}{Numeric or character, which sets the library variable for xmap}

\item{tar}{Numeric or character, which sets the target variable for xmap}

\item{lib}{Vector or 2-column matrix, which sets the knot position in library data}

\item{E}{Integer, which is the dimension of ref's embedding matrix}

\item{Tp}{Integer, which is the prediction-time horizon in ref xmap tar}

\item{libMin}{Numeric, which setsthe minumum Library size}

\item{numSamples}{Integer, which is the boot-strap sampling size
(works when randomLibs = TRUE)}

\item{seed}{Integer, which sets the RND seed (works when randomLibs = TRUE)}

\item{surr}{A list that consists of the surrogate data of the ref's time series}
}
\description{
\code{ccm_test} returns the (tibble) result of causality test with CCM
Algorithm is adopted from Sugihara et al. 2012, Science
}
