% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ChainedError}
\alias{ChainedError}
\alias{ChainedError$new}
\alias{ChainedError$without_source}
\alias{$.ChainedError}
\alias{[[.ChainedError}
\alias{ChainedError$RError$error_message}
\alias{ChainedError$RError$error_chain}
\alias{ChainedError$RError$error_chain_length}
\title{ChainedError Class}
\usage{
ChainedError

\method{$}{ChainedError}(self, name)

\method{[[}{ChainedError}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
ChainedError inherent methods: constructors with and without source errors.

Arguments of \code{ChainedError$new()}:

\describe{
\item{\code{outer_msg}}{Character string for the outer error message.}
\item{\code{inner_msg}}{Character string for the inner (source) error message.}
}
}
