% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{channel_dir_name}
\alias{channel_dir_name}
\title{Create a directory name for a channel}
\usage{
channel_dir_name(channel)
}
\arguments{
\item{channel}{Channel object}
}
\value{
String to use as file name
}
\description{
Create a directory name for a channel
}
