% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{check_make_valid_coordinate_matrix}
\alias{check_make_valid_coordinate_matrix}
\title{Make or Check coordinate matrix}
\usage{
check_make_valid_coordinate_matrix(
  coordinate_matrix,
  y,
  parameter_name = "",
  check_inputs = TRUE
)
}
\arguments{
\item{coordinate_matrix}{Must be one of:
1. Non-negative scalar: Indicates that the coordinate matrix should be 
coordinates (i, j) of yty st |yty[i, j]| >= t
2. "full": Indicates that the coordinate matrix should be the full 
upper triangle coordinate matrix 
3. Integer matrix of shape (m, 2) such that coordinate_matrix[i, ] is
a coordinate (i, j). Must be sorted lexigraphically.}

\item{y}{The scaled data matrix of shape (n, p) where each row x[i, ],
x being the un-scaled data matrix, is believed to be drawn from N(0, theta)
Only used when `coordinate_matrix` is a non-negative scalar.}

\item{parameter_name}{The name of the parameter being checked. 
Used for error propagation}

\item{check_inputs}{Flag whether or not to check user provided input
If TRUE, checks `coordinate_matrix` for validity.
If FALSE, does not check `coordinate_matrix` for validity.}
}
\description{
TODO: Fill in description
}
