% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_numeric_minmax}
\alias{check_numeric_minmax}
\title{Check that a column is strictly numeric.}
\usage{
check_numeric_minmax(x, xname)
}
\arguments{
\item{x}{Column values, a vector}

\item{xname}{Column name}
}
\value{
Nothing (run for its warning side effect).
}
\description{
Check that a column is strictly numeric.
}
\keyword{internal}
