% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_template_info_columns}
\alias{check_template_info_columns}
\title{Check that column names in the info and template files match.}
\usage{
check_template_info_columns(template, template_info, outfile, verbose = TRUE)
}
\arguments{
\item{template}{Template structure read in from \code{ISRaD_Master_Template.xlsx}}

\item{template_info}{Template info structure read from \code{ISRaD_Template_Info.xlsx}}

\item{outfile}{File output is being written to}

\item{verbose}{Print output? Logical}
}
\value{
TRUE if any mismatches occur.
}
\description{
Check that column names in the info and template files match.
}
\note{
This is typically called only from \code{\link{checkTemplateFiles}}.
}
\keyword{internal}
