% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{checked_add}
\alias{checked_add}
\title{S3 generic for \code{checked_add}}
\source{
Generated by miniextendr from \verb{impl Counter for S3TraitCounter}
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments passed to methods}
}
\description{
S3 generic for \code{checked_add}
}
