% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{checked_bump}
\alias{checked_bump}
\title{Add an amount in place, rejecting negative values}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Add an amount in place, rejecting negative values
}
