% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTemplateFiles.R
\name{checkTemplateFiles}
\alias{checkTemplateFiles}
\title{Check ISRaD Template/Info files}
\usage{
checkTemplateFiles(outfile = "", verbose = TRUE)
}
\arguments{
\item{outfile}{file to dump the output report. Defaults to an empty string that will print
to standard output}

\item{verbose}{if TRUE (default) will print output to specified outfile}
}
\value{
Nothing (run for side effects).
}
\description{
Check that the template information file and the template file match appropriately.
}
\details{
Used in compile() function, but primarily a development tool
}
\examples{
checkTemplateFiles()
}
