% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{choices_color}
\alias{choices_color}
\title{choices_color}
\usage{
choices_color(color = c("red", "green", "blue"))
}
\arguments{
\item{color}{One of "red", "green", "blue".}
}
\description{
choices_color
}
