% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{choices_correlation}
\alias{choices_correlation}
\title{choices_correlation}
\usage{
choices_correlation(x, y, method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{x}{First value.}

\item{y}{Second value.}

\item{method}{One of "pearson", "kendall", "spearman".}
}
\description{
choices_correlation
}
