% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{choices_mixed}
\alias{choices_mixed}
\title{choices_mixed}
\usage{
choices_mixed(n, mode = c("fast", "safe"), verbose)
}
\arguments{
\item{n}{Count.}

\item{mode}{One of "fast", "safe".}

\item{verbose}{Whether to print details.}
}
\description{
choices_mixed
}
