% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{choices_multi_color}
\alias{choices_multi_color}
\title{choices_multi_color}
\usage{
choices_multi_color(colors = c("red", "green", "blue"))
}
\arguments{
\item{colors}{One or more colors.}
}
\description{
Select multiple colors.
}
