% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{choices_multi_metrics}
\alias{choices_multi_metrics}
\title{choices_multi_metrics}
\usage{
choices_multi_metrics(n, metrics = c("mean", "median", "sd", "var"))
}
\arguments{
\item{n}{Count.}

\item{metrics}{One or more metrics.}
}
\description{
Select metrics with several.ok.
}
