% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{choices_no_default_color}
\alias{choices_no_default_color}
\title{choices_no_default_color}
\usage{
choices_no_default_color(color)
}
\arguments{
\item{color}{One of "red", "green", "blue", or NULL.}
}
\description{
Inline choice without a default on a \verb{Missing<Option<String>>}.
}
