% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{choices_omitted_colors}
\alias{choices_omitted_colors}
\title{choices_omitted_colors}
\usage{
choices_omitted_colors(colors = c("red", "green", "blue"))
}
\arguments{
\item{colors}{One or more of "red", "green", "blue"; omitting the argument means no choice.}
}
\description{
Omittable several_ok inline choice on a \verb{Missing<Vec<String>>}.
}
