% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI2_help.R
\name{CI2_help}
\alias{CI2_help}
\title{help function to calculate 1-alpha confidence interval of concordance with weighted variances according to sample sizes.}
\usage{
CI2_help(theta, thetahat, alpha, N1, N2)
}
\arguments{
\item{theta}{confidence limit - to be found by uniroot function}

\item{thetahat}{estimated empirical concordance}

\item{alpha}{significance level}

\item{N1}{sample size of stage 1}

\item{N2}{sample size of stage 2}
}
\value{
variance of concordance
}
\description{
help function to calculate 1-alpha confidence interval of concordance with pooled variances (modified Hanley-McNeil approach) according to Newcombe, 2006.
}
\details{
eWHORM simulations
}
\author{
Sonja Zehetmayer
}
\keyword{internal}
