% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_help.R
\name{CI_help}
\alias{CI_help}
\title{help function to calculate 1-alpha confidence interval of concordance with pooled variances (modified Hanley-McNeil approach) according to Newcombe, 2006.}
\usage{
CI_help(theta, thetahat, N1, N2, alpha)
}
\arguments{
\item{theta}{unknown confidence limit - to be found by uniroot function}

\item{thetahat}{estimated empirical concordance}

\item{N1}{sample size of stage 1}

\item{N2}{sample size of stage 2}

\item{alpha}{significance level}
}
\description{
help function to calculate 1-alpha confidence interval of concordance with pooled variances (modified Hanley-McNeil approach) according to Newcombe, 2006.
}
\details{
eWHORM simulations
}
\author{
Sonja Zehetmayer
}
\keyword{internal}
