% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{classed_condition_vec}
\alias{classed_condition_vec}
\title{classed_condition_vec}
\usage{
classed_condition_vec(classes)
}
\arguments{
\item{classes}{(no documentation available)}
}
\description{
\verb{rust_condition!} with a class vector supplied as a \verb{Vec<String>}.
}
