% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{classed_error_vec}
\alias{classed_error_vec}
\title{classed_error_vec}
\usage{
classed_error_vec(member, family)
}
\arguments{
\item{member}{(no documentation available)}

\item{family}{(no documentation available)}
}
\description{
\verb{rust_error!(class = [member, family], ...)}.
}
