% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_coerce_widen}
\alias{test_coerce_widen}
\alias{test_coerce_attr_f32}
\alias{test_coerce_attr_i16}
\alias{test_coerce_attr_u16}
\alias{rpkg_coercion_tests}
\alias{test_coerce_identity}
\alias{test_coerce_bool_to_int}
\alias{test_coerce_via_helper}
\alias{test_rnative_newtype}
\alias{test_per_arg_coerce_vec}
\alias{test_coerce_attr_vec_u16}
\alias{test_per_arg_coerce_both}
\alias{test_rnative_named_field}
\alias{test_per_arg_coerce_first}
\alias{test_per_arg_coerce_second}
\alias{test_try_coerce_f64_to_i32}
\alias{test_coerce_attr_with_invisible}
\title{Test widening coercion from i32 to f64 via Coerce trait}
\source{
Generated by miniextendr from Rust fn `test_coerce_widen`

Generated by miniextendr from Rust fn `test_coerce_attr_f32`

Generated by miniextendr from Rust fn `test_coerce_attr_i16`

Generated by miniextendr from Rust fn `test_coerce_attr_u16`

Generated by miniextendr from Rust fn `test_coerce_identity`

Generated by miniextendr from Rust fn `test_rnative_newtype`

Generated by miniextendr from Rust fn `test_coerce_via_helper`

Generated by miniextendr from Rust fn `test_coerce_bool_to_int`

Generated by miniextendr from Rust fn `test_per_arg_coerce_vec`

Generated by miniextendr from Rust fn `test_coerce_attr_vec_u16`

Generated by miniextendr from Rust fn `test_per_arg_coerce_both`

Generated by miniextendr from Rust fn `test_rnative_named_field`

Generated by miniextendr from Rust fn `test_per_arg_coerce_first`

Generated by miniextendr from Rust fn `test_per_arg_coerce_second`

Generated by miniextendr from Rust fn `test_try_coerce_f64_to_i32`

Generated by miniextendr from Rust fn `test_coerce_attr_with_invisible`
}
\usage{
test_coerce_widen(x)

test_coerce_attr_f32(x)

test_coerce_attr_i16(x)

test_coerce_attr_u16(x)

test_coerce_identity(x)

test_rnative_newtype(id)

test_coerce_via_helper(x)

test_coerce_bool_to_int(x)

test_per_arg_coerce_vec(x, y)

test_coerce_attr_vec_u16(x)

test_per_arg_coerce_both(x, y)

test_rnative_named_field(temp)

test_per_arg_coerce_first(x, y)

test_per_arg_coerce_second(x, y)

test_try_coerce_f64_to_i32(x)

test_coerce_attr_with_invisible(x)
}
\arguments{
\item{x}{Integer scalar coerced to u16.}

\item{id}{Integer input coerced to UserId.}

\item{y}{Integer scalar coerced to u16.}

\item{temp}{Numeric input coerced to Temperature.}
}
\description{
Coercion and RNativeType tests
}
\examples{
test_coerce_identity(1L)
test_coerce_widen(1L)
test_try_coerce_f64_to_i32(1.2)
test_coerce_attr_u16(10L)
test_per_arg_coerce_first(10L, 5L)
}
