% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_collect_empty}
\alias{test_collect_empty}
\alias{test_collect_range}
\alias{test_collect_sines}
\alias{test_collect_na_f64}
\alias{test_collect_na_i32}
\alias{test_collect_squares}
\alias{test_collect_strings_upper}
\alias{test_collect_strings_numbered}
\title{test_collect_empty}
\source{
Generated by miniextendr from Rust fn \code{test_collect_empty}

Generated by miniextendr from Rust fn \code{test_collect_range}

Generated by miniextendr from Rust fn \code{test_collect_sines}

Generated by miniextendr from Rust fn \code{test_collect_na_f64}

Generated by miniextendr from Rust fn \code{test_collect_na_i32}

Generated by miniextendr from Rust fn \code{test_collect_squares}

Generated by miniextendr from Rust fn \code{test_collect_strings_upper}

Generated by miniextendr from Rust fn \code{test_collect_strings_numbered}
}
\usage{
test_collect_empty()

test_collect_range()

test_collect_sines(n)

test_collect_na_f64(n)

test_collect_na_i32(n)

test_collect_squares(n)

test_collect_strings_upper(words)

test_collect_strings_numbered(n)
}
\arguments{
\item{n}{(no documentation available)}

\item{words}{(no documentation available)}
}
\description{
test_collect_empty

test_collect_range

test_collect_sines

test_collect_na_f64

test_collect_na_i32

test_collect_squares

test_collect_strings_upper

test_collect_strings_numbered
}
