% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{Collections}
\alias{Collections}
\alias{Collections$new}
\alias{Collections$to_r}
\alias{Collections$empty}
\alias{Collections$from_r}
\alias{$.Collections}
\alias{[[.Collections}
\title{Collections Class}
\usage{
Collections

\method{$}{Collections}(self, name)

\method{[[}{Collections}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Collections Class
}
