% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{color}
\alias{color}
\title{The \code{color()} generic}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{color()} is an S7 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{S7GoldenRetriever}}
}
}
