% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{colSums}
\alias{colMeans,fm-method}
\alias{colSums}
\alias{colSums,fm-method}
\alias{fm.colSums}
\alias{fm.rowSums}
\alias{rowMeans,fm-method}
\alias{rowSums,fm-method}
\title{Form Row and Column Sums and Means}
\usage{
fm.rowSums(x, lazy = FALSE)

fm.colSums(x, lazy = FALSE)

\S4method{rowSums}{fm}(x, na.rm)

\S4method{colSums}{fm}(x, na.rm)

\S4method{rowMeans}{fm}(x, na.rm)

\S4method{colMeans}{fm}(x, na.rm)
}
\arguments{
\item{x}{a FlashR matrix.}

\item{lazy}{logical. indicates whether to evaluate the expression lazily.}

\item{na.rm}{logical. Should missing values (including NaN) be omitted
from the calculations?}
}
\value{
a FlashR vector.
}
\description{
Form row and column sums and means for numeric arrays.
}
\examples{
mat <- fm.runif.matrix(100, 10)
mat <- fm.rowSums(mat)
mat <- fm.colSums(mat)
mat <- rowSums(mat)
mat <- colSums(mat)
mat <- rowMeans(mat)
mat <- colMeans(mat)
}

