% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread.R
\name{comment_to_string}
\alias{comment_to_string}
\title{Convert comment to string}
\usage{
comment_to_string(comment, timezone = "UTC")
}
\arguments{
\item{comment}{Comment from list returned by \code{get_comments()}}

\item{timezone}{Timezone for timestamps (default: "UTC")}
}
\description{
Get the text representation of a comment for writing to a thread file.
}
