% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{complex_add}
\alias{complex_add}
\title{complex_add}
\usage{
complex_add(a, b)
}
\arguments{
\item{a}{First complex number.}

\item{b}{Second complex number.}
}
\description{
Test adding two complex numbers.
}
