% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{complex_from_parts}
\alias{complex_from_parts}
\title{complex_from_parts}
\usage{
complex_from_parts(re, im)
}
\arguments{
\item{re}{Real part.}

\item{im}{Imaginary part.}
}
\description{
Test constructing a complex number from real and imaginary parts.
}
