% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{complex_im}
\alias{complex_im}
\title{complex_im}
\usage{
complex_im(c)
}
\arguments{
\item{c}{Complex number from R.}
}
\description{
Test extracting the imaginary part of a complex number.
}
