% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{complex_is_finite}
\alias{complex_is_finite}
\title{complex_is_finite}
\usage{
complex_is_finite(c)
}
\arguments{
\item{c}{Complex number from R.}
}
\description{
Test whether a complex number has finite real and imaginary parts.
}
