% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{complex_norm}
\alias{complex_norm}
\title{complex_norm}
\usage{
complex_norm(c)
}
\arguments{
\item{c}{Complex number from R.}
}
\description{
Test computing the norm (absolute value) of a complex number.
}
