% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{complex_roundtrip}
\alias{complex_roundtrip}
\title{complex_roundtrip}
\usage{
complex_roundtrip(c)
}
\arguments{
\item{c}{Complex number from R.}
}
\description{
Test Complex\if{html}{\out{<f64>}} roundtrip through R.
}
