% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{complex_roundtrip_vec}
\alias{complex_roundtrip_vec}
\title{complex_roundtrip_vec}
\usage{
complex_roundtrip_vec(v)
}
\arguments{
\item{v}{Complex vector from R.}
}
\description{
Test Vec<Complex\if{html}{\out{<f64>}}> roundtrip through R complex vector.
}
