% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complm.R
\name{complm}
\alias{complm}
\title{Stratified clustered bootstrapping of (multivariate) linear models}
\usage{
complm(
  observed,
  formula,
  data = sample_data(observed),
  contrasts = NULL,
  boot = FALSE,
  times = 1000,
  clusters = seq(nrow(data)),
  strata = rep(1, nrow(data)),
  log_obs = FALSE
)
}
\arguments{
\item{observed}{A phyloseq object or phyloseq-otu_table object}

\item{data}{A data.frame or object coerceable to one}

\item{log_obs}{Set to TRUE if \code{observed} is already on the log scale}
}
\description{
coef(fit, boot = TRUE) returns a 3d array where the first dimention is
the response variables, second dimension is the coefficients, and third
dimension is bootstrap replicates.
}
