% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{condition_error_long_message}
\alias{condition_error_long_message}
\title{condition_error_long_message}
\usage{
condition_error_long_message(n_chunks)
}
\arguments{
\item{n_chunks}{(no documentation available)}
}
\description{
Long message exercises the \code{CString} + STRSXP encoding path.
}
