% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_condition_ok}
\alias{test_condition_ok}
\alias{test_condition_chained}
\alias{test_condition_parse_int}
\title{test condition ok}
\source{
Generated by miniextendr from Rust fn \code{test_condition_ok}

Generated by miniextendr from Rust fn \code{test_condition_chained}

Generated by miniextendr from Rust fn \code{test_condition_parse_int}
}
\usage{
test_condition_ok()

test_condition_chained(s)

test_condition_parse_int(s)
}
\arguments{
\item{s}{(no documentation available)}
}
\description{
test condition ok

test condition chained

test condition parse int
}
