% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{console_r_print}
\alias{console_r_print}
\title{console_r_print}
\usage{
console_r_print(msg)
}
\arguments{
\item{msg}{Message to print.}
}
\description{
Print a message to R's console via \verb{r_print!} (no trailing newline).
}
