% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{console_r_println}
\alias{console_r_println}
\title{console_r_println}
\usage{
console_r_println(msg)
}
\arguments{
\item{msg}{Message to print.}
}
\description{
Print a message plus newline via \verb{r_println!}.
}
