% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{console_r_println_empty}
\alias{console_r_println_empty}
\title{console_r_println_empty}
\usage{
console_r_println_empty()
}
\description{
Print just a newline via the zero-argument \verb{r_println!()} arm.
}
