% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{constant_logical}
\alias{constant_logical}
\title{constant logical}
\source{
Generated by miniextendr from Rust fn \code{constant_logical}
}
\usage{
constant_logical(value, n)
}
\arguments{
\item{value}{(no documentation available)}

\item{n}{(no documentation available)}
}
\description{
constant logical
}
