% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_as_named_list_array}
\alias{conv_as_named_list_array}
\title{conv_as_named_list_array}
\usage{
conv_as_named_list_array()
}
\description{
Return test: \verb{AsNamedList<[(String, f64); 2]>} -> R named list from fixed-size array.
}
