% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_as_named_list_duplicate_names}
\alias{conv_as_named_list_duplicate_names}
\title{conv_as_named_list_duplicate_names}
\usage{
conv_as_named_list_duplicate_names()
}
\description{
Return test: \code{AsNamedList} with duplicate key names -> R named list (duplicates preserved).
}
