% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_as_named_list_empty}
\alias{conv_as_named_list_empty}
\title{conv_as_named_list_empty}
\usage{
conv_as_named_list_empty()
}
\description{
Return test: empty \code{AsNamedList} -> R empty named list.
}
