% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_as_named_list_ext_trait}
\alias{conv_as_named_list_ext_trait}
\title{conv_as_named_list_ext_trait}
\usage{
conv_as_named_list_ext_trait()
}
\description{
Return test: \code{AsNamedListExt} trait method -> R named list.
}
