% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_as_named_list_slice}
\alias{conv_as_named_list_slice}
\title{conv_as_named_list_slice}
\usage{
conv_as_named_list_slice()
}
\description{
Return test: \code{AsNamedList} from a borrowed slice -> R named list via SEXP.
}
