% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_as_named_list_str_keys}
\alias{conv_as_named_list_str_keys}
\title{conv_as_named_list_str_keys}
\usage{
conv_as_named_list_str_keys()
}
\description{
Return test: \verb{AsNamedList<Vec<(&str, i32)>>} -> R named list with \verb{&str} keys.
}
