% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_as_named_list_vec}
\alias{conv_as_named_list_vec}
\title{conv_as_named_list_vec}
\usage{
conv_as_named_list_vec()
}
\description{
Return test: \verb{AsNamedList<Vec<(String, i32)>>} -> R named list.
}
