% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_as_named_vector_array}
\alias{conv_as_named_vector_array}
\title{conv_as_named_vector_array}
\usage{
conv_as_named_vector_array()
}
\description{
Return test: \code{AsNamedVector} from fixed-size array -> R named double vector.
}
