% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_as_named_vector_empty}
\alias{conv_as_named_vector_empty}
\title{conv_as_named_vector_empty}
\usage{
conv_as_named_vector_empty()
}
\description{
Return test: empty \code{AsNamedVector} -> R empty named double vector.
}
