% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_as_named_vector_f64}
\alias{conv_as_named_vector_f64}
\title{conv_as_named_vector_f64}
\usage{
conv_as_named_vector_f64()
}
\description{
Return test: \verb{AsNamedVector<Vec<(&str, f64)>>} -> R named double vector.
}
