% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_as_named_vector_option_i32}
\alias{conv_as_named_vector_option_i32}
\title{conv_as_named_vector_option_i32}
\usage{
conv_as_named_vector_option_i32()
}
\description{
Return test: \verb{AsNamedVector<Vec<(&str, Option<i32>)>>} -> R named integer vector with NA.
}
