% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_as_named_vector_slice}
\alias{conv_as_named_vector_slice}
\title{conv_as_named_vector_slice}
\usage{
conv_as_named_vector_slice()
}
\description{
Return test: \code{AsNamedVector} from a borrowed slice -> R named double vector via SEXP.
}
