% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_as_named_vector_string}
\alias{conv_as_named_vector_string}
\title{conv_as_named_vector_string}
\usage{
conv_as_named_vector_string()
}
\description{
Return test: \verb{AsNamedVector<Vec<(String, String)>>} -> R named character vector.
}
