% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_btreemap_f64_len}
\alias{conv_btreemap_f64_len}
\title{conv_btreemap_f64_len}
\usage{
conv_btreemap_f64_len(x)
}
\arguments{
\item{x}{Input named list of doubles (sorted by key).}
}
\description{
Test \verb{BTreeMap<String, f64>} argument: returns the number of entries.
}
