% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_btreemap_f64_ret}
\alias{conv_btreemap_f64_ret}
\title{conv_btreemap_f64_ret}
\usage{
conv_btreemap_f64_ret()
}
\description{
Return test: Rust \verb{BTreeMap<String, f64>} -> R named list of doubles (sorted).
}
