% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_btreemap_i32_len}
\alias{conv_btreemap_i32_len}
\title{conv_btreemap_i32_len}
\usage{
conv_btreemap_i32_len(x)
}
\arguments{
\item{x}{Input named list of integers (sorted by key).}
}
\description{
Test \verb{BTreeMap<String, i32>} argument: returns the number of entries.
}
