% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_btreemap_i32_ret}
\alias{conv_btreemap_i32_ret}
\title{conv_btreemap_i32_ret}
\usage{
conv_btreemap_i32_ret()
}
\description{
Return test: Rust \verb{BTreeMap<String, i32>} -> R named list of integers (sorted).
}
