% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_btreeset_i32_ret}
\alias{conv_btreeset_i32_ret}
\title{conv_btreeset_i32_ret}
\usage{
conv_btreeset_i32_ret()
}
\description{
Return test: Rust \verb{BTreeSet<i32>} -> R integer vector (sorted).
}
