% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_btreeset_string_len}
\alias{conv_btreeset_string_len}
\title{conv_btreeset_string_len}
\usage{
conv_btreeset_string_len(x)
}
\arguments{
\item{x}{Input character vector (duplicates removed, sorted).}
}
\description{
Test \verb{BTreeSet<String>} argument: returns the number of unique sorted elements.
}
