% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_btreeset_string_ret}
\alias{conv_btreeset_string_ret}
\title{conv_btreeset_string_ret}
\usage{
conv_btreeset_string_ret()
}
\description{
Return test: Rust \verb{BTreeSet<String>} -> R character vector (sorted).
}
