% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_btreeset_u8_ret}
\alias{conv_btreeset_u8_ret}
\title{conv_btreeset_u8_ret}
\usage{
conv_btreeset_u8_ret()
}
\description{
Return test: Rust \verb{BTreeSet<u8>} -> R raw vector (sorted).
}
