% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_char_arg}
\alias{conv_char_arg}
\title{conv_char_arg}
\usage{
conv_char_arg(x)
}
\arguments{
\item{x}{Input single character.}
}
\description{
Roundtrip test: R character scalar -> Rust \code{char} -> R character scalar.
}
