% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_hashmap_f64_len}
\alias{conv_hashmap_f64_len}
\title{conv_hashmap_f64_len}
\usage{
conv_hashmap_f64_len(x)
}
\arguments{
\item{x}{Input named list of doubles.}
}
\description{
Test \verb{HashMap<String, f64>} argument: returns the number of entries.
}
