% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_hashmap_f64_ret}
\alias{conv_hashmap_f64_ret}
\title{conv_hashmap_f64_ret}
\usage{
conv_hashmap_f64_ret()
}
\description{
Return test: Rust \verb{HashMap<String, f64>} -> R named list of doubles.
}
