% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_hashmap_i32_len}
\alias{conv_hashmap_i32_len}
\title{conv_hashmap_i32_len}
\usage{
conv_hashmap_i32_len(x)
}
\arguments{
\item{x}{Input named list of integers.}
}
\description{
Test \verb{HashMap<String, i32>} argument: returns the number of entries.
}
