% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_hashmap_i32_ret}
\alias{conv_hashmap_i32_ret}
\title{conv_hashmap_i32_ret}
\usage{
conv_hashmap_i32_ret()
}
\description{
Return test: Rust \verb{HashMap<String, i32>} -> R named list of integers.
}
