% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_hashmap_rlog_len}
\alias{conv_hashmap_rlog_len}
\title{conv_hashmap_rlog_len}
\usage{
conv_hashmap_rlog_len(x)
}
\arguments{
\item{x}{Input named list of logicals.}
}
\description{
Test \verb{HashMap<String, RLogical>} argument: returns the number of entries.
}
