% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_hashmap_rlog_ret}
\alias{conv_hashmap_rlog_ret}
\title{conv_hashmap_rlog_ret}
\usage{
conv_hashmap_rlog_ret()
}
\description{
Return test: Rust \verb{HashMap<String, RLogical>} -> R named list of logicals.
}
