% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_hashmap_string_len}
\alias{conv_hashmap_string_len}
\title{conv_hashmap_string_len}
\usage{
conv_hashmap_string_len(x)
}
\arguments{
\item{x}{Input named list of strings.}
}
\description{
Test \verb{HashMap<String, String>} argument: returns the number of entries.
}
