% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_hashmap_string_ret}
\alias{conv_hashmap_string_ret}
\title{conv_hashmap_string_ret}
\usage{
conv_hashmap_string_ret()
}
\description{
Return test: Rust \verb{HashMap<String, String>} -> R named list of strings.
}
