% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_hashset_i16_ret}
\alias{conv_hashset_i16_ret}
\title{conv_hashset_i16_ret}
\usage{
conv_hashset_i16_ret()
}
\description{
Return test: Rust \verb{HashSet<i16>} -> R integer vector (coerced via i32).
}
