% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_hashset_i32_len}
\alias{conv_hashset_i32_len}
\title{conv_hashset_i32_len}
\usage{
conv_hashset_i32_len(x)
}
\arguments{
\item{x}{Input integer vector (duplicates removed).}
}
\description{
Test \verb{HashSet<i32>} argument: returns the number of unique elements.
}
