% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_hashset_rlog_len}
\alias{conv_hashset_rlog_len}
\title{conv_hashset_rlog_len}
\usage{
conv_hashset_rlog_len(x)
}
\arguments{
\item{x}{Input logical vector (duplicates removed).}
}
\description{
Test \verb{HashSet<RLogical>} argument: returns the number of unique elements.
}
