% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_hashset_rlog_ret}
\alias{conv_hashset_rlog_ret}
\title{conv_hashset_rlog_ret}
\usage{
conv_hashset_rlog_ret()
}
\description{
Return test: Rust \verb{HashSet<RLogical>} -> R logical vector.
}
