% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_hashset_string_ret}
\alias{conv_hashset_string_ret}
\title{conv_hashset_string_ret}
\usage{
conv_hashset_string_ret()
}
\description{
Return test: Rust \verb{HashSet<String>} -> R character vector.
}
