% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_hashset_u8_len}
\alias{conv_hashset_u8_len}
\title{conv_hashset_u8_len}
\usage{
conv_hashset_u8_len(x)
}
\arguments{
\item{x}{Input raw vector (duplicates removed).}
}
\description{
Test \verb{HashSet<u8>} argument: returns the number of unique elements.
}
