% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_i64_arg}
\alias{conv_i64_arg}
\title{conv_i64_arg}
\usage{
conv_i64_arg(x)
}
\arguments{
\item{x}{Input value.}
}
\description{
Roundtrip test: R numeric -> Rust \code{i64} -> R numeric.
}
