% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_isize_arg}
\alias{conv_isize_arg}
\title{conv_isize_arg}
\usage{
conv_isize_arg(x)
}
\arguments{
\item{x}{Input value.}
}
\description{
Roundtrip test: R numeric -> Rust \code{isize} -> R numeric.
}
