% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_named_list_contains}
\alias{conv_named_list_contains}
\title{conv_named_list_contains}
\usage{
conv_named_list_contains(config)
}
\arguments{
\item{config}{Input named list.}
}
\description{
Test \code{NamedList::contains}: checks existence of "a", "b", and "missing".
}
