% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_named_list_len}
\alias{conv_named_list_len}
\title{conv_named_list_len}
\usage{
conv_named_list_len(config)
}
\arguments{
\item{config}{Input named list.}
}
\description{
Test \code{NamedList} length queries: returns total length and named-element count.
}
