% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_named_list_roundtrip}
\alias{conv_named_list_roundtrip}
\title{conv_named_list_roundtrip}
\usage{
conv_named_list_roundtrip(config)
}
\arguments{
\item{config}{Input named list.}
}
\description{
Test \code{NamedList} roundtrip: converts to \code{List} and back to R.
}
