% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_opt_btreemap_i32_some_ret}
\alias{conv_opt_btreemap_i32_some_ret}
\title{conv_opt_btreemap_i32_some_ret}
\usage{
conv_opt_btreemap_i32_some_ret()
}
\description{
Return test: \code{Some(BTreeMap)} as \verb{Option<BTreeMap<String, i32>>} -> R named list.
}
