% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_opt_hashset_i32_arg}
\alias{conv_opt_hashset_i32_arg}
\title{conv_opt_hashset_i32_arg}
\usage{
conv_opt_hashset_i32_arg(x)
}
\arguments{
\item{x}{Input integer vector or NULL.}
}
\description{
Test \verb{Option<HashSet<i32>>} argument: returns sum if Some, -999 if None.
}
