% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_opt_hashset_i32_none_ret}
\alias{conv_opt_hashset_i32_none_ret}
\title{conv_opt_hashset_i32_none_ret}
\usage{
conv_opt_hashset_i32_none_ret()
}
\description{
Return test: \code{None} as \verb{Option<HashSet<i32>>} -> R NULL.
}
