% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_opt_rbool_is_some}
\alias{conv_opt_rbool_is_some}
\title{conv_opt_rbool_is_some}
\usage{
conv_opt_rbool_is_some(x)
}
\arguments{
\item{x}{Input value or NULL.}
}
\description{
Test \verb{Option<Rboolean>} argument: returns 1 if Some, 0 if None (NA).
}
