% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_opt_ref_i32_some_ret}
\alias{conv_opt_ref_i32_some_ret}
\title{conv_opt_ref_i32_some_ret}
\usage{
conv_opt_ref_i32_some_ret()
}
\description{
Return test: \verb{Some(&42)} as \verb{Option<&i32>} -> R integer.
}
