% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_opt_string_none}
\alias{conv_opt_string_none}
\title{conv_opt_string_none}
\usage{
conv_opt_string_none()
}
\description{
Return test: \code{None} as \verb{Option<String>} -> R NA_character_.
}
