% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_opt_string_some}
\alias{conv_opt_string_some}
\title{conv_opt_string_some}
\usage{
conv_opt_string_some()
}
\description{
Return test: \code{Some("opt")} as \verb{Option<String>} -> R character scalar.
}
