% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_opt_vec_string_arg}
\alias{conv_opt_vec_string_arg}
\title{conv_opt_vec_string_arg}
\usage{
conv_opt_vec_string_arg(x)
}
\arguments{
\item{x}{Input character vector or NULL.}
}
\description{
Test \verb{Option<Vec<String>>} argument: returns length if Some, -999 if None (NULL).
}
