% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_opt_vec_string_some_ret}
\alias{conv_opt_vec_string_some_ret}
\title{conv_opt_vec_string_some_ret}
\usage{
conv_opt_vec_string_some_ret()
}
\description{
Return test: \verb{Some(vec!["a", "b"])} as \verb{Option<Vec<String>>} -> R character vector.
}
