% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_option_f32_some}
\alias{conv_option_f32_some}
\title{conv_option_f32_some}
\usage{
conv_option_f32_some()
}
\description{
Return test: \code{Some(1.5)} as \verb{Option<f32>} -> R double.
}
