% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_option_i64_some_big}
\alias{conv_option_i64_some_big}
\title{conv_option_i64_some_big}
\usage{
conv_option_i64_some_big()
}
\description{
Return test: \code{Some(i64::MAX)} as \verb{Option<i64>} -> R double (too large for i32).
}
