% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_option_i64_some_small}
\alias{conv_option_i64_some_small}
\title{conv_option_i64_some_small}
\usage{
conv_option_i64_some_small()
}
\description{
Return test: \code{Some(42)} as \verb{Option<i64>} -> R integer (fits in i32).
}
