% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_option_u32_some}
\alias{conv_option_u32_some}
\title{conv_option_u32_some}
\usage{
conv_option_u32_some()
}
\description{
Return test: \code{Some(100)} as \verb{Option<u32>} -> R integer.
}
