% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_ref_i32_arg}
\alias{conv_ref_i32_arg}
\title{conv_ref_i32_arg}
\usage{
conv_ref_i32_arg(x)
}
\arguments{
\item{x}{Input integer reference.}
}
\description{
Test \verb{&i32} reference argument: dereferences and returns the value.
}
