% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_ref_mut_i32_add_one}
\alias{conv_ref_mut_i32_add_one}
\title{conv_ref_mut_i32_add_one}
\usage{
conv_ref_mut_i32_add_one(x)
}
\arguments{
\item{x}{Input integer (mutated in-place).}
}
\description{
Test \verb{&mut i32} mutable reference: increments by one and returns.
}
