% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_result_i32_arg}
\alias{conv_result_i32_arg}
\title{conv_result_i32_arg}
\usage{
conv_result_i32_arg(x)
}
\arguments{
\item{x}{Input value (NULL becomes Err).}
}
\description{
Test \verb{Result<i32, ()>} argument: returns 1 if Ok, 0 if Err (NULL).
}
