% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_result_string_err}
\alias{conv_result_string_err}
\title{conv_result_string_err}
\usage{
conv_result_string_err()
}
\description{
Return test: \verb{Err(())} as \verb{Result<String, ()>} -> R NULL.
}
