% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_result_string_ok}
\alias{conv_result_string_ok}
\title{conv_result_string_ok}
\usage{
conv_result_string_ok()
}
\description{
Return test: \code{Ok("ok")} as \verb{Result<String, ()>} -> R character scalar.
}
