% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_result_vec_i32_err}
\alias{conv_result_vec_i32_err}
\title{conv_result_vec_i32_err}
\usage{
conv_result_vec_i32_err()
}
\description{
Return test: \verb{Err(())} as \verb{Result<Vec<i32>, ()>} -> R NULL.
}
