% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_result_vec_i32_ok}
\alias{conv_result_vec_i32_ok}
\title{conv_result_vec_i32_ok}
\usage{
conv_result_vec_i32_ok()
}
\description{
Return test: \verb{Ok(vec![1, 2])} as \verb{Result<Vec<i32>, ()>} -> R integer vector.
}
