% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_rlog_arg}
\alias{conv_rlog_arg}
\title{conv_rlog_arg}
\usage{
conv_rlog_arg(x)
}
\arguments{
\item{x}{Input value.}
}
\description{
Roundtrip test: R logical -> Rust \code{RLogical} -> R logical.
}
