% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_sexp_arg}
\alias{conv_sexp_arg}
\title{conv_sexp_arg}
\usage{
conv_sexp_arg(x)
}
\arguments{
\item{x}{Input value.}
}
\description{
Roundtrip test: R SEXP -> Rust \code{SEXP} -> R SEXP (passthrough).
}
