% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_slice_f64_len}
\alias{conv_slice_f64_len}
\title{conv_slice_f64_len}
\usage{
conv_slice_f64_len(x)
}
\arguments{
\item{x}{Input double vector.}
}
\description{
Test \verb{&[f64]} slice argument: returns the length of the slice.
}
