% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_slice_i32_len}
\alias{conv_slice_i32_len}
\title{conv_slice_i32_len}
\usage{
conv_slice_i32_len(x)
}
\arguments{
\item{x}{Input integer vector.}
}
\description{
Test \verb{&[i32]} slice argument: returns the length of the slice.
}
