% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_slice_mut_u8_add_one}
\alias{conv_slice_mut_u8_add_one}
\title{conv_slice_mut_u8_add_one}
\usage{
conv_slice_mut_u8_add_one(x)
}
\arguments{
\item{x}{Input raw vector (mutated in-place).}
}
\description{
Test \verb{&mut [u8]} mutable slice: wrapping-adds one to each element.
}
