% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_slice_rlog_len}
\alias{conv_slice_rlog_len}
\title{conv_slice_rlog_len}
\usage{
conv_slice_rlog_len(x)
}
\arguments{
\item{x}{Input logical vector.}
}
\description{
Test \verb{&[RLogical]} slice argument: returns the length of the slice.
}
