% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_slice_u8_len}
\alias{conv_slice_u8_len}
\title{conv_slice_u8_len}
\usage{
conv_slice_u8_len(x)
}
\arguments{
\item{x}{Input raw vector.}
}
\description{
Test \verb{&[u8]} slice argument: returns the length of the slice.
}
