% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_string_arg}
\alias{conv_string_arg}
\title{conv_string_arg}
\usage{
conv_string_arg(x)
}
\arguments{
\item{x}{Input value.}
}
\description{
Roundtrip test: R character scalar -> Rust \code{String} -> R character scalar.
}
