% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_try_from_list_field_x}
\alias{conv_try_from_list_field_x}
\title{conv_try_from_list_field_x}
\usage{
conv_try_from_list_field_x(input)
}
\arguments{
\item{input}{A list with an \code{x} entry.}
}
\description{
Test \verb{#[derive(TryFromList)]} on a list with repeated names: a repeated
field name (\code{x}) is an error, a repeated other name is ignored.
}
