% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_tuple1_roundtrip}
\alias{conv_tuple1_roundtrip}
\title{conv_tuple1_roundtrip}
\usage{
conv_tuple1_roundtrip(solo)
}
\arguments{
\item{solo}{Unnamed list containing a single integer.}
}
\description{
Test 1-tuple argument round-trip: reads a 1-element unnamed list as \verb{(i32,)} and returns it.
}
