% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_u16_arg}
\alias{conv_u16_arg}
\title{conv_u16_arg}
\usage{
conv_u16_arg(x)
}
\arguments{
\item{x}{Input value.}
}
\description{
Roundtrip test: R integer -> Rust \code{u16} -> R integer.
}
