% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_u64_arg}
\alias{conv_u64_arg}
\title{conv_u64_arg}
\usage{
conv_u64_arg(x)
}
\arguments{
\item{x}{Input value.}
}
\description{
Roundtrip test: R numeric -> Rust \code{u64} -> R numeric.
}
