% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_u8_arg}
\alias{conv_u8_arg}
\title{conv_u8_arg}
\usage{
conv_u8_arg(x)
}
\arguments{
\item{x}{Input value.}
}
\description{
Roundtrip test: R raw element -> Rust \code{u8} -> R raw element.
}
