% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_f32_len}
\alias{conv_vec_f32_len}
\title{conv_vec_f32_len}
\usage{
conv_vec_f32_len(x)
}
\arguments{
\item{x}{Input double vector (coerced from R double).}
}
\description{
Test \verb{Vec<f32>} argument: returns the length of the vector.
}
