% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_f64_len}
\alias{conv_vec_f64_len}
\title{conv_vec_f64_len}
\usage{
conv_vec_f64_len(x)
}
\arguments{
\item{x}{Input double vector.}
}
\description{
Test \verb{Vec<f64>} argument: returns the length of the vector.
}
