% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_i32_len}
\alias{conv_vec_i32_len}
\title{conv_vec_i32_len}
\usage{
conv_vec_i32_len(x)
}
\arguments{
\item{x}{Input integer vector.}
}
\description{
Test \verb{Vec<i32>} argument: returns the length of the vector.
}
