% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_i64_len}
\alias{conv_vec_i64_len}
\title{conv_vec_i64_len}
\usage{
conv_vec_i64_len(x)
}
\arguments{
\item{x}{Input numeric vector (coerced from R numeric).}
}
\description{
Test \verb{Vec<i64>} argument: returns the length of the vector.
}
