% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_i8_len}
\alias{conv_vec_i8_len}
\title{conv_vec_i8_len}
\usage{
conv_vec_i8_len(x)
}
\arguments{
\item{x}{Input integer vector (coerced from R integer).}
}
\description{
Test \verb{Vec<i8>} argument: returns the length of the vector.
}
