% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_mut_slice_i32_add_one}
\alias{conv_vec_mut_slice_i32_add_one}
\title{conv_vec_mut_slice_i32_add_one}
\usage{
conv_vec_mut_slice_i32_add_one(x)
}
\arguments{
\item{x}{Input list of integer vectors (mutated in-place).}
}
\description{
Test \verb{Vec<&mut [i32]>} argument: increments all elements and returns total length.
}
