% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_opt_bool_len}
\alias{conv_vec_opt_bool_len}
\title{conv_vec_opt_bool_len}
\usage{
conv_vec_opt_bool_len(x)
}
\arguments{
\item{x}{Input logical vector (may contain NA).}
}
\description{
Test \verb{Vec<Option<bool>>} argument: returns the length of the vector.
}
