% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_opt_bool_ret}
\alias{conv_vec_opt_bool_ret}
\title{conv_vec_opt_bool_ret}
\usage{
conv_vec_opt_bool_ret()
}
\description{
Return test: Rust \verb{Vec<Option<bool>>} -> R logical vector with NA.
}
