% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_opt_f64_ret}
\alias{conv_vec_opt_f64_ret}
\title{conv_vec_opt_f64_ret}
\usage{
conv_vec_opt_f64_ret()
}
\description{
Return test: Rust \verb{Vec<Option<f64>>} -> R double vector with NA.
}
