% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_opt_i32_ret}
\alias{conv_vec_opt_i32_ret}
\title{conv_vec_opt_i32_ret}
\usage{
conv_vec_opt_i32_ret()
}
\description{
Return test: Rust \verb{Vec<Option<i32>>} -> R integer vector with NA.
}
