% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_opt_rlog_ret}
\alias{conv_vec_opt_rlog_ret}
\title{conv_vec_opt_rlog_ret}
\usage{
conv_vec_opt_rlog_ret()
}
\description{
Return test: Rust \verb{Vec<Option<RLogical>>} -> R logical vector with NA.
}
