% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_opt_u8_len}
\alias{conv_vec_opt_u8_len}
\title{conv_vec_opt_u8_len}
\usage{
conv_vec_opt_u8_len(x)
}
\arguments{
\item{x}{Input raw vector (may contain NA).}
}
\description{
Test \verb{Vec<Option<u8>>} argument: returns the length of the vector.
}
