% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_option_f32_ret}
\alias{conv_vec_option_f32_ret}
\title{conv_vec_option_f32_ret}
\usage{
conv_vec_option_f32_ret()
}
\description{
Return test: Rust \verb{Vec<Option<f32>>} -> R double vector with NA.
}
