% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_option_i8_ret}
\alias{conv_vec_option_i8_ret}
\title{conv_vec_option_i8_ret}
\usage{
conv_vec_option_i8_ret()
}
\description{
Return test: Rust \verb{Vec<Option<i8>>} -> R integer vector with NA.
}
