% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_option_u64_ret_big}
\alias{conv_vec_option_u64_ret_big}
\title{conv_vec_option_u64_ret_big}
\usage{
conv_vec_option_u64_ret_big()
}
\description{
Return test: Rust \verb{Vec<Option<u64>>} with large values -> R numeric vector with NA.
}
