% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_option_u64_ret_small}
\alias{conv_vec_option_u64_ret_small}
\title{conv_vec_option_u64_ret_small}
\usage{
conv_vec_option_u64_ret_small()
}
\description{
Return test: Rust \verb{Vec<Option<u64>>} with small values -> R integer vector with NA.
}
