% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_slice_i32_total_len}
\alias{conv_vec_slice_i32_total_len}
\title{conv_vec_slice_i32_total_len}
\usage{
conv_vec_slice_i32_total_len(x)
}
\arguments{
\item{x}{Input list of integer vectors.}
}
\description{
Test \verb{Vec<&[i32]>} argument: returns the total length across all slices.
}
