% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_string_len}
\alias{conv_vec_string_len}
\title{conv_vec_string_len}
\usage{
conv_vec_string_len(x)
}
\arguments{
\item{x}{Input character vector.}
}
\description{
Test \verb{Vec<String>} argument: returns the length of the vector.
}
