% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_vec_i32_arg}
\alias{conv_vec_vec_i32_arg}
\title{conv_vec_vec_i32_arg}
\usage{
conv_vec_vec_i32_arg(x)
}
\arguments{
\item{x}{Input list of integer vectors.}
}
\description{
Test \verb{Vec<Vec<i32>>} argument: returns the sum across all inner vectors.
}
