% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_vec_i32_ret}
\alias{conv_vec_vec_i32_ret}
\title{conv_vec_vec_i32_ret}
\usage{
conv_vec_vec_i32_ret()
}
\description{
Return test: Rust \verb{Vec<Vec<i32>>} -> R list of integer vectors.
}
