% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_vec_vec_string_ret}
\alias{conv_vec_vec_string_ret}
\title{conv_vec_vec_string_ret}
\usage{
conv_vec_vec_string_ret()
}
\description{
Return test: Rust \verb{Vec<Vec<String>>} -> R list of character vectors.
}
