% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conversion_value}
\alias{conversion_value}
\title{The \code{conversion_value()} generic}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{conversion_value()} is an S7 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{ConvertInput}}
\item \code{\link{TypedConversion}}
\item \code{\link{AttributeConversion}}
}
}
