% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_fm_d14c.R
\name{convert_fm_d14c}
\alias{convert_fm_d14c}
\title{convert_fm_d14c}
\usage{
convert_fm_d14c(fm = NA, d14c = NA, obs_date_y)
}
\arguments{
\item{fm}{fraction modern (default NA)}

\item{d14c}{delta 14C in per mille (default NA)}

\item{obs_date_y}{year of observation/sample collection}
}
\description{
convert fraction modern to d14c and d14c to fraction modern
}
\details{
convenience function for radiocarbon unit conversions
}
\examples{
convert_fm_d14c(fm = 0.97057, obs_date_y = 2005)
convert_fm_d14c(d14c = -35.86611, obs_date_y = 2005)
}
\author{
J. Beem-Miller
}
