% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{coordinates}
\alias{coordinates}
\title{Creates a data frame with the coordinates of the sites}
\usage{
coordinates(database)
}
\arguments{
\item{database}{A list with the SIDb structure}
}
\value{
A data frame with the longitude and latitude of sites from the database
}
\description{
Creates a data frame with the coordinates of the sites
}
\examples{
coor=coordinates(database=sidb)
}
