% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_stats.R
\docType{methods}
\name{cor}
\alias{cor}
\alias{cor,fm-method}
\alias{cov,fm-method}
\alias{var,fm-method}
\title{Correlation, Variance and Covariance (Matrices)}
\usage{
\S4method{var}{fm}(x, y = NULL, na.rm = FALSE, use)

\S4method{cor}{fm}(x, y = NULL, use = "everything", method = c("pearson",
  "kendall", "spearman"))

\S4method{cov}{fm}(x, y = NULL, use = "everything", method = c("pearson",
  "kendall", "spearman"))
}
\arguments{
\item{x}{a numeric vector or matrix}

\item{y}{\code{NULL} (default) or a vector or matrix with compatible
dimensions to \code{x}. The default is equivalent to \code{y=x}.}

\item{use}{an optional character string giving a method for computing
covariances in the presence of missing values.  This must be
(an abbreviation of) one of the strings \code{"everything"},
\code{"all.obs"}, \code{"complete.obs"}, \code{"na.or.complete"}, or
\code{"pairwise.complete.obs"}.}

\item{method}{a character string indicating which correlation coefficient
(or covariance) is to be computed.  One of \code{"pearson"}
(default), \code{"kendall"}, or \code{"spearman"}: can be abbreviated.
Right now this argument isn't used.}
}
\description{
\code{cov} and \code{cor} compute the covariance or correlation of \code{x}
and \code{y} if these are vectors.  If \code{x}
and \code{y} are matrices then the covariances (or correlations) between
the columns of \code{x} and the columns of \code{y} are computed.
}
\examples{
mat <- fm.runif.matrix(100, 10)
var(mat)
cor(mat)
cov(mat)
}

