% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoncnt_functions.R
\name{count_bipartitions_both}
\alias{count_bipartitions_both}
\title{Count exonic part reads for both setdiff1 and setdiff2 of bipartition events}
\usage{
count_bipartitions_both(
  bipartition_file,
  countmat,
  sampleinfo,
  outdir,
  outfilesuffix
)
}
\description{
Count exonic part reads for both setdiff1 and setdiff2 of bipartition events
}
\examples{
\dontrun{
countmat <- read.table("GENE1.dexseq.counts.txt", header = TRUE)
sampleinfo <- c(sampleA = "group1", sampleB = "group1",
                sampleC = "group2", sampleD = "group2")
grase::count_bipartitions_both(
  bipartition_file = "GENE1.bipartition.txt",
  countmat = countmat, sampleinfo = sampleinfo,
  outdir = tempdir(), outfilesuffix = "bipartition"
)
}
}
