% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoncnt_functions.R
\name{count_multinomial}
\alias{count_multinomial}
\title{Count exonic part reads for multinomial events}
\usage{
count_multinomial(multinomial_file, countmat, sampleinfo, outdir)
}
\description{
Count exonic part reads for multinomial events
}
\examples{
\dontrun{
countmat <- read.table("GENE1.dexseq.counts.txt", header = TRUE)
sampleinfo <- c(sampleA = "group1", sampleB = "group1",
                sampleC = "group2", sampleD = "group2")
grase::count_multinomial(
  multinomial_file = "GENE1.multinomial.txt",
  countmat = countmat, sampleinfo = sampleinfo, outdir = tempdir()
)
}
}
