% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{counter_connection}
\alias{counter_connection}
\title{counter_connection}
\usage{
counter_connection(start, end)
}
\arguments{
\item{start}{First integer to generate}

\item{end}{Last integer to generate (inclusive)}
}
\value{
A read-only connection that generates one integer per line
}
\description{
Create a connection that generates sequential integers.
}
\examples{
\dontrun{
conn <- counter_connection(1L, 5L)
readLines(conn)  # c("1", "2", "3", "4", "5")
close(conn)
}
}
