% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{Counter_TraitEnv}
\alias{Counter_TraitEnv}
\alias{Counter_TraitEnv$new}
\alias{Counter_TraitEnv$get_value}
\alias{$.Counter_TraitEnv}
\alias{[[.Counter_TraitEnv}
\alias{Counter_TraitEnv$MatrixCounter$custom_get}
\alias{Counter_TraitEnv$MatrixCounter$custom_add}
\alias{Counter_TraitEnv$MatrixCounter$default_value}
\title{Counter_TraitEnv Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type \code{Counter_TraitEnv}
}
\usage{
Counter_TraitEnv

\method{$}{Counter_TraitEnv}(self, name)

\method{[[}{Counter_TraitEnv}(self, name)
}
\description{
Counter_TraitEnv Class
}
\keyword{datasets}
