% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{Counter_TraitR6}
\alias{Counter_TraitR6}
\alias{Counter_TraitR6$new}
\alias{Counter_TraitR6$get_value}
\alias{$.Counter_TraitR6}
\alias{[[.Counter_TraitR6}
\alias{r6_trait_MatrixCounter_custom_get}
\alias{r6_trait_MatrixCounter_custom_add}
\alias{Counter_TraitR6$MatrixCounter$default_value}
\title{Counter_TraitR6 Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type \code{Counter_TraitR6}

Generated by miniextendr from \verb{impl MatrixCounter for Counter_TraitR6}

Generated by miniextendr from \verb{impl MatrixCounter for Counter_TraitR6}
}
\usage{
Counter_TraitR6

\method{$}{Counter_TraitR6}(self, name)

\method{[[}{Counter_TraitR6}(self, name)

r6_trait_MatrixCounter_custom_get(x)

r6_trait_MatrixCounter_custom_add(x, n)
}
\arguments{
\item{x}{A \code{Counter_TraitR6} object}
}
\description{
Counter_TraitR6 Class

R6 trait method \code{MatrixCounter::custom_get} for Counter_TraitR6

R6 trait method \code{MatrixCounter::custom_add} for Counter_TraitR6

Static trait method MatrixCounter::default_value()
}
\keyword{datasets}
