% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{Counter_TraitS3}
\alias{Counter_TraitS3}
\alias{Counter_TraitS3$new}
\alias{Counter_TraitS3$get_value}
\alias{$.Counter_TraitS3}
\alias{[[.Counter_TraitS3}
\alias{custom_get.Counter_TraitS3}
\alias{custom_add.Counter_TraitS3}
\alias{Counter_TraitS3$MatrixCounter$default_value}
\title{Counter_TraitS3 Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type \code{Counter_TraitS3}
}
\usage{
Counter_TraitS3

\method{$}{Counter_TraitS3}(self, name)

\method{[[}{Counter_TraitS3}(self, name)

\method{custom_get}{Counter_TraitS3}(x, ...)

\method{custom_add}{Counter_TraitS3}(x, n, ...)
}
\description{
Counter_TraitS3 Class

Static trait method MatrixCounter::default_value()
}
\keyword{datasets}
