% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{Counter_TraitS4}
\alias{Counter_TraitS4}
\alias{Counter_TraitS4$new}
\alias{Counter_TraitS4$get_value}
\alias{$.Counter_TraitS4}
\alias{[[.Counter_TraitS4}
\alias{s4_trait_MatrixCounter_custom_get}
\alias{s4_trait_MatrixCounter_custom_add}
\alias{Counter_TraitS4_MatrixCounter_default_value}
\title{Counter_TraitS4 Class}
\format{
An object of class \code{environment} of length 2.
}
\source{
Generated by miniextendr from Rust type \code{Counter_TraitS4}

Generated by miniextendr from \verb{impl MatrixCounter for Counter_TraitS4}

Generated by miniextendr from \verb{impl MatrixCounter for Counter_TraitS4}
}
\usage{
Counter_TraitS4

\method{$}{Counter_TraitS4}(self, name)

\method{[[}{Counter_TraitS4}(self, name)

Counter_TraitS4_MatrixCounter_default_value()
}
\description{
Counter_TraitS4 Class

S4 generic for trait method \code{MatrixCounter::custom_get}

S4 generic for trait method \code{MatrixCounter::custom_add}

Static trait method MatrixCounter::default_value() for Counter_TraitS4
}
\keyword{datasets}
