% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{Counter_TraitS7}
\alias{Counter_TraitS7}
\alias{Counter_TraitS7$new}
\alias{Counter_TraitS7$get_value}
\alias{$.Counter_TraitS7}
\alias{[[.Counter_TraitS7}
\alias{s7_trait_MatrixCounter_custom_get}
\alias{s7_trait_MatrixCounter_custom_add}
\alias{Counter_TraitS7$MatrixCounter$default_value}
\title{Counter_TraitS7 Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type \code{Counter_TraitS7}

Generated by miniextendr from \verb{impl MatrixCounter for Counter_TraitS7}

Generated by miniextendr from \verb{impl MatrixCounter for Counter_TraitS7}
}
\usage{
Counter_TraitS7

\method{$}{Counter_TraitS7}(self, name)

\method{[[}{Counter_TraitS7}(self, name)
}
\description{
Counter_TraitS7 Class

S7 generic for trait method \code{MatrixCounter::custom_get}

S7 generic for trait method \code{MatrixCounter::custom_add}

Static trait method MatrixCounter::default_value()
}
\keyword{datasets}
