% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{CounterTraitR6}
\alias{CounterTraitR6}
\alias{CounterTraitR6$MatrixCounter$custom_get}
\alias{CounterTraitR6$MatrixCounter$custom_add}
\alias{CounterTraitR6$MatrixCounter$default_value}
\title{CounterTraitR6 R6 Class}
\description{
CounterTraitR6 R6 Class

Static trait method MatrixCounter::default_value()
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-CounterTraitR6-initialize}{\code{CounterTraitR6$new()}}
    \item \href{#method-CounterTraitR6-get_value}{\code{CounterTraitR6$get_value()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CounterTraitR6-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-CounterTraitR6-initialize}{}}}
\subsection{\code{CounterTraitR6$new()}}{
  Create a new counter with initial value.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CounterTraitR6$new(v, .ptr = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{v}}{Initial counter value.}
      \item{\code{.ptr}}{Internal pointer (used by static methods, not for direct use).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CounterTraitR6-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-CounterTraitR6-get_value}{}}}
\subsection{\code{CounterTraitR6$get_value()}}{
  Get the current counter value.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CounterTraitR6$get_value()}
    \if{html}{\out{</div>}}
  }
}

}
