% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{CounterTraitS3}
\alias{CounterTraitS3}
\alias{new_countertraits3}
\alias{get_value.CounterTraitS3}
\alias{custom_get.CounterTraitS3}
\alias{custom_add.CounterTraitS3}
\alias{CounterTraitS3$MatrixCounter$default_value}
\title{CounterTraitS3 S3 Class}
\usage{
new_countertraits3(v)

\method{get_value}{CounterTraitS3}(x, ...)

CounterTraitS3

\method{custom_get}{CounterTraitS3}(x, ...)

\method{custom_add}{CounterTraitS3}(x, n, ...)
}
\arguments{
\item{v}{Initial counter value.}

\item{x}{An object}

\item{...}{Additional arguments passed to methods}

\item{n}{Amount to add.}
}
\description{
S3 generic for \code{get_value}

S3 generic for \code{custom_get}

S3 generic for \code{custom_add}
}
