% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{class}
\name{CounterTraitS4}
\alias{CounterTraitS4}
\alias{CounterTraitS4-s4_get_value}
\alias{s4_get_value,CounterTraitS4-method}
\alias{s4_trait_MatrixCounter_custom_get}
\alias{s4_trait_MatrixCounter_custom_get,CounterTraitS4-method}
\alias{s4_trait_MatrixCounter_custom_add}
\alias{s4_trait_MatrixCounter_custom_add,CounterTraitS4-method}
\alias{CounterTraitS4_MatrixCounter_default_value}
\title{CounterTraitS4 S4 Class}
\usage{
CounterTraitS4(v)

\S4method{s4_trait_MatrixCounter_custom_get}{CounterTraitS4}(x, ...)

\S4method{s4_trait_MatrixCounter_custom_add}{CounterTraitS4}(x, n, ...)

CounterTraitS4_MatrixCounter_default_value()
}
\arguments{
\item{v}{Initial counter value.}

\item{x}{A \code{CounterTraitS4} object}

\item{...}{Additional arguments passed to methods}

\item{n}{Amount to add.}
}
\description{
S4 generic for trait method \code{MatrixCounter::custom_get}

S4 generic for trait method \code{MatrixCounter::custom_add}
}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to Rust \code{CounterTraitS4} struct}
}}

\keyword{internal}
