% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countNA.R
\name{countNA}
\alias{countNA}
\title{CountNA}
\usage{
countNA(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
This function accepts a data frame and counts the number of NAs
}
\examples{
countNA(insertNA(genProportions(ncol=10, nrow=10)  , .1))
}
