% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covsel.embed.R
\name{covsel.embed}
\alias{covsel.embed}
\title{covsel.embed}
\usage{
covsel.embed(
  covdata,
  pa,
  weights = NULL,
  force = NULL,
  algorithms = c("glm", "gam", "rf"),
  ncov = ceiling(log2(length(which(pa == 1)))),
  maxncov = 12,
  nthreads = detectCores()/2
)
}
\arguments{
\item{covdata}{data.frame containing covariate data (continuous values) extracted at 'pa' locations}

\item{pa}{numeric vector of species presences (1) and absences (0)}

\item{weights}{numeric vector containing the weights for each value in 'pa' (of length 'pa')}

\item{force}{optional character vector indicating the name(s) of the covariate(s) to be forced in the final set}

\item{algorithms}{character vector indicating the name(s) of the algorithms(s) to be used for the embedding procedure (options: 'glm', 'gam', 'rf')}

\item{ncov}{target number of covariates to include in the final set}

\item{maxncov}{maximum possible number of covariates in the final set}

\item{nthreads}{number of cores to be used during parallel operations}
}
\value{
A list with three objects: (i) a data.frame with the covariates selected after the regularization/penalization and ranking procedures (covdata), (ii) a data.frame with the individual ranks of all covariates for each target algorithm (ranks_1), (iii) a data.frame with the final average ranks of selected covariates (ranks_2)
}
\description{
Covariate selection with model-specific embedding (Step-2)
}
\examples{
library(covsel)
covdata<-data_covfilter
dim(covdata)
covdata_embed<-covsel.embed(covdata, pa=data_covsel$pa, algorithms=c('glm','gam','rf'))
dim(covdata_embed$covdata)
}
\author{
Antoine Adde (antoine.adde@unil.ch)
}
