% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{create_events_df}
\alias{create_events_df}
\title{Create a data frame from an aligned enum}
\source{
Generated by miniextendr from Rust fn \code{create_events_df}
}
\usage{
create_events_df()
}
\description{
Create a data frame from an aligned enum
}
