% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{create_expanded_points_df}
\alias{create_expanded_points_df}
\title{create_expanded_points_df}
\usage{
create_expanded_points_df()
}
\description{
Create a data frame with expanded array columns.
}
