% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{create_large_par_events}
\alias{create_large_par_events}
\title{create_large_par_events}
\usage{
create_large_par_events(n)
}
\arguments{
\item{n}{Number of rows to create.}
}
\description{
Create a large parallel events data frame.
}
