% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{create_large_par_points}
\alias{create_large_par_points}
\title{create_large_par_points}
\usage{
create_large_par_points(n)
}
\arguments{
\item{n}{Number of rows to create.}
}
\description{
Create a large parallel points data frame.
}
