% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{create_people_df}
\alias{create_people_df}
\title{Create a data frame with heterogeneous types}
\source{
Generated by miniextendr from Rust fn \code{create_people_df}
}
\usage{
create_people_df()
}
\description{
Create a data frame with heterogeneous types
}
