% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{create_points_df}
\alias{create_points_df}
\title{Create a data frame using derive macro}
\source{
Generated by miniextendr from Rust fn \code{create_points_df}
}
\usage{
create_points_df()
}
\description{
Create a data frame using derive macro
}
