% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{create_scored_items_df}
\alias{create_scored_items_df}
\title{create_scored_items_df}
\usage{
create_scored_items_df()
}
\description{
Create a data frame with skip, rename, and Vec expansion.
}
