% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{create_shapes_df}
\alias{create_shapes_df}
\title{Create a data frame from shapes (no tag column)}
\source{
Generated by miniextendr from Rust fn \code{create_shapes_df}
}
\usage{
create_shapes_df()
}
\description{
Create a data frame from shapes (no tag column)
}
