% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{create_shapes_split}
\alias{create_shapes_split}
\title{create_shapes_split}
\usage{
create_shapes_split()
}
\description{
Unzip shapes to a named list of data.frames (2 variants, shared field).
}
