% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{create_tuple_sig_split}
\alias{create_tuple_sig_split}
\title{create_tuple_sig_split}
\usage{
create_tuple_sig_split()
}
\description{
Unzip tuple-variant enum to a named list of data.frames.
}
