% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{create_workspace_dirs}
\alias{create_workspace_dirs}
\title{Create directory structure for workspace}
\usage{
create_workspace_dirs(
  workspace_id = twist_workspace_id(),
  token = twist_token(),
  workspace_dir = twist_workspace_dir()
)
}
\arguments{
\item{workspace_id}{ID of the workspace (default: from twist_workspace_id())}

\item{token}{Authentication token (default: from twist_token())}

\item{workspace_dir}{Directory to store workspace data (default: from
twist_workspace_dir())}
}
\value{
Vector of created directory paths
}
\description{
Creates a directory for each channel in the workspace and returns the
created paths.
}
