% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{crossprod}
\alias{crossprod}
\alias{crossprod,fm,ANY-method}
\alias{tcrossprod,fm,ANY-method}
\title{Matrix Crossproduct}
\usage{
\S4method{crossprod}{fm,ANY}(x, y = NULL)

\S4method{tcrossprod}{fm,ANY}(x, y = NULL)
}
\arguments{
\item{x, y}{numeric matrices (or vectors): \code{y = NULL} is taken
to be the same matrix as \code{x}. Vectors are promoted to single-column
or single-row matrices, depending on the context.}

\item{lazy}{a logical value, indicating to perform the computation lazily.}
}
\value{
a double matrix
}
\description{
Given matrices \code{x} and \code{y} as arguments, return a matrix
cross-product. This is formally equivalent to (but usually slightly
faster than) the call \code{t(x) \%*\% y} (\code{crossprod}) or
\code{x \%*\% t(y)} (\code{tcrossprod}).
}
\examples{
mat <- fm.runif.matrix(100, 10)
crossprod(mat)
tcrossprod(mat)
}

