% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{custom_add}
\alias{custom_add}
\title{Add to a counter through a custom-named trait method}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Add to a counter through a custom-named trait method
}
