% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit.R
\name{custom_knit}
\alias{custom_knit}
\title{Custom Knit function for specifying output directory}
\usage{
custom_knit(input, output_dir, intermediates_dir, ...)
}
\arguments{
\item{input}{File input}

\item{output_dir}{Custom output directory}

\item{intermediates_dir}{Intermediate files directory}
}
\description{
Custom Knit function for specifying output directory
}
