% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_df_chain}
\alias{test_df_chain}
\alias{test_df_count}
\alias{test_df_select}
\alias{test_df_columns}
\alias{test_df_aggregate}
\alias{test_df_sql_query}
\alias{test_df_global_agg}
\alias{test_df_sort_limit}
\title{test df chain}
\source{
Generated by miniextendr from Rust fn \code{test_df_chain}

Generated by miniextendr from Rust fn \code{test_df_count}

Generated by miniextendr from Rust fn \code{test_df_select}

Generated by miniextendr from Rust fn \code{test_df_columns}

Generated by miniextendr from Rust fn \code{test_df_aggregate}

Generated by miniextendr from Rust fn \code{test_df_sql_query}

Generated by miniextendr from Rust fn \code{test_df_global_agg}

Generated by miniextendr from Rust fn \code{test_df_sort_limit}
}
\usage{
test_df_chain(df)

test_df_count(df)

test_df_select(df, cols)

test_df_columns(df)

test_df_aggregate(df)

test_df_sql_query(df, sql)

test_df_global_agg(df)

test_df_sort_limit(df, col, asc, n)
}
\arguments{
\item{df}{(no documentation available)}

\item{cols}{(no documentation available)}

\item{sql}{(no documentation available)}

\item{col}{(no documentation available)}

\item{asc}{(no documentation available)}

\item{n}{(no documentation available)}
}
\description{
test df chain

test df count

test df select

test df columns

test df aggregate

test df sql query

test df global agg

test df sort limit
}
