% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-dbplyr__duckdb_connection.R
\name{db_connection_describe}
\alias{db_connection_describe}
\title{Description of the database connection}
\arguments{
\item{con}{A \code{\link{dbConnect}} object, as returned by \code{dbConnect()}}
}
\value{
String consisting of DuckDB version, user login name, operating system, R version and the name of database
}
\description{
Description of the database connection
}
