% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-dbplyr__duckdb_connection.R
\name{dbplyr_edition}
\alias{dbplyr_edition}
\title{Declare which version of dbplyr API is being called.}
\arguments{
\item{con}{A \code{\link{dbConnect}} object, as returned by \code{dbConnect()}}
}
\description{
Declare which version of dbplyr API is being called.
}
