% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{decimal_add}
\alias{decimal_add}
\title{decimal_add}
\usage{
decimal_add(a, b)
}
\arguments{
\item{a}{First decimal as string.}

\item{b}{Second decimal as string.}
}
\description{
Test Decimal addition of two string-encoded values.
}
