% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{decimal_is_zero}
\alias{decimal_is_zero}
\title{decimal_is_zero}
\usage{
decimal_is_zero(s)
}
\arguments{
\item{s}{Decimal value as string.}
}
\description{
Test whether a Decimal value parsed from string is zero.
}
