% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{decimal_mul}
\alias{decimal_mul}
\title{decimal_mul}
\usage{
decimal_mul(a, b)
}
\arguments{
\item{a}{First decimal as string.}

\item{b}{Second decimal as string.}
}
\description{
Test Decimal multiplication of two string-encoded values.
}
