% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{decimal_round}
\alias{decimal_round}
\title{decimal_round}
\usage{
decimal_round(s, dp)
}
\arguments{
\item{s}{Decimal value as string.}

\item{dp}{Number of decimal places.}
}
\description{
Test rounding a Decimal to a given number of decimal places.
}
