% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{decimal_roundtrip}
\alias{decimal_roundtrip}
\title{decimal_roundtrip}
\usage{
decimal_roundtrip(s)
}
\arguments{
\item{s}{Decimal value as string.}
}
\description{
Test Decimal string roundtrip via parsing and formatting.
}
