% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{decimal_scale}
\alias{decimal_scale}
\title{decimal_scale}
\usage{
decimal_scale(s)
}
\arguments{
\item{s}{Decimal value as string.}
}
\description{
Test extracting the scale (number of decimal digits) of a Decimal.
}
