% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{default_connection}
\alias{default_connection}
\title{Get the default connection}
\usage{
default_connection()
}
\value{
A DuckDB connection object
}
\description{
\code{default_connection()} returns a default, built-in connection
}
\examples{
conn <- default_connection()
print(duckdb::sql("SELECT 42", conn=conn))
}
