% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twistr.R
\name{defaults}
\alias{defaults}
\title{Twistr default options}
\arguments{
\item{workspace_id}{ID of the workspace (default: from twist_workspace_id())}

\item{token}{Authentication token (default: from twist_token())}

\item{workspace_dir}{Directory to store workspace data (default: from
twist_workspace_dir())}
}
\description{
Twistr default options
}
