% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{deferred_condition_value}
\alias{deferred_condition_value}
\title{deferred_condition_value}
\usage{
deferred_condition_value()
}
\description{
\verb{defer_condition!} with a custom class: silent without a handler.
}
