% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{deferred_finalizer_pointer}
\alias{deferred_finalizer_pointer}
\title{deferred_finalizer_pointer}
\usage{
deferred_finalizer_pointer()
}
\description{
An external pointer whose destructor attempts to queue a warning.
}
