% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{deferred_then_panic}
\alias{deferred_then_panic}
\title{deferred_then_panic}
\usage{
deferred_then_panic()
}
\description{
A deferred warning followed by \verb{rust_error!} (the panic transport).
}
