% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{defunct_fn}
\alias{defunct_fn}
\title{defunct_fn}
\usage{
defunct_fn(x)
}
\arguments{
\item{x}{Integer input value (unused, function is defunct).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A defunct function that should error.
}
