% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorityFlow_data.R
\docType{data}
\name{DEM}
\alias{DEM}
\title{Example watershed DEM for PrioirtyFlow}
\format{a matrix of elevaiton values with nrow=nx and ncol=ny dimensions of the domain}
\usage{
data(DEM)
}
\description{
This is a small elevation dataset (215km by 172 km at 1km spatial resolution)
}
\keyword{datasets}
