% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{demo_condition}
\alias{demo_condition}
\title{demo_condition}
\usage{
demo_condition(msg)
}
\arguments{
\item{msg}{(no documentation available)}
}
\description{
Signal a generic rust_condition (no-op if unhandled).
}
