% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{demo_condition_custom_class}
\alias{demo_condition_custom_class}
\title{demo_condition_custom_class}
\usage{
demo_condition_custom_class(class, msg)
}
\arguments{
\item{class}{(no documentation available)}

\item{msg}{(no documentation available)}
}
\description{
Signal a rust_condition with a custom class.
}
