% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{demo_error}
\alias{demo_error}
\title{demo_error}
\usage{
demo_error(msg)
}
\arguments{
\item{msg}{(no documentation available)}
}
\description{
Raise a rust_error with the standard class layering.
}
