% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{demo_error_custom_class}
\alias{demo_error_custom_class}
\title{demo_error_custom_class}
\usage{
demo_error_custom_class(class, msg)
}
\arguments{
\item{class}{(no documentation available)}

\item{msg}{(no documentation available)}
}
\description{
Raise a rust_error with a custom class prepended.
}
