% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{demo_error_data_keyed}
\alias{demo_error_data_keyed}
\title{demo_error_data_keyed}
\usage{
demo_error_data_keyed(value, code)
}
\arguments{
\item{value}{(no documentation available)}

\item{code}{(no documentation available)}
}
\description{
Raise an error using the keyed builder sugar (\verb{data = \{ k = v, ... \}}).
}
