% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{demo_warning_custom_class}
\alias{demo_warning_custom_class}
\title{demo_warning_custom_class}
\usage{
demo_warning_custom_class(class, msg)
}
\arguments{
\item{class}{(no documentation available)}

\item{msg}{(no documentation available)}
}
\description{
Raise a rust_warning with a custom class prepended.
}
