% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{read_csv_guinsoodb}
\alias{read_csv_guinsoodb}
\title{Deprecated functions}
\usage{
read_csv_guinsoodb(conn, files, tablename, ...)
}
\description{
\code{read_csv_guinsoodb()} has been superseded by \code{guinsoodb_read_csv()}.
The order of the arguments has changed.
}
\keyword{internal}
