% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{derived_percent_class_info}
\alias{derived_percent_class_info}
\title{Verify VctrsClass trait constants}
\source{
Generated by miniextendr from Rust fn \code{derived_percent_class_info}
}
\usage{
derived_percent_class_info()
}
\description{
Verify VctrsClass trait constants
}
