% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{derived_rational_class_info}
\alias{derived_rational_class_info}
\title{derived_rational_class_info}
\usage{
derived_rational_class_info()
}
\description{
Verify VctrsClass trait constants for rational.
}
