% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{detect_bubbles_from_mat}
\alias{detect_bubbles_from_mat}
\title{Main bubble detection from matrix}
\usage{
detect_bubbles_from_mat(txpathmat)
}
\arguments{
\item{txpathmat}{A logical matrix of transcript paths (rows = transcripts, columns = nodes).}
}
\description{
Main bubble detection from matrix
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_txpath_to_vertex_attr(g)
txmat <- grase::make_matrix_from_txpath_igraph(g)
bubbles_df <- grase::detect_bubbles_from_mat(txmat)
}
}
