% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{detect_bubbles_i_j_igraph}
\alias{detect_bubbles_i_j_igraph}
\title{bubble detection between v_start and v_end}
\usage{
detect_bubbles_i_j_igraph(v_start_idx, v_end_idx, g, v_start_name, v_end_name)
}
\description{
bubble detection between v_start and v_end
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_txpath_to_vertex_attr(g)
v_names <- igraph::V(g)$name
result <- grase::detect_bubbles_i_j_igraph(2, 6, g, v_names[2], v_names[6])
}
}
