% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{detect_bubbles_igraph_wrapper}
\alias{detect_bubbles_igraph_wrapper}
\title{Wrapper: detect bubbles on igraph DAG}
\usage{
detect_bubbles_igraph_wrapper(g)
}
\description{
Wrapper: detect bubbles on igraph DAG
}
