% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational.R
\name{df_is_materialized}
\alias{df_is_materialized}
\title{Checks if a lazy data frame created using rel_to_altrep has been materialized yet}
\usage{
df_is_materialized(df)
}
\arguments{
\item{df}{an altrep-backed lazy data frame}
}
\value{
true if materialization has happened
}
\description{
Checks if a lazy data frame created using rel_to_altrep has been materialized yet
}
\examples{
con <- DBI::dbConnect(duckdb())
rel <- rel_from_df(con, mtcars)
df <- rel_to_altrep(rel)
stopifnot(!df_is_materialized(df))
str(df)
stopifnot(df_is_materialized(df))
}
