% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{dict_node2pos}
\alias{dict_node2pos}
\title{Generate dictionary that translates between node IDs and positions}
\usage{
dict_node2pos(sg, gene)
}
\arguments{
\item{sg}{A SplicingGraph object}
}
\value{
A named vector mapping node IDs to genomic positions
}
\description{
Generate dictionary that translates between node IDs and positions
}
