% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartition_analysis.R
\name{diff_exons_gene_ovr}
\alias{diff_exons_gene_ovr}
\title{Compute diff exons with one-vs-rest approach for each partition}
\usage{
diff_exons_gene_ovr(gene, g, outdir)
}
\arguments{
\item{gene}{Character string. Gene identifier (e.g., Ensembl gene ID).}

\item{g}{An \code{igraph} directed acyclic graph representing a gene splicing graph.}

\item{outdir}{Character string. Path to output directory.}
}
\description{
Compute diff exons with one-vs-rest approach for each partition
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
result <- grase::diff_exons_gene_ovr(
  gene = "ENSG00000000003.14", g = g, outdir = tempdir()
)
}
}
