% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{dim}
\alias{dim}
\title{Dimensions of an FlashMatrix Object}
\usage{
\S4method{dim}{fm}(x)
}
\arguments{
\item{x}{A FlashMatrix matrix.}
}
\description{
Retrieve the dimension of an FlashMatrix object.
}

