% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{dimnames}
\alias{dimnames}
\alias{dimnames,fm-method}
\alias{dimnames,fmV-method}
\alias{dimnames<-,fm,list-method}
\alias{dimnames<-,fmV,list-method}
\title{Dimnames of an Object}
\usage{
\S4method{dimnames}{fm}(x)

\S4method{dimnames}{fmV}(x)

\S4method{dimnames}{fm,list}(x) <- value

\S4method{dimnames}{fmV,list}(x) <- value
}
\arguments{
\item{x}{a FlashR matrix.}

\item{value}{a list that provides values for dimension names.}
}
\value{
\code{dimnames} return NULL if there aren't dimension names.
        Otherwise, return the dimension names.
}
\description{
Retrieve or set the dimnames of a matrix.
}
\examples{
mat <- fm.runif.matrix(100, 10)
dimnames(mat)
dimnames(mat) <- list("dim1", "dim2")
}

