% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_display_ip}
\alias{test_display_ip}
\alias{test_fromstr_ip}
\alias{test_fromstr_int}
\alias{test_display_bool}
\alias{test_display_number}
\alias{test_display_vec_ips}
\alias{test_fromstr_vec_ips}
\alias{test_display_vec_ints}
\alias{test_fromstr_vec_ints}
\alias{test_fromstr_bad_input}
\title{test_display_ip}
\source{
Generated by miniextendr from Rust fn \code{test_display_ip}

Generated by miniextendr from Rust fn \code{test_fromstr_ip}

Generated by miniextendr from Rust fn \code{test_fromstr_int}

Generated by miniextendr from Rust fn \code{test_display_bool}

Generated by miniextendr from Rust fn \code{test_display_number}

Generated by miniextendr from Rust fn \code{test_display_vec_ips}

Generated by miniextendr from Rust fn \code{test_fromstr_vec_ips}

Generated by miniextendr from Rust fn \code{test_display_vec_ints}

Generated by miniextendr from Rust fn \code{test_fromstr_vec_ints}

Generated by miniextendr from Rust fn \code{test_fromstr_bad_input}
}
\usage{
test_display_ip()

test_fromstr_ip(addr)

test_fromstr_int(s)

test_display_bool()

test_display_number()

test_display_vec_ips()

test_fromstr_vec_ips(addrs)

test_display_vec_ints()

test_fromstr_vec_ints(nums)

test_fromstr_bad_input(s)
}
\arguments{
\item{addr}{(no documentation available)}

\item{s}{(no documentation available)}

\item{addrs}{(no documentation available)}

\item{nums}{(no documentation available)}
}
\description{
test_display_ip

test_fromstr_ip

test_fromstr_int

test_display_bool

test_display_number

test_display_vec_ips

test_fromstr_vec_ips

test_display_vec_ints

test_fromstr_vec_ints

test_fromstr_bad_input
}
