% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_pce_baseline.R
\name{do_pce_baseline}
\alias{do_pce_baseline}
\title{Help function to simulate trial}
\usage{
do_pce_baseline(
  n_trials,
  n_arms = 4,
  N1,
  N2,
  mu_0m,
  mu_6m,
  mu_12m,
  sg,
  alpha1,
  alpha,
  sel_scen,
  side,
  test,
  dropout,
  rr,
  bound
)
}
\arguments{
\item{n_trials}{number of simulation runs}

\item{n_arms}{number of arms (including control)}

\item{N1}{sample size stage 1}

\item{N2}{sample size stage 2}

\item{mu_0m}{Baseline value per arm (vector of length \code{n_arm})}

\item{mu_6m}{6-month value per arm (vector of length \code{n_arm})}

\item{mu_12m}{12-month value per arm (vector of length \code{n_arm})}

\item{sg}{covariance matrix between 6- and 12-month mean differences assumed equal across arms (matrix of dim 2x2)}

\item{alpha1}{significance level for dose selection (futility boundary)}

\item{alpha}{significance level for selected dose vs control comparison}

\item{sel_scen}{choose between two different options in case that in interim analysis low dose is promising, but median dose not: 0: do not continue with low dose or median dose; 1: continue with low and median doses}

\item{side}{TRUE/FALSE referring to the side for 1-side testing (if TRUE then lower = side)}

\item{test}{defines type of analysis: "t" calculates a t-test, "l" a linear model with baseline values as covariables, "w" Wilcoxon test of differences, and "w1" Wilcoxon test of follow-up values}

\item{dropout}{dropoutrate, between 0 and 1}

\item{rr}{responder rate for each dose (vector of length \code{n_arm}), which gives the proportion of patients with value 0 at follow-up}

\item{bound}{lower bound to define total responder in simulation study}
}
\value{
A vector consisting of summary measures for data simulation with n_trials repetitions: frequency selected for stage 2, conditional power, power, disunctive power, power of MA1, power of MA2, power of ma1a, concordances
}
\description{
Help function to simulate trial
}
\details{
eWHORM simulations
}
\author{
Marta Bofill Roig, Sonja Zehetmayer
}
\keyword{internal}
