% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{doc_attr_no_params}
\alias{doc_attr_no_params}
\title{No-param doc}
\source{
Generated by miniextendr from Rust fn \code{doc_attr_no_params}
}
\usage{
doc_attr_no_params()
}
\description{
A function with custom doc and no params.
}
