% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{doc_attr_basic}
\alias{doc_attr_basic}
\alias{doc_attr_no_params}
\title{Custom title from doc attr}
\usage{
doc_attr_basic(x)

doc_attr_no_params()
}
\arguments{
\item{x}{A numeric value.}
}
\value{
The value doubled.
}
\description{
This description was set via the doc attribute.

A function with custom doc and no params.
}
