% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{dog_breed}
\alias{dog_breed}
\title{The \code{dog_breed()} generic}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{dog_breed()} is an S7 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{S7Dog}}
}
}
