% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redivis-metadata.R
\name{.fetch_comps_biblio_table}
\alias{.fetch_comps_biblio_table}
\title{Fetch Competition Bibliography Table}
\usage{
.fetch_comps_biblio_table()
}
\value{
A tibble containing filtered comps biblio information.
}
\description{
Retrieves the comps_biblio table from Redivis user("datapages")$dataset("irw_meta").
Only fetches new data if the dataset version tag has changed.
This version filters out any tables that do not exist in the IRW competitions database.
}
\keyword{internal}
