% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redivis-metadata.R
\name{.fetch_metadata_table}
\alias{.fetch_metadata_table}
\title{Fetch Metadata Table}
\usage{
.fetch_metadata_table()
}
\value{
A cached or newly fetched tibble containing metadata information.
}
\description{
Retrieves the metadata table from Redivis user("datapages")$dataset("irw_meta")$table("metadata").
Only fetches new data if the table version tag has changed.
Rows are filtered to tables that currently exist in the IRW production datasets.
}
\keyword{internal}
