% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redivis-connection.R, R/redivis-metadata.R
\name{.get_irw_itemtext_dataset}
\alias{.get_irw_itemtext_dataset}
\title{Access the IRW item text dataset object}
\usage{
.get_irw_itemtext_dataset()

.get_irw_itemtext_dataset()
}
\value{
A Redivis dataset object.

A Redivis dataset object.
}
\description{
Returns the Redivis dataset object for IRW item text metadata,
and ensures metadata is loaded via \code{get()}.

Returns the Redivis dataset object for IRW item text metadata,
and ensures metadata is loaded via \code{get()}.
}
\keyword{internal}
