% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dots_pos_s3_collect}
\alias{dots_pos_s3_collect}
\title{Report a count with a formal after the dots, via S3 dispatch}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Report a count with a formal after the dots, via S3 dispatch
}
