% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{DoublerEnv}
\alias{DoublerEnv}
\alias{DoublerEnv$new}
\alias{DoublerEnv$value}
\alias{$.DoublerEnv}
\alias{[[.DoublerEnv}
\alias{DoublerEnv$Doubler$doubled}
\alias{DoublerEnv$Doubler$duplicate}
\alias{DoublerEnv$Doubler$spawn}
\title{DoublerEnv Class}
\usage{
DoublerEnv

\method{$}{DoublerEnv}(self, name)

\method{[[}{DoublerEnv}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Create a new DoublerEnv.

\describe{
\item{\code{v}}{Initial value.}
}

Get the stored value.
}
