% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{download_data}
\alias{download_data}
\title{Download a Dataset}
\usage{
download_data(table, path = NULL, overwrite = FALSE)
}
\arguments{
\item{table}{A Redivis table object from which data should be downloaded.
Ensure that this is a valid Redivis table with download capabilities.}

\item{path}{A string specifying the file path where the data should be saved.
If \code{NULL}, the dataset will be saved in the current working directory with the table's name as the file name.}

\item{overwrite}{A logical indicating whether to overwrite an existing file at the specified path.
Default is \code{FALSE}.}
}
\value{
A message confirming the file download location.
}
\description{
This function downloads a specified dataset from the IRW datasource.
}
