% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{drop}
\alias{drop}
\title{Drop Redundant Extent Information}
\usage{
\S4method{drop}{fm}(x)
}
\arguments{
\item{x}{a FlashR matrix.}
}
\value{
a FlashR matrix or vector.
}
\description{
Delete the dimensions of a FlashR matrix which have only one level.
}
\details{
If the input matrix has only one row or one column, it works the same as
fm.as.vector. Otherwise, it returns the original matrix.
}
\examples{
mat <- fm.runif.matrix(100, 1)
vec <- drop(mat)
}

