% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{drop_message_on_success}
\alias{drop_message_on_success}
\title{drop_message_on_success}
\usage{
drop_message_on_success()
}
\description{
Test that MsgOnDrop prints its message on normal function return.
}
