% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{drop_on_panic_with_move}
\alias{drop_on_panic_with_move}
\title{drop_on_panic_with_move}
\usage{
drop_on_panic_with_move()
}
\description{
Test that MsgOnDrop runs its destructor when an Rf_error (R longjmp) occurs.
}
