% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guinsoodb.R
\docType{package}
\name{guinsoodb-package}
\alias{guinsoodb-package}
\title{GuinsooDB client package for R}
\description{
R client package for GuinsooDB: an embeddable SQL OLAP Database Management System.
}
\seealso{
\code{\link[=guinsoodb]{guinsoodb()}} for connection instructions.

\url{https://guinsoodb.org/} for the project website.
}
