% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\docType{class}
\name{guinsoodb_connection-class}
\alias{guinsoodb_connection-class}
\alias{guinsoodb_connection}
\alias{show,guinsoodb_connection-method}
\alias{dbIsValid,guinsoodb_connection-method}
\alias{dbSendQuery,guinsoodb_connection,character-method}
\alias{dbDataType,guinsoodb_connection-method}
\alias{dbWriteTable,guinsoodb_connection,character,data.frame-method}
\alias{dbListTables,guinsoodb_connection-method}
\alias{dbExistsTable,guinsoodb_connection,character-method}
\alias{dbListFields,guinsoodb_connection,character-method}
\alias{dbRemoveTable,guinsoodb_connection,character-method}
\alias{dbGetInfo,guinsoodb_connection-method}
\alias{dbBegin,guinsoodb_connection-method}
\alias{dbCommit,guinsoodb_connection-method}
\alias{dbRollback,guinsoodb_connection-method}
\title{GuinsooDB connection class}
\usage{
\S4method{show}{guinsoodb_connection}(object)

\S4method{dbIsValid}{guinsoodb_connection}(dbObj, ...)

\S4method{dbSendQuery}{guinsoodb_connection,character}(conn, statement, params = NULL, ...)

\S4method{dbDataType}{guinsoodb_connection}(dbObj, obj, ...)

\S4method{dbWriteTable}{guinsoodb_connection,character,data.frame}(
  conn,
  name,
  value,
  row.names = FALSE,
  overwrite = FALSE,
  append = FALSE,
  field.types = NULL,
  temporary = FALSE,
  ...
)

\S4method{dbListTables}{guinsoodb_connection}(conn, ...)

\S4method{dbExistsTable}{guinsoodb_connection,character}(conn, name, ...)

\S4method{dbListFields}{guinsoodb_connection,character}(conn, name, ...)

\S4method{dbRemoveTable}{guinsoodb_connection,character}(conn, name, ...)

\S4method{dbGetInfo}{guinsoodb_connection}(dbObj, ...)

\S4method{dbBegin}{guinsoodb_connection}(conn, ...)

\S4method{dbCommit}{guinsoodb_connection}(conn, ...)

\S4method{dbRollback}{guinsoodb_connection}(conn, ...)
}
\arguments{
\item{object}{Any R object}

\item{dbObj}{An object inheriting from \linkS4class{DBIObject},
i.e. \linkS4class{DBIDriver}, \linkS4class{DBIConnection},
or a \linkS4class{DBIResult}}

\item{...}{Other arguments to methods.}

\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{statement}{a character string containing SQL.}

\item{params}{A list of bindings, named or unnamed.}

\item{obj}{An R object whose SQL type we want to determine.}

\item{name}{A character string specifying the unquoted DBMS table name,
or the result of a call to \code{\link[DBI:dbQuoteIdentifier]{dbQuoteIdentifier()}}.}

\item{value}{a \link{data.frame} (or coercible to data.frame).}

\item{row.names}{Whether the row.names of the data.frame should be preserved}

\item{overwrite}{If a table with the given name already exists, should it be overwritten?}

\item{append}{If a table with the given name already exists, just try to append the passed data to it}

\item{field.types}{Override the auto-generated SQL types}

\item{temporary}{Should the created table be temporary?}
}
\description{
Implements \linkS4class{DBIConnection}.
}
\keyword{internal}
