% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R, R/Driver.R
\docType{class}
\name{guinsoodb_driver-class}
\alias{guinsoodb_driver-class}
\alias{guinsoodb_driver}
\alias{show,guinsoodb_driver-method}
\alias{dbDataType,guinsoodb_driver-method}
\alias{dbIsValid,guinsoodb_driver-method}
\alias{dbGetInfo,guinsoodb_driver-method}
\title{GuinsooDB driver class}
\usage{
\S4method{show}{guinsoodb_driver}(object)

\S4method{dbDataType}{guinsoodb_driver}(dbObj, obj, ...)

\S4method{dbIsValid}{guinsoodb_driver}(dbObj, ...)

\S4method{dbGetInfo}{guinsoodb_driver}(dbObj, ...)
}
\arguments{
\item{object}{Any R object}

\item{dbObj}{An object inheriting from \linkS4class{DBIObject},
i.e. \linkS4class{DBIDriver}, \linkS4class{DBIConnection},
or a \linkS4class{DBIResult}}

\item{...}{Other arguments to methods.}
}
\description{
Implements \linkS4class{DBIDriver}.
}
\keyword{internal}
