% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print__duckdb_explain.R
\docType{class}
\name{duckdb_explain-class}
\alias{duckdb_explain-class}
\alias{duckdb_explain}
\alias{print.duckdb_explain}
\title{DuckDB EXPLAIN query tree}
\description{
DuckDB EXPLAIN query tree
}
