% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrait.R
\name{duckdb_get_substrait}
\alias{duckdb_get_substrait}
\title{Get the Substrait plan for a SQL query
Transforms a SQL query into a raw vector containing the serialized Substrait query blob}
\usage{
duckdb_get_substrait(conn, query)
}
\arguments{
\item{conn}{A DuckDB connection, created by \code{dbConnect()}.}

\item{query}{The query string in SQL}
}
\value{
A raw vector containing the substrait protobuf blob
}
\description{
Get the Substrait plan for a SQL query
Transforms a SQL query into a raw vector containing the serialized Substrait query blob
}
