% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrait.R
\name{duckdb_prepare_substrait}
\alias{duckdb_prepare_substrait}
\title{Query DuckDB using Substrait
Method for interpreting a Substrait BLOB plan as a DuckDB Query Plan
It interprets and executes the query.}
\usage{
duckdb_prepare_substrait(conn, query, arrow = FALSE)
}
\arguments{
\item{conn}{A DuckDB connection, created by \code{dbConnect()}.}

\item{query}{The Protobuf-encoded Substrait Query Plan. Qack!}

\item{arrow}{Whether the result should be in Arrow format}
}
\value{
A DuckDB Query Result
}
\description{
Query DuckDB using Substrait
Method for interpreting a Substrait BLOB plan as a DuckDB Query Plan
It interprets and executes the query.
}
