% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrait.R
\name{duckdb_prepare_substrait_json}
\alias{duckdb_prepare_substrait_json}
\title{Query DuckDB using Substrait
Method for interpreting a Substrait JSON plan as a DuckDB Query Plan
It interprets and executes the query.}
\usage{
duckdb_prepare_substrait_json(conn, json, arrow = FALSE)
}
\arguments{
\item{conn}{A DuckDB connection, created by \code{dbConnect()}.}

\item{json}{The Json Query Plan. Qack!}

\item{arrow}{Whether the result should be in Arrow format}
}
\value{
A DuckDB Query Result
}
\description{
Query DuckDB using Substrait
Method for interpreting a Substrait JSON plan as a DuckDB Query Plan
It interprets and executes the query.
}
