% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{guinsoodb_register}
\alias{guinsoodb_register}
\alias{guinsoodb_unregister}
\title{Register a data frame as a virtual table}
\usage{
guinsoodb_register(conn, name, df)

guinsoodb_unregister(conn, name)
}
\arguments{
\item{conn}{A GuinsooDB connection, created by \code{dbConnect()}.}

\item{name}{The name for the virtual table that is registered or unregistered}

\item{df}{A \code{data.frame} with the data for the virtual table}
}
\value{
These functions are called for their side effect.
}
\description{
\code{guinsoodb_register()} registers a data frame as a virtual table (view)
in a GuinsooDB connection.
No data is copied.
}
\details{
\code{guinsoodb_unregister()} unregisters a previously registered data frame.
}
\examples{
con <- dbConnect(guinsoodb())

data <- data.frame(a = 1:3, b = letters[1:3])

guinsoodb_register(con, "data", data)
dbReadTable(con, "data")

guinsoodb_unregister(con, "data")
try(dbReadTable(con, "data"))

dbDisconnect(con)
}
