% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{duckdb_register_parquet}
\alias{duckdb_register_parquet}
\alias{duckdb_unregister_parquet}
\title{Register Parquet file(s) as a virtual table}
\usage{
duckdb_register_parquet(
  conn,
  name,
  path,
  binary_as_string = FALSE,
  replace = FALSE,
  temporary = FALSE
)

duckdb_unregister_parquet(conn, name, if_exists = FALSE)
}
\arguments{
\item{conn}{A DuckDB connection, created by \code{dbConnect()}.}

\item{name}{The name for the virtual table that is registered}

\item{path}{A (vector of) strings pointing to the files to be registered}

\item{binary_as_string}{Convert binary data to strings (boolean)}

\item{replace}{Replace view if it is already existing (boolean)}

\item{temporary}{Do not save view to the catalog permanently (boolean)}

\item{if_exists}{Do not throw error if no view with given name is found (boolean)}
}
\value{
These functions are called for their side effect.
}
\description{
\code{duckdb_register_parquet()} registers Parquet file(s) as a virtual table (view)
in a DuckDB connection.
No data is copied.
}
\details{
\code{duckdb_unregister_parquet()} drops a view from the catalog.
}
