% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\docType{class}
\name{guinsoodb_result-class}
\alias{guinsoodb_result-class}
\alias{guinsoodb_result}
\alias{show,guinsoodb_result-method}
\alias{dbClearResult,guinsoodb_result-method}
\alias{dbFetch,guinsoodb_result-method}
\alias{dbHasCompleted,guinsoodb_result-method}
\alias{dbGetInfo,guinsoodb_result-method}
\alias{dbIsValid,guinsoodb_result-method}
\alias{dbGetStatement,guinsoodb_result-method}
\alias{dbColumnInfo,guinsoodb_result-method}
\alias{dbGetRowCount,guinsoodb_result-method}
\alias{dbGetRowsAffected,guinsoodb_result-method}
\alias{dbBind,guinsoodb_result-method}
\title{GuinsooDB Result Set}
\usage{
\S4method{show}{guinsoodb_result}(object)

\S4method{dbClearResult}{guinsoodb_result}(res, ...)

\S4method{dbFetch}{guinsoodb_result}(res, n = -1, ...)

\S4method{dbHasCompleted}{guinsoodb_result}(res, ...)

\S4method{dbGetInfo}{guinsoodb_result}(dbObj, ...)

\S4method{dbIsValid}{guinsoodb_result}(dbObj, ...)

\S4method{dbGetStatement}{guinsoodb_result}(res, ...)

\S4method{dbColumnInfo}{guinsoodb_result}(res, ...)

\S4method{dbGetRowCount}{guinsoodb_result}(res, ...)

\S4method{dbGetRowsAffected}{guinsoodb_result}(res, ...)

\S4method{dbBind}{guinsoodb_result}(res, params, ...)
}
\arguments{
\item{object}{Any R object}

\item{res}{An object inheriting from \linkS4class{DBIResult}.}

\item{...}{Other arguments passed on to methods.}

\item{n}{maximum number of records to retrieve per fetch. Use \code{n = -1}
or \code{n = Inf}
to retrieve all pending records.  Some implementations may recognize other
special values.}

\item{dbObj}{An object inheriting from \linkS4class{DBIObject},
i.e. \linkS4class{DBIDriver}, \linkS4class{DBIConnection},
or a \linkS4class{DBIResult}}

\item{params}{A list of bindings, named or unnamed.}
}
\description{
Methods for accessing result sets for queries on GuinsooDB connections.
Implements \linkS4class{DBIResult}.
}
\keyword{internal}
