% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrsa.r
\docType{data}
\name{dvix}
\alias{dvix}
\title{Daily VIX Close}
\format{
An \code{xts} object with 9,069 observations and 1 columns. 
The data run from January, 1990 through November, 2025.
The columns are defined as follows:
\describe{
 \item{\code{VIX}}{The closing value of the VIX index. 
 From January 2004 onward, the data are from the modern definition of the VIX index.
 The data from before 2004 are sourced from the back-computed data archive on CBOE.}
}
}
\source{
CBOE. See 
\url{https://www.cboe.com/tradable_products/vix/vix_historical_data}.
}
\usage{
dvix
}
\description{
The daily closing value of the CBOE VIX index.
}
\examples{
\dontrun{
data(dvix)
str(dvix)
}
}
\author{
Steven E. Pav \email{steven@sharperat.io}
}
\keyword{datasets}
