% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseflow.R
\name{eckhardt_filter}
\alias{eckhardt_filter}
\title{Eckhardt recursive digital filter for baseflow separation}
\usage{
eckhardt_filter(
  Q,
  BFImax = pkg_env$eckhardt_bfimax,
  a = pkg_env$eckhardt_alpha
)
}
\arguments{
\item{Q}{Numeric vector of daily streamflow.}

\item{BFImax}{Maximum baseflow index (default from config).}

\item{a}{Recession constant (default from config).}
}
\value{
Numeric vector of baseflow values.
}
\description{
Eckhardt recursive digital filter for baseflow separation
}
\keyword{internal}
