% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{edge_in_txpath}
\alias{edge_in_txpath}
\title{Check if edge is in transcript path}
\usage{
edge_in_txpath(g, trans, e)
}
\description{
Check if edge is in transcript path
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_edge_names(g)
trans <- grase::transcripts_from_igraph(g)
eid <- as.integer(igraph::E(g))[1]
grase::edge_in_txpath(g, trans, eid)
}
}
