% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{either_dbl_or_vec}
\alias{either_dbl_or_vec}
\title{either_dbl_or_vec}
\usage{
either_dbl_or_vec(value)
}
\arguments{
\item{value}{Either a double or an integer vector from R.}
}
\description{
Test dispatching an Either<f64, Vec\if{html}{\out{<i32>}}> and returning a tagged string.
}
