% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{either_grade}
\alias{either_grade}
\title{Describe a grade name or number through an S3 trait method}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Describe a grade name or number through an S3 trait method
}
