% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{either_grades}
\alias{either_grades}
\title{Describe grade names or a number through an S3 trait method}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Describe grade names or a number through an S3 trait method
}
