% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{either_int_or_str}
\alias{either_int_or_str}
\title{either_int_or_str}
\usage{
either_int_or_str(value)
}
\arguments{
\item{value}{Either an integer or a string from R.}
}
\description{
Test dispatching an Either<i32, String> and returning a tagged string.
}
