% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{either_is_left}
\alias{either_is_left}
\title{either_is_left}
\usage{
either_is_left(value)
}
\arguments{
\item{value}{Either an integer or a string from R.}
}
\description{
Test whether an Either value was parsed as Left (integer).
}
