% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{either_is_right}
\alias{either_is_right}
\title{either_is_right}
\usage{
either_is_right(value)
}
\arguments{
\item{value}{Either an integer or a string from R.}
}
\description{
Test whether an Either value was parsed as Right (string).
}
