% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{either_make_left}
\alias{either_make_left}
\title{either_make_left}
\usage{
either_make_left(n)
}
\arguments{
\item{n}{Integer value for the Left variant.}
}
\description{
Test creating a Left(i32) variant of Either.
}
