% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{either_make_right}
\alias{either_make_right}
\title{either_make_right}
\usage{
either_make_right(s)
}
\arguments{
\item{s}{String value for the Right variant.}
}
\description{
Test creating a Right(String) variant of Either.
}
