% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{either_nested}
\alias{either_nested}
\title{either_nested}
\usage{
either_nested(value)
}
\arguments{
\item{value}{Nested Either value from R.}
}
\description{
Test nested Either dispatch: Either<bool, Either<i32, String>>.
}
