% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{class}
\name{EitherRouteS4}
\alias{EitherRouteS4}
\alias{EitherRouteS4-s4_plan}
\alias{s4_plan,EitherRouteS4-method}
\title{EitherRouteS4 S4 Class}
\usage{
EitherRouteS4()
}
\description{
A route name or a data frame of doses, and an omittable level name or
number.
}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to Rust \code{EitherRouteS4} struct}
}}

