% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{class}
\name{EitherRoutesS4}
\alias{EitherRoutesS4}
\alias{EitherRoutesS4-s4_plan_routes}
\alias{s4_plan_routes,EitherRoutesS4-method}
\title{EitherRoutesS4 S4 Class}
\usage{
EitherRoutesS4()
}
\description{
Route names or a data frame of doses, and omittable tier names or a number.
}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to Rust \code{EitherRoutesS4} struct}
}}

