% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-strings.R
\name{encrypt_string}
\alias{encrypt_string}
\title{Encrypt String}
\usage{
encrypt_string(string, secret_key = NA)
}
\arguments{
\item{string}{string}

\item{secret_key}{secret}
}
\value{
encrypted string
}
\description{
Encrypt String
}
