% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{endpoint}
\alias{endpoint}
\title{Create a custom endpoint}
\usage{
endpoint(model, preprocess = NULL, predict = NULL, postprocess = NULL)
}
\description{
This creates an API endpoint. Model specific preprocessing and postprocessing
are added for several model types
}

