% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pols503-package.R
\name{equicor}
\alias{equicor}
\title{Create an equicorrelated correlation matrix.}
\usage{
equicor(n, rho)
}
\arguments{
\item{n}{Size of the matrix.}

\item{rho}{The off diagonal correlations}
}
\value{
A correlation matrix.
}
\description{
A correlation matrix with all off diagonal correlations equal to \eqn{rho}.
}
\examples{
n <- 3
rho <- 0.25
equicor(n, rho)
}

