% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{error_in_r_i32_ok}
\alias{error_in_r_i32_ok}
\title{Function returning i32 -- tests numeric return}
\source{
Generated by miniextendr from Rust fn \code{error_in_r_i32_ok}
}
\usage{
error_in_r_i32_ok()
}
\description{
Function returning i32 -- tests numeric return
}
