% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{error_in_r_normal}
\alias{error_in_r_normal}
\title{Normal function (no error) -- works fine}
\source{
Generated by miniextendr from Rust fn \code{error_in_r_normal}
}
\usage{
error_in_r_normal()
}
\description{
Normal function (no error) -- works fine
}
