% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{error_in_r_option_none}
\alias{error_in_r_option_none}
\title{error_in_r_option_none}
\usage{
error_in_r_option_none()
}
\description{
Standalone function that returns Option::None (unit) -- becomes error value.
}
