% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{error_in_r_option_some}
\alias{error_in_r_option_some}
\title{error_in_r_option_some}
\usage{
error_in_r_option_some()
}
\description{
Standalone function that returns Option::Some(()) -- returns NULL.
}
