% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{error_in_r_panic}
\alias{error_in_r_panic}
\title{Standalone function that panics -- error_in_r catches this}
\source{
Generated by miniextendr from Rust fn \code{error_in_r_panic}
}
\usage{
error_in_r_panic()
}
\description{
Standalone function that panics -- error_in_r catches this
}
