% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{error_in_r_panic_custom}
\alias{error_in_r_panic_custom}
\title{error_in_r_panic_custom}
\usage{
error_in_r_panic_custom(msg)
}
\arguments{
\item{msg}{Custom panic message.}
}
\description{
Panic with custom message.
}
