% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{error_in_r_result_err}
\alias{error_in_r_result_err}
\title{error_in_r_result_err}
\usage{
error_in_r_result_err()
}
\description{
Standalone function that returns Result::Err -- becomes error value.
}
