% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{error_in_r_result_ok}
\alias{error_in_r_result_ok}
\title{Standalone function that returns Result::Ok -- works normally}
\source{
Generated by miniextendr from Rust fn \code{error_in_r_result_ok}
}
\usage{
error_in_r_result_ok()
}
\description{
Standalone function that returns Result::Ok -- works normally
}
