% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ErrorInRCounter}
\alias{ErrorInRCounter}
\alias{ErrorInRCounter$new}
\alias{ErrorInRCounter$get}
\alias{ErrorInRCounter$inc}
\alias{ErrorInRCounter$panic_method}
\alias{ErrorInRCounter$failing_method}
\alias{$.ErrorInRCounter}
\alias{[[.ErrorInRCounter}
\title{ErrorInRCounter Class}
\usage{
ErrorInRCounter

\method{$}{ErrorInRCounter}(self, name)

\method{[[}{ErrorInRCounter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Create a new counter.

Get the current value -- should always succeed.

Increment -- mutable method, should succeed and allow chaining.

Deliberately panic in a method.

Return Result::Err in a method.
}
