% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exontest_functions.R
\name{estimate_phi_for_comparison}
\alias{estimate_phi_for_comparison}
\title{Estimate phi (overdispersion) for all events in one comparison using glmmTMB}
\usage{
estimate_phi_for_comparison(sc, comp_name, outdir, mc_cores = 1L)
}
\arguments{
\item{sc}{split counts data frame for one comparison}

\item{comp_name}{comparison label used for log file names}

\item{outdir}{output directory for progress and error logs}

\item{mc_cores}{number of parallel cores}
}
\description{
Estimate phi (overdispersion) for all events in one comparison using glmmTMB
}
