% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial-helpers.R
\name{example_dge_data}
\alias{example_dge_data}
\title{Generates example data for three categories of differential expression}
\usage{
example_dge_data(seed = 300)
}
\description{
Generates example data for three categories of differential expression
}
