% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch.R
\name{execute}
\alias{execute}
\title{Run remote commands}
\usage{
execute(connection, command)
}
\arguments{
\item{connection}{a connection object as returned by \code{\link{get_connection}}}

\item{command}{a valid command for the remote system's shell environment}
}
\description{
Run remote commands
}
\examples{
\dontrun{
connection <- get_connection(host = "some.remote.system", user = "you")
execute(connection, "ps -ef | grep R /*")
}
}
