% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exm.r
\name{exm_lookup}
\alias{exm_lookup}
\title{Lookup the exm backing file for an object}
\usage{
exm_lookup(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
Lookup the exm backing file for an object
}
