% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exm.r
\name{exm_madvise}
\alias{exm_madvise}
\title{Lookup the exm backing file for an object}
\usage{
exm_madvise(object, advice = c("normal", "random", "sequential"))
}
\arguments{
\item{object}{Any R object}

\item{advice}{Mapping type, one of "normal", "random", or "sequential"}
}
\value{
Integer return code from the OS-level madvise call, zero means success
}
\description{
Lookup the exm backing file for an object
}
