% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exm.r
\name{exm_path}
\alias{exm_path}
\title{Set or retrieve the exm data file path}
\usage{
exm_path(path)
}
\arguments{
\item{path}{If missing, return the current data path. Otherwise set the
data path to the specified value.}
}
\description{
Set or retrieve the exm data file path
}
