% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exm_set_pattern}
\alias{exm_set_pattern}
\title{Set the exm temp file pattern.}
\usage{
exm_set_pattern(pattern = "fm_XXXXXX")
}
\arguments{
\item{pattern}{A valid tmp file pattern containing the string "XXXXXX"}
}
\description{
Set the exm temp file pattern.
}

