% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exm.r
\name{exm_threshold}
\alias{exm_threshold}
\title{Retrieve and/or change the threshold beyond which allocations are
file-backed.}
\usage{
exm_threshold(nbytes = 0)
}
\arguments{
\item{nbytes}{the new threshold (in bytes)}
}
\description{
Retrieve and/or change the threshold beyond which allocations are
file-backed.
}
\examples{
# Add another 1000 bytes to the threshold.
\dontrun{
exm_threshold(exm_threshold()+1000)
}
}
\seealso{
\code{\link{exm_template}}
}
