% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{experimental_feature}
\alias{experimental_feature}
\title{experimental_feature}
\usage{
experimental_feature(x)
}
\arguments{
\item{x}{Numeric input value.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} An experimental function that may change.
}
\keyword{internal}
