% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{export_control_internal}
\alias{export_control_internal}
\title{export_control_internal}
\usage{
export_control_internal()
}
\description{
An internal function -- callable but not exported, with @keywords internal.
}
\keyword{internal}
