% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{export_control_normal}
\alias{export_control_normal}
\title{A normal exported function for comparison}
\source{
Generated by miniextendr from Rust fn \code{export_control_normal}
}
\usage{
export_control_normal()
}
\description{
A normal exported function for comparison
}
