% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational.R
\name{expr_constant}
\alias{expr_constant}
\title{Create a constant expression}
\usage{
expr_constant(val)
}
\arguments{
\item{val}{the constant value}
}
\value{
a constant expression
}
\description{
Create a constant expression
}
\examples{
const_int_expr <- expr_constant(42)
const_str_expr <- expr_constant("Hello, World")
}
