% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational.R
\name{expr_function}
\alias{expr_function}
\title{Create a function call expression}
\usage{
expr_function(name, args)
}
\arguments{
\item{name}{the function name}

\item{args}{the a list of expressions for the function arguments}
}
\value{
a function call expression
}
\description{
Create a function call expression
}
\examples{
call_expr <- expr_function("ABS", list(expr_constant(-42)))
}
