% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational.R
\name{expr_reference}
\alias{expr_reference}
\title{Create a column reference expression}
\usage{
expr_reference(ref)
}
\arguments{
\item{ref}{the column name to be referenced}
}
\value{
a column reference expression
}
\description{
Create a column reference expression
}
\examples{
col_ref_expr <- expr_reference("some_column_name")
}
