% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational.R
\name{expr_set_alias}
\alias{expr_set_alias}
\title{Set the alias for an expression}
\usage{
expr_set_alias(expr, alias)
}
\arguments{
\item{expr}{the expression}

\item{alias}{the alias}
}
\description{
NOTICE: This function is experimental and may change/disappear without warning in future versions
}
\examples{
expr_set_alias(expr_constant(42), "my_alias")
}
