% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational.R
\name{expr_tostring}
\alias{expr_tostring}
\title{Convert an expression to a string for debugging purposes}
\usage{
expr_tostring(expr)
}
\arguments{
\item{expr}{the expression}
}
\value{
a string representation of the expression
}
\description{
Convert an expression to a string for debugging purposes
}
\examples{
expr_str <- expr_tostring(expr_constant(42))
}
