% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{external_slice_sum}
\alias{external_slice_sum}
\title{external_slice_sum}
\usage{
external_slice_sum(data)
}
\arguments{
\item{data}{(no documentation available)}
}
\description{
Create an ExternalSlice, access elements via as_slice.
}
