% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ptr_pick_larger}
\alias{ptr_pick_larger}
\title{ptr_pick_larger}
\usage{
ptr_pick_larger(a, b)
}
\arguments{
\item{a}{An ExternalPtr wrapping a PtrIdentityTest.}

\item{b}{An ExternalPtr wrapping a PtrIdentityTest.}
}
\description{
Takes two ExternalPtrs, returns the one with the larger value.
The returned SEXP should be the same as the input's SEXP.
}
