% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{extptr_any_erased_is}
\alias{extptr_any_erased_is}
\title{extptr_any_erased_is}
\usage{
extptr_any_erased_is(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
Test type-erased downcast via ErasedExternalPtr.
}
