% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{extptr_any_into_inner}
\alias{extptr_any_into_inner}
\title{extptr_any_into_inner}
\usage{
extptr_any_into_inner(val)
}
\arguments{
\item{val}{(no documentation available)}
}
\description{
Test that into_inner recovers the value via Any::downcast.
}
