% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{extptr_any_wrong_type_is}
\alias{extptr_any_wrong_type_is}
\title{extptr_any_wrong_type_is}
\usage{
extptr_any_wrong_type_is(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
Test that wrong-type downcast returns None (not crash).
}
