% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{Extract}
\alias{Extract}
\alias{[,fm,ANY,ANY,ANY-method}
\alias{[,fm,ANY,missing,ANY-method}
\alias{[,fm,missing,ANY,ANY-method}
\alias{[,fmV,ANY,ANY,ANY-method}
\title{Extract Parts of an object}
\usage{
\S4method{[}{fm,ANY,missing,ANY}(x, i, j, drop = TRUE)

\S4method{[}{fm,missing,ANY,ANY}(x, i, j, drop = TRUE)

\S4method{[}{fm,ANY,ANY,ANY}(x, i, j, drop = TRUE)

\S4method{[}{fmV,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{object from which to extract elements.}

\item{i,j}{indices specifying elements to extract. Indices are integer
or numeric vectors.}

\item{drop}{for matrices. If \code{TRUE}, the result is coerced to a vector.}
}
\description{
Operators acting on FlashR vectors and matrices to extract parts of
the objects.
}
\details{
FlashR only supports extracting data from an object. It doesn't support
replace data in an object.
}
\examples{
mat <- fm.runif.matrix(100, 10)
mat[1:5,]
mat[,1:5]
mat[1:5,1:5]
}

