% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{Extremes}
\alias{Extremes}
\alias{pmax2}
\alias{pmax2,ANY,fm-method}
\alias{pmax2,ANY,fmV-method}
\alias{pmax2,fm,ANY-method}
\alias{pmax2,fm,fm-method}
\alias{pmax2,fm,fmV-method}
\alias{pmax2,fm,matrix-method}
\alias{pmax2,fmV,ANY-method}
\alias{pmax2,fmV,fm-method}
\alias{pmax2,fmV,fmV-method}
\alias{pmax2,matrix,fm-method}
\alias{pmin2,ANY,fm-method}
\alias{pmin2,ANY,fmV-method}
\alias{pmin2,fm,ANY-method}
\alias{pmin2,fm,fm-method}
\alias{pmin2,fm,fmV-method}
\alias{pmin2,fm,matrix-method}
\alias{pmin2,fmV,ANY-method}
\alias{pmin2,fmV,fm-method}
\alias{pmin2,fmV,fmV-method}
\alias{pmin2,matrix,fm-method}
\title{Maxima and Minima}
\usage{
pmax2(e1, e2)

\S4method{pmax2}{fm,fm}(e1, e2)

\S4method{pmax2}{fmV,fmV}(e1, e2)

\S4method{pmax2}{fm,fmV}(e1, e2)

\S4method{pmax2}{fmV,fm}(e1, e2)

\S4method{pmax2}{fm,matrix}(e1, e2)

\S4method{pmax2}{matrix,fm}(e1, e2)

\S4method{pmax2}{fm,ANY}(e1, e2)

\S4method{pmax2}{ANY,fm}(e1, e2)

\S4method{pmax2}{fmV,ANY}(e1, e2)

\S4method{pmax2}{ANY,fmV}(e1, e2)

\S4method{pmin2}{fm,fm}(e1, e2)

\S4method{pmin2}{fmV,fmV}(e1, e2)

\S4method{pmin2}{fm,fmV}(e1, e2)

\S4method{pmin2}{fmV,fm}(e1, e2)

\S4method{pmin2}{fm,matrix}(e1, e2)

\S4method{pmin2}{matrix,fm}(e1, e2)

\S4method{pmin2}{fm,ANY}(e1, e2)

\S4method{pmin2}{ANY,fm}(e1, e2)

\S4method{pmin2}{fmV,ANY}(e1, e2)

\S4method{pmin2}{ANY,fmV}(e1, e2)
}
\arguments{
\item{e1,e2}{One of the operands need to be a FlashMatrix object. If one operand
is a matrix and the other is a vector, we perform the arithmetic operation
on the vector and every column of the matrix. If one operand is a scalar,
we perform the operation on the scalar with every element in the matrix or
the vector.}
}
\description{
Returns the (parallel) maxima and minima of the input values.
}
\details{
pmax2 returns the maximum of two objects.
pmin2 returns the minimum of two objects.
}

