% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{factor_colors_with_na}
\alias{factor_colors_with_na}
\title{factor_colors_with_na}
\usage{
factor_colors_with_na(colors)
}
\arguments{
\item{colors}{A factor vector of Color values that may contain NA.}
}
\description{
Test handling NA values in a FactorOptionVec of Colors.
}
