% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{factor_count_colors}
\alias{factor_count_colors}
\title{factor_count_colors}
\usage{
factor_count_colors(colors)
}
\arguments{
\item{colors}{A factor vector of Color values.}
}
\description{
Test counting occurrences of each Color in a FactorVec.
}
