% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{factor_describe_priority}
\alias{factor_describe_priority}
\title{factor_describe_priority}
\usage{
factor_describe_priority(priority)
}
\arguments{
\item{priority}{A Priority factor value.}
}
\description{
Test describing a Priority factor with individually renamed variants.
}
