% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{factor_describe_status}
\alias{factor_describe_status}
\title{factor_describe_status}
\usage{
factor_describe_status(status)
}
\arguments{
\item{status}{A Status factor value.}
}
\description{
Test describing a Status factor with snake_case rename_all.
}
