% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{factor_get_color}
\alias{factor_get_color}
\title{factor_get_color}
\usage{
factor_get_color(name)
}
\arguments{
\item{name}{Character name of the color ("red", "green", or "blue").}
}
\description{
Test constructing a Color factor from a string name.
}
