% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{factor_outcome_levels}
\alias{factor_outcome_levels}
\title{factor_outcome_levels}
\usage{
factor_outcome_levels()
}
\description{
Test retrieving the Outcome factor level names via CHOICES.
}
