% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{factor_outcome_round_trip}
\alias{factor_outcome_round_trip}
\title{factor_outcome_round_trip}
\usage{
factor_outcome_round_trip(outcome)
}
\arguments{
\item{outcome}{An Outcome factor value.}
}
\description{
Test round-tripping an Outcome factor through R.
}
