% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{factor_outcome_stage_levels}
\alias{factor_outcome_stage_levels}
\title{factor_outcome_stage_levels}
\usage{
factor_outcome_stage_levels()
}
\description{
Test retrieving the OutcomeStage interaction levels via CHOICES.
}
