% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{factor_outcome_stage_round_trip}
\alias{factor_outcome_stage_round_trip}
\title{factor_outcome_stage_round_trip}
\usage{
factor_outcome_stage_round_trip(stage)
}
\arguments{
\item{stage}{An OutcomeStage factor value.}
}
\description{
Test round-tripping an OutcomeStage interaction factor through R.
}
