% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{factor_priority_levels}
\alias{factor_priority_levels}
\title{factor_priority_levels}
\usage{
factor_priority_levels()
}
\description{
Test retrieving the Priority factor level names via CHOICES.
}
