% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{failing_result}
\alias{failing_result}
\title{The \code{failing_result()} generic}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{failing_result()} is an S7 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{ErrorInRS7Gauge}}
}
}
