% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{fdefault_no_coerce_flag}
\alias{fdefault_no_coerce_flag}
\title{fdefault_no_coerce_flag}
\usage{
fdefault_no_coerce_flag(x)
}
\arguments{
\item{x}{Logical scalar.}
}
\description{
Same as \code{fdefault_coerce_flag} with \code{no_coerce}: requires a logical even
under \code{coerce-default}.
}
