% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{fdefault_no_strict_i64}
\alias{fdefault_no_strict_i64}
\title{fdefault_no_strict_i64}
\usage{
fdefault_no_strict_i64(x)
}
\arguments{
\item{x}{Integer-like scalar.}
}
\description{
Same as \code{fdefault_strict_i64} with \code{no_strict}: stays lenient even under
\code{strict-default}.
}
