% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{fdefault_no_worker_thread_name}
\alias{fdefault_no_worker_thread_name}
\title{fdefault_no_worker_thread_name}
\usage{
fdefault_no_worker_thread_name()
}
\description{
Same body as \code{fdefault_worker_thread_name} with \code{no_worker}: pinned to the
R main thread even under \code{worker-default}.
}
