% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{fdefault_worker_thread_name}
\alias{fdefault_worker_thread_name}
\title{fdefault_worker_thread_name}
\usage{
fdefault_worker_thread_name()
}
\description{
Name of the thread the body runs on under the build's default dispatch:
"miniextendr-worker" under \code{worker-default}, the R main thread otherwise.
}
