% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{FdefaultTraitTarget}
\alias{FdefaultTraitTarget}
\alias{FdefaultTraitTarget$new}
\alias{FdefaultTraitTarget$tag}
\alias{$.FdefaultTraitTarget}
\alias{[[.FdefaultTraitTarget}
\alias{FdefaultTraitTarget$FdefaultTraitDefaults$coerce_flag}
\alias{FdefaultTraitTarget$FdefaultTraitDefaults$no_coerce_flag}
\alias{FdefaultTraitTarget$FdefaultTraitDefaults$strict_i64}
\alias{FdefaultTraitTarget$FdefaultTraitDefaults$no_strict_i64}
\title{FdefaultTraitTarget Class}
\usage{
FdefaultTraitTarget

\method{$}{FdefaultTraitTarget}(self, name)

\method{[[}{FdefaultTraitTarget}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Creates the probe.

Returns the probe's tag (always 0).
}
