% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ffi_guard_catch_unwind_ok}
\alias{ffi_guard_catch_unwind_ok}
\title{ffi_guard_catch_unwind_ok}
\usage{
ffi_guard_catch_unwind_ok()
}
\description{
Test guarded_ffi_call with CatchUnwind on a non-panicking closure.
}
