% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ffi_guard_fallback_ok}
\alias{ffi_guard_fallback_ok}
\title{ffi_guard_fallback_ok}
\usage{
ffi_guard_fallback_ok()
}
\description{
Test guarded_ffi_call_with_fallback on a non-panicking closure.
}
