% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ffi_guard_fallback_panic}
\alias{ffi_guard_fallback_panic}
\title{ffi_guard_fallback_panic}
\usage{
ffi_guard_fallback_panic()
}
\description{
Test guarded_ffi_call_with_fallback on a panicking closure — returns fallback.
}
