% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fg.ase}
\alias{fg.ASE.igraph}
\alias{fg.ase}
\title{Spectral embedding}
\usage{
fg.ASE.igraph(fg, num.eigen, which = c("A, AcD, L, nL, nL_tau"),
  which.eigen = c("LM, LA, SM, SA"), c = 1, tau = 1, tol = 1e-12)
}
\arguments{
\item{fg}{The FlashGraphR object}

\item{num.eigen}{The number of eigenvalues/vectors required.}

\item{which}{The type of the embedding.}

\item{which.eigen}{Specify which eigenvalues/vectors to compute,
character constant with exactly two characters.}

\item{c}{The constant used in the AcD variant of embedding.}

\item{tau}{The value used in the regularized Laplacian embedding.}

\item{tol}{Numeric scalar. Stopping criterion: the relative accuracy
of the Ritz value is considered acceptable if its error is less than
tol times its estimated value. If this is set to zero then machine
precision is used.}
}
\value{
A named list with the following members:

values

   Numeric vector, the desired eigenvalues.

vectors

   Numeric matrix, the desired eigenvectors as columns. If complex=TRUE
  (the default for non-symmetric problems), then the matrix is complex.

options

  A named list with the supplied options and some information about
  the performed calculation, including an ARPACK exit code.
}
\description{
This computes spectral embedding of a given graph.
}
\details{
It supports five variants.

Adjacency matrix (A),

A - c * D (AcD), where c is a constant provided by a user.

Laplacian matrix (L), L = D - A,

Normalized Laplacian matrix (nL), nL = I - D^(-1/2) * A * D^(-1/2)

Regularized Laplacian matrix (nL_tau), nL_tau = D_tau^(-1/2) A D_tau^(-1/2)
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

