% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fg.betweenness}
\alias{fg.betweenness}
\title{Vertex betweenness centrality.}
\usage{
fg.betweenness(fg, vids = 0:(fg$vcount - 1))
}
\arguments{
\item{fg}{The FlashGraphR object.}

\item{vids}{A vector of vertex IDs. Default runs it on the entire
        graph.}
}
\value{
A vector with betweenness centrality values for all vertices
			with respect to `vids`.
}
\description{
The vertex betweenness centrality can be defined as the
number of geodesics (shortest paths) going through a vertex.
}
\examples{
fg <- fg.load.graph("edge_list.txt")
res <- fg.betweenness(fg, c(1,10))
}
\author{
Disa Mhembere <disa@jhu.edu>
}

