% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashgraph.R
\name{fg.cc}
\alias{fg.cc}
\alias{fg.clusters}
\title{Connected components of a graph}
\usage{
fg.clusters(graph, mode = c("weak", "strong"))
}
\arguments{
\item{graph}{The FlashGraph object}

\item{mode}{Character string, either "weak" or "strong". For directed
graphs "weak" implies weakly, "strong" strongly c
components to search. It is ignored for undirected graphs.}
}
\value{
A numeric vector that indicates the cluster id to which each vertex
        blongs to.
}
\description{
Compute all (weakly or strongly) connected components of a graph.
}
\details{
For an undirected graph, this function only computes connected components
and ignores the argument `mode'.

For strongly connected components, we use a customized version of the
algorithm described in the paper

Sungpack Hong, Nicole C. Rodia, Kunle Olukotun, On Fast Parallel Detection
of Strongly Connected Components (SCC) in Small-World Graphs, Proceedings
of the International Conference on High Performance Computing, Networking,
Storage and Analysis, 2013
}
\author{
Da Zheng <dzheng5@jhu.edu>
}
\references{
Sungpack Hong, Nicole C. Rodia, Kunle Olukotun, On Fast Parallel Detection
of Strongly Connected Components (SCC) in Small-World Graphs, Proceedings
of the International Conference on High Performance Computing, Networking,
Storage and Analysis, 2013
}

