% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fg.degree}
\alias{fg.degree}
\title{Degree of the vertices in a graph}
\usage{
fg.degree(graph, type = "both")
}
\arguments{
\item{graph}{The FlashGraphR object}

\item{type}{Character string. "out" for out-degree, "in" for in-degree,
"both" for the sum of the two. This argument is ignored for
undirected graphs.}
}
\value{
A numeric vector with the degree of each vertex in the graph.
}
\description{
Get the degree of vertices in a graph.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

