% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashgraph.R
\name{fg.diameter}
\alias{fg.diameter}
\title{Diameter estimation}
\usage{
fg.diameter(graph, directed = FALSE)
}
\arguments{
\item{graph}{The FlashGraph object}

\item{directed}{Indicates whether or not to respect the direction of edges
in a graph when traversing the graph. It is ignored for
undirected graphs.}
}
\value{
A single number
}
\description{
Estimate the diameter of a graph, the longest distance between two vertices
in a graph.
}
\details{
This implementation uses the double sweep method described in the paper
below to estimate the lower bound of the diameter.

Cl茅mence Magnienand Matthieu Latapy and Michel Habib: Fast computation of
empirically tight bounds for the diameter of massive graphs, Journal of
Experimental Algorithmics (JEA), 2009
}
\author{
Da Zheng <dzheng5@jhu.edu>
}
\references{
Cl茅mence Magnienand Matthieu Latapy and Michel Habib: Fast computation of
empirically tight bounds for the diameter of massive graphs, Journal of
Experimental Algorithmics (JEA), 2009
}

