% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fg.eigen}
\alias{fg.eigen}
\title{Eigensolver}
\usage{
fg.eigen(graph, which = "LM", nev = 1, ncv = 2)
}
\arguments{
\item{graph}{The FlashGraphR object}

\item{which}{Specify which eigenvalues/vectors to compute, character
             constant with exactly two characters.
Possible values for symmetric input matrices:
"LA' Compute 'nev' largest (algebraic) eigenvalues.
"SA" Compute "nev" smallest (algebraic) eigenvalues.
"LM" Compute `nev' largest (in magnitude) eigenvalues.
"SM" Compute `nev' smallest (in magnitude) eigenvalues.}

\item{nev}{Numeric scalar. The number of eigenvalues to be computed.}

\item{ncv}{Number of Lanczos vectors to be generated.}
}
\value{
A named list with the following members:
        values: Numeric vector, the desired eigenvalues.
        vectors: Numeric matrix, the desired eigenvectors as columns.
}
\description{
Compute eigenvalues/vectors of the adjacency matrix of an undirected graph.
}
\details{
This implements Implicitly Restart Lanczos method described in the paper

D. Calvetti and L. Reichel and D. C. Sorensen: An Implicitly Restarted
Lanczos method for Large Symmetric Eigenvalue problems, Jounral of ETNA,
1994.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}
\references{
D. Calvetti and L. Reichel and D. C. Sorensen: An Implicitly Restarted
Lanczos method for Large Symmetric Eigenvalue problems, Jounral of ETNA,
1994.
}

