% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashgraph.R
\name{fg.export.graph}
\alias{fg.export.graph}
\title{Export graph image.}
\usage{
fg.export.graph(graph, graph.file, index.file)
}
\arguments{
\item{graph}{The FlashGraphR object}

\item{graph.file}{The graph file in the local filesystem to which
the adjacency lists of the graph is exported to.}

\item{index.file}{The index file in the local filesystem to which
the index of the graph is exported to.}
}
\value{
true if the graph is exported to the local filesystem correctly;
false, otherwise.
}
\description{
This function exports a graph image in FlashGraphR to the local filesystem.
}

