% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fg.local.scan}
\alias{fg.local.scan}
\alias{fg.topK.scan}
\title{Locality statistic}
\usage{
fg.topK.scan(graph, order = 1, K = 1)

fg.local.scan(graph, order = 1)
}
\arguments{
\item{graph}{The FlashGraphR object}

\item{order}{An integer scalar, the size of the local neighborhood for
each vertex. Should be non-negative.}
}
\value{
A numeric vector that contains locality statistic of each vertex.
}
\description{
Compute locality statistic of vertices in a graph.
}
\details{
`fg.topK.scan' finds the top K vertices with the largest locality
statistics in a graph and computes their locality statistic.

`fg.local.scan' computes locality statistic of each vertex in a graph.

Locality statistic is defined as the number of edges in the neighborhood
of a vertex.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}
\references{
C.E. Priebe, J.M. Conroy, D.J. Marchette, and Y. Park, Scan Statistics on
Enron Graphs," Computational and Mathematical Organization Theory, 2005.
}

