% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fg.multiply}
\alias{fg.multiply}
\title{Sparse matrix vector multiplication}
\usage{
fg.multiply(graph, vec, transpose = FALSE)
}
\arguments{
\item{graph}{The FlashGraphR object}

\item{vec}{A numueric vector}

\item{transpose}{Indicate whether or not to multiply with the transpose of
the matrix. It is ignored by an undirected graph.}
}
\value{
A numeric vector
}
\description{
Multiply a sparse matrix and dense vector.
}
\details{
Note that the sparse matrix is represented by a graph. A symmetric matrix
is represented by an undirected graph and an asymmetric matrix is
represented by a directed graph. For an asymmetric matrix, we can multiply
the vector with the transpose of the matrix without really transposing
the matrix. Right now it only supports multiplication on the adjacency
matrix of the graph.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

