% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fg.pagerank}
\alias{fg.page.rank}
\alias{fg.pagerank}
\title{PageRank}
\usage{
fg.page.rank(graph, no.iters = 1000, damping = 0.85)
}
\arguments{
\item{graph}{The FlashGraphR object}

\item{no.iters}{The number of iterations}

\item{damping}{The damping factor ('d' in the original p)}
}
\value{
A numeric vector that contains PageRank values of each vertex.
}
\description{
Compute the Google PageRank for a graph.
}
\details{
This implementation computes PageRank values in the original PageRank
paper below and does not normalize PageRank values in each iteration.

Sergey Brin and Larry Page: The Anatomy of a Large-Scale
Hypertextual Web Search Engine. Proceedings of the 7th World-Wide
Web Conference, Brisbane, Australia, April 1998.

To improve performance, a vertex only sends the difference of its PageRank
value between the previous iteration and the current iteration to its
neighbors in each iteration. If the difference is smaller than a threshold,
a vertex does not send the difference to its neighbors. The algorithm
converges if all vertices stop sending messages.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}
\references{
Sergey Brin and Larry Page: The Anatomy of a Large-Scale
Hypertextual Web Search Engine. Proceedings of the 7th World-Wide
Web Conference, Brisbane, Australia, April 1998.
}

