% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashgraph.R
\name{fg.sem.kmeans}
\alias{fg.sem.kmeans}
\title{Perform semi-extenal memory k-means clustering on a data matrix in FG format.}
\usage{
fg.sem.kmeans(mat, k, max.iters = 10, init = c("random", "forgy",
  "kmeanspp"), tol = -1)
}
\arguments{
\item{mat}{A numeric matrix of data in FG format.}

\item{k}{The number of clusters.}

\item{max.iters}{The maximum number of iterations allowed.}

\item{init}{The form of initialization to use when the algorithm begins.
The default is "forgy". For a desciption of each see:
http://en.wikipedia.org/wiki/K-means_clustering#Initialization_methods.}

\item{tol}{The tolerance for convergence. Between 0 and 1 and is the minimum fraction
of cluster changes necessary to cause non-convergence. Default is -1 which
represents no cluster changes.}
}
\value{
A named list with the following members:
        iters: The number of (outer) iterations performed.
        centers: A matrix of cluster centers.
        cluster: A vector of integers (from 1:k) indicating the cluster to which each point is allocated.
        sizes: The number of points in each cluster.
}
\description{
Assign each row of a matrix to a cluster denoted by a numeric. The clusters
are based on the distance metric of each row to one another. The assigned
cluster will have the smallest distance from the cluster center mean.
}
\examples{
num.clusts <- 5
mat <- fg.load.graph("mat-adj", "mat-index") # TODO: how to load
kms <- fg.kmeans(mat, num.clusts)
}
\author{
Disa Mhembere <disa@jhu.edu>
}

