% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fg.set.conf}
\alias{fg.set.conf}
\title{Reconfigure FlashGraphR}
\usage{
fg.set.conf(conf.file)
}
\arguments{
\item{conf.file}{The configuration file.}
}
\description{
This reconfigures FlashGraphR with the settings in the configuration file.
The configuration file contains a list of key-value pairs. Each line in
the file is a key-value pair in the form of "key_name=value".
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

