% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fg.subgraph}
\alias{fg.fetch.subgraph}
\alias{fg.fetch.subgraph.igraph}
\alias{fg.subgraph}
\title{Generate an induced subgraph}
\usage{
fg.fetch.subgraph.igraph(graph, vertices)

fg.fetch.subgraph(graph, vertices, name = paste(graph$name, "-sub", sep = ""))
}
\arguments{
\item{graph}{The FlashGraphR object}

\item{vertices}{A numeric vector that contains the ids of vertices in
the induced subgraph.}

\item{name}{The name of the FlashGraphR object.}
}
\value{
`fg.fetch.subgraph.igraph' returns an iGraph object,
`fg.fetch.subgraph' returns a FlashGraph object
}
\description{
Generate an induced subgraph that contains the specified vertices from
the input given graph.
}
\details{
`fg.fetch.subgraph.igraph' generates the induced subgraph and converts
it into an iGraph object

`fg.fetch.subgraph' generates the induced subgraph and returns
a FlashGraphR object.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

