% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fg.transitivity}
\alias{fg.transitivity}
\title{Transitivity of a graph}
\usage{
fg.transitivity(graph, type = c("global", "local"))
}
\arguments{
\item{graph}{The FlashGraphR object}

\item{type}{The type of the transitivity measure}
}
\value{
A numeric vector that contains transitivity of each vertex if
`type' is "local" and a single value if `type' is "global".
}
\description{
Transitivity measures the probability that the adjacent vertices
of a vertex are connected. This is sometimes also called the clustering
coefficient.
}
\details{
There are essentially two types of transitivity measures, one is
a vertex-level, the other a graph level property. This function can compute
both types of transitivity measures and works for both directed and
undirected graphs.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

