% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fg.triangle}
\alias{fg.triangle}
\alias{fg.triangles}
\title{Triangle counting}
\usage{
fg.triangles(graph, type = "cycle")
}
\arguments{
\item{graph}{The FlashGraphR object}

\item{type}{The type of triangles. It is ignored for undirected graphs.}
}
\value{
A numeric vector that contains the number of triangles associated
        with each vertex.
}
\description{
Count the number of triangles on each vertex in a graph.
}
\details{
Triangle counting works for both directed and undirected graphs.
For directed graphs, this counts the number of cycle triangles, shown
below.
A -> B
^   /
| v
C
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

