% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ntsa.R
\name{find_best_dim}
\alias{find_best_dim}
\title{Grid search of best embedding dimension with simplex projection}
\usage{
find_best_dim(
  df,
  cols,
  lib = matrix(c(1, nrow(df)), nrow = 1),
  pred = NULL,
  range = 1:10,
  Tps = rep(1, length(cols)),
  criterion = "rmse",
  both = TRUE,
  threadNo = detectCores()
)
}
\arguments{
\item{df}{vector, matrix, data.frame or tibble,
which contains time series to be analyzed}

\item{cols}{Numeric or character vector,
which specifies the df's column for the analysis}

\item{lib}{Vector or 2-column matrix, which sets the knot position in library data}

\item{pred}{Vector or 2-column matrix, which sets the knot position in test data}

\item{range}{Integer vector, which sets the range of dimension search}

\item{Tps}{Integer vector, which sets the prediction time horizon for each variable}

\item{criterion}{String, which switch the cost function (Rho, MAE & RMSE)
for the dimension search (default = "rmse")}

\item{both}{Logical, which switches the back/forward prediction or
forward only for the dimension search (defualt = TRUE)}

\item{threadNo}{Integer, which sets the thread number of parallel computing}
}
\description{
\code{find_best_dim} returns the best embedding dimension
}
