% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartition_analysis.R
\name{find_diff_and_ref_exparts_for_split}
\alias{find_diff_and_ref_exparts_for_split}
\title{Find differential and reference exonic parts for binary splits}
\usage{
find_diff_and_ref_exparts_for_split(
  g,
  source,
  sink,
  split,
  parsed_partitions,
  parsed_paths,
  tx_ex_parts,
  bipartitions_df,
  gene = gene,
  g_exon = NULL,
  g_expart = NULL,
  ex_or_in_gexpart = NULL,
  dexseq_frag = NULL
)
}
\description{
Find differential and reference exonic parts for binary splits
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_edge_names(g)
result <- grase::bipartition_paths("ENSG00000000003.14", g, outdir = tempdir())
}
}
