% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_analysis.R
\name{find_diff_and_ref_exparts_general}
\alias{find_diff_and_ref_exparts_general}
\title{Find differential and reference exonic parts for multinomial partitions}
\usage{
find_diff_and_ref_exparts_general(
  g,
  source,
  sink,
  groups,
  parsed_partitions,
  parsed_paths,
  tx_ex_parts,
  multipaths_list,
  gene = gene
)
}
\description{
Find differential and reference exonic parts for multinomial partitions
}
