% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{find_draft_comments}
\alias{find_draft_comments}
\title{Find draft comment sections at the end of file}
\usage{
find_draft_comments(lines)
}
\arguments{
\item{lines}{Character vector of file lines}
}
\value{
List of draft section info (start_line, end_line)
}
\description{
Find draft comment sections at the end of file
}
