% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ntsa.R
\name{find_knot}
\alias{find_knot}
\title{Find knot positions in data}
\usage{
find_knot(df, key_col = NULL, time_col = 1, diff = 1)
}
\arguments{
\item{df}{data.frame or tibble that contains the time-series data}

\item{key_col}{Numeric or character, which contains the lables of the different data group}

\item{time_col}{Numeric or character, which specifies the time-point column}

\item{diff}{Numeric, which sets the value of increment step in 'time_col'}
}
\description{
\code{find_knot} returns a 2-column matrix, where start and end point of an i'th knot in X
}
