% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartition_analysis.R
\name{find_reference_exonic_part}
\alias{find_reference_exonic_part}
\title{Find reference exonic part near bubble region (legacy version)}
\usage{
find_reference_exonic_part(
  source,
  sink,
  ex_part1_set,
  ex_part2_set,
  tx_ex_part1_set,
  tx_ex_part2_set,
  g
)
}
\description{
Find reference exonic part near bubble region (legacy version)
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_edge_names(g)
ref <- grase::find_reference_exonic_part(
  source = "3", sink = "7",
  ex_part1_set = integer(0), ex_part2_set = integer(0),
  tx_ex_part1_set = integer(0), tx_ex_part2_set = integer(0),
  g = g
)
}
}
