% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_analysis.R
\name{find_reference_exonic_part_general}
\alias{find_reference_exonic_part_general}
\title{Find reference exonic parts for multiple partitions (generalization of simple binary case)}
\usage{
find_reference_exonic_part_general(
  source,
  sink,
  ex_part_sets,
  tx_ex_part_sets,
  g_expart
)
}
\description{
Find reference exonic parts for multiple partitions (generalization of simple binary case)
}
