% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartitions.R
\name{find_reference_exonic_part_simple}
\alias{find_reference_exonic_part_simple}
\title{Find reference exonic part near bubble region between two partitions}
\usage{
find_reference_exonic_part_simple(
  source,
  sink,
  ex_part1_set,
  ex_part2_set,
  tx_ex_part1_set,
  tx_ex_part2_set,
  g_expart
)
}
\description{
Find reference exonic part near bubble region between two partitions
}
