% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{find_tx_with_epath}
\alias{find_tx_with_epath}
\title{Find transcripts that use given edge paths}
\usage{
find_tx_with_epath(g, trans, edges_list)
}
\arguments{
\item{g}{An \code{igraph} directed acyclic graph representing a gene splicing graph.}

\item{trans}{Character vector of transcript names (edge attribute names on \code{g}).}

\item{edges_list}{A list of character vectors, where each vector contains edge names
(as returned by \code{set_edge_names}) defining an exon path through the splicing graph.}
}
\description{
Find transcripts that use given edge paths
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_edge_names(g)
trans <- grase::transcripts_from_igraph(g)
edge_names <- igraph::E(g)$name[igraph::E(g)$ex_or_in == "ex"]
result <- grase::find_tx_with_epath(g, trans, list(edge_names[1:2]))
}
}
