% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_Orphan.R
\name{FindOrphan}
\alias{FindOrphan}
\title{Find orphan branches}
\usage{
FindOrphan(dem, mask, marked)
}
\arguments{
\item{dem}{Elevation matrix}

\item{mask}{Mask with ones for cells to be processed and zeros for everything else - defaults to a mask of all 1's}

\item{marked}{Matrix of grid cells that have been processed}
}
\description{
Function to look for unprocessed river cells that have d8 neigbors on the river network or on the boundary
}
