% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{fit_models}
\usage{
fit_models(formula, data)
}
\arguments{
\item{formula}{PARAM_DESCRIPTION}

\item{data}{input the data, family = Poisson by default and it has an offset by default}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
This function analyzes the data and outputs a summary
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[lme4]{glmer}}
 \code{\link[sjPlot]{tab_model}}
}
