% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recession.R
\name{fit_recession_event}
\alias{fit_recession_event}
\title{Fit recession power law for a single event}
\usage{
fit_recession_event(Q_values, remove_first_day = TRUE)
}
\description{
Fit recession power law for a single event
}
\keyword{internal}
