% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recession.R
\name{fit_sinusoidal_model}
\alias{fit_sinusoidal_model}
\title{Fit sinusoidal model to recession log(a) vs day-of-year}
\usage{
fit_sinusoidal_model(doy_values, log_a_values)
}
\description{
Fit sinusoidal model to recession log(a) vs day-of-year
}
\keyword{internal}
