% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fix_Drainage.R
\name{FixDrainage}
\alias{FixDrainage}
\title{Walk upstream from a point ensuring DEM is increasing by a minum epsilon}
\usage{
FixDrainage(dem, direction, mask, bank.epsilon, startpoint, d4 = c(1, 2,
  3, 4))
}
\arguments{
\item{dem}{Elevation matrix}

\item{direction}{a matrix of the flow directions for cells that have been processed - defaults to all zeros}

\item{mask}{Mask with ones for cells to be processed and zeros for everything else - defaults to a mask of all 1's}

\item{bank.epsilon}{the minimum elevation difference between cells walking up the banks from the river network.}

\item{startpoint}{the x,y index of a grid cell to start walking upstream from}

\item{d4}{directional numbering system: the numbers you want to assigne to down, left, top,right (defaults to 1,2,3,4)}
}
\description{
This walks upstream from a point following a flow direction file and checks that the elevation upstream
is greater than or equal to the elvation at the point + epsilon
Once the function reaches a point where upstream cells pass it stops 
NOTE: this is only processing the immediate neigborhood and does not recurse over the entire domain, 
For example if you did the entire overal prioirty flow processing with an espilon of 0 then ran this
functon starting at a river point with an epsilon of 0.1 this function will traverse upstream from the 
river bottom checking that every cell conneting to the river cell is higher by at least this ammout but 
once it reaches a point where every connected cell passes this test it will stop. Therefore there could
still be locations higher up on the hillslope with the original epsilon of zero still.  This  is on purpose
and this script is not intended to gobally ensure a given epsilon.
}
