% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flat_Fix.R
\name{FixFlat}
\alias{FixFlat}
\title{Indentify and fix stagnation points}
\usage{
FixFlat(direction, slopex, slopey, adj_th, adj_ratio = -1, mask)
}
\arguments{
\item{direction}{Nx by ny matrix of flow directions following the following convention (1=down, 2=left, 3=up, 4=right)}

\item{slopex}{Nx by ny matrix of slopes in the x direction (Note: these should be face centered slopes as woudl be calcualted with SlopeCalcStan)}

\item{slopey}{Nx by ny matrix of slopes in the y direction (Note: these should be face centered slopes as woudl be calcualted with SlopeCalcStan)}

\item{adj_th}{Threshold for slope adjustment. If the total slopes out divided by the total slopes into a given cell is less than adj_th the outlet will be scale by adj_ratio.}

\item{adj_ratio}{Scaler value for slope adjustment. New outlet slopes will be set to inital outlet slope time adj_ratio. If no ad_ratio is provided
then the Outlet slope will be set to the total inlets slopws times the adjustment theshold}

\item{mask}{Nx by Ny matrix indicating the active domain, 1=active, 0 in inactive. If not mask is provided the function will be applied over the entire input matrices.}
}
\value{
This will generate the following outputs:
   - outslope- Nx by ny matrix with the sum of the slopes pointed out of every cell (Note: this is the sum of slop magnitudes so it will all be positive)
   - inslope- Nx by ny matrix with the sum of the slopes pointed into every cell (Note: this is the sum of slop magnitudes so it will all be positive)
   - OutIn_ratio - Nx by Ny matrix with the ratio of outslope to inslope (Note: cells outside the mask or with 0 inslope are assigned a ratio of -1)
   - slopex - The adjusted nx by ny matrix of slopex values
   - slopey - The adjusted nx by ny matrix of slopex values
   - adj_mask - Nx by ny matrix indicating cells that were ajusted (1=adjusted, 0=not adjsuted). This will be all the cells where inslope/outslop<adj_th
   - SlopeOutlet - Nx by ny matrix of the outlet slope for every grid cell 
   - SlopeOutletNew - Nx by ny matrix of the outlet slope for every grid cell after processing
}
\description{
A funtion that finds cells where the total output slope is much less than the input
slopes and adjusts the outlet slope to address this
}
