% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatterSIDB.R
\name{flatterSIDB}
\alias{flatterSIDB}
\title{Flattens sidb database object for easier querying}
\usage{
flatterSIDB(database)
}
\arguments{
\item{database}{A database with the sidb structure}
}
\value{
R list with two elements, linked by shared column "ID":\cr
1) \emph{timeseries}, a list of unique timeseries, each a 2-column dataframe (time, response)\cr
2) \emph{vars}, a list of all site, incubation and variables fields flattened into a dataframe for each entry
}
\description{
Flattens sidb database object for easier querying
}
\examples{
db <- loadEntries(path="~/sidb/data/")
sidb.flatter <- flatterSIDB(db)
}
