% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexjson2.R
\name{flexsurv_json}
\alias{flexsurv_json}
\title{FUNCTION_TITLE}
\usage{
flexsurv_json(
  model = cffpm.list,
  vartime = seq(365.25, 730.5, by = 365.25),
  qmat = tmat,
  process = "Markov",
  totlos = TRUE,
  ci.json = TRUE,
  cl.json = 0.95,
  B.json = 100,
  tcovs = NULL,
  Mjson = 100,
  variance = FALSE,
  covariates_list = list(pat1, pat2, pat3),
 
    jsonpath = "C:/Users/niksko/Desktop/mstate3/datasets/json/flexsurv/json_present_flexsurv",
  name = "predictions_EBMT_flex_fw.json"
)
}
\arguments{
\item{model}{PARAM_DESCRIPTION, Default: cffpm.list}

\item{vartime}{PARAM_DESCRIPTION, Default: seq(365.25, 730.5, by = 365.25)}

\item{qmat}{PARAM_DESCRIPTION, Default: tmat}

\item{process}{PARAM_DESCRIPTION, Default: 'Markov'}

\item{totlos}{PARAM_DESCRIPTION, Default: TRUE}

\item{ci.json}{PARAM_DESCRIPTION, Default: TRUE}

\item{cl.json}{PARAM_DESCRIPTION, Default: 0.95}

\item{B.json}{PARAM_DESCRIPTION, Default: 100}

\item{tcovs}{PARAM_DESCRIPTION, Default: NULL}

\item{Mjson}{PARAM_DESCRIPTION, Default: 100}

\item{variance}{PARAM_DESCRIPTION, Default: FALSE}

\item{covariates_list}{PARAM_DESCRIPTION, Default: list(pat1, pat2, pat3)}

\item{jsonpath}{PARAM_DESCRIPTION, Default: 'C:/Users/niksko/Desktop/mstate3/datasets/json/flexsurv/json_present_flexsurv'}

\item{name}{PARAM_DESCRIPTION, Default: 'predictions_EBMT_flex_fw.json'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringi]{stri_sort}}
}
