% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{float_is_finite}
\alias{float_is_finite}
\title{float_is_finite}
\usage{
float_is_finite(x)
}
\arguments{
\item{x}{Numeric scalar.}
}
\description{
Test whether a float is finite via RFloat trait.
}
