% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\docType{class}
\name{fm-class}
\alias{fm-class}
\title{An S4 class to represent a FlashR matrix.}
\description{
An S4 class to represent a FlashR matrix.
}
\section{Slots}{

\describe{
\item{\code{pointer}}{points a matrix object in C.}

\item{\code{name}}{a string indicating the name of the matrix.}

\item{\code{nrow}}{a numeric value indicating the number of rows.}

\item{\code{ncol}}{a numeric value indicating the number of columns.}

\item{\code{type}}{a string indicating the type of the matrix. e.g., sparse or dense.}

\item{\code{ele_type}}{a string indicating the element type in the matrix.}

\item{\code{attrs}}{a list that stores the attributes of the matrix.}
}}

