% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fm.agg.impl}
\alias{fm.agg.impl}
\alias{fm.max}
\alias{fm.min}
\alias{fm.sum}
\title{Aggregation on a FlashMatrixR vector/matrix.}
\usage{
fm.sum(m)

fm.min(m)

fm.max(m)
}
\arguments{
\item{m}{The input vector/matrix}
}
\value{
an integer if the input is an integer vector/matrix; a numeric
value if the input is a numeric vector/matrix.
}
\description{
`fm.sum' returns the sum of all the values in the input vector/matrix.
}
\details{
`fm.min' returns the minimum of all values in the input vector/matrix.

`fm.max' returns the maximum of all values in the input vector/matrix.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

