% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\docType{class}
\name{fm.agg.op-class}
\alias{fm.agg.op-class}
\title{An S4 class to represent an aggregation operator used in aggregation
operations in FlashMatrix.}
\description{
An S4 class to represent an aggregation operator used in aggregation
operations in FlashMatrix.
}
\section{Slots}{

\describe{
\item{\code{agg}}{an integer indicating the operator for computing partial
aggregation results.}

\item{\code{combine}}{an integer indicating the operator for combining partial
aggregation results and computing the final result.}

\item{\code{name}}{a string indicating the name of the aggregation operation.}
}}

