% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\docType{class}
\name{fm.apply.op-class}
\alias{fm.apply.op-class}
\title{An S4 class to represent an apply operator on rows/columns in a matrix.
This operator is used in \code{fm.apply}.}
\description{
An S4 class to represent an apply operator on rows/columns in a matrix.
This operator is used in \code{fm.apply}.
}
\section{Slots}{

\describe{
\item{\code{info}}{an integer indicating the apply operator registered in the system.}

\item{\code{name}}{a string indicating the name of the operator.}
}}

