% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eigen.R
\name{fm.arpack}
\alias{fm.arpack}
\title{Compute eigenvalues/vectors on a symmetric matrix.}
\usage{
fm.arpack(A.multiply, n, k, m, which, tol)
}
\arguments{
\item{A.multiply}{the function to perform matrix vector multiplication.}

\item{n}{Numeric scalar. The dimension of the eigenproblem.}

\item{k}{Numeric scalar. The number of eigenvalues to be computed.}

\item{m}{Number of Lanczos vectors to be generated.}

\item{which}{Specify which eigenvalues/vectors to compute, character
constant with exactly two characters.
Possible values for symmetric input matrices:
"LA" Compute "nev" largest (algebraic) eigenvalues.
"SA" Compute "nev" smallest (algebraic) eigenvalues.
"LM" Compute "nev" largest (in magnitude) eigenvalues.
"SM" Compute "nev" smallest (in magnitude) eigenvalues.}

\item{Numeric}{scalar. Stopping criterion: the relative accuracy of
the Ritz value is considered acceptable if its error is less than 'tol'
times its estimated value. If this is set to zero then machine precision is used.}
}
\description{
This method reimplements implicitly restarted Lanczos method to compute
eigenvalues and eigenvectors. The method is implemented completely in R
and follows the implementation of ARPACK library.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

