% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.as.factor}
\alias{fm.as.factor}
\title{Convert a FlashR vector to a FlashR factor vector.}
\usage{
fm.as.factor(fm, num.levels = -1)
}
\arguments{
\item{fm}{a FlashR vector.}

\item{num.levels}{The number of levels in the factor vector.}
}
\value{
a FlashR factor vector.
}
\description{
Convert a FlashR vector to a FlashR factor vector.
}

