% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fm.as.vector}
\alias{fm.as.vector}
\title{Convert a FlashMatrixR matrix to a FlashMatrixR vector.}
\usage{
fm.as.vector(fm)
}
\arguments{
\item{fm}{a FlashMatrixR matrix}
}
\value{
a FlashMatrixR vector
}
\description{
The matrix must have only one row or one column. Otherwise, the function
returns an error.
}

