% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ASE.R
\name{fm.ase}
\alias{fm.ASE}
\alias{fm.ase}
\title{Spectral embedding}
\usage{
fm.ASE(fm, num.eigen, which = c("A, Aug, L, nL"),
  which.eigen = c("LM, LA, SM, SA"), c = 1, tol = 1e-12)
}
\arguments{
\item{fm}{The FlashMatrixR object}

\item{num.eigen}{The number of eigenvalues/vectors required.}

\item{which}{The type of the embedding.}

\item{which.eigen}{Specify which eigenvalues/vectors to compute,
character constant with exactly two characters.}

\item{c}{The constant used in the Aug variant of embedding.}

\item{tol}{Numeric scalar. Stopping criterion: the relative accuracy
of the Ritz value is considered acceptable if its error is less than
tol times its estimated value. If this is set to zero then machine
precision is used.}

\item{tau}{The value used in the regularized Laplacian embedding.}
}
\value{
A named list with the following members:

values

   Numeric vector, the desired eigenvalues.

vectors

   Numeric matrix, the desired eigenvectors as columns. If complex=TRUE
  (the default for non-symmetric problems), then the matrix is complex.

options

  A named list with the supplied options and some information about
  the performed calculation, including an ARPACK exit code.
}
\description{
This computes spectral embedding of a given graph.
It supports four variants:
}
\details{
Adjacency matrix (A),

Augmented matrix (Aug), A - c * D, where c is a constant provided by a user.

Laplacian matrix (L), L = D - A,

Normalized Laplacian matrix (nL), nL = I - D^(-1/2) * A * D^(-1/2)

For the Laplacian matrix and the normalized Laplacian matrix, this function
only accepts a symmetric sparse matrix.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

