% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\docType{class}
\name{fm.bo-class}
\alias{fm.bo-class}
\title{An S4 class to represent a binary operator used in generalized matrix
operations.}
\description{
An S4 class to represent a binary operator used in generalized matrix
operations.
}
\section{Slots}{

\describe{
\item{\code{info}}{an integer indicating the binary operator registered
in the system.}

\item{\code{name}}{a string indicating the name of the operator.}
}}

