% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.conv.store}
\alias{fm.conv.store}
\title{Convert the Storage of an Object.}
\usage{
fm.conv.store(fm, in.mem, name = "")
}
\arguments{
\item{fm}{a FlashMatrix object.}

\item{in.mem}{a logical value indicating whether to store the FlashMatrix
object in memory.}

\item{name}{a string to indicate the name of the new FlashMatrix object.}
}
\value{
a new FlashMatrix object with data stored in specified storage.
}
\description{
This function converts the storage of a FlashMatrix object. The storage
can be memory or disks. This function is used for finely optimizing
the computation of FlashMatrix code. It shouldn't be used to most of
the users.
}

