% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fm.exist.matrix}
\alias{fm.exist.matrix}
\title{Indicate whether a matrix has been loaded to FlashMatrixR}
\usage{
fm.exist.matrix(name)
}
\arguments{
\item{name}{A matrix name.}
}
\value{
A boolean value: true if the matrix has been loaded to FlashMatrixR.
}
\description{
This function indicates whether a matrix has been loaded to FlashMatrixR.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

