% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fm.get.cols}
\alias{fm.get.cols}
\title{Get a submatrix composed of some columns from a FlashMatrixR matrix}
\usage{
fm.get.cols(fm, idxs)
}
\arguments{
\item{fm}{A FlashMatrixR matrix}

\item{idxs}{an array of column indices in fm.}
}
\value{
a submatrix
}
\description{
This only works on a column-wise matrix. The submatrix is also organized
column-wise. The submatrix shares the same data as the input matrix, so
there is no memory copy.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

