% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.get.eles}
\alias{fm.get.cols}
\alias{fm.get.eles}
\alias{fm.get.eles.vec}
\alias{fm.get.rows}
\title{Get a submatrix from a FlashR matrix}
\usage{
fm.get.cols(fm, idxs)

fm.get.rows(fm, idxs)

fm.get.eles.vec(fm, idxs)
}
\arguments{
\item{fm}{A FlashR matrix}

\item{idxs}{an array of column indices in fm.}
}
\value{
a FlashR vector if getting one row or column;
a FlashR matrix if getting more than one row or column.
}
\description{
\code{fm.get.rows} gets specified rows in a FlashR matrix.
\code{fm.get.cols} gets specified columns in a FlashR matrix.
\code{fm.get.eles.vec} gets specified elements from a FlashR vector.
}
\examples{
mat <- fm.runif.matrix(100, 10)
sub <- fm.get.cols(mat, as.integer(runif(5, min=1, max=10)))
sub <- fm.get.rows(mat, as.integer(runif(5, min=1, max=100)))
vec <- fm.runif(100)
sub <- fm.get.eles.vec(vec, as.integer(runif(5, min=1, max=100)))
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

