% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashmatrix.R
\name{fm.get.layout}
\alias{fm.get.layout}
\title{Get the data layout of a FlashMatrixR matrix.}
\usage{
fm.get.layout(fm)
}
\arguments{
\item{fm}{a FlashMatrixR matrix}
}
\value{
a string that indicates the data layout.
}
\description{
Get the data layout of a FlashMatrixR matrix.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

