% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.get.matrix}
\alias{fm.get.dense.matrix}
\alias{fm.get.matrix}
\alias{fm.get.sparse.matrix}
\alias{fm.load.dense.matrix}
\alias{fm.load.dense.matrix.bin}
\alias{fm.load.sparse.matrix}
\title{Load a sparse matrix to FlashR.}
\usage{
fm.get.sparse.matrix(fg)

fm.get.dense.matrix(name)

fm.load.dense.matrix(name, in.mem, ele.type = "D", delim = ",",
  ncol = .Machine$integer.max, mat.name = "")

fm.load.dense.matrix.bin(name, in.mem, nrow, ncol, byrow, ele.type,
  mat.name = "")

fm.load.sparse.matrix(spm, spm.idx, t.spm = NULL, t.spm.idx = NULL,
  in.mem = TRUE)
}
\arguments{
\item{fg}{A FlashGraph object.}

\item{in.mem}{Determine the loaded matrix is stored in memory or on SAFS.}

\item{ele.type}{The element type in a matrix.}

\item{delim}{The delimiter of separating elements in the text format.}

\item{spm}{The file that stores the sparse matrix.}

\item{spm.idx}{The file that stores the index of the sparse matrix.}

\item{t.spm}{The file that stores the transpose of the sparse matrix.}

\item{t.spm.idx}{The file that stores the index of the transpose of the sparse matrix.}
}
\value{
a FlashMatrix matrix.
}
\description{
Load a sparse matrix to FlashR from difference sources.
}
\details{
`fm.get.sparse.matrix' gets a FlashMatrix sparse matrix that references
a matrix represented by a FlashGraph object.

`fm.get.dense.matrix' gets a FlashMatrix dense matrix from SAFS.

`fm.load.dense.matrix' loads a dense matrix in the text format from
the Linux filesystem.

`fm.load.sparse.matrix' loads a FlashMatrix sparse matrix from files.
The matrix in the file is in the FlashMatrix format.
}
\examples{
fg <- fg.load.graph("graph.adj", "graph.index")
fm <- fm.get.sparse.matrix(fg)
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

