% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMM.R
\name{fm.GMM}
\alias{fm.GMM}
\title{Gaussian Mixture Model}
\usage{
fm.GMM(X, k, maxiters, verbose = FALSE)
}
\arguments{
\item{X}{a n x p data matrix, where each row is a data point.}

\item{k}{the number of clusters}

\item{maxiters}{the maximal number of iterations.}

\item{verbose}{indicate whether to print extra information.}
}
\value{
This function returns a list of components:
  \item{P}{A n x k matrix, where each row stores the probability that
        a corresponding data point belongs to each cluster.}
  \item{clust.ids}{A vector of n elements, where each element indicates
        the cluster where a corresponding data point most likely belongs to.}
  \item{niters}{a scalar that indicates the number of iterations}
  \item{phi}{A vector of k elements, indicating the weight of Gaussian
        distributions.}
  \item{mus}{A k x p matrix, where each row is the mean of a Gaussian
        distribution.}
  \item{covars}{A list of k covariance matrices.}
}
\description{
This clusters data points with an assumption that data is sampled from
mixtures of Gaussian distribution. It fits Gaussian mixture model with
EM algorithm.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

