% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.groupby}
\alias{fm.groupby}
\alias{fm.sgroupby}
\title{Groupby on a FlashR vector.}
\usage{
fm.sgroupby(obj, FUN)

fm.groupby(obj, margin, factor, FUN)
}
\arguments{
\item{obj}{a FlashR vector or matrix}

\item{FUN}{an aggregation operator returned by \code{fm.create.agg.op}.}

\item{margin}{the subscript which the function will be applied over.
E.g., for a matrix, \code{1} indicates rows, \code{2} indicates columns.}

\item{factor}{a FlashR factor vector that indicates how rows/columns
in a matrix should be grouped.}
}
\value{
\code{fm.sgroupby} returns a data frame, where the column \code{val}
stores all of the unique values in the original data container, and the column
\code{agg} stores the aggregate result of the corresponding value.
}
\description{
\code{fm.sgroupby} groups elements in a vector based on corresponding
\code{labels} and applies \code{FUN} to the elements in each group.
\code{FUN} is an aggregation operator.
}
\details{
\code{fm.groupby} groups rows/columns of a matrix based on corresponding
\code{labels} and applies \code{FUN} to the rows/columns in each group.
\code{FUN} is an aggregation operator.
}
\examples{
vec <- fm.runif(100)
res <- fm.sgroupby(vec, "+")
mat <- fm.runif.matrix(100, 10)
fact <- fm.as.factor(as.integer(fm.runif(nrow(mat), min=0, max=3)))
res <- fm.groupby(mat, 2, fact, "+")
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

