% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.inner.prod}
\alias{fm.inner.prod}
\title{Matrix inner product}
\usage{
fm.inner.prod(fm, mat, Fun1, Fun2)
}
\arguments{
\item{fm}{A FlashR matrix}

\item{mat}{A FlashR dense matrix.}

\item{Fun1}{The reference or the name of one of the predefined basic binary
operators.}

\item{Fun2}{The reference or the name of one of the predefined basic binary
operators.}
}
\value{
a FlashR vector if the second argument is a vector;
a FlashR matrix if the second argument is a matrix.
}
\description{
It takes two operators and performs inner product on a dense matrix
}
\examples{
mat1 <- fm.runif.matrix(1000, 100)
mat2 <- fm.runif.matrix(100, 10)
mat <- fm.inner.prod(mat1, mat2, "*", "+")
mat <- fm.inner.prod(mat1, mat2, fm.bo.mul, fm.bo.add)
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

