% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KMeans.R
\name{fm.KMeans}
\alias{fm.KMeans}
\title{KMeans clustering}
\usage{
fm.KMeans(data, K, max.iters = 10, debug = FALSE)
}
\arguments{
\item{data}{the input data matrix where each row is a data point.}

\item{K}{the number of clusters}

\item{max.iters}{the maximal number of iterations.}

\item{debug}{This indicates whether to print debug info.}
}
\value{
a vector that contains cluster Ids for each data point.
}
\description{
Perform k-means clustering on a data matrix.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

