% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\docType{methods}
\name{fm.mapply2}
\alias{fm.mapply.col}
\alias{fm.mapply.row}
\alias{fm.mapply2}
\alias{fm.mapply2,ANY,fm,ANY,logical-method}
\alias{fm.mapply2,ANY,fmV,ANY,logical-method}
\alias{fm.mapply2,fm,ANY,ANY,logical-method}
\alias{fm.mapply2,fm,fm,ANY,logical-method}
\alias{fm.mapply2,fm,fmV,ANY,logical-method}
\alias{fm.mapply2,fm,matrix,ANY,logical-method}
\alias{fm.mapply2,fmV,ANY,ANY,logical-method}
\alias{fm.mapply2,fmV,fm,ANY,logical-method}
\alias{fm.mapply2,fmV,fmV,ANY,logical-method}
\alias{fm.mapply2,matrix,fm,ANY,logical-method}
\title{Apply a Function to two FlashMatrix vectors/matrices.}
\usage{
fm.mapply.row(o1, o2, FUN, set.na = TRUE)

fm.mapply.col(o1, o2, FUN, set.na = TRUE)

\S4method{fm.mapply2}{fm,fm,ANY,logical}(o1, o2, FUN, set.na = TRUE)

\S4method{fm.mapply2}{fmV,fmV,ANY,logical}(o1, o2, FUN, set.na = TRUE)

\S4method{fm.mapply2}{fm,fmV,ANY,logical}(o1, o2, FUN, set.na = TRUE)

\S4method{fm.mapply2}{fmV,fm,ANY,logical}(o1, o2, FUN, set.na)

\S4method{fm.mapply2}{fm,matrix,ANY,logical}(o1, o2, FUN, set.na = TRUE)

\S4method{fm.mapply2}{matrix,fm,ANY,logical}(o1, o2, FUN, set.na = TRUE)

\S4method{fm.mapply2}{fm,ANY,ANY,logical}(o1, o2, FUN, set.na = TRUE)

\S4method{fm.mapply2}{ANY,fm,ANY,logical}(o1, o2, FUN, set.na = TRUE)

\S4method{fm.mapply2}{fmV,ANY,ANY,logical}(o1, o2, FUN, set.na = TRUE)

\S4method{fm.mapply2}{ANY,fmV,ANY,logical}(o1, o2, FUN, set.na = TRUE)
}
\arguments{
\item{o1,o2}{a FlashMatrix vector/matrix.}

\item{FUN}{the reference or the name of one of the predefined basic binary
operators.}

\item{set.na}{a logical value indicating whether to set the value in
the output matrix to NA if the value in the corresponding
location of the input matrix is NA.}
}
\value{
a FlashMatrix vector/matrix.
}
\description{
\code{fm.mapply2} applies \code{FUN} to the first elements of each
vector/matrix, the second elements, the third elements, and so on.
Two vectors/matrices should have the same shape. Currently,
\code{fm.mapply2} only accepts predefined basic operators returned
by \code{fm.get.basic.op}.
}
\details{
\code{fm.mapply.row} and \code{fm.mapply.col} applies to a matrix and
a vector. \code{fm.mapply.row} applies \code{FUN} element-wise to
each row of the matrix in the left argument and the vector in the right
argument. \code{fm.mapply.col} applies \code{FUN} element-wise to
each column of the matrix in the left argument and the vector in the right
argument.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

