% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashmatrix.R
\name{fm.matrix}
\alias{fm.matrix}
\title{Create a matrix from the given set of values.}
\usage{
fm.matrix(vec, nrow, ncol, byrow = FALSE)
}
\arguments{
\item{vec}{a given vector.}

\item{nrow}{the number of rows in the output matrix.}

\item{ncol}{the number of columns in the output matrix.}

\item{byrow}{logical. If FALSE (the default) the matrix is filly
columns, otherwise the matrix is filled by rows.}
}
\value{
a FlashMatrixR matrix
}
\description{
The given set has to be a FlashGraphR matrix or FlashGraphR vector.
}

