% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.multiply}
\alias{fm.multiply}
\title{Matrix multiplication}
\usage{
fm.multiply(fm, mat)
}
\arguments{
\item{fm}{A FlashR matrix}

\item{mat}{A FlashR dense matrix.}
}
\value{
a FlashR vector if the second argument is a vector;
a FlashR matrix if the second argument is a matrix.
}
\description{
Multiply a sparse/dense matrix with a dense vector/matrix.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

