% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph-algs.R
\name{fm.PageRank}
\alias{fm.PageRank}
\title{Compute PageRank on a graph.}
\usage{
fm.PageRank(A, d = 0.15, max.niters = 30, epsilon = 0.01,
  verbose = FALSE)
}
\arguments{
\item{A}{a sparse matrix that stores the graph.}

\item{d}{a damping factor.}

\item{max.niters}{the maximal number of iterations.}

\item{epsilon}{determines the precision of PageRank values.}

\item{verbose}{indicates whether to print extra information.}
}
\value{
a vector that contains the PageRank value for each vertex.
}
\description{
This function assumes a graph is stored in a sparse matrix and computes
exact PageRank with a sequence of sparse matrix multiplication.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

