% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.print.mat.info}
\alias{fm.print.mat.info}
\title{Print the information of a FlashMatrix object}
\usage{
fm.print.mat.info(fm)
}
\arguments{
\item{fm}{a FlashMatrix object}
}
\description{
Print the information of the internal storage representation of
a FlashMatrix object. For a virtualized computation result,
\code{fm.print.mat.info} prints the internal computation dependancy.
}

