% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.rep.int}
\alias{fm.rep.int}
\title{Create a FlashR vector with replicated elements.}
\usage{
fm.rep.int(x, times)
}
\arguments{
\item{x}{A constant initial value}

\item{times}{The length of the vector to be generated.}
}
\value{
A FlashR vector
}
\description{
Create a FlashR vector with replicated elements.
}
\examples{
vec <- fm.rep.int(0, 10) # create a FlashR vector with 10 elements.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

