% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.rnorm}
\alias{fm.rnorm}
\title{Create a FlashMatrix vector with random numbers from normal distribution.}
\usage{
fm.rnorm(n, mean = 0, sd = 1, in.mem = TRUE, name = "")
}
\arguments{
\item{n}{the number of random numbers to be generated.}

\item{mean}{the mean of the distribution.}

\item{sd}{the standard deviation of the distribution.}

\item{in.mem}{whether the vector is stored in memory.}
}
\description{
Create a FlashMatrix vector with random numbers from normal distribution.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

