% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.rsparse.proj}
\alias{fm.rsparse.proj}
\title{Create a sparse projection matrix.}
\usage{
fm.rsparse.proj(nrow, ncol, density, name = "")
}
\arguments{
\item{nrow}{the number of rows in the generated matrix.}

\item{ncol}{the number of columns in the generated matrix.}

\item{density}{the ratio of non-zero entries to the total number of elements.}

\item{name}{the name of the matrix. It's stored on disks, it's used as
the file name.}
}
\value{
a FlashR matrix.
}
\description{
\code{fm.rsparse.proj} creates a sparse projection matrix stored in memory.
}
\details{
The non-zero values in the sparse projection matrix are either 1 or -1.
Their values are uniformly randomly chosen to be 1 or -1.
}
\examples{
mat <- fm.rsparse.proj(10000, 100, 0.001) # a sparse projection matrix.
}

