% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\docType{methods}
\name{fm.sapply}
\alias{fm.sapply}
\alias{fm.sapply,fm,ANY,logical-method}
\alias{fm.sapply,fmV,ANY,logical-method}
\title{Apply a Function to a FlashMatrix vector/matrix.}
\usage{
fm.sapply(o, FUN, set.na)

\S4method{fm.sapply}{fm,ANY,logical}(o, FUN, set.na = TRUE)

\S4method{fm.sapply}{fmV,ANY,logical}(o, FUN, set.na = TRUE)
}
\arguments{
\item{o}{a FlashMatrix vector/matrix.}

\item{FUN}{the reference or the name of a predefined uniary operator.}

\item{set.na}{a logical value indicating whether to set the value in
the output vector or matrix to NA if the value in
the corresponding location of the input vector or matrix
is NA.}
}
\value{
a FlashMatrix vector/matrix.
}
\description{
\code{sapply} applies \code{FUN} to every element of a vector/matrix.
Currently, \code{sapply} only accepts predefined basic operators
returned by \code{fm.get.basic.uop}.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

