% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.seq}
\alias{fm.seq}
\alias{fm.seq.int}
\alias{fm.seq.matrix}
\title{Sequence Generation}
\usage{
fm.seq.int(from, to, by)

fm.seq.matrix(from, to, nrow, ncol, byrow = FALSE)
}
\arguments{
\item{from}{the starting value of the sequence.}

\item{to}{the end value of the sequence.}

\item{by}{number. increment of the sequence.}

\item{nrow}{the number of rows in the generated matrix.}

\item{ncol}{the number of columns in the generated matrix.}

\item{byrow}{logical. If \code{FALSE} (the default) the matrix is filled by
columns, otherwise the matrix is filled by rows.}
}
\value{
a FlashMatrix vector or matrix filled with a sequence of numbers
        that belongs to [from, to]
}
\description{
\code{fm.seq.int} creates a FlashMatrix vector with a sequence of numbers.
\code{fm.seq.matrix} creates a FlashMatrix matrix with a sequence of numbers.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

