% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.set.conf}
\alias{fm.print.conf}
\alias{fm.set.conf}
\alias{fm.set.log.level}
\title{Reconfigure FlashMatrix}
\usage{
fm.set.conf(conf.file)

fm.set.log.level(level)

fm.print.conf()
}
\arguments{
\item{conf.file}{The configuration file.}

\item{level}{a string, including "debug", "info", "warning", "error", "fatal".}
}
\description{
\code{fm.set.conf} reconfigures FlashMatrix with the settings in
the configuration file. \code{fm.print.conf} prints the current configurations.
\code{fm.set.log.level} sets what levels of messages should be logged.
}
\details{
The configuration file contains a list of key-value pairs. Each line in
the file is a key-value pair in the form of "key_name=value".
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

