% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.t}
\alias{fm.t}
\title{Transpose a FlashR matrix.}
\usage{
fm.t(m)
}
\arguments{
\item{m}{a FlashR matrix}
}
\value{
a FlashR matrix
}
\description{
Transpose a FlashR matrix.
}
\examples{
mat <- fm.t(fm.runif.matrix(100, 10))
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

