% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR_base.R
\name{fm.table}
\alias{fm.table}
\title{Count the number of elements}
\usage{
fm.table(x)
}
\arguments{
\item{x}{a FlashR vector.}
}
\value{
a list of elements.
\itemize{
\item{val}{The unique values in the vector.}
\item{Freq}{The number of occurences of each unique value.}
}
}
\description{
This function counts the number of occurences of each unique value
in a FlashR vector.
}

