% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.write.obj}
\alias{fm.write.obj}
\title{Write a FlashR object (vector/matrix) to a file}
\usage{
fm.write.obj(fm, file)
}
\arguments{
\item{fm}{a FlashR object.}

\item{file}{a file in the local filesystem.}
}
\value{
a logical value. True if the object is written to a file
successfully. Otherwise, FALSE.
}
\description{
Write a FlashR object (vector/matrix) to a file
}
\examples{
mat <- fm.runif.matrix(100, 10)
fm.write.obj(mat, "/tmp/tmp.mat")
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

