% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\docType{class}
\name{fmSink-class}
\alias{fmSink-class}
\title{An S4 class to represent a FlashMatrix sink matrix. The definition of a sink
matrix can be found in the paper https://arxiv.org/abs/1604.06414}
\description{
An S4 class to represent a FlashMatrix sink matrix. The definition of a sink
matrix can be found in the paper https://arxiv.org/abs/1604.06414
}
\section{Slots}{

\describe{
\item{\code{pointer}}{points a matrix object in C.}

\item{\code{name}}{a string indicating the name of the matrix.}

\item{\code{nrow}}{a numeric value indicating the number of rows.}

\item{\code{ncol}}{a numeric value indicating the number of columns.}

\item{\code{type}}{a string indicating the type of the matrix. This field isn't used.}

\item{\code{ele_type}}{a string indicating the element type in the matrix.}
}}

