% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\docType{class}
\name{fmV-class}
\alias{fmV-class}
\title{An S4 class to represent a FlashR vector.}
\description{
An S4 class to represent a FlashR vector.
}
\section{Slots}{

\describe{
\item{\code{pointer}}{points a vector object in C.}

\item{\code{name}}{a string indicating the name of the vector.}

\item{\code{len}}{a numeric value indicating the number of elements.}

\item{\code{type}}{a string indicating the type of the vector. This field isn't
used for a vector.}

\item{\code{ele_type}}{a string indicating the element type in the vector.}

\item{\code{attrs}}{a list that stores the attributes of the matrix.}
}}

