% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focalexons.R
\name{focal_exons_gene_ovr}
\alias{focal_exons_gene_ovr}
\title{Compute focal exons from graph and splicing structure}
\usage{
focal_exons_gene_ovr(gene, g, sg, outdir)
}
\description{
Compute focal exons from graph and splicing structure
}
