% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focalexons.R
\name{focal_exons_gene_power}
\alias{focal_exons_gene_power}
\title{Compute focal exons from graph and splicing structure}
\usage{
focal_exons_gene_power(gene, g, sg, node2pos, outdir, max_powerset = 10000)
}
\description{
Compute focal exons from graph and splicing structure
}
