% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focalexons.R
\name{focal_exons_gene_powerset}
\alias{focal_exons_gene_powerset}
\title{Compute focal exons from graph and splicing structure}
\usage{
focal_exons_gene_powerset(
  gene,
  g,
  sg,
  outdir,
  max_path = 30,
  collapse_bubbles = TRUE
)
}
\description{
Compute focal exons from graph and splicing structure
}
