% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{force_invisible_i32}
\alias{force_invisible_i32}
\title{force_invisible_i32}
\usage{
force_invisible_i32()
}
\description{
Test that \verb{#[miniextendr(invisible)]} forces an i32 return to be invisible in R.
}
