% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{force_visible_unit}
\alias{force_visible_unit}
\title{force_visible_unit}
\usage{
force_visible_unit()
}
\description{
Test that \verb{#[miniextendr(visible)]} forces a unit return to be visible in R.
}
