% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fork}
\alias{fork}
\title{Copy the counter, emitting a deferred message}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Copy the counter, emitting a deferred message
}
