% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_total.R, R/msboxes_R.R
\name{freq_func_total}
\alias{freq_func_total}
\title{FUNCTION_TITLE}
\usage{
freq_func_total(
  msdata,
  msid,
  names_of_ststates,
  values_ststates,
  names_of_nastates,
  values_nastates,
  names_of_abstates,
  values_abstates,
  names_of_transitions,
  values_of_transitions,
  time,
  timevar,
  scale_inner = 1
)

freq_func_total(
  msdata,
  msid,
  names_of_ststates,
  values_ststates,
  names_of_nastates,
  values_nastates,
  names_of_abstates,
  values_abstates,
  names_of_transitions,
  values_of_transitions,
  time,
  timevar,
  scale_inner = 1
)
}
\arguments{
\item{msdata}{PARAM_DESCRIPTION}

\item{msid}{PARAM_DESCRIPTION}

\item{names_of_ststates}{PARAM_DESCRIPTION}

\item{values_ststates}{PARAM_DESCRIPTION}

\item{names_of_nastates}{PARAM_DESCRIPTION}

\item{values_nastates}{PARAM_DESCRIPTION}

\item{names_of_abstates}{PARAM_DESCRIPTION}

\item{values_abstates}{PARAM_DESCRIPTION}

\item{names_of_transitions}{PARAM_DESCRIPTION}

\item{values_of_transitions}{PARAM_DESCRIPTION}

\item{time}{PARAM_DESCRIPTION}

\item{timevar}{PARAM_DESCRIPTION}

\item{scale_inner}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
