% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_total_msm.R, R/msboxes_R.R
\name{freq_func_total_msm}
\alias{freq_func_total_msm}
\title{FUNCTION_TITLE}
\usage{
freq_func_total_msm(data_msm, id_msm, timevar, scale_msm = 1, Ntransitions)

freq_func_total_msm(data_msm, id_msm, timevar, scale_msm = 1, Ntransitions)
}
\arguments{
\item{data_msm}{PARAM_DESCRIPTION}

\item{id_msm}{PARAM_DESCRIPTION}

\item{timevar}{PARAM_DESCRIPTION}

\item{scale_msm}{PARAM_DESCRIPTION, Default: 1}

\item{Ntransitions}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
