% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dummies.R
\name{from_dummies}
\alias{from_dummies}
\title{Convert dummy variable columns to a categorical variable column}
\usage{
from_dummies(.data, col, from, default = NA_character_, remove = TRUE)
}
\arguments{
\item{.data}{A data frame}

\item{col}{The bare column name which will be created.}

\item{from}{The columns which contain the dummy variables.}

\item{default}{The value to use if no dummy variables match.}

\item{remove}{If \code{TRUE}, then remove the input column from the output
data frame.}
}
\value{
A data frame.
}
\description{
\code{from_dummies} converts dummy variable columns to a categorical variable
column.
}
\examples{
example_df <- data.frame(a = c(1, 0, 0), b = c(0, 1, 0))
from_dummies(example_df, x, c("a", "b"))
from_dummies(example_df, x, c("a", "b"), remove = FALSE)
from_dummies(example_df, x, c("a", "b"), default = "c")
}

