% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dummies.R
\name{from_dummies_}
\alias{from_dummies_}
\title{Standard evaluation version of \code{from_dummies}.}
\usage{
from_dummies_(.data, col, from, default = NA_character_, remove = TRUE)
}
\arguments{
\item{.data}{A data frame}

\item{col}{The bare column name which will be created.}

\item{from}{The columns which contain the dummy variables.}

\item{default}{The value to use if no dummy variables match.}

\item{remove}{If \code{TRUE}, then remove the input column from the output
data frame.}
}
\description{
This is an S3 generic.
}

