% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{from_exon_path_to_exonic_part_path}
\alias{from_exon_path_to_exonic_part_path}
\title{Convert exon path to exonic part path}
\usage{
from_exon_path_to_exonic_part_path(g, exonpath)
}
\arguments{
\item{g}{An \code{igraph} directed acyclic graph representing a gene splicing graph.}

\item{exonpath}{Integer vector. Sequence of edge indices (exon-level edges) forming a path
through the splicing graph, as returned by \code{from_vpath_to_exon_path}.}
}
\description{
Convert exon path to exonic part path
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_edge_names(g)
bubblepath <- c("R", "3", "5", "L")
exonpath <- grase::from_vpath_to_exon_path(g, bubblepath)
expart_path <- grase::from_exon_path_to_exonic_part_path(g, exonpath)
}
}
