% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{from_vpath_to_edge_path_simple}
\alias{from_vpath_to_edge_path_simple}
\title{convert vertex path to exon path}
\usage{
from_vpath_to_edge_path_simple(g, bubblepath)
}
\description{
convert vertex path to exon path
}
