% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{from_vpath_to_exon_path}
\alias{from_vpath_to_exon_path}
\title{Convert vertex path to exon path}
\usage{
from_vpath_to_exon_path(g, bubblepath)
}
\arguments{
\item{g}{An \code{igraph} directed acyclic graph representing a gene splicing graph.}

\item{bubblepath}{Character vector. Sequence of vertex names forming a path through a bubble.}
}
\description{
Convert vertex path to exon path
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_edge_names(g)
bubblepath <- c("R", "3", "5", "L")
epath <- grase::from_vpath_to_exon_path(g, bubblepath)
}
}
