% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{fully_deprecated}
\alias{fully_deprecated}
\title{fully_deprecated}
\usage{
fully_deprecated(x)
}
\arguments{
\item{x}{Integer input value.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A deprecated function with full lifecycle spec.
}
