% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future14C.R
\docType{data}
\name{future14C}
\alias{future14C}
\title{Future atmospheric 14C dataset for delta-delta calculation}
\format{
dataframe
}
\usage{
data(future14C)
}
\description{
Predicted atmospheric 14C data for 2020-2025.
}
\details{
Future atmospheric 14C predictions for the period 2020 to 2025 are projected using a time series  model trained on data covering the period 2000-2019. See Sierra (2018) for model details.\cr\cr
}
\references{
Sierra, C. "Forecasting atmospheric radiocarbon decline to pre-bomb values", Radiocarbon, Vol 60, Nr 4, 2018, p 1055.1066 DOI:10.1017/RDC.2018.33
}
\keyword{datasets}
