% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaudinski_2001.R
\docType{data}
\name{Gaudinski_2001}
\alias{Gaudinski_2001}
\title{Gaudinski 2001 example dataset}
\format{
list
}
\usage{
data(Gaudinski_2001)
}
\description{
Gaudinski_2001 data set
}
\details{
Example of an ISRaD entry. Selected as an example dataset as it contains data in all 8 tables of the ISRaD data structure: metadata, site, profile, flux, layer, interstitial, fraction, incubation.
}
\references{
Gaudinski, J., 2001, Belowground carbon cycling in three temperate forests of the eastern United States, University of California Irvine, Ph.D. thesis
}
\keyword{datasets}
