% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ntsa.R
\name{gen_emat}
\alias{gen_emat}
\title{Embedding matrix generation with multivariate and multi-timelags}
\usage{
gen_emat(df, cols, lags, knot = matrix(c(1, nrow(df)), nrow = 1))
}
\arguments{
\item{df}{vector, matrix, data.frame or tibble,
which contains time series to be embedded}

\item{cols}{Numeric or character vector,
which selects the df's column for the embedding matrix reconstruction}

\item{lags}{Integer vector, which sets
the time delay value for each coordinate in embedding matrix}

\item{knot}{vector or 2-column matrix, which specifies the knot
positions in the data}
}
\description{
\code{gen_emat} returns an embedding matrix with multivariate and multi-timelags
}
