% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_surrogate.R
\name{gen_rec_mat}
\alias{gen_rec_mat}
\title{Calculate recurrence matrix}
\usage{
gen_rec_mat(dmat, s = 0.1)
}
\arguments{
\item{dmat}{A distance matrix}

\item{s}{Numeric, which sets neighbourhood radius based on the proportion 's' of dmat}
}
\description{
\code{gen_rec_mat} returns (binary) 0-1 matrix based on distance matrix
}
