% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_surrogate.R
\name{gen_surr_ts}
\alias{gen_surr_ts}
\title{Surrogate generation with twin-surrogate}
\usage{
gen_surr_ts(
  X,
  col,
  knot = matrix(c(1, nrow(X)), nrow = 1),
  iter = 100,
  seed = NULL,
  twin_op = "RM",
  time_idx,
  phase_lock = TRUE,
  period = 12,
  lim = 10,
  range = c(0.125, seq(0.12, 0.05, -0.01), seq(0.15, 0.2, 0.01), 0.04),
  theta = 1
)
}
\arguments{
\item{X}{data matrix}

\item{col}{Integer, which specifies the column position of surrogate target variable in 'X'}

\item{knot}{Vector or 2-column matrix, which specifies the knot positions in the data}

\item{iter}{Integer, No. of surrogate iteration}

\item{seed, }{Integer, which sets RNG seed}

\item{twin_op}{Character, which can be "RM" or "NN"
if "RM", twin search is based on recurrence matrix,
otherwise is based on NN twin method}

\item{time_idx}{Vector of time index}

\item{phase_lock}{Logical, if TRUE, twin search is limited to the points having
time index same with the target (a.k.a. phase-locked twin surrogate, Ushio et al. 2018)}

\item{period}{Integer, the phase period (only works when phase_lock == TRUE)}

\item{lim}{Integer, the minimum No. of twins (only works when twin_op == "RM")}

\item{range}{Vector, the range of threshold in recurrence matrix
(only works when twin_op == "RM", see also \code{\link{gen_rec_mat}})}

\item{theta}{Numeric, which determines the locality parameter (works if twin_op == 'simplex')}
}
\description{
\code{gen_surr_ts} returns a column-wide piled surrogate data with TS algorithm
}
