% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ntsa.R
\name{gen_valid_idx}
\alias{gen_valid_idx}
\title{Generate valid time-index from knot data}
\usage{
gen_valid_idx(knot, idx_all)
}
\arguments{
\item{knot}{vector or 2-column matrix, which specifies the knot position in data matrix}

\item{idx_all}{a vector contains valid position in the whole data matrix}
}
\description{
\code{gen_valid_idx} returns a vector, which contains positions of valid rows in knot
}
