% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencode.TxDb.R
\name{gencode_txdb}
\alias{gencode_txdb}
\alias{gencode_source_url}
\title{Create a Gencode TxDb object}
\usage{
gencode_txdb(
  gtf_file,
  genome = c("hg19", "hg38"),
  chrs = paste0("chr", c(seq_len(22), "X", "Y", "M"))
)

gencode_source_url(version = "34", genome = c("hg38", "hg19"))
}
\arguments{
\item{gtf_file}{A \code{character(1)} with the path to a local Gencode GTF annotation file
(e.g., \code{"path/to/gencode.v31.annotation.gtf.gz"}).}

\item{genome}{A \code{character(1)} with the human genome version number. Valid
options are \code{'hg38'} or \code{'hg19'}.}

\item{chrs}{A \code{character()} vector with the chromosome (contig) names to
keep.}

\item{version}{A \code{character(1)} with the Gencode version number.}
}
\value{
A \link[GenomicFeatures:TxDb-class]{GenomicFeatures::TxDb} object.

A \code{character(1)} with the URL for the GTF Gencode file of interest.
}
\description{
This function builds a transcript database (\code{TxDb}) object which you can then
use to build a Gencode \code{GenomicState} object. This function will download
the data from Gencode, import it into R, process it and build the \code{TxDb}
object.
}
\examples{
\dontrun{
## Build from a local GTF file:
txdb <- gencode_txdb("path/to/gencode.v31.annotation.gtf.gz",
    genome = "hg19", chrs = "chr21")
txdb
}

## Locate the GTF file for Gencode version 31 for hg19
gencode_source_url(version = "31", genome = "hg19")
}
\references{
Based on code for the \code{brainflowprobes} package at:
\url{https://github.com/LieberInstitute/brainflowprobes/blob/master/data-raw/create_sysdata.R}
}
\author{
Leonardo Collado-Torres
}
