% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.r
\name{genDF}
\alias{genDF}
\title{genDF}
\usage{
genDF(n = 10, type = wide, digits = 2, proportion = FALSE,
  na.rate = 0)
}
\arguments{
\item{n}{numeric number of rows}

\item{type}{either wide or long, defaults to wide}

\item{digits}{for rounding, defaults to 2}

\item{proportion}{defaults to FALSE}

\item{na.rate}{numeric, defaults to 0}
}
\description{
GENERATE A RANDOM DATA SET.  CAN BE SET TO LONG OR WIDE. DATA SET HAS FACTORS
AND NUMERIC VARIABLES AND CAN. OPTIONALLY GIVE BUDGET EXPENDITURES AS A
PROPORTION. CAN ALSO TELL A PROPORTION OF CELLS TO BE MISSING VALUES
}
\examples{
genDF()
genDF(type="long")
genDF(20000)
genDF(prop=T)
genDF(na.rate=.3)
}
