% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_choose_2_analysis.R
\name{generate_path_pairs}
\alias{generate_path_pairs}
\title{Generate all n-choose-2 combinations of path indices}
\usage{
generate_path_pairs(n_paths)
}
\arguments{
\item{n_paths}{Integer. Number of paths to generate pair combinations from.}
}
\description{
Generate all n-choose-2 combinations of path indices
}
