% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{get_api_version}
\alias{get_api_version}
\title{Get API Version}
\usage{
get_api_version(...)
}
\arguments{
\item{...}{Not used}
}
\value{
character
}
\description{
Utility function to retrieve the version of the API from the plumber \verb{#* @apiVersion} roxygen tag.
}
