% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{get_bubble_depths}
\alias{get_bubble_depths}
\title{Get bubble depths from intervals}
\usage{
get_bubble_depths(intervals)
}
\arguments{
\item{intervals}{A list of integer vectors of length 2, each giving the
\code{c(source_idx, sink_idx)} topological positions of a bubble.}
}
\description{
Get bubble depths from intervals
}
\examples{
intervals <- list(c(1, 10), c(3, 7), c(12, 15))
depths <- grase::get_bubble_depths(intervals)
depths
}
