% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{get_bubble_variants}
\alias{get_bubble_variants}
\title{copied from SplicingGraphs}
\usage{
get_bubble_variants(txpathmat, sgnodetypes, i, j)
}
\description{
copied from SplicingGraphs
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_txpath_to_vertex_attr(g)
txmat <- grase::make_matrix_from_txpath_igraph(g)
sgtypes <- grase::get_sgnodetypes(txmat)
variants <- grase::get_bubble_variants(txmat, sgtypes, 2, 6)
}
}
