% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{get_bubble_variants_igraph}
\alias{get_bubble_variants_igraph}
\title{Get bubble variant partitions and paths from igraph splice graph}
\usage{
get_bubble_variants_igraph(g, v_start_idx, v_end_idx)
}
\arguments{
\item{g}{An \code{igraph} directed acyclic graph representing a gene splicing graph.}

\item{v_start_idx}{Integer. Topological index of the bubble source vertex.}

\item{v_end_idx}{Integer. Topological index of the bubble sink vertex.}
}
\description{
Get bubble variant partitions and paths from igraph splice graph
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_txpath_to_vertex_attr(g)
variants <- grase::get_bubble_variants_igraph(g, 2, 6)
# Returns list with $partition and $path elements
}
}
