% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{get_channel}
\alias{get_channel}
\title{Get a single channel by ID}
\usage{
get_channel(channel_id, token = twist_token())
}
\arguments{
\item{channel_id}{ID of the channel}

\item{token}{Authentication token}
}
\value{
Channel object
}
\description{
Get a single channel by ID
}
