% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{get_channel_threads}
\alias{get_channel_threads}
\title{Get threads in a channel}
\usage{
get_channel_threads(
  channel_id,
  token = twist_token(),
  as_ids = NULL,
  filter_by = NULL,
  limit = NULL,
  newer_than_ts = NULL,
  older_than_ts = NULL,
  before_id = NULL,
  after_id = NULL,
  workspace_id = NULL,
  is_pinned = NULL,
  is_starred = NULL,
  order_by = NULL,
  exclude_thread_ids = NULL
)
}
\arguments{
\item{channel_id}{Integer. The ID of the channel.}

\item{token}{Character. Authentication token.}

\item{as_ids}{Logical. If TRUE, only the IDs of the threads are returned. Default is FALSE.}

\item{filter_by}{Character. A filter can be one of "attached_to_me" or "everyone". Default is "everyone".}

\item{limit}{Integer. Limits the number of threads returned. Default is 20, maximum is 500.}

\item{newer_than_ts}{Integer. Limits threads to those newer than the specified Unix time.}

\item{older_than_ts}{Integer. Limits threads to those older than the specified Unix time.}

\item{before_id}{Integer. Limits threads to those with a lower ID than specified.}

\item{after_id}{Integer. Limits threads to those with a higher ID than specified.}

\item{workspace_id}{Integer. The ID of the workspace.}

\item{is_pinned}{Logical. If TRUE, only pinned threads are returned.}

\item{is_starred}{Logical. If TRUE, only starred threads are returned.}

\item{order_by}{Character. The order of the threads returned. Either "desc" (default, descending) or "asc" (ascending).}

\item{exclude_thread_ids}{List of Integers. The thread IDs that should be excluded from the results.}
}
\value{
List of thread objects.
}
\description{
Get threads in a channel
}
