% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch.R
\name{get_connection}
\alias{get_connection}
\title{Connect to a remote server}
\usage{
get_connection(host, user = credentials::guess_user(), password, port = 22)
}
\arguments{
\item{host}{the name of the remote system}

\item{user}{a string for the username on the remote system.  The default is to use the same name as the current user.}

\item{password}{the password for the user on the remote system}

\item{port}{this is usually 22 for ssh connections}
}
\value{
an object of class Rsch_connection
}
\description{
Connect to a remote server
}
\examples{
\dontrun{
connection <- get_connection(host = "cuckoo.lbnl.gov", user = "cliff", password = "TopSecret")
}
}
