% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_metadata.R
\name{get_database_metadata}
\alias{get_database_metadata}
\title{Get Database Metadata}
\usage{
get_database_metadata()
}
\value{
A list containing the database metadata for programmatic use.
}
\description{
Retrieves and prints comprehensive metadata about the IRW database, including version, number of tables,
data size, access information, and relevant URLs.
}
