% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{get_existing_thread_files}
\alias{get_existing_thread_files}
\title{Get existing thread files with metadata}
\usage{
get_existing_thread_files(channel_dir)
}
\arguments{
\item{channel_dir}{Directory for channel files}
}
\value{
Named list of thread file info
}
\description{
Get existing thread files with metadata
}
