% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_annotation.R
\name{get_gene_annotation}
\alias{get_gene_annotation}
\title{Retrieves the gene-level anntation file for this organism}
\usage{
get_gene_annotation(organism = c("mouse", "fly", "fish"))
}
\arguments{
\item{organism}{either "mouse", "fly", or "fish"}
}
\value{
the gene-level annotation data.frame for the organism
}
\description{
Retrieves the gene-level anntation file for this organism
}
