% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hyp_mat.R
\name{get_hyp_mat}
\alias{get_hyp_mat}
\title{Function to compute the hypotheses to test (closed test)}
\usage{
get_hyp_mat(n_hypothesis = 3, selected_hypothesis = 1)
}
\arguments{
\item{n_hypothesis}{num elementary hypotheses}

\item{selected_hypothesis}{selected hypothesis for closed test}
}
\value{
maximum value in a row
}
\description{
Function to compute the hypotheses to test (closed test)
}
\details{
eWHORM simulations
}
\author{
Marta Bofill Roig
}
