% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{get_mae}
\alias{get_mae}
\title{R wrapper function of getMAE in func_cpp.cpp}
\usage{
get_mae(x, y)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}
}
\description{
\code{get_mae} returns the MAE (Mean Absolute Error) between x & y
}
