% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_col_index.R
\name{get_max_col_index}
\alias{get_max_col_index}
\title{Function to get the column index of the maximum value in a row}
\usage{
get_max_col_index(row)
}
\arguments{
\item{row}{selected row}
}
\value{
maximum value in a row
}
\description{
Function to get the column index of the maximum value in a row
}
\details{
eWHORM simulations
}
\author{
Marta Bofill Roig
}
