% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{get_metadata}
\alias{get_metadata}
\title{Retrieve Metadata}
\usage{
get_metadata()
}
\description{
Retrieves the \code{metadata} dataset dynamically loaded into the package namespace.
}
\details{
The \code{metadata} dataset contains precomputed information about tables in the database,
such as \code{id_count}, \code{item_count}, \code{resp_count}, and others. This function ensures
that the dataset is accessed in a controlled manner.
}
\examples{
metadata <- get_metadata()
head(metadata)

}
