% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{get_ned}
\alias{get_ned}
\title{R wrapper function of getNEDist in func_cpp.cpp}
\usage{
get_ned(X)
}
\arguments{
\item{X}{Data matrix}
}
\description{
\code{get_NED} returns the Normal Euclidean Distance (NED) of the matrix X
}
