% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{get_rho}
\alias{get_rho}
\title{R wrapper function of getRho in func_cpp.cpp}
\usage{
get_rho(x, y)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}
}
\description{
\code{get_rho} returns the correlation coefficient between x & y
}
