% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{get_rmse}
\alias{get_rmse}
\title{R wrapper function of getRMSE in func_cpp.cpp}
\usage{
get_rmse(x, y)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}
}
\description{
\code{get_rmse} returns the RMSE (Root Mean Squared Error) between x & y
}
