% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_annotation.R
\name{get_sample_annotation}
\alias{get_sample_annotation}
\title{Retrieve the sample annotation}
\usage{
get_sample_annotation(organism = NULL, dataset = c("provided", "generated"))
}
\arguments{
\item{organism}{either "mouse", "fly", or "fish". If `NULL`, annotations
for all experiments are returned.}

\item{dataset}{either "provided" or "generated"}
}
\value{
a data.frame of sample annotations
}
\description{
Returns the sample annotations for a particular dataset.
}
