% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{get_sgnodetypes}
\alias{get_sgnodetypes}
\title{identifies nodes that are exon starts and ends and marks them with 1 and 2 respectively.
copied from SplicingGraphs}
\usage{
get_sgnodetypes(txpathmat, check.matrix = FALSE)
}
\arguments{
\item{txpathmat}{A logical matrix of transcript paths (rows = transcripts, columns = nodes).}

\item{check.matrix}{Logical. If \code{TRUE}, validate that no node is simultaneously an exon
start and an exon end. Default is \code{FALSE}.}
}
\description{
identifies nodes that are exon starts and ends and marks them with 1 and 2 respectively.
copied from SplicingGraphs
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_txpath_to_vertex_attr(g)
txmat <- grase::make_matrix_from_txpath_igraph(g)
node_types <- grase::get_sgnodetypes(txmat)
}
}
