% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{get_sgnodetypes_igraph}
\alias{get_sgnodetypes_igraph}
\title{Get node types (exon start/end) from igraph splice graph}
\usage{
get_sgnodetypes_igraph(g_tmp, check.graph = FALSE)
}
\arguments{
\item{g_tmp}{An \code{igraph} directed acyclic graph representing a gene splicing graph.}

\item{check.graph}{Logical. If \code{TRUE}, validate that no vertex is simultaneously an exon
start and an exon end. Default is \code{FALSE}.}
}
\description{
Get node types (exon start/end) from igraph splice graph
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
node_types <- grase::get_sgnodetypes_igraph(g)
}
}
