% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread.R
\name{get_thread}
\alias{get_thread}
\title{Get a single thread by ID}
\usage{
get_thread(thread_id, token = twist_token())
}
\arguments{
\item{thread_id}{ID of the thread}

\item{token}{Authentication token}
}
\value{
Thread as a list
}
\description{
Get a single thread by ID
}
