% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread.R
\name{get_thread_comments}
\alias{get_thread_comments}
\title{Get comments for a thread}
\usage{
get_thread_comments(thread_id, token = twist_token())
}
\arguments{
\item{thread_id}{ID of the thread}

\item{token}{Authentication token}
}
\value{
List of comments
}
\description{
Get comments for a thread
}
