% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/miniextendr-wrappers.R
\name{get_value}
\alias{get_value}
\title{Get value}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{get_value()} is an S7 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{CounterTraitS7}}
\item \code{\link{S7TraitCounter}}
}
}
