% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{get_workspace_channels}
\alias{get_workspace_channels}
\title{Get all channels in a workspace}
\usage{
get_workspace_channels(workspace_id, token = twist_token(), archived = FALSE)
}
\arguments{
\item{workspace_id}{ID of the workspace}

\item{token}{Authentication token}

\item{archived}{Whether to include archived channels (default: FALSE)}
}
\value{
List of channels
}
\description{
Get all channels in a workspace
}
