% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Border.R
\name{GetBorder}
\alias{GetBorder}
\title{Identify border cells
Function that reads in a mask and returns a mask of the border cells for the irrigular boundary.}
\usage{
GetBorder(mask_mat)
}
\arguments{
\item{mask_mat}{Matrix mask with values of 0 for cells outside the domain 1 for cells inside the domain}
}
\description{
Identify border cells
Function that reads in a mask and returns a mask of the border cells for the irrigular boundary.
}
