% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Myfunctions.R
\name{getresidual}
\alias{getresidual}
\title{Get residual of two integer numbers}
\usage{
getresidual(a, b)
}
\arguments{
\item{a,}{b are integer numbers with a > b}
}
\value{
the residual of a divided by b
}
\description{
Take an integer number divided by another and get the residual
}
\examples{
getresidual(111,11)             # gets residual 1
getresidual(12,3)                         # gets residual 0
}

