\name{gL0Learn-package}
\alias{gL0Learn-package}
\alias{gL0Learn}
\docType{package}
\title{
\packageTitle{gL0Learn}
}
\description{
\packageDescription{gL0Learn}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gL0Learn}
\packageIndices{gL0Learn}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{gL0Learn}

Maintainer: \packageMaintainer{gL0Learn}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
