% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle.R
\name{gL0Learn.bounds}
\alias{gL0Learn.bounds}
\title{gL0Learn.bounds}
\usage{
gL0Learn.bounds(lows = -Inf, highs = Inf)
}
\arguments{
\item{lows}{See `gL0Learn.fit` l0 penalty documentation.}

\item{highs}{See `gL0Learn.fit` l0 penalty documentation.}
}
\description{
Creates a C++ bounds object
 **Warning** This function does not perform value checking and assumes all 
 values conform to `gL0Learn.fit`'s checks. If this is not true, bare C++ 
 errors or segfaults may occur!
 The values must satisfy all of the following:
   1. `lows` and `highs` are of the same type
   2. Type must be real scalar or real matrix
}
