% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_constant.R
\name{gL0Learn.generate_constant}
\alias{gL0Learn.generate_constant}
\title{Generate a constantly correlated data set for gL0Learn}
\usage{
gL0Learn.generate_constant(n, p, rho, normalize, seed = 1, ...)
}
\arguments{
\item{n}{The number of observations to generated.
This will create a data matrix of shape (n, p)}

\item{p}{The number of features to generated.
This will create a data matrix of shape (n, p)}

\item{rho}{The correlation for ...}

\item{normalize}{The method for normalizing data
Currently only "covariance" is supported}

\item{seed}{A seed to a random number generated}
}
\description{
Computes the ...
}
