% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_independent.R
\name{gL0Learn.generate_independent}
\alias{gL0Learn.generate_independent}
\title{Generate a independent correlated data set for gL0Learn}
\usage{
gL0Learn.generate_independent(n, p, normalize, seed = 1)
}
\arguments{
\item{n}{See `gL0Learn.generate_synthetic` for details}

\item{p}{See `gL0Learn.generate_synthetic` for details}

\item{normalize}{See `gL0Learn.generate_synthetic` for details}

\item{seed}{See `gL0Learn.generate_synthetic` for details}
}
\description{
Computes the ...
}
