% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_regression.R
\name{gL0Learn.generate_regression}
\alias{gL0Learn.generate_regression}
\title{Generate a regression data set for gL0Learn}
\usage{
gL0Learn.generate_regression(n, p, k, val, normalize, seed = 1, ...)
}
\arguments{
\item{n}{The number of observations to generated.
This will create a data matrix of shape (n, p)}

\item{p}{The number of features to generated.
This will create a data matrix of shape (n, p)}

\item{normalize}{The method for normalizing data
Currently only "covariance" is supported}

\item{seed}{A seed to a random number generated}

\item{rwo}{The correlation for ...}
}
\description{
Computes the ...
}
