% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gL0Learn.is.real_matrix}
\alias{gL0Learn.is.real_matrix}
\title{gL0Learn.is.scalar}
\usage{
gL0Learn.is.real_matrix(x, dims)
}
\arguments{
\item{x}{The value to check for "scalar"-ness}
}
\description{
Determine if a value is a real matrix with dim `dims`
Source: # https://stackoverflow.com/questions/38088392/how-do-you-check-for-a-scalar-in-r/38088874 # nolint
}
\examples{
MISSING
}
