% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gL0Learn.is.real_vector}
\alias{gL0Learn.is.real_vector}
\title{gL0Learn.is.real_vector}
\usage{
gL0Learn.is.real_vector(x)
}
\arguments{
\item{x}{The value to check for "vector"-ness and element-wise scalar"-ness}
}
\description{
Determine if a value is a vector of scalar values
Source: # https://stackoverflow.com/questions/38088392/how-do-you-check-for-a-scalar-in-r/38088874 # nolint
}
\examples{
MISSING
}
