% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_search.R
\name{gL0Learn.linear_plot}
\alias{gL0Learn.linear_plot}
\title{Linear Plot of gL0Learn regularized loss.}
\usage{
gL0Learn.linear_plot(
  theta_opt,
  l0 = 0,
  l1 = 0,
  l2 = 0,
  lows = -1,
  highs = +1,
  atol = 0.001
)
}
\arguments{
\item{theta_opt}{See `gL0Learn.linear_search`}

\item{l0}{See `gL0Learn.linear_search`}

\item{l1}{See `gL0Learn.linear_search`}

\item{l2}{See `gL0Learn.linear_search`}

\item{lows}{See `gL0Learn.linear_search`}

\item{highs}{See `gL0Learn.linear_search`}

\item{atol}{See `gL0Learn.linear_search`}
}
\description{
Plots value of L(theta_opt, x) over a linspace from `lows` to 
`highs` to display shape of L for:
 L(theta_opt, x) = (|theta_opt - x|)**2 + l0|x|_0 + l1|x|_1 + + l2|x|_2
}
