% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{globset_is_match_opts}
\alias{globset_is_match_opts}
\title{globset_is_match_opts}
\usage{
globset_is_match_opts(
  patterns,
  paths,
  literal_separator,
  case_insensitive,
  backslash_escape
)
}
\arguments{
\item{patterns}{Character vector of glob patterns.}

\item{paths}{Character vector of paths to test.}

\item{literal_separator}{If TRUE, \code{*} cannot match \code{/}.}

\item{case_insensitive}{If TRUE, matching ignores case.}

\item{backslash_escape}{If TRUE, backslash escapes meta-characters.}
}
\description{
Match paths with explicit builder options.
}
