% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch.R
\name{goodbye}
\alias{goodbye}
\title{Closes the remote connection}
\usage{
goodbye(connection)
}
\arguments{
\item{connection}{an Rsch_connection object
 @examples
\dontrun{
connection <- get_connection(host = "cuckoo.lbnl.gov", user = "cliff", password = "TopSecret")
execute(connection,"grep \"nuclear\" /*")
results <- printout(connection)
goodbye(connection)
}}
}
\description{
Closes the remote connection
}
