% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{\%mx_scale\%}
\alias{\%mx_scale\%}
\title{The \verb{\%mx_scale\%} generic}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\verb{\%mx_scale\%} is an S7 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{MxS7Bag}}
}
}
