% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{greetings_last_as_named_and_unused_dots}
\alias{greetings_last_as_named_and_unused_dots}
\title{greetings_last_as_named_and_unused_dots}
\usage{
greetings_last_as_named_and_unused_dots(exclamations, ...)
}
\arguments{
\item{exclamations}{Integer count of exclamation marks.}

\item{...}{Additional arguments (captured but unused).}
}
\description{
Test dots as the last parameter with a named but unused dots argument.
}
