% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{greetings_with_nameless_dots}
\alias{greetings_with_nameless_dots}
\title{greetings_with_nameless_dots}
\usage{
greetings_with_nameless_dots(...)
}
\arguments{
\item{...}{Additional arguments (captured but unused).}
}
\description{
Test dots handling with nameless dots (\verb{_: ...}).
}
