% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exontest_functions.R
\name{group_by_event}
\alias{group_by_event}
\title{Group exon count data by gene and event}
\usage{
group_by_event(dat, col_y, col_n)
}
\description{
Group exon count data by gene and event
}
\examples{
\dontrun{
splitcnts <- read.table("bipartition.internal.exoncnt.combined.txt",
                        header = TRUE, row.names = NULL)
grouped_data <- grase::group_by_event(splitcnts, col_y = "diff", col_n = "n")
length(grouped_data)
}
}
