% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{group_by_sizes}
\alias{group_by_sizes}
\title{group_by_sizes}
\usage{
group_by_sizes(df, col)
}
\arguments{
\item{df}{A data.frame.}

\item{col}{The key column name.}
}
\description{
Group sizes in group order.
}
