% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{handle_channel_directory}
\alias{handle_channel_directory}
\title{Handle channel directory creation or renaming}
\usage{
handle_channel_directory(channel, workspace_dir)
}
\arguments{
\item{channel}{Channel object}

\item{workspace_dir}{Base workspace directory}
}
\value{
Path to the channel directory
}
\description{
Handle channel directory creation or renaming
}
