% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{head}
\alias{head}
\alias{head,fm-method}
\alias{head,fmV-method}
\alias{tail,fm-method}
\alias{tail,fmV-method}
\title{Return the First or Last part of an Object}
\usage{
\S4method{head}{fm}(x, n = 6L)

\S4method{tail}{fm}(x, n = 6L)

\S4method{head}{fmV}(x, n = 6L)

\S4method{tail}{fmV}(x, n = 6L)
}
\arguments{
\item{x}{a FlashR vector or matrix.}

\item{n}{a positive integer.}
}
\value{
An object (usually) like \code{x} but generally smaller.
}
\description{
Returns the first or last parts of a FlashR vector or matrix.
}
\examples{
mat <- fm.runif.matrix(100, 10)
head(mat)
tail(mat)
}

