% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hua2021.R
\docType{data}
\name{Hua2021}
\alias{Hua2021}
\title{Hua 2021 dataset for delta-delta calculation}
\format{
dataframe
}
\usage{
Hua2021
}
\description{
Atmospheric 14C data for the period 1941-2019.
}
\details{
Data from Hua et al., 2021
}
\references{
Hua, Q., Turnbull, J., Santos, G., Rakowski, A., Ancapichún, S., De Pol-Holz, R., . . . Turney, C. (2022). ATMOSPHERIC RADIOCARBON FOR THE PERIOD 1950–2019. Radiocarbon, 64(4), 723-745. doi:10.1017/RDC.2021.95
}
\keyword{datasets}
