% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{hybrid_as_list}
\alias{hybrid_as_list}
\title{Prefer list conversion when multiple IntoR paths exist}
\source{
Generated by miniextendr from Rust fn \code{hybrid_as_list}
}
\usage{
hybrid_as_list(x)
}
\description{
Wraps a type implementing ExternalPtr, IntoList, and RNativeType with \code{AsList} to force VECSXP.
}
\examples{
hybrid_as_list(1L)
}
\keyword{internal}
