% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{hybrid_as_native}
\alias{hybrid_as_native}
\title{Prefer native scalar conversion when multiple IntoR paths exist}
\source{
Generated by miniextendr from Rust fn \code{hybrid_as_native}
}
\usage{
hybrid_as_native(x)
}
\description{
Uses \code{AsRNative} to force a length-1 integer vector.
}
\examples{
hybrid_as_native(1L)
}
\keyword{internal}
