% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{hybrid_as_ptr}
\alias{hybrid_as_ptr}
\title{Prefer external pointer conversion when multiple IntoR paths exist}
\source{
Generated by miniextendr from Rust fn \code{hybrid_as_ptr}
}
\usage{
hybrid_as_ptr(x)
}
\description{
Uses \code{AsExternalPtr} to force EXTPTRSXP even though list/native are available.
}
\examples{
hybrid_as_ptr(1L)
}
\keyword{internal}
