% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulses.R
\name{identify_pulses}
\alias{identify_pulses}
\title{Identify pulse events (runs above/below threshold)}
\usage{
identify_pulses(flow_vector, threshold, above = TRUE)
}
\description{
Identify pulse events (runs above/below threshold)
}
\keyword{internal}
