% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recession.R
\name{identify_recession_events}
\alias{identify_recession_events}
\title{Identify recession events using look-ahead algorithm}
\usage{
identify_recession_events(Q_vector, min_length = NULL)
}
\description{
Identify recession events using look-ahead algorithm
}
\keyword{internal}
