% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{impl_dots_s3_add_with_dots}
\alias{impl_dots_s3_add_with_dots}
\title{Add a value and the number of method dots via S3 dispatch}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Add a value and the number of method dots via S3 dispatch
}
