% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{impl_dots_s3_ctor_dots}
\alias{impl_dots_s3_ctor_dots}
\title{Return the constructor dots count via S3 dispatch}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Return the constructor dots count via S3 dispatch
}
