% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{impl_return_string}
\alias{impl_return_string}
\alias{impl_return_vec}
\title{impl_return_string}
\usage{
impl_return_string()

impl_return_vec()
}
\description{
\verb{-> impl IntoR} with String behind it.

\verb{-> impl IntoR} with \verb{Vec<i32>} behind it.
}
