% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incubationTime.R
\name{incubationTime}
\alias{incubationTime}
\title{Returns the incubation time of the entries in the database}
\usage{
incubationTime(database)
}
\arguments{
\item{database}{A list with the SIDb structure}
}
\value{
A data frame of incubation times and their units
}
\description{
Returns the incubation time of the entries in the database
}
\examples{
incubTimes=incubationTime(database=sidb)
}
