% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{indexmap_duplicate_key}
\alias{indexmap_duplicate_key}
\title{indexmap_duplicate_key}
\usage{
indexmap_duplicate_key()
}
\description{
Test that duplicate key insertion keeps the last value.
}
