% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{indexmap_keys}
\alias{indexmap_keys}
\title{indexmap_keys}
\usage{
indexmap_keys(map)
}
\arguments{
\item{map}{Named integer list from R.}
}
\description{
Test extracting keys from an IndexMap in insertion order.
}
