% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{indexmap_len}
\alias{indexmap_len}
\title{indexmap_len}
\usage{
indexmap_len(map)
}
\arguments{
\item{map}{Named integer list from R.}
}
\description{
Test getting the number of entries in an IndexMap.
}
