% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{indexmap_order_preserved}
\alias{indexmap_order_preserved}
\title{indexmap_order_preserved}
\usage{
indexmap_order_preserved()
}
\description{
Test that insertion order is preserved in IndexMap keys.
}
