% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{indexmap_roundtrip_dbl}
\alias{indexmap_roundtrip_dbl}
\title{indexmap_roundtrip_dbl}
\usage{
indexmap_roundtrip_dbl(map)
}
\arguments{
\item{map}{Named double list from R.}
}
\description{
Test IndexMap<String, f64> roundtrip through R named list.
}
