% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{indexmap_roundtrip_int}
\alias{indexmap_roundtrip_int}
\title{indexmap_roundtrip_int}
\usage{
indexmap_roundtrip_int(map)
}
\arguments{
\item{map}{Named integer list from R.}
}
\description{
Test IndexMap<String, i32> roundtrip through R named list.
}
