% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{indexmap_roundtrip_str}
\alias{indexmap_roundtrip_str}
\title{indexmap_roundtrip_str}
\usage{
indexmap_roundtrip_str(map)
}
\arguments{
\item{map}{Named character list from R.}
}
\description{
Test IndexMap<String, String> roundtrip through R named list.
}
