% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartition_analysis.R
\name{infer_diff_exons_from_tx_groups}
\alias{infer_diff_exons_from_tx_groups}
\title{Infer differential exonic parts given two groups of transcripts (e.g. changed vs constant)}
\usage{
infer_diff_exons_from_tx_groups(gene, g, tx_changed, tx_constant, outdir)
}
\description{
Infer differential exonic parts given two groups of transcripts (e.g. changed vs constant)
}
