% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_Queue.R
\name{InitQueue}
\alias{InitQueue}
\title{Initilze queue for topographic processing}
\usage{
InitQueue(dem, initmask, domainmask, border, d4 = c(1, 2, 3, 4))
}
\arguments{
\item{dem}{Elevation matrix}

\item{initmask}{Mask of the same dimensions of dem denoting a subset of cells to be considered for the queue (e.g. if you want to setup a run starting with only river cells). 
Note: if no init mask is included every border cell will be added to the queue}

\item{domainmask}{Mask of the domain extent to be considered. If no domain mask is provided boundaries will be calcualted from the rectangular extent}

\item{border}{Alternatively you can input your own border rathern than having it be calculated from the domain mask. For example if you want to have the river network and the borders combined you can input this as a border.}

\item{d4}{directional numbering system: the numbers you want to assigne to down, left, top,right (defaults to 1,2,3,4)}
}
\value{
This function returns five outputs:

1.marked - A matrix indicating the outlet cells that were identified (1=outlet, 0=not outlet)

2.queue - A list of the outlet cells. This has three columns: x, y, elevation 
This mask should match the river reach mask you provide the function if all cells were appropriately processed.

3.initmask- A matrix indicating the cells that were input as potential output points.

4.basin- A matrix indicating the basin number for each outlet point (each outlet is assigned a unique basin number)

5.direciton- A matrix indicating the flow direction for each outlet point. The numbering scheme will follow the d4 numbering scheme provided as input.
}
\description{
Sets up a queue and initilizes marked and step matrices for DEM processing
}
