% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.r
\name{insertNA}
\alias{insertNA}
\title{insertNA}
\usage{
insertNA(df, prop = 0.1)
}
\arguments{
\item{df}{a data frame}

\item{prop}{numeric}
}
\description{
RANDOMLY INSERT A CERTAIN PROPORTION OF NAs INTO A DATAFRAME
}
\examples{
insertNA(genProportions(ncol=10, nrow=10)  , .1)
}
