% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.rpkg.R
\name{install.rpkg}
\alias{install.rpkg}
\title{Install R packages}
\usage{
install.rpkg(cran.pkg = NULL, bioconductor.pkg = NULL, github.pkg = NULL)
}
\arguments{
\item{cran.pkg}{character. names of packages available in
\href{https://cran.r-project.org/web/packages/available_packages_by_name.html}{cran repository}.}

\item{bioconductor.pkg}{character. names of packages available in
\href{https://www.bioconductor.org/}{bioconductor}.}

\item{github.pkg}{character. names of packages available in
\href{https://github.com}{github}.}
}
\value{
Returns a character vector of newly installed packages.
A warning is generated if any package could not be installed.
}
\description{
Install R packages
}
\examples{
cran.pkg = c(
  "Rdpack", "Rcpp",
  "foreach", "doParallel", "flock",
  "igraph", "data.table", "dplyr", "tidyr", "reshape2",
  "argparser",
  "MASS", "PRROC", "lme4", "svd", "stringr", "feather",
  "ggplot2", "cowplot", "corrplot", "RColorBrewer", "gridExtra",
  "MatrixEQTL",
  "Seurat",
  "msigdbr",
  "corpcor", "glasso", "ppcor", "infotheo"
)
bioconductor.pkg = c(
  "SharedObject",
  "STRINGdb", "KEGGgraph", "biomaRt", "Biostrings",
  "limma", "edgeR", "DESeq2", "sva",
  "IRanges", "GenomicRanges",
  "GENIE3", "minet", "WGCNA"
)
github.pkg = c(
  "alorchhota/autil/ioutil",
  "alorchhota/autil/miscutil",
  "alorchhota/autil/genomicsutil",
  "alorchhota/autil/mappabilityutil"
)
install.rpkg(cran.pkg, bioconductor.pkg, github.pkg)

}
