% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{integer}
\alias{as.integer,fm-method}
\alias{as.integer,fmV-method}
\alias{integer}
\title{Integer Vectors}
\usage{
\S4method{as.integer}{fm}(x)

\S4method{as.integer}{fmV}(x)
}
\arguments{
\item{x}{a FlashR object to be coerced or tested.}
}
\value{
\code{as.integer} returns an integer FlashR object,
\code{is.integer} returns a logical value.
}
\description{
\code{as.integer} coerces objects of type \code{"integer"}.
\code{is.integer} is a more general test of an object being
interpretable as integers.
}
\examples{
vec <- as.integer(fm.runif(1000, min=0, max=10))
}

