% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{integer_sequence_list}
\alias{integer_sequence_list}
\title{integer_sequence_list}
\usage{
integer_sequence_list(n)
}
\arguments{
\item{n}{Number of elements in the list.}
}
\value{
A list where element i contains the vector 1:i.
}
\description{
Create a list ALTREP where each element is an integer sequence.
}
\examples{
lst <- integer_sequence_list(3L)
lst[[1]]  # c(1L)
lst[[2]]  # c(1L, 2L)
lst[[3]]  # c(1L, 2L, 3L)
}
