% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{unsafe_C_check_interupt_after}
\alias{unsafe_C_check_interupt_after}
\alias{unsafe_C_check_interupt_unwind}
\title{unsafe_C_check_interupt_after}
\source{
Generated by miniextendr from Rust fn \code{C_check_interupt_after}

Generated by miniextendr from Rust fn \code{C_check_interupt_unwind}
}
\usage{
unsafe_C_check_interupt_after()

unsafe_C_check_interupt_unwind()
}
\description{
Test R_CheckUserInterrupt after a 2-second sleep.

Test R_CheckUserInterrupt inside with_r_unwind_protect_or_raise after a 2-second sleep.
}
