% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{into_r_error_length_overflow}
\alias{into_r_error_length_overflow}
\title{into_r_error_length_overflow}
\usage{
into_r_error_length_overflow()
}
\description{
Trigger a LengthOverflow error and return its message.
}
