% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{into_r_error_string_too_long}
\alias{into_r_error_string_too_long}
\title{into_r_error_string_too_long}
\usage{
into_r_error_string_too_long()
}
\description{
Trigger a StringTooLong error and return its message.
}
