% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{into_r_error_string_too_long}
\alias{into_r_error_string_too_long}
\alias{into_r_error_length_overflow}
\alias{into_r_error_inner}
\title{into_r_error_string_too_long}
\usage{
into_r_error_string_too_long()

into_r_error_length_overflow()

into_r_error_inner()
}
\description{
Trigger a StringTooLong error and return its message.

Trigger a LengthOverflow error and return its message.

Trigger an Inner error and return its message.
}
