% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{IntVecIter}
\alias{IntVecIter}
\alias{IntVecIter$new}
\alias{$.IntVecIter}
\alias{[[.IntVecIter}
\alias{IntVecIter$RIterator$next_item}
\alias{IntVecIter$RIterator$count}
\alias{IntVecIter$RIterator$collect_n}
\alias{IntVecIter$RIterator$skip}
\alias{IntVecIter$RIterator$nth}
\title{IntVecIter Class}
\usage{
IntVecIter

\method{$}{IntVecIter}(self, name)

\method{[[}{IntVecIter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
IntVecIter inherent methods: constructor from integer vector.

Arguments of \code{IntVecIter$new()}:

\describe{
\item{\code{data}}{Integer vector of elements to iterate over.}
}
}
