% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{invisibly_option_return_none}
\alias{invisibly_option_return_none}
\title{invisibly_option_return_none}
\usage{
invisibly_option_return_none()
}
\description{
Test that returning None from Option<()> produces an R error.
}
