% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{invisibly_option_return_some}
\alias{invisibly_option_return_some}
\title{invisibly_option_return_some}
\usage{
invisibly_option_return_some()
}
\description{
Test that returning Some(()) from Option<()> is invisible in R.
}
