% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{invisibly_result_return_ok}
\alias{invisibly_result_return_ok}
\title{invisibly_result_return_ok}
\usage{
invisibly_result_return_ok()
}
\description{
Test that returning Ok(()) from Result<(), ()> is invisible in R.
}
