% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{invisibly_return_arrow}
\alias{invisibly_return_arrow}
\title{invisibly_return_arrow}
\usage{
invisibly_return_arrow()
}
\description{
Test that a function with explicit -> () return is invisible in R.
}
