% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite.r
\name{irw_cite}
\alias{irw_cite}
\title{Generate Citation or BibTeX for a Table}
\usage{
irw_cite(table_name, if_bibtex = FALSE)
}
\arguments{
\item{table_name}{A character string specifying the name of the table.}

\item{if_bibtex}{Logical; if TRUE, returns the BibTeX entry instead of the citation. Defaults to FALSE.}
}
\value{
A character string containing the citation or BibTeX entry, NA if neither is available, or NULL if the table does not exist.
}
\description{
Generates a citation or BibTeX entry for a table in the data warehouse.
}
