% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{irw_db_metadata}
\alias{irw_db_metadata}
\title{Retrieve Database Metadata}
\usage{
irw_db_metadata()
}
\value{
A list containing the database metadata, including:
\item{version}{The version of the database.}
\item{table_count}{The total number of tables in the database.}
\item{created_at}{The creation timestamp of the database (formatted as a string).}
\item{updated_at}{The last update timestamp of the database (formatted as a string).}
\item{total_size_gb}{The total size of the database in gigabytes.}
\item{active_data_size_gb}{The size of active tabular data in gigabytes.}
\item{doi}{The DOI of the database.}
\item{dataset_url}{The URL to the database.}
\item{documentation}{The link to the database documentation.}
\item{methodology}{The link to the database methodology.}
\item{usage_info}{The link to usage information.}
}
\description{
Fetches and displays comprehensive metadata for the IRW database, including the database version,
number of tables, total data size, and relevant URLs.
}
\examples{
\dontrun{
  db_metadata <- irw_db_metadata()
  print(db_metadata)
}
}
