% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{irw_recode_key}
\alias{irw_recode_key}
\title{Retrieve the Recode Key from an Object Returned by irw_recode()}
\usage{
irw_recode_key(x)
}
\arguments{
\item{x}{An object returned by \code{irw_recode()}.}
}
\value{
A data frame with columns \code{column}, \code{original}, and
\code{code}.
}
\description{
Retrieve the Recode Key from an Object Returned by irw_recode()
}
\seealso{
\code{\link{irw_recode}}, \code{\link{irw_decode}}
}
