% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{irw_rename}
\alias{irw_rename}
\title{Apply renaming to an object with names}
\usage{
irw_rename(x, ...)
}
\arguments{
\item{x}{an object with names to clean}

\item{...}{additional arguments to be passed down to irw_name_fix}
}
\value{
the object with cleaned names
}
\description{
Apply renaming to an object with names
}
\details{
This function is the main standardization function for IRW objects. It applies irw_name_fix to the names of an object.
}
\examples{
irw_rename(data.frame("A B" = 1:3, "C D" = 4:6))
irw_rename(c("A B", "C D"))
}
