% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{irw_visualize}
\alias{irw_visualize}
\title{Visualize Metadata Distributions}
\usage{
irw_visualize(ranges = list())
}
\arguments{
\item{ranges}{A named list specifying the lower and upper range for each attribute to visualize.
If not provided, the full range of data will be used for each attribute. Valid attributes include:
- \code{"id_count"}: Number of unique IDs per dataset.
- \code{"item_count"}: Number of unique items per dataset.
- \code{"resp_count"}: Total number of responses per dataset.
- \code{"sparsity"}: Sparsity measure (between 0 and 1).}
}
\value{
No return value. Outputs histograms to the console.
}
\description{
Generates histograms for the distributions of key attributes in the IRW metadata.
}
\examples{
# Example with default ranges:
irw_visualize()

# Example with custom ranges:
irw_visualize(ranges = list(id_count = c(0, 1000), sparsity = c(0.1, 0.9)))
}
