% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\docType{methods}
\name{is.finite}
\alias{is.finite}
\alias{is.finite,fm-method}
\alias{is.finite,fmV-method}
\alias{is.infinite,fm-method}
\alias{is.infinite,fmV-method}
\alias{is.nan,fm-method}
\alias{is.nan,fmV-method}
\title{Finite, Infinite and NaN Numbers}
\usage{
\S4method{is.nan}{fm}(x)

\S4method{is.nan}{fmV}(x)

\S4method{is.infinite}{fm}(x)

\S4method{is.infinite}{fmV}(x)

\S4method{is.finite}{fm}(x)

\S4method{is.finite}{fmV}(x)
}
\arguments{
\item{x}{a FlashR object}
}
\value{
A logical vector of the same length as \code{x}
}
\description{
\code{is.finite} and \code{is.infinite} return a vector of the same length
as \code{x}, indicating which elements are finite (not infinite and not
missing) or infinite.
}
\examples{
mat <- fm.runif.matrix(100, 10)
is.finite(mat)
is.infinite(mat)
is.nan(mat)
}

