% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{is.widget}
\alias{is.widget}
\title{Check if an object is a widget}
\source{
Generated by miniextendr from Rust fn \code{is_widget}
}
\usage{
is.widget(x)
}
\arguments{
\item{x}{Value to check.}
}
\description{
Check if an object is a widget
}
