% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{is_proper_subset}
\alias{is_proper_subset}
\title{Check if set a is a proper subset of b}
\usage{
is_proper_subset(a, b)
}
\arguments{
\item{a}{A vector. The candidate subset.}

\item{b}{A vector. The candidate superset.}
}
\description{
Check if set a is a proper subset of b
}
