% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.Cstocks.R
\name{ISRaD.extra.Cstocks}
\alias{ISRaD.extra.Cstocks}
\title{ISRaD.extra.Cstocks}
\usage{
ISRaD.extra.Cstocks(database)
}
\arguments{
\item{database}{ISRaD dataset object.}
}
\value{
returns ISRaD_data object with filled columns
}
\description{
Calculates soil organic carbon stock
}
\details{
Function first fills lyr_bd_samp and lyr_c_org. Notes: 1) SOC stocks can only be calculated if organic carbon concentration and bulk density data are available, 2) SOC stocks are calculated for the fine earth fraction (<2mm).
}
\examples{
# Load example dataset Gaudinski_2001
database <- ISRaD::Gaudinski_2001
database.x <- ISRaD.extra.Cstocks(database)
}
\author{
J. Beem-Miller
}
