% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.fill_14c.R
\name{ISRaD.extra.fill_14c}
\alias{ISRaD.extra.fill_14c}
\title{ISRaD.extra.fill_14c}
\usage{
ISRaD.extra.fill_14c(database)
}
\arguments{
\item{database}{ISRaD dataset object.}
}
\value{
returns ISRaD_data object with filled delta 14C columns
}
\description{
: Fills delta 14C from fraction modern if delta 14C not reported.
}
\details{
: Warning: xxx_obs_date_y columns must be filled for this to work!
}
\examples{
# Load example dataset Gaudinski_2001
database <- ISRaD::Gaudinski_2001
# Fill dates
database.x <- ISRaD.extra.fill_dates(database)
# Fill delta 14C from fraction modern
database.x <- ISRaD.extra.fill_14c(database.x)
}
\references{
: Stuiver and Polach, 1977
}
\author{
: J. Beem-Miller & A. Hoyt
}
