% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.fill_CN.R
\name{ISRaD.extra.fill_CN}
\alias{ISRaD.extra.fill_CN}
\title{ISRaD.extra.fill_c_to_n}
\usage{
ISRaD.extra.fill_CN(database)
}
\arguments{
\item{database}{ISRaD dataset object}
}
\value{
ISRaD database object with gap-filled C:N ratios in a new column
}
\description{
Checks for existing C:N ratios and simply calculates missing values
}
\details{
If missing, C:N ratio is calculated from data provided, avoiding inorganic C if possible
}
\examples{
\donttest{
# Load example dataset Gaudinski_2001
database <- ISRaD::Gaudinski_2001
# Fill ISRaD.extra data
database.x <- ISRaD.extra.fill_CN(database)
)
}
}
\author{
Shane Stoner
}
