% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.fill_country.R
\name{ISRaD.extra.fill_country}
\alias{ISRaD.extra.fill_country}
\title{ISRaD.extra.fill_country}
\usage{
ISRaD.extra.fill_country(database, continent = FALSE, region = FALSE)
}
\arguments{
\item{database}{ISRaD dataset object.}

\item{continent}{Boolean noting whether a column should be added for extracted continent (6 continent model: "Eurasia")}

\item{region}{Boolean noting whether a column should be added for extracted region (7 continent model: "Europe", "Asia")}
}
\value{
ISRaD_data object with extracted country names.
}
\description{
Fills country code from profile and site coordinates. Copied mostly from https://stackoverflow.com/questions/14334970/convert-latitude-and-longitude-coordinates-to-country-name-in-r
}
\examples{
# Load example dataset Gaudinski_2001
database <- ISRaD::Gaudinski_2001
# Fill profile coordinates
database.x <- ISRaD.extra.fill_country(database)
}
\author{
Shane Stoner
}
