% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.fill_fm.R
\name{ISRaD.extra.fill_fm}
\alias{ISRaD.extra.fill_fm}
\title{ISRaD.extra.fill_fm}
\usage{
ISRaD.extra.fill_fm(database)
}
\arguments{
\item{database}{ISRaD dataset object.}
}
\value{
returns ISRaD_data object with filled fraction modern columns
}
\description{
Fills fraction modern from delta 14C if fraction modern not reported.
}
\details{
: Warning: xxx_obs_date_y columns must be filled for this to work!
}
\examples{
# Load example dataset Gaudinski_2001
database <- ISRaD::Gaudinski_2001
# Fill dates
database.x <- ISRaD.extra.fill_dates(database)
# Fill fraction modern from delta 14C
database.x <- ISRaD.extra.fill_fm(database.x)
}
\references{
: Stuiver and Polach, 1977
}
\author{
: J. Beem-Miller & A. Hoyt
}
