% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.fill_soilorders.R
\name{ISRaD.extra.fill_soilorders}
\alias{ISRaD.extra.fill_soilorders}
\title{ISRaD.extra.fill_soilorders}
\usage{
ISRaD.extra.fill_soilorders(database)
}
\arguments{
\item{database}{ISRaD dataset object.}
}
\value{
returns ISRaD_data object with filled pro_usda_soil_order column
}
\description{
Fills pro_usda_soil_order field from pro_soil_taxon field.
}
\details{
This function is a static conversion script written at the Fall 2018 Powell Center workshop and therefore performance is not guaranteed for new entries.

Back fills pro_usda_soil_order based on USDA classifications
}
\examples{
# Obtain current ISRaD data
database <- ISRaD.getdata(tempdir(), dataset = "full", extra = FALSE)
database.x <- ISRaD.extra.fill_soilorders(database)
}
