% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.geospatial.keys.R
\name{ISRaD.extra.geospatial.keys}
\alias{ISRaD.extra.geospatial.keys}
\title{ISRaD.extra.geospatial.keys}
\usage{
ISRaD.extra.geospatial.keys(database, geodata_keys)
}
\arguments{
\item{database}{ISRaD dataset object}

\item{geodata_keys}{directory where geospatial data are found}
}
\value{
returns updated ISRaD_extra object with recoded columns
}
\description{
Recode numeric values from categorical geospatial data products
}
\details{
Generic function that reads .csv files paired with categorical raster data and recodes extracted data in the ISRaD_extra object.
For the function to work, the .csv filenames must be identical to the corresponding raster filenames, except for the file extension.
Additionally, the first column of the .csv file must contain the numeric identifier and the second column the corresponding character value.
}
