% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.geospatial.R
\name{ISRaD.extra.geospatial}
\alias{ISRaD.extra.geospatial}
\title{ISRaD.extra.geospatial}
\usage{
ISRaD.extra.geospatial(database, geodata_directory,
  crs = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0",
  fillWorldClim = TRUE)
}
\arguments{
\item{database}{ISRaD dataset object}

\item{geodata_directory}{directory where geospatial data are found}

\item{crs}{coordinate reference system used for geospatial datasets}

\item{fillWorldClim}{option to fill climate data from the Worldclim dataset (downloads data from web)}
}
\value{
returns updated ISRaD_extra object with new columns at the profile level
}
\description{
Extracts data from a user-supplied raster file and adds data as a new variable at the profile level
}
\details{
Generic function that uses geographic coordinates of profiles to extract data from one or more raster files.
Raster data will be added as new variables at the profile level.
The new variable name will be a concatenation of "pro_", and the file name (excluding the file extension).
The ISRaD recommended file name convention uses a 6 component string, separated by "_". Missing components can be replaced with "x" ("x"s will be dropped before creating variable names). The 6 components are as follows:
1) Short description of the data type, e.g. "bd" for bulk density
2) Top layer depth (numeric, cm)
3) Bottom layer depth (numeric, cm)
4) Year of data observation (numeric)
5) Data units (e.g. mmyr for mean annual precipitation)
6) Any relevant notes
Coordinate reference system can be specified with the "crs" argument; default is WGS84. Note that all files in geodata_directory must use the same crs.
Option "fillWorldClim" fills climate data from worldclim V1.4 at 2.5 resolution (http://www.worldclim.org/bioclim). Variable descriptions are as follows:
BIO1 = Annual Mean Temperature,
BIO2 = Mean Diurnal Range (Mean of monthly (max temp - min temp)),
BIO3 = Isothermality (BIO2/BIO7) (* 100),
BIO4 = Temperature Seasonality (standard deviation *100),
BIO5 = Max Temperature of Warmest Month,
BIO6 = Min Temperature of Coldest Month,
BIO7 = Temperature Annual Range (BIO5-BIO6),
BIO8 = Mean Temperature of Wettest Quarter,
BIO9 = Mean Temperature of Driest Quarter,
BIO10 = Mean Temperature of Warmest Quarter,
BIO11 = Mean Temperature of Coldest Quarter,
BIO12 = Annual Precipitation,
BIO13 = Precipitation of Wettest Month,
BIO14 = Precipitation of Driest Month,
BIO15 = Precipitation Seasonality (Coefficient of Variation),
BIO16 = Precipitation of Wettest Quarter,
BIO17 = Precipitation of Driest Quarter,
BIO18 = Precipitation of Warmest Quarter,
BIO19 = Precipitation of Coldest Quarter
}
