% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.geospatial.Zheng.R
\name{ISRaD.extra.geospatial.Zheng}
\alias{ISRaD.extra.geospatial.Zheng}
\title{ISRaD.extra.geospatial.Zheng}
\usage{
ISRaD.extra.geospatial.Zheng(database, geodata_soil_directory)
}
\arguments{
\item{database}{soilcarbon dataset object}

\item{geodata_soil_directory}{directory where 0.5 degree geospatial soil and climate data are located}
}
\value{
returns new ISRaD_extra object with extracted 0.5 degree MAT, MAP, land cover, and soil order for every profile
}
\description{
Extracts MAT, MAP, MODIS land cover, and soil order from global 0.5 degree gridded data products
}
\details{
Uses geographic coordinates of profiles (including those filled from site-level coordinates) to extract MAT, MAP, land cover, and soil order at 0.5 degree spatial resolution. These products were derived for global mapping purposes by Yujie He and Zheng Shi. Note: MODIS 0.5 degree land cover (pro_0.5_landCover_MODIS) was reclassified from 16 classes to 10 classes (pro_0.5_landCover) to match observations for He et al. (2016) (doi: 10.1126/science.aad4273)
}
