% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.R
\name{ISRaD.extra}
\alias{ISRaD.extra}
\title{ISRaD.extra}
\usage{
ISRaD.extra(database, geodata_clim_directory, geodata_soil_directory,
  geodata_pet_directory)
}
\arguments{
\item{database}{soilcarbon dataset object}

\item{geodata_clim_directory}{directory where geospatial climate data are found}

\item{geodata_soil_directory}{directory where geospatial soil data are found}

\item{geodata_pet_directory}{directory where geospatial pet data are found}
}
\value{
returns new ISRaD_extra object with derived, transformed, and filled columns
}
\description{
Fills in transformed and geospatial data where possible, generating an enhanced version of ISRaD.
}
\details{
Fills fraction modern, delta 14C, delta-delta values, profile coordinates, and SOC stocks frmo entered data; fills soil and climatic data from external geospatial data products
}
\examples{
# Load example dataset Gaudinski_2001
database <- ISRaD::Gaudinski_2001
# Fill ISRaD.extra data
database.x <- ISRaD.extra(database,
                          geodata_clim_directory = tempdir(),
                          geodata_soil_directory = system.file("extdata", "geodata_soil_directory", package = "ISRaD"))
                          geodata_pet_directory = system.file("extdata", "geodata_pet_directory", package = "ISRaD"))
}
