% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.rep.count.all.R
\name{ISRaD.rep.count.all}
\alias{ISRaD.rep.count.all}
\title{ISRaD.rep.count.all}
\usage{
ISRaD.rep.count.all(database = NULL)
}
\arguments{
\item{database}{ISRaD data object}
}
\description{
Generates a report of counts of observations at each level of the database
}
\examples{
# Obtain current ISRaD data
database <- ISRaD.getdata(tempdir(), dataset = "full", extra = F, force_downlaod = F)
ISRaD.rep.count.all(database)
}
