% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.rep.entry.stats.R
\name{ISRaD.rep.entry.stats}
\alias{ISRaD.rep.entry.stats}
\title{ISRaD.rep.entry.stats}
\usage{
ISRaD.rep.entry.stats(database = NULL)
}
\arguments{
\item{database}{ISRaD data object}
}
\description{
Generates a report of metadata statistics for all entries
}
\examples{
# Load example dataset Gaudinski_2001
database <- ISRaD::Gaudinski_2001
ISRaD.rep.entry.stats(database)
}
