% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.report.R
\name{ISRaD.reports}
\alias{ISRaD.reports}
\title{ISRaD.report}
\usage{
ISRaD.reports(database = NULL, report = "count.all")
}
\arguments{
\item{database}{ISRaD data object}

\item{report}{Parameter to define which type of report you want. The default is "count.all" other options include "entry.stats", "count.frc", or "site.map".}
}
\description{
Generate basic summary reports of ISRaD data
}
\examples{
# Obtain current ISRaD data
database <- ISRaD.getdata(tempdir(), dataset = "full", extra = F, force_downlaod = F)
# Report metadata statistics
ISRaD.report(database, report = "entry.stats")
# Report summary statistics for all levels of the database
ISRaD.report(database, report = "count.all")
# Generate a map of all ISRaD sites
ISRaD.report(database, report = "site.map")
}
