% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{Fit models}
\usage{
fit_models(formula, data)
}
\arguments{
\item{formula}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}
}
\description{
analyzes the data and outputs a summary of the results in the form of an HTML table.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[lme4]{glmer}}
 \code{\link[sjPlot]{tab_model}}
}
