% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{IterableVec}
\alias{IterableVec}
\alias{IterableVec$new}
\alias{IterableVec$len}
\alias{IterableVec$to_vec}
\alias{$.IterableVec}
\alias{[[.IterableVec}
\alias{IterableVec$RMakeIter$make_iter}
\title{IterableVec Class}
\usage{
IterableVec

\method{$}{IterableVec}(self, name)

\method{[[}{IterableVec}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
IterableVec inherent methods: constructor, length, and conversion.
}
