% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{IterableVecIter}
\alias{IterableVecIter}
\alias{IterableVecIter$collect_all}
\alias{$.IterableVecIter}
\alias{[[.IterableVecIter}
\alias{IterableVecIter$RIterator$next_item}
\alias{IterableVecIter$RIterator$count}
\alias{IterableVecIter$RIterator$collect_n}
\alias{IterableVecIter$RIterator$skip}
\alias{IterableVecIter$RIterator$nth}
\title{IterableVecIter  Class}
\usage{
IterableVecIter

\method{$}{IterableVecIter}(self, name)

\method{[[}{IterableVecIter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
IterableVecIter blanket impl: collect all remaining items.
}
