% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_altrep_elt}
\alias{jiff_altrep_elt}
\title{jiff_altrep_elt}
\usage{
jiff_altrep_elt(x, i)
}
\arguments{
\item{x}{POSIXct vector (may be ALTREP).}

\item{i}{0-based index.}
}
\description{
Check the i-th element (0-based) of a Timestamp POSIXct vector as f64 secs.
}
