% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_altrep_len}
\alias{jiff_altrep_len}
\title{jiff_altrep_len}
\usage{
jiff_altrep_len(x)
}
\arguments{
\item{x}{POSIXct vector (may be ALTREP).}
}
\description{
Return the length of a POSIXct vector materialized from Timestamp.
}
