% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_altrep_timestamps}
\alias{jiff_altrep_timestamps}
\title{jiff_altrep_timestamps}
\usage{
jiff_altrep_timestamps(n)
}
\arguments{
\item{n}{Number of elements.}
}
\description{
Create a JiffTimestampVec ALTREP containing n timestamps (0..n seconds after epoch).

Returns a POSIXct (UTC) ALTREP REALSXP vector backed by Arc<Vec\if{html}{\out{<Timestamp>}}>.
}
