% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_counted_altrep}
\alias{jiff_counted_altrep}
\title{jiff_counted_altrep}
\usage{
jiff_counted_altrep(n)
}
\arguments{
\item{n}{Number of elements (each is i seconds after epoch).}
}
\description{
Create a JiffTimestampVecCounted ALTREP and reset the elt call counter.
}
