% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_counted_altrep_elt_count}
\alias{jiff_counted_altrep_elt_count}
\title{jiff_counted_altrep_elt_count}
\usage{
jiff_counted_altrep_elt_count()
}
\description{
Return the number of times elt() has been called on the last counted ALTREP.
}
