% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_date_weekday}
\alias{jiff_date_weekday}
\title{jiff_date_weekday}
\usage{
jiff_date_weekday(date)
}
\arguments{
\item{date}{Date scalar from R.}
}
\description{
Return the ISO weekday of a civil::Date (1=Mon … 7=Sun).
}
