% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_fractional_timestamp}
\alias{jiff_fractional_timestamp}
\title{jiff_fractional_timestamp}
\usage{
jiff_fractional_timestamp()
}
\description{
Return a Timestamp with sub-second precision (1.5 seconds after epoch).
}
