% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_half_second_before_epoch}
\alias{jiff_half_second_before_epoch}
\title{jiff_half_second_before_epoch}
\usage{
jiff_half_second_before_epoch()
}
\description{
Return a Timestamp just before midnight negative (tests floor correctness).
-0.5 seconds before epoch = -1 second floor + 0.5 seconds subsec.
}
