% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_negative_timestamp}
\alias{jiff_negative_timestamp}
\title{jiff_negative_timestamp}
\usage{
jiff_negative_timestamp()
}
\description{
Return a negative Timestamp (1960-01-01 00:00:00 UTC) to test floor-based split.
}
