% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_one_hour_duration}
\alias{jiff_one_hour_duration}
\title{jiff_one_hour_duration}
\usage{
jiff_one_hour_duration()
}
\description{
Return a positive SignedDuration (3600 seconds = 1 hour).
}
