% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_roundtrip_date}
\alias{jiff_roundtrip_date}
\title{jiff_roundtrip_date}
\usage{
jiff_roundtrip_date(date)
}
\arguments{
\item{date}{Date scalar from R.}
}
\description{
Roundtrip a civil::Date through R Date.
}
