% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_roundtrip_date_vec}
\alias{jiff_roundtrip_date_vec}
\title{jiff_roundtrip_date_vec}
\usage{
jiff_roundtrip_date_vec(dates)
}
\arguments{
\item{dates}{Date vector from R.}
}
\description{
Roundtrip a \verb{Vec<civil::Date>} through R Date vector.
}
