% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_roundtrip_timestamp_vec}
\alias{jiff_roundtrip_timestamp_vec}
\title{jiff_roundtrip_timestamp_vec}
\usage{
jiff_roundtrip_timestamp_vec(ts)
}
\arguments{
\item{ts}{POSIXct (UTC) vector from R.}
}
\description{
Roundtrip a vector of UTC POSIXct values through Vec\if{html}{\out{<Timestamp>}}.
}
