% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_roundtrip_zoned}
\alias{jiff_roundtrip_zoned}
\title{jiff_roundtrip_zoned}
\usage{
jiff_roundtrip_zoned(zdt)
}
\arguments{
\item{zdt}{POSIXct with tzone attribute from R.}
}
\description{
Roundtrip a POSIXct with timezone through Zoned.
}
