% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_roundtrip_zoned_vec}
\alias{jiff_roundtrip_zoned_vec}
\title{jiff_roundtrip_zoned_vec}
\usage{
jiff_roundtrip_zoned_vec(zdts)
}
\arguments{
\item{zdts}{POSIXct vector with tzone attribute from R.}
}
\description{
Roundtrip a Vec\if{html}{\out{<Zoned>}} through R POSIXct.
}
