% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_span_days}
\alias{jiff_span_days}
\title{jiff_span_days}
\usage{
jiff_span_days(span)
}
\arguments{
\item{span}{ExternalPtr wrapping a JiffSpan.}
}
\description{
Extract the days component from a JiffSpan.
}
