% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_span_is_negative}
\alias{jiff_span_is_negative}
\title{jiff_span_is_negative}
\usage{
jiff_span_is_negative(span)
}
\arguments{
\item{span}{ExternalPtr wrapping a JiffSpan.}
}
\description{
Check whether a JiffSpan is negative.
}
