% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_span_is_zero}
\alias{jiff_span_is_zero}
\title{jiff_span_is_zero}
\usage{
jiff_span_is_zero(span)
}
\arguments{
\item{span}{ExternalPtr wrapping a JiffSpan.}
}
\description{
Check whether a JiffSpan is zero.
}
