% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_span_months}
\alias{jiff_span_months}
\title{jiff_span_months}
\usage{
jiff_span_months(span)
}
\arguments{
\item{span}{ExternalPtr wrapping a JiffSpan.}
}
\description{
Extract the months component from a JiffSpan.
}
