% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_span_new}
\alias{jiff_span_new}
\title{jiff_span_new}
\usage{
jiff_span_new(years, months, days)
}
\arguments{
\item{years}{Number of years (integer).}

\item{months}{Number of months (integer).}

\item{days}{Number of days (integer).}
}
\description{
Create a new JiffSpan ExternalPtr from year/month/day components.
}
