% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_span_rcrd_demo}
\alias{jiff_span_rcrd_demo}
\title{jiff_span_rcrd_demo}
\usage{
jiff_span_rcrd_demo()
}
\value{
A vctrs rcrd of class \code{jiff_span}.
}
\description{
Build a \code{jiff_span} vctrs rcrd from three hard-coded spans.

Spans: 1y2m, 3m15d, zero.
}
