% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_span_years}
\alias{jiff_span_years}
\title{jiff_span_years}
\usage{
jiff_span_years(span)
}
\arguments{
\item{span}{ExternalPtr wrapping a JiffSpan.}
}
\description{
Extract the years component from a JiffSpan.
}
