% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_time_new}
\alias{jiff_time_new}
\title{jiff_time_new}
\usage{
jiff_time_new(hour, minute, second)
}
\arguments{
\item{hour}{Hour (0–23).}

\item{minute}{Minute (0–59).}

\item{second}{Second (0–59).}
}
\description{
Build a civil::Time ExternalPtr from hour/minute/second components.
}
