% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_timestamp_as_millisecond}
\alias{jiff_timestamp_as_millisecond}
\title{jiff_timestamp_as_millisecond}
\usage{
jiff_timestamp_as_millisecond(ts)
}
\arguments{
\item{ts}{POSIXct (UTC) scalar from R.}
}
\description{
Return the milliseconds-since-Unix-epoch for a Timestamp.
}
