% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_timestamp_strftime}
\alias{jiff_timestamp_strftime}
\title{jiff_timestamp_strftime}
\usage{
jiff_timestamp_strftime(ts, fmt)
}
\arguments{
\item{ts}{POSIXct (UTC) scalar from R.}

\item{fmt}{strftime format string.}
}
\description{
Format a Timestamp (UTC) using a strftime format string.
}
