% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_zoned_month}
\alias{jiff_zoned_month}
\title{jiff_zoned_month}
\usage{
jiff_zoned_month(zdt)
}
\arguments{
\item{zdt}{POSIXct with tzone attribute from R.}
}
\description{
Extract the month from a Zoned datetime (1–12).
}
