% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_zoned_rcrd_demo}
\alias{jiff_zoned_rcrd_demo}
\title{jiff_zoned_rcrd_demo}
\usage{
jiff_zoned_rcrd_demo()
}
\value{
A vctrs rcrd of class \code{jiff_zoned}.
}
\description{
Build a \code{jiff_zoned} vctrs rcrd from two zones: UTC and Europe/Paris.
}
