% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_zoned_start_of_day}
\alias{jiff_zoned_start_of_day}
\title{jiff_zoned_start_of_day}
\usage{
jiff_zoned_start_of_day(zdt)
}
\arguments{
\item{zdt}{POSIXct with tzone attribute from R.}
}
\description{
Return the start of day for a Zoned datetime (preserves timezone).
}
