% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_zoned_strftime}
\alias{jiff_zoned_strftime}
\title{jiff_zoned_strftime}
\usage{
jiff_zoned_strftime(zdt, fmt)
}
\arguments{
\item{zdt}{POSIXct with tzone attribute from R.}

\item{fmt}{strftime format string.}
}
\description{
Format a Zoned datetime using a strftime format string.
}
