% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_zoned_tz_name}
\alias{jiff_zoned_tz_name}
\title{jiff_zoned_tz_name}
\usage{
jiff_zoned_tz_name(zdt)
}
\arguments{
\item{zdt}{POSIXct with tzone attribute from R.}
}
\description{
Extract the IANA timezone name from a Zoned.
}
