% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{jiff_zoned_year}
\alias{jiff_zoned_year}
\title{jiff_zoned_year}
\usage{
jiff_zoned_year(zdt)
}
\arguments{
\item{zdt}{POSIXct with tzone attribute from R.}
}
\description{
Extract the year from a Zoned datetime.
}
