% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{json_is_number}
\alias{json_is_number}
\title{json_is_number}
\usage{
json_is_number(value)
}
\arguments{
\item{value}{JSON value from R.}
}
\description{
Test whether a JSON value is a number.
}
