% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{json_is_object}
\alias{json_is_object}
\title{json_is_object}
\usage{
json_is_object(value)
}
\arguments{
\item{value}{JSON value from R.}
}
\description{
Test whether a JSON value is an object.
}
