% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{json_is_string}
\alias{json_is_string}
\title{json_is_string}
\usage{
json_is_string(value)
}
\arguments{
\item{value}{JSON value from R.}
}
\description{
Test whether a JSON value is a string.
}
