% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{json_object_keys}
\alias{json_object_keys}
\title{json_object_keys}
\usage{
json_object_keys(value)
}
\arguments{
\item{value}{JSON object from R.}
}
\description{
Test extracting keys from a JSON object.
}
