% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{json_options_builder}
\alias{json_options_builder}
\title{json_options_builder}
\usage{
json_options_builder(value)
}
\arguments{
\item{value}{R object to serialize.}
}
\description{
Exercise the \code{JsonOptions} builder methods (\code{new} + \code{na}/\code{nan}/\code{inf}):
NA serializes as the string \code{"NA"}, NaN/Inf as \code{null}.
}
