% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{json_serialize_point}
\alias{json_serialize_point}
\title{json_serialize_point}
\usage{
json_serialize_point(x, y)
}
\arguments{
\item{x}{X coordinate.}

\item{y}{Y coordinate.}
}
\description{
Test serializing a Point struct to JSON string.
}
