% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_json_point}
\alias{test_json_point}
\alias{test_json_config}
\alias{test_fromjson_bad}
\alias{test_fromjson_config}
\alias{test_json_vec_points}
\alias{test_json_pretty_point}
\alias{test_fromjson_point_sum}
\title{test json point}
\source{
Generated by miniextendr from Rust fn \code{test_json_point}

Generated by miniextendr from Rust fn \code{test_json_config}

Generated by miniextendr from Rust fn \code{test_fromjson_bad}

Generated by miniextendr from Rust fn \code{test_fromjson_config}

Generated by miniextendr from Rust fn \code{test_json_vec_points}

Generated by miniextendr from Rust fn \code{test_json_pretty_point}

Generated by miniextendr from Rust fn \code{test_fromjson_point_sum}
}
\usage{
test_json_point()

test_json_config()

test_fromjson_bad(json)

test_fromjson_config(json)

test_json_vec_points()

test_json_pretty_point()

test_fromjson_point_sum(json)
}
\arguments{
\item{json}{(no documentation available)}
}
\description{
test json point

test json config

test fromjson bad

test fromjson config

test json vec points

test json pretty point

test fromjson point sum
}
