% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{json_type_name}
\alias{json_type_name}
\title{json_type_name}
\usage{
json_type_name(value)
}
\arguments{
\item{value}{JSON value from R.}
}
\description{
Test getting the JSON type name of a value.
}
