% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lapply_df}
\alias{lapply_df}
\title{Run \code{\link{lapply}}, converting result to a data frame.}
\usage{
lapply_df(...)
}
\arguments{
\item{...}{Parameters to pass to lapply}
}
\value{
A data frame of the results with stringsAsFactors=FALSE.
}
\description{
Run \code{\link{lapply}}, converting result to a data frame.
}
\keyword{internal}
