% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{large_vec_altrep}
\alias{large_vec_altrep}
\title{large_vec_altrep}
\source{
Generated by miniextendr from Rust fn \code{large_vec_altrep}
}
\usage{
large_vec_altrep()
}
\description{
Example: Large data - ALTREP avoids copy
}
