% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{lazy_s3_report}
\alias{lazy_s3_report}
\title{Describe the unforced dots via S3 dispatch}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Describe the unforced dots via S3 dispatch
}
