% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{lazy_s3_sum}
\alias{lazy_s3_sum}
\title{Sum the dots that are not empty via S3 dispatch}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Sum the dots that are not empty via S3 dispatch
}
