% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{lazy_string}
\alias{lazy_string}
\title{lazy string}
\source{
Generated by miniextendr from Rust fn \code{lazy_string}
}
\usage{
lazy_string(prefix, n)
}
\arguments{
\item{prefix}{(no documentation available)}

\item{n}{(no documentation available)}
}
\description{
lazy string
}
