% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{lbfgsb_infinite_objective}
\alias{lbfgsb_infinite_objective}
\title{lbfgsb_infinite_objective}
\usage{
lbfgsb_infinite_objective()
}
\description{
Call \code{lbfgsb} with an objective returning \code{Inf}, which it rejects with an
R error.
}
