% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{length}
\alias{length}
\alias{length,fm-method}
\alias{length,fmV-method}
\title{Length of an Object}
\usage{
\S4method{length}{fmV}(x)

\S4method{length}{fm}(x)
}
\arguments{
\item{x}{an FlashR object.}
}
\value{
an integer of length 1 or a double if the object has more than
2^31-1 elements.
}
\description{
Get the length of a vector or a matrix.
}
\examples{
mat <- fm.runif.matrix(100, 10)
length(mat)
}

