% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\docType{methods}
\name{levels}
\alias{levels}
\title{Levels Attributes}
\usage{
\S4method{levels}{fmVFactor}(x)
}
\arguments{
\item{x}{a FlashR factor.}
}
\value{
a FlashR vector that contains the value of the levels.
}
\description{
\code{levels} returns the levels attribute of a variable.
}
\details{
Currently, levels of a factor vector is a sequence number from 1 to
the number of levels. FlashR doesn't support other factor values.
}
\examples{
vec <- fm.as.factor(as.integer(fm.runif(100, min=0, max=100)))
levels(vec)
}

