% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{LifecycleDemo}
\alias{LifecycleDemo}
\title{LifecycleDemo R6 Class}
\description{
R6 class demonstrating lifecycle attributes on methods.
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-LifecycleDemo-initialize}{\code{LifecycleDemo$new()}}
    \item \href{#method-LifecycleDemo-get_value}{\code{LifecycleDemo$get_value()}}
    \item \href{#method-LifecycleDemo-old_value}{\code{LifecycleDemo$old_value()}}
    \item \href{#method-LifecycleDemo-experimental_method}{\code{LifecycleDemo$experimental_method()}}
    \item \href{#method-LifecycleDemo-legacy_value}{\code{LifecycleDemo$legacy_value()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LifecycleDemo-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-LifecycleDemo-initialize}{}}}
\subsection{\code{LifecycleDemo$new()}}{
  Creates a new LifecycleDemo.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LifecycleDemo$new(value, .ptr = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{value}}{Integer initial value.}
      \item{\code{.ptr}}{Internal pointer (used by static methods, not for direct use).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LifecycleDemo-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-LifecycleDemo-get_value}{}}}
\subsection{\code{LifecycleDemo$get_value()}}{
  Returns the value.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LifecycleDemo$get_value()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LifecycleDemo-old_value"></a>}}
\if{latex}{\out{\hypertarget{method-LifecycleDemo-old_value}{}}}
\subsection{\code{LifecycleDemo$old_value()}}{
  \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A deprecated method using lifecycle attribute.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LifecycleDemo$old_value()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LifecycleDemo-experimental_method"></a>}}
\if{latex}{\out{\hypertarget{method-LifecycleDemo-experimental_method}{}}}
\subsection{\code{LifecycleDemo$experimental_method()}}{
  \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} An experimental method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LifecycleDemo$experimental_method()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LifecycleDemo-legacy_value"></a>}}
\if{latex}{\out{\hypertarget{method-LifecycleDemo-legacy_value}{}}}
\subsection{\code{LifecycleDemo$legacy_value()}}{
  \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A method marked with Rust's \verb{#[deprecated]} attribute.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LifecycleDemo$legacy_value()}
    \if{html}{\out{</div>}}
  }
}

}
