% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lint.R
\name{lint_knit}
\alias{lint_knit}
\title{Lint code chunks in a knitr document}
\usage{
lint_knit(filename, linters = default_linters,
  encoding = getOption("encoding"), cache = FALSE)
}
\arguments{
\item{filename}{the given filename to lint.}

\item{linters}{a list of linter functions to apply see \code{\link{linters}}
for a full list of default and available linters.}

\item{encoding}{The encoding fo the input file; see \code{\link{file}}.}

\item{cache}{toggle caching of lint results}
}
\description{
Lint any document that knitr will read.
The markup used for R chunks is recognized and defined the in the same manner
as \code{\link[knitr]{knit}}; which this function internally uses.
The extension will be matched against the patterns in \code{knitr::all_paterns},
The patttern can be manually set using the \code{\link{knit_patterns}} object.
}
\details{
Note that this does not lint any inline code in the knitr document.
}

