% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{list_column_starts_df}
\alias{list_column_starts_df}
\title{list_column_starts_df}
\usage{
list_column_starts_df()
}
\description{
Data frame with one list column of start IDs; rows 2 and 4 have none.
}
