% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{list_column_starts_event_df}
\alias{list_column_starts_event_df}
\title{list_column_starts_event_df}
\usage{
list_column_starts_event_df()
}
\description{
Enum data frame: the \code{Miss} rows have \code{NULL} list cells.
}
