% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{list_column_starts_rows_wide}
\alias{list_column_starts_rows_wide}
\title{list_column_starts_rows_wide}
\usage{
list_column_starts_rows_wide()
}
\description{
\verb{Vec<Vec<usize>>} returned directly (default, lossy widening).
}
