% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_endpoint.R
\name{lm_endpoint}
\alias{lm_endpoint}
\title{Create an endpoint from a \code{\link{lm}} object}
\usage{
lm_endpoint(model, preprocess, predict, postprocess)
}
\description{
Create an endpoint from a \code{\link{lm}} object
}

