% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rnaseq.R
\name{load_rnaseq}
\alias{load_rnaseq}
\title{Loads RNA-seq datasets}
\usage{
load_rnaseq(organism = c("mouse", "fly", "fish"), dataset = c("provided",
  "generated"), ...)
}
\arguments{
\item{organism}{either "mouse", "fly", or "fish"}

\item{dataset}{either "provided" or "generated"}
}
\value{
a DGEList of the data.
}
\description{
This function will load an R object that contains all of the expression data
for a given organism from the the data that was generated before the
workshop, or during the first few days of it.
}
