% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadEntries.R
\name{loadEntries}
\alias{loadEntries}
\title{Load all entries of the aridec dataset}
\usage{
loadEntries(path = "~/Repos/aridec/data/")
}
\arguments{
\item{path}{character string with the path where isdb is stored}
}
\value{
R list with all entries
}
\description{
Load all entries of the aridec dataset
}
\examples{
aridec=loadEntries()
}
