% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{log}
\alias{exp,fm-method}
\alias{exp,fmV-method}
\alias{log}
\alias{log,fm-method}
\alias{log,fmV-method}
\alias{log10,fm-method}
\alias{log10,fmV-method}
\alias{log2,fm-method}
\alias{log2,fmV-method}
\title{Logarithms and Exponentials}
\usage{
\S4method{log10}{fm}(x)

\S4method{log10}{fmV}(x)

\S4method{log2}{fm}(x)

\S4method{log2}{fmV}(x)

\S4method{exp}{fm}(x)

\S4method{exp}{fmV}(x)

\S4method{log}{fm}(x, base = exp(1))

\S4method{log}{fmV}(x, base = exp(1))
}
\arguments{
\item{x}{a numeric vector.}

\item{base}{a positive number.}
}
\value{
A vector of the same length as \code{x} containing the transformed
values. \code{log(0)} gives \code{-Inf}, and \code{log(x)} for negative
values of \code{x} is \code{NaN}.  \code{exp(-Inf)} is 0.
}
\description{
\code{log} computes logarithms, by default natural logarithms, \code{log10}
computes common (i.e., base 10) logarithms, and \code{log2} computes binary
(i.e., base 2) logarithms. The general form log(x, base) computes logarithms
with \code{base}.
}
\details{
\code{exp} computes the exponential function.
}
\examples{
mat <- log(fm.runif.matrix(100, 10))
mat <- log10(fm.runif.matrix(100, 10))
mat <- log2(fm.runif.matrix(100, 10))
mat <- exp(fm.runif.matrix(100, 10))
}

