% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_log_info}
\alias{test_log_info}
\alias{test_log_warn}
\alias{test_log_debug}
\alias{test_log_error}
\alias{test_log_set_level}
\title{test log info}
\source{
Generated by miniextendr from Rust fn \code{test_log_info}

Generated by miniextendr from Rust fn \code{test_log_warn}

Generated by miniextendr from Rust fn \code{test_log_debug}

Generated by miniextendr from Rust fn \code{test_log_error}

Generated by miniextendr from Rust fn \code{test_log_set_level}
}
\usage{
test_log_info(msg)

test_log_warn(msg)

test_log_debug(msg)

test_log_error(msg)

test_log_set_level(level)
}
\arguments{
\item{msg}{(no documentation available)}

\item{level}{(no documentation available)}
}
\description{
test log info

test log warn

test log debug

test log error

test log set level
}
