% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{Logic}
\alias{!.fm}
\alias{!.fmV}
\alias{&,ANY,fm-method}
\alias{&,ANY,fmV-method}
\alias{&,fm,ANY-method}
\alias{&,fm,fm-method}
\alias{&,fm,fmV-method}
\alias{&,fm,matrix-method}
\alias{&,fmV,ANY-method}
\alias{&,fmV,fm-method}
\alias{&,fmV,fmV-method}
\alias{&,matrix,fm-method}
\alias{Logic}
\alias{|,ANY,fm-method}
\alias{|,ANY,fmV-method}
\alias{|,fm,ANY-method}
\alias{|,fm,fm-method}
\alias{|,fm,fmV-method}
\alias{|,fm,matrix-method}
\alias{|,fmV,ANY-method}
\alias{|,fmV,fm-method}
\alias{|,fmV,fmV-method}
\alias{|,matrix,fm-method}
\title{Logical Operators}
\usage{
\S4method{|}{fm,fm}(e1, e2)

\S4method{|}{fmV,fmV}(e1, e2)

\S4method{|}{fm,fmV}(e1, e2)

\S4method{|}{fmV,fm}(e1, e2)

\S4method{|}{fm,matrix}(e1, e2)

\S4method{|}{matrix,fm}(e1, e2)

\S4method{|}{fm,ANY}(e1, e2)

\S4method{|}{ANY,fm}(e1, e2)

\S4method{|}{fmV,ANY}(e1, e2)

\S4method{|}{ANY,fmV}(e1, e2)

\S4method{&}{fm,fm}(e1, e2)

\S4method{&}{fmV,fmV}(e1, e2)

\S4method{&}{fm,fmV}(e1, e2)

\S4method{&}{fmV,fm}(e1, e2)

\S4method{&}{fm,matrix}(e1, e2)

\S4method{&}{matrix,fm}(e1, e2)

\S4method{&}{fm,ANY}(e1, e2)

\S4method{&}{ANY,fm}(e1, e2)

\S4method{&}{fmV,ANY}(e1, e2)

\S4method{&}{ANY,fmV}(e1, e2)

\method{!}{fm}(e1)

\method{!}{fmV}(e1)
}
\arguments{
\item{e1,e2}{One of the operands need to be a FlashR object. If one operand
is a matrix and the other is a vector, we perform the arithmetic operation
on the vector and every column of the matrix. If one operand is a scalar,
we perform the operation on the scalar with every element in the matrix or
the vector.}
}
\description{
These operators act on logical and number-like vectors.
}
\details{
\code{!} indicates logical negation (NOT).
\code{&} indicates logical AND and \code{|} indicates logical OR.
}

