% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Main pipeline}
\usage{
main(otu_filename, plot_filename)
}
\arguments{
\item{otu_filename}{path to a file with OTU relative abundance data in wide format.}

\item{plot_filename}{path to save the box plot.}
}
\description{
Read a file containing OTU relative abundance data with \code{readr::read_csv()},
convert it to a long format with \code{wide_to_long()},
make a box plot of the relative abundances with \code{plot_relabun_box()},
and save the plot with \code{ggplot2::ggsave()}.
}
