% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP.R
\name{MAP}
\alias{MAP}
\title{Creates a data frame with mean annual precipitation values of the sites}
\usage{
MAP(database)
}
\arguments{
\item{database}{A list with the aridec structure}
}
\value{
A data frame with the mean annual precipitation values (mm) from the database
}
\description{
Creates a data frame with mean annual precipitation values of the sites
}
\examples{
MAP=MAP(database=aridec)
}
