% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMATS.R
\name{map_rMATS_original}
\alias{map_rMATS_original}
\title{Original map_rMATS function preserved for backward compatibility}
\usage{
map_rMATS_original(
  g,
  gene,
  gff,
  fromGTF_A3SS,
  fromGTF_A5SS,
  fromGTF_SE,
  fromGTF_RI,
  bipartitions,
  grase_output_dir
)
}
\description{
Original map_rMATS function preserved for backward compatibility
}
