% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{MarkerReader}
\alias{MarkerReader}
\alias{MarkerReader$new}
\alias{$.MarkerReader}
\alias{[[.MarkerReader}
\alias{MarkerReader$MarkerArgs$marker_number}
\alias{MarkerReader$MarkerArgs$marker_missing}
\alias{MarkerReader$MarkerArgs$marker_label}
\alias{MarkerReader$MarkerArgs$marker_labels}
\alias{MarkerReader$MarkerArgs$marker_maybe}
\alias{MarkerReader$MarkerArgs$marker_ip}
\alias{MarkerReader$MarkerArgs$marker_ips}
\alias{MarkerReader$MarkerArgs$marker_echo}
\alias{MarkerReader$MarkerArgs$marker_dose}
\title{MarkerReader Class}
\usage{
MarkerReader

\method{$}{MarkerReader}(self, name)

\method{[[}{MarkerReader}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Env class whose \code{MarkerArgs} trait methods take the reading markers.

Create a reader.
}
