% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_alias_shade}
\alias{match_arg_alias_shade}
\title{match_arg_alias_shade}
\usage{
match_arg_alias_shade(color = c("red", "gray", "blue"))
}
\arguments{
\item{color}{A colour.}
}
\description{
\code{color} matched by the generated wrapper.
}
\keyword{internal}
