% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_alias_shade_default}
\alias{match_arg_alias_shade_default}
\title{match_arg_alias_shade_default}
\usage{
match_arg_alias_shade_default(color = c("blue", "red", "gray"))
}
\arguments{
\item{color}{A colour, \code{"blue"} by default.}
}
\description{
\code{color} whose default is \code{"blue"}.
}
\keyword{internal}
