% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_alias_shade_or_number}
\alias{match_arg_alias_shade_or_number}
\title{match_arg_alias_shade_or_number}
\usage{
match_arg_alias_shade_or_number(color = c("red", "gray", "blue"))
}
\arguments{
\item{color}{A colour, or a number.}
}
\description{
A colour or a number.
}
\keyword{internal}
