% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_alias_shade_param}
\alias{match_arg_alias_shade_param}
\title{match_arg_alias_shade_param}
\usage{
match_arg_alias_shade_param(color)
}
\arguments{
\item{color}{A colour.}
}
\description{
\code{color} matched by the body with \code{match_arg_param()}.
}
\keyword{internal}
