% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_either_route}
\alias{match_arg_either_route}
\title{match_arg_either_route}
\usage{
match_arg_either_route(route = c("oral", "bolus", "infusion"))
}
\arguments{
\item{route}{One of "oral", "bolus", "infusion", or a data frame.}
}
\description{
A route name or a data frame of doses.
}
