% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_either_route_optional}
\alias{match_arg_either_route_optional}
\title{match_arg_either_route_optional}
\usage{
match_arg_either_route_optional(maybe_route = NULL)
}
\arguments{
\item{maybe_route}{One of "oral", "bolus", "infusion", a data frame, or NULL for no choice.}
}
\description{
An optional route name or data frame: \code{NULL} is no choice.
}
