% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_either_route_or_list}
\alias{match_arg_either_route_or_list}
\title{match_arg_either_route_or_list}
\usage{
match_arg_either_route_or_list(route = c("oral", "bolus", "infusion"))
}
\arguments{
\item{route}{One of "oral", "bolus", "infusion", or a list.}
}
\description{
A route name or a list of doses: a \code{List} as the other arm.
}
