% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_either_routes}
\alias{match_arg_either_routes}
\title{match_arg_either_routes}
\usage{
match_arg_either_routes(routes = c("oral", "bolus", "infusion"))
}
\arguments{
\item{routes}{One or more of "oral", "bolus", "infusion", or a data frame.}
}
\description{
Route names or a data frame of doses.
}
