% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_either_routes_boxed}
\alias{match_arg_either_routes_boxed}
\title{match_arg_either_routes_boxed}
\usage{
match_arg_either_routes_boxed(route_set = c("oral", "bolus", "infusion"))
}
\arguments{
\item{route_set}{One or more of "oral", "bolus", "infusion", or a data frame.}
}
\description{
Route names or a data frame of doses, as a boxed slice.
}
