% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_either_routes_optional}
\alias{match_arg_either_routes_optional}
\title{match_arg_either_routes_optional}
\usage{
match_arg_either_routes_optional(maybe_routes = NULL)
}
\arguments{
\item{maybe_routes}{One or more of "oral", "bolus", "infusion", a data frame, or NULL for no choice.}
}
\description{
Optional route names or a data frame: \code{NULL} is no choice.
}
