% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_mixed}
\alias{match_arg_mixed}
\title{match_arg_mixed}
\usage{
match_arg_mixed(x, mode = c("Fast", "Safe", "Debug"))
}
\arguments{
\item{x}{An integer input.}

\item{mode}{A Mode enum value.}
}
\description{
Function with match_arg + regular param to test mixing.
}
