% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_mode_choices}
\alias{match_arg_mode_choices}
\title{Return the choices for Mode (for testing from R)}
\source{
Generated by miniextendr from Rust fn \code{match_arg_mode_choices}
}
\usage{
match_arg_mode_choices()
}
\description{
Return the choices for Mode (for testing from R)
}
