% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_multi_priority}
\alias{match_arg_multi_priority}
\title{match_arg_multi_priority}
\usage{
match_arg_multi_priority(n, priorities = c("lo", "med", "hi"))
}
\arguments{
\item{n}{Count.}

\item{priorities}{One or more Priority values.}
}
\description{
Select multiple priorities with a regular param.
}
