% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_omitted_modes_boxed}
\alias{match_arg_omitted_modes_boxed}
\title{match_arg_omitted_modes_boxed}
\usage{
match_arg_omitted_modes_boxed(modes = c("Fast", "Safe", "Debug"))
}
\arguments{
\item{modes}{One or more of "Fast", "Safe", "Debug"; omitting the argument means no choice.}
}
\description{
Omittable several_ok list in a boxed slice.
}
