% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_on_failure}
\alias{match_arg_on_failure}
\title{match_arg_on_failure}
\usage{
match_arg_on_failure(on_failure = c("error", "value", "output"))
}
\arguments{
\item{on_failure}{An OnFailure choice.}
}
\description{
Round-trip an \code{OnFailure} choice through R.
}
