% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_on_failure_choices}
\alias{match_arg_on_failure_choices}
\title{match_arg_on_failure_choices}
\usage{
match_arg_on_failure_choices()
}
\description{
Return the choices for OnFailure.
}
