% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_on_failures}
\alias{match_arg_on_failures}
\title{match_arg_on_failures}
\usage{
match_arg_on_failures(on_failure = c("error", "value", "output"))
}
\arguments{
\item{on_failure}{One or more OnFailure choices.}
}
\description{
Round-trip several \code{OnFailure} choices through R.
}
