% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_priority_choices}
\alias{match_arg_priority_choices}
\title{Return the choices for Priority}
\source{
Generated by miniextendr from Rust fn \code{match_arg_priority_choices}
}
\usage{
match_arg_priority_choices()
}
\description{
Return the choices for Priority
}
