% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_set_mode}
\alias{match_arg_set_mode}
\title{match_arg_set_mode}
\usage{
match_arg_set_mode(mode = c("Fast", "Safe", "Debug"))
}
\arguments{
\item{mode}{A Mode enum value.}
}
\description{
Set mode using match_arg.
}
