% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_set_priority}
\alias{match_arg_set_priority}
\title{match_arg_set_priority}
\usage{
match_arg_set_priority(priority = c("lo", "med", "hi"))
}
\arguments{
\item{priority}{A Priority enum value.}
}
\description{
Set priority using match_arg.
}
