% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_set_status}
\alias{match_arg_set_status}
\title{match_arg_set_status}
\usage{
match_arg_set_status(status = c("in_progress", "completed", "not_started"))
}
\arguments{
\item{status}{A BuildStatus enum value.}
}
\description{
Set status using match_arg.
}
