% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_status_choices}
\alias{match_arg_status_choices}
\title{Return the choices for BuildStatus}
\source{
Generated by miniextendr from Rust fn \code{match_arg_status_choices}
}
\usage{
match_arg_status_choices()
}
\description{
Return the choices for BuildStatus
}
