% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_with_default}
\alias{match_arg_with_default}
\title{match_arg_with_default}
\usage{
match_arg_with_default(mode = c("Safe", "Fast", "Debug"))
}
\arguments{
\item{mode}{A Mode enum value.}
}
\description{
Function with explicit default override for match_arg param.
}
