% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{MathFun}
\alias{MathFun}
\alias{abs,fm-method}
\alias{abs,fmV-method}
\alias{sqrt,fm-method}
\alias{sqrt,fmV-method}
\title{Miscellaneous Mathematical Functions}
\usage{
\S4method{abs}{fm}(x)

\S4method{abs}{fmV}(x)

\S4method{sqrt}{fm}(x)

\S4method{sqrt}{fmV}(x)
}
\arguments{
\item{x}{a numeric vector or array.}
}
\description{
\code{abs(x)} computes the absolute value of x, \code{sqrt(x)} computes the square
root of x.
}
\examples{
mat <- abs(fm.runif.matrix(100, 10))
mat <- sqrt(fm.runif.matrix(100, 10))
}

