% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{matmult}
\alias{\%*\%,fm,ANY-method}
\alias{\%*\%,fm,fm-method}
\alias{\%*\%,fm,fmV-method}
\alias{matmult}
\title{Matrix multiplication}
\usage{
\S4method{\%*\%}{fm,fm}(x, y)

\S4method{\%*\%}{fm,fmV}(x, y)

\S4method{\%*\%}{fm,ANY}(x, y)
}
\arguments{
\item{x}{a FlashMatrix matrix.}

\item{y}{can be a FlashMatrix vector or matrix, an R vector or matrix.}
}
\value{
a FlashMatrix matrix.
}
\description{
Multiplies two matrices, if they are conformable. If one argument is
a vector, it will be promoted to either a row or column matrix to make
the two arguments conformable.
}

