% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R, R/FlashR_base.R
\docType{methods}
\name{matrix}
\alias{as.matrix,fm-method}
\alias{fm.as.matrix}
\alias{is.matrix,fm-method}
\alias{matrix}
\title{Matrices}
\usage{
\S4method{as.matrix}{fm}(x)

fm.as.matrix(x)

\S4method{is.matrix}{fm}(x)
}
\arguments{
\item{x}{an R object}
}
\description{
\code{as.matrix} attempts to turn a FlashMatrix matrix to an R matrix.
\code{fm.as.matrix} attempts to turn its argument into a FlashMatrix matrix.
\code{is.matrix} tests if its argument is a matrix.
}

