% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{maybe_bump}
\alias{maybe_bump}
\title{Add an amount in place, returning NULL past the cap}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Add an amount in place, returning NULL past the cap
}
