% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl_aesthetics.R
\name{mbl_create_shape_map}
\alias{mbl_create_shape_map}
\title{Maps shapes to categorical values}
\usage{
mbl_create_shape_map(vals, map = NULL)
}
\arguments{
\item{vals}{a vector of categorical values}

\item{map}{a map definition. By default we use pch symbol identifiers.}
}
\value{
a named vector. `names()` are the unique values in `vals`, and values
  are the different shapes (pch integers)
}
\description{
Map unique leves of `vals` to different shapes. Only works for categorical
variables.
}
\details{
TODO: Use plotly shapes (currently we use base R pch). This webpage shows
you the symbols and how to generate them:
http://www.r-graph-gallery.com/125-the-plotlys-symbols/
}
\examples{
# This isn't a real example. It is the code from the aforementioned page
# that generates the plotly shapes.
library(plotly)
data=expand.grid(c(1:6) , c(1:6))
data=cbind(data , my_symbol=c(1:36))
data=data[data$my_symbol<33 , ]

# Make the graph
my_graph=plot_ly(data , x=~Var1 , y=~Var2 , type="scatter",
                 mode="markers+text" , hoverinfo="text", text=~my_symbol,
                 textposition = "bottom right",
                 marker=list(symbol=~my_symbol, size=40, color="red",
                             opacity=0.7)) \%>\%
  layout(
    hovermode="closest",
    yaxis=list(autorange="reversed", title="",
               tickfont=list(color="white")) ,
    xaxis=list( title="" , tickfont=list(color="white"))
  )
# show graph
my_graph
}
