% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl_plot_expression.R
\name{mbl_fuzzy_find_gene_row}
\alias{mbl_fuzzy_find_gene_row}
\title{Fast and loose way to identify which row a query search for a string across all columns of}
\usage{
mbl_fuzzy_find_gene_row(y, query)
}
\description{
Fast and loose way to identify which row a query search for a string across all columns of
}
