% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl_get_gene_annotation.R
\name{mbl_get_gene_annotation}
\alias{mbl_get_gene_annotation}
\alias{mbl_get_transcript_annotation}
\title{Retrieves the gene-level anntation file for this organism}
\usage{
mbl_get_gene_annotation(organism = c("mouse", "fly", "fish"),
  rm.description = TRUE)

mbl_get_transcript_annotation(organism = c("mouse", "fly", "fish"),
  rm.description = TRUE)
}
\arguments{
\item{organism}{either "mouse", "fly", or "fish"}
}
\value{
the gene-level annotation data.frame for the organism
}
\description{
Retrieves the gene-level anntation file for this organism
}
