% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl_get_sample_annotation.R
\name{mbl_get_sample_annotation}
\alias{mbl_get_sample_annotation}
\title{Retrieve the sample annotation}
\usage{
mbl_get_sample_annotation(organism = c("mouse", "fly", "fish"),
  dataset = c("may", "generated"), kallisto_parent_dir = NULL,
  trim = TRUE)
}
\arguments{
\item{organism}{either "mouse", "fly", or "fish".}

\item{dataset}{either "may" or "generated"}
}
\value{
a data.frame of sample annotations
}
\description{
Returns the sample annotations for a particular dataset.
}
