% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl_load_rnaseq.R
\name{mbl_load_rnaseq}
\alias{mbl_load_rnaseq}
\title{Loads RNA-seq datasets}
\usage{
mbl_load_rnaseq(organism = c("mouse", "fly", "fish"),
  dataset = c("provided", "generated"), ...)
}
\arguments{
\item{organism}{either "mouse", "fly", or "fish"}

\item{dataset}{either "provided" or "generated"}
}
\value{
a DGEList of the data.
}
\description{
This function will load an R object that contains all of the expression data
for a given organism from the the data that was generated before the
workshop, or during the first few days of it.
}
