% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl_plots.R
\name{mbl_plot_expression}
\alias{mbl_plot_expression}
\title{Plot expression of a gene/feature}
\usage{
mbl_plot_expression(y, gene, group, color_by = NULL, ...)
}
\arguments{
\item{y}{the DGEList or voomed object your data is in}

\item{gene}{the name of the gene identifier or symbol you want to plot}

\item{group}{what column in the $samples (or $targets, for vm) pheno table
to use to plot split expression across the x axis}

\item{color_by}{name of column (like `group` param) to color your points by}
}
\value{
a ggplot object
}
\description{
This function returns a raw ggplot object that you can decorate
further
}
