% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl_tidy.R
\name{mbl_tidy}
\alias{mbl_tidy}
\alias{mbl_tidy.DGEList}
\alias{mbl_tidy.EList}
\title{Converts a DGEList into a tidy data.frame}
\usage{
mbl_tidy(x, ...)

\method{mbl_tidy}{DGEList}(x, normalized.lib.sizes = TRUE, prior.count = 3,
  ...)

\method{mbl_tidy}{EList}(x, ...)
}
\arguments{
\item{x}{The expression object to tidy}
}
\description{
This makes downstream exploratory data analysis a bit easier.
}
