% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{mean}
\alias{mean}
\alias{mean,fm-method}
\alias{mean,fmV-method}
\title{Arithmetic Mean}
\usage{
\S4method{mean}{fm}(x, ...)

\S4method{mean}{fmV}(x, ...)
}
\arguments{
\item{x}{A FlashR vector or matrix.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Compute arithmetic mean.
}
\examples{
mat <- fm.runif.matrix(100, 10)
mean(mat)
}

