% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metawards.R
\docType{package}
\name{metawards}
\alias{metawards}
\title{metawards: Metapopulation Disease Modelling}
\description{
This package provides a convenient R wrapper around the Python
metawards package. For more information about MetaWards see
\url{https://metawards.org}. For a quick-start guide see
\url{https://metawards.org/quickstart}.
}
\details{
This package uses reticulate to automatically load and wrap
the Python metawards package. The easiest way to install metawards
is to run the \code{\link{py_install_metawards}} function. This will
install metawards into the default python associated with
reticulate. Alternatively, you can install python manually into
your own python environment and then tell reticulate to use
that python so that it can find the metawards module.
}
\section{Python functions}{

The python functions that are wrapped are all documented
at \url{https://metawards.org/api}.
}

\section{R-specific functions}{

\itemize{
\item \code{\link{py_install_metawards}}
\item \code{\link{py_metawards_available}}
\item \code{\link{py_version_metawards}}
}
}

