% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrsa.r
\docType{data}
\name{mff6}
\alias{mff6}
\title{Monthly Fama French 6 Factor Returns}
\format{
An \code{xts} object with 749 observations and 7 columns. 
The data run from July, 1963 through November, 2025.
As in the upstream source, the data are given in \emph{percents}, meaning a
value of 1.00 corresponds to a 1\% movement.
Note also that returns presumably are \sQuote{simple} returns, not log
returns, though this is not clarified by the upstream source.
The columns are defined as follows:
\describe{
 \item{\code{Mkt}}{The Market monthly return. Note that the risk free rate has been added back to the excess returns published by the upstream source.}
 \item{\code{SMB}}{The cap factor monthly return.}
 \item{\code{HML}}{The growth factor monthly return.}
 \item{\code{UMD}}{The momentum factor monthly return.}
 \item{\code{RMW}}{The profitability factor monthly return.}
 \item{\code{CMA}}{The investment factor monthly return.}
 \item{\code{RF}}{The risk-free rate, presumably as a monthly rate. The average value is arround 0.39, corresponding to an annualized rate of around 4.7\%.}
}
}
\source{
Kenneth French data library.
See \url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html},
and data description at \url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/Data_Library/f-f_5_factors_2x3.html}.
}
\usage{
mff6
}
\description{
The monthly returns of the 6 Fama French Factors: 
Market, 
the cap factor SMB, 
the growth factor HML, 
the momentum factor UMD,
the profitability factor RMW,
and the invesment factor CMA.
}
\examples{
\dontrun{
data(mff6)
str(mff6)
}
}
\author{
Steven E. Pav \email{steven@sharperat.io}
}
\keyword{datasets}
