% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrsa.r
\docType{data}
\name{mind5}
\alias{mind5}
\title{Monthly Fama French 5 Industry Returns}
\format{
An \code{xts} object with 1,187 observations and 5 columns. 
The data run from January, 1927 through November, 2025.
As in the upstream source, the data are given in \emph{percents}, meaning a
value of 1.00 corresponds to a 1\% movement.
Note also that returns presumably are \sQuote{simple} returns, not log
returns, though this is not clarified by the upstream source.
The columns are defined as follows:
\describe{
 \item{\code{Consumer}}{The monthly returns of the Consumer industry, published as \dQuote{Cnsmr}.}
 \item{\code{Manufacturing}}{The monthly returns of the Manufacturing industry, published as \dQuote{Manuf}.}
 \item{\code{Technology}}{The monthly returns of the Technology industry, published as \dQuote{HiTec}.}
 \item{\code{Healthcare}}{The monthly returns of the Healthcare industry, published as \dQuote{Hlth}.}
 \item{\code{Other}}{The monthly returns of the Other industry, published as \dQuote{Other}.}
}
}
\source{
Kenneth French data library.
See \url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html},
data description at \url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/Data_Library/det_5_ind_port.html}.
}
\usage{
mind5
}
\description{
The monthly returns of 5 industries: Consumer, Manufacturing, High Technology, Healthcare and Other.
}
\note{
These are not the \dQuote{ex dividend} returns series.
}
\examples{
\dontrun{
data(mind5)
str(mind5)
}
}
\author{
Steven E. Pav \email{steven@sharperat.io}
}
\keyword{datasets}
