% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpkg-package.R
\docType{package}
\name{miniextendr-package}
\alias{miniextendr}
\alias{miniextendr-package}
\title{miniextendr: Rust-R Interoperability Runtime}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
