% Generated by roxygen2 (4.0.2): do not edit by hand
\name{minmax}
\alias{fm.pmax2}
\alias{fm.pmin2}
\alias{minmax}
\title{Maxima and Minima}
\usage{
fm.pmin2(o1, o2)

fm.pmax2(o1, o2)
}
\arguments{
\item{o1}{the input vector/matrix.}

\item{o2}{the input vector/matrix.}
}
\value{
a vector/matrix.
}
\description{
Returns the (parallel) maxima and minima of the input values.
}
\details{
`fm.pmin2' and `fm.pmax2' works on two input vectors and returns
return a single vector/matrix giving the 'parallel' maxima (or minima) of
the vectors/matrices. The first element of the result is the maximum (minimum)
of the first elements of all the arguments, the second element of
the result is the maximum (minimum) of the second elements of all
the arguments and so on.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

