% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{missing_test_list}
\alias{missing_test_list}
\title{missing_test_list}
\usage{
missing_test_list(x)
}
\arguments{
\item{x}{An optional list.}
}
\description{
Test \verb{Missing<List>} -- a list argument that may be omitted.
}
