% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{missing_test_present}
\alias{missing_test_present}
\title{missing_test_present}
\usage{
missing_test_present(x)
}
\arguments{
\item{x}{An optional integer value.}
}
\description{
Test \verb{Missing<i32>} with is_present / is_absent.
}
