% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingFields.R
\name{missingFields}
\alias{missingFields}
\title{Reports what fields are missing from an entry in relation to a template}
\usage{
missingFields(entry, template)
}
\arguments{
\item{entry}{An entry from the sidb database}

\item{template}{A template of sidb created from a yaml file}
}
\value{
A data.frame with the latitude and longitude of sites from the database
}
\description{
Reports what fields are missing from an entry in relation to a template
}
\examples{
database=loadEntries()
template=yaml::yaml.load_file(input="~/sidb/data/template_metadata.yaml")
missingFields(entry=database[[1]], template = template)
}
