% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exontest_functions.R
\name{moderate_phi_log_scale}
\alias{moderate_phi_log_scale}
\title{Moderate phi estimates on the log scale using empirical Bayes shrinkage}
\usage{
moderate_phi_log_scale(phi_table, trimming_limit = 1e+10)
}
\description{
Moderate phi estimates on the log scale using empirical Bayes shrinkage
}
\examples{
\dontrun{
phi_df <- read.table("phi.glmmtmb.txt", header = TRUE, row.names = NULL)
phi_table <- grase::moderate_phi_log_scale(phi_df)
head(phi_table[, c("gene", "event", "phi", "phi_mod", "w")])
}
}
