% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exontest_functions.R
\name{moderate_prec_log_scale}
\alias{moderate_prec_log_scale}
\title{Moderate Dirichlet-Multinomial precision estimates using empirical Bayes
shrinkage}
\usage{
moderate_prec_log_scale(prec_table)
}
\arguments{
\item{prec_table}{A data frame with columns \code{gene}, \code{event},
\code{log_prec}, and \code{var_log_prec}, as returned by
\code{prec_estimate_vgam}.}
}
\description{
Moderate Dirichlet-Multinomial precision estimates using empirical Bayes
shrinkage
}
\examples{
\dontrun{
prec_table <- read.table("prec_dm.txt", header = TRUE, row.names = NULL)
prec_table <- grase::moderate_prec_log_scale(prec_table)
head(prec_table[, c("gene", "event", "log_prec", "log_prec_mod", "rho_mod")])
}
}
