% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moyenne.R
\name{moyenne}
\alias{moyenne}
\title{Ma super fonction}
\usage{
moyenne(x, na_rm = FALSE)
}
\arguments{
\item{x}{Vecteur numérique}

\item{na_rm}{Boolean to remove NA (default is FALSE)}
}
\value{
Un vecteur de longueur 1 qui sera la moyenne calculee
}
\description{
Ceci est ma super
fonction.
}
\examples{
moyenne(1:3)
\dontrun{
moyenne(c(1, 2, 3, NA)) # Error
}
moyenne(c(1, 2, 3, NA), na_rm = TRUE)
}
