% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msboxes_R.R
\name{msboxes_R}
\alias{msboxes_R}
\title{FUNCTION_TITLE}
\usage{
msboxes_R(
  data,
  id,
  yb,
  xb,
  boxwidth,
  boxheight,
  time_study,
  vartime,
  tmat.,
  scale,
  msm = FALSE,
  jsonpath = "~",
  name = "msboxes_R.json"
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{id}{PARAM_DESCRIPTION}

\item{yb}{PARAM_DESCRIPTION}

\item{xb}{PARAM_DESCRIPTION}

\item{boxwidth}{PARAM_DESCRIPTION}

\item{boxheight}{PARAM_DESCRIPTION}

\item{time_study}{PARAM_DESCRIPTION}

\item{vartime}{PARAM_DESCRIPTION}

\item{tmat.}{PARAM_DESCRIPTION}

\item{scale}{PARAM_DESCRIPTION}

\item{msm}{PARAM_DESCRIPTION, Default: FALSE}

\item{jsonpath}{PARAM_DESCRIPTION, Default: '~'}

\item{name}{PARAM_DESCRIPTION, Default: 'msboxes_R.json'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
