% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmjson2.R
\name{msmjson}
\alias{msmjson}
\title{FUNCTION_TITLE}
\usage{
msmjson(
  msm.model,
  vartime = seq(1, 10, 1),
  mat.init = q.crude,
  totlos = TRUE,
  visit = TRUE,
  sojourn = TRUE,
  pnext = TRUE,
  efpt = TRUE,
  envisits = TRUE,
  ci.json = "normal",
  cl.json = 0.95,
  B.json = 1000,
  cores.json = NULL,
  piecewise.times.json = NULL,
  piecewise.covariates.json = NULL,
  num.integ.json = FALSE,
  covariates_list = list(list(sex = 1), list(sex = 0)),
  jsonpath = "~",
  name = "predictions_R.json"
)
}
\arguments{
\item{msm.model}{PARAM_DESCRIPTION}

\item{vartime}{PARAM_DESCRIPTION, Default: seq(1, 10, 1)}

\item{mat.init}{PARAM_DESCRIPTION, Default: q.crude}

\item{totlos}{PARAM_DESCRIPTION, Default: TRUE}

\item{visit}{PARAM_DESCRIPTION, Default: TRUE}

\item{sojourn}{PARAM_DESCRIPTION, Default: TRUE}

\item{pnext}{PARAM_DESCRIPTION, Default: TRUE}

\item{efpt}{PARAM_DESCRIPTION, Default: TRUE}

\item{envisits}{PARAM_DESCRIPTION, Default: TRUE}

\item{ci.json}{PARAM_DESCRIPTION, Default: 'normal'}

\item{cl.json}{PARAM_DESCRIPTION, Default: 0.95}

\item{B.json}{PARAM_DESCRIPTION, Default: 1000}

\item{cores.json}{PARAM_DESCRIPTION, Default: NULL}

\item{piecewise.times.json}{PARAM_DESCRIPTION, Default: NULL}

\item{piecewise.covariates.json}{PARAM_DESCRIPTION, Default: NULL}

\item{num.integ.json}{PARAM_DESCRIPTION, Default: FALSE}

\item{covariates_list}{PARAM_DESCRIPTION, Default: list(list(sex = 1), list(sex = 0))}

\item{jsonpath}{PARAM_DESCRIPTION, Default: '~'}

\item{name}{PARAM_DESCRIPTION, Default: 'predictions_R.json'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringi]{stri_sort}}
}
