% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstatejson.R
\name{mstatejson}
\alias{mstatejson}
\title{FUNCTION_TITLE}
\usage{
mstatejson(
  x = cfcox,
  qmat = tmat,
  process = "Markov",
  totlos = TRUE,
  ci.json = TRUE,
  cl.json = 0.95,
  B.json = 100,
  variance = TRUE,
  vartype = "aalen",
  covariates_list = list(),
  M = 100,
  jsonpath = "~",
  name = "predictions_R_mstate.json"
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION, Default: cfcox}

\item{qmat}{PARAM_DESCRIPTION, Default: tmat}

\item{process}{PARAM_DESCRIPTION, Default: 'Markov'}

\item{totlos}{PARAM_DESCRIPTION, Default: TRUE}

\item{ci.json}{PARAM_DESCRIPTION, Default: TRUE}

\item{cl.json}{PARAM_DESCRIPTION, Default: 0.95}

\item{B.json}{PARAM_DESCRIPTION, Default: 100}

\item{variance}{PARAM_DESCRIPTION, Default: TRUE}

\item{vartype}{PARAM_DESCRIPTION, Default: 'aalen'}

\item{covariates_list}{PARAM_DESCRIPTION, Default: list()}

\item{M}{PARAM_DESCRIPTION, Default: 100}

\item{jsonpath}{PARAM_DESCRIPTION, Default: '~'}

\item{name}{PARAM_DESCRIPTION, Default: 'predictions_R_mstate.json'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringi]{stri_sort}}
}
