% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.multi_input_class.R
\docType{class}
\name{multi.input-class}
\alias{multi.input-class}
\alias{nsdm.multi.input}
\title{Generate an S4 class to store muti.input data}
\description{
Not to be called directly by the user
}
\author{
Philipp Brun (philipp.brun@wsl.ch)
}
