% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{mx_fit_create}
\alias{mx_fit_create}
\title{mx_fit_create}
\usage{
mx_fit_create(estimate, sigma)
}
\arguments{
\item{estimate}{Numeric vector of parameter estimates.}

\item{sigma}{Numeric residual scale.}
}
\description{
Construct an \code{MxFit} and return it as an R list (exercises \code{IntoList}).
}
