% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{mx_point_new}
\alias{mx_point_new}
\title{mx_point_new}
\usage{
mx_point_new(x, y)
}
\arguments{
\item{x}{Numeric x-coordinate.}

\item{y}{Numeric y-coordinate.}
}
\description{
Test creating an MxPoint via ExternalPtr generated by \verb{#[miniextendr]} on struct.
}
