% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{mx_point_sum}
\alias{mx_point_sum}
\title{mx_point_sum}
\usage{
mx_point_sum(p)
}
\arguments{
\item{p}{An ExternalPtr wrapping an MxPoint.}
}
\description{
Test summing fields of an MxPoint ExternalPtr.
}
