% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{mx_record_create}
\alias{mx_record_create}
\title{mx_record_create}
\usage{
mx_record_create(name, value)
}
\arguments{
\item{name}{Character name field.}

\item{value}{Integer value field.}
}
\description{
Test creating an MxRecord and returning as list via PreferList.
}
