% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{mx_s7_pair}
\alias{mx_s7_pair}
\title{The \code{mx_s7_pair()} generic}
\arguments{
\item{x}{An object.}

\item{other}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{mx_s7_pair()} is an S7 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{MxS7Left}}
\item \code{\link{MxS7Right}}
}
}
