% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{mx_season_name}
\alias{mx_season_name}
\title{mx_season_name}
\usage{
mx_season_name(s)
}
\arguments{
\item{s}{An MxSeason factor value.}
}
\description{
Test converting an MxSeason factor back to its debug string.
}
