% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{mx_verbosity_check}
\alias{mx_verbosity_check}
\title{mx_verbosity_check}
\usage{
mx_verbosity_check(v)
}
\arguments{
\item{v}{An MxVerbosity enum value.}
}
\description{
Test converting an MxVerbosity match_arg enum to its debug string.
}
