% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{MxNargsProbe}
\alias{MxNargsProbe}
\alias{MxNargsProbe$new}
\alias{$.MxNargsProbe}
\alias{[[.MxNargsProbe}
\alias{MxNargsProbe$ArgCount$count_args}
\title{MxNargsProbe Class}
\usage{
MxNargsProbe

\method{$}{MxNargsProbe}(self, name)

\method{[[}{MxNargsProbe}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Create a probe.
}
