% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{MyFloat}
\alias{MyFloat}
\alias{MyFloat$new}
\alias{MyFloat$value}
\alias{MyFloat$nan}
\alias{$.MyFloat}
\alias{[[.MyFloat}
\alias{MyFloat$RPartialOrd$partial_cmp}
\title{MyFloat Class}
\usage{
MyFloat

\method{$}{MyFloat}(self, name)

\method{[[}{MyFloat}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
MyFloat inherent methods: constructor and accessor.
}
