% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_choose_2_analysis.R
\name{n_choose_2_paths}
\alias{n_choose_2_paths}
\title{Compute diff exons from graph using n-choose-2 pairwise analysis}
\usage{
n_choose_2_paths(gene, g, sg, outdir, max_path = 20, collapse_bubbles = FALSE)
}
\description{
Compute diff exons from graph using n-choose-2 pairwise analysis
}
