% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\docType{methods}
\name{NA}
\alias{NA}
\alias{is.na,fm-method}
\alias{is.na,fmV-method}
\title{'Not Available' / Missing Values}
\usage{
\S4method{is.na}{fm}(x)

\S4method{is.na}{fmV}(x)
}
\arguments{
\item{x}{an FlashR vector or matrix.}
}
\value{
a logical FlashR vector or matrix.
}
\description{
This function indicates which elements are missing.
}
\examples{
mat <- fm.runif.matrix(100, 10)
is.na(mat)
}

