% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{nalgebra_dmatrix_ncols}
\alias{nalgebra_dmatrix_ncols}
\title{nalgebra_dmatrix_ncols}
\usage{
nalgebra_dmatrix_ncols(m)
}
\arguments{
\item{m}{Numeric matrix from R.}
}
\description{
Test getting the number of columns in a DMatrix.
}
