% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{nalgebra_dmatrix_nrows}
\alias{nalgebra_dmatrix_nrows}
\title{nalgebra_dmatrix_nrows}
\usage{
nalgebra_dmatrix_nrows(m)
}
\arguments{
\item{m}{Numeric matrix from R.}
}
\description{
Test getting the number of rows in a DMatrix.
}
