% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{nalgebra_dmatrix_roundtrip}
\alias{nalgebra_dmatrix_roundtrip}
\title{nalgebra_dmatrix_roundtrip}
\usage{
nalgebra_dmatrix_roundtrip(m)
}
\arguments{
\item{m}{Numeric matrix from R.}
}
\description{
Test DMatrix\if{html}{\out{<f64>}} roundtrip through R.
}
