% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{nalgebra_dmatrix_trace}
\alias{nalgebra_dmatrix_trace}
\title{nalgebra_dmatrix_trace}
\usage{
nalgebra_dmatrix_trace(m)
}
\arguments{
\item{m}{Numeric square matrix from R.}
}
\description{
Test computing the trace of a DMatrix.
}
