% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{nalgebra_dmatrix_transpose}
\alias{nalgebra_dmatrix_transpose}
\title{nalgebra_dmatrix_transpose}
\usage{
nalgebra_dmatrix_transpose(m)
}
\arguments{
\item{m}{Numeric matrix from R.}
}
\description{
Test transposing a DMatrix.
}
