% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{nalgebra_dvector_dot}
\alias{nalgebra_dvector_dot}
\title{nalgebra_dvector_dot}
\usage{
nalgebra_dvector_dot(a, b)
}
\arguments{
\item{a}{First numeric vector from R.}

\item{b}{Second numeric vector from R.}
}
\description{
Test computing the dot product of two DVectors.
}
