% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{nalgebra_dvector_len}
\alias{nalgebra_dvector_len}
\title{nalgebra_dvector_len}
\usage{
nalgebra_dvector_len(v)
}
\arguments{
\item{v}{Numeric vector from R.}
}
\description{
Test getting the length of a DVector.
}
