% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{nalgebra_dvector_roundtrip}
\alias{nalgebra_dvector_roundtrip}
\title{nalgebra_dvector_roundtrip}
\usage{
nalgebra_dvector_roundtrip(v)
}
\arguments{
\item{v}{Numeric vector from R.}
}
\description{
Test DVector\if{html}{\out{<f64>}} roundtrip through R.
}
