% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{nalgebra_from_fn}
\alias{nalgebra_from_fn}
\title{nalgebra_from_fn}
\usage{
nalgebra_from_fn(nrow, ncol)
}
\arguments{
\item{nrow}{Number of rows.}

\item{ncol}{Number of columns.}
}
\description{
Construct a DMatrix where each element equals row * ncol + col.
}
