% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{nalgebra_svector3_roundtrip}
\alias{nalgebra_svector3_roundtrip}
\title{nalgebra_svector3_roundtrip}
\usage{
nalgebra_svector3_roundtrip(x, y, z)
}
\arguments{
\item{x}{First element.}

\item{y}{Second element.}

\item{z}{Third element.}
}
\description{
Roundtrip a static SVector<f64, 3> through Rust.
}
