% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{cli_active_progress_bars}
\alias{cli_active_progress_bars}
\title{Returns the number of currently active cli progress bars}
\source{
Generated by miniextendr from Rust fn `cli_active_progress_bars`
}
\usage{
cli_active_progress_bars()
}
\description{
Returns the number of currently active cli progress bars
}
