% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{native_list_as_list}
\alias{native_list_as_list}
\title{Prefer list when both RNativeType and IntoList exist}
\source{
Generated by miniextendr from Rust fn \code{native_list_as_list}
}
\usage{
native_list_as_list(x)
}
\description{
Forces VECSXP even though a native vector would be possible.
}
\examples{
native_list_as_list(3L)
}
\keyword{internal}
