% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{native_list_as_native}
\alias{native_list_as_native}
\title{Prefer native vector when both RNativeType and IntoList exist}
\source{
Generated by miniextendr from Rust fn \code{native_list_as_native}
}
\usage{
native_list_as_native(x)
}
\description{
Forces an integer vector via \code{AsRNative}.
}
\examples{
native_list_as_native(3L)
}
\keyword{internal}
