% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ndarray_roundtrip_array}
\alias{ndarray_roundtrip_array}
\title{ndarray_roundtrip_array}
\usage{
ndarray_roundtrip_array(data)
}
\arguments{
\item{data}{An N-dimensional numeric array to roundtrip.}
}
\description{
Test roundtripping an N-dimensional numeric array through R and back.
}
