% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ndarray_roundtrip_int_matrix}
\alias{ndarray_roundtrip_int_matrix}
\title{ndarray_roundtrip_int_matrix}
\usage{
ndarray_roundtrip_int_matrix(data)
}
\arguments{
\item{data}{An integer matrix to roundtrip.}
}
\description{
Test roundtripping a 2D integer matrix through R and back.
}
