% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ndarray_roundtrip_int_vec}
\alias{ndarray_roundtrip_int_vec}
\title{ndarray_roundtrip_int_vec}
\usage{
ndarray_roundtrip_int_vec(data)
}
\arguments{
\item{data}{An integer vector to roundtrip.}
}
\description{
Test roundtripping a 1D integer array through R and back.
}
