% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ndarray_roundtrip_matrix}
\alias{ndarray_roundtrip_matrix}
\title{ndarray_roundtrip_matrix}
\usage{
ndarray_roundtrip_matrix(data)
}
\arguments{
\item{data}{A numeric matrix to roundtrip.}
}
\description{
Test roundtripping a 2D numeric matrix through R and back.
}
