% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ndarray_roundtrip_string_array}
\alias{ndarray_roundtrip_string_array}
\title{ndarray_roundtrip_string_array}
\usage{
ndarray_roundtrip_string_array(data)
}
\arguments{
\item{data}{An N-dimensional character array to roundtrip (may contain NA).}
}
\description{
Test roundtripping an N-dimensional character array through R and back (NA-safe).
}
