% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ndarray_roundtrip_string_matrix_lossy}
\alias{ndarray_roundtrip_string_matrix_lossy}
\title{ndarray_roundtrip_string_matrix_lossy}
\usage{
ndarray_roundtrip_string_matrix_lossy(data)
}
\arguments{
\item{data}{A character matrix to roundtrip.}
}
\description{
Test roundtripping a character matrix through \verb{Array2<String>}.
This path is lossy for missing values: \code{NA_character_} becomes \code{""}.
}
