% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ndarray_roundtrip_vec}
\alias{ndarray_roundtrip_vec}
\title{ndarray_roundtrip_vec}
\usage{
ndarray_roundtrip_vec(data)
}
\arguments{
\item{data}{A numeric vector to roundtrip.}
}
\description{
Test roundtripping a 1D numeric array through R and back.
}
