% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{needs_update}
\alias{needs_update}
\title{Determine if a thread file needs to be updated}
\usage{
needs_update(thread, thread_file)
}
\arguments{
\item{thread}{Thread object from API}

\item{thread_file}{Thread file info (from get_existing_thread_files)}
}
\value{
TRUE if update needed, FALSE otherwise
}
\description{
Determine if a thread file needs to be updated
}
