% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{nested_factor_split_1v1r}
\alias{nested_factor_split_1v1r}
\title{nested_factor_split_1v1r}
\usage{
nested_factor_split_1v1r()
}
\description{
1v1r split (as_factor): single Move row, dir as factor column.
}
