% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{nested_factor_split_nv1r}
\alias{nested_factor_split_nv1r}
\title{nested_factor_split_nv1r}
\usage{
nested_factor_split_nv1r()
}
\description{
Nv1r split (as_factor): one Move and one Stop row.
}
