% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{nested_factor_split_nvnr}
\alias{nested_factor_split_nvnr}
\title{nested_factor_split_nvnr}
\usage{
nested_factor_split_nvnr()
}
\description{
NvNr split (as_factor): multiple rows across both variants.
}
