% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{nested_flatten_align_1vnr}
\alias{nested_flatten_align_1vnr}
\title{nested_flatten_align_1vnr}
\usage{
nested_flatten_align_1vnr()
}
\description{
1vNr align (flatten): multiple Tracked rows, aligned data frame.
}
