% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{nested_flatten_split_1v1r}
\alias{nested_flatten_split_1v1r}
\title{nested_flatten_split_1v1r}
\usage{
nested_flatten_split_1v1r()
}
\description{
1v1r split (flatten): single Tracked/Ok row.
}
