% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{nested_flatten_split_1vnr}
\alias{nested_flatten_split_1vnr}
\title{nested_flatten_split_1vnr}
\usage{
nested_flatten_split_1vnr()
}
\description{
1vNr split (flatten): multiple Tracked rows (mix of Ok and Err).
}
