% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{nested_flatten_split_nv1r}
\alias{nested_flatten_split_nv1r}
\title{nested_flatten_split_nv1r}
\usage{
nested_flatten_split_nv1r()
}
\description{
Nv1r split (flatten): one Tracked and one Other row.
}
