% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{nested_flatten_split_nvnr}
\alias{nested_flatten_split_nvnr}
\title{nested_flatten_split_nvnr}
\usage{
nested_flatten_split_nvnr()
}
\description{
NvNr split (flatten): multiple rows across both variants.
}
