% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{nested_list_split_nvnr}
\alias{nested_list_split_nvnr}
\title{nested_list_split_nvnr}
\usage{
nested_list_split_nvnr()
}
\description{
NvNr split (as_list): multiple rows across both variants.
}
