% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{new_derived_point}
\alias{new_derived_point}
\title{new_derived_point}
\usage{
new_derived_point(x, y)
}
\arguments{
\item{x}{X coordinates.}

\item{y}{Y coordinates.}
}
\value{
A derived_point record vector with proxy methods.
}
\description{
Create a new derived_point record vector.
}
