% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{new_derived_rational}
\alias{new_derived_rational}
\title{new_derived_rational}
\usage{
new_derived_rational(n, d)
}
\arguments{
\item{n}{Numerator values.}

\item{d}{Denominator values.}
}
\value{
A derived_rational record vector.
}
\description{
Create a new derived_rational vector.
}
