% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrogen.R
\name{nitrogen}
\alias{nitrogen}
\title{Creates a data frame with the nitrogen content of litter samples}
\usage{
nitrogen(database)
}
\arguments{
\item{database}{A list with the aridec structure}
}
\value{
A data frame with the nitrogen content (%) of the litter samples from the database
}
\description{
Creates a data frame with the nitrogen content of litter samples
}
\examples{
N=nitrogen(database=aridec)
}
