% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\docType{methods}
\name{nlevels}
\alias{nlevels}
\title{The Number of Levels of a Factor}
\usage{
\S4method{nlevels}{fmVFactor}(x)
}
\arguments{
\item{x}{a FlashR factor}
}
\value{
The length of \code{levels(x)}, which is zero if \code{x}
has no levels.
}
\description{
Return the number of levels which its argument has.
}
\details{
This is applied to a FlashR factor.
}
\examples{
vec <- fm.as.factor(as.integer(fm.runif(100, min=0, max=100)))
nlevels(vec)
}

