% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{nrow}
\alias{ncol,fm-method}
\alias{nrow}
\alias{nrow,fm-method}
\title{The Number of Rows/Columns of a matrix.}
\usage{
\S4method{nrow}{fm}(x)

\S4method{ncol}{fm}(x)
}
\arguments{
\item{x}{a matrix.}
}
\value{
an integer of length 1 or NULL.
}
\description{
\code{nrow} and \code{ncol} return the number of rows or columns
present in \code{x}.
}
\examples{
mat <- fm.runif.matrix(100, 10)
nrow(mat)
ncol(mat)
}

