% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.bigextract.R
\name{nsdm.bigextract}
\alias{nsdm.bigextract}
\title{nsdm.bigextract}
\usage{
nsdm.bigextract(
  cov,
  data,
  rst_ref,
  cov_info,
  t_match = FALSE,
  tmatch_scheme = "npts",
  nzvt = 12,
  nsplits = ncores
)
}
\arguments{
\item{cov}{A character vector specifying the file paths of individual raster layers.}

\item{data}{An \code{nsdm.pseudoabsences} object used for covariate extraction.}

\item{rst_ref}{A raster object used as a reference (for cell identifiers, CRS, extent, etc.).}

\item{cov_info}{A \code{data.frame} containing key covariate information, as generated by \code{nsdm.covinfo}.}

\item{t_match}{Logical; whether to perform temporal matching (\code{TRUE}) or not (\code{FALSE}).}

\item{tmatch_scheme}{A character string specifying the temporal matching scheme:
\itemize{
\item \code{"npts"}: nearest previous time
\item \code{"nts"}: nearest time
}}

\item{nzvt}{A numeric threshold specifying the minimum number of unique values required for a covariate to be retained in the candidate set.}

\item{nsplits}{An integer indicating the number of cores to use for parallel extraction.}
}
\value{
An updated \code{nsdm.pseudoabsences} object.
}
\description{
Parallelized loading and spatiotemporal extraction of large raster datasets.
}
\author{
Antoine Adde (antoine.adde@eawag.ch)
}
