% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.ensembleeval.R
\name{nsdm.ensembleeval}
\alias{nsdm.ensembleeval}
\title{nsdm.ensembleeval}
\usage{
nsdm.ensembleeval(
  sets,
  level,
  model_names,
  species_name,
  nesting_name = NA,
  posthoc_nesting_name = NA,
  scratch_path
)
}
\arguments{
\item{sets}{A list of replicates; each replicate contains \code{reg_train}, \code{reg_test},
\code{glo_train}, and \code{glo_test}.}

\item{level}{Character string, either \code{"glo"} or \code{"reg"}, selecting which SIDs to use
for testing.}

\item{model_names}{Character vector of modeling algorithms to ensemble
(e.g. \code{c("glm", "rf", "esm")}).}

\item{species_name}{Character string, the name of the species or taxon being evaluated.}

\item{nesting_name}{Character string (default = \code{NA}), indicating the nesting strategies evaluated
\code{"posthoc"} or \code{"covariate"}.}

\item{posthoc_nesting_name}{Character string (default = \code{NA}), indicating the posthoc_nesting strategies evaluated
\code{"multiply"}, \code{"average"}, etc.}

\item{x}{An \code{nsdm.pseudoabsences} object that holds the master covariates, SIDs,
and presence values.}
}
\value{
A named list of ensemble evaluation scores, averaged across replicates.
Elements correspond to ensemble types (e.g. \code{"GLO"}, \code{"REG"}, \code{"MUL"}, \code{"MULW"}, \code{"COV"}).

Evaluation is performed with \code{\link[=nsdm.ceval]{nsdm.ceval()}}, returning mean values over replicates.
}
\description{
Evaluate ensemble predictions for a given species, level, and nesting strategy.
}
\details{
This function loads the best model(s) for each algorithm, applies them to test data,
and evaluates ensemble performance across replicates using specified ensemble criteria.
}
\author{
Antoine Adde (antoine.adde@eawag.ch)
}
