% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.partialPlot.gbm.R
\name{nsdm.partialPlot.gbm}
\alias{nsdm.partialPlot.gbm}
\title{nsdm.partialPlot.gbm}
\usage{
nsdm.partialPlot.gbm(
  obj,
  pred.data,
  xname,
  n.pt = 19,
  discrete.x = FALSE,
  subsample = pmin(1, n.pt * 100/nrow(pred.data)),
  which.class = NULL,
  xlab = deparse(substitute(xname)),
  ylab = "",
  type = if (discrete.x) "p" else "b",
  main = "",
  rug = TRUE,
  seed = NULL,
  ...
)
}
\arguments{
\item{obj}{model object}

\item{pred.data}{Matrix to be used in prediction (no xgb.DMatrix, no lgb.Data)}

\item{xname}{Name of column in \code{pred.data} according to that dependency plot is calculated}

\item{n.pt}{Evaluation grid size (used only if x is not discrete)}

\item{subsample}{Fraction of random lines in pred.data to be used in prediction}

\item{which.class}{Which class if objective is "multi:softprob" (value from 0 to num_class - 1)}

\item{xlab, }{ylab, main, type, ... Parameters passed to \code{plot}}

\item{seed}{Random seed used if \code{subsample < 1}}

\item{x.discrete}{If TRUE, the evaluation grid is set to the unique values of x}
}
\value{
List of prepared objects
}
\description{
Partial dependency plot for xgboost, lightgbm, or ranger object
}
\author{
Michael Mayer, \email{mayer@consultag.ch}
}
