% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.preps3.R
\name{nsdm.preps3}
\alias{nsdm.preps3}
\title{nsdm.preps3}
\usage{
nsdm.preps3(pseu_reg = NULL, pseu_glo, n_reps = 10, ratio_abs = 1)
}
\arguments{
\item{pseu_reg}{An \code{nsdm.pseudoabsences} object for the regional level (can be NULL)}

\item{pseu_glo}{An \code{nsdm.pseudoabsences} object for the global level}

\item{n_reps}{Integer, number of replicates (also number of spatial clusters)}

\item{ratio_abs}{Numeric, absences per presence, set to \code{0} to skip absences}
}
\value{
A named list of length \code{n_reps}. Each element has \code{reg_train}, \code{reg_test},
\code{glo_train}, \code{glo_test}, each an \code{nsdm.pseudoabsences} object or \code{NULL}
}
\description{
Build spatially stratified train/test sets for regional and global pseudoabsence data.
Presences are split by k-means spatial clustering. Each replicate selects one cluster
as the test set and uses the others as training. Absences are randomly sampled to
balance the sets according to \code{ratio_abs}.
}
\author{
Antoine Adde (antoine.adde@eawag.ch)
}
