% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.savemap.R
\name{nsdm.savemap}
\alias{nsdm.savemap}
\title{nsdm.savemap}
\usage{
nsdm.savemap(maps, save_path, species_name, model_name = NULL, format = "both")
}
\arguments{
\item{maps}{A list of rasters (or a single raster) filled with predicted values, generated by \code{nsdm.map}.}

\item{save_path}{A character string specifying the directory where output maps will be saved.}

\item{species_name}{A character string indicating the name of the target taxon.}

\item{model_name}{A character string indicating the name of the target modeling algorithm.}

\item{format}{A character string specifying the file format(s) to save (\code{"rds"} for RDS files, \code{"tif"} for GeoTIFF, or \code{"both"} for both formats).}
}
\value{
This function does not return a value; it saves maps as \code{.rds} and/or \code{.tif} files in the specified directory.
}
\description{
Save individual predictive maps to disk.
}
\author{
Antoine Adde (antoine.adde@eawag.ch)
}
