% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.savethis.R
\name{nsdm.savethis}
\alias{nsdm.savethis}
\title{nsdm.savethis}
\usage{
nsdm.savethis(
  object,
  model_name = NULL,
  species_name = NULL,
  tag = NULL,
  compression = FALSE,
  save_path
)
}
\arguments{
\item{object}{Object to be saved}

\item{model_name}{A character string indicating the name of the target modelling algorithm}

\item{species_name}{A character string  indicating the name of the target taxon}

\item{tag}{A character string  indicating the tag for the target hyperparameter X modelling algorithm combination}

\item{compression}{Logical (TRUE FALSE) for saveRDS "compress" argument}

\item{save_path}{A character string indicating the upper path where to save the object}
}
\value{
Saved object in .rds format
}
\description{
Easy saving of .rds nsdm.ine objects
}
\author{
Antoine Adde (aadde@unil.ch)
}
