% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{nudge}
\alias{nudge}
\title{Nudge the counter up by one, warning once it passes the limit}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Nudge the counter up by one, warning once it passes the limit
}
