% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{nudge_gauge}
\alias{nudge_gauge}
\title{Nudge the gauge up by one, returning the gauge visibly}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Nudge the gauge up by one, returning the gauge visibly
}
