% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{numeric}
\alias{as.numeric,fm-method}
\alias{as.numeric,fmV-method}
\alias{is.numeric,fm-method}
\alias{is.numeric,fmV-method}
\alias{numeric}
\title{Numeric Vectors}
\usage{
\S4method{as.numeric}{fm}(x)

\S4method{as.numeric}{fmV}(x)

\S4method{is.numeric}{fm}(x)

\S4method{is.numeric}{fmV}(x)
}
\arguments{
\item{x}{a FlashR object to be coerced or tested.}
}
\value{
\code{as.numeric} returns a numeric FlashR object,
\code{is.numeric} returns a logical value.
}
\description{
Coerces or test objects of type \code{"numeric"}. \code{is.numeric}
is a more general test of an object being interpretable as numbers.
}
\examples{
vec <- as.numeric(fm.runif(1000, min=0, max=10))
is.numeric(vec)
}

