% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{old_deprecated_fn}
\alias{old_deprecated_fn}
\title{old_deprecated_fn}
\usage{
old_deprecated_fn(x)
}
\arguments{
\item{x}{Integer input value.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A function marked with Rust's \verb{#[deprecated]} attribute.
This should generate lifecycle::deprecate_warn in the R wrapper.
}
