% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{omit_grade}
\alias{omit_grade}
\title{Report an omittable inline choice through an S3 trait method}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Report an omittable inline choice through an S3 trait method
}
