% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{omit_grades}
\alias{omit_grades}
\title{Report omittable inline choices through an S3 trait method}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Report omittable inline choices through an S3 trait method
}
