% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{omit_pick_s3}
\alias{omit_pick_s3}
\title{Report omittable choice parameters through an S3 method}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Report omittable choice parameters through an S3 method
}
