% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{on_exit_lifo}
\alias{on_exit_lifo}
\title{Test on.exit cleanup with after = false (LIFO order)}
\source{
Generated by miniextendr from Rust fn \code{on_exit_lifo}
}
\usage{
on_exit_lifo(x)
}
\arguments{
\item{x}{Input value.}
}
\description{
Test on.exit cleanup with after = false (LIFO order)
}
