% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{on_exit_no_add}
\alias{on_exit_no_add}
\title{on_exit_no_add}
\usage{
on_exit_no_add(x)
}
\arguments{
\item{x}{Input value.}
}
\description{
Test on.exit cleanup with add = false (overwrite previous).
}
