% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{on_exit_short}
\alias{on_exit_short}
\title{Test on.exit cleanup with short form}
\source{
Generated by miniextendr from Rust fn \code{on_exit_short}
}
\usage{
on_exit_short(x)
}
\arguments{
\item{x}{Input value.}
}
\description{
Test on.exit cleanup with short form
}
