% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onepFitFlux.R
\name{onepFitFlux}
\alias{onepFitFlux}
\title{Fits a one pool model to a time-series to obtain the vector of release fluxes out of the pools for all times}
\usage{
onepFitFlux(timeSeries, initialCarbon)
}
\arguments{
\item{timeSeries}{A time series of respiration values over time}

\item{initialCarbon}{The initial amount of carbon in units that correspond to the time series data}
}
\value{
R list with an FME model object, a SoilR model object, and the AIC value
}
\description{
Fits a one pool model to a time-series to obtain the vector of release fluxes out of the pools for all times
}
\examples{
a=onepFitFlux(timeSeries = sidb[[19]]$timeSeries[,1:2],
initialCarbon=sidb[[19]]$initConditions[1,"carbonMean"]*10)
}
