% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{optional_string}
\alias{optional_string}
\title{The \code{optional_string()} generic}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{optional_string()} is an S7 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{ScalarOptionS7}}
}
}
