% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ordered_float_is_finite}
\alias{ordered_float_is_finite}
\title{ordered_float_is_finite}
\usage{
ordered_float_is_finite(x)
}
\arguments{
\item{x}{Numeric scalar.}
}
\description{
Test whether a value is finite using OrderedFloat.
}
