% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ordered_float_is_nan}
\alias{ordered_float_is_nan}
\title{ordered_float_is_nan}
\usage{
ordered_float_is_nan(x)
}
\arguments{
\item{x}{Numeric scalar.}
}
\description{
Test whether a value is NaN using OrderedFloat.
}
