% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ordered_float_ops_predicates}
\alias{ordered_float_ops_predicates}
\title{ordered_float_ops_predicates}
\usage{
ordered_float_ops_predicates(x)
}
\arguments{
\item{x}{Numeric scalar.}
}
\description{
Predicate methods of \code{ROrderedFloatOps}, via the trait.
}
