% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ordered_float_roundtrip}
\alias{ordered_float_roundtrip}
\title{ordered_float_roundtrip}
\usage{
ordered_float_roundtrip(x)
}
\arguments{
\item{x}{Numeric scalar.}
}
\description{
Test OrderedFloat\if{html}{\out{<f64>}} scalar roundtrip through R.
}
