% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ordered_float_roundtrip_vec}
\alias{ordered_float_roundtrip_vec}
\title{ordered_float_roundtrip_vec}
\usage{
ordered_float_roundtrip_vec(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\description{
Test Vec<OrderedFloat\if{html}{\out{<f64>}}> roundtrip through R.
}
