% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ordered_float_sort}
\alias{ordered_float_sort}
\title{ordered_float_sort}
\usage{
ordered_float_sort(x)
}
\arguments{
\item{x}{Numeric vector to sort.}
}
\description{
Test sorting a numeric vector using OrderedFloat total ordering.
}
