% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ordered_float_sort_special}
\alias{ordered_float_sort_special}
\title{ordered_float_sort_special}
\usage{
ordered_float_sort_special(x)
}
\arguments{
\item{x}{Numeric vector with special values.}
}
\description{
Test sorting a vector containing Inf, -Inf, NaN, and normal values.
}
