% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageLoad.r
\name{packageLoad}
\alias{packageLoad}
\title{Package Load Function}
\usage{
packageLoad(package_list = "tictoc")
}
\arguments{
\item{package_list}{String or vector of strings. Defaults to "tictoc".}
}
\description{
This function accepts a string (or vector of strings) representing a package name. It checks to see if any of the packages are not installed and installs them. Then, it loads all the packages.
}
\examples{
packageLoad("beepr")

list_of_packages = c("chron","corrplot","data.table")
packageLoad(list_of_packages)
}
\keyword{packages}
