% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{panic_location_worker_nested}
\alias{panic_location_worker_nested}
\title{panic_location_worker_nested}
\usage{
panic_location_worker_nested()
}
\description{
Nested panic on the WORKER thread — delegates to \code{plain_boom}.
}
