% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{panic_method}
\alias{panic_method}
\title{The \code{panic_method()} generic}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{panic_method()} is an S7 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{ErrorInRS7Gauge}}
}
}
