% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{telemetry_install_counter}
\alias{telemetry_install_counter}
\alias{telemetry_get_count}
\alias{telemetry_clear_hook}
\title{telemetry_install_counter}
\usage{
telemetry_install_counter()

telemetry_get_count()

telemetry_clear_hook()
}
\description{
Install a telemetry hook that counts panics.

Get the telemetry counter value.

Clear the telemetry hook.
}
