% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{PanickyCounter}
\alias{PanickyCounter}
\alias{PanickyCounter$get_value}
\alias{PanickyCounter$try_add}
\alias{PanickyCounter$new_panicky}
\alias{$.PanickyCounter}
\alias{[[.PanickyCounter}
\alias{PanickyCounter$Counter$value}
\alias{PanickyCounter$Counter$increment}
\alias{PanickyCounter$Counter$checked_add}
\alias{PanickyCounter$Counter$default_initial}
\alias{PanickyCounter$Counter$MAX_VALUE}
\title{PanickyCounter Class}
\usage{
PanickyCounter

\method{$}{PanickyCounter}(self, name)

\method{[[}{PanickyCounter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Get value.

Add (may panic).

Create a new panicky counter.
}
