% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{parse_draft_comment}
\alias{parse_draft_comment}
\title{Parse a single draft comment section}
\usage{
parse_draft_comment(lines, section)
}
\arguments{
\item{lines}{Character vector of file lines}

\item{section}{Section info from find_draft_comments}
}
\value{
List with 'params' and 'content' elements
}
\description{
Parse a single draft comment section
}
