% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{path_subset_relation}
\alias{path_subset_relation}
\title{Create subset relation graph from sets}
\usage{
path_subset_relation(sets)
}
\arguments{
\item{sets}{A named list of integer or character vectors representing sets (e.g., exonic part paths).}
}
\description{
Create subset relation graph from sets
}
\examples{
sets <- list("1" = c(1L, 2L), "2" = c(1L, 2L, 3L), "3" = c(4L, 5L))
dag <- grase::path_subset_relation(sets)
igraph::E(dag)
}
