% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Extract.R
\name{PathExtract}
\alias{PathExtract}
\title{Walk downstream from a point and extract values from a matrix}
\usage{
PathExtract(input, direction, mask, startpoint, d4 = c(1, 2, 3, 4))
}
\arguments{
\item{input}{The matrix of values that you would like to extract the streampath from}

\item{direction}{a matrix of the flow directions for cells that have been processed - defaults to all zeros}

\item{mask}{Mask with ones for cells to be processed and zeros for everything else - defaults to a mask of all 1's}

\item{startpoint}{The x,y index of the grid cell you would like to start from}

\item{d4}{directional numbering system: the numbers you want to assigne to down, left, top,right (defaults to 1,2,3,4)}
}
\description{
This function grabs out values from a matrix by walking downstream from a point
}
