% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_Distance.R
\name{PeakDist}
\alias{PeakDist}
\title{Calculate distance to topographic peaks}
\usage{
PeakDist(direction, mask, d4 = c(1, 2, 3, 4), printflag = F)
}
\arguments{
\item{direction}{a matrix of the flow directions for cells that have been processed - defaults to all zeros}

\item{mask}{Mask with ones for cells to be processed and zeros for everything else - defaults to a mask of all 1's}

\item{d4}{directional numbering system: the numbers you want to assigne to down, left, top,right (defaults to 1,2,3,4)}

\item{printflag}{optional flag to print function progress. this is a test}
}
\description{
Calculates the maximum and minimum distance from a head water cell
}
