% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exontest_functions.R
\name{phi_estimate_glmmTMB}
\alias{phi_estimate_glmmTMB}
\title{Estimate beta-binomial dispersion (phi) per event using glmmTMB}
\usage{
phi_estimate_glmmTMB(dd)
}
\arguments{
\item{dd}{A data frame of exon count data.}
}
\description{
Estimate beta-binomial dispersion (phi) per event using glmmTMB
}
\examples{
\dontrun{
splitcnts <- read.table("bipartition.internal.exoncnt.combined.txt",
                        header = TRUE, row.names = NULL)
grouped_data <- grase::group_by_event(splitcnts, "diff", "n")
phi_result <- grase::phi_estimate_glmmTMB(grouped_data[[1]])
phi_result
}
}
