% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exontest_functions.R
\name{phi_map_glmmTMB}
\alias{phi_map_glmmTMB}
\title{Estimate MAP log-phi per event using glmmTMB with an empirical Bayes prior}
\usage{
phi_map_glmmTMB(dd, prior_disp)
}
\arguments{
\item{dd}{A data frame of exon count data.}

\item{prior_disp}{A data frame specifying the glmmTMB prior on the dispersion parameter.}
}
\description{
Estimate MAP log-phi per event using glmmTMB with an empirical Bayes prior
}
\examples{
\dontrun{
splitcnts <- read.table("bipartition.internal.exoncnt.combined.txt",
                        header = TRUE, row.names = NULL)
grouped_data <- grase::group_by_event(splitcnts, "diff", "n")
prior_disp <- data.frame(prior = "normal(-1.5, 1.2)", class = "fixef_disp",
                         coef = "", stringsAsFactors = FALSE)
phi_map <- grase::phi_map_glmmTMB(grouped_data[[1]], prior_disp)
phi_map
}
}
