% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/info.R
\name{pkg_bug}
\alias{pkg_bug}
\title{Report a bug for a package}
\usage{
pkg_bug(pkg)
}
\arguments{
\item{pkg}{Name of the package.}
}
\description{
Open the web page of the package's bug tracker, or an email
client to email the package's maintainer.
}
\details{
This command is similar to \code{utils::bug.report}.
}

