% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/download.R
\name{pkg_download}
\alias{pkg_download}
\title{Download packages from the CRAN mirror at Github}
\usage{
pkg_download(pkgs, dest_dir = ".")
}
\arguments{
\item{pkgs}{Package(s) to download. By default the latest
versions are downloaded, a specific version number can
be given after a dash.}

\item{dest_dir}{Destination directory.}
}
\value{
A character vector containing the full paths of the
  downloaded packages.
}
\description{
The files will be named according to the package names and
version numbers.
}
\examples{
dest_dir <- tempdir()
pkg_download("testthat", dest_dir = dest_dir)
pkg_download("testthat-0.8.1", dest_dir = dest_dir)
}

