% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/install.R
\name{pkg_install}
\alias{pkg_install}
\title{Install packages}
\usage{
pkg_install(pkgs, global = FALSE, ask = FALSE, all = FALSE,
  download_dir = tempfile())
}
\arguments{
\item{pkgs}{The package(s) to install. Specific versions
can be specified after a dash.}

\item{global}{Whether to install packages globally.}

\item{ask}{Whether to ask for confirmation before proceeding
with the installation.}

\item{all}{Whether to install suggested and enhancaed packages
as well, not only just the ones \code{pkgs} depend on, import from and
link to. Set this to \code{TRUE} to make sure that you can run
all examples and vignettes from the package(s).}

\item{download_dir}{Directory to store the downloaded
packages in.}
}
\value{
Invisibly a logical vector which is \code{TRUE} for
  successfully installed packages, and \code{FALSE} for others.
}
\description{
Install packages
}

