% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/paths.R
\name{pkg_paths}
\alias{pkg_paths}
\title{Default library directories}
\usage{
pkg_paths(new_value)
}
\arguments{
\item{new_value}{If specified, then the user-defined library
directory list will be set to this value.}
}
\value{
The current value of the default library directories.
}
\description{
If called without an argument, then it returns the library
directories. If called with an argument then it sets them.
}
\details{
Note that the system-default R library directory (or directories
possibly) are always searched (and returned by \code{pkg_paths}.
}
\examples{
pkg_paths()
}

