% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/tree.R
\name{pkg_tree}
\alias{pkg_tree}
\title{Dependency tree of a package}
\usage{
pkg_tree(pkg, include_base = FALSE)
}
\arguments{
\item{pkg}{Package name, optionally with version number after a dash.}

\item{include_base}{Whether to include base R packages in the tree.}
}
\description{
Dependency tree of a package
}

