% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/upgrade.R
\name{pkg_upgrade}
\alias{pkg_upgrade}
\title{Upgrade packages}
\usage{
pkg_upgrade(filter = "", lib = pkg_paths(), ...)
}
\arguments{
\item{filter}{Regular expression to filter the packages to update.}

\item{lib}{Library paths to look at.}

\item{...}{Extra arguments are passed to \code{\link{pkg_install}}.}
}
\value{
Invisibly a logical vector which is \code{TRUE} for
  successfully installed packages, and \code{FALSE} for others.
}
\description{
Upgrade packages
}

