% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plan_routes_s3}
\alias{plan_routes_s3}
\title{Describe route names or a data frame, and tiers, through an S3 method}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Describe route names or a data frame, and tiers, through an S3 method
}
