% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plan_s3}
\alias{plan_s3}
\title{Describe a route name or data frame, and a level, through an S3 method}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Describe a route name or data frame, and a level, through an S3 method
}
