% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{plan_s7}
\alias{plan_s7}
\title{The \code{plan_s7()} generic}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{plan_s7()} is an S7 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{EitherRouteS7}}
}
}
