% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_relabun_box}
\alias{plot_relabun_box}
\title{Plot relative abundances as a box plot}
\usage{
plot_relabun_box(dat, xcol = rel_abun, ycol = otu, colorcol = dx)
}
\arguments{
\item{dat}{a long data frame (e.g from \code{wide_to_long()})}

\item{xcol}{column for the x aesthetic}

\item{ycol}{column for the y aesthetic}

\item{colorcol}{column for the color aesthetic}
}
\value{
a ggplot2 box plot of OTU relative abundances
}
\description{
Plot relative abundances as a box plot
}
