% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEntry.R
\name{plotEntry}
\alias{plotEntry}
\title{Plot individual entries of the aridec dataset}
\usage{
plotEntry(entry)
}
\arguments{
\item{entry}{character string with the name of the entry to be plotted}
}
\value{
A plot
}
\description{
Plot individual entries of the aridec dataset
}
\examples{
plotEntry(entry=aridec[["Adair2017"]])
}
