% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{pm_s3_reseed}
\alias{pm_s3_reseed}
\title{Reseed with a \code{Checked} parameter, via S3 dispatch}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Reseed with a \code{Checked} parameter, via S3 dispatch
}
