% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{Point}
\alias{Point}
\alias{Point$new}
\alias{Point$x}
\alias{Point$y}
\alias{$.Point}
\alias{[[.Point}
\alias{Point$RDebug$debug_str}
\alias{Point$RDebug$debug_str_pretty}
\alias{Point$RDisplay$as_r_string}
\alias{Point$RHash$hash}
\alias{Point$RClone$clone}
\alias{Point$RDefault$default}
\alias{Point$RFromStr$from_str}
\alias{Point$RCopy$copy}
\alias{Point$RCopy$is_copy}
\alias{Point$ROrd$cmp}
\title{Point  Class}
\usage{
Point

\method{$}{Point}(self, name)

\method{[[}{Point}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Point inherent methods: constructor and field accessors.
}
