% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_surrogate.R
\name{point_nex}
\alias{point_nex}
\title{Transition procedure to the next point from twins}
\usage{
point_nex(curr, twins, knot, probs)
}
\arguments{
\item{curr}{Integer, the position of current state}

\item{twins}{A list, that preserves twins for each state (calculated by \code{\link{find_twins}})}

\item{knot}{Vector or 2-column matrix, which specifies the knot positions in the data}

\item{probs}{List of twin probabilitities (see \code{\link{twin_probs}})}
}
\description{
\code{point_nex} proceeds the time series with swapping between original and twin states
If valid data does not exist, this function returns 0
}
