% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{post_comment}
\alias{post_comment}
\title{Post a comment to a thread}
\usage{
post_comment(
  thread_id,
  content,
  token = twist_token(),
  recipients = "EVERYONE_IN_THREAD",
  direct_mentions = NULL,
  direct_group_mentions = NULL,
  groups = NULL,
  mark_thread_position = TRUE,
  thread_action = NULL
)
}
\arguments{
\item{thread_id}{ID of the thread to comment on}

\item{content}{Content of the comment (supports Markdown)}

\item{token}{Authentication token (default: from twist_token())}

\item{recipients}{Users to notify. Can be a list of user IDs, "EVERYONE",
or "EVERYONE_IN_THREAD" (default: "EVERYONE_IN_THREAD")}

\item{direct_mentions}{List of user IDs that are directly mentioned}

\item{direct_group_mentions}{List of group IDs that are directly mentioned}

\item{groups}{List of group IDs to notify}

\item{mark_thread_position}{Whether to mark the thread position (default: TRUE)}

\item{thread_action}{Optional action: "close" or "reopen"}
}
\value{
Comment object from the API
}
\description{
Post a comment to a thread
}
