% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exontest_functions.R
\name{prec_estimate_vgam}
\alias{prec_estimate_vgam}
\title{Estimate Dirichlet-Multinomial precision per event using VGAM}
\usage{
prec_estimate_vgam(dd)
}
\description{
Estimate Dirichlet-Multinomial precision per event using VGAM
}
\examples{
\dontrun{
splitcnts <- read.table("multinomial.exoncnt.combined.txt",
                        header = TRUE, row.names = NULL)
grouped_counts <- dplyr::group_split(dplyr::group_by(splitcnts, gene, event))
prec_result <- grase::prec_estimate_vgam(grouped_counts[[1]])
prec_result
}
}
