% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch.R
\name{print.Rsch_connection}
\alias{print.Rsch_connection}
\title{Print properties of an Rsch_connection object}
\usage{
\method{print}{Rsch_connection}(x, ...)
}
\arguments{
\item{x}{an Rsch_connection object}

\item{...}{currently ignored}
}
\description{
Print properties of an Rsch_connection object
}
