% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch.R
\name{printout}
\alias{printout}
\title{Read back from the remote system}
\usage{
printout(connection, split = "\\n")
}
\arguments{
\item{connection}{An Rsch_connection object}

\item{split}{passed on to \code{\link[base]{strsplit}} UNLESS the value is \code{NA}
in which case a single string will be returned.  The default is \code{\\n}}
}
\value{
generally a vector of strings representing the console output resulting from the last
command issued to \code{\link{execute}}.
}
\description{
This function returns a vector of strings that is the entire output of the remote system since
the last call to \code{\link{execute}}
}
\examples{
\dontrun{
connection <- get_connection(host = "cuckoo.lbnl.gov", user = "cliff", password = "TopSecret")
execute(connection,"grep \\"nuclear\\" /*")
results <- printout(connection)

}
}
