% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorityFlow.R
\docType{package}
\name{PriorityFlow-package}
\alias{PriorityFlow}
\alias{PriorityFlow-package}
\title{PriorityFlow: What the package does (short line)}
\description{
More about what it does (maybe more than one line)
}
\details{
PriorityFlow is a toolkit for topographic processing for hydrologic models. 
This repo contains a set of R Functions and example workflow scripts.
}
\keyword{internal}
