% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\name{profile}
\alias{fm.start.profiler}
\alias{fm.stop.profiler}
\alias{profile}
\title{Google profiler}
\usage{
fm.start.profiler(file)

fm.stop.profiler()
}
\arguments{
\item{file}{a string that indicates the file where the profiling result
is saved.}
}
\description{
This uses the Google profiler to profile the execution of the code.
https://github.com/gperftools/gperftools
}

