% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{ptr_list_as_ptr}
\alias{ptr_list_as_ptr}
\title{Prefer external pointer when both ExternalPtr and IntoList exist}
\source{
Generated by miniextendr from Rust fn \code{ptr_list_as_ptr}
}
\usage{
ptr_list_as_ptr(x)
}
\description{
\code{AsExternalPtr} wins over list conversion.
}
\examples{
ptr_list_as_ptr(2L)
}
\keyword{internal}
