% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalueAdjustment.R
\name{pvalueAdjustment}
\alias{pvalueAdjustment}
\title{pvalueAdjustment function}
\usage{
pvalueAdjustment(
  res,
  independentFiltering,
  filter,
  theta,
  alpha,
  pAdjustMethod
)
}
\description{
pvalueAdjustment function
}
