% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metawards.R
\name{py_install_metawards}
\alias{py_install_metawards}
\title{Install the metawards python module}
\usage{
py_install_metawards(method = "auto", python_version = NULL)
}
\arguments{
\item{method}{The installation method to use (see reticulate). Defaults
to auto, which means that it will install into the
main environment (no conda-env or pipenv)}

\item{python_version}{The version of Python to use. This will be
be automatically set, and should be greater
than or equal to 3.7. Only use if you want to force
a particular python version to be used.}
}
\description{
Install the metawards python module
}
