% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metawards.R
\name{py_metawards_update_available}
\alias{py_metawards_update_available}
\title{Return whether a newer version of MetaWards is available.
If there is, this returns the version number. If not, this
returns NULL}
\usage{
py_metawards_update_available()
}
\value{
The version number of the update, if available
}
\description{
Return whether a newer version of MetaWards is available.
If there is, this returns the version number. If not, this
returns NULL
}
