% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metawards.R
\name{py_update_metawards}
\alias{py_update_metawards}
\title{Update the MetaWards Python module to the latest available
version.}
\usage{
py_update_metawards()
}
\value{
Returns TRUE if MetaWards was updated
}
\description{
Update the MetaWards Python module to the latest available
version.
}
