% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{quiet_nudge_gauge}
\alias{quiet_nudge_gauge}
\title{Nudge the gauge up by one, returning the gauge invisibly}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Nudge the gauge up by one, returning the gauge invisibly
}
