% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rpkg_r6_standalone}
\alias{rpkg_r6_standalone}
\alias{r6_standalone_add}
\title{r6_standalone_add}
\usage{
r6_standalone_add(a, b)
}
\arguments{
\item{a}{(no documentation available)}

\item{b}{(no documentation available)}
}
\description{
Test standalone function mixed with R6 impl blocks.
}
\examples{
r6_standalone_add(1L, 2L)
}
