% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{R6Cloneable}
\alias{R6Cloneable}
\title{R6Cloneable R6 Class}
\description{
R6 cloneable class with locked class definition.
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-R6Cloneable-initialize}{\code{R6Cloneable$new()}}
    \item \href{#method-R6Cloneable-get_value}{\code{R6Cloneable$get_value()}}
    \item \href{#method-R6Cloneable-set_value}{\code{R6Cloneable$set_value()}}
    \item \href{#method-R6Cloneable-clone}{\code{R6Cloneable$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Cloneable-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-R6Cloneable-initialize}{}}}
\subsection{\code{R6Cloneable$new()}}{
  Creates a new instance.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{R6Cloneable$new(value, .ptr = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{value}}{Integer value to store.}
      \item{\code{.ptr}}{Internal pointer (used by static methods, not for direct use).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Cloneable-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-R6Cloneable-get_value}{}}}
\subsection{\code{R6Cloneable$get_value()}}{
  Returns the value.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{R6Cloneable$get_value()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Cloneable-set_value"></a>}}
\if{latex}{\out{\hypertarget{method-R6Cloneable-set_value}{}}}
\subsection{\code{R6Cloneable$set_value()}}{
  Sets the value.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{R6Cloneable$set_value(value)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{value}}{Integer value to store.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Cloneable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-R6Cloneable-clone}{}}}
\subsection{\code{R6Cloneable$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{R6Cloneable$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
