% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{R6NonPortable}
\alias{R6NonPortable}
\title{R6NonPortable R6 Class}
\description{
R6 non-portable class for testing portable = FALSE flag.
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-R6NonPortable-initialize}{\code{R6NonPortable$new()}}
    \item \href{#method-R6NonPortable-get_value}{\code{R6NonPortable$get_value()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6NonPortable-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-R6NonPortable-initialize}{}}}
\subsection{\code{R6NonPortable$new()}}{
  Create a new \code{R6NonPortable}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{R6NonPortable$new(value, .ptr = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{value}}{Integer value to store.}
      \item{\code{.ptr}}{Internal pointer (used by static methods, not for direct use).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6NonPortable-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-R6NonPortable-get_value}{}}}
\subsection{\code{R6NonPortable$get_value()}}{
  Method \code{get_value}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{R6NonPortable$get_value()}
    \if{html}{\out{</div>}}
  }
}

}
