% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{r_backed_rdmatrix_roundtrip}
\alias{r_backed_rdmatrix_roundtrip}
\title{r_backed_rdmatrix_roundtrip}
\usage{
r_backed_rdmatrix_roundtrip(m)
}
\arguments{
\item{m}{Numeric matrix input.}
}
\description{
Test nalgebra RDMatrix\if{html}{\out{<f64>}} identity roundtrip through R.
}
