% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{r_backed_rdmatrix_scale}
\alias{r_backed_rdmatrix_scale}
\title{r_backed_rdmatrix_scale}
\usage{
r_backed_rdmatrix_scale(m, factor)
}
\arguments{
\item{m}{Numeric matrix input.}

\item{factor}{Numeric scaling factor.}
}
\description{
Test scaling all elements of an RDMatrix\if{html}{\out{<f64>}} by a scalar factor.
}
