% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{r_backed_rdmatrix_sum}
\alias{r_backed_rdmatrix_sum}
\title{r_backed_rdmatrix_sum}
\usage{
r_backed_rdmatrix_sum(m)
}
\arguments{
\item{m}{Numeric matrix input.}
}
\description{
Test summing all elements of an RDMatrix\if{html}{\out{<f64>}}.
}
