% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{r_backed_rdvector_dot}
\alias{r_backed_rdvector_dot}
\title{r_backed_rdvector_dot}
\usage{
r_backed_rdvector_dot(a, b)
}
\arguments{
\item{a}{Numeric vector input.}

\item{b}{Numeric vector input.}
}
\description{
Test computing the dot product of two RDVector\if{html}{\out{<f64>}}.
}
