% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{r_backed_rdvector_int_sum}
\alias{r_backed_rdvector_int_sum}
\title{r_backed_rdvector_int_sum}
\usage{
r_backed_rdvector_int_sum(v)
}
\arguments{
\item{v}{Integer vector input.}
}
\description{
Test summing all elements of an RDVector\if{html}{\out{<i32>}}.
}
