% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{r_backed_rdvector_norm}
\alias{r_backed_rdvector_norm}
\title{r_backed_rdvector_norm}
\usage{
r_backed_rdvector_norm(v)
}
\arguments{
\item{v}{Numeric vector input.}
}
\description{
Test computing the Euclidean norm of an RDVector\if{html}{\out{<f64>}}.
}
