% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{r_backed_rdvector_roundtrip}
\alias{r_backed_rdvector_roundtrip}
\title{r_backed_rdvector_roundtrip}
\usage{
r_backed_rdvector_roundtrip(v)
}
\arguments{
\item{v}{Numeric vector input.}
}
\description{
Test nalgebra RDVector\if{html}{\out{<f64>}} identity roundtrip through R.
}
