% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{r_backed_rdvector_scale}
\alias{r_backed_rdvector_scale}
\title{r_backed_rdvector_scale}
\usage{
r_backed_rdvector_scale(v, factor)
}
\arguments{
\item{v}{Numeric vector input.}

\item{factor}{Numeric scaling factor.}
}
\description{
Test scaling an RDVector\if{html}{\out{<f64>}} by a scalar factor.
}
