% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{r_backed_rndmat_fill}
\alias{r_backed_rndmat_fill}
\title{r_backed_rndmat_fill}
\usage{
r_backed_rndmat_fill(m, value)
}
\arguments{
\item{m}{Numeric matrix input.}

\item{value}{Numeric fill value.}
}
\description{
Test filling all elements of an RndMat\if{html}{\out{<f64>}} with a constant value.
}
