% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{r_backed_rndmat_nrow}
\alias{r_backed_rndmat_nrow}
\title{r_backed_rndmat_nrow}
\usage{
r_backed_rndmat_nrow(m)
}
\arguments{
\item{m}{Numeric matrix input.}
}
\description{
Test getting the number of rows of an RndMat\if{html}{\out{<f64>}}.
}
