% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{r_backed_rndmat_trace}
\alias{r_backed_rndmat_trace}
\title{r_backed_rndmat_trace}
\usage{
r_backed_rndmat_trace(m)
}
\arguments{
\item{m}{Numeric square matrix input.}
}
\description{
Test computing the trace (sum of diagonal) of an RndMat\if{html}{\out{<f64>}}.
}
