% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{r_backed_rndvec_double}
\alias{r_backed_rndvec_double}
\title{r_backed_rndvec_double}
\usage{
r_backed_rndvec_double(v)
}
\arguments{
\item{v}{Numeric vector input.}
}
\description{
Test doubling each element of an RndVec\if{html}{\out{<f64>}} via ndarray operations.
}
