% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{r_backed_rndvec_empty_sum}
\alias{r_backed_rndvec_empty_sum}
\title{r_backed_rndvec_empty_sum}
\usage{
r_backed_rndvec_empty_sum(v)
}
\arguments{
\item{v}{Numeric vector input (expected empty).}
}
\description{
Test summing an empty RndVec\if{html}{\out{<f64>}} (should return 0.0).
}
