% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{r_backed_rndvec_int_roundtrip}
\alias{r_backed_rndvec_int_roundtrip}
\title{r_backed_rndvec_int_roundtrip}
\usage{
r_backed_rndvec_int_roundtrip(v)
}
\arguments{
\item{v}{Integer vector input.}
}
\description{
Test ndarray RndVec\if{html}{\out{<i32>}} identity roundtrip through R.
}
