% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{r_backed_rndvec_len}
\alias{r_backed_rndvec_len}
\title{r_backed_rndvec_len}
\usage{
r_backed_rndvec_len(v)
}
\arguments{
\item{v}{Numeric vector input.}
}
\description{
Test getting the length of an RndVec\if{html}{\out{<f64>}}.
}
