% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{r_backed_rndvec_sum}
\alias{r_backed_rndvec_sum}
\title{r_backed_rndvec_sum}
\usage{
r_backed_rndvec_sum(v)
}
\arguments{
\item{v}{Numeric vector input.}
}
\description{
Test summing all elements of an RndVec\if{html}{\out{<f64>}}.
}
