% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{r_entry_demo}
\alias{r_entry_demo}
\title{Coerce x to integer before passing to Rust}
\source{
Generated by miniextendr from Rust fn \code{r_entry_demo}
}
\usage{
r_entry_demo(x)
}
\arguments{
\item{x}{Value to coerce.}
}
\description{
Coerce x to integer before passing to Rust
}
