% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{r_post_checks_demo}
\alias{r_post_checks_demo}
\title{Log a message after checks pass}
\source{
Generated by miniextendr from Rust fn \code{r_post_checks_demo}
}
\usage{
r_post_checks_demo(x)
}
\arguments{
\item{x}{Input value.}
}
\description{
Log a message after checks pass
}
