% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rand_rdistributions_sample}
\alias{rand_rdistributions_sample}
\title{rand_rdistributions_sample}
\usage{
rand_rdistributions_sample()
}
\description{
Draw one value from each \code{RDistributions} method, in order:
c(standard_normal, standard_exp, uniform_index(100), uniform_f64).
}
