% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rand_rrng_uniform}
\alias{rand_rrng_uniform}
\title{rand_rrng_uniform}
\usage{
rand_rrng_uniform(n)
}
\arguments{
\item{n}{Integer number of values to generate.}
}
\description{
Sample n uniform values in [0, 1) from R's RNG via \code{RngExt::random}.
}
