% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{range}
\alias{range}
\alias{range,fm-method}
\alias{range,fmV-method}
\title{Range of Values}
\usage{
\S4method{range}{fm}(x, ..., na.rm = FALSE)

\S4method{range}{fmV}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{a FlashMatrix vector or matrix.}

\item{...}{any FlashMatrix vectors or matrices.}

\item{na.rm}{logical, indicating if \code{NA} should be omitted.}
}
\description{
\code{range} returns a vector containing the minimum and maximum of all
the given arguments.
}

