% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rarray_matrix_column}
\alias{rarray_matrix_column}
\title{rarray_matrix_column}
\usage{
rarray_matrix_column(x, col)
}
\arguments{
\item{x}{A numeric matrix.}

\item{col}{1-based column index. Errors if \code{col} is not in \verb{1..=ncol}.}
}
\description{
Get a specific column from a numeric matrix as a Vec.
}
