% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rarray_matrix_len}
\alias{rarray_matrix_len}
\title{rarray_matrix_len}
\usage{
rarray_matrix_len(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
Get total length of a matrix.
}
