% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rarray_vector_sum}
\alias{rarray_vector_sum}
\title{rarray_vector_sum}
\usage{
rarray_vector_sum(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
Sum all elements of a numeric vector via RVector.
}
