% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{raw_ident_args}
\alias{raw_ident_args}
\title{raw_ident_args}
\usage{
raw_ident_args(where, type, ref = 1L)
}
\arguments{
\item{where}{Character filter.}

\item{type}{Integer type code.}

\item{ref}{Integer reference (defaults to \code{1L}).}
}
\description{
Free function whose parameters are all keywords.
}
