% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{raw_ident_choice}
\alias{raw_ident_choice}
\title{raw_ident_choice}
\usage{
raw_ident_choice(type = c("fast", "slow"))
}
\arguments{
\item{type}{One of \code{"fast"} or \code{"slow"}.}
}
\description{
\code{choices(...)} (\code{match.arg}) on a keyword parameter.
}
