% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{raw_ident_dots}
\alias{raw_ident_dots}
\title{raw_ident_dots}
\usage{
raw_ident_dots(type, ...)
}
\arguments{
\item{type}{Integer added to the number of extra arguments.}

\item{...}{Extra arguments; only counted.}
}
\description{
Keyword-named dots: R still sees plain \code{...}.
}
