% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rayon_in_thread}
\alias{rayon_in_thread}
\title{rayon_in_thread}
\usage{
rayon_in_thread()
}
\description{
Check if the current thread is a rayon worker (should be FALSE when called from R).
}
