% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rayon_par_map}
\alias{rayon_par_map}
\title{rayon_par_map}
\usage{
rayon_par_map(x)
}
\arguments{
\item{x}{Numeric vector input.}
}
\description{
Test par_map: parallel transform of an input slice into an R numeric vector.
}
