% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rayon_par_map2}
\alias{rayon_par_map2}
\title{rayon_par_map2}
\usage{
rayon_par_map2(a, b)
}
\arguments{
\item{a}{Numeric vector input.}

\item{b}{Numeric vector input.}
}
\description{
Test par_map2: element-wise parallel addition of two numeric vectors.
}
