% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rayon_parallel_filter_positive}
\alias{rayon_parallel_filter_positive}
\title{rayon_parallel_filter_positive}
\usage{
rayon_parallel_filter_positive(x)
}
\arguments{
\item{x}{Numeric vector input.}
}
\description{
Test parallel filter keeping only positive values.
}
