% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rayon_parallel_sqrt}
\alias{rayon_parallel_sqrt}
\title{rayon_parallel_sqrt}
\usage{
rayon_parallel_sqrt(x)
}
\arguments{
\item{x}{Numeric vector input.}
}
\description{
Test parallel map computing the square root of each element.
}
