% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rayon_parallel_sum}
\alias{rayon_parallel_sum}
\title{rayon_parallel_sum}
\usage{
rayon_parallel_sum(x)
}
\arguments{
\item{x}{Numeric vector to sum.}
}
\description{
Test parallel sum of a numeric vector using rayon.
}
