% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rayon_parallel_sum_int}
\alias{rayon_parallel_sum_int}
\title{rayon_parallel_sum_int}
\usage{
rayon_parallel_sum_int(x)
}
\arguments{
\item{x}{Integer vector to sum.}
}
\description{
Test parallel sum of an integer vector using rayon.
}
