% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rayon_vec_collect}
\alias{rayon_vec_collect}
\title{rayon_vec_collect}
\usage{
rayon_vec_collect(n)
}
\arguments{
\item{n}{Number of elements to generate.}
}
\description{
Test collecting square roots from a parallel range iterator into a Vec.
}
