% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rayon_with_r_matrix}
\alias{rayon_with_r_matrix}
\title{rayon_with_r_matrix}
\usage{
rayon_with_r_matrix(nrow, ncol)
}
\arguments{
\item{nrow}{Number of rows.}

\item{ncol}{Number of columns.}
}
\description{
Test with_r_matrix for parallel column-wise fill of an R matrix.
}
