% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rayon_with_r_vec}
\alias{rayon_with_r_vec}
\title{rayon_with_r_vec}
\usage{
rayon_with_r_vec(n)
}
\arguments{
\item{n}{Length of the output vector.}
}
\description{
Test with_r_vec for chunk-based parallel fill of an R numeric vector.
}
