% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rayon_with_r_vec_map}
\alias{rayon_with_r_vec_map}
\title{rayon_with_r_vec_map}
\usage{
rayon_with_r_vec_map(n)
}
\arguments{
\item{n}{Length of the output vector.}
}
\description{
Test with_r_vec_map for element-wise parallel fill of an R numeric vector.
}
