% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rbufmut_trait_clear}
\alias{rbufmut_trait_clear}
\title{rbufmut_trait_clear}
\usage{
rbufmut_trait_clear()
}
\description{
Exercise \code{RBufMut::len} / \code{is_empty} / \code{clear} through the trait.
}
