% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rdata_sidecar_env_new}
\alias{rdata_sidecar_env_new}
\title{rdata_sidecar_env_new}
\usage{
rdata_sidecar_env_new(count, score, flag, name)
}
\arguments{
\item{count}{Integer sidecar field.}

\item{score}{Numeric sidecar field.}

\item{flag}{Logical sidecar field.}

\item{name}{Character sidecar field.}
}
\description{
Test creating a SidecarEnv with all sidecar field types.
}
