% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rdata_sidecar_panicky_new}
\alias{rdata_sidecar_panicky_new}
\title{rdata_sidecar_panicky_new}
\usage{
rdata_sidecar_panicky_new()
}
\description{
Test creating a SidecarPanicky whose \code{boom} getter panics on access.
}
