% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rdata_sidecar_raw_new}
\alias{rdata_sidecar_raw_new}
\title{rdata_sidecar_raw_new}
\usage{
rdata_sidecar_raw_new(byte_val)
}
\arguments{
\item{byte_val}{Raw byte value for the sidecar field.}
}
\description{
Test creating a SidecarRaw with a u8 scalar sidecar field.
}
