% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rdata_sidecar_rawsexp_new}
\alias{rdata_sidecar_rawsexp_new}
\title{rdata_sidecar_rawsexp_new}
\usage{
rdata_sidecar_rawsexp_new()
}
\description{
Test creating a SidecarRawSexp with all SEXP slots initialized to NULL.
}
