% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rdata_sidecar_s3_new}
\alias{rdata_sidecar_s3_new}
\title{rdata_sidecar_s3_new}
\usage{
rdata_sidecar_s3_new(data)
}
\arguments{
\item{data}{Numeric sidecar field.}
}
\description{
Test creating a SidecarS3 with S3 $ method dispatch for sidecar fields.
}
