% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rdata_sidecar_s4_new}
\alias{rdata_sidecar_s4_new}
\title{rdata_sidecar_s4_new}
\usage{
rdata_sidecar_s4_new(slot_int, slot_real, slot_str)
}
\arguments{
\item{slot_int}{Integer sidecar field.}

\item{slot_real}{Numeric sidecar field.}

\item{slot_str}{Character sidecar field.}
}
\description{
Test creating a SidecarS4 with S4 slot accessors.
}
