% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rdata_sidecar_s7_new}
\alias{rdata_sidecar_s7_new}
\title{rdata_sidecar_s7_new}
\usage{
rdata_sidecar_s7_new(prop_int, prop_flag, prop_name)
}
\arguments{
\item{prop_int}{Integer sidecar field.}

\item{prop_flag}{Logical sidecar field.}

\item{prop_name}{Character sidecar field.}
}
\description{
Test creating a SidecarS7 with S7 property-based sidecar accessors.
}
