% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rdata_sidecar_vctrs_new}
\alias{rdata_sidecar_vctrs_new}
\title{rdata_sidecar_vctrs_new}
\usage{
rdata_sidecar_vctrs_new(vec_data, vec_label)
}
\arguments{
\item{vec_data}{Numeric vector sidecar field.}

\item{vec_label}{Character sidecar field.}
}
\description{
Test creating a SidecarVctrs with vctrs-compatible sidecar fields.
}
