% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{re_expand_roundtrip}
\alias{re_expand_roundtrip}
\title{re_expand_roundtrip}
\usage{
re_expand_roundtrip(df)
}
\arguments{
\item{df}{data.frame from \code{re_expand_align(...)}.}
}
\description{
Column-expansion round-trip. Columns: \verb{_type}, \code{id}, \code{c_1..c_2}, \code{s_1..s_3}.
}
