% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{re_loc_roundtrip}
\alias{re_loc_roundtrip}
\title{re_loc_roundtrip}
\usage{
re_loc_roundtrip(df)
}
\arguments{
\item{df}{data.frame from \code{re_loc_align(...)}.}
}
\description{
Struct-flatten variant field round-trip. Columns: \verb{_type}, \code{id}, \code{p_x}, \code{p_y}.
}
