% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{re_move_roundtrip}
\alias{re_move_roundtrip}
\title{re_move_roundtrip}
\usage{
re_move_roundtrip(df)
}
\arguments{
\item{df}{data.frame from \code{re_move_align(...)}.}
}
\description{
as_factor nested enum round-trip. Columns: \verb{_type}, \code{id}, \code{dir} (factor).
}
