% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{re_scalar_roundtrip_par}
\alias{re_scalar_roundtrip_par}
\title{re_scalar_roundtrip_par}
\usage{
re_scalar_roundtrip_par(df)
}
\arguments{
\item{df}{(no documentation available)}
}
\description{
Parallel REScalar round-trip. @export
}
