% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{re_scalar_roundtrip_zero}
\alias{re_scalar_roundtrip_zero}
\title{re_scalar_roundtrip_zero}
\usage{
re_scalar_roundtrip_zero(df)
}
\arguments{
\item{df}{(no documentation available)}
}
\description{
Zero-row REScalar round-trip (empty frame). @export
}
