% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_parquet}
\alias{read_parquet}
\title{Read a parquet file into a data.table}
\usage{
read_parquet(path)
}
\arguments{
\item{path}{Path to parquet file.}
}
\value{
A data.table.
}
\description{
Read a parquet file into a data.table
}
