% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in read_redatam.R
\name{read_redatam}
\alias{read_redatam}
\title{Read REDATAM dictionary}
\usage{
read_redatam(dictionary)
}

\arguments{
\item{dictionary}{A character string with the path to the dictionary file. It
 allows relative paths.}
}

\value{
A list of tables of data.frame type, each one representing a table with the census
 entities (or levels) and their attributes (or variables).
}

\description{
Read a DIC or DICX dictionary file directly into R.
}

\examples{
# Read a dictionary file (Uruguay 2011, aggregated)
zip <- system.file("extdata", "uru2011mini.zip", package = "redatam")
dout <- paste(tempdir(), "uru2011mini", sep = "/")
unzip(zip, exdir = dout)
read_redatam(paste(dout, "uru2011mini.dic", sep = "/"))
# or equivalently
read_redatam(paste(dout, "uru2011mini.dicx", sep = "/"))
}

