% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_template_file}
\alias{read_template_file}
\title{Read in the template file.}
\usage{
read_template_file(template_file)
}
\arguments{
\item{template_file}{Filename, character; if not provided, load the default from \code{extdata/}}
}
\value{
A list with sheets of the template info file.
}
\description{
Read in the template file.
}
\keyword{internal}
