% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread.R
\name{read_yaml_header}
\alias{read_yaml_header}
\title{Read YAML header from a Markdown file}
\usage{
read_yaml_header(file, n_max = 20)
}
\arguments{
\item{file}{Path to the Markdown file}
}
\value{
List containing the YAML metadata
}
\description{
Read YAML header from a Markdown file
}
