% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_YujieHe2016.R
\name{read_YujieHe2016}
\alias{read_YujieHe2016}
\title{Read He 2016}
\usage{
read_YujieHe2016(Yujie_file = NULL)
}
\arguments{
\item{Yujie_file}{The raw csv data}
}
\value{
ISRaD compliant file structure with only columns that overlap with original data
}
\description{
Read in the data from Yujie He's 2016 Science paper as a raw csv file
}
