% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEntry.R
\name{readEntry}
\alias{readEntry}
\title{Read single entry of the aridec database}
\usage{
readEntry(path, entryName)
}
\arguments{
\item{path}{character string with the path where aridec is stored}

\item{entryName}{character string with the name of the entry in the database}
}
\value{
R list with the entry
}
\description{
Read single entry of the aridec database
}
\examples{
Adair2017=readEntry(path="~/aridec/data/", entryName="Adair2017")
}
