% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{reader_auto_box_roundtrip}
\alias{reader_auto_box_roundtrip}
\title{reader_auto_box_roundtrip}
\usage{
reader_auto_box_roundtrip(df)
}
\arguments{
\item{df}{data.frame with \code{tag} and \verb{xs_*} columns.}
}
\description{
Auto-expand \verb{Box<[i32]>} round-trip. Columns \code{tag}, \code{xs_1..xs_k}.
}
