% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{reader_auto_roundtrip}
\alias{reader_auto_roundtrip}
\title{reader_auto_roundtrip}
\usage{
reader_auto_roundtrip(df)
}
\arguments{
\item{df}{data.frame with \code{name} and \verb{values_*} columns.}
}
\description{
Auto-expand \verb{Vec<f64>} round-trip. Columns \code{name}, \code{values_1..values_k}.
}
