% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFB-ReadFcn.R
\name{readpfb}
\alias{readpfb}
\title{Read PFB files
Function to Read ParFlow Binary Files into R}
\usage{
readpfb(filename, verbose)
}
\arguments{
\item{filename}{Name of file to read}

\item{verbose}{Optional to turn on print statement}
}
\description{
Read PFB files
Function to Read ParFlow Binary Files into R
}
