% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch.R
\name{reconnect}
\alias{reconnect}
\title{Reconnect after calling \code{\link{goodbye}}}
\usage{
reconnect(connection, password)
}
\arguments{
\item{connection}{and Rsch_connection object}

\item{password}{the password associated with the account in the Rsch_connection object}
}
\value{
an Rsch_connection object
 @examples
\dontrun{
connection <- get_connection(host = "cuckoo.lbnl.gov", user = "cliff", password = "TopSecret")
execute(connection,"grep \"nuclear\" /*")
results <- printout(connection)
goodbye(connection)

connection <- reconnect(connection, "TopSecret")
execute(connection,"grep \"SDI\" /*")

}
}
\description{
Reconnect after calling \code{\link{goodbye}}
}
