% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in redatam-package.R
\name{redatam-package}
\alias{redatam-package}
\title{Import 'REDATAM' Files}
\description{
Import 'REDATAM' formats into R via the 'Open REDATAM' C++ library.
 The full context of this project and details about the implementation are
 available in <doi:10.1017/dap.2025.4> (Open Access).
}

\keyword{internal}
