% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{regex_find}
\alias{regex_find}
\title{regex_find}
\usage{
regex_find(pattern, text)
}
\arguments{
\item{pattern}{Regular expression pattern.}

\item{text}{String to search within.}
}
\description{
Test finding the first regex match in the text.
}
