% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{regex_find_all}
\alias{regex_find_all}
\title{regex_find_all}
\usage{
regex_find_all(pattern, text)
}
\arguments{
\item{pattern}{Regular expression pattern.}

\item{text}{String to search within.}
}
\description{
Test finding all regex matches in the text.
}
