% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{regex_replace_all}
\alias{regex_replace_all}
\title{regex_replace_all}
\usage{
regex_replace_all(pattern, text, replacement)
}
\arguments{
\item{pattern}{Regular expression pattern.}

\item{text}{String to search within.}

\item{replacement}{Replacement string.}
}
\description{
Test replacing all regex matches with a replacement string.
}
