% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{regex_replace_first}
\alias{regex_replace_first}
\title{regex_replace_first}
\usage{
regex_replace_first(pattern, text, replacement)
}
\arguments{
\item{pattern}{Regular expression pattern.}

\item{text}{String to search within.}

\item{replacement}{Replacement string.}
}
\description{
Test replacing the first regex match with a replacement string.
}
