% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational.R
\name{rel_alias}
\alias{rel_alias}
\title{Get the internal alias for a DuckDB relation object}
\usage{
rel_alias(rel)
}
\arguments{
\item{rel}{the DuckDB relation object}
}
\description{
Get the internal alias for a DuckDB relation object
}
\examples{
con <- DBI::dbConnect(duckdb())
rel <- rel_from_df(con, mtcars)
rel_alias(rel)
}
