% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational.R
\name{rel_filter}
\alias{rel_filter}
\title{Lazily filter a DuckDB relation object}
\usage{
rel_filter(rel, expr)
}
\arguments{
\item{rel}{the DuckDB relation object}

\item{expr}{a filter condition expression}
}
\value{
the now filtered \code{duckdb_relation} object
}
\description{
Lazily filter a DuckDB relation object
}
\examples{
con <- DBI::dbConnect(duckdb::duckdb())
rel <- rel_from_df(con, mtcars)
rel2 <- rel_filter(rel, expr_function(">", list(expr_reference("cyl"), expr_constant("6"))))
}
