% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational.R
\name{rel_limit}
\alias{rel_limit}
\title{Lazily retrieve the top-n rows of a DuckDB relation object}
\usage{
rel_limit(rel, n)
}
\arguments{
\item{rel}{the DuckDB relation object}

\item{n}{the amount of rows to retrieve}
}
\value{
the now limited \code{duckdb_relation} object
}
\description{
NOTICE: This function is experimental and may change/disappear without warning in future versions
}
\examples{
con <- DBI::dbConnect(duckdb())
rel <- rel_from_df(con, mtcars)
rel2 <- rel_limit(rel, 10)
}
