% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational.R
\name{rel_order}
\alias{rel_order}
\title{Lazily reorder a DuckDB relation object}
\usage{
rel_order(rel, orders)
}
\arguments{
\item{rel}{the DuckDB relation object}

\item{orders}{a list of DuckDB expressions to order by}
}
\value{
the now aggregated \code{duckdb_relation} object
}
\description{
NOTICE: This function is experimental and may change/disappear without warning in future versions
}
\examples{
con <- DBI::dbConnect(duckdb())
rel <- rel_from_df(con, mtcars)
rel2 <- rel_order(rel, list(expr_reference("hp")))
}
