% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational.R
\name{rel_set_alias}
\alias{rel_set_alias}
\title{Set the internal alias for a DuckDB relation object}
\usage{
rel_set_alias(rel, alias)
}
\arguments{
\item{rel}{the DuckDB relation object}

\item{alias}{the new alias}
}
\description{
Set the internal alias for a DuckDB relation object
}
\examples{
con <- DBI::dbConnect(duckdb())
rel <- rel_from_df(con, mtcars)
rel_set_alias(rel, "my_new_alias")
}
