% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational.R
\name{rel_sql}
\alias{rel_sql}
\title{Run a SQL query on a DuckDB relation object}
\usage{
rel_sql(rel, sql)
}
\arguments{
\item{rel}{the DuckDB relation object}

\item{sql}{a SQL query to run, use \verb{_} to refer back to the relation}
}
\value{
the now aggregated \code{duckdb_relation} object
}
\description{
Run a SQL query on a DuckDB relation object
}
\examples{
con <- DBI::dbConnect(duckdb())
rel <- rel_from_df(con, mtcars)
rel2 <- rel_sql(rel, "SELECT hp, cyl FROM _ WHERE hp > 100")
}
