% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational.R
\name{rel_to_altrep}
\alias{rel_to_altrep}
\title{Transforms a relation object to a lazy data frame using altrep}
\usage{
rel_to_altrep(rel)
}
\arguments{
\item{rel}{the DuckDB relation object}
}
\value{
a data frame
}
\description{
NOTICE: This function is experimental and may change/disappear without warning in future versions
}
\examples{
con <- DBI::dbConnect(duckdb())
rel <- rel_from_df(con, mtcars)
print(rel_to_altrep(rel))
}
