% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational.R
\name{rel_union_all}
\alias{rel_union_all}
\title{UNION ALL on two DuckDB relation objects}
\usage{
rel_union_all(rel_a, rel_b)
}
\arguments{
\item{rel_a}{a DuckDB relation object}

\item{rel_b}{a DuckDB relation object
rel_a <- rel_from_df(con, mtcars)
rel_b <- rel_from_df(con, mtcars)
rel_union_all(rel_a, rel_b)}
}
\description{
UNION ALL on two DuckDB relation objects
}
