% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{remove_draft_from_file}
\alias{remove_draft_from_file}
\title{Remove a draft section from end of file}
\usage{
remove_draft_from_file(file_path, section)
}
\arguments{
\item{file_path}{Path to the file}

\item{section}{Section info from find_draft_comments}
}
\description{
Remove a draft section from end of file
}
