% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{repeating_raw}
\alias{repeating_raw}
\title{repeating raw}
\source{
Generated by miniextendr from Rust fn \code{repeating_raw}
}
\usage{
repeating_raw(pattern, n)
}
\arguments{
\item{pattern}{(no documentation available)}

\item{n}{(no documentation available)}
}
\description{
repeating raw
}
