% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{reports}
\alias{reports}
\title{reports}
\usage{
reports(database = NULL, report = "count_data")
}
\arguments{
\item{database}{ISRaD data object}

\item{report}{Parameter to define which type of report you want. The default is "count_data" other options include "entry_stats" and "site_map".}
}
\description{
generate reports of ISRaD data
}
