% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rerror_parse}
\alias{rerror_parse}
\title{rerror_parse}
\usage{
rerror_parse(s)
}
\arguments{
\item{s}{(no documentation available)}
}
\description{
\verb{Result<i32, RError>} built from a \code{ParseIntError} via \code{From}, then classed.
}
