% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{result_null_on_err}
\alias{result_null_on_err}
\title{result_null_on_err}
\usage{
result_null_on_err(x)
}
\arguments{
\item{x}{Integer input (negative triggers Err).}
}
\description{
Test that Result Err(()) returns NULL in R, Ok returns doubled value.
}
