% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorityFlow_data.R
\docType{data}
\name{river.mask}
\alias{river.mask}
\title{River Mask}
\format{
a matrix of 0's and 1' showing the location of the river cells (1=river, 0=nonriver), where nrow=nx
and ncol=ny for the domain.
}
\usage{
river.mask
}
\description{
A mask showing and example river network for  the test domain
}
\keyword{datasets}
