% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/River_Smoothing.R
\name{RiverSmooth}
\alias{RiverSmooth}
\title{Apply smoothing to a DEM along a pre-defined stream network}
\usage{
RiverSmooth(dem, direction, mask, river.summary, river.segments,
  epsilon = 0.01, d4 = c(1, 2, 3, 4))
}
\arguments{
\item{dem}{Elevation matrix}

\item{direction}{a matrix of the flow directions for cells that have been processed - defaults to all zeros}

\item{mask}{Mask with ones for cells to be processed and zeros for everything else - defaults to a mask of all 1's}

\item{river.summary}{A table summarizing the stream segments in a domain with the following. 
it should have 1 row per stream segmentand the following 7 columns: (1) subbasin number, (2) x and (3) y index
of the upstream end of the stream segment, (4) x and (5) y index of the downstream eand of the river segement,
(6) the subbasin number for the downstream basin, -1 indicates a subbasin draining out of the domain
(7) drainage area of the subbasin}

\item{river.segments}{A nx by ny matrix indicating the subbasin number for with for all grid cells on the river network 
( (all cells not on the river network shoudl be 0)}

\item{epsilon}{the minimum elevation difference between cells walking upstream from the river network.}

\item{d4}{directional numbering system: the numbers you want to assigne to down, left, top,right (defaults to 1,2,3,4)}
}
\description{
This function will smooth a DEM along a stream network. It requires pre-defined stream segments 
and subbasins which can be obtained using the \code{\link{CalcSubbasins}} function.
}
