% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/River_Slope.R
\name{RivSlope}
\alias{RivSlope}
\title{Apply minimum slope and secondary scaling to river cells}
\usage{
RivSlope(direction, slopex, slopey, minslope, RiverMask, Remove.Sec = FALSE)
}
\arguments{
\item{direction}{Nx by ny matrix of flow directions following the following convention (1=down, 2=left, 3=up, 4=right)}

\item{slopex}{Nx by ny matrix of slopes in the x direction (Note: these should be face centered slopes as woudl be calcualted with SlopeCalcStan)}

\item{slopey}{Nx by ny matrix of slopes in the y direction (Note: these should be face centered slopes as woudl be calcualted with SlopeCalcStan)}

\item{minslope}{Threshold for slope adjustment. Any primary direciton slope for a river cell will be adjusted such that abs(slope)>=minslope}

\item{RiverMask}{Nx by Ny matrix indicating the mask of river cells that the min slope will be applied to.}

\item{Remove.Sec}{Flag, if set to True any secondary outlfows on river cells will be set to zero. Defaults to False.}
}
\value{
This will generate the following outputs:
   - slopex - The adjusted nx by ny matrix of slopex values
   - slopey - The adjusted nx by ny matrix of slopex values
   - adj_mask - Nx by ny matrix indicating cells that were ajusted (1=adjusted, 0=not adjsuted). This will be all the cells where inslope/outslop<adj_th
   - SlopeOutlet - Nx by ny matrix of the outlet slope for every grid cell 
   - SlopeOutletNew - Nx by ny matrix of the outlet slope for every grid cell after processing
}
\description{
A funtion that will apply a minimum slope threshold for the primary flow direciton along river cells. This function can also limit secondary slopes out of river cells to zero.
}
