% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlmtest.R
\name{rlmtest}
\alias{rlmtest}
\title{Title}
\usage{
rlmtest(
  y,
  x = NA,
  a.order = 1,
  sig.level = 0.05,
  crit.type = c("asymptotic", "nonparametric"),
  nonp.method = c("bootstrap", "permutation"),
  num.resamp = 1000,
  num.discard = 0,
  CPP = T
)
}
\arguments{
\item{CPP}{}
}
\value{

}
\description{
Title
}
