% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rng_chi_sq_approx}
\alias{rng_chi_sq_approx}
\title{rng_chi_sq_approx}
\usage{
rng_chi_sq_approx(n, df)
}
\arguments{
\item{n}{Number of values.}

\item{df}{Degrees of freedom (integer, used as count of squared normals).}
}
\description{
Generate n chi-squared-like values using the sum-of-normals method.
}
