% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rng_exponential}
\alias{rng_exponential}
\title{rng_exponential}
\usage{
rng_exponential(n)
}
\arguments{
\item{n}{Integer number of values to generate.}
}
\description{
Test generating n exponential random numbers using R's RNG.
}
