% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rng_guard_test}
\alias{rng_guard_test}
\title{rng_guard_test}
\usage{
rng_guard_test(n)
}
\arguments{
\item{n}{Integer number of uniform values to generate.}
}
\description{
Test manual RNG management using RngGuard RAII pattern.
}
