% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rng_int}
\alias{rng_int}
\title{rng_int}
\usage{
rng_int(n, max)
}
\arguments{
\item{n}{Integer number of values to generate.}

\item{max}{Numeric upper bound (exclusive) for the random integers.}
}
\description{
Test generating n random integers in [0, max) using R's RNG.
}
