% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rng_normal}
\alias{rng_normal}
\title{rng_normal}
\usage{
rng_normal(n)
}
\arguments{
\item{n}{Integer number of values to generate.}
}
\description{
Test generating n standard normal random numbers using R's RNG.
}
