% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rng_range}
\alias{rng_range}
\title{rng_range}
\usage{
rng_range(min, max)
}
\arguments{
\item{min}{Lower bound.}

\item{max}{Upper bound.}
}
\description{
Generate a random double in [min, max) using R's RNG.
}
