% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rng_uniform}
\alias{rng_uniform}
\title{rng_uniform}
\usage{
rng_uniform(n)
}
\arguments{
\item{n}{Integer number of values to generate.}
}
\description{
Test generating n uniform random numbers using R's RNG.
}
