% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rng_with_interrupt}
\alias{rng_with_interrupt}
\title{rng_with_interrupt}
\usage{
rng_with_interrupt(n)
}
\arguments{
\item{n}{Integer number of uniform values to generate.}
}
\description{
Test RNG with interrupt checking enabled.
}
