% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rng_with_rng_test}
\alias{rng_with_rng_test}
\title{rng_with_rng_test}
\usage{
rng_with_rng_test(n)
}
\arguments{
\item{n}{Integer number of uniform values to generate.}
}
\description{
Test manual RNG management using the with_rng closure API.
}
