% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rng_worker_uniform}
\alias{rng_worker_uniform}
\title{rng_worker_uniform}
\usage{
rng_worker_uniform(n)
}
\arguments{
\item{n}{Integer number of uniform values to generate.}
}
\description{
Test RNG on worker thread (requires worker-thread feature).
}
