% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{rok_available}
\alias{rok_available}
\title{Whether the rok binary is ready}
\usage{
rok_available()
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Checks that a rok binary with the same version as this package can be found (see
\code{\link[=rok_binary]{rok_binary()}}). If not, \code{\link[=setup]{setup()}} installs one.
}
\examples{
rok_available()
}
