% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{round}
\alias{ceiling,fm-method}
\alias{ceiling,fmV-method}
\alias{floor,fm-method}
\alias{floor,fmV-method}
\alias{round}
\alias{round,fm-method}
\alias{round,fmV-method}
\title{Rounding of Numbers}
\usage{
\S4method{ceiling}{fm}(x)

\S4method{ceiling}{fmV}(x)

\S4method{floor}{fm}(x)

\S4method{floor}{fmV}(x)

\S4method{round}{fm}(x)

\S4method{round}{fmV}(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\description{
\code{ceiling} takes a single numeric argument \code{x} and returns
a numeric vector containing the smallest integers not less than
the corresponding elements of \code{x}.
}
\details{
\code{floor} takes a single numeric argument \code{x} and returns a numeric
vector containing the largest integers not greater than the corresponding
elements of \code{x}.

\code{round} rounds the values in its first argument to the specified number
of decimal places (default 0).
}
\examples{
mat <- ceiling(fm.runif.matrix(100, 10))
mat <- floor(fm.runif.matrix(100, 10))
mat <- round(fm.runif.matrix(100, 10))
}

