% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{routes_given}
\alias{routes_given}
\title{The \code{routes_given()} generic}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{routes_given()} is an S7 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{EitherRoutesS7}}
}
}
