% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpkg-package.R
\docType{package}
\name{rpkg}
\alias{rpkg}
\title{rpkg: an R package for "Flexibility of species interactions and ecological stability"
This package includes functions for causality test and Jacobian-estimation}
