% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{rpkg_altrep_examples}
\alias{rpkg_altrep_examples}
\alias{unit_circle}
\alias{lazy_int_seq}
\alias{boxed_ints}
\alias{static_ints}
\alias{leaked_ints}
\alias{static_strings}
\title{ALTREP Example Constructors}
\source{
Generated by miniextendr from Rust fn \code{unit_circle}
}
\usage{
unit_circle(n)
}
\arguments{
\item{n}{Number of roots of unity (complex numbers on the unit circle).}
}
\value{
An ALTREP vector.
}
\description{
ALTREP Example Constructors
}
\examples{
unit_circle(8L)
lazy <- lazy_int_seq(1L, 5L, 1L)
lazy[1:3]
boxed_ints(3L)
static_strings()
}
