% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{rpkg_altrep_helpers}
\alias{rpkg_altrep_helpers}
\alias{altrep_compact_int}
\alias{altrep_from_doubles}
\alias{altrep_from_strings}
\alias{altrep_from_logicals}
\alias{altrep_from_raw}
\alias{altrep_from_list}
\alias{altrep_constant_int}
\alias{altrep_lazy_int_seq_is_materialized}
\title{ALTREP Helpers}
\source{
Generated by miniextendr from Rust fn \code{altrep_compact_int}

Generated by miniextendr from Rust fn \code{altrep_from_doubles}

Generated by miniextendr from Rust fn \code{altrep_from_strings}

Generated by miniextendr from Rust fn \code{altrep_from_logicals}

Generated by miniextendr from Rust fn \code{altrep_from_raw}

Generated by miniextendr from Rust fn \code{altrep_from_list}

Generated by miniextendr from Rust fn \code{altrep_constant_int}

Generated by miniextendr from Rust fn \code{altrep_lazy_int_seq_is_materialized}
}
\usage{
altrep_compact_int(n, start, step)

altrep_from_doubles(x)

altrep_from_strings(x)

altrep_from_logicals(x)

altrep_from_raw(x)

altrep_from_list(x)

altrep_constant_int()

altrep_lazy_int_seq_is_materialized(x)
}
\description{
ALTREP convenience wrappers (internal)
}
\examples{
\dontrun{
x <- altrep_compact_int(5L, 1L, 2L)
y <- altrep_from_doubles(c(1, 2, 3))
z <- altrep_from_strings(c("a", "b"))
altrep_lazy_int_seq_is_materialized(lazy_int_seq(1L, 5L, 1L))
}
}
\keyword{internal}
