% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rpkg_arithmetic}
\alias{rpkg_arithmetic}
\alias{add}
\title{Arithmetic Tests}
\usage{
add(left, right)
}
\arguments{
\item{left}{Integer input.}

\item{right}{Integer input.}
}
\value{
A scalar integer result.
}
\description{
Arithmetic and return-value tests
}
\examples{
add(1L, 2L)
}
