% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{rpkg_conversion_tests}
\alias{rpkg_conversion_tests}
\alias{test_i32_identity}
\alias{test_i32_add_one}
\alias{test_i32_sum}
\alias{test_f64_identity}
\alias{test_f64_add_one}
\title{Conversion Tests}
\source{
Generated by miniextendr from Rust fn \code{test_i32_identity}
}
\usage{
test_i32_identity(x)
}
\description{
Scalar and slice conversion tests
}
\examples{
test_i32_add_one(1L)
test_f64_multiply(2, 3)
test_u8_identity(as.raw(255))
test_i32_slice_sum(1:5)
test_f64_slice_mean(c(1, 2, 3))
}
\keyword{internal}
