% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rpkg_enabled_features}
\alias{rpkg_enabled_features}
\title{Returns a vector of enabled feature names for this build}
\source{
Generated by miniextendr from Rust fn \code{rpkg_enabled_features}
}
\usage{
rpkg_enabled_features()
}
\value{
A character vector of enabled feature names.
}
\description{
Returns a vector of enabled feature names for this build.
}
\examples{
rpkg_enabled_features()
}
