% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{rpkg_r6_standalone}
\alias{rpkg_r6_standalone}
\alias{r6_standalone_add}
\title{R6 Standalone Function}
\source{
Generated by miniextendr from Rust fn \code{r6_standalone_add}
}
\usage{
r6_standalone_add(a, b)
}
\description{
Standalone helper in the R6 test module.
}
\examples{
r6_standalone_add(1L, 2L)
}
\keyword{internal}
