% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{rpkg_thread_builder}
\alias{rpkg_thread_builder}
\alias{unsafe_C_test_r_thread_builder}
\alias{unsafe_C_test_r_thread_builder_spawn_join}
\title{Thread Builder Tests}
\source{
Generated by miniextendr from Rust fn \code{C_test_r_thread_builder}
}
\usage{
unsafe_C_test_r_thread_builder()
}
\description{
Thread Builder Tests
}
\examples{
\dontrun{
unsafe_C_test_r_thread_builder()
unsafe_C_test_r_thread_builder_spawn_join()
unsafe_C_test_spawn_with_r_lean_stack()
unsafe_C_test_stack_check_guard_lean()
}
}
\keyword{internal}
