% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{rpkg_unwind_protect}
\alias{rpkg_unwind_protect}
\alias{unsafe_C_unwind_protect_normal}
\alias{unsafe_C_unwind_protect_r_error}
\title{Unwind Protect Tests}
\source{
Generated by miniextendr from Rust fn \code{C_unwind_protect_normal}
}
\usage{
unsafe_C_unwind_protect_normal()
}
\description{
Unwind Protect Tests
}
\examples{
\dontrun{
unsafe_C_unwind_protect_normal()
unsafe_C_unwind_protect_r_error()
unsafe_C_unwind_protect_lowlevel_test()
}
}
\keyword{internal}
