% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{rpkg_worker_tests}
\alias{rpkg_worker_tests}
\alias{test_worker_return_i32}
\alias{test_worker_return_string}
\alias{test_worker_return_f64}
\title{Worker Thread Tests}
\source{
Generated by miniextendr from Rust fn \code{test_worker_return_i32}
}
\usage{
test_worker_return_i32()
}
\description{
Worker Thread Tests
}
\examples{
test_worker_return_i32()
test_worker_return_string()
test_worker_return_f64()
try(test_main_thread_r_error())
\dontrun{
unsafe_C_test_worker_simple()
}
}
\keyword{internal}
