% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber-run_api.R
\name{run_api}
\alias{run_api}
\title{Run the Plumber API}
\usage{
run_api(.host = "0.0.0.0", .port = 8080, ...)
}
\arguments{
\item{.host}{Host - defaults to "0.0.0.0" (localhost)}

\item{.port}{Post - defaults to 8080}

\item{...}{
  Arguments passed on to \code{\link[plumber:pr_run]{plumber::pr_run}}
  \describe{
    \item{\code{pr}}{A Plumber API. Note: The supplied Plumber API object will also be updated in place as well as returned by the function.}
    \item{\code{host}}{A string that is a valid IPv4 or IPv6 address that is owned by
this server, which the application will listen on. "0.0.0.0" represents
all IPv4 addresses and "::/0" represents all IPv6 addresses.}
    \item{\code{port}}{A number or integer that indicates the server port that should
be listened on. Note that on most Unix-like systems including Linux and
Mac OS X, port numbers smaller than 1025 require root privileges.}
    \item{\code{debug}}{If \code{TRUE}, it will provide more insight into your API errors.
Using this value will only last for the duration of the run.
If \code{\link[plumber:pr_set_debug]{pr_set_debug()}} has not been called, \code{debug} will default to \code{interactive()} at \code{\link[plumber:pr_run]{pr_run()}} time}
    \item{\code{docs}}{Visual documentation value to use while running the API.
This value will only be used while running the router.
If missing, defaults to information previously set with \code{\link[plumber:pr_set_docs]{pr_set_docs()}}.
For more customization, see \code{\link[plumber:pr_set_docs]{pr_set_docs()}} for examples.}
    \item{\code{swaggerCallback}}{An optional single-argument function that is called
back with the URL to an OpenAPI user interface when one becomes ready. If
missing, defaults to information set with \code{\link[plumber:pr_set_docs_callback]{pr_set_docs_callback()}}.
This value will only be used while running the router.}
    \item{\code{quiet}}{If \code{TRUE}, don't print routine startup messages.}
  }}
}
\value{
Plumber API
}
\description{
Run the Plumber API
}
