% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_grase.R
\name{run_grase}
\alias{run_grase}
\title{Run grase API}
\usage{
run_grase(task, gene, indir, outdir, tx_ids = NULL)
}
\arguments{
\item{task}{Task to perform ("plot_graphs" or "find_focal_exons")}

\item{gene}{Gene ID}

\item{indir}{Input directory containing GTF and GraphML}

\item{outdir}{Output directory}

\item{tx_ids}{Vector of transcript IDs (required for "find_focal_exons")}
}
\description{
Run grase API
}
