% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exontest_functions.R
\name{run_one_comparison}
\alias{run_one_comparison}
\title{Run a test function over events in chunked parallel batches with checkpointing}
\usage{
run_one_comparison(
  sc,
  test_fn,
  err_log,
  ...,
  chunk_size = 10000L,
  checkpoint_prefix = NULL,
  mc_cores = 1L
)
}
\arguments{
\item{sc}{split counts data frame}

\item{test_fn}{function applied per event}

\item{err_log}{path to error log file}

\item{chunk_size}{number of events per parallel batch}

\item{checkpoint_prefix}{prefix for checkpoint RDS files; NULL disables checkpointing}

\item{mc_cores}{number of parallel cores}
}
\description{
Run a test function over events in chunked parallel batches with checkpointing
}
