% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial-helpers.R
\name{run_tail_length_experiment}
\alias{run_tail_length_experiment}
\title{Simulation of data measured for "mouse tail length experiment"}
\usage{
run_tail_length_experiment(num_mice = 3, actual_increase = 1,
  variance = 1)
}
\arguments{
\item{actual_increase}{The actual increase in length (cm) the supplement
has on the tail length of mice. The default value (1) indicates
**an actuall** increase of 1cm in tail length of the treatment mice.}

\item{mice_per_group}{The number of mice used per group}
}
\value{
a data.frame of tail length measurements for the mice in the
  experiment.
}
\description{
This function simulates measurements of tail lengths from two different
groups of mice. Mice in the `"control"` group were given a normal diet, and
mice in the `"treatment"` group were given a diet with a supplement that
claims to increase tail length.

Explore how the number of mice measured per group effects the conclusion from
the experiment by playing with (1) the number of mice used per experimental
group; and (2) the *actual* difference tail length in the treatment group.

This function assumes that the average tail length of a mouse is 8cm.
}
