% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{s3_value}
\alias{s3_value}
\title{S3 generic for \code{s3_value}}
\source{
Generated by miniextendr from \code{S3Counter::s3_value}
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments passed to methods}
}
\description{
S3 generic for \code{s3_value}
}
