% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{s4_class_name_test}
\alias{s4_class_name_test}
\title{s4_class_name_test}
\usage{
s4_class_name_test(obj)
}
\arguments{
\item{obj}{(no documentation available)}
}
\description{
Get the class name of an S4 object.
}
