% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{s4_get_value}
\alias{s4_get_value}
\title{The \code{s4_get_value()} generic}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{s4_get_value()} is an S4 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{CounterTraitS4}}
\item \code{\link{S4TraitCounter}}
}
}
