% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{s4_is_s4}
\alias{s4_is_s4}
\alias{s4_has_slot_test}
\alias{s4_class_name_test}
\title{s4_is_s4}
\usage{
s4_is_s4(obj)

s4_has_slot_test(obj, slot_name)

s4_class_name_test(obj)
}
\arguments{
\item{obj}{(no documentation available)}

\item{slot_name}{(no documentation available)}
}
\description{
Check if an object is S4.

Check if an S4 object has a named slot.

Get the class name of an S4 object.
}
