% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{s4_is_s4}
\alias{s4_is_s4}
\title{Check if an object is S4}
\source{
Generated by miniextendr from Rust fn \code{s4_is_s4}
}
\usage{
s4_is_s4(obj)
}
\arguments{
\item{obj}{(no documentation available)}
}
\description{
Check if an object is S4
}
