% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/miniextendr-wrappers.R
\name{s4_mode_current}
\alias{s4_mode_current}
\title{Get current mode (S4)}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{s4_mode_current()} is an S4 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{S4MatchArgHolder}}
}
}
