% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{s4_optional_integer}
\alias{s4_optional_integer}
\title{The \code{s4_optional_integer()} generic}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{s4_optional_integer()} is an S4 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{ScalarOptionS4}}
}
}
