% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{s4_raise_warning}
\alias{s4_raise_warning}
\title{The \code{s4_raise_warning()} generic}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{s4_raise_warning()} is an S4 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{S4Raiser}}
}
}
