% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{class}
\name{S4MatchArgHolder}
\alias{S4MatchArgHolder}
\alias{S4MatchArgHolder-s4_mode_current}
\alias{s4_mode_current,S4MatchArgHolder-method}
\alias{S4MatchArgHolder-s4_mode_set}
\alias{s4_mode_set,S4MatchArgHolder-method}
\title{S4MatchArgHolder S4 Class}
\usage{
S4MatchArgHolder(mode = c("Fast", "Safe", "Debug"))
}
\arguments{
\item{mode}{One of "Fast", "Safe", "Debug".}
}
\description{
S4MatchArgHolder S4 Class
}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to Rust \code{S4MatchArgHolder} struct}
}}

