% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{class}
\name{S4Raiser}
\alias{S4Raiser}
\alias{s4_raise_error,S4Raiser-method}
\alias{s4_raise_error_classed,S4Raiser-method}
\alias{s4_raise_warning,S4Raiser-method}
\alias{s4_raise_warning_classed,S4Raiser-method}
\alias{s4_raise_message,S4Raiser-method}
\alias{s4_raise_condition,S4Raiser-method}
\alias{s4_raise_condition_classed,S4Raiser-method}
\alias{s4_id,S4Raiser-method}
\alias{S4Raiser-s4_id}
\alias{S4Raiser-s4_raise_error}
\alias{S4Raiser-s4_raise_error_classed}
\alias{S4Raiser-s4_raise_warning}
\alias{S4Raiser-s4_raise_warning_classed}
\alias{S4Raiser-s4_raise_message}
\alias{S4Raiser-s4_raise_condition}
\alias{S4Raiser-s4_raise_condition_classed}
\title{S4Raiser S4 Class}
\usage{
S4Raiser(id)
}
\arguments{
\item{id}{(undocumented)}
}
\description{
S4Raiser S4 Class
}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to Rust \code{S4Raiser} struct}
}}

\keyword{internal}
