% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{class}
\name{S4TraitCounter}
\alias{S4TraitCounter}
\alias{S4TraitCounter-s4_get_value}
\alias{s4_get_value,S4TraitCounter-method}
\alias{S4TraitCounter_new_s4trait}
\alias{s4_trait_Counter_value}
\alias{s4_trait_Counter_value,S4TraitCounter-method}
\alias{s4_trait_Counter_increment}
\alias{s4_trait_Counter_increment,S4TraitCounter-method}
\alias{s4_trait_Counter_checked_add}
\alias{s4_trait_Counter_checked_add,S4TraitCounter-method}
\alias{S4TraitCounter_Counter_default_initial}
\alias{S4TraitCounter_Counter_MAX_VALUE}
\title{S4TraitCounter S4 Class}
\usage{
S4TraitCounter_new_s4trait(initial)

\S4method{s4_trait_Counter_value}{S4TraitCounter}(x, ...)

\S4method{s4_trait_Counter_increment}{S4TraitCounter}(x, ...)

\S4method{s4_trait_Counter_checked_add}{S4TraitCounter}(x, n, ...)

S4TraitCounter_Counter_default_initial()

S4TraitCounter_Counter_MAX_VALUE()
}
\arguments{
\item{initial}{(undocumented)}

\item{x}{A \code{S4TraitCounter} object}

\item{...}{Additional arguments passed to methods}

\item{n}{Amount to add.}
}
\description{
S4TraitCounter inherent methods (constructor and getters).

Get value via inherent method.

Create a new S4 trait counter.

S4 generic for trait method \code{Counter::value}

S4 generic for trait method \code{Counter::increment}

S4 generic for trait method \code{Counter::checked_add}
}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to Rust \code{S4TraitCounter} struct}
}}

\keyword{internal}
