% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ScalarOptionEnv}
\alias{ScalarOptionEnv}
\alias{ScalarOptionEnv$new}
\alias{ScalarOptionEnv$optional_integer}
\alias{ScalarOptionEnv$optional_real}
\alias{ScalarOptionEnv$optional_logical}
\alias{ScalarOptionEnv$optional_string}
\alias{ScalarOptionEnv$optional_str}
\alias{ScalarOptionEnv$optional_unit}
\alias{ScalarOptionEnv$optional_i8}
\alias{ScalarOptionEnv$optional_i16}
\alias{ScalarOptionEnv$optional_u16}
\alias{ScalarOptionEnv$optional_u32}
\alias{ScalarOptionEnv$optional_f32}
\alias{ScalarOptionEnv$optional_i64}
\alias{ScalarOptionEnv$optional_u64}
\alias{ScalarOptionEnv$optional_isize}
\alias{ScalarOptionEnv$optional_usize}
\alias{ScalarOptionEnv$optional_rboolean}
\alias{ScalarOptionEnv$optional_rlogical}
\alias{ScalarOptionEnv$optional_complex}
\alias{ScalarOptionEnv$optional_path}
\alias{ScalarOptionEnv$optional_os_string}
\alias{ScalarOptionEnv$optional_self}
\alias{ScalarOptionEnv$optional_strict}
\alias{ScalarOptionEnv$optional_worker}
\alias{ScalarOptionEnv$optional_worker_strict}
\title{ScalarOptionEnv Class}
\usage{
ScalarOptionEnv
}
\description{
ScalarOptionEnv Class
}
\keyword{internal}
