% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ScalarOptionR6}
\alias{ScalarOptionR6}
\alias{ScalarOptionR6$optional_self}
\title{ScalarOptionR6 R6 Class}
\description{
ScalarOptionR6 R6 Class
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-ScalarOptionR6-initialize}{\code{ScalarOptionR6$new()}}
    \item \href{#method-ScalarOptionR6-optional_integer}{\code{ScalarOptionR6$optional_integer()}}
    \item \href{#method-ScalarOptionR6-optional_real}{\code{ScalarOptionR6$optional_real()}}
    \item \href{#method-ScalarOptionR6-optional_logical}{\code{ScalarOptionR6$optional_logical()}}
    \item \href{#method-ScalarOptionR6-optional_string}{\code{ScalarOptionR6$optional_string()}}
    \item \href{#method-ScalarOptionR6-optional_str}{\code{ScalarOptionR6$optional_str()}}
    \item \href{#method-ScalarOptionR6-optional_unit}{\code{ScalarOptionR6$optional_unit()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScalarOptionR6-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-ScalarOptionR6-initialize}{}}}
\subsection{\code{ScalarOptionR6$new()}}{
  Create a new \code{ScalarOptionR6}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScalarOptionR6$new(.ptr = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{.ptr}}{Internal pointer (used by static methods, not for direct use).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScalarOptionR6-optional_integer"></a>}}
\if{latex}{\out{\hypertarget{method-ScalarOptionR6-optional_integer}{}}}
\subsection{\code{ScalarOptionR6$optional_integer()}}{
  Method \code{optional_integer}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScalarOptionR6$optional_integer(present)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{present}}{(no documentation available)}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScalarOptionR6-optional_real"></a>}}
\if{latex}{\out{\hypertarget{method-ScalarOptionR6-optional_real}{}}}
\subsection{\code{ScalarOptionR6$optional_real()}}{
  Method \code{optional_real}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScalarOptionR6$optional_real(present)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{present}}{(no documentation available)}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScalarOptionR6-optional_logical"></a>}}
\if{latex}{\out{\hypertarget{method-ScalarOptionR6-optional_logical}{}}}
\subsection{\code{ScalarOptionR6$optional_logical()}}{
  Method \code{optional_logical}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScalarOptionR6$optional_logical(present)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{present}}{(no documentation available)}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScalarOptionR6-optional_string"></a>}}
\if{latex}{\out{\hypertarget{method-ScalarOptionR6-optional_string}{}}}
\subsection{\code{ScalarOptionR6$optional_string()}}{
  Method \code{optional_string}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScalarOptionR6$optional_string(present)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{present}}{(no documentation available)}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScalarOptionR6-optional_str"></a>}}
\if{latex}{\out{\hypertarget{method-ScalarOptionR6-optional_str}{}}}
\subsection{\code{ScalarOptionR6$optional_str()}}{
  Method \code{optional_str}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScalarOptionR6$optional_str(present)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{present}}{(no documentation available)}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScalarOptionR6-optional_unit"></a>}}
\if{latex}{\out{\hypertarget{method-ScalarOptionR6-optional_unit}{}}}
\subsection{\code{ScalarOptionR6$optional_unit()}}{
  Method \code{optional_unit}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScalarOptionR6$optional_unit(present)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{present}}{(no documentation available)}
    }
    \if{html}{\out{</div>}}
  }
}

}
