% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ScalarOptionS3}
\alias{ScalarOptionS3}
\alias{new_scalaroptions3}
\alias{optional_integer.ScalarOptionS3}
\alias{optional_real.ScalarOptionS3}
\alias{optional_logical.ScalarOptionS3}
\alias{optional_string.ScalarOptionS3}
\alias{optional_str.ScalarOptionS3}
\alias{optional_unit.ScalarOptionS3}
\alias{scalaroptions3_optional_self}
\title{ScalarOptionS3 S3 Class}
\usage{
new_scalaroptions3()

\method{optional_integer}{ScalarOptionS3}(x, present, ...)

\method{optional_real}{ScalarOptionS3}(x, present, ...)

\method{optional_logical}{ScalarOptionS3}(x, present, ...)

\method{optional_string}{ScalarOptionS3}(x, present, ...)

\method{optional_str}{ScalarOptionS3}(x, present, ...)

\method{optional_unit}{ScalarOptionS3}(x, present, ...)

scalaroptions3_optional_self(present)

ScalarOptionS3
}
\arguments{
\item{x}{An object.}

\item{present}{(undocumented)}

\item{...}{Additional arguments.}
}
\description{
S3 generic for \code{optional_integer}

S3 generic for \code{optional_real}

S3 generic for \code{optional_logical}

S3 generic for \code{optional_string}

S3 generic for \code{optional_str}

S3 generic for \code{optional_unit}
}
\keyword{internal}
