% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{class}
\name{ScalarOptionS4}
\alias{ScalarOptionS4}
\alias{ScalarOptionS4-s4_optional_integer}
\alias{s4_optional_integer,ScalarOptionS4-method}
\alias{ScalarOptionS4-s4_optional_real}
\alias{s4_optional_real,ScalarOptionS4-method}
\alias{ScalarOptionS4-s4_optional_logical}
\alias{s4_optional_logical,ScalarOptionS4-method}
\alias{ScalarOptionS4-s4_optional_string}
\alias{s4_optional_string,ScalarOptionS4-method}
\alias{ScalarOptionS4-s4_optional_str}
\alias{s4_optional_str,ScalarOptionS4-method}
\alias{ScalarOptionS4-s4_optional_unit}
\alias{s4_optional_unit,ScalarOptionS4-method}
\alias{ScalarOptionS4_optional_self}
\title{ScalarOptionS4 S4 Class}
\usage{
ScalarOptionS4()

ScalarOptionS4_optional_self(present)
}
\arguments{
\item{present}{(undocumented)}
}
\description{
ScalarOptionS4 S4 Class
}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to Rust \code{ScalarOptionS4} struct}
}}

\keyword{internal}
