% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ScalarOptionVctrs}
\alias{ScalarOptionVctrs}
\alias{new_scalaroptionvctrs}
\alias{vec_ptype_abbr.ScalarOptionVctrs}
\alias{vec_ptype2.ScalarOptionVctrs.ScalarOptionVctrs}
\alias{vec_cast.ScalarOptionVctrs.ScalarOptionVctrs}
\alias{scalaroptionvctrs_optional_integer}
\alias{scalaroptionvctrs_optional_real}
\alias{scalaroptionvctrs_optional_logical}
\alias{scalaroptionvctrs_optional_string}
\alias{scalaroptionvctrs_optional_str}
\alias{scalaroptionvctrs_optional_unit}
\title{ScalarOptionVctrs vctrs S3 Class}
\usage{
new_scalaroptionvctrs(values)

\method{vec_ptype_abbr}{ScalarOptionVctrs}(x, ...)

\method{vec_ptype2}{ScalarOptionVctrs.ScalarOptionVctrs}(x, y, ...)

\method{vec_cast}{ScalarOptionVctrs.ScalarOptionVctrs}(x, to, ...)

scalaroptionvctrs_optional_integer(values, present)

scalaroptionvctrs_optional_real(values, present)

scalaroptionvctrs_optional_logical(values, present)

scalaroptionvctrs_optional_string(values, present)

scalaroptionvctrs_optional_str(values, present)

scalaroptionvctrs_optional_unit(values, present)
}
\arguments{
\item{values}{(undocumented)}

\item{x}{A ScalarOptionVctrs vector to cast.}

\item{...}{Additional arguments (unused).}

\item{y}{A ScalarOptionVctrs vector.}

\item{to}{A ScalarOptionVctrs prototype.}

\item{present}{(undocumented)}
}
\description{
ScalarOptionVctrs vctrs S3 Class
}
\keyword{internal}
