% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{scale}
\alias{scale}
\title{Scaling and Centering of Matrix}
\usage{
\S4method{scale}{fm}(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{a FlashR matrix}

\item{center}{either a logical value or a numeric vector of length equal to
the number of columns of \code{x}.}

\item{sclae}{either a logical value or a numeric vector of length equal to
the number of columns of \code{x}.}
}
\value{
a FlashR matrix.
}
\description{
\code{scale} centers and/or scales the columns of a FlashR matrix.
}
\details{
The value of \code{center} determines how column centering is
performed.  If \code{center} is a numeric vector with length equal to
the number of columns of \code{x}, then each column of \code{x} has the
corresponding value from \code{center} subtracted from it.  If \code{center}
is \code{TRUE} then centering is done by subtracting the column means
(omitting \code{NA}s) of \code{x} from their corresponding columns, and if
\code{center} is \code{FALSE}, no centering is done.

The value of \code{scale} determines how column scaling is performed
(after centering).  If \code{scale} is a numeric vector with length
equal to the number of columns of \code{x}, then each column of \code{x} is
divided by the corresponding value from \code{scale}.  If \code{scale} is
\code{TRUE} then scaling is done by dividing the (centered) columns of
\code{x} by their standard deviations if \code{center} is \code{TRUE},
and the root mean square otherwise.  If \code{scale} is \code{FALSE},
no scaling is done.

The root-mean-square for a (possibly centered) column is defined
as sqrt(sum(x^2)/(n-1)), where x is a vector of the non-missing
values and n is the number of non-missing values.  In the case
\code{center = TRUE}, this is the same as the standard deviation, but
in general it is not.
}
\examples{
mat <- fm.runif.matrix(100, 10)
mat <- scale(mat)
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

