% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ScalerR6}
\alias{ScalerR6}
\alias{ScalerR6$Scaler$scale}
\alias{ScalerR6$Scaler$bump}
\alias{ScalerR6$Scaler$set_mode}
\title{ScalerR6 R6 Class}
\description{
ScalerR6 R6 Class
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-ScalerR6-initialize}{\code{ScalerR6$new()}}
    \item \href{#method-ScalerR6-value}{\code{ScalerR6$value()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScalerR6-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-ScalerR6-initialize}{}}}
\subsection{\code{ScalerR6$new()}}{
  Create a new \code{ScalerR6}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScalerR6$new(v, .ptr = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{v}}{(no documentation available)}
      \item{\code{.ptr}}{Internal pointer (used by static methods, not for direct use).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScalerR6-value"></a>}}
\if{latex}{\out{\hypertarget{method-ScalerR6-value}{}}}
\subsection{\code{ScalerR6$value()}}{
  Method \code{value}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScalerR6$value()}
    \if{html}{\out{</div>}}
  }
}

}
