% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{class}
\name{ScalerS4}
\alias{ScalerS4}
\alias{ScalerS4-s4_value}
\alias{s4_value,ScalerS4-method}
\alias{s4_trait_Scaler_scale}
\alias{s4_trait_Scaler_scale,ScalerS4-method}
\alias{s4_trait_Scaler_bump}
\alias{s4_trait_Scaler_bump,ScalerS4-method}
\alias{s4_trait_Scaler_set_mode}
\alias{s4_trait_Scaler_set_mode,ScalerS4-method}
\title{ScalerS4 S4 Class}
\usage{
ScalerS4(v)

\S4method{s4_trait_Scaler_scale}{ScalerS4}(x, x_factor, ...)

\S4method{s4_trait_Scaler_bump}{ScalerS4}(x, amount, ...)

\S4method{s4_trait_Scaler_set_mode}{ScalerS4}(x, mode = c("fast", "slow"), ...)
}
\arguments{
\item{v}{(undocumented)}

\item{x}{A \code{ScalerS4} object}

\item{...}{Additional arguments passed to methods}
}
\description{
S4 generic for trait method \code{Scaler::scale}

S4 generic for trait method \code{Scaler::bump}

S4 generic for trait method \code{Scaler::set_mode}
}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to Rust \code{ScalerS4} struct}
}}

\keyword{internal}
