% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{score_total}
\alias{score_total}
\title{The \code{score_total()} generic}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{score_total()} is an S7 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{SidecarS7}}
}
}
