% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_stats.R
\docType{methods}
\name{sd}
\alias{sd}
\alias{sd,fm-method}
\alias{sd,fmV-method}
\title{Standard Deviation}
\usage{
\S4method{sd}{fm}(x, na.rm = FALSE)

\S4method{sd}{fmV}(x, na.rm = FALSE)
}
\arguments{
\item{x}{a FlashR vector or matrix.}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
This function computes the standard deviation of the values in \code{x}.
If \code{na.rm} is \code{TRUE} then missing values are removed before
computation proceeds.
}
\examples{
mat <- fm.runif.matrix(100, 10)
sd(mat)
}

