% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pols503-package.R
\name{sdcor2cov}
\alias{sdcor2cov}
\title{Create covariance matrix from standard deviations and correlation matrix}
\usage{
sdcor2cov(sd, cor = diag(length(sd)))
}
\arguments{
\item{sd}{\code{numeric} A vector of standard deviations}

\item{cor}{\code{matrix} A correlation matrix.}
}
\value{
A covariance matrix.
}
\description{
A covariance matrix can be specified as,
\deqn{\Sigma = diag(s) R diag(s)}
where \eqn{\Sigma} is the covariance matrix, \eqn{diag(s)} is a diagonal matrix
with the standard deviation on its diagonal, and \eqn{R} is the correlation matrix.
}
\examples{
s <- c(1, 2)
R <- matrix(c(1, 0.5, 0.5, 1), nrow = 2, ncol = 2)
sdcor2cov(s, R)
}

