% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{SeededNormal}
\alias{SeededNormal}
\alias{SeededNormal$new}
\alias{$.SeededNormal}
\alias{[[.SeededNormal}
\alias{SeededNormal$RDistributionOps$sample}
\alias{SeededNormal$RDistributionOps$sample_n}
\alias{SeededNormal$RDistributionOps$sample_vec}
\alias{SeededNormal$RDistributionOps$mean}
\alias{SeededNormal$RDistributionOps$variance}
\alias{SeededNormal$RDistributionOps$std_dev}
\title{SeededNormal Class}
\usage{
SeededNormal

\method{$}{SeededNormal}(self, name)

\method{[[}{SeededNormal}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
SeededNormal inherent methods: constructor.

Arguments of \code{SeededNormal$new()}:

\describe{
\item{\code{mean}}{Mean of the normal distribution.}
\item{\code{sd}}{Standard deviation (must be finite).}
\item{\code{seed}}{Integer seed (absolute value is used).}
}

Arguments of \code{SeededNormal$RDistributionOps$sample_n()}:

\describe{
\item{\code{n}}{Integer number of samples to draw.}
}

Arguments of \code{SeededNormal$RDistributionOps$sample_vec()}:

\describe{
\item{\code{n}}{Integer number of samples to draw.}
}
}
