% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_err_external}
\alias{serde_err_external}
\title{serde_err_external}
\usage{
serde_err_external(which)
}
\arguments{
\item{which}{One of \code{"bad"}, \code{"plain"}, \code{"unit"}.}
}
\description{
Raise one of the externally tagged variants.
}
