% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_err_missing}
\alias{serde_err_missing}
\title{serde_err_missing}
\usage{
serde_err_missing(field)
}
\arguments{
\item{field}{Field name to report missing.}
}
\description{
Struct variant with one string field.
}
