% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_deserialize_f64}
\alias{serde_r_deserialize_f64}
\title{serde_r_deserialize_f64}
\usage{
serde_r_deserialize_f64(sexp)
}
\arguments{
\item{sexp}{R object to deserialize.}
}
\description{
Test deserializing an R numeric to f64 via serde_r.
}
