% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_deserialize_i32}
\alias{serde_r_deserialize_i32}
\title{serde_r_deserialize_i32}
\usage{
serde_r_deserialize_i32(sexp)
}
\arguments{
\item{sexp}{R object to deserialize.}
}
\description{
Test deserializing an R integer to i32 via serde_r.
}
