% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_deserialize_missing_field}
\alias{serde_r_deserialize_missing_field}
\title{serde_r_deserialize_missing_field}
\usage{
serde_r_deserialize_missing_field(sexp)
}
\arguments{
\item{sexp}{R list that is missing a field required by SerdeRPoint.}
}
\description{
Test that deserializing a struct with a missing required field returns an error.
}
