% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_deserialize_string}
\alias{serde_r_deserialize_string}
\title{serde_r_deserialize_string}
\usage{
serde_r_deserialize_string(sexp)
}
\arguments{
\item{sexp}{R object to deserialize.}
}
\description{
Test deserializing an R character to String via serde_r.
}
