% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_deserialize_vec_f64}
\alias{serde_r_deserialize_vec_f64}
\title{serde_r_deserialize_vec_f64}
\usage{
serde_r_deserialize_vec_f64(sexp)
}
\arguments{
\item{sexp}{R numeric vector to roundtrip.}
}
\description{
Test deserializing an R numeric vector to Vec\if{html}{\out{<f64>}} then re-serializing.
}
