% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_deserialize_vec_i32}
\alias{serde_r_deserialize_vec_i32}
\title{serde_r_deserialize_vec_i32}
\usage{
serde_r_deserialize_vec_i32(sexp)
}
\arguments{
\item{sexp}{R integer vector to roundtrip.}
}
\description{
Test deserializing an R integer vector to Vec\if{html}{\out{<i32>}} then re-serializing.
}
