% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_roundtrip_collections}
\alias{serde_r_roundtrip_collections}
\title{serde_r_roundtrip_collections}
\usage{
serde_r_roundtrip_collections()
}
\description{
Test serialize-then-deserialize roundtrip of a struct with multiple collection types.
}
