% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_roundtrip_deep_nest}
\alias{serde_r_roundtrip_deep_nest}
\title{serde_r_roundtrip_deep_nest}
\usage{
serde_r_roundtrip_deep_nest()
}
\description{
Test serialize-then-deserialize roundtrip of a deeply nested struct.
}
