% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_roundtrip_optionals_none}
\alias{serde_r_roundtrip_optionals_none}
\title{serde_r_roundtrip_optionals_none}
\usage{
serde_r_roundtrip_optionals_none()
}
\description{
Test roundtrip of a struct where all optional fields are None (NA).
}
