% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_roundtrip_optionals_present}
\alias{serde_r_roundtrip_optionals_present}
\title{serde_r_roundtrip_optionals_present}
\usage{
serde_r_roundtrip_optionals_present()
}
\description{
Test roundtrip of a struct where all optional fields are present (Some).
}
