% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_roundtrip_point}
\alias{serde_r_roundtrip_point}
\title{serde_r_roundtrip_point}
\usage{
serde_r_roundtrip_point(x, y)
}
\arguments{
\item{x}{Numeric x-coordinate.}

\item{y}{Numeric y-coordinate.}
}
\description{
Test serialize-then-deserialize roundtrip of a SerdeRPoint struct.
}
