% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_roundtrip_rectangle}
\alias{serde_r_roundtrip_rectangle}
\title{serde_r_roundtrip_rectangle}
\usage{
serde_r_roundtrip_rectangle(x1, y1, x2, y2)
}
\arguments{
\item{x1}{Numeric top-left x.}

\item{y1}{Numeric top-left y.}

\item{x2}{Numeric bottom-right x.}

\item{y2}{Numeric bottom-right y.}
}
\description{
Test serialize-then-deserialize roundtrip of a Rectangle struct.
}
