% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_serialize_hashmap}
\alias{serde_r_serialize_hashmap}
\title{serde_r_serialize_hashmap}
\usage{
serde_r_serialize_hashmap()
}
\description{
Test serializing a HashMap<String, i32> to an R named list via serde_r.
}
