% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_serialize_i32}
\alias{serde_r_serialize_i32}
\title{serde_r_serialize_i32}
\usage{
serde_r_serialize_i32(x)
}
\arguments{
\item{x}{Integer scalar to serialize.}
}
\description{
Test serializing a scalar i32 to an R value via serde_r.
}
