% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_serialize_option_i32}
\alias{serde_r_serialize_option_i32}
\title{serde_r_serialize_option_i32}
\usage{
serde_r_serialize_option_i32(x)
}
\arguments{
\item{x}{Integer scalar or NA.}
}
\description{
Test serializing an Option\if{html}{\out{<i32>}} to an R integer (NA for None).
}
