% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_serialize_string}
\alias{serde_r_serialize_string}
\title{serde_r_serialize_string}
\usage{
serde_r_serialize_string(x)
}
\arguments{
\item{x}{Character scalar to serialize.}
}
\description{
Test serializing a String to an R character via serde_r.
}
