% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_serialize_tuple}
\alias{serde_r_serialize_tuple}
\title{serde_r_serialize_tuple}
\usage{
serde_r_serialize_tuple()
}
\description{
Test serializing a Rust tuple (i32, f64, String) to an R list.
}
