% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_serialize_tuple_struct}
\alias{serde_r_serialize_tuple_struct}
\title{serde_r_serialize_tuple_struct}
\usage{
serde_r_serialize_tuple_struct()
}
\description{
Test serializing a Rust tuple struct (Pair) to an R list.
}
