% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_serialize_vec_bool}
\alias{serde_r_serialize_vec_bool}
\title{serde_r_serialize_vec_bool}
\usage{
serde_r_serialize_vec_bool(x)
}
\arguments{
\item{x}{Logical vector input.}
}
\description{
Test roundtrip of Vec\if{html}{\out{<bool>}} through serde_r (deserialize then serialize).
}
