% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_serialize_vec_f64}
\alias{serde_r_serialize_vec_f64}
\title{serde_r_serialize_vec_f64}
\usage{
serde_r_serialize_vec_f64(x)
}
\arguments{
\item{x}{Numeric vector input.}
}
\description{
Test roundtrip of Vec\if{html}{\out{<f64>}} through serde_r (deserialize then serialize).
}
