% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_serialize_vec_i32}
\alias{serde_r_serialize_vec_i32}
\title{serde_r_serialize_vec_i32}
\usage{
serde_r_serialize_vec_i32(x)
}
\arguments{
\item{x}{Integer vector input.}
}
\description{
Test roundtrip of Vec\if{html}{\out{<i32>}} through serde_r (deserialize then serialize).
}
