% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_serialize_vec_string}
\alias{serde_r_serialize_vec_string}
\title{serde_r_serialize_vec_string}
\usage{
serde_r_serialize_vec_string(x)
}
\arguments{
\item{x}{Character vector input.}
}
\description{
Test roundtrip of Vec\if{html}{\out{<String>}} through serde_r (deserialize then serialize).
}
