% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serde_r_trait_to_r}
\alias{serde_r_trait_to_r}
\title{serde_r_trait_to_r}
\usage{
serde_r_trait_to_r()
}
\description{
Serialize through the \code{RSerializeNative} trait (blanket impl over
\code{serde::Serialize}).
}
