% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serialized_invisible_attr}
\alias{serialized_invisible_attr}
\title{serialized_invisible_attr}
\usage{
serialized_invisible_attr()
}
\description{
Combine serialization with an outer visibility marker.
}
