% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serialized_result_attr}
\alias{serialized_result_attr}
\title{serialized_result_attr}
\usage{
serialized_result_attr(fail)
}
\arguments{
\item{fail}{Whether to return the Err variant as serde data.}
}
\description{
Serialize a complete Result, including its variant.
}
