% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{serialized_result_type}
\alias{serialized_result_type}
\title{serialized_result_type}
\usage{
serialized_result_type(fail)
}
\arguments{
\item{fail}{Whether to return the Err variant.}
}
\description{
The equivalent complete-Result type spelling.
}
