% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/miniextendr-wrappers.R
\name{set}
\alias{set}
\title{Set mode (S7)}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{set()} is an S7 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{S7MatchArgHolder}}
}
}
