% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{set_edge_names}
\alias{set_edge_names}
\title{Set edge names}
\usage{
set_edge_names(g)
}
\description{
Set edge names
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_edge_names(g)
head(igraph::E(g)$name)
}
}
