% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{set_level}
\alias{set_level}
\title{The \code{set_level()} generic}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{set_level()} is an S7 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{ErrorInRS7Gauge}}
}
}
