% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complm.R
\name{set_pairwise}
\alias{set_pairwise}
\title{Convert response (taxa) to pairwise logratios}
\usage{
set_pairwise(x, diag = FALSE, upper = FALSE)
}
\description{
\code{upper} and \code{diag} arguments patterned after \code{dist}, and refer to the
corresponding entries in the pairwise difference matrix, even though we're
giving the pairwise differences stacked in a vector.
}
\examples{
fit[1:3,,1:2] \%>\% set_pairwise
fit[1:3,,1:2] \%>\% set_pairwise(upper = TRUE)

}
