% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complm.R
\name{set_ref}
\alias{set_ref}
\title{Set the reference taxa or coefficient}
\usage{
set_ref(x, margin, ref)
}
\description{
Note, I'm not sure what conditions its actually safe to do this
}
\examples{
set_ref(fit1, 1, 1)
set_ref(fit1, 1, c("T1", "T2"))
set_ref(fit1, 2, c("protocolP2"))
set_ref(fit1, 2, str_c("protocol", protocols))
set_ref(fit1, 3, 1) 

}
