% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complm.R
\name{set_ref_taxa}
\alias{set_ref_taxa}
\title{Set the reference taxon/taxa}
\usage{
set_ref_taxa(x, ref)
}
\description{
Set the reference taxon/taxa
}
\examples{
smaller array for testing
fit0 <- fit[1:5,,1:2]
set_denom(fit0, 1)
set_denom(fit0, 1:2)
set_denom(fit0, c("T1", "T2"))

}
