% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{set_txpath_to_vertex_attr}
\alias{set_txpath_to_vertex_attr}
\title{Set transcript paths as vertex attributes on an igraph splice graph}
\usage{
set_txpath_to_vertex_attr(g)
}
\description{
Set transcript paths as vertex attributes on an igraph splice graph
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_txpath_to_vertex_attr(g)
igraph::vertex_attr_names(g)
}
}
