% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{SG2igraph}
\alias{SG2igraph}
\title{Convert splicingGraph to igraph}
\usage{
SG2igraph(geneID, gene_sg, gene_graph)
}
\arguments{
\item{geneID}{Character string. Gene identifier (e.g., Ensembl gene ID).}

\item{gene_sg}{A \code{SplicingGraphs} object subsetted to the gene of interest.}

\item{gene_graph}{A \code{GRanges} or data frame of splice graph edges for the gene.}
}
\description{
Convert splicingGraph to igraph
}
\examples{
\dontrun{
library(SplicingGraphs)
gr <- rtracklayer::import("ENSG00000000003.gtf")
gr <- gr[!(rtracklayer::mcols(gr)$type \%in\% c("start_codon", "stop_codon"))]
sg <- SplicingGraphs::SplicingGraphs(txdbmaker::makeTxDbFromGRanges(gr), min.ntx = 1)
edges_by_gene <- SplicingGraphs::sgedgesByGene(sg)
geneID <- "ENSG00000000003.14"
g <- grase::SG2igraph(geneID, sg[geneID], edges_by_gene[[geneID]])
}
}
