% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{SharedData}
\alias{SharedData}
\title{SharedData R6 Class}
\description{
SharedData R6 Class
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-SharedData-initialize}{\code{SharedData$new()}}
    \item \href{#method-SharedData-get_x}{\code{SharedData$get_x()}}
    \item \href{#method-SharedData-get_y}{\code{SharedData$get_y()}}
    \item \href{#method-SharedData-get_label}{\code{SharedData$get_label()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharedData-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-SharedData-initialize}{}}}
\subsection{\code{SharedData$new()}}{
  Create a new \code{SharedData}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SharedData$new(x, y, label, .ptr = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{x}}{Numeric x-coordinate.}
      \item{\code{y}}{Numeric y-coordinate.}
      \item{\code{label}}{Character label.}
      \item{\code{.ptr}}{Internal pointer (used by static methods, not for direct use).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharedData-get_x"></a>}}
\if{latex}{\out{\hypertarget{method-SharedData-get_x}{}}}
\subsection{\code{SharedData$get_x()}}{
  Method \code{get_x}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SharedData$get_x()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharedData-get_y"></a>}}
\if{latex}{\out{\hypertarget{method-SharedData-get_y}{}}}
\subsection{\code{SharedData$get_y()}}{
  Method \code{get_y}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SharedData$get_y()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharedData-get_label"></a>}}
\if{latex}{\out{\hypertarget{method-SharedData-get_label}{}}}
\subsection{\code{SharedData$get_label()}}{
  Method \code{get_label}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SharedData$get_label()}
    \if{html}{\out{</div>}}
  }
}

}
