% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{SharedSimpleCounter}
\alias{SharedSimpleCounter}
\alias{SharedSimpleCounter$new}
\alias{SharedSimpleCounter$get_value}
\alias{$.SharedSimpleCounter}
\alias{[[.SharedSimpleCounter}
\alias{SharedSimpleCounter$SharedCounter$value}
\alias{SharedSimpleCounter$SharedCounter$increment}
\alias{SharedSimpleCounter$SharedCounter$add}
\alias{SharedSimpleCounter$SharedCounter$reset}
\title{SharedSimpleCounter Class}
\usage{
SharedSimpleCounter

\method{$}{SharedSimpleCounter}(self, name)

\method{[[}{SharedSimpleCounter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
SharedSimpleCounter Class
}
