% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{shift}
\alias{shift}
\title{Shift a vector forward or backward depending on the lag value}
\usage{
shift(v, lag)
}
\arguments{
\item{v:}{input vector}

\item{lag:}{lag value for shifting v}
}
\description{
\code{shift} shifts an input vector v backward or forward
if lag < 0: backward and otherwise forward
}
