% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{shortest_edges}
\alias{shortest_edges}
\title{Find the shortest edges in an igraph layout for curvature adjustment}
\usage{
shortest_edges(g, coords)
}
\arguments{
\item{g}{An \code{igraph} directed acyclic graph representing a gene splicing graph.}

\item{coords}{A numeric matrix of layout coordinates with one row per vertex (as returned by
igraph layout functions such as \code{igraph::layout_with_sugiyama}).}
}
\description{
Find the shortest edges in an igraph layout for curvature adjustment
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
coords <- igraph::layout_with_sugiyama(g)$layout
pairs <- grase::shortest_edges(g, coords)
}
}
