% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_metadata.R
\name{show_overall_statistics}
\alias{show_overall_statistics}
\title{Show Overall Statistics for Metadata}
\usage{
show_overall_statistics()
}
\description{
Computes and displays overall statistics for the datasets stored in the internal metadata file.
}
\details{
The function dynamically loads the \code{metadata.csv} file from the package's \verb{data/} folder
and computes:
\itemize{
\item Range (min, max) and mean values for \code{id_count}, \code{item_count}, \code{resp_count}, and \code{sparsity}.
\item Number of tables with \code{rater}, \verb{response time (rt)}, and \code{date} information.
}

The function prints the statistics to the console without returning any value.
}
\examples{
\dontrun{
  show_overall_statistics()
}

}
