% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidb.R
\docType{data}
\name{sidb}
\alias{sidb}
\title{Soil incubation data included in the R package}
\format{A list with metadata, time series, and initial conditions. Each database entry contains:
\describe{
  \item{metadata}{This is a list with metadata associated with the entry}
  \item{timeSeries}{A data.frame with time series of respiration fluxes}
  \item{initConditions}{A data.frame with details of the initial conditions of the experiment}
}}
\source{
\url{https://github.com/SoilBGC-Datashare/sidb/tree/master/data}
}
\usage{
sidb
}
\description{
A dataset containing the time series of respired carbon from incubation experiments and associeted metadata
}
\keyword{datasets}
