% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{sidecar_computed_keys}
\alias{sidecar_computed_keys}
\title{sidecar_computed_keys}
\usage{
sidecar_computed_keys(ptr)
}
\arguments{
\item{ptr}{A \code{SidecarComputed} pointer.}
}
\description{
The \code{keys} of a \code{SidecarComputed}, through an \verb{ExternalPtr<T>} argument.
}
