% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{SidecarEnv_get_name}
\alias{SidecarEnv_get_name}
\title{Get \code{name} field from SidecarEnv}
\usage{
SidecarEnv_get_name(x)
}
\arguments{
\item{x}{The SidecarEnv external pointer}
}
\value{
The value of the \code{name} field
}
\description{
Get \code{name} field from SidecarEnv
}
