% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{SidecarEnv_get_raw_slot}
\alias{SidecarEnv_get_raw_slot}
\title{Get \code{raw_slot} field from SidecarEnv}
\usage{
SidecarEnv_get_raw_slot(x)
}
\arguments{
\item{x}{The SidecarEnv external pointer}
}
\value{
The value of the \code{raw_slot} field
}
\description{
Get \code{raw_slot} field from SidecarEnv
}
