% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{SidecarEnv_get_score}
\alias{SidecarEnv_get_score}
\title{Get \code{score} field from SidecarEnv}
\usage{
SidecarEnv_get_score(x)
}
\arguments{
\item{x}{The SidecarEnv external pointer}
}
\value{
The value of the \code{score} field
}
\description{
Get \code{score} field from SidecarEnv
}
