% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{SidecarEnv_set_score}
\alias{SidecarEnv_set_score}
\title{Set \code{score} field on SidecarEnv}
\usage{
SidecarEnv_set_score(x, value)
}
\arguments{
\item{x}{The SidecarEnv external pointer}

\item{value}{The new value to set}
}
\value{
The SidecarEnv pointer (invisibly)
}
\description{
Set \code{score} field on SidecarEnv
}
