% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{SidecarHandmade}
\alias{SidecarHandmade}
\alias{SidecarHandmade_get_id}
\alias{SidecarHandmade_set_id}
\alias{SidecarHandmade_get_ids}
\alias{SidecarHandmade_set_ids}
\alias{SidecarHandmade_get_label}
\alias{SidecarHandmade_get_extra}
\alias{SidecarHandmade_set_extra}
\title{SidecarHandmade Sidecar Accessors}
\usage{
SidecarHandmade_get_id(x)

SidecarHandmade_set_id(x, value)

SidecarHandmade_get_ids(x)

SidecarHandmade_set_ids(x, value)

SidecarHandmade_get_label(x)

SidecarHandmade_get_extra(x)

SidecarHandmade_set_extra(x, value)
}
\arguments{
\item{x}{The SidecarHandmade external pointer}

\item{value}{The new value to set}
}
\value{
The value of the \code{id} field

The SidecarHandmade pointer (invisibly)

The value of the \code{ids} field

The SidecarHandmade pointer (invisibly)

The value of the \code{label} field

The value of the \code{extra} field

The SidecarHandmade pointer (invisibly)
}
\description{
Getter and setter functions for \verb{#[r_data]} fields on \code{SidecarHandmade}.
}
