% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{SidecarR6_get_value}
\alias{SidecarR6_get_value}
\title{Get \code{value} field from SidecarR6 (for R6)}
\usage{
SidecarR6_get_value(x)
}
\arguments{
\item{x}{The SidecarR6 external pointer}
}
\value{
The value of the \code{value} field
}
\description{
Get \code{value} field from SidecarR6 (for R6)
}
