% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{SidecarR6_set_label}
\alias{SidecarR6_set_label}
\title{Set \code{label} field on SidecarR6 (for R6)}
\usage{
SidecarR6_set_label(x, value)
}
\arguments{
\item{x}{The SidecarR6 external pointer}

\item{value}{The new value to set}
}
\value{
The SidecarR6 pointer (invisibly)
}
\description{
Set \code{label} field on SidecarR6 (for R6)
}
