% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{SidecarRawSexp}
\alias{SidecarRawSexp}
\alias{SidecarRawSexp_get_int_vec}
\alias{SidecarRawSexp_set_int_vec}
\alias{SidecarRawSexp_get_real_vec}
\alias{SidecarRawSexp_set_real_vec}
\alias{SidecarRawSexp_get_char_vec}
\alias{SidecarRawSexp_set_char_vec}
\alias{SidecarRawSexp_get_list_val}
\alias{SidecarRawSexp_set_list_val}
\alias{SidecarRawSexp_get_func_val}
\alias{SidecarRawSexp_set_func_val}
\alias{SidecarRawSexp_get_env_val}
\alias{SidecarRawSexp_set_env_val}
\alias{$.SidecarRawSexp}
\alias{[[.SidecarRawSexp}
\title{SidecarRawSexp Sidecar Accessors}
\usage{
SidecarRawSexp_get_int_vec(x)

SidecarRawSexp_set_int_vec(x, value)

SidecarRawSexp_get_real_vec(x)

SidecarRawSexp_set_real_vec(x, value)

SidecarRawSexp_get_char_vec(x)

SidecarRawSexp_set_char_vec(x, value)

SidecarRawSexp_get_list_val(x)

SidecarRawSexp_set_list_val(x, value)

SidecarRawSexp_get_func_val(x)

SidecarRawSexp_set_func_val(x, value)

SidecarRawSexp_get_env_val(x)

SidecarRawSexp_set_env_val(x, value)

SidecarRawSexp

\method{$}{SidecarRawSexp}(self, name)

\method{[[}{SidecarRawSexp}(self, name)
}
\arguments{
\item{x}{The SidecarRawSexp external pointer}

\item{value}{The new value to set}

\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\value{
The value of the \code{int_vec} field

The SidecarRawSexp pointer (invisibly)

The value of the \code{real_vec} field

The SidecarRawSexp pointer (invisibly)

The value of the \code{char_vec} field

The SidecarRawSexp pointer (invisibly)

The value of the \code{list_val} field

The SidecarRawSexp pointer (invisibly)

The value of the \code{func_val} field

The SidecarRawSexp pointer (invisibly)

The value of the \code{env_val} field

The SidecarRawSexp pointer (invisibly)
}
\description{
Getter and setter functions for \verb{#[r_data]} fields on \code{SidecarRawSexp}.

Env class registration for SidecarRawSexp (raw SEXP slot testing).
}
