% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{SidecarVctrs}
\alias{SidecarVctrs}
\alias{SidecarVctrs_get_vec_data}
\alias{SidecarVctrs_set_vec_data}
\alias{SidecarVctrs_get_vec_label}
\alias{SidecarVctrs_set_vec_label}
\alias{vec_ptype2.SidecarVctrs.SidecarVctrs}
\alias{vec_cast.SidecarVctrs.SidecarVctrs}
\title{SidecarVctrs Sidecar Accessors}
\usage{
SidecarVctrs_get_vec_data(x)

SidecarVctrs_set_vec_data(x, value)

SidecarVctrs_get_vec_label(x)

SidecarVctrs_set_vec_label(x, value)

\method{vec_ptype2}{SidecarVctrs.SidecarVctrs}(x, y, ...)

\method{vec_cast}{SidecarVctrs.SidecarVctrs}(x, to, ...)
}
\arguments{
\item{x}{A SidecarVctrs vector to cast.}

\item{value}{The new value to set}

\item{y}{A SidecarVctrs vector.}

\item{...}{Additional arguments (unused).}

\item{to}{A SidecarVctrs prototype.}
}
\value{
The value of the \code{vec_data} field

The SidecarVctrs pointer (invisibly)

The value of the \code{vec_label} field

The SidecarVctrs pointer (invisibly)
}
\description{
Getter and setter functions for \verb{#[r_data]} fields on \code{SidecarVctrs}.
}
