% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate data from a multi-arm trial with shared control}
\usage{
sim_data(n_arms, N, mu_6m, mu_12m, sigma, rmonth)
}
\arguments{
\item{n_arms}{number of arms (including control)}

\item{N}{total sample size}

\item{mu_6m}{6-month mean response per arm (vector of length \code{n_arm})}

\item{mu_12m}{12-month mean response per arm (vector of length \code{n_arm})}

\item{sigma}{covariance matrix between 6- and 12-month responses assumed equal across arms (matrix of dim 2x2)}

\item{rmonth}{recruitment per month (recruitment speed assumed constant over time)}
}
\value{
simulated data consisting of the responses at 6 and 12 months, treatment arm, and recruitment time for each subject.
}
\description{
Function to simulate trial data (1-stage, multiple arms)
}
\details{
eWHORM simulations
}
\author{
Marta Bofill Roig
}
