% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_trial.R
\name{sim_trial}
\alias{sim_trial}
\title{Simulate data from a multi-arm multi-stage trial with shared control and dose selection}
\usage{
sim_trial(
  n_arms = 4,
  N1 = 30 * 4,
  N2 = 30 * 2,
  mu_6m,
  mu_12m,
  sigma,
  rmonth,
  alpha1 = 0.5,
  alpha = 0.05,
  p_safety = c(0.9, 0.8, 0.7),
  safety = T
)
}
\arguments{
\item{n_arms}{number of arms (including control)}

\item{N1}{sample size stage 1}

\item{N2}{sample size stage 2}

\item{mu_6m}{6-month mean response per arm (vector of length \code{n_arm})}

\item{mu_12m}{12-month mean response per arm (vector of length \code{n_arm})}

\item{sigma}{covariance matrix between 6- and 12-month responses assumed equal across arms (matrix of dim 2x2)}

\item{rmonth}{recruitment per month (recruitment speed assumed constant over time)}

\item{alpha1}{significance level for dose selection}

\item{alpha}{significance level for selected dose vs control comparison}

\item{p_safety}{probability of each dose to be safe}

\item{safety}{indicator - if true, it simulates safety according to  \code{p_safety}}
}
\value{
Combined p-value, selected dose and safety for each dose (if argument \code{safety=TRUE})
}
\description{
Function to simulate trial data (2-stages, with dose selection)
}
\details{
eWHORM simulations
}
\author{
Marta Bofill Roig
}
