% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_trial_pce.R
\name{sim_trial_pce}
\alias{sim_trial_pce}
\title{Simulate data from a multi-arm multi-stage trial with shared control and two initial doses, where an additional dose could be added after the interim analysis}
\usage{
sim_trial_pce(
  n_arms = 4,
  N1 = 30 * 4,
  N2 = 30 * 2,
  mu_6m,
  mu_12m,
  sigma,
  rmonth,
  alpha1 = 0.1,
  alpha = 0.05
)
}
\arguments{
\item{n_arms}{number of arms (including control)}

\item{N1}{sample size stage 1}

\item{N2}{sample size stage 2}

\item{mu_6m}{6-month mean response per arm (vector of length \code{n_arm})}

\item{mu_12m}{12-month mean response per arm (vector of length \code{n_arm})}

\item{sigma}{covariance matrix between 6- and 12-month responses assumed equal across arms (matrix of dim 2x2)}

\item{rmonth}{recruitment per month (recruitment speed assumed constant over time)}

\item{alpha1}{significance level for dose selection}

\item{alpha}{significance level for selected dose vs control comparison}
}
\value{
A list consisting of pvalues at stage 1, pvalues at stage 2, the decision at stages 1 and 2, the selected dose at stage 1, and the time at which the last patient was recruited in stage 1 and 2.
}
\description{
Function to simulate trial data (2-stages, with dose selection). The analyses are performed using partial conditional error.
}
\details{
eWHORM simulations
}
\author{
Marta Bofill Roig
}
