% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_trial_pceind_test.R
\name{sim_trial_pceind_test}
\alias{sim_trial_pceind_test}
\title{Simulate data from a multi-arm multi-stage trial with shared control and two initial doses, where an additional dose could be added after the interim analysis; individual observations are simulated}
\usage{
sim_trial_pceind_test(
  n_arms = 4,
  N1,
  N2,
  mu_0m,
  mu_6m,
  mu_12m,
  sg,
  alpha1,
  alpha = 0.025,
  sel_scen,
  side = T,
  test,
  dropout,
  rr,
  bound
)
}
\arguments{
\item{n_arms}{number of arms (including control)}

\item{N1}{sample size stage 1}

\item{N2}{sample size stage 2}

\item{mu_0m}{Baseline value per arm (vector of length \code{n_arm})}

\item{mu_6m}{6-month value per arm (vector of length \code{n_arm})}

\item{mu_12m}{12-month value per arm (vector of length \code{n_arm})}

\item{sg}{sigma covariance matrix between 6- and 12-month mean differences assumed equal across arms (matrix of dim 2x2)}

\item{alpha1}{significance level for dose selection (futility boundary)}

\item{alpha}{significance level for selected dose vs control comparison}

\item{sel_scen}{choose between two different options in case that in interim analysis low dose is promising, but median dose not: 0: do not continue with low dose or median dose; 1: continue with low and median doses}

\item{side}{TRUE/FALSE referring to the side for 1-side testing (if TRUE then lower = side)}

\item{test}{defines type of analysis: "t" calculates a t-test, "l" a linear model with baseline values as covariables, "w" Wilcoxon test of differences, and "w1" Wilcoxon test of follow-up values}

\item{dropout}{dropoutrate, between 0 and 1}

\item{rr}{responder rate for each dose (vector of length \code{n_arm}), which gives the proportion of patients with value 0 at follow-up}

\item{bound}{lower bound to define total responder in simulation study}
}
\value{
A list consisting of pvalues at stage 1, pvalues at stage 2, the decision at stages 1 and 2, the selected dose at stage 1, concordance values, corresponding bias and confidence limits and corresponding values on the multi-armed trials.
}
\description{
Function to simulate trial data (2-stages, with dose selection). The analyses are performed using partial conditional error rates.
}
\details{
eWHORM simulations
}
\author{
Marta Bofill Roig, Sonja Zehetmayer
}
