% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_res.R
\name{simul_res}
\alias{simul_res}
\title{Simulate trial}
\usage{
simul_res(
  mu_raw_0,
  sd_raw_0,
  r0_6,
  r1_6,
  r2_6,
  r3_6,
  r0_12,
  r1_12,
  r2_12,
  r3_12,
  rho,
  n_trials,
  n_arms = 4,
  N1,
  N2,
  alpha1,
  alpha,
  sel_scen,
  side1,
  test1,
  dropout,
  rr0,
  rr1,
  rr2,
  rr3,
  bound
)
}
\arguments{
\item{mu_raw_0}{num mean value of original data at baseline}

\item{sd_raw_0}{num standard deviation of original data at baseline}

\item{r0_6}{reduction rate of control dose at Month 6 with values between 0 and 1}

\item{r1_6}{reduction rate of low dose at Month 6 with values between 0 and 1}

\item{r2_6}{reduction rate of medium dose at Month 6 with values between 0 and 1}

\item{r3_6}{reduction rate of high dose at Month 6 with values between 0 and 1}

\item{r0_12}{reduction rate of control dose at Month 12 with values between 0 and 1}

\item{r1_12}{reduction rate of low dose at Month 12 with values between 0 and 1}

\item{r2_12}{reduction rate of medium dose at Month 12 with values between 0 and 1}

\item{r3_12}{reduction rate of high dose at Month 12 with values between 0 and 1}

\item{rho}{correlation between baseline and 12 months follow-up observations and between 6 months and 12 months follow-up observations}

\item{n_trials}{number of simulation runs}

\item{n_arms}{number of arms (including control)}

\item{N1}{sample size stage 1}

\item{N2}{sample size stage 2}

\item{alpha1}{significance level for dose selection (futility boundary)}

\item{alpha}{significance level for selected dose vs control comparison}

\item{sel_scen}{choose between two different options in case that in interim analysis low dose is promising, but median dose not: 0: do not continue with low dose or median dose; 1: continue with low and median doses}

\item{side1}{TRUE/FALSE referring to the side for 1-side testing (if TRUE then lower = side)}

\item{test1}{defines type of analysis: "t" calculates a t-test, "l" a linear model with baseline values as covariables, "w" Wilcoxon test of differences, and "w1" Wilcoxon test of follow-up values}

\item{dropout}{dropoutrate with values between 0 and 1}

\item{rr0}{total responder rate for control dose, which gives the proportion of patients with value 0 at follow-up}

\item{rr1}{total responder rate for low dose, which gives the proportion of patients with value 0 at follow-up}

\item{rr2}{total responder rate for medium dose, which gives the proportion of patients with value 0 at follow-up}

\item{rr3}{total responder rate for high dose, which gives the proportion of patients with value 0 at follow-up}

\item{bound}{lower bound to define total responder in simulation study}
}
\value{
A vector consisting of summary measures for data simulation with n_trials repetitions: frequency selected for stage 2, conditional power, power ,disjunctive power, power of MA1, power of MA2, power of ma1a, concordances
}
\description{
Simulate trial
}
\details{
eWHORM simulations
}
\author{
Marta Bofill Roig, Sonja Zehetmayer
}
\keyword{internal}
