% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_dataset}
\alias{simulate_dataset}
\title{Simulate a differential-bias dataset}
\usage{
simulate_dataset(n, sd, log = TRUE, phyloseq = FALSE)
}
\arguments{
\item{n}{5-vector specifying the numbers of taxa, specimens, protocols,
conditions, and replicates}

\item{sd}{5-vector specifying the standard deviation for the draws of the
specimen, protocol, condition, and protocol:condition effects and finally
the residuals}
}
\description{
Note, assumes that specimen is nested within condition, like the vivo-vitro
experiment. Should think of condition as a treatment-var of the specimens.
The first condition is always named "Ref". Tries to even splits specimen
among the conditions. (Perhaps I should just call the conditions C1, ...)
}
\details{
The returned \code{coefficients} are the coefficents relative to the Ref-Ref
protocol, condition combination.
}
