% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slope_Calc_Upwind.R
\name{SlopeCalcUP}
\alias{SlopeCalcUP}
\title{Calculate the slopes from a DEM}
\usage{
SlopeCalcUP(dem, direction, dx, dy, mask, borders, borderdir = 1,
  d4 = c(1, 2, 3, 4), minslope = 1e-05, maxslope = -1,
  secondaryTH = -1, river_method = 0, river_secondaryTH = 0,
  rivermask, subbasins, printflag = F, upflag = T)
}
\arguments{
\item{dem}{Elevation matrix}

\item{direction}{a matrix of the flow directions for cells that have been processed - defaults to all zeros}

\item{dx, dy}{Lateral grid cell resolution}

\item{mask}{Mask with ones for cells to be processed and zeros for everything else - defaults to a mask of all 1's}

\item{borders}{A matrix with 1's for borders cells to default to pointing out, 2 to default to pointing in, and 0 for all non-border cells.}

\item{borderdir}{Default value for border cells 1=point out, 2= point in}

\item{d4}{directional numbering system: the numbers you want to assigne to down, left, top,right (defaults to 1,2,3,4)}

\item{minslope}{Minimum absolute slope value to apply to flat cells if needed. Defaults to 1e-5}

\item{maxslope}{Maximum absolute value of slopes. If this is set to -1 the slopes will not be limited. Default value is -1}

\item{secondaryTH}{secondary threshold - maximum ratio of |secondary|/|primary| to be enforced. NOTE - this scaling occurs after any max threholds are applied. If this is set to -1 no scaling will be applied.}

\item{river_method}{Optional method to treat river cells differently from the rest of the domain}

\item{river_secondaryTH}{secondary threshold  to apply to the river cells if river method 1-3 is chosen- maximum ratio of |secondary|/|primary| to be enforced. NOTE - this scaling occurs after any max threholds are applied and the river smoothing is done if you are usnig optons 2 or 3.  The dafault value is zero. NOTE- there is not currently a '-1' equivalend for this threshold. secondary slopes along the river cells must either be scaled to 0 or some ratio of the primary slopes if one of the river methods is chosen.}

\item{rivermask}{Mask with 1 for river cells and 0 for other cells}

\item{subbasins}{An optional matrix of subbasin values}

\item{printflag}{Print function progress}

\item{upflag}{A flag indicating whether slope calc should be upwinded, defaults to T generating slopes that are consistent with ParFlow. If set to F then all slopes will be calcualted as [i+1]-[i]}
}
\description{
This function will calculate slopes using standard or upwinding options and apply a range of smoothing options
}
\section{River Methods}{

0: default value, no special treatment for river cells

1: Scale secondary slopes along the river (Note this requries a river mask and you must set a river_secondaryTH if you want this to be something other than 0)

2: Apply watershed mean slope to each river reach (requires river mask and subbasins)

3: Apply the stream mean slope to each reach (requires river mask and subbasins)

NOTE: the river mask can be different from the rivers that were used to create the subbasins if desired (i.e. if you want to use a threshold of 100 to create subbasins but then apply to river cells with a threshold of 50)
}

