% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slope_Calc_Standard.R
\name{SlopeCalStan}
\alias{SlopeCalStan}
\title{Calculate the slopes from a DEM}
\usage{
SlopeCalStan(
  dem,
  direction,
  dx,
  dy,
  mask,
  d4 = c(1, 2, 3, 4),
  minslope = 0,
  maxslope = -1,
  secondaryTH = -1,
  printflag = F
)
}
\arguments{
\item{dem}{Elevation matrix}

\item{direction}{a matrix of the flow directions for cells that have been processed - defaults to all zeros}

\item{dx, dy}{Lateral grid cell resolution}

\item{mask}{Mask with ones for cells to be processed and zeros for everything else - defaults to a mask of all 1's}

\item{d4}{directional numbering system: the numbers you want to assigne to down, left, top,right (defaults to 1,2,3,4)}

\item{minslope}{Minimum absolute slope value to apply to flat cells if needed. Defaults to 1e-5}

\item{maxslope}{Maximum absolute value of slopes. If this is set to -1 the slopes will not be limited. Default value is -1}

\item{secondaryTH}{secondary threshold - maximum ratio of |secondary|/|primary| to be enforced. 
NOTE - this scaling occurs after any max threholds are applied. Currently this is only working for two options:
(1) If this is set to -1 no scaling will be applied, (2) If  this is set to zero all  seconeary slopes will be zero}

\item{printflag}{Print function progress}
}
\description{
This function will calculate slopes using standard or upwinding options and apply a range of smoothing options
}
\section{River Methods}{

0: default value, no special treatment for river cells

1: Scale secondary slopes along the river (Note this requries a river mask and you must set a river_secondaryTH if you want this to be something other than 0)

2: Apply watershed mean slope to each river reach (requires river mask and subbasins)

3: Apply the stream mean slope to each reach (requires river mask and subbasins)

NOTE: the river mask can be different from the rivers that were used to create the subbasins if desired (i.e. if you want to use a threshold of 100 to create subbasins but then apply to river cells with a threshold of 50)
}

