% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{small_vec_copy}
\alias{small_vec_copy}
\title{Example: Small data - regular copy is fine}
\source{
Generated by miniextendr from Rust fn `small_vec_copy`
}
\usage{
small_vec_copy()
}
\description{
Example: Small data - regular copy is fine
}
