% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ntsa.R
\name{smap_net}
\alias{smap_net}
\title{S-map with elastic net (a.k.a. Regularized S-map)}
\usage{
smap_net(
  X,
  col = 1,
  lib = matrix(c(1, nrow(X)), nrow = 1),
  pred = NULL,
  Tp = 1,
  threadNo = detectCores(),
  range = seq(0, 10, 1),
  seed = NULL,
  criterion = "rmse",
  s = "lambda.1se",
  lambda = NULL,
  alpha = 0
)
}
\arguments{
\item{X}{A matrix, which is constructed by the function \code{\link{gen_emat}}}

\item{col}{Integer, which specifies the position of prediction target in X}

\item{lib}{Vector or 2-column matrix, which sets the knot position in library data}

\item{pred}{Vector or 2-column matrix, which sets the knot position in test data}

\item{Tp}{Integer, which is the prediction-time horizon}

\item{threadNo}{Integer, which sets the thread number of parallel computing}

\item{range}{Numeric vector, which sets the range of nonlinear parameter theta for S-map}

\item{seed}{Integer, which sets the RNG seed}

\item{criterion}{String, which switch the cost function (Rho, MAE & RMSE)
for the dimension search (default = "rmse")}

\item{s}{String, which switches the criterion of the best theta in elastic net
(for more detail, see original function \code{\link[glmnet]{cv.glmnet}})}

\item{lambda}{Numeric vector, which sets the sequence of penalty strength for glmnet
if NULL (default), then the adaptive search with cv.glmnet is adopted
(for more detail, see original function \code{\link[glmnet]{cv.glmnet}})}

\item{alpha}{Numeric, which controls the Elastic net parameter
\code{alpha = 0} is the ridge penalty and \code{alpha = 1} is the lasso penalty
(for more detail, see original function \code{\link[glmnet]{cv.glmnet}})}
}
\description{
\code{smap_net} returns the (tibble) result of Regularized S-map
Algorithm is adopted from Cenci et al. (2019), MEE
}
