% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow.R
\name{snow_spells}
\alias{snow_spells}
\title{Maximal runs of consecutive TRUE values}
\usage{
snow_spells(mask)
}
\arguments{
\item{mask}{Logical vector.}
}
\value{
list of integer vectors (1-based index runs).
}
\description{
Maximal runs of consecutive TRUE values
}
\keyword{internal}
