% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{soft_deprecated_fn}
\alias{soft_deprecated_fn}
\title{soft_deprecated_fn}
\usage{
soft_deprecated_fn(x)
}
\arguments{
\item{x}{Integer input value.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A soft-deprecated function (warning on first use only).
}
