% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-dbplyr__duckdb_connection.R
\name{sql_escape_date}
\alias{sql_escape_date}
\title{Customized escape translation for date objects}
\arguments{
\item{con}{A \code{\link{dbConnect}} object, as returned by \code{dbConnect()}}

\item{x}{Date object to be escaped}
}
\description{
Customized escape translation for date objects
}
