% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-dbplyr__duckdb_connection.R
\name{sql_escape_datetime}
\alias{sql_escape_datetime}
\alias{sql_escape_datetime.duckdb_connection}
\title{Customized escape translation for datetime objects}
\usage{
sql_escape_datetime.duckdb_connection(con, x)
}
\arguments{
\item{con}{A \code{\link{dbConnect}} object, as returned by \code{dbConnect()}}

\item{x}{Datetime object to be escaped}
}
\description{
Customized escape translation for datetime objects
}
