% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-dbplyr__duckdb_connection.R
\name{sql_expr_matches}
\alias{sql_expr_matches}
\alias{sql_expr_matches.duckdb_connection}
\title{Customized translation for comparing to objects in DuckDB SQL}
\usage{
sql_expr_matches.duckdb_connection(con, x, y)
}
\arguments{
\item{con}{A \code{\link{dbConnect}} object, as returned by \code{dbConnect()}}

\item{x}{First object to be compared}

\item{y}{Second object to be compared}
}
\description{
Customized translation for comparing to objects in DuckDB SQL
}
