% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_analysis.R
\name{standardize_multinomial_columns}
\alias{standardize_multinomial_columns}
\title{Standardize multinomial results from list to data frame with consistent column structure}
\usage{
standardize_multinomial_columns(row_list, max_paths)
}
\arguments{
\item{row_list}{List of row-lists with variable multinomial columns}

\item{max_paths}{Maximum number of paths to standardize to}
}
\description{
Standardize multinomial results from list to data frame with consistent column structure
}
\examples{
row_list <- list(
  list(gene="GENE1", source="3", sink="7", ref_ex_part="E001",
       setdiff1="E002", transcripts1="TX1", path1="3-5-7",
       setdiff2="E003", transcripts2="TX2", path2="3-6-7"),
  list(gene="GENE1", source="3", sink="9", ref_ex_part="E004",
       setdiff1="E005", transcripts1="TX1", path1="3-8-9",
       setdiff2="",     transcripts2="TX3", path2="3-9")
)
df <- grase::standardize_multinomial_columns(row_list, max_paths = 2)
colnames(df)
}
