% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_choose_2_analysis.R
\name{standardize_pairwise_columns}
\alias{standardize_pairwise_columns}
\title{Convert pairwise analysis results to standardized data frame}
\usage{
standardize_pairwise_columns(pairwise_list)
}
\description{
Convert pairwise analysis results to standardized data frame
}
