% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{stem_page_docs}
\alias{stem_page_docs}
\alias{stem_scale}
\alias{stem_shift}
\alias{stem_floor}
\title{Values transformed in place}
\usage{
stem_scale(values, factor)

stem_shift(values, offset, direction = c("up", "down"))

stem_floor(values, floor)
}
\arguments{
\item{values}{Numbers to transform.}

\item{factor}{Multiplier applied to every value.}

\item{offset}{Amount added to every value.}

\item{direction}{One of "up", "down".}

\item{floor}{Smallest value kept.}
}
\description{
\code{stem_scale()} multiplies the values by a factor.

\code{stem_shift()} moves the values by an offset.

\code{stem_floor()} raises the values to at least a floor.
}
