% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stream_Distance.R
\name{StreamDist}
\alias{StreamDist}
\title{Find the distance to th nearest stream point following drainag directions}
\usage{
StreamDist(direction, streammask, domainmask, d4 = c(1, 2, 3, 4))
}
\arguments{
\item{direction}{a matrix of the flow directions for cells that have been processed - defaults to all zeros}

\item{streammask}{Mask with a value of 1 for every stream cell and 0 for all non stream cells. Refer to the
\code{\link{CalcSubbasins}} function to derive this mask.}

\item{domainmask}{Option mask of the domain area with a value of 1 for cells inside the domain and 0 for cells
outside the domain. If no maks is provided it will default to using the entire rectangular domain.}

\item{d4}{directional numbering system: the numbers you want to assigne to down, left, top,right (defaults to 1,2,3,4)}
}
\description{
This Function uses a stream mask and a flow direction file to determine the overland 
flow distance from any point in the domain to its nearest stream neigbor following
the definined primary flow directions.
}
