% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stream_Traverse.R
\name{StreamTraverse}
\alias{StreamTraverse}
\title{DEM processing of stream networks}
\usage{
StreamTraverse(
  dem,
  mask,
  queue,
  marked,
  step,
  direction,
  basins,
  d4 = c(1, 2, 3, 4),
  printstep = F,
  epsilon = 0
)
}
\arguments{
\item{dem}{Elevation matrix}

\item{mask}{Mask with zeros for non river cells and 1 for river cells}

\item{queue}{a priority queue to start from three columns, x, y, elevation}

\item{marked}{a matrix of which cells have been marked already}

\item{step}{a matrix of the step number for cells that have been processed - defaults to all zeros}

\item{direction}{a matrix of hte flow directions for cells that have been processed - defaults to all zeros}

\item{basins}{a matrix of basin numbers that can be created by the initilizaiton script. If you input this every cell will be assigned the same basin as the cell that adds it}

\item{d4}{directional numbering system: the numbers you want to assigne to down, left, top,right (defaults to 1,2,3,4)}

\item{printstep}{if true it will print out the step number and the size of the queue}

\item{epsilon}{amount to add to filled areas to avoid creating flats, defaults to zero}
}
\description{
Function to process stream networks walking upstream on d4 neigbors in a river mask. Where no D4 neigbhors exist it looks for d8 neigbors and creates d4 bridges to these diagonal cells
}
