% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{strict_echo_vec_i64}
\alias{strict_echo_vec_i64}
\title{strict_echo_vec_i64}
\usage{
strict_echo_vec_i64(x)
}
\arguments{
\item{x}{Integer or numeric vector (must be INTSXP or REALSXP in strict mode).}
}
\description{
Test strict-mode \verb{Vec<i64>} identity roundtrip.
}
