% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{strict_length}
\alias{strict_length}
\title{The \code{strict_length()} generic}
\arguments{
\item{x}{An object.}
}
\description{
\code{strict_length()} is an S7 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{S7Strict}}
}
}
