% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{strict_maybe_starts_none}
\alias{strict_maybe_starts_none}
\title{strict_maybe_starts_none}
\usage{
strict_maybe_starts_none()
}
\description{
\verb{Option<Vec<usize>>} returned under strict mode as \code{None}.
}
