% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{strict_maybe_starts_overflow}
\alias{strict_maybe_starts_overflow}
\title{strict_maybe_starts_overflow}
\usage{
strict_maybe_starts_overflow()
}
\description{
\verb{Option<Vec<usize>>} returned under strict mode with a value past \code{i32::MAX}.
}
