% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string and return a character vector
A simplified wrapper around base R's function strsplit() that returns the first element.}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character string to split.}

\item{split}{A character string containing the regular expression to use for splitting.}
}
\value{
A character vector of the split pieces.
}
\description{
Split a string and return a character vector
A simplified wrapper around base R's function strsplit() that returns the first element.
}
\examples{
strsplit1("alfa,bravo,charlie,delta", split = ",")
}
