% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{struct_flatten_align_nvnr}
\alias{struct_flatten_align_nvnr}
\title{struct_flatten_align_nvnr}
\usage{
struct_flatten_align_nvnr()
}
\description{
NvNr align: aligned data frame with NA-fill for absent variant.
}
