% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{struct_flatten_split_1vnr}
\alias{struct_flatten_split_1vnr}
\title{struct_flatten_split_1vnr}
\usage{
struct_flatten_split_1vnr()
}
\description{
1vNr split: multiple Located rows, all same variant.
}
