% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{struct_flatten_split_nvnr}
\alias{struct_flatten_split_nvnr}
\title{struct_flatten_split_nvnr}
\usage{
struct_flatten_split_nvnr()
}
\description{
NvNr split: multiple rows across both variants.
}
