% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{struct_list_align_nvnr}
\alias{struct_list_align_nvnr}
\title{struct_list_align_nvnr}
\usage{
struct_list_align_nvnr()
}
\description{
NvNr align (as_list): aligned data frame, origin as list-column.
}
