% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{struct_list_split_1v1r}
\alias{struct_list_split_1v1r}
\title{struct_list_split_1v1r}
\usage{
struct_list_split_1v1r()
}
\description{
1v1r split (as_list): single Located row, origin as list-column.
}
