% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{struct_list_split_nvnr}
\alias{struct_list_split_nvnr}
\title{struct_list_split_nvnr}
\usage{
struct_list_split_nvnr()
}
\description{
NvNr split (as_list): multiple rows across both variants.
}
