% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{summary}
\alias{.summary}
\alias{summary}
\alias{summary,fm-method}
\alias{summary,fmV-method}
\title{FlashR Summaries}
\usage{
.summary(x)

\S4method{summary}{fm}(object, ...)

\S4method{summary}{fmV}(object, ...)
}
\arguments{
\item{x}{a FlashR vector or matrix.}
}
\value{
A list containing the following named components:
\itemize{
\item{min}{The minimum value}
\item{max}{The maximum value}
\item{mean}{The average}
\item{normL1}{The L1 norm}
\item{normL2}{The L2 norm}
\item{numNonzeros}{The number of non-zero values}
}
}
\description{
\code{summary} produces summaries of a FlashR object.
}
\details{
It computes the min, max, sum, mean, L1, L2, number of non-zero values if
the argument is a vector, or these statistics for each column if the argument
is a matrix.
}
\examples{
mat <- fm.runif.matrix(100, 10)
summary(mat)
}

