% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{superseded_fn}
\alias{superseded_fn}
\title{superseded_fn}
\usage{
superseded_fn(x)
}
\arguments{
\item{x}{Integer input value.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}} A superseded function with a better alternative.
}
