% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{sweep}
\alias{sweep}
\title{Sweep out Array Summaries}
\usage{
\S4method{sweep}{fm}(x, MARGIN, STATS, FUN = "-", check.margin = TRUE, ...)
}
\arguments{
\item{x}{a FlashR matrix.}

\item{MARGIN}{an integer giving the extent of \code{x} which
correspond to \code{STATS}.}

\item{STATS}{the summary statistic which is to be swept out.}

\item{FUN}{the function to be used to carry out the sweep.}

\item{check.margin}{logical. It's ignored right now.}

\item{...}{optional arguments to \code{FUN}.}
}
\value{
A matrix with the same shape as \code{x}, but with the summary
statistics swept out.
}
\description{
Return an array obtained from an input array by sweeping out a
summary statistic.
}
\examples{
mat <- fm.runif.matrix(100, 10)
sweep(mat, 1, runif(100))
}

