% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{sync_channel}
\alias{sync_channel}
\title{Sync all threads in a channel}
\usage{
sync_channel(
  channel,
  token = twist_token(),
  workspace_dir = twist_workspace_dir(),
  options = list()
)
}
\arguments{
\item{channel}{Channel object or ID}

\item{token}{Authentication token}

\item{workspace_dir}{Directory containing workspace data}

\item{options}{List of options (see sync_workspace)}
}
\value{
List of updated thread files
}
\description{
Sync all threads in a channel
}
