% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{t,fm-method}
\alias{t,fm-method}
\title{Matrix Transpose}
\usage{
\S4method{t}{fm}(x)
}
\arguments{
\item{x}{a FlashMatrix matrix.}
}
\value{
a matrix.
}
\description{
Given a matrix \code{x}, \code{t} returns the transpose of \code{x}.
}

