% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{tabled_compact}
\alias{tabled_compact}
\title{tabled_compact}
\usage{
tabled_compact(col1, col2)
}
\arguments{
\item{col1}{First column values.}

\item{col2}{Second column values.}
}
\description{
Create a compact table from raw string vectors.
}
