% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{tabled_many_columns}
\alias{tabled_many_columns}
\title{tabled_many_columns}
\usage{
tabled_many_columns(headers)
}
\arguments{
\item{headers}{Character vector of column headers.}
}
\description{
Test creating a table with many columns.
}
