% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{tabled_to_string_empty}
\alias{tabled_to_string_empty}
\title{tabled_to_string_empty}
\usage{
tabled_to_string_empty()
}
\description{
\code{table_to_string} over an empty slice (headers only).
}
