% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{tabled_with_max_width}
\alias{tabled_with_max_width}
\title{tabled_with_max_width}
\usage{
tabled_with_max_width(max)
}
\arguments{
\item{max}{Maximum column width.}
}
\description{
Truncate columns to a maximum width.
}
