% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{take_and_return_nothing}
\alias{take_and_return_nothing}
\title{take_and_return_nothing}
\usage{
take_and_return_nothing()
}
\description{
Test that a function returning unit () works correctly (invisible NULL in R).
}
