% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{telemetry_install_counter}
\alias{telemetry_install_counter}
\title{telemetry_install_counter}
\usage{
telemetry_install_counter()
}
\description{
Install a telemetry hook that counts panics.
}
