% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_collect_empty}
\alias{test_collect_empty}
\title{test collect empty}
\source{
Generated by miniextendr from Rust fn \code{test_collect_empty}
}
\usage{
test_collect_empty()
}
\description{
test collect empty
}
