% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_collect_na_f64}
\alias{test_collect_na_f64}
\title{test collect na f64}
\source{
Generated by miniextendr from Rust fn \code{test_collect_na_f64}
}
\usage{
test_collect_na_f64(n)
}
\description{
test collect na f64
}
