% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_collect_na_i32}
\alias{test_collect_na_i32}
\title{test collect na i32}
\source{
Generated by miniextendr from Rust fn \code{test_collect_na_i32}
}
\usage{
test_collect_na_i32(n)
}
\description{
test collect na i32
}
