% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_collect_range}
\alias{test_collect_range}
\title{test collect range}
\source{
Generated by miniextendr from Rust fn \code{test_collect_range}
}
\usage{
test_collect_range()
}
\description{
test collect range
}
