% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_collect_sines}
\alias{test_collect_sines}
\title{test collect sines}
\source{
Generated by miniextendr from Rust fn \code{test_collect_sines}
}
\usage{
test_collect_sines(n)
}
\description{
test collect sines
}
