% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_collect_squares}
\alias{test_collect_squares}
\title{test collect squares}
\source{
Generated by miniextendr from Rust fn \code{test_collect_squares}
}
\usage{
test_collect_squares(n)
}
\arguments{
\item{n}{(no documentation available)}
}
\description{
test collect squares
}
