% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_collect_strings_upper}
\alias{test_collect_strings_upper}
\title{test collect strings upper}
\source{
Generated by miniextendr from Rust fn \code{test_collect_strings_upper}
}
\usage{
test_collect_strings_upper(words)
}
\description{
test collect strings upper
}
