% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_columnar_deep_nesting}
\alias{test_columnar_deep_nesting}
\title{test columnar deep nesting}
\source{
Generated by miniextendr from Rust fn \code{test_columnar_deep_nesting}
}
\usage{
test_columnar_deep_nesting()
}
\description{
test columnar deep nesting
}
