% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_columnar_drop}
\alias{test_columnar_drop}
\title{test columnar drop}
\source{
Generated by miniextendr from Rust fn \code{test_columnar_drop}
}
\usage{
test_columnar_drop()
}
\description{
test columnar drop
}
