% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_columnar_empty}
\alias{test_columnar_empty}
\title{test columnar empty}
\source{
Generated by miniextendr from Rust fn \code{test_columnar_empty}
}
\usage{
test_columnar_empty()
}
\description{
test columnar empty
}
