% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_columnar_nested}
\alias{test_columnar_nested}
\title{test columnar nested}
\source{
Generated by miniextendr from Rust fn \code{test_columnar_nested}
}
\usage{
test_columnar_nested()
}
\description{
test columnar nested
}
