% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_columnar_optional_struct}
\alias{test_columnar_optional_struct}
\title{test columnar optional struct}
\source{
Generated by miniextendr from Rust fn \code{test_columnar_optional_struct}
}
\usage{
test_columnar_optional_struct()
}
\description{
test columnar optional struct
}
