% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_columnar_rename}
\alias{test_columnar_rename}
\title{test columnar rename}
\source{
Generated by miniextendr from Rust fn \code{test_columnar_rename}
}
\usage{
test_columnar_rename()
}
\description{
test columnar rename
}
