% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_columnar_rename_noop}
\alias{test_columnar_rename_noop}
\title{test columnar rename noop}
\source{
Generated by miniextendr from Rust fn \code{test_columnar_rename_noop}
}
\usage{
test_columnar_rename_noop()
}
\description{
test columnar rename noop
}
