% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_columnar_select}
\alias{test_columnar_select}
\title{test columnar select}
\source{
Generated by miniextendr from Rust fn \code{test_columnar_select}
}
\usage{
test_columnar_select()
}
\description{
test columnar select
}
