% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_columnar_serde_flatten}
\alias{test_columnar_serde_flatten}
\title{test columnar serde flatten}
\source{
Generated by miniextendr from Rust fn \code{test_columnar_serde_flatten}
}
\usage{
test_columnar_serde_flatten()
}
\description{
test columnar serde flatten
}
