% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_columnar_skip_serializing_if}
\alias{test_columnar_skip_serializing_if}
\title{test columnar skip serializing if}
\source{
Generated by miniextendr from Rust fn \code{test_columnar_skip_serializing_if}
}
\usage{
test_columnar_skip_serializing_if()
}
\description{
test columnar skip serializing if
}
