% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_columnar_strip_prefix}
\alias{test_columnar_strip_prefix}
\title{strip_prefix: remove "point_" from column names}
\source{
Generated by miniextendr from Rust fn \code{test_columnar_strip_prefix}
}
\usage{
test_columnar_strip_prefix()
}
\description{
strip_prefix: remove "point_" from column names
}
