% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_columnar_tagged_enum}
\alias{test_columnar_tagged_enum}
\title{Internally tagged enum: "kind" column acts as discriminator}
\source{
Generated by miniextendr from Rust fn \code{test_columnar_tagged_enum}
}
\usage{
test_columnar_tagged_enum()
}
\description{
Internally tagged enum: "kind" column acts as discriminator
}
