% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_columnar_untagged_enum}
\alias{test_columnar_untagged_enum}
\title{Untagged enum: Ok rows have status+size, Err rows have error}
\source{
Generated by miniextendr from Rust fn \code{test_columnar_untagged_enum}
}
\usage{
test_columnar_untagged_enum()
}
\description{
Untagged enum: Ok rows have status+size, Err rows have error
}
