% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_condition_chained}
\alias{test_condition_chained}
\title{test condition chained}
\source{
Generated by miniextendr from Rust fn \code{test_condition_chained}
}
\usage{
test_condition_chained(s)
}
\arguments{
\item{s}{(no documentation available)}
}
\description{
test condition chained
}
