% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_condition_ok}
\alias{test_condition_ok}
\title{test condition ok}
\source{
Generated by miniextendr from Rust fn \code{test_condition_ok}
}
\usage{
test_condition_ok()
}
\description{
test condition ok
}
