% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_condition_parse_int}
\alias{test_condition_parse_int}
\title{test condition parse int}
\source{
Generated by miniextendr from Rust fn \code{test_condition_parse_int}
}
\usage{
test_condition_parse_int(s)
}
\description{
test condition parse int
}
