% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_df_aggregate}
\alias{test_df_aggregate}
\title{test df aggregate}
\source{
Generated by miniextendr from Rust fn \code{test_df_aggregate}
}
\usage{
test_df_aggregate(df)
}
\description{
test df aggregate
}
