% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_df_columns}
\alias{test_df_columns}
\title{test df columns}
\source{
Generated by miniextendr from Rust fn \code{test_df_columns}
}
\usage{
test_df_columns(df)
}
\arguments{
\item{df}{(no documentation available)}
}
\description{
test df columns
}
