% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_df_global_agg}
\alias{test_df_global_agg}
\title{test df global agg}
\source{
Generated by miniextendr from Rust fn \code{test_df_global_agg}
}
\usage{
test_df_global_agg(df)
}
\description{
test df global agg
}
