% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_df_select}
\alias{test_df_select}
\title{test df select}
\source{
Generated by miniextendr from Rust fn \code{test_df_select}
}
\usage{
test_df_select(df, cols)
}
\description{
test df select
}
