% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_df_sort_limit}
\alias{test_df_sort_limit}
\title{test df sort limit}
\source{
Generated by miniextendr from Rust fn \code{test_df_sort_limit}
}
\usage{
test_df_sort_limit(df, col, asc, n)
}
\arguments{
\item{df}{(no documentation available)}

\item{col}{(no documentation available)}

\item{asc}{(no documentation available)}

\item{n}{(no documentation available)}
}
\description{
test df sort limit
}
