% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_display_bool}
\alias{test_display_bool}
\title{test display bool}
\source{
Generated by miniextendr from Rust fn \code{test_display_bool}
}
\usage{
test_display_bool()
}
\description{
test display bool
}
