% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_display_ip}
\alias{test_display_ip}
\title{test display ip}
\source{
Generated by miniextendr from Rust fn \code{test_display_ip}
}
\usage{
test_display_ip()
}
\description{
test display ip
}
