% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_display_number}
\alias{test_display_number}
\title{test display number}
\source{
Generated by miniextendr from Rust fn \code{test_display_number}
}
\usage{
test_display_number()
}
\description{
test display number
}
