% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_display_vec_ints}
\alias{test_display_vec_ints}
\title{test display vec ints}
\source{
Generated by miniextendr from Rust fn \code{test_display_vec_ints}
}
\usage{
test_display_vec_ints()
}
\description{
test display vec ints
}
