% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_display_vec_ips}
\alias{test_display_vec_ips}
\title{test display vec ips}
\source{
Generated by miniextendr from Rust fn \code{test_display_vec_ips}
}
\usage{
test_display_vec_ips()
}
\description{
test display vec ips
}
