% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_fromjson_bad}
\alias{test_fromjson_bad}
\title{test fromjson bad}
\source{
Generated by miniextendr from Rust fn \code{test_fromjson_bad}
}
\usage{
test_fromjson_bad(json)
}
\description{
test fromjson bad
}
