% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_fromjson_config}
\alias{test_fromjson_config}
\title{test fromjson config}
\source{
Generated by miniextendr from Rust fn \code{test_fromjson_config}
}
\usage{
test_fromjson_config(json)
}
\arguments{
\item{json}{(no documentation available)}
}
\description{
test fromjson config
}
