% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_fromjson_point_sum}
\alias{test_fromjson_point_sum}
\title{test fromjson point sum}
\source{
Generated by miniextendr from Rust fn \code{test_fromjson_point_sum}
}
\usage{
test_fromjson_point_sum(json)
}
\arguments{
\item{json}{(no documentation available)}
}
\description{
test fromjson point sum
}
