% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_fromstr_bad_input}
\alias{test_fromstr_bad_input}
\title{test fromstr bad input}
\source{
Generated by miniextendr from Rust fn \code{test_fromstr_bad_input}
}
\usage{
test_fromstr_bad_input(s)
}
\arguments{
\item{s}{(no documentation available)}
}
\description{
test fromstr bad input
}
