% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_fromstr_int}
\alias{test_fromstr_int}
\title{test fromstr int}
\source{
Generated by miniextendr from Rust fn \code{test_fromstr_int}
}
\usage{
test_fromstr_int(s)
}
\description{
test fromstr int
}
