% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_fromstr_vec_ints}
\alias{test_fromstr_vec_ints}
\title{test fromstr vec ints}
\source{
Generated by miniextendr from Rust fn \code{test_fromstr_vec_ints}
}
\usage{
test_fromstr_vec_ints(nums)
}
\description{
test fromstr vec ints
}
