% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_fromstr_vec_ips}
\alias{test_fromstr_vec_ips}
\title{test fromstr vec ips}
\source{
Generated by miniextendr from Rust fn \code{test_fromstr_vec_ips}
}
\usage{
test_fromstr_vec_ips(addrs)
}
\description{
test fromstr vec ips
}
