% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_json_config}
\alias{test_json_config}
\title{test json config}
\source{
Generated by miniextendr from Rust fn \code{test_json_config}
}
\usage{
test_json_config()
}
\description{
test json config
}
