% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_json_pretty_point}
\alias{test_json_pretty_point}
\title{test json pretty point}
\source{
Generated by miniextendr from Rust fn \code{test_json_pretty_point}
}
\usage{
test_json_pretty_point()
}
\description{
test json pretty point
}
