% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_json_vec_points}
\alias{test_json_vec_points}
\title{test json vec points}
\source{
Generated by miniextendr from Rust fn \code{test_json_vec_points}
}
\usage{
test_json_vec_points()
}
\description{
test json vec points
}
