% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_lazy_arrow_bool}
\alias{test_lazy_arrow_bool}
\title{test lazy arrow bool}
\source{
Generated by miniextendr from Rust fn \code{test_lazy_arrow_bool}
}
\usage{
test_lazy_arrow_bool(n)
}
\arguments{
\item{n}{(no documentation available)}
}
\description{
test lazy arrow bool
}
