% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_lazy_arrow_string}
\alias{test_lazy_arrow_string}
\title{test lazy arrow string}
\source{
Generated by miniextendr from Rust fn \code{test_lazy_arrow_string}
}
\usage{
test_lazy_arrow_string()
}
\description{
test lazy arrow string
}
