% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_lazy_nalgebra_f64}
\alias{test_lazy_nalgebra_f64}
\title{test lazy nalgebra f64}
\source{
Generated by miniextendr from Rust fn \code{test_lazy_nalgebra_f64}
}
\usage{
test_lazy_nalgebra_f64(n)
}
\arguments{
\item{n}{(no documentation available)}
}
\description{
test lazy nalgebra f64
}
