% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_lazy_nalgebra_i32}
\alias{test_lazy_nalgebra_i32}
\title{test lazy nalgebra i32}
\source{
Generated by miniextendr from Rust fn \code{test_lazy_nalgebra_i32}
}
\usage{
test_lazy_nalgebra_i32(n)
}
\description{
test lazy nalgebra i32
}
