% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_log_debug}
\alias{test_log_debug}
\title{test log debug}
\source{
Generated by miniextendr from Rust fn \code{test_log_debug}
}
\usage{
test_log_debug(msg)
}
\description{
test log debug
}
