% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_log_error}
\alias{test_log_error}
\title{test log error}
\source{
Generated by miniextendr from Rust fn \code{test_log_error}
}
\usage{
test_log_error(msg)
}
\description{
test log error
}
