% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_log_info}
\alias{test_log_info}
\title{test log info}
\source{
Generated by miniextendr from Rust fn \code{test_log_info}
}
\usage{
test_log_info(msg)
}
\arguments{
\item{msg}{(no documentation available)}
}
\description{
test log info
}
