% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_log_set_level}
\alias{test_log_set_level}
\title{test log set level}
\source{
Generated by miniextendr from Rust fn \code{test_log_set_level}
}
\usage{
test_log_set_level(level)
}
\description{
test log set level
}
