% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_log_warn}
\alias{test_log_warn}
\title{test log warn}
\source{
Generated by miniextendr from Rust fn \code{test_log_warn}
}
\usage{
test_log_warn(msg)
}
\arguments{
\item{msg}{(no documentation available)}
}
\description{
test log warn
}
