% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exontest_functions.R
\name{test_model_glmmTMB_without_prior}
\alias{test_model_glmmTMB_without_prior}
\title{Test differential exon usage with a beta-binomial glmmTMB model, no prior on dispersion.
Intended for comparison with \code{test_model_glmmTMB_EB}.}
\usage{
test_model_glmmTMB_without_prior(dd)
}
\arguments{
\item{dd}{A data frame of exon count data.}
}
\description{
Test differential exon usage with a beta-binomial glmmTMB model, no prior on dispersion.
Intended for comparison with \code{test_model_glmmTMB_EB}.
}
\examples{
\dontrun{
splitcnts <- read.table("bipartition.internal.exoncnt.combined.txt",
                        header = TRUE, row.names = NULL)
splitcnts$groups <- factor(splitcnts$groups)
grouped_data <- grase::group_by_event(splitcnts, "diff", "n")
result <- grase::test_model_glmmTMB_without_prior(grouped_data[[1]])
result
}
}
