% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exontest_functions.R
\name{test_model_multinomial_vgam_EB}
\alias{test_model_multinomial_vgam_EB}
\title{Test differential exon usage with a VGAM Dirichlet-multinomial model using empirical Bayes fixed precision.}
\usage{
test_model_multinomial_vgam_EB(dd)
}
\arguments{
\item{dd}{A data frame of exon count data.}
}
\description{
Test differential exon usage with a VGAM Dirichlet-multinomial model using empirical Bayes fixed precision.
}
\examples{
\dontrun{
splitcnts <- read.table("multinomial.exoncnt.combined.txt",
                        header = TRUE, row.names = NULL)
prec_table <- grase::moderate_prec_log_scale(
  read.table("prec_dm.txt", header = TRUE, row.names = NULL)
)
splitcnts_eb <- dplyr::left_join(splitcnts, prec_table, by = c("gene", "event"))
grouped_eb <- dplyr::group_split(dplyr::group_by(splitcnts_eb, gene, event))
result <- grase::test_model_multinomial_vgam_EB(grouped_eb[[1]])
result
}
}
