% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exontest_functions.R
\name{test_model_wilcoxon}
\alias{test_model_wilcoxon}
\title{Test for differential exon usage using Wilcoxon rank-sum test}
\usage{
test_model_wilcoxon(dd)
}
\arguments{
\item{dd}{A data frame of exon count data.}
}
\description{
Test for differential exon usage using Wilcoxon rank-sum test
}
\examples{
\dontrun{
splitcnts <- read.table("bipartition.internal.exoncnt.combined.txt",
                        header = TRUE, row.names = NULL)
splitcnts$groups <- factor(splitcnts$groups)
grouped_data <- grase::group_by_event(splitcnts, "diff", "n")
result <- grase::test_model_wilcoxon(grouped_data[[1]])
result
}
}
