% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_new_list_of_ptype}
\alias{test_new_list_of_ptype}
\title{Test new_list_of with a ptype}
\source{
Generated by miniextendr from Rust fn \code{test_new_list_of_ptype}
}
\usage{
test_new_list_of_ptype(x, ptype, class)
}
\arguments{
\item{x}{(no documentation available)}

\item{ptype}{(no documentation available)}

\item{class}{(no documentation available)}
}
\description{
Test new_list_of with a ptype
}
