% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_new_list_of_size}
\alias{test_new_list_of_size}
\title{Test new_list_of with a size}
\source{
Generated by miniextendr from Rust fn \code{test_new_list_of_size}
}
\usage{
test_new_list_of_size(x, size, class)
}
\arguments{
\item{x}{(no documentation available)}

\item{size}{(no documentation available)}

\item{class}{(no documentation available)}
}
\description{
Test new_list_of with a size
}
