% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_new_rcrd}
\alias{test_new_rcrd}
\title{Test new_rcrd with the given field list and class name(s)}
\source{
Generated by miniextendr from Rust fn \code{test_new_rcrd}
}
\usage{
test_new_rcrd(fields, class)
}
\arguments{
\item{fields}{(no documentation available)}

\item{class}{(no documentation available)}
}
\description{
Test new_rcrd with the given field list and class name(s)
}
