% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_new_vctr}
\alias{test_new_vctr}
\title{Test new_vctr with default inherit_base_type}
\source{
Generated by miniextendr from Rust fn \code{test_new_vctr}
}
\usage{
test_new_vctr(data, class)
}
\arguments{
\item{data}{(no documentation available)}

\item{class}{(no documentation available)}
}
\description{
Test new_vctr with default inherit_base_type
}
