% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_new_vctr_inherit}
\alias{test_new_vctr_inherit}
\title{Test new_vctr with explicit inherit_base_type}
\source{
Generated by miniextendr from Rust fn \code{test_new_vctr_inherit}
}
\usage{
test_new_vctr_inherit(data, class, inherit_base_type)
}
\arguments{
\item{data}{(no documentation available)}

\item{class}{(no documentation available)}

\item{inherit_base_type}{(no documentation available)}
}
\description{
Test new_vctr with explicit inherit_base_type
}
