% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_vctrs_build_error_message}
\alias{test_vctrs_build_error_message}
\title{Get the VctrsBuildError message for a specific error type}
\source{
Generated by miniextendr from Rust fn \code{test_vctrs_build_error_message}
}
\usage{
test_vctrs_build_error_message(error_type)
}
\arguments{
\item{error_type}{(no documentation available)}
}
\description{
Get the VctrsBuildError message for a specific error type
}
