% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread.R
\name{thread_file_name}
\alias{thread_file_name}
\title{Create a file name for a thread}
\usage{
thread_file_name(thread)
}
\arguments{
\item{thread}{Thread returned by \code{get_thread()}}
}
\value{
String to use as file name
}
\description{
Create a file name for a thread
}
