% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rpkg_thread_builder}
\alias{rpkg_thread_builder}
\alias{unsafe_C_test_r_thread_builder}
\alias{unsafe_C_test_r_thread_builder_spawn_join}
\title{Test RThreadBuilder spawn with explicit stack size and thread name}
\source{
Generated by miniextendr from Rust fn \code{C_test_r_thread_builder}

Generated by miniextendr from Rust fn \code{C_test_r_thread_builder_spawn_join}
}
\usage{
unsafe_C_test_r_thread_builder()

unsafe_C_test_r_thread_builder_spawn_join()
}
\description{
Test RThreadBuilder spawn with explicit stack size and thread name

thread tests
}
\examples{
\dontrun{
unsafe_C_test_r_thread_builder()
unsafe_C_test_r_thread_builder_spawn_join()
unsafe_C_test_spawn_with_r_lean_stack()
unsafe_C_test_stack_check_guard_lean()
}
}
