% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeppFit.R
\name{threeppFit}
\alias{threeppFit}
\title{Fits a three pool model with parallel structure to a time-series}
\usage{
threeppFit(timeSeries, initialCarbon, inipars = c(1, 0.5, 0.5, 0.5, 0.5))
}
\arguments{
\item{timeSeries}{A time series of respiration values over time}

\item{initialCarbon}{The initial amount of carbon in units that correspond to the time series data}

\item{inipars}{vector of parameter values for the initial search of the optimization algorithm}
}
\value{
R list with an FME model object, a SoilR model object, and the AIC value
}
\description{
Fits a three pool model with parallel structure to a time-series
}
\examples{
\donttest{
incubation=sidb[["Crow2019a"]]
a=threeppFit(timeSeries = incubation$timeSeries[,c(1,79)],
initialCarbon=incubation$initConditions[78,"carbonMean"]*10000,
inipars = c(0.05, 0.01, 0.001, 0.1, 0.1))
}
}
