% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.DGEList}
\alias{tidy.DGEList}
\title{Converts a DGEList into a tidy data.frame}
\usage{
\method{tidy}{DGEList}(x, normalized.lib.sizes = TRUE, prior.count = 3, ...)
}
\description{
This makes downstream exploratory data analysis a bit easier.
}
