% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{time_epoch_posixct}
\alias{time_epoch_posixct}
\title{time_epoch_posixct}
\usage{
time_epoch_posixct()
}
\description{
Test creating and roundtripping the epoch datetime (1970-01-01 00:00:00 UTC).
}
