% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{time_format_date}
\alias{time_format_date}
\title{time_format_date}
\usage{
time_format_date(date)
}
\arguments{
\item{date}{Date from R.}
}
\description{
Test formatting a Date as a YYYY-MM-DD string.
}
