% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dates.R
\name{time_now_utc}
\alias{time_now_utc}
\title{Create a UTC Timestamp}
\usage{
time_now_utc()
}
\arguments{
\item{time}{POSIXct, POSIXlt, Date, chron date-time object}
}
\value{
a POSIXct object in the updated time zone
}
\description{
Create a UTC Timestamp
}
