% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{time_roundtrip_date}
\alias{time_roundtrip_date}
\title{time_roundtrip_date}
\usage{
time_roundtrip_date(date)
}
\arguments{
\item{date}{Date from R.}
}
\description{
Test Date roundtrip through R Date.
}
