% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{time_roundtrip_posixct}
\alias{time_roundtrip_posixct}
\title{time_roundtrip_posixct}
\usage{
time_roundtrip_posixct(dt)
}
\arguments{
\item{dt}{POSIXct datetime from R.}
}
\description{
Test OffsetDateTime roundtrip through R POSIXct.
}
