% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{tinyvec_roundtrip_dbl}
\alias{tinyvec_roundtrip_dbl}
\title{tinyvec_roundtrip_dbl}
\usage{
tinyvec_roundtrip_dbl(v)
}
\arguments{
\item{v}{Numeric vector from R.}
}
\description{
Test \verb{TinyVec<[f64; 8]>} roundtrip through R.
}
