% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{tinyvec_roundtrip_int}
\alias{tinyvec_roundtrip_int}
\title{tinyvec_roundtrip_int}
\usage{
tinyvec_roundtrip_int(v)
}
\arguments{
\item{v}{Integer vector from R.}
}
\description{
Test \verb{TinyVec<[i32; 8]>} roundtrip through R.
}
