% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{to_celsius}
\alias{to_celsius}
\title{The \code{to_celsius()} generic}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{to_celsius()} is an S7 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{S7Fahrenheit}}
}
}
