% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dummies.R
\name{to_dummies}
\alias{to_dummies}
\title{Convert column to dummy variables}
\usage{
to_dummies(.data, col, contrasts = NULL, remove = TRUE,
  drop_level = FALSE, sep = "_")
}
\arguments{
\item{.data}{A data frame}

\item{col}{The bare column name}

\item{contrasts}{A character string naming a function or a numeric matrix,
to be used as a replacement value for the \code{\link{contrasts}} replacement
function.}

\item{remove}{If \code{TRUE}, then remove the input column from the output
data frame.}

\item{drop_level}{If \code{TRUE}, then drop a level when creating the variables.
This is the default for models.}

\item{sep}{\code{character} string to use as the separator in the new column
names.}
}
\value{
A data frame.
}
\description{
\code{to_dummies} converts a column into multiple columns of dummy variables.
}
\examples{
example_df <- data.frame(x = c(rep("a", 2), rep("b", 2)))
to_dummies(example_df, x)
to_dummies(example_df, x, remove = FALSE)
to_dummies(example_df, x, drop_level = TRUE)
}

