% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dummies.R
\name{to_dummies_}
\alias{to_dummies_}
\title{Standard evaluation version of \code{to_dummies}.}
\usage{
to_dummies_(.data, col, contrasts = NULL, remove = TRUE,
  drop_level = FALSE, sep = "_")
}
\arguments{
\item{.data}{A data frame}

\item{col}{The bare column name}

\item{contrasts}{A character string naming a function or a numeric matrix,
to be used as a replacement value for the \code{\link{contrasts}} replacement
function.}

\item{remove}{If \code{TRUE}, then remove the input column from the output
data frame.}

\item{drop_level}{If \code{TRUE}, then drop a level when creating the variables.
This is the default for models.}

\item{sep}{\code{character} string to use as the separator in the new column
names.}
}
\description{
This is an S3 generic.
}

