% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{to_source}
\alias{to_source}
\title{The \code{to_source()} generic}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{to_source()} is an S7 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{TypedConversion}}
\item \code{\link{AttributeConversion}}
}
}
