% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{toml_array_count}
\alias{toml_array_count}
\title{toml_array_count}
\usage{
toml_array_count(input, key)
}
\arguments{
\item{input}{TOML string containing an array of tables.}

\item{key}{Key name of the array of tables.}
}
\description{
Test counting entries in a TOML array of tables.
}
