% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{toml_pretty}
\alias{toml_pretty}
\title{toml_pretty}
\usage{
toml_pretty(input)
}
\arguments{
\item{input}{TOML string to parse and pretty-print.}
}
\description{
Test pretty-printing a TOML string.
}
