% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{toml_type_name}
\alias{toml_type_name}
\title{toml_type_name}
\usage{
toml_type_name(input)
}
\arguments{
\item{input}{TOML value from R.}
}
\description{
Test getting the TOML type name of a value.
}
