% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{transcripts_from_igraph}
\alias{transcripts_from_igraph}
\title{get transcript names from igraph edge attributes}
\usage{
transcripts_from_igraph(g)
}
\description{
get transcript names from igraph edge attributes
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
tx_names <- grase::transcripts_from_igraph(g)
tx_names
}
}
