% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{transpose}
\alias{t,fm-method}
\alias{t,fmV-method}
\alias{transpose}
\title{Matrix Transpose}
\usage{
\S4method{t}{fm}(x)

\S4method{t}{fmV}(x)
}
\arguments{
\item{x}{a FlashR matrix.}
}
\value{
a matrix.
}
\description{
Given a matrix \code{x}, \code{t} returns the transpose of \code{x}.
}
\examples{
mat <- fm.runif.matrix(100, 10)
t(mat)
}

