% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentNames.R
\name{treatmentNames}
\alias{treatmentNames}
\title{Returns names of all treatments found in supplied database}
\usage{
treatmentNames(database)
}
\arguments{
\item{database}{a list with the SIDb structure}
}
\value{
list of unique treatment names
}
\description{
Returns names of all treatments found in supplied database
}
\details{
This function looks for unique field names in the variables and
 incubationInfo lists in an object with a SIDb-like structure, e.g.
 \link[sidb]{sidb}. Fields that are not considered "treatments" are excluded.
  Specifically: "name", "varDesc", "site", "gasMeasured", "units", and
 "primaryVariable".
}
\examples{
# return temperature levels from all entries
treatmentLevels(sidb, "temperature")
# return moisture levels from all entries
treatmentLevels(sidb, "moisture")
# return moisture levels from all entries with units "percentFieldCapacity"
treatmentLevels(sidb, "moisture", "percentFieldCapacity")
}
