% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentUnitNames.R
\name{treatmentUnitNames}
\alias{treatmentUnitNames}
\title{Returns names of units associated with specified treatment}
\usage{
treatmentUnitNames(database, treatment)
}
\arguments{
\item{database}{a list with the SIDb structure}

\item{treatment}{character vector of the treatment of interest. Must be in quotes.}
}
\value{
list of unique treatment names
}
\description{
Returns names of units associated with specified treatment
}
\examples{
# return moisture units from all entries
treatmentUnitNames(sidb, "moisture")
# return units of glucose additions from all entries
treatmentUnitNames(sidb, "glucose")
}
