% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{try_amount}
\alias{try_amount}
\title{Add an amount, rejecting negative values}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Add an amount, rejecting negative values
}
