% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_surrogate.R
\name{twin_probs}
\alias{twin_probs}
\title{Calculate twin probabilities}
\usage{
twin_probs(twins, dmat, theta = 1, twin_op)
}
\arguments{
\item{twins}{A list, that preserves twins for each state (calculated by \code{\link{find_twins}})}

\item{dmat}{A distance matrix}

\item{theta}{Numeric, the locality parameter (works if twin_op == "NN")}

\item{twin_op}{Character, which can be "RM" or "NN"
if "RM", twin search is based on recurrence matrix,
otherwise is based on NN twin method}
}
\description{
\code{twin_probs} returns a list of swapping probabilities of twins for each state
}
