% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{twist_request}
\alias{twist_request}
\title{Make API request to Twist}
\usage{
twist_request(endpoint, params = list(), token = twist_token(), method = "GET")
}
\arguments{
\item{params}{List of parameters to include in the request}

\item{token}{Authentication token}

\item{method}{HTTP method to use (default: "GET")}

\item{API}{endpoint to call}
}
\value{
httr2 response object
}
\description{
The supplied endpoint is appended to the base url
"https://api.twist.com/api/". The endpoint must start with a specific
version prefix like "v3/" or "v4/".
}
