% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{twist_token}
\alias{twist_token}
\title{Set or get the Twist API token}
\usage{
twist_token(token = NULL)
}
\arguments{
\item{token}{API token to use (optional)}
}
\value{
The current API token (invisibly if setting)
}
\description{
Set or get the Twist API token
}
