% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{twist_workspace_dir}
\alias{twist_workspace_dir}
\title{Set or get the Twist workspace directory}
\usage{
twist_workspace_dir(workspace_dir = NULL)
}
\arguments{
\item{workspace_dir}{Workspace directory to use (optional)}
}
\value{
The current workspace directory (invisibly if setting)
}
\description{
Set or get the Twist workspace directory
}
