% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{twist_workspace_id}
\alias{twist_workspace_id}
\title{Set or get the Twist workspace ID}
\usage{
twist_workspace_id(workspace_id = NULL)
}
\arguments{
\item{workspace_id}{Workspace ID to use (optional)}
}
\value{
The current workspace ID (invisibly if setting)
}
\description{
Set or get the Twist workspace ID
}
