% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twopfFit.R
\name{twopfFit}
\alias{twopfFit}
\title{Fits a two pool model with feedback structure to a time series}
\usage{
twopfFit(timeSeries, initialCarbon, inipars = c(1, 0.5, 0.5, 0.5, 0.3))
}
\arguments{
\item{timeSeries}{A time series of respiration values}

\item{initialCarbon}{The initial amount of carbon in units that correspond to the time series data}

\item{inipars}{vector of parameter values for the initial search of the optimization algorithm}
}
\value{
R list with an FME model object, a SoilR model object, and the AIC value
}
\description{
Fits a two pool model with feedback structure to a time series
}
\examples{
\donttest{
aridec=loadEntries()
entry=aridec[["20"]]
b=twopfFit(timeSeries = entry$timeSeries[,c(1,2)],
initialCarbon=100,
inipars=c(0.005, 0.00001, 0.1, 0.01, 0.01))
}
}
