% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoppFit.R
\name{twoppFit}
\alias{twoppFit}
\title{Fits a two pool model with parallel structure to a time-series}
\usage{
twoppFit(timeSeries, initialCarbon)
}
\arguments{
\item{timeSeries}{A time series of respiration values over time}

\item{initialCarbon}{The initial amount of carbon in units that correspond to the time series data}
}
\value{
R list with an FME model object, a SoilR model object, and the AIC value
}
\description{
Fits a two pool model with parallel structure to a time-series
}
\examples{
db=loadEntries(path="~/sidb/clean/")
a=twoppFit(timeSeries = db[[20]]$timeSeries[,1:2], initialCarbon=db[[20]]$initConditions[1,"carbonMean"]*10)
}
