% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twopsFit.R
\name{twopsFit}
\alias{twopsFit}
\title{Fits a two pool model with series structure to a time series}
\usage{
twopsFit(timeSeries, initialCarbon, inipars = c(1, 0.5, 0.5, 0.3))
}
\arguments{
\item{timeSeries}{A time series of decomposition values}

\item{initialCarbon}{The initial amount of carbon in units that correspond to the time series data}

\item{inipars}{vector of parameter values for the initial search of the optimization algorithm}
}
\value{
R list with an FME model object, a SoilR model object, and the AIC value
}
\description{
Fits a two pool model with series structure to a time series
}
\examples{
\donttest{
decomposition=aridec[["Crow2019a"]]
b=twopsFit(timeSeries = decomposition$timeSeries[,c(1,79)],
initialCarbon=decomposition$initConditions[78,"carbonMean"]*10000,
inipars=c(0.005, 0.00001, 0.1, 0.01))
}
}
