% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{txpath_from_edgeattr}
\alias{txpath_from_edgeattr}
\title{Generate transcript-paths from igraph edge attributes}
\usage{
txpath_from_edgeattr(g, type = "exon")
}
\arguments{
\item{g}{An \code{igraph} directed acyclic graph representing a gene splicing graph.}

\item{type}{Character string. Edge type to use for path extraction; either
\code{"exon"} (default) to use exon-level edges or \code{"ex_part"} to use
exonic-part edges.}
}
\description{
Generate transcript-paths from igraph edge attributes
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
txpaths <- grase::txpath_from_edgeattr(g)
names(txpaths)
txpaths[[1]]
}
}
