% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{TypeB}
\alias{TypeB}
\alias{TypeB$new}
\alias{TypeB$get_val}
\alias{$.TypeB}
\alias{[[.TypeB}
\title{TypeB  Class}
\usage{
TypeB

\method{$}{TypeB}(self, name)

\method{[[}{TypeB}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
TypeB  Class
}
