% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{typed_df_theoph_has_flag}
\alias{typed_df_theoph_has_flag}
\title{typed_df_theoph_has_flag}
\usage{
typed_df_theoph_has_flag(df)
}
\arguments{
\item{df}{A data.frame matching the \code{TheophDf} shape.}
}
\value{
Logical scalar.
}
\description{
Whether the optional \code{flag} column was present.
}
