% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{unsafe_C_altrep_ensure_materialized_str}
\alias{unsafe_C_altrep_ensure_materialized_str}
\title{Use ensure_materialized on any SEXP, then extract as string vector}
\source{
Generated by miniextendr from Rust fn \code{C_altrep_ensure_materialized_str}
}
\usage{
unsafe_C_altrep_ensure_materialized_str(x)
}
\arguments{
\item{x}{A SEXP to materialize.}
}
\description{
Use ensure_materialized on any SEXP, then extract as string vector
}
