% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{unsafe_C_altrep_sexp_is_altrep}
\alias{unsafe_C_altrep_sexp_is_altrep}
\title{Return whether a SEXP is ALTREP}
\source{
Generated by miniextendr from Rust fn \code{C_altrep_sexp_is_altrep}
}
\usage{
unsafe_C_altrep_sexp_is_altrep(x)
}
\arguments{
\item{x}{A SEXP to check.}
}
\description{
Takes raw SEXP (extern "C-unwind") to bypass the ALTREP rejection.
}
