% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rpkg_unwind_protect}
\alias{rpkg_unwind_protect}
\alias{unsafe_C_unwind_protect_normal}
\alias{unsafe_C_unwind_protect_r_error}
\alias{unsafe_C_unwind_protect_lowlevel_test}
\title{Test with_r_unwind_protect normal execution with RAII resource cleanup}
\source{
Generated by miniextendr from Rust fn \code{C_unwind_protect_normal}

Generated by miniextendr from Rust fn \code{C_unwind_protect_r_error}

Generated by miniextendr from Rust fn \code{C_unwind_protect_lowlevel_test}
}
\usage{
unsafe_C_unwind_protect_normal()

unsafe_C_unwind_protect_r_error()

unsafe_C_unwind_protect_lowlevel_test()
}
\description{
Test with_r_unwind_protect normal execution with RAII resource cleanup

unwind protect tests

unwind protect tests
}
\examples{
\dontrun{
unsafe_C_unwind_protect_normal()
unsafe_C_unwind_protect_r_error()
unsafe_C_unwind_protect_lowlevel_test()
}
}
