% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread.R
\name{update_thread_file}
\alias{update_thread_file}
\title{Update a thread file if it has been updated in Twist}
\usage{
update_thread_file(path, token = twist_token(), force = FALSE, timezone = NULL)
}
\arguments{
\item{path}{Path to the thread file}

\item{token}{Authentication token}

\item{force}{Logical. Whether to force rewriting even if timestamps indicate
no update (default: FALSE)}

\item{timezone}{Timezone for timestamps; if NULL, use timezone from existing
thread file}
}
\value{
Path to updated file, or NULL if file not updated
}
\description{
If the thread title has been updated, then the thread file will be renamed
to match. If the file doesn't exist, it will be created.
}
