% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{url_fragment}
\alias{url_fragment}
\title{url_fragment}
\usage{
url_fragment(url)
}
\arguments{
\item{url}{Parsed URL from R string.}
}
\description{
Test extracting the fragment from a URL.
}
