% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{url_is_valid}
\alias{url_is_valid}
\title{url_is_valid}
\usage{
url_is_valid(s)
}
\arguments{
\item{s}{String to validate as URL.}
}
\description{
Test validating whether a string is a valid URL.
}
