% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{url_port_or_default}
\alias{url_port_or_default}
\title{url_port_or_default}
\usage{
url_port_or_default(url)
}
\arguments{
\item{url}{Parsed URL from R string.}
}
\description{
Test extracting the port or known default port from a URL.
}
