% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{url_roundtrip_vec}
\alias{url_roundtrip_vec}
\title{url_roundtrip_vec}
\usage{
url_roundtrip_vec(urls)
}
\arguments{
\item{urls}{Character vector of URLs from R.}
}
\description{
Test Vec\if{html}{\out{<Url>}} roundtrip through R character vector.
}
