% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{uuid_roundtrip}
\alias{uuid_roundtrip}
\title{uuid_roundtrip}
\usage{
uuid_roundtrip(uuid)
}
\arguments{
\item{uuid}{UUID parsed from R string.}
}
\description{
Test UUID roundtrip through R string.
}
