% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{uuid_roundtrip_vec}
\alias{uuid_roundtrip_vec}
\title{uuid_roundtrip_vec}
\usage{
uuid_roundtrip_vec(uuids)
}
\arguments{
\item{uuids}{Character vector of UUIDs from R.}
}
\description{
Test Vec\if{html}{\out{<Uuid>}} roundtrip through R character vector.
}
