% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{uuid_version}
\alias{uuid_version}
\title{uuid_version}
\usage{
uuid_version(uuid)
}
\arguments{
\item{uuid}{UUID parsed from R string.}
}
\description{
Test extracting the version number from a UUID.
}
