% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{valid_partitions_idx}
\alias{valid_partitions_idx}
\title{Generate valid partition indices from DAG}
\usage{
valid_partitions_idx(dag)
}
\description{
Generate valid partition indices from DAG
}
\examples{
dag <- igraph::graph_from_edgelist(
  matrix(c("A","B", "B","C"), ncol=2, byrow=TRUE), directed = TRUE
)
igraph::V(dag)$name <- c("A","B","C")
idx_splits <- grase::valid_partitions_idx(dag)
idx_splits
}
