% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_partitions_slow}
\alias{valid_partitions_slow}
\title{Generate valid partitions with powerset approach (slower)}
\usage{
valid_partitions_slow(g, max_powerset)
}
\description{
Generate valid partitions with powerset approach (slower)
}
