% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var2.R
\name{var2}
\alias{var2}
\title{modified Hanley–McNeil approach for variance  estimation  (Newcombe, 2006)}
\usage{
var2(theta, m, n)
}
\arguments{
\item{theta}{concordance}

\item{m}{sample size of group 1}

\item{n}{sample size of group 2}
}
\value{
variance of concordance
}
\description{
#modified Hanley–McNeil approach for variance  estimation  (Newcombe, 2006)
}
\details{
eWHORM simulations
}
\author{
Sonja Zehetmayer
}
\keyword{internal}
