% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reformat.R
\docType{data}
\name{var_roles}
\alias{var_roles}
\title{var_roles list is used to create a catalog object to keep track of and catalog all the variables in the tibble}
\format{
A list of lists
}
\usage{
var_roles
}
\description{
var_roles list is used to create a catalog object to keep track of and catalog all the variables in the tibble
}
\keyword{datasets}
