% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_new_rcrd}
\alias{test_new_rcrd}
\alias{test_new_vctr}
\alias{test_new_list_of_size}
\alias{test_new_vctr_inherit}
\alias{test_new_list_of_ptype}
\alias{test_vctrs_build_error_message}
\title{Test new_rcrd with the given field list and class name(s)}
\source{
Generated by miniextendr from Rust fn \code{test_new_rcrd}

Generated by miniextendr from Rust fn \code{test_new_vctr}

Generated by miniextendr from Rust fn \code{test_new_list_of_size}

Generated by miniextendr from Rust fn \code{test_new_vctr_inherit}

Generated by miniextendr from Rust fn \code{test_new_list_of_ptype}

Generated by miniextendr from Rust fn \code{test_vctrs_build_error_message}
}
\usage{
test_new_rcrd(fields, class)

test_new_vctr(data, class)

test_new_list_of_size(x, size, class)

test_new_vctr_inherit(data, class, inherit_base_type)

test_new_list_of_ptype(x, ptype, class)

test_vctrs_build_error_message(error_type)
}
\arguments{
\item{fields}{(no documentation available)}

\item{class}{(no documentation available)}

\item{data}{(no documentation available)}

\item{x}{(no documentation available)}

\item{size}{(no documentation available)}

\item{inherit_base_type}{(no documentation available)}

\item{ptype}{(no documentation available)}

\item{error_type}{(no documentation available)}
}
\description{
Creates a vctr with the given class name(s) from data. Uses default inherit_base_type (true for lists, false otherwise).
}
