% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{vec_cast.double.percent}
\alias{vec_cast.double.percent}
\title{vec_cast.double.percent}
\usage{
\method{vec_cast}{double.percent}(x, to, ...)
}
\arguments{
\item{x}{(no documentation available)}

\item{to}{(no documentation available)}

\item{...}{Additional arguments (ignored, for vctrs compatibility).}
}
\description{
Cast from percent to double.
}
