% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{vec_cast.percent.double}
\alias{vec_cast.percent.double}
\title{vec_cast.percent.double}
\usage{
\method{vec_cast}{percent.double}(x, to, ...)
}
\arguments{
\item{x}{(no documentation available)}

\item{to}{(no documentation available)}

\item{...}{Additional arguments (ignored, for vctrs compatibility).}
}
\description{
Cast from double to percent.
}
