% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{vec_ptype2.percent.double}
\alias{vec_ptype2.percent.double}
\title{vec_ptype2.percent.double}
\usage{
\method{vec_ptype2}{percent.double}(x, y, ...)
}
\arguments{
\item{x}{(no documentation available)}

\item{y}{(no documentation available)}

\item{...}{Additional arguments (ignored, for vctrs compatibility).}
}
\description{
Coercion: double + percent = percent.
}
