% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\docType{methods}
\name{vector}
\alias{as.vector,fm-method}
\alias{as.vector,fmV-method}
\alias{fm.as.vector}
\alias{fm.is.vector}
\alias{vector}
\title{Vector}
\usage{
fm.as.vector(obj)

\S4method{as.vector}{fmV}(x)

\S4method{as.vector}{fm}(x)

fm.is.vector(x)
}
\arguments{
\item{obj}{an object}

\item{x}{a FlashR vector}
}
\value{
a FlashR vector
}
\description{
\code{fm.as.vector} converts an object to a FlashR vector.
\code{as.vector} converts a FlashR vector to an R vector.
\code{fm.is.vector} test whether the input argument is a FlashR vector.
}
\details{
Right now, \code{fm.as.vector} can only convert a matrix with only one row
or one column. Otherwise, the function returns NULL.
}
\examples{
vec <- fm.as.vector(runif(100)) # Convert an R vector to a FlashR vector.
vec <- as.vector(fm.runif(100)) # Convert a FlashR vector to an R vector.
res <- fm.is.vector(vec) # Test if an object is a FlashR vector.
}

