% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{veptr_make_bags}
\alias{veptr_make_bags}
\title{veptr_make_bags}
\usage{
veptr_make_bags(n)
}
\arguments{
\item{n}{Number of handles to create.}
}
\description{
Build a \code{list()} of \code{Bag} handles with values \verb{1..=n}.

Return side: \verb{Vec<ExternalPtr<Bag>>} becomes an R \code{list()} of external
pointers.
}
