% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorityFlow_data.R
\docType{data}
\name{watershed.mask}
\alias{watershed.mask}
\title{Watershed Mask}
\format{a matrix of 0's and 1' showing the watershed extent (1=inside the watershed, 0=outside the watershed), 
where nrow=nx and ncol=ny for the domain.}
\usage{
watershed.mask
}
\description{
A mask showing the watershed drainage area for the test domain
}
\keyword{datasets}
