% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{wide_to_long}
\alias{wide_to_long}
\title{Convert a wide data frame of OTU relative abundances to a long data frame}
\usage{
wide_to_long(dat, otu_prefix = "Otu")
}
\arguments{
\item{dat}{a wide data frame}

\item{otu_prefix}{column name prefix for OTU columns (default: 'Otu')}
}
\value{
a long dataframe
}
\description{
Convert a wide data frame of OTU relative abundances to a long data frame
}
