% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{widget.create}
\alias{widget.create}
\title{widget.create}
\usage{
widget.create(n)
}
\arguments{
\item{n}{Number of widgets.}
}
\description{
Combined test: r_name + r_entry + r_post_checks.
}
