% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{with_amount}
\alias{with_amount}
\title{Add an amount, consuming and returning the builder}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Add an amount, consuming and returning the builder
}
