% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{with_flag}
\alias{with_flag}
\title{with_flag}
\usage{
with_flag(flag = FALSE)
}
\arguments{
\item{flag}{Logical flag (defaults to FALSE).}
}
\description{
Test boolean parameter with a default value.
}
