% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{with_interrupt_check}
\alias{with_interrupt_check}
\title{with_interrupt_check}
\usage{
with_interrupt_check(x)
}
\arguments{
\item{x}{Integer scalar input.}
}
\description{
Test that \verb{#[miniextendr(check_interrupt)]} checks for user interrupts before execution.
}
