% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{WithOptionals}
\alias{WithOptionals}
\alias{WithOptionals$to_r}
\alias{WithOptionals$all_present}
\alias{WithOptionals$all_none}
\alias{WithOptionals$mixed}
\alias{WithOptionals$from_r}
\alias{$.WithOptionals}
\alias{[[.WithOptionals}
\title{WithOptionals Class}
\usage{
WithOptionals

\method{$}{WithOptionals}(self, name)

\method{[[}{WithOptionals}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
WithOptionals Class
}
