% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{wrapped_board}
\alias{wrapped_board}
\title{wrapped_board}
\usage{
wrapped_board(w, h)
}
\arguments{
\item{w, h}{Board dimensions.}
}
\description{
Explicit wrapping also applies to free functions.
}
