% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{wrapped_board_attr}
\alias{wrapped_board_attr}
\title{wrapped_board_attr}
\usage{
wrapped_board_attr(w, h)
}
\arguments{
\item{w, h}{Board dimensions.}
}
\description{
Attribute equivalent for a free function.
}
