% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{wrapped_optional_boards}
\alias{wrapped_optional_boards}
\title{wrapped_optional_boards}
\usage{
wrapped_optional_boards(present)
}
\arguments{
\item{present}{Whether to return a vector of boards.}
}
\description{
Explicit optional class returns raise on None, including free functions.
}
