% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{wrapped_optional_boards_attr}
\alias{wrapped_optional_boards_attr}
\title{wrapped_optional_boards_attr}
\usage{
wrapped_optional_boards_attr(present)
}
\arguments{
\item{present}{Whether to return a vector of boards.}
}
\description{
The equivalent attribute form of an optional free-function vector.
}
