% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write_Raster.R
\name{write.raster}
\alias{write.raster}
\title{Write raster text file for ArcGIS}
\usage{
write.raster(data, fout, xllcorner = 0, yllcorner = 0, dx = 1000,
  naval = -999)
}
\arguments{
\item{data}{Matrix of values for the raster}

\item{fout}{file name for output}

\item{xllcorner}{xlocation of lower left corner of domain (defaults to 0.0)}

\item{yllcorner}{ylocation of lower left corner of domain (defaults to 0.0)}

\item{dx}{grid cell lateral dimension (defaults to 1000)}

\item{naval}{value assigned to NAs in the raster (defaults to -999)}
}
\description{
Write.Raster - This function writes out a matrix with the six header rows needed for GIS rasters
}
