% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoncnt_functions.R
\name{write_exoncnt_wide}
\alias{write_exoncnt_wide}
\title{Write exon counts in wide format for 3-way testing}
\usage{
write_exoncnt_wide(
  ref_counts_df,
  diff1_counts_df,
  diff2_counts_df,
  events,
  gene_name,
  sampleinfo,
  outfilesuffix,
  outdir
)
}
\arguments{
\item{ref_counts_df}{A data frame of reference exonic part read counts.}

\item{diff1_counts_df}{A data frame of differential exonic part 1 read counts.}

\item{diff2_counts_df}{A data frame of differential exonic part 2 read counts.}

\item{events}{Character vector of event identifiers to include.}

\item{gene_name}{Character string. Gene identifier for the output file name.}

\item{sampleinfo}{A named character vector mapping sample names to group labels.}

\item{outfilesuffix}{Character string. Suffix for output file names.}

\item{outdir}{Character string. Path to output directory.}
}
\description{
Write exon counts in wide format for 3-way testing
}
