% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_signatures}
\alias{write_signatures}
\title{Write signature results to CSV}
\usage{
write_signatures(results, path)
}
\arguments{
\item{results}{A data.frame or data.table of signature results.}

\item{path}{Output CSV path.}
}
\description{
Write signature results to CSV
}
