% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread.R
\name{write_thread}
\alias{write_thread}
\title{Write a thread to a Markdown file}
\usage{
write_thread(thread, token = twist_token(), dir = ".", timezone = "UTC")
}
\arguments{
\item{thread}{Thread object or ID}

\item{token}{Authentication token}

\item{dir}{Directory to write the file to (default: current directory)}

\item{timezone}{Timezone for timestamps (default: "UTC")}
}
\value{
Path to the created file
}
\description{
Write a thread to a Markdown file
}
