% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFB-WriteFcn.R
\name{writepfb}
\alias{writepfb}
\title{Write PFB files
Function to Write a ParFlow Binary Files}
\usage{
writepfb(to.write, filename, dx, dy, dz)
}
\arguments{
\item{to.write}{The array to be written. This must be a 3D array with dimensions nx, ny, nz.
values should be input to the array such that [1,1,1] is the lower left corner of the domain and [nx,ny,nz] is the upper right}

\item{filename}{The name for the file to be written out}

\item{dx}{Grid cell size in x dimension.}

\item{dy}{Grid cell size in y dimension.}

\item{dz}{Grid cell size in z dimension.}
}
\description{
Write PFB files
Function to Write a ParFlow Binary Files
}
