% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ntsa.R
\name{xmap}
\alias{xmap}
\title{Cross-mapping with the information of nearest-neighbour points}
\usage{
xmap(
  df,
  ref,
  tar,
  lib = matrix(c(1, nrow(df)), nrow = 1),
  E,
  Tp = 0,
  libSize = 10,
  randomLibs = TRUE,
  numSamples = 100,
  replace = FALSE,
  seed = NULL
)
}
\arguments{
\item{df}{vector, matrix, data.frame or tibble,
which contains time series to be analyzed}

\item{ref}{Numeric or character, which sets the library variable for xmap}

\item{tar}{Numeric or character, which sets the target variable for xmap}

\item{lib}{Vector or 2-column matrix, which sets the knot position in library data}

\item{E}{Integer, which is the dimension of ref's embedding matrix}

\item{Tp}{Integer, which is the prediction-time horizon in ref xmap tar}

\item{libSize}{Integer, which sets the size of library data used}

\item{randomLibs}{Logical
if TRUE, library data is assinged from random boot-strap sampling}

\item{numSamples}{Integer, which is the boot-strap sampling size
(works when randomLibs = TRUE)}

\item{replace}{Logical
if TRUE, random sampling can select same data (works when randomLibs = TRUE)}

\item{seed}{Integer, which sets the RND seed (works when randomLibs = TRUE)}
}
\description{
\code{xmap} returns the prediction skill of ref cross-mapping tar
cross-mapping skill is calculated with lmdSmplx in ands_ntsa.cpp
}
