% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{zero_copy_arrow_f64_identity}
\alias{zero_copy_arrow_f64_identity}
\title{zero_copy_arrow_f64_identity}
\usage{
zero_copy_arrow_f64_identity(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
Returns TRUE if Float64Array round-trip returns the same R object.
}
