% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{zero_copy_arrow_i32_roundtrip}
\alias{zero_copy_arrow_i32_roundtrip}
\title{zero_copy_arrow_i32_roundtrip}
\usage{
zero_copy_arrow_i32_roundtrip(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
Round-trip Int32Array through Arrow and return the result.
}
