% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{zero_copy_cow_f64_roundtrip}
\alias{zero_copy_cow_f64_roundtrip}
\title{zero_copy_cow_f64_roundtrip}
\usage{
zero_copy_cow_f64_roundtrip(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
Round-trip \verb{Cow<[f64]>} and return the result.
}
