% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{zero_copy_cow_i32_identity}
\alias{zero_copy_cow_i32_identity}
\title{Returns TRUE if Cow<\link{i32}> round-trip returns the same R object}
\source{
Generated by miniextendr from Rust fn \code{zero_copy_cow_i32_identity}
}
\usage{
zero_copy_cow_i32_identity(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
Returns TRUE if Cow<\link{i32}> round-trip returns the same R object
}
