% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{zero_copy_cow_str_is_borrowed}
\alias{zero_copy_cow_str_is_borrowed}
\title{zero_copy_cow_str_is_borrowed}
\usage{
zero_copy_cow_str_is_borrowed(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
Returns TRUE if \verb{Cow<str>} from R is Cow::Borrowed (zero-copy).
}
