% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{zero_copy_protected_strvec_unique}
\alias{zero_copy_protected_strvec_unique}
\title{zero_copy_protected_strvec_unique}
\usage{
zero_copy_protected_strvec_unique(strings)
}
\arguments{
\item{strings}{(no documentation available)}
}
\description{
Count unique non-NA strings via ProtectedStrVec.
}
