% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{zero_copy_sexprec_offset}
\alias{zero_copy_sexprec_offset}
\title{Verifies r_memory::sexprec_data_offset is nonzero (init ran)}
\source{
Generated by miniextendr from Rust fn \code{zero_copy_sexprec_offset}
}
\usage{
zero_copy_sexprec_offset()
}
\description{
Verifies r_memory::sexprec_data_offset is nonzero (init ran)
}
