% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{add}
\alias{add}
\title{Add together two numbers}
\usage{
add(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
A number
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 11)
}
