% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{AliasGuardProbe}
\alias{AliasGuardProbe}
\alias{AliasGuardProbe$new}
\alias{AliasGuardProbe$check}
\alias{$.AliasGuardProbe}
\alias{[[.AliasGuardProbe}
\title{AliasGuardProbe  Class}
\source{
Generated by miniextendr from Rust type \code{AliasGuardProbe}

Generated by miniextendr from \code{AliasGuardProbe::new}

Generated by miniextendr from \code{AliasGuardProbe::check}
}
\usage{
AliasGuardProbe

\method{$}{AliasGuardProbe}(self, name)

\method{[[}{AliasGuardProbe}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Create an alias-check probe.

Check list and direct-slice arguments through an instance method.

\describe{
\item{\code{a}}{List of integer vectors.}
\item{\code{b}}{Shared integer vector.}
}
}
