% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{altrep_sexp_materialize_int}
\alias{altrep_sexp_materialize_int}
\title{Try to wrap a SEXP in AltrepSexp and materialize it as integers}
\source{
Generated by miniextendr from Rust fn \code{altrep_sexp_materialize_int}
}
\usage{
altrep_sexp_materialize_int(x)
}
\arguments{
\item{x}{An ALTREP integer vector.}
}
\description{
Try to wrap a SEXP in AltrepSexp and materialize it as integers
}
