% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{altrep_sexp_materialize_real}
\alias{altrep_sexp_materialize_real}
\title{Try to wrap a SEXP in AltrepSexp and materialize it as doubles}
\source{
Generated by miniextendr from Rust fn \code{altrep_sexp_materialize_real}
}
\usage{
altrep_sexp_materialize_real(x)
}
\arguments{
\item{x}{An ALTREP real vector.}
}
\description{
Uses AltrepSexp parameter -- only accepts ALTREP vectors.
}
