% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{altrep_sexp_materialize_strings}
\alias{altrep_sexp_materialize_strings}
\title{Try to wrap a SEXP in AltrepSexp and materialize strings}
\source{
Generated by miniextendr from Rust fn \code{altrep_sexp_materialize_strings}
}
\usage{
altrep_sexp_materialize_strings(x)
}
\arguments{
\item{x}{An ALTREP string vector.}
}
\description{
Try to wrap a SEXP in AltrepSexp and materialize strings
}
