% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{altrep_sexp_materialize_int}
\alias{altrep_sexp_materialize_int}
\alias{altrep_sexp_materialize_real}
\alias{altrep_sexp_materialize_strings}
\title{altrep_sexp_materialize_int}
\source{
Generated by miniextendr from Rust fn \code{altrep_sexp_materialize_int}

Generated by miniextendr from Rust fn \code{altrep_sexp_materialize_real}

Generated by miniextendr from Rust fn \code{altrep_sexp_materialize_strings}
}
\usage{
altrep_sexp_materialize_int(x)

altrep_sexp_materialize_real(x)

altrep_sexp_materialize_strings(x)
}
\arguments{
\item{x}{An ALTREP string vector.}
}
\description{
Try to wrap a SEXP in AltrepSexp and materialize it as integers.

Uses AltrepSexp parameter -- only accepts ALTREP vectors.

Try to wrap a SEXP in AltrepSexp and materialize it as doubles.

Uses AltrepSexp parameter -- only accepts ALTREP vectors.

Try to wrap a SEXP in AltrepSexp and materialize strings.

Uses AltrepSexp parameter -- only accepts ALTREP vectors.
}
