% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashiness.R
\name{analyze_flashiness_trends}
\alias{analyze_flashiness_trends}
\title{Calculate Richards-Baker flashiness index trends}
\usage{
analyze_flashiness_trends(
  streamflow_data,
  trend_completeness = NULL,
  decade_completeness = NULL,
  min_values_for_stats = NULL,
  changepoint = NULL,
  collector = NULL
)
}
\arguments{
\item{streamflow_data}{A data.frame with columns: water_year, Q, dowy.}
}
\value{
Named list of 8 statistics for flashinessRB.
}
\description{
Calculate Richards-Baker flashiness index trends
}
