% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{arith_seq}
\alias{arith_seq}
\title{arith seq}
\source{
Generated by miniextendr from Rust fn `arith_seq`
}
\usage{
arith_seq(from, step, length_out)
}
\arguments{
\item{from}{(no documentation available)}

\item{step}{(no documentation available)}

\item{length_out}{(no documentation available)}
}
\description{
arith seq
}
