% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{AsCoerceErrorTest}
\alias{AsCoerceErrorTest}
\alias{new_ascoerceerrortest}
\alias{as_data_frame.AsCoerceErrorTest}
\alias{as_list.AsCoerceErrorTest}
\alias{as_character.AsCoerceErrorTest}
\alias{as.data.frame.AsCoerceErrorTest}
\alias{as.list.AsCoerceErrorTest}
\alias{as.character.AsCoerceErrorTest}
\title{as.\if{html}{\out{<class>}}() Error Test Type}
\format{
An object of class \code{environment} of length 1.
}
\source{
Generated by miniextendr from Rust type \code{AsCoerceErrorTest}

Generated by miniextendr from \code{AsCoerceErrorTest::as_data_frame}

Generated by miniextendr from \code{AsCoerceErrorTest::as_data_frame}

Generated by miniextendr from \code{AsCoerceErrorTest::as_list}

Generated by miniextendr from \code{AsCoerceErrorTest::as_list}

Generated by miniextendr from \code{AsCoerceErrorTest::as_character}

Generated by miniextendr from \code{AsCoerceErrorTest::as_character}

Generated by miniextendr from \code{AsCoerceErrorTest::as_data_frame}

Generated by miniextendr from \code{AsCoerceErrorTest::as_list}

Generated by miniextendr from \code{AsCoerceErrorTest::as_character}
}
\usage{
new_ascoerceerrortest(is_empty)

\method{as_data_frame}{AsCoerceErrorTest}(x, ...)

\method{as_list}{AsCoerceErrorTest}(x, ...)

\method{as_character}{AsCoerceErrorTest}(x, ...)

AsCoerceErrorTest

\method{as.data.frame}{AsCoerceErrorTest}(x, ...)

\method{as.list}{AsCoerceErrorTest}(x, ...)

\method{as.character}{AsCoerceErrorTest}(x, ...)
}
\arguments{
\item{is_empty}{Logical; if TRUE, triggers error on as.data.frame.}

\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
S3 generic for \code{as_data_frame}

Convert to data.frame - fails when empty.

S3 generic for \code{as_list}

Convert to list - always returns "not supported" error.

S3 generic for \code{as_character}

Convert to character - custom error message.

Convert to data.frame - fails when empty.

Convert to list - always returns "not supported" error.

Convert to character - custom error message.
}
\keyword{datasets}
