% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{AtomicCounter}
\alias{AtomicCounter}
\alias{AtomicCounter$new_atomic}
\alias{$.AtomicCounter}
\alias{[[.AtomicCounter}
\alias{AtomicCounter$Counter$value}
\alias{AtomicCounter$Counter$increment}
\alias{AtomicCounter$Counter$add}
\alias{AtomicCounter$Counter$reset}
\title{AtomicCounter Class}
\format{
An object of class \code{environment} of length 2.
}
\source{
Generated by miniextendr from Rust type \code{AtomicCounter}
}
\usage{
AtomicCounter

\method{$}{AtomicCounter}(self, name)

\method{[[}{AtomicCounter}(self, name)
}
\description{
AtomicCounter Class
}
\keyword{datasets}
