% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{backtrace_install_hook}
\alias{backtrace_install_hook}
\title{backtrace tests}
\source{
Generated by miniextendr from Rust fn `backtrace_install_hook`
}
\usage{
backtrace_install_hook()
}
\description{
backtrace tests
}
