% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bench_vec_altrep}
\alias{bench_vec_altrep}
\title{Create a vector of given size using ALTREP zero-copy}
\source{
Generated by miniextendr from Rust fn `bench_vec_altrep`
}
\usage{
bench_vec_altrep(n)
}
\arguments{
\item{n}{Length of the vector.}
}
\description{
Create a vector of given size using ALTREP zero-copy
}
