% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bench_vec_copy}
\alias{bench_vec_copy}
\title{Create a vector of given size using regular copy (IntoR)}
\source{
Generated by miniextendr from Rust fn `bench_vec_copy`
}
\usage{
bench_vec_copy(n)
}
\arguments{
\item{n}{Length of the vector.}
}
\description{
Create a vector of given size using regular copy (IntoR)
}
