% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartition_analysis.R
\name{bipartition_paths}
\alias{bipartition_paths}
\title{Compute diff exons from graph using binary partitions}
\usage{
bipartition_paths(gene, g, outdir, max_path = 20, collapse_bubbles = FALSE)
}
\description{
Compute diff exons from graph using binary partitions
}
