% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartitions.R
\name{bipartition_paths}
\alias{bipartition_paths}
\title{Compute diff exons from graph and splicing structure}
\usage{
bipartition_paths(gene, g, sg, outdir, max_path = 20, collapse_bubbles = FALSE)
}
\description{
Compute diff exons from graph and splicing structure
}
