% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitflags_all}
\alias{bitflags_all}
\alias{bitflags_xor}
\alias{bitflags_count}
\alias{bitflags_empty}
\alias{bitflags_names}
\alias{bitflags_union}
\alias{bitflags_display}
\alias{bitflags_has_read}
\alias{bitflags_has_write}
\alias{bitflags_intersect}
\alias{bitflags_roundtrip}
\alias{bitflags_complement}
\alias{bitflags_from_strict}
\alias{bitflags_has_execute}
\alias{bitflags_from_truncate}
\title{bitflags adapter tests}
\source{
Generated by miniextendr from Rust fn `bitflags_all`

Generated by miniextendr from Rust fn `bitflags_xor`

Generated by miniextendr from Rust fn `bitflags_count`

Generated by miniextendr from Rust fn `bitflags_empty`

Generated by miniextendr from Rust fn `bitflags_names`

Generated by miniextendr from Rust fn `bitflags_union`

Generated by miniextendr from Rust fn `bitflags_display`

Generated by miniextendr from Rust fn `bitflags_has_read`

Generated by miniextendr from Rust fn `bitflags_has_write`

Generated by miniextendr from Rust fn `bitflags_intersect`

Generated by miniextendr from Rust fn `bitflags_roundtrip`

Generated by miniextendr from Rust fn `bitflags_complement`

Generated by miniextendr from Rust fn `bitflags_from_strict`

Generated by miniextendr from Rust fn `bitflags_has_execute`

Generated by miniextendr from Rust fn `bitflags_from_truncate`
}
\usage{
bitflags_all()

bitflags_xor(a, b)

bitflags_count(flags)

bitflags_empty()

bitflags_names(flags)

bitflags_union(a, b)

bitflags_display(flags)

bitflags_has_read(flags)

bitflags_has_write(flags)

bitflags_intersect(a, b)

bitflags_roundtrip(flags)

bitflags_complement(flags)

bitflags_from_strict(value)

bitflags_has_execute(flags)

bitflags_from_truncate(value)
}
\arguments{
\item{a}{First permission flags.}

\item{b}{Second permission flags.}

\item{flags}{Permission flags as RFlags wrapper.}

\item{value}{Integer bitmask to convert.}
}
\description{
bitflags adapter tests

Compute the symmetric difference (XOR) of two flag sets

Count the number of flags that are set

bitflags adapter tests

Get the names of all set flags as a character vector

Test computing the union (bitwise OR) of two flag sets

Display the Debug format of a flag set

Test checking the READ flag on a permission set

Test checking the WRITE flag on a permission set

Test computing the intersection (bitwise AND) of two flag sets

Test RFlags roundtrip through R

Compute the complement (NOT) of a flag set (within defined bits)

Test strict conversion from integer to flags, rejecting unknown bits

Test checking the EXECUTE flag on a permission set

Test truncating conversion from integer to flags, ignoring unknown bits
}
