% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{bitvec_len}
\alias{bitvec_len}
\alias{bitvec_ones}
\alias{bitvec_empty}
\alias{bitvec_zeros}
\alias{bitvec_toggle}
\alias{bitvec_to_vec}
\alias{bitvec_all_ones}
\alias{bitvec_from_vec}
\alias{bitvec_all_zeros}
\alias{bitvec_roundtrip}
\title{Test getting the length of a bit vector}
\source{
Generated by miniextendr from Rust fn `bitvec_len`

Generated by miniextendr from Rust fn `bitvec_ones`

Generated by miniextendr from Rust fn `bitvec_empty`

Generated by miniextendr from Rust fn `bitvec_zeros`

Generated by miniextendr from Rust fn `bitvec_toggle`

Generated by miniextendr from Rust fn `bitvec_to_vec`

Generated by miniextendr from Rust fn `bitvec_all_ones`

Generated by miniextendr from Rust fn `bitvec_from_vec`

Generated by miniextendr from Rust fn `bitvec_all_zeros`

Generated by miniextendr from Rust fn `bitvec_roundtrip`
}
\usage{
bitvec_len(bits)

bitvec_ones(bits)

bitvec_empty()

bitvec_zeros(bits)

bitvec_toggle(bits)

bitvec_to_vec(bits)

bitvec_all_ones(n)

bitvec_from_vec(bools)

bitvec_all_zeros(n)

bitvec_roundtrip(bits)
}
\arguments{
\item{bits}{Bit vector to roundtrip.}

\item{n}{Number of bits.}

\item{bools}{Logical vector to convert.}
}
\description{
Test getting the length of a bit vector

Test counting the number of set bits (ones) in a bit vector

bitvec adapter tests

Test counting the number of unset bits (zeros) in a bit vector

Test toggling (flipping) all bits in a bit vector

Test converting a bit vector back to a logical vector

Test creating an all-ones bit vector of given length

Test creating a bit vector from a logical vector

Test creating an all-zeros bit vector of given length

Test RBitVec roundtrip through R
}
