% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{box_slice_i32_roundtrip}
\alias{box_slice_i32_roundtrip}
\title{\verb{Box<[i32]>} roundtrip: R integer → Rust → R integer}
\source{
Generated by miniextendr from Rust fn \code{box_slice_i32_roundtrip}
}
\usage{
box_slice_i32_roundtrip(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
\verb{Box<[i32]>} roundtrip: R integer → Rust → R integer
}
