% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{boxed_data_altrep}
\alias{boxed_data_altrep}
\title{Example: Using as_altrep() to store wrapper}
\source{
Generated by miniextendr from Rust fn \code{boxed_data_altrep}
}
\usage{
boxed_data_altrep(n)
}
\description{
Example: Using as_altrep() to store wrapper
}
