% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focalexons.R
\name{bubble_ordering}
\alias{bubble_ordering}
\title{hierarchical ordering of bubbles in the DAG grase graph from innermost to outermost based on nesting relationships.}
\usage{
bubble_ordering(g, bubbles_df)
}
\description{
hierarchical ordering of bubbles in the DAG grase graph from innermost to outermost based on nesting relationships.
}
