% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{bubble_paths_igraph}
\alias{bubble_paths_igraph}
\title{Get all simple paths through a bubble in an igraph splice graph}
\usage{
bubble_paths_igraph(g, v_start, v_end)
}
\description{
Get all simple paths through a bubble in an igraph splice graph
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_edge_names(g)
paths <- grase::bubble_paths_igraph(g, "3", "7")
}
}
