% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{call_attr_with}
\alias{call_attr_with}
\alias{unsafe_C_call_attr_without}
\title{Wrapped path}
\source{
Generated by miniextendr from Rust fn \code{call_attr_with}

Generated by miniextendr from Rust fn \code{C_call_attr_without}
}
\usage{
call_attr_with(left, right)

unsafe_C_call_attr_without(left, right)
}
\arguments{
\item{left}{Ignored.}

\item{right}{Ignored.}
}
\description{
Wrapped path

Unwrapped path
}
\keyword{internal}
