% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalStream_Order.R
\name{CalStreamOrder}
\alias{CalStreamOrder}
\title{Defining Stream order}
\usage{
CalStreamOrder(basinID, dsID, segments)
}
\arguments{
\item{basinID}{a column list of basin IDs}

\item{dsID}{downstream basin IDs, the downstream ID of each basin, should be correpsonding to the basin ID}

\item{segments}{(optional) a channel mask with segments assigned with basin IDs 
if PriorityFlow is used to generate subbasins, all inputs can be obtained with function CalcSubbasins in PriporityFlow.}
}
\value{
summary - A summary table with a row for every basin with three columns "Basin_ID" ""Downstream_ID" "StreamOrder_number"

channel_orders (optional)- only available if segments is an input,this will output a spatial map with the channel orders
}
\description{
Function to calculate Strahler stream orers using stream segments delineated using the CalcSubbasins Function
}
