% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{Create categorical column from expressions}
\usage{
categorize(.data, .col, ..., .default = NA_character_)
}
\arguments{
\item{.data}{A data frame}

\item{.col}{Bare column name}

\item{...}{Named expressions evaluating a logical value. The name will
be used for the categorical variable. For}

\item{.default}{The value to use for \code{col} if none of the logical
predicates in \code{...} are true.}
}
\value{
A data frame consisting of \code{.data} with the additional column
  \code{col}.
}
\description{
Create a categorical variable from logical expressions on a data frame.
}
\examples{
categorize(data.frame(a = 1:10, b = rep(0:1, 5)),
             x,
             "lt 5" = a < 5,
             "ge 5, b" = a >= 5 & b == 1,
             "gt 5, not b" = a >= 5 & b == 0)
}
\seealso{
\code{\link[plyr]{revalue}}, \code{\link[plyr]{mapvalues}},
  \code{\link{cut}}, \code{\link{switch}}
}

