% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{ChainedError}
\alias{ChainedError}
\alias{ChainedError$new}
\alias{ChainedError$without_source}
\alias{$.ChainedError}
\alias{[[.ChainedError}
\alias{ChainedError$RError$error_message}
\alias{ChainedError$RError$error_chain}
\alias{ChainedError$RError$error_chain_length}
\title{ChainedError inherent methods: constructors with and without source errors}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type \code{ChainedError}

Generated by miniextendr from \code{ChainedError::new}

Generated by miniextendr from \code{ChainedError::without_source}
}
\usage{
ChainedError

\method{$}{ChainedError}(self, name)

\method{[[}{ChainedError}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}

\item{outer_msg}{Character string for the outer error message.}

\item{inner_msg}{Character string for the inner (source) error message.}
}
\description{
ChainedError inherent methods: constructors with and without source errors
}
\keyword{datasets}
