% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{ChainedError}
\alias{ChainedError}
\alias{ChainedError$new}
\alias{ChainedError$error_message}
\alias{ChainedError$error_chain}
\alias{ChainedError$error_chain_length}
\alias{ChainedError$without_source}
\alias{$.ChainedError}
\alias{[[.ChainedError}
\title{Error wrapper for testing RError}
\format{
An object of class \code{environment} of length 5.
}
\source{
Generated by miniextendr from Rust type \code{ChainedError}

Generated by miniextendr from \code{ChainedError::new}

Generated by miniextendr from \code{ChainedError::error_message}

Generated by miniextendr from \code{ChainedError::error_chain}

Generated by miniextendr from \code{ChainedError::error_chain_length}

Generated by miniextendr from \code{ChainedError::without_source}
}
\usage{
ChainedError

\method{$}{ChainedError}(self, name)

\method{[[}{ChainedError}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\value{
An ExternalPtr to a ChainedError.
}
\description{
An error type that demonstrates error chain walking.
}
\examples{
err <- ChainedError$new("outer", "inner")
err$error_message()
err$error_chain()
}
\keyword{datasets}
