% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{ChainedError}
\alias{ChainedError}
\alias{ChainedError$new}
\alias{ChainedError$without_source}
\alias{$.ChainedError}
\alias{[[.ChainedError}
\alias{ChainedError$RError$error_message}
\alias{ChainedError$RError$error_chain}
\alias{ChainedError$RError$error_chain_length}
\title{ChainedError  Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type `ChainedError`

Generated by miniextendr from `ChainedError::new`

Generated by miniextendr from `ChainedError::without_source`
}
\usage{
ChainedError

\method{$}{ChainedError}(self, name)

\method{[[}{ChainedError}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
ChainedError  Class

\describe{
  \item{\code{outer_msg}}{Character string for the outer error message.}
  \item{\code{inner_msg}}{Character string for the inner (source) error message.}
}
}
\keyword{datasets}
