% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{choices_mixed}
\alias{choices_mixed}
\title{choices mixed}
\source{
Generated by miniextendr from Rust fn \code{choices_mixed}
}
\usage{
choices_mixed(n, mode = c("fast", "safe"), verbose)
}
\arguments{
\item{mode}{One of "fast", "safe".}
}
\description{
choices mixed
}
