% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{Collections}
\alias{Collections}
\alias{Collections$new}
\alias{Collections$to_r}
\alias{Collections$empty}
\alias{Collections$from_r}
\alias{$.Collections}
\alias{[[.Collections}
\title{Collections  Class}
\format{
An object of class \code{environment} of length 4.
}
\source{
Generated by miniextendr from Rust type \code{Collections}

Generated by miniextendr from \code{Collections::new}

Generated by miniextendr from \code{Collections::to_r}

Generated by miniextendr from \code{Collections::empty}

Generated by miniextendr from \code{Collections::from_r}
}
\usage{
Collections

\method{$}{Collections}(self, name)

\method{[[}{Collections}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Collections  Class
}
\keyword{datasets}
