% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{flatten_enum_field_fixture}
\alias{flatten_enum_field_fixture}
\alias{flatten_enum_option_none_fixture}
\alias{flatten_enum_unit_variant_fixture}
\title{columnar flatten enum tests}
\source{
Generated by miniextendr from Rust fn \code{flatten_enum_field_fixture}

Generated by miniextendr from Rust fn \code{flatten_enum_option_none_fixture}

Generated by miniextendr from Rust fn \code{flatten_enum_unit_variant_fixture}
}
\usage{
flatten_enum_field_fixture()

flatten_enum_option_none_fixture()

flatten_enum_unit_variant_fixture()
}
\description{
columnar flatten enum tests

columnar flatten enum tests

columnar flatten enum tests
}
