% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_i32_arg}
\alias{conv_i32_arg}
\alias{conv_i32_ret}
\alias{conv_f64_arg}
\alias{conv_f64_ret}
\alias{conv_u8_arg}
\alias{conv_u8_ret}
\alias{conv_rbool_arg}
\alias{conv_rbool_ret}
\alias{conv_rlog_arg}
\alias{conv_rlog_ret}
\alias{conv_string_arg}
\alias{conv_string_ret}
\alias{conv_str_ret}
\alias{conv_sexp_arg}
\alias{conv_sexp_ret}
\alias{conv_i64_arg}
\alias{conv_i64_ret}
\alias{conv_u64_arg}
\alias{conv_u64_ret}
\alias{conv_isize_arg}
\alias{conv_isize_ret}
\alias{conv_usize_arg}
\alias{conv_usize_ret}
\alias{conv_i8_arg}
\alias{conv_i8_ret}
\alias{conv_i16_arg}
\alias{conv_i16_ret}
\alias{conv_u16_arg}
\alias{conv_u16_ret}
\alias{conv_u32_arg}
\alias{conv_u32_ret}
\alias{conv_f32_arg}
\alias{conv_f32_ret}
\alias{conv_opt_i32_is_some}
\alias{conv_opt_i32_some}
\alias{conv_opt_i32_none}
\alias{conv_opt_f64_is_some}
\alias{conv_opt_f64_some}
\alias{conv_opt_f64_none}
\alias{conv_opt_bool_is_some}
\alias{conv_opt_bool_some}
\alias{conv_opt_bool_none}
\alias{conv_opt_string_is_some}
\alias{conv_opt_string_some}
\alias{conv_opt_string_none}
\alias{conv_opt_i8_is_some}
\alias{conv_opt_i16_is_some}
\alias{conv_opt_i64_is_some}
\alias{conv_opt_isize_is_some}
\alias{conv_opt_u16_is_some}
\alias{conv_opt_u32_is_some}
\alias{conv_opt_u64_is_some}
\alias{conv_opt_usize_is_some}
\alias{conv_opt_f32_is_some}
\alias{conv_opt_u8_is_some}
\alias{conv_opt_rbool_is_some}
\alias{conv_opt_rlog_is_some}
\alias{conv_slice_i32_len}
\alias{conv_slice_f64_len}
\alias{conv_slice_u8_len}
\alias{conv_slice_rlog_len}
\alias{conv_ref_i32_arg}
\alias{conv_ref_mut_i32_add_one}
\alias{conv_slice_mut_i32_add_one}
\alias{conv_slice_mut_u8_add_one}
\alias{conv_opt_ref_i32_is_some}
\alias{conv_opt_mut_slice_i32_is_some}
\alias{conv_vec_ref_i32_len}
\alias{conv_vec_slice_i32_total_len}
\alias{conv_vec_mut_slice_i32_add_one}
\alias{alias_probe}
\alias{debug_assertions_enabled}
\alias{conv_vec_i32_len}
\alias{conv_vec_i32_ret}
\alias{conv_vec_f64_len}
\alias{conv_vec_f64_ret}
\alias{conv_vec_u8_len}
\alias{conv_vec_u8_ret}
\alias{conv_vec_rlog_len}
\alias{conv_vec_rlog_ret}
\alias{conv_vec_bool_len}
\alias{conv_vec_bool_ret}
\alias{conv_vec_string_len}
\alias{conv_vec_string_ret}
\alias{conv_vec_i8_len}
\alias{conv_vec_i16_len}
\alias{conv_vec_i64_len}
\alias{conv_vec_isize_len}
\alias{conv_vec_u16_len}
\alias{conv_vec_u32_len}
\alias{conv_vec_u64_len}
\alias{conv_vec_usize_len}
\alias{conv_vec_f32_len}
\alias{conv_vec_opt_i32_len}
\alias{conv_vec_opt_i32_ret}
\alias{conv_vec_opt_f64_len}
\alias{conv_vec_opt_f64_ret}
\alias{conv_vec_opt_bool_len}
\alias{conv_vec_opt_bool_ret}
\alias{conv_vec_opt_string_len}
\alias{conv_vec_opt_string_ret}
\alias{conv_vec_opt_rlog_ret}
\alias{conv_vec_opt_rbool_ret}
\alias{conv_vec_opt_i8_len}
\alias{conv_vec_opt_u8_len}
\alias{conv_hashset_i32_len}
\alias{conv_hashset_i32_ret}
\alias{conv_hashset_u8_len}
\alias{conv_hashset_u8_ret}
\alias{conv_hashset_string_len}
\alias{conv_hashset_string_ret}
\alias{conv_hashset_rlog_len}
\alias{conv_hashset_rlog_ret}
\alias{conv_btreeset_i32_len}
\alias{conv_btreeset_i32_ret}
\alias{conv_btreeset_u8_len}
\alias{conv_btreeset_u8_ret}
\alias{conv_btreeset_string_len}
\alias{conv_btreeset_string_ret}
\alias{conv_hashmap_i32_len}
\alias{conv_hashmap_i32_ret}
\alias{conv_hashmap_f64_len}
\alias{conv_hashmap_f64_ret}
\alias{conv_hashmap_string_len}
\alias{conv_hashmap_string_ret}
\alias{conv_hashmap_rlog_len}
\alias{conv_hashmap_rlog_ret}
\alias{conv_btreemap_i32_len}
\alias{conv_btreemap_i32_ret}
\alias{conv_btreemap_f64_len}
\alias{conv_btreemap_f64_ret}
\alias{conv_btreemap_string_len}
\alias{conv_btreemap_string_ret}
\alias{conv_btreemap_rlog_len}
\alias{conv_btreemap_rlog_ret}
\alias{conv_list_mut_set_first}
\alias{conv_result_i32_arg}
\alias{conv_result_vec_i32_arg}
\alias{conv_result_i32_ok}
\alias{conv_result_i32_err}
\alias{conv_result_f64_ok}
\alias{conv_result_f64_err}
\alias{conv_result_string_ok}
\alias{conv_result_string_err}
\alias{conv_result_vec_i32_ok}
\alias{conv_result_vec_i32_err}
\alias{conv_char_arg}
\alias{conv_char_ret}
\alias{conv_vec_i8_ret}
\alias{conv_vec_i16_ret}
\alias{conv_vec_u16_ret}
\alias{conv_vec_f32_ret}
\alias{conv_hashset_i8_ret}
\alias{conv_btreeset_i8_ret}
\alias{conv_hashset_i16_ret}
\alias{conv_btreeset_i16_ret}
\alias{conv_hashset_u16_ret}
\alias{conv_btreeset_u16_ret}
\alias{conv_opt_ref_i32_some_ret}
\alias{conv_opt_ref_i32_none_ret}
\alias{conv_opt_vec_i32_arg}
\alias{conv_opt_vec_i32_some_ret}
\alias{conv_opt_vec_i32_none_ret}
\alias{conv_opt_vec_string_arg}
\alias{conv_opt_vec_string_some_ret}
\alias{conv_opt_vec_string_none_ret}
\alias{conv_opt_hashmap_i32_arg}
\alias{conv_opt_hashmap_i32_some_ret}
\alias{conv_opt_hashmap_i32_none_ret}
\alias{conv_opt_hashset_i32_arg}
\alias{conv_opt_hashset_i32_some_ret}
\alias{conv_opt_hashset_i32_none_ret}
\alias{conv_opt_btreemap_i32_arg}
\alias{conv_opt_btreemap_i32_some_ret}
\alias{conv_opt_btreemap_i32_none_ret}
\alias{conv_vec_hashmap_i32_arg}
\alias{conv_vec_hashmap_i32_ret}
\alias{conv_vec_btreemap_i32_arg}
\alias{conv_vec_btreemap_i32_ret}
\alias{conv_vec_vec_i32_arg}
\alias{conv_vec_vec_i32_ret}
\alias{conv_vec_vec_string_ret}
\alias{conv_vec_option_i64_ret_small}
\alias{conv_vec_option_i64_ret_big}
\alias{conv_vec_option_u32_ret}
\alias{conv_vec_option_f32_ret}
\alias{conv_vec_option_i64_roundtrip}
\alias{conv_vec_option_u64_ret_small}
\alias{conv_vec_option_u64_ret_big}
\alias{conv_vec_option_isize_ret_small}
\alias{conv_vec_option_usize_ret_small}
\alias{conv_vec_option_i8_ret}
\alias{conv_vec_option_i16_ret}
\alias{conv_vec_option_u16_ret}
\alias{conv_option_i64_some_small}
\alias{conv_option_i64_some_big}
\alias{conv_option_i64_none}
\alias{conv_option_f32_some}
\alias{conv_option_u32_some}
\alias{conv_as_named_list_vec}
\alias{conv_as_named_list_array}
\alias{conv_as_named_list_heterogeneous}
\alias{conv_as_named_list_str_keys}
\alias{conv_as_named_list_empty}
\alias{conv_as_named_list_duplicate_names}
\alias{conv_as_named_vector_i32}
\alias{conv_as_named_vector_f64}
\alias{conv_as_named_vector_string}
\alias{conv_as_named_vector_option_i32}
\alias{conv_as_named_vector_array}
\alias{conv_as_named_vector_empty}
\alias{conv_as_named_vector_ext_trait}
\alias{conv_as_named_list_ext_trait}
\alias{conv_as_named_list_slice}
\alias{conv_as_named_vector_slice}
\alias{conv_named_list_get}
\alias{conv_named_list_contains}
\alias{conv_named_list_len}
\alias{conv_named_list_roundtrip}
\alias{conv_tuple1_roundtrip}
\alias{conv_tuple2_roundtrip}
\alias{conv_tuple3_sum}
\title{conv_i32_arg}
\usage{
conv_i32_arg(x)

conv_i32_ret()

conv_f64_arg(x)

conv_f64_ret()

conv_u8_arg(x)

conv_u8_ret()

conv_rbool_arg(x)

conv_rbool_ret()

conv_rlog_arg(x)

conv_rlog_ret()

conv_string_arg(x)

conv_string_ret()

conv_str_ret()

conv_sexp_arg(x)

conv_sexp_ret()

conv_i64_arg(x)

conv_i64_ret()

conv_u64_arg(x)

conv_u64_ret()

conv_isize_arg(x)

conv_isize_ret()

conv_usize_arg(x)

conv_usize_ret()

conv_i8_arg(x)

conv_i8_ret()

conv_i16_arg(x)

conv_i16_ret()

conv_u16_arg(x)

conv_u16_ret()

conv_u32_arg(x)

conv_u32_ret()

conv_f32_arg(x)

conv_f32_ret()

conv_opt_i32_is_some(x)

conv_opt_i32_some()

conv_opt_i32_none()

conv_opt_f64_is_some(x)

conv_opt_f64_some()

conv_opt_f64_none()

conv_opt_bool_is_some(x)

conv_opt_bool_some()

conv_opt_bool_none()

conv_opt_string_is_some(x)

conv_opt_string_some()

conv_opt_string_none()

conv_opt_i8_is_some(x)

conv_opt_i16_is_some(x)

conv_opt_i64_is_some(x)

conv_opt_isize_is_some(x)

conv_opt_u16_is_some(x)

conv_opt_u32_is_some(x)

conv_opt_u64_is_some(x)

conv_opt_usize_is_some(x)

conv_opt_f32_is_some(x)

conv_opt_u8_is_some(x)

conv_opt_rbool_is_some(x)

conv_opt_rlog_is_some(x)

conv_slice_i32_len(x)

conv_slice_f64_len(x)

conv_slice_u8_len(x)

conv_slice_rlog_len(x)

conv_ref_i32_arg(x)

conv_ref_mut_i32_add_one(x)

conv_slice_mut_i32_add_one(x)

conv_slice_mut_u8_add_one(x)

conv_opt_ref_i32_is_some(x)

conv_opt_mut_slice_i32_is_some(x)

conv_vec_ref_i32_len(x)

conv_vec_slice_i32_total_len(x)

conv_vec_mut_slice_i32_add_one(x)

alias_probe(a, b)

debug_assertions_enabled()

conv_vec_i32_len(x)

conv_vec_i32_ret()

conv_vec_f64_len(x)

conv_vec_f64_ret()

conv_vec_u8_len(x)

conv_vec_u8_ret()

conv_vec_rlog_len(x)

conv_vec_rlog_ret()

conv_vec_bool_len(x)

conv_vec_bool_ret()

conv_vec_string_len(x)

conv_vec_string_ret()

conv_vec_i8_len(x)

conv_vec_i16_len(x)

conv_vec_i64_len(x)

conv_vec_isize_len(x)

conv_vec_u16_len(x)

conv_vec_u32_len(x)

conv_vec_u64_len(x)

conv_vec_usize_len(x)

conv_vec_f32_len(x)

conv_vec_opt_i32_len(x)

conv_vec_opt_i32_ret()

conv_vec_opt_f64_len(x)

conv_vec_opt_f64_ret()

conv_vec_opt_bool_len(x)

conv_vec_opt_bool_ret()

conv_vec_opt_string_len(x)

conv_vec_opt_string_ret()

conv_vec_opt_rlog_ret()

conv_vec_opt_rbool_ret()

conv_vec_opt_i8_len(x)

conv_vec_opt_u8_len(x)

conv_hashset_i32_len(x)

conv_hashset_i32_ret()

conv_hashset_u8_len(x)

conv_hashset_u8_ret()

conv_hashset_string_len(x)

conv_hashset_string_ret()

conv_hashset_rlog_len(x)

conv_hashset_rlog_ret()

conv_btreeset_i32_len(x)

conv_btreeset_i32_ret()

conv_btreeset_u8_len(x)

conv_btreeset_u8_ret()

conv_btreeset_string_len(x)

conv_btreeset_string_ret()

conv_hashmap_i32_len(x)

conv_hashmap_i32_ret()

conv_hashmap_f64_len(x)

conv_hashmap_f64_ret()

conv_hashmap_string_len(x)

conv_hashmap_string_ret()

conv_hashmap_rlog_len(x)

conv_hashmap_rlog_ret()

conv_btreemap_i32_len(x)

conv_btreemap_i32_ret()

conv_btreemap_f64_len(x)

conv_btreemap_f64_ret()

conv_btreemap_string_len(x)

conv_btreemap_string_ret()

conv_btreemap_rlog_len(x)

conv_btreemap_rlog_ret()

conv_list_mut_set_first(x)

conv_result_i32_arg(x)

conv_result_vec_i32_arg(x)

conv_result_i32_ok()

conv_result_i32_err()

conv_result_f64_ok()

conv_result_f64_err()

conv_result_string_ok()

conv_result_string_err()

conv_result_vec_i32_ok()

conv_result_vec_i32_err()

conv_char_arg(x)

conv_char_ret()

conv_vec_i8_ret()

conv_vec_i16_ret()

conv_vec_u16_ret()

conv_vec_f32_ret()

conv_hashset_i8_ret()

conv_btreeset_i8_ret()

conv_hashset_i16_ret()

conv_btreeset_i16_ret()

conv_hashset_u16_ret()

conv_btreeset_u16_ret()

conv_opt_ref_i32_some_ret()

conv_opt_ref_i32_none_ret()

conv_opt_vec_i32_arg(x)

conv_opt_vec_i32_some_ret()

conv_opt_vec_i32_none_ret()

conv_opt_vec_string_arg(x)

conv_opt_vec_string_some_ret()

conv_opt_vec_string_none_ret()

conv_opt_hashmap_i32_arg(x)

conv_opt_hashmap_i32_some_ret()

conv_opt_hashmap_i32_none_ret()

conv_opt_hashset_i32_arg(x)

conv_opt_hashset_i32_some_ret()

conv_opt_hashset_i32_none_ret()

conv_opt_btreemap_i32_arg(x)

conv_opt_btreemap_i32_some_ret()

conv_opt_btreemap_i32_none_ret()

conv_vec_hashmap_i32_arg(x)

conv_vec_hashmap_i32_ret()

conv_vec_btreemap_i32_arg(x)

conv_vec_btreemap_i32_ret()

conv_vec_vec_i32_arg(x)

conv_vec_vec_i32_ret()

conv_vec_vec_string_ret()

conv_vec_option_i64_ret_small()

conv_vec_option_i64_ret_big()

conv_vec_option_u32_ret()

conv_vec_option_f32_ret()

conv_vec_option_i64_roundtrip(x)

conv_vec_option_u64_ret_small()

conv_vec_option_u64_ret_big()

conv_vec_option_isize_ret_small()

conv_vec_option_usize_ret_small()

conv_vec_option_i8_ret()

conv_vec_option_i16_ret()

conv_vec_option_u16_ret()

conv_option_i64_some_small()

conv_option_i64_some_big()

conv_option_i64_none()

conv_option_f32_some()

conv_option_u32_some()

conv_as_named_list_vec()

conv_as_named_list_array()

conv_as_named_list_heterogeneous()

conv_as_named_list_str_keys()

conv_as_named_list_empty()

conv_as_named_list_duplicate_names()

conv_as_named_vector_i32()

conv_as_named_vector_f64()

conv_as_named_vector_string()

conv_as_named_vector_option_i32()

conv_as_named_vector_array()

conv_as_named_vector_empty()

conv_as_named_vector_ext_trait()

conv_as_named_list_ext_trait()

conv_as_named_list_slice()

conv_as_named_vector_slice()

conv_named_list_get(config)

conv_named_list_contains(config)

conv_named_list_len(config)

conv_named_list_roundtrip(config)

conv_tuple1_roundtrip(solo)

conv_tuple2_roundtrip(pair)

conv_tuple3_sum(triple)
}
\arguments{
\item{x}{Input numeric vector (may contain NA).}

\item{a}{First integer vector (mutated in-place).}

\item{b}{Second integer vector (mutated in-place).}

\item{config}{Input named list.}

\item{solo}{Unnamed list containing a single integer.}

\item{pair}{Unnamed list of an integer and a string.}

\item{triple}{Unnamed list of a double, a logical, and an integer vector.}
}
\value{
\code{TRUE} if \code{debug_assertions} is on (debug build), else \code{FALSE}.
}
\description{
Roundtrip test: R integer -> Rust \code{i32} -> R integer.

Return test: Rust \code{i32} -> R integer.

Roundtrip test: R double -> Rust \code{f64} -> R double.

Return test: Rust \code{f64} -> R double.

Roundtrip test: R raw element -> Rust \code{u8} -> R raw element.

Return test: Rust \code{u8} -> R raw element.

Roundtrip test: R logical -> Rust \code{Rboolean} -> R logical.

Return test: Rust \code{Rboolean} -> R logical.

Roundtrip test: R logical -> Rust \code{RLogical} -> R logical.

Return test: Rust \code{RLogical} -> R logical.

Roundtrip test: R character scalar -> Rust \code{String} -> R character scalar.

Return test: Rust \code{String} -> R character scalar.

Return test: Rust \verb{&str} -> R character scalar.

Roundtrip test: R SEXP -> Rust \code{SEXP} -> R SEXP (passthrough).

Return test: Rust \code{SEXP} -> R (via \code{IntoR}).

Roundtrip test: R numeric -> Rust \code{i64} -> R numeric.

Return test: Rust \code{i64} -> R numeric.

Roundtrip test: R numeric -> Rust \code{u64} -> R numeric.

Return test: Rust \code{u64} -> R numeric.

Roundtrip test: R numeric -> Rust \code{isize} -> R numeric.

Return test: Rust \code{isize} -> R numeric.

Roundtrip test: R numeric -> Rust \code{usize} -> R numeric.

Return test: Rust \code{usize} -> R numeric.

Roundtrip test: R integer -> Rust \code{i8} -> R integer.

Return test: Rust \code{i8} -> R integer.

Roundtrip test: R integer -> Rust \code{i16} -> R integer.

Return test: Rust \code{i16} -> R integer.

Roundtrip test: R integer -> Rust \code{u16} -> R integer.

Return test: Rust \code{u16} -> R integer.

Roundtrip test: R integer -> Rust \code{u32} -> R integer.

Return test: Rust \code{u32} -> R integer.

Roundtrip test: R double -> Rust \code{f32} -> R double.

Return test: Rust \code{f32} -> R double.

Test \verb{Option<i32>} argument: returns 1 if Some, 0 if None (NA).

Return test: \code{Some(10)} as \verb{Option<i32>} -> R integer.

Return test: \code{None} as \verb{Option<i32>} -> R NA_integer_.

Test \verb{Option<f64>} argument: returns 1 if Some, 0 if None (NA).

Return test: \code{Some(2.5)} as \verb{Option<f64>} -> R double.

Return test: \code{None} as \verb{Option<f64>} -> R NA_real_.

Test \verb{Option<bool>} argument: returns 1 if Some, 0 if None (NA).

Return test: \code{Some(true)} as \verb{Option<bool>} -> R logical.

Return test: \code{None} as \verb{Option<bool>} -> R NA.

Test \verb{Option<String>} argument: returns 1 if Some, 0 if None (NA).

Return test: \code{Some("opt")} as \verb{Option<String>} -> R character scalar.

Return test: \code{None} as \verb{Option<String>} -> R NA_character_.

Test \verb{Option<i8>} argument: returns 1 if Some, 0 if None (NA).

Test \verb{Option<i16>} argument: returns 1 if Some, 0 if None (NA).

Test \verb{Option<i64>} argument: returns 1 if Some, 0 if None (NA).

Test \verb{Option<isize>} argument: returns 1 if Some, 0 if None (NA).

Test \verb{Option<u16>} argument: returns 1 if Some, 0 if None (NA).

Test \verb{Option<u32>} argument: returns 1 if Some, 0 if None (NA).

Test \verb{Option<u64>} argument: returns 1 if Some, 0 if None (NA).

Test \verb{Option<usize>} argument: returns 1 if Some, 0 if None (NA).

Test \verb{Option<f32>} argument: returns 1 if Some, 0 if None (NA).

Test \verb{Option<u8>} argument: returns 1 if Some, 0 if None (NA).

Test \verb{Option<Rboolean>} argument: returns 1 if Some, 0 if None (NA).

Test \verb{Option<RLogical>} argument: returns 1 if Some, 0 if None (NA).

Test \verb{&[i32]} slice argument: returns the length of the slice.

Test \verb{&[f64]} slice argument: returns the length of the slice.

Test \verb{&[u8]} slice argument: returns the length of the slice.

Test \verb{&[RLogical]} slice argument: returns the length of the slice.

Test \verb{&i32} reference argument: dereferences and returns the value.

Test \verb{&mut i32} mutable reference: increments by one and returns.

Test \verb{&mut [i32]} mutable slice: increments each element and returns length.

Test \verb{&mut [u8]} mutable slice: wrapping-adds one to each element.

Test \verb{Option<&i32>} argument: returns 1 if Some, 0 if None.

Test \verb{Option<&mut [i32]>} argument: returns 1 if Some, 0 if None.

Test \verb{Vec<&i32>} argument: returns the length of the vector.

Test \verb{Vec<&[i32]>} argument: returns the total length across all slices.

Test \verb{Vec<&mut [i32]>} argument: increments all elements and returns total length.

Aliasing probe for the pre-conversion mutable-slice guard (#1104, #1252).

Takes two mutable integer slices and increments each. Passing the same non-empty R vector to both parameters raises an error before conversion in both debug and release builds.

Whether this build of the package compiled with \code{debug_assertions} enabled.

Runtime tests use this to verify which compiler profile is installed.

Test \verb{Vec<i32>} argument: returns the length of the vector.

Return test: Rust \verb{Vec<i32>} -> R integer vector.

Test \verb{Vec<f64>} argument: returns the length of the vector.

Return test: Rust \verb{Vec<f64>} -> R double vector.

Test \verb{Vec<u8>} argument: returns the length of the vector.

Return test: Rust \verb{Vec<u8>} -> R raw vector.

Test \verb{Vec<RLogical>} argument: returns the length of the vector.

Return test: Rust \verb{Vec<RLogical>} -> R logical vector.

Test \verb{Vec<bool>} argument: returns the length of the vector.

Return test: Rust \verb{Vec<bool>} -> R logical vector.

Test \verb{Vec<String>} argument: returns the length of the vector.

Return test: Rust \verb{Vec<String>} -> R character vector.

Test \verb{Vec<i8>} argument: returns the length of the vector.

Test \verb{Vec<i16>} argument: returns the length of the vector.

Test \verb{Vec<i64>} argument: returns the length of the vector.

Test \verb{Vec<isize>} argument: returns the length of the vector.

Test \verb{Vec<u16>} argument: returns the length of the vector.

Test \verb{Vec<u32>} argument: returns the length of the vector.

Test \verb{Vec<u64>} argument: returns the length of the vector.

Test \verb{Vec<usize>} argument: returns the length of the vector.

Test \verb{Vec<f32>} argument: returns the length of the vector.

Test \verb{Vec<Option<i32>>} argument: returns the length of the vector.

Return test: Rust \verb{Vec<Option<i32>>} -> R integer vector with NA.

Test \verb{Vec<Option<f64>>} argument: returns the length of the vector.

Return test: Rust \verb{Vec<Option<f64>>} -> R double vector with NA.

Test \verb{Vec<Option<bool>>} argument: returns the length of the vector.

Return test: Rust \verb{Vec<Option<bool>>} -> R logical vector with NA.

Test \verb{Vec<Option<String>>} argument: returns the length of the vector.

Return test: Rust \verb{Vec<Option<String>>} -> R character vector with NA.

Return test: Rust \verb{Vec<Option<RLogical>>} -> R logical vector with NA.

Return test: Rust \verb{Vec<Option<Rboolean>>} -> R logical vector with NA.

Test \verb{Vec<Option<i8>>} argument: returns the length of the vector.

Test \verb{Vec<Option<u8>>} argument: returns the length of the vector.

Test \verb{HashSet<i32>} argument: returns the number of unique elements.

Return test: Rust \verb{HashSet<i32>} -> R integer vector.

Test \verb{HashSet<u8>} argument: returns the number of unique elements.

Return test: Rust \verb{HashSet<u8>} -> R raw vector.

Test \verb{HashSet<String>} argument: returns the number of unique elements.

Return test: Rust \verb{HashSet<String>} -> R character vector.

Test \verb{HashSet<RLogical>} argument: returns the number of unique elements.

Return test: Rust \verb{HashSet<RLogical>} -> R logical vector.

Test \verb{BTreeSet<i32>} argument: returns the number of unique sorted elements.

Return test: Rust \verb{BTreeSet<i32>} -> R integer vector (sorted).

Test \verb{BTreeSet<u8>} argument: returns the number of unique sorted elements.

Return test: Rust \verb{BTreeSet<u8>} -> R raw vector (sorted).

Test \verb{BTreeSet<String>} argument: returns the number of unique sorted elements.

Return test: Rust \verb{BTreeSet<String>} -> R character vector (sorted).

Test \verb{HashMap<String, i32>} argument: returns the number of entries.

Return test: Rust \verb{HashMap<String, i32>} -> R named list of integers.

Test \verb{HashMap<String, f64>} argument: returns the number of entries.

Return test: Rust \verb{HashMap<String, f64>} -> R named list of doubles.

Test \verb{HashMap<String, String>} argument: returns the number of entries.

Return test: Rust \verb{HashMap<String, String>} -> R named list of strings.

Test \verb{HashMap<String, RLogical>} argument: returns the number of entries.

Return test: Rust \verb{HashMap<String, RLogical>} -> R named list of logicals.

Test \verb{BTreeMap<String, i32>} argument: returns the number of entries.

Return test: Rust \verb{BTreeMap<String, i32>} -> R named list of integers (sorted).

Test \verb{BTreeMap<String, f64>} argument: returns the number of entries.

Return test: Rust \verb{BTreeMap<String, f64>} -> R named list of doubles (sorted).

Test \verb{BTreeMap<String, String>} argument: returns the number of entries.

Return test: Rust \verb{BTreeMap<String, String>} -> R named list of strings (sorted).

Test \verb{BTreeMap<String, RLogical>} argument: returns the number of entries.

Return test: Rust \verb{BTreeMap<String, RLogical>} -> R named list of logicals (sorted).

Test \code{ListMut} argument: sets the first element to 99 and returns the length.

Test \verb{Result<i32, ()>} argument: returns 1 if Ok, 0 if Err (NULL).

Test \verb{Result<Vec<i32>, ()>} argument: returns 1 if Ok, 0 if Err (NULL).

Return test: \code{Ok(9)} as \verb{Result<i32, ()>} -> R integer.

Return test: \verb{Err(())} as \verb{Result<i32, ()>} -> R NULL.

Return test: \code{Ok(9.5)} as \verb{Result<f64, ()>} -> R double.

Return test: \verb{Err(())} as \verb{Result<f64, ()>} -> R NULL.

Return test: \code{Ok("ok")} as \verb{Result<String, ()>} -> R character scalar.

Return test: \verb{Err(())} as \verb{Result<String, ()>} -> R NULL.

Return test: \verb{Ok(vec![1, 2])} as \verb{Result<Vec<i32>, ()>} -> R integer vector.

Return test: \verb{Err(())} as \verb{Result<Vec<i32>, ()>} -> R NULL.

Roundtrip test: R character scalar -> Rust \code{char} -> R character scalar.

Return test: Rust \code{char} (Unicode) -> R character scalar.

Return test: Rust \verb{Vec<i8>} -> R integer vector (coerced via i32).

Return test: Rust \verb{Vec<i16>} -> R integer vector (coerced via i32).

Return test: Rust \verb{Vec<u16>} -> R integer vector (coerced via i32).

Return test: Rust \verb{Vec<f32>} -> R double vector (coerced via f64).

Return test: Rust \verb{HashSet<i8>} -> R integer vector (coerced via i32).

Return test: Rust \verb{BTreeSet<i8>} -> R integer vector (coerced via i32, sorted).

Return test: Rust \verb{HashSet<i16>} -> R integer vector (coerced via i32).

Return test: Rust \verb{BTreeSet<i16>} -> R integer vector (coerced via i32, sorted).

Return test: Rust \verb{HashSet<u16>} -> R integer vector (coerced via i32).

Return test: Rust \verb{BTreeSet<u16>} -> R integer vector (coerced via i32, sorted).

Return test: \verb{Some(&42)} as \verb{Option<&i32>} -> R integer.

Return test: \code{None} as \verb{Option<&i32>} -> R NULL.

Test \verb{Option<Vec<i32>>} argument: returns sum if Some, -999 if None (NULL).

Return test: \verb{Some(vec![1, 2, 3])} as \verb{Option<Vec<i32>>} -> R integer vector.

Return test: \code{None} as \verb{Option<Vec<i32>>} -> R NULL.

Test \verb{Option<Vec<String>>} argument: returns length if Some, -999 if None (NULL).

Return test: \verb{Some(vec!["a", "b"])} as \verb{Option<Vec<String>>} -> R character vector.

Return test: \code{None} as \verb{Option<Vec<String>>} -> R NULL.

Test \verb{Option<HashMap<String, i32>>} argument: returns sum if Some, -999 if None.

Return test: \code{Some(HashMap)} as \verb{Option<HashMap<String, i32>>} -> R named list.

Return test: \code{None} as \verb{Option<HashMap<String, i32>>} -> R NULL.

Test \verb{Option<HashSet<i32>>} argument: returns sum if Some, -999 if None.

Return test: \code{Some(HashSet)} as \verb{Option<HashSet<i32>>} -> R integer vector.

Return test: \code{None} as \verb{Option<HashSet<i32>>} -> R NULL.

Test \verb{Option<BTreeMap<String, i32>>} argument: returns sum if Some, -999 if None.

Return test: \code{Some(BTreeMap)} as \verb{Option<BTreeMap<String, i32>>} -> R named list.

Return test: \code{None} as \verb{Option<BTreeMap<String, i32>>} -> R NULL.

Test \verb{Vec<HashMap<String, i32>>} argument: returns the sum across all maps.

Return test: Rust \verb{Vec<HashMap<String, i32>>} -> R list of named lists.

Test \verb{Vec<BTreeMap<String, i32>>} argument: returns the sum across all maps.

Return test: Rust \verb{Vec<BTreeMap<String, i32>>} -> R list of named lists (sorted).

Test \verb{Vec<Vec<i32>>} argument: returns the sum across all inner vectors.

Return test: Rust \verb{Vec<Vec<i32>>} -> R list of integer vectors.

Return test: Rust \verb{Vec<Vec<String>>} -> R list of character vectors.

Return test: Rust \verb{Vec<Option<i64>>} with small values -> R integer vector with NA.

Return test: Rust \verb{Vec<Option<i64>>} with large values -> R numeric vector with NA.

Return test: Rust \verb{Vec<Option<u32>>} -> R integer vector with NA.

Return test: Rust \verb{Vec<Option<f32>>} -> R double vector with NA.

Roundtrip test: R numeric vector with NA -> Rust \verb{Vec<Option<i64>>} -> R.

Return test: Rust \verb{Vec<Option<u64>>} with small values -> R integer vector with NA.

Return test: Rust \verb{Vec<Option<u64>>} with large values -> R numeric vector with NA.

Return test: Rust \verb{Vec<Option<isize>>} with small values -> R integer vector with NA.

Return test: Rust \verb{Vec<Option<usize>>} with small values -> R integer vector with NA.

Return test: Rust \verb{Vec<Option<i8>>} -> R integer vector with NA.

Return test: Rust \verb{Vec<Option<i16>>} -> R integer vector with NA.

Return test: Rust \verb{Vec<Option<u16>>} -> R integer vector with NA.

Return test: \code{Some(42)} as \verb{Option<i64>} -> R integer (fits in i32).

Return test: \code{Some(i64::MAX)} as \verb{Option<i64>} -> R double (too large for i32).

Return test: \code{None} as \verb{Option<i64>} -> R NA.

Return test: \code{Some(1.5)} as \verb{Option<f32>} -> R double.

Return test: \code{Some(100)} as \verb{Option<u32>} -> R integer.

Return test: \verb{AsNamedList<Vec<(String, i32)>>} -> R named list.

Return test: \verb{AsNamedList<[(String, f64); 2]>} -> R named list from fixed-size array.

Return test: \code{AsNamedList} with heterogeneous typed values -> R named list.

Previously this fixture used raw \code{SEXP} values (\verb{Vec<(String, SEXP)>}) and suppressed MXL302, because the deferred \code{AsNamedList::into_sexp} ran in a later frame than the value construction — a structural GC hazard with no fixable call site (#1030). It is migrated to typed values via the framework \code{RValue} tree: each value's \code{into_sexp()} runs \emph{inside} \code{AsNamedList::into_sexp}, immediately before being protected and placed into the list, so no sibling is ever unprotected across an allocation. The MXL302 suppression is gone because there is no longer an \code{into_sexp()} call inside the \verb{vec!} literal.

Return test: \verb{AsNamedList<Vec<(&str, i32)>>} -> R named list with \verb{&str} keys.

Return test: empty \code{AsNamedList} -> R empty named list.

Return test: \code{AsNamedList} with duplicate key names -> R named list (duplicates preserved).

Return test: \verb{AsNamedVector<Vec<(String, i32)>>} -> R named integer vector.

Return test: \verb{AsNamedVector<Vec<(&str, f64)>>} -> R named double vector.

Return test: \verb{AsNamedVector<Vec<(String, String)>>} -> R named character vector.

Return test: \verb{AsNamedVector<Vec<(&str, Option<i32>)>>} -> R named integer vector with NA.

Return test: \code{AsNamedVector} from fixed-size array -> R named double vector.

Return test: empty \code{AsNamedVector} -> R empty named double vector.

Return test: \code{AsNamedVectorExt} trait method -> R named integer vector.

Return test: \code{AsNamedListExt} trait method -> R named list.

Return test: \code{AsNamedList} from a borrowed slice -> R named list via SEXP.

Return test: \code{AsNamedVector} from a borrowed slice -> R named double vector via SEXP.

Test \code{NamedList} O(1) lookup: retrieves "width" and "height" and returns their product.

Test \code{NamedList::contains}: checks existence of "a", "b", and "missing".

Test \code{NamedList} length queries: returns total length and named-element count.

Test \code{NamedList} roundtrip: converts to \code{List} and back to R.

Test 1-tuple argument round-trip: reads a 1-element unnamed list as \verb{(i32,)} and returns it.

Test tuple argument round-trip: reads an unnamed 2-element list as \verb{(i32, String)} and returns it.

Test 3-tuple argument: reads \verb{(f64, bool, Vec<i32>)} and returns the sum of all numeric parts.
}
