% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{conv_i8_arg}
\alias{conv_i8_arg}
\alias{conv_i8_ret}
\alias{conv_u8_arg}
\alias{conv_u8_ret}
\alias{conv_f32_arg}
\alias{conv_f32_ret}
\alias{conv_f64_arg}
\alias{conv_f64_ret}
\alias{conv_i16_arg}
\alias{conv_i16_ret}
\alias{conv_i32_arg}
\alias{conv_i32_ret}
\alias{conv_i64_arg}
\alias{conv_i64_ret}
\alias{conv_str_ret}
\alias{conv_u16_arg}
\alias{conv_u16_ret}
\alias{conv_u32_arg}
\alias{conv_u32_ret}
\alias{conv_u64_arg}
\alias{conv_u64_ret}
\alias{conv_char_arg}
\alias{conv_char_ret}
\alias{conv_rlog_arg}
\alias{conv_rlog_ret}
\alias{conv_sexp_arg}
\alias{conv_sexp_ret}
\alias{conv_isize_arg}
\alias{conv_isize_ret}
\alias{conv_rbool_arg}
\alias{conv_rbool_ret}
\alias{conv_usize_arg}
\alias{conv_usize_ret}
\alias{conv_string_arg}
\alias{conv_string_ret}
\alias{conv_vec_i8_len}
\alias{conv_vec_i8_ret}
\alias{conv_vec_u8_len}
\alias{conv_vec_u8_ret}
\alias{conv_ref_i32_arg}
\alias{conv_vec_f32_len}
\alias{conv_vec_f32_ret}
\alias{conv_vec_f64_len}
\alias{conv_vec_f64_ret}
\alias{conv_vec_i16_len}
\alias{conv_vec_i16_ret}
\alias{conv_vec_i32_len}
\alias{conv_vec_i32_ret}
\alias{conv_vec_i64_len}
\alias{conv_vec_u16_len}
\alias{conv_vec_u16_ret}
\alias{conv_vec_u32_len}
\alias{conv_vec_u64_len}
\alias{conv_opt_f64_none}
\alias{conv_opt_f64_some}
\alias{conv_opt_i32_none}
\alias{conv_opt_i32_some}
\alias{conv_slice_u8_len}
\alias{conv_vec_bool_len}
\alias{conv_vec_bool_ret}
\alias{conv_vec_rlog_len}
\alias{conv_vec_rlog_ret}
\alias{conv_opt_bool_none}
\alias{conv_opt_bool_some}
\alias{conv_result_f64_ok}
\alias{conv_result_i32_ok}
\alias{conv_slice_f64_len}
\alias{conv_slice_i32_len}
\alias{conv_vec_isize_len}
\alias{conv_vec_usize_len}
\alias{conv_hashset_i8_ret}
\alias{conv_hashset_u8_len}
\alias{conv_hashset_u8_ret}
\alias{conv_named_list_get}
\alias{conv_named_list_len}
\alias{conv_opt_i8_is_some}
\alias{conv_opt_u8_is_some}
\alias{conv_result_f64_err}
\alias{conv_result_i32_arg}
\alias{conv_result_i32_err}
\alias{conv_slice_rlog_len}
\alias{conv_vec_opt_i8_len}
\alias{conv_vec_opt_u8_len}
\alias{conv_vec_string_len}
\alias{conv_vec_string_ret}
\alias{conv_btreeset_i8_ret}
\alias{conv_btreeset_u8_len}
\alias{conv_btreeset_u8_ret}
\alias{conv_hashmap_f64_len}
\alias{conv_hashmap_f64_ret}
\alias{conv_hashmap_i32_len}
\alias{conv_hashmap_i32_ret}
\alias{conv_hashset_i16_ret}
\alias{conv_hashset_i32_len}
\alias{conv_hashset_i32_ret}
\alias{conv_hashset_u16_ret}
\alias{conv_option_f32_some}
\alias{conv_option_i64_none}
\alias{conv_option_u32_some}
\alias{conv_opt_f32_is_some}
\alias{conv_opt_f64_is_some}
\alias{conv_opt_i16_is_some}
\alias{conv_opt_i32_is_some}
\alias{conv_opt_i64_is_some}
\alias{conv_opt_string_none}
\alias{conv_opt_string_some}
\alias{conv_opt_u16_is_some}
\alias{conv_opt_u32_is_some}
\alias{conv_opt_u64_is_some}
\alias{conv_opt_vec_i32_arg}
\alias{conv_vec_opt_f64_len}
\alias{conv_vec_opt_f64_ret}
\alias{conv_vec_opt_i32_len}
\alias{conv_vec_opt_i32_ret}
\alias{conv_vec_ref_i32_len}
\alias{conv_vec_vec_i32_arg}
\alias{conv_vec_vec_i32_ret}
\alias{conv_btreemap_f64_len}
\alias{conv_btreemap_f64_ret}
\alias{conv_btreemap_i32_len}
\alias{conv_btreemap_i32_ret}
\alias{conv_btreeset_i16_ret}
\alias{conv_btreeset_i32_len}
\alias{conv_btreeset_i32_ret}
\alias{conv_btreeset_u16_ret}
\alias{conv_hashmap_rlog_len}
\alias{conv_hashmap_rlog_ret}
\alias{conv_hashset_rlog_len}
\alias{conv_hashset_rlog_ret}
\alias{conv_opt_bool_is_some}
\alias{conv_opt_rlog_is_some}
\alias{conv_result_string_ok}
\alias{conv_vec_opt_bool_len}
\alias{conv_vec_opt_bool_ret}
\alias{conv_vec_opt_rlog_ret}
\alias{conv_as_named_list_vec}
\alias{conv_btreemap_rlog_len}
\alias{conv_btreemap_rlog_ret}
\alias{conv_opt_isize_is_some}
\alias{conv_opt_rbool_is_some}
\alias{conv_opt_usize_is_some}
\alias{conv_result_string_err}
\alias{conv_result_vec_i32_ok}
\alias{conv_vec_option_i8_ret}
\alias{conv_vec_opt_rbool_ret}
\alias{conv_hashmap_string_len}
\alias{conv_hashmap_string_ret}
\alias{conv_hashset_string_len}
\alias{conv_hashset_string_ret}
\alias{conv_list_mut_set_first}
\alias{conv_opt_string_is_some}
\alias{conv_opt_vec_string_arg}
\alias{conv_result_vec_i32_arg}
\alias{conv_result_vec_i32_err}
\alias{conv_vec_option_f32_ret}
\alias{conv_vec_option_i16_ret}
\alias{conv_vec_option_u16_ret}
\alias{conv_vec_option_u32_ret}
\alias{conv_vec_opt_string_len}
\alias{conv_vec_opt_string_ret}
\alias{conv_vec_vec_string_ret}
\alias{conv_as_named_list_array}
\alias{conv_as_named_list_empty}
\alias{conv_as_named_list_slice}
\alias{conv_as_named_vector_f64}
\alias{conv_as_named_vector_i32}
\alias{conv_btreemap_string_len}
\alias{conv_btreemap_string_ret}
\alias{conv_btreeset_string_len}
\alias{conv_btreeset_string_ret}
\alias{conv_named_list_contains}
\alias{conv_option_i64_some_big}
\alias{conv_opt_hashmap_i32_arg}
\alias{conv_opt_hashset_i32_arg}
\alias{conv_opt_ref_i32_is_some}
\alias{conv_ref_mut_i32_add_one}
\alias{conv_vec_hashmap_i32_arg}
\alias{conv_vec_hashmap_i32_ret}
\alias{conv_named_list_roundtrip}
\alias{conv_opt_btreemap_i32_arg}
\alias{conv_opt_ref_i32_none_ret}
\alias{conv_opt_ref_i32_some_ret}
\alias{conv_opt_vec_i32_none_ret}
\alias{conv_opt_vec_i32_some_ret}
\alias{conv_slice_mut_u8_add_one}
\alias{conv_vec_btreemap_i32_arg}
\alias{conv_vec_btreemap_i32_ret}
\alias{conv_as_named_vector_array}
\alias{conv_as_named_vector_empty}
\alias{conv_as_named_vector_slice}
\alias{conv_option_i64_some_small}
\alias{conv_slice_mut_i32_add_one}
\alias{conv_as_named_list_str_keys}
\alias{conv_as_named_vector_string}
\alias{conv_vec_option_i64_ret_big}
\alias{conv_vec_option_u64_ret_big}
\alias{conv_as_named_list_ext_trait}
\alias{conv_opt_vec_string_none_ret}
\alias{conv_opt_vec_string_some_ret}
\alias{conv_vec_slice_i32_total_len}
\alias{conv_opt_hashmap_i32_none_ret}
\alias{conv_opt_hashmap_i32_some_ret}
\alias{conv_opt_hashset_i32_none_ret}
\alias{conv_opt_hashset_i32_some_ret}
\alias{conv_vec_option_i64_ret_small}
\alias{conv_vec_option_i64_roundtrip}
\alias{conv_vec_option_u64_ret_small}
\alias{conv_as_named_vector_ext_trait}
\alias{conv_opt_btreemap_i32_none_ret}
\alias{conv_opt_btreemap_i32_some_ret}
\alias{conv_opt_mut_slice_i32_is_some}
\alias{conv_vec_mut_slice_i32_add_one}
\alias{conv_as_named_vector_option_i32}
\alias{conv_vec_option_isize_ret_small}
\alias{conv_vec_option_usize_ret_small}
\alias{conv_as_named_list_heterogeneous}
\alias{conv_as_named_list_duplicate_names}
\title{Roundtrip test: R integer -> Rust `i8` -> R integer}
\source{
Generated by miniextendr from Rust fn `conv_i8_arg`

Generated by miniextendr from Rust fn `conv_i8_ret`

Generated by miniextendr from Rust fn `conv_u8_arg`

Generated by miniextendr from Rust fn `conv_u8_ret`

Generated by miniextendr from Rust fn `conv_f32_arg`

Generated by miniextendr from Rust fn `conv_f32_ret`

Generated by miniextendr from Rust fn `conv_f64_arg`

Generated by miniextendr from Rust fn `conv_f64_ret`

Generated by miniextendr from Rust fn `conv_i16_arg`

Generated by miniextendr from Rust fn `conv_i16_ret`

Generated by miniextendr from Rust fn `conv_i32_arg`

Generated by miniextendr from Rust fn `conv_i32_ret`

Generated by miniextendr from Rust fn `conv_i64_arg`

Generated by miniextendr from Rust fn `conv_i64_ret`

Generated by miniextendr from Rust fn `conv_str_ret`

Generated by miniextendr from Rust fn `conv_u16_arg`

Generated by miniextendr from Rust fn `conv_u16_ret`

Generated by miniextendr from Rust fn `conv_u32_arg`

Generated by miniextendr from Rust fn `conv_u32_ret`

Generated by miniextendr from Rust fn `conv_u64_arg`

Generated by miniextendr from Rust fn `conv_u64_ret`

Generated by miniextendr from Rust fn `conv_char_arg`

Generated by miniextendr from Rust fn `conv_char_ret`

Generated by miniextendr from Rust fn `conv_rlog_arg`

Generated by miniextendr from Rust fn `conv_rlog_ret`

Generated by miniextendr from Rust fn `conv_sexp_arg`

Generated by miniextendr from Rust fn `conv_sexp_ret`

Generated by miniextendr from Rust fn `conv_isize_arg`

Generated by miniextendr from Rust fn `conv_isize_ret`

Generated by miniextendr from Rust fn `conv_rbool_arg`

Generated by miniextendr from Rust fn `conv_rbool_ret`

Generated by miniextendr from Rust fn `conv_usize_arg`

Generated by miniextendr from Rust fn `conv_usize_ret`

Generated by miniextendr from Rust fn `conv_string_arg`

Generated by miniextendr from Rust fn `conv_string_ret`

Generated by miniextendr from Rust fn `conv_vec_i8_len`

Generated by miniextendr from Rust fn `conv_vec_i8_ret`

Generated by miniextendr from Rust fn `conv_vec_u8_len`

Generated by miniextendr from Rust fn `conv_vec_u8_ret`

Generated by miniextendr from Rust fn `conv_ref_i32_arg`

Generated by miniextendr from Rust fn `conv_vec_f32_len`

Generated by miniextendr from Rust fn `conv_vec_f32_ret`

Generated by miniextendr from Rust fn `conv_vec_f64_len`

Generated by miniextendr from Rust fn `conv_vec_f64_ret`

Generated by miniextendr from Rust fn `conv_vec_i16_len`

Generated by miniextendr from Rust fn `conv_vec_i16_ret`

Generated by miniextendr from Rust fn `conv_vec_i32_len`

Generated by miniextendr from Rust fn `conv_vec_i32_ret`

Generated by miniextendr from Rust fn `conv_vec_i64_len`

Generated by miniextendr from Rust fn `conv_vec_u16_len`

Generated by miniextendr from Rust fn `conv_vec_u16_ret`

Generated by miniextendr from Rust fn `conv_vec_u32_len`

Generated by miniextendr from Rust fn `conv_vec_u64_len`

Generated by miniextendr from Rust fn `conv_opt_f64_none`

Generated by miniextendr from Rust fn `conv_opt_f64_some`

Generated by miniextendr from Rust fn `conv_opt_i32_none`

Generated by miniextendr from Rust fn `conv_opt_i32_some`

Generated by miniextendr from Rust fn `conv_slice_u8_len`

Generated by miniextendr from Rust fn `conv_vec_bool_len`

Generated by miniextendr from Rust fn `conv_vec_bool_ret`

Generated by miniextendr from Rust fn `conv_vec_rlog_len`

Generated by miniextendr from Rust fn `conv_vec_rlog_ret`

Generated by miniextendr from Rust fn `conv_opt_bool_none`

Generated by miniextendr from Rust fn `conv_opt_bool_some`

Generated by miniextendr from Rust fn `conv_result_f64_ok`

Generated by miniextendr from Rust fn `conv_result_i32_ok`

Generated by miniextendr from Rust fn `conv_slice_f64_len`

Generated by miniextendr from Rust fn `conv_slice_i32_len`

Generated by miniextendr from Rust fn `conv_vec_isize_len`

Generated by miniextendr from Rust fn `conv_vec_usize_len`

Generated by miniextendr from Rust fn `conv_hashset_i8_ret`

Generated by miniextendr from Rust fn `conv_hashset_u8_len`

Generated by miniextendr from Rust fn `conv_hashset_u8_ret`

Generated by miniextendr from Rust fn `conv_named_list_get`

Generated by miniextendr from Rust fn `conv_named_list_len`

Generated by miniextendr from Rust fn `conv_opt_i8_is_some`

Generated by miniextendr from Rust fn `conv_opt_u8_is_some`

Generated by miniextendr from Rust fn `conv_result_f64_err`

Generated by miniextendr from Rust fn `conv_result_i32_arg`

Generated by miniextendr from Rust fn `conv_result_i32_err`

Generated by miniextendr from Rust fn `conv_slice_rlog_len`

Generated by miniextendr from Rust fn `conv_vec_opt_i8_len`

Generated by miniextendr from Rust fn `conv_vec_opt_u8_len`

Generated by miniextendr from Rust fn `conv_vec_string_len`

Generated by miniextendr from Rust fn `conv_vec_string_ret`

Generated by miniextendr from Rust fn `conv_btreeset_i8_ret`

Generated by miniextendr from Rust fn `conv_btreeset_u8_len`

Generated by miniextendr from Rust fn `conv_btreeset_u8_ret`

Generated by miniextendr from Rust fn `conv_hashmap_f64_len`

Generated by miniextendr from Rust fn `conv_hashmap_f64_ret`

Generated by miniextendr from Rust fn `conv_hashmap_i32_len`

Generated by miniextendr from Rust fn `conv_hashmap_i32_ret`

Generated by miniextendr from Rust fn `conv_hashset_i16_ret`

Generated by miniextendr from Rust fn `conv_hashset_i32_len`

Generated by miniextendr from Rust fn `conv_hashset_i32_ret`

Generated by miniextendr from Rust fn `conv_hashset_u16_ret`

Generated by miniextendr from Rust fn `conv_option_f32_some`

Generated by miniextendr from Rust fn `conv_option_i64_none`

Generated by miniextendr from Rust fn `conv_option_u32_some`

Generated by miniextendr from Rust fn `conv_opt_f32_is_some`

Generated by miniextendr from Rust fn `conv_opt_f64_is_some`

Generated by miniextendr from Rust fn `conv_opt_i16_is_some`

Generated by miniextendr from Rust fn `conv_opt_i32_is_some`

Generated by miniextendr from Rust fn `conv_opt_i64_is_some`

Generated by miniextendr from Rust fn `conv_opt_string_none`

Generated by miniextendr from Rust fn `conv_opt_string_some`

Generated by miniextendr from Rust fn `conv_opt_u16_is_some`

Generated by miniextendr from Rust fn `conv_opt_u32_is_some`

Generated by miniextendr from Rust fn `conv_opt_u64_is_some`

Generated by miniextendr from Rust fn `conv_opt_vec_i32_arg`

Generated by miniextendr from Rust fn `conv_vec_opt_f64_len`

Generated by miniextendr from Rust fn `conv_vec_opt_f64_ret`

Generated by miniextendr from Rust fn `conv_vec_opt_i32_len`

Generated by miniextendr from Rust fn `conv_vec_opt_i32_ret`

Generated by miniextendr from Rust fn `conv_vec_ref_i32_len`

Generated by miniextendr from Rust fn `conv_vec_vec_i32_arg`

Generated by miniextendr from Rust fn `conv_vec_vec_i32_ret`

Generated by miniextendr from Rust fn `conv_btreemap_f64_len`

Generated by miniextendr from Rust fn `conv_btreemap_f64_ret`

Generated by miniextendr from Rust fn `conv_btreemap_i32_len`

Generated by miniextendr from Rust fn `conv_btreemap_i32_ret`

Generated by miniextendr from Rust fn `conv_btreeset_i16_ret`

Generated by miniextendr from Rust fn `conv_btreeset_i32_len`

Generated by miniextendr from Rust fn `conv_btreeset_i32_ret`

Generated by miniextendr from Rust fn `conv_btreeset_u16_ret`

Generated by miniextendr from Rust fn `conv_hashmap_rlog_len`

Generated by miniextendr from Rust fn `conv_hashmap_rlog_ret`

Generated by miniextendr from Rust fn `conv_hashset_rlog_len`

Generated by miniextendr from Rust fn `conv_hashset_rlog_ret`

Generated by miniextendr from Rust fn `conv_opt_bool_is_some`

Generated by miniextendr from Rust fn `conv_opt_rlog_is_some`

Generated by miniextendr from Rust fn `conv_result_string_ok`

Generated by miniextendr from Rust fn `conv_vec_opt_bool_len`

Generated by miniextendr from Rust fn `conv_vec_opt_bool_ret`

Generated by miniextendr from Rust fn `conv_vec_opt_rlog_ret`

Generated by miniextendr from Rust fn `conv_as_named_list_vec`

Generated by miniextendr from Rust fn `conv_btreemap_rlog_len`

Generated by miniextendr from Rust fn `conv_btreemap_rlog_ret`

Generated by miniextendr from Rust fn `conv_opt_isize_is_some`

Generated by miniextendr from Rust fn `conv_opt_rbool_is_some`

Generated by miniextendr from Rust fn `conv_opt_usize_is_some`

Generated by miniextendr from Rust fn `conv_result_string_err`

Generated by miniextendr from Rust fn `conv_result_vec_i32_ok`

Generated by miniextendr from Rust fn `conv_vec_option_i8_ret`

Generated by miniextendr from Rust fn `conv_vec_opt_rbool_ret`

Generated by miniextendr from Rust fn `conv_hashmap_string_len`

Generated by miniextendr from Rust fn `conv_hashmap_string_ret`

Generated by miniextendr from Rust fn `conv_hashset_string_len`

Generated by miniextendr from Rust fn `conv_hashset_string_ret`

Generated by miniextendr from Rust fn `conv_list_mut_set_first`

Generated by miniextendr from Rust fn `conv_opt_string_is_some`

Generated by miniextendr from Rust fn `conv_opt_vec_string_arg`

Generated by miniextendr from Rust fn `conv_result_vec_i32_arg`

Generated by miniextendr from Rust fn `conv_result_vec_i32_err`

Generated by miniextendr from Rust fn `conv_vec_option_f32_ret`

Generated by miniextendr from Rust fn `conv_vec_option_i16_ret`

Generated by miniextendr from Rust fn `conv_vec_option_u16_ret`

Generated by miniextendr from Rust fn `conv_vec_option_u32_ret`

Generated by miniextendr from Rust fn `conv_vec_opt_string_len`

Generated by miniextendr from Rust fn `conv_vec_opt_string_ret`

Generated by miniextendr from Rust fn `conv_vec_vec_string_ret`

Generated by miniextendr from Rust fn `conv_as_named_list_array`

Generated by miniextendr from Rust fn `conv_as_named_list_empty`

Generated by miniextendr from Rust fn `conv_as_named_list_slice`

Generated by miniextendr from Rust fn `conv_as_named_vector_f64`

Generated by miniextendr from Rust fn `conv_as_named_vector_i32`

Generated by miniextendr from Rust fn `conv_btreemap_string_len`

Generated by miniextendr from Rust fn `conv_btreemap_string_ret`

Generated by miniextendr from Rust fn `conv_btreeset_string_len`

Generated by miniextendr from Rust fn `conv_btreeset_string_ret`

Generated by miniextendr from Rust fn `conv_named_list_contains`

Generated by miniextendr from Rust fn `conv_option_i64_some_big`

Generated by miniextendr from Rust fn `conv_opt_hashmap_i32_arg`

Generated by miniextendr from Rust fn `conv_opt_hashset_i32_arg`

Generated by miniextendr from Rust fn `conv_opt_ref_i32_is_some`

Generated by miniextendr from Rust fn `conv_ref_mut_i32_add_one`

Generated by miniextendr from Rust fn `conv_vec_hashmap_i32_arg`

Generated by miniextendr from Rust fn `conv_vec_hashmap_i32_ret`

Generated by miniextendr from Rust fn `conv_named_list_roundtrip`

Generated by miniextendr from Rust fn `conv_opt_btreemap_i32_arg`

Generated by miniextendr from Rust fn `conv_opt_ref_i32_none_ret`

Generated by miniextendr from Rust fn `conv_opt_ref_i32_some_ret`

Generated by miniextendr from Rust fn `conv_opt_vec_i32_none_ret`

Generated by miniextendr from Rust fn `conv_opt_vec_i32_some_ret`

Generated by miniextendr from Rust fn `conv_slice_mut_u8_add_one`

Generated by miniextendr from Rust fn `conv_vec_btreemap_i32_arg`

Generated by miniextendr from Rust fn `conv_vec_btreemap_i32_ret`

Generated by miniextendr from Rust fn `conv_as_named_vector_array`

Generated by miniextendr from Rust fn `conv_as_named_vector_empty`

Generated by miniextendr from Rust fn `conv_as_named_vector_slice`

Generated by miniextendr from Rust fn `conv_option_i64_some_small`

Generated by miniextendr from Rust fn `conv_slice_mut_i32_add_one`

Generated by miniextendr from Rust fn `conv_as_named_list_str_keys`

Generated by miniextendr from Rust fn `conv_as_named_vector_string`

Generated by miniextendr from Rust fn `conv_vec_option_i64_ret_big`

Generated by miniextendr from Rust fn `conv_vec_option_u64_ret_big`

Generated by miniextendr from Rust fn `conv_as_named_list_ext_trait`

Generated by miniextendr from Rust fn `conv_opt_vec_string_none_ret`

Generated by miniextendr from Rust fn `conv_opt_vec_string_some_ret`

Generated by miniextendr from Rust fn `conv_vec_slice_i32_total_len`

Generated by miniextendr from Rust fn `conv_opt_hashmap_i32_none_ret`

Generated by miniextendr from Rust fn `conv_opt_hashmap_i32_some_ret`

Generated by miniextendr from Rust fn `conv_opt_hashset_i32_none_ret`

Generated by miniextendr from Rust fn `conv_opt_hashset_i32_some_ret`

Generated by miniextendr from Rust fn `conv_vec_option_i64_ret_small`

Generated by miniextendr from Rust fn `conv_vec_option_i64_roundtrip`

Generated by miniextendr from Rust fn `conv_vec_option_u64_ret_small`

Generated by miniextendr from Rust fn `conv_as_named_vector_ext_trait`

Generated by miniextendr from Rust fn `conv_opt_btreemap_i32_none_ret`

Generated by miniextendr from Rust fn `conv_opt_btreemap_i32_some_ret`

Generated by miniextendr from Rust fn `conv_opt_mut_slice_i32_is_some`

Generated by miniextendr from Rust fn `conv_vec_mut_slice_i32_add_one`

Generated by miniextendr from Rust fn `conv_as_named_vector_option_i32`

Generated by miniextendr from Rust fn `conv_vec_option_isize_ret_small`

Generated by miniextendr from Rust fn `conv_vec_option_usize_ret_small`

Generated by miniextendr from Rust fn `conv_as_named_list_heterogeneous`

Generated by miniextendr from Rust fn `conv_as_named_list_duplicate_names`
}
\usage{
conv_i8_arg(x)

conv_i8_ret()

conv_u8_arg(x)

conv_u8_ret()

conv_f32_arg(x)

conv_f32_ret()

conv_f64_arg(x)

conv_f64_ret()

conv_i16_arg(x)

conv_i16_ret()

conv_i32_arg(x)

conv_i32_ret()

conv_i64_arg(x)

conv_i64_ret()

conv_str_ret()

conv_u16_arg(x)

conv_u16_ret()

conv_u32_arg(x)

conv_u32_ret()

conv_u64_arg(x)

conv_u64_ret()

conv_char_arg(x)

conv_char_ret()

conv_rlog_arg(x)

conv_rlog_ret()

conv_sexp_arg(x)

conv_sexp_ret()

conv_isize_arg(x)

conv_isize_ret()

conv_rbool_arg(x)

conv_rbool_ret()

conv_usize_arg(x)

conv_usize_ret()

conv_string_arg(x)

conv_string_ret()

conv_vec_i8_len(x)

conv_vec_i8_ret()

conv_vec_u8_len(x)

conv_vec_u8_ret()

conv_ref_i32_arg(x)

conv_vec_f32_len(x)

conv_vec_f32_ret()

conv_vec_f64_len(x)

conv_vec_f64_ret()

conv_vec_i16_len(x)

conv_vec_i16_ret()

conv_vec_i32_len(x)

conv_vec_i32_ret()

conv_vec_i64_len(x)

conv_vec_u16_len(x)

conv_vec_u16_ret()

conv_vec_u32_len(x)

conv_vec_u64_len(x)

conv_opt_f64_none()

conv_opt_f64_some()

conv_opt_i32_none()

conv_opt_i32_some()

conv_slice_u8_len(x)

conv_vec_bool_len(x)

conv_vec_bool_ret()

conv_vec_rlog_len(x)

conv_vec_rlog_ret()

conv_opt_bool_none()

conv_opt_bool_some()

conv_result_f64_ok()

conv_result_i32_ok()

conv_slice_f64_len(x)

conv_slice_i32_len(x)

conv_vec_isize_len(x)

conv_vec_usize_len(x)

conv_hashset_i8_ret()

conv_hashset_u8_len(x)

conv_hashset_u8_ret()

conv_named_list_get(config)

conv_named_list_len(config)

conv_opt_i8_is_some(x)

conv_opt_u8_is_some(x)

conv_result_f64_err()

conv_result_i32_arg(x)

conv_result_i32_err()

conv_slice_rlog_len(x)

conv_vec_opt_i8_len(x)

conv_vec_opt_u8_len(x)

conv_vec_string_len(x)

conv_vec_string_ret()

conv_btreeset_i8_ret()

conv_btreeset_u8_len(x)

conv_btreeset_u8_ret()

conv_hashmap_f64_len(x)

conv_hashmap_f64_ret()

conv_hashmap_i32_len(x)

conv_hashmap_i32_ret()

conv_hashset_i16_ret()

conv_hashset_i32_len(x)

conv_hashset_i32_ret()

conv_hashset_u16_ret()

conv_option_f32_some()

conv_option_i64_none()

conv_option_u32_some()

conv_opt_f32_is_some(x)

conv_opt_f64_is_some(x)

conv_opt_i16_is_some(x)

conv_opt_i32_is_some(x)

conv_opt_i64_is_some(x)

conv_opt_string_none()

conv_opt_string_some()

conv_opt_u16_is_some(x)

conv_opt_u32_is_some(x)

conv_opt_u64_is_some(x)

conv_opt_vec_i32_arg(x)

conv_vec_opt_f64_len(x)

conv_vec_opt_f64_ret()

conv_vec_opt_i32_len(x)

conv_vec_opt_i32_ret()

conv_vec_ref_i32_len(x)

conv_vec_vec_i32_arg(x)

conv_vec_vec_i32_ret()

conv_btreemap_f64_len(x)

conv_btreemap_f64_ret()

conv_btreemap_i32_len(x)

conv_btreemap_i32_ret()

conv_btreeset_i16_ret()

conv_btreeset_i32_len(x)

conv_btreeset_i32_ret()

conv_btreeset_u16_ret()

conv_hashmap_rlog_len(x)

conv_hashmap_rlog_ret()

conv_hashset_rlog_len(x)

conv_hashset_rlog_ret()

conv_opt_bool_is_some(x)

conv_opt_rlog_is_some(x)

conv_result_string_ok()

conv_vec_opt_bool_len(x)

conv_vec_opt_bool_ret()

conv_vec_opt_rlog_ret()

conv_as_named_list_vec()

conv_btreemap_rlog_len(x)

conv_btreemap_rlog_ret()

conv_opt_isize_is_some(x)

conv_opt_rbool_is_some(x)

conv_opt_usize_is_some(x)

conv_result_string_err()

conv_result_vec_i32_ok()

conv_vec_option_i8_ret()

conv_vec_opt_rbool_ret()

conv_hashmap_string_len(x)

conv_hashmap_string_ret()

conv_hashset_string_len(x)

conv_hashset_string_ret()

conv_list_mut_set_first(x)

conv_opt_string_is_some(x)

conv_opt_vec_string_arg(x)

conv_result_vec_i32_arg(x)

conv_result_vec_i32_err()

conv_vec_option_f32_ret()

conv_vec_option_i16_ret()

conv_vec_option_u16_ret()

conv_vec_option_u32_ret()

conv_vec_opt_string_len(x)

conv_vec_opt_string_ret()

conv_vec_vec_string_ret()

conv_as_named_list_array()

conv_as_named_list_empty()

conv_as_named_list_slice()

conv_as_named_vector_f64()

conv_as_named_vector_i32()

conv_btreemap_string_len(x)

conv_btreemap_string_ret()

conv_btreeset_string_len(x)

conv_btreeset_string_ret()

conv_named_list_contains(config)

conv_option_i64_some_big()

conv_opt_hashmap_i32_arg(x)

conv_opt_hashset_i32_arg(x)

conv_opt_ref_i32_is_some(x)

conv_ref_mut_i32_add_one(x)

conv_vec_hashmap_i32_arg(x)

conv_vec_hashmap_i32_ret()

conv_named_list_roundtrip(config)

conv_opt_btreemap_i32_arg(x)

conv_opt_ref_i32_none_ret()

conv_opt_ref_i32_some_ret()

conv_opt_vec_i32_none_ret()

conv_opt_vec_i32_some_ret()

conv_slice_mut_u8_add_one(x)

conv_vec_btreemap_i32_arg(x)

conv_vec_btreemap_i32_ret()

conv_as_named_vector_array()

conv_as_named_vector_empty()

conv_as_named_vector_slice()

conv_option_i64_some_small()

conv_slice_mut_i32_add_one(x)

conv_as_named_list_str_keys()

conv_as_named_vector_string()

conv_vec_option_i64_ret_big()

conv_vec_option_u64_ret_big()

conv_as_named_list_ext_trait()

conv_opt_vec_string_none_ret()

conv_opt_vec_string_some_ret()

conv_vec_slice_i32_total_len(x)

conv_opt_hashmap_i32_none_ret()

conv_opt_hashmap_i32_some_ret()

conv_opt_hashset_i32_none_ret()

conv_opt_hashset_i32_some_ret()

conv_vec_option_i64_ret_small()

conv_vec_option_i64_roundtrip(x)

conv_vec_option_u64_ret_small()

conv_as_named_vector_ext_trait()

conv_opt_btreemap_i32_none_ret()

conv_opt_btreemap_i32_some_ret()

conv_opt_mut_slice_i32_is_some(x)

conv_vec_mut_slice_i32_add_one(x)

conv_as_named_vector_option_i32()

conv_vec_option_isize_ret_small()

conv_vec_option_usize_ret_small()

conv_as_named_list_heterogeneous()

conv_as_named_list_duplicate_names()
}
\arguments{
\item{x}{Input list of integer vectors (mutated in-place).}

\item{config}{Input named list.}
}
\description{
Roundtrip test: R integer -> Rust `i8` -> R integer

conversions

Roundtrip test: R raw element -> Rust `u8` -> R raw element

conversions

Roundtrip test: R double -> Rust `f32` -> R double

conversions

Roundtrip test: R double -> Rust `f64` -> R double

conversions

Roundtrip test: R integer -> Rust `i16` -> R integer

conversions

Roundtrip test: R integer -> Rust `i32` -> R integer

conversions

Roundtrip test: R numeric -> Rust `i64` -> R numeric

conversions

conversions

Roundtrip test: R integer -> Rust `u16` -> R integer

conversions

Roundtrip test: R integer -> Rust `u32` -> R integer

conversions

Roundtrip test: R numeric -> Rust `u64` -> R numeric

conversions

Roundtrip test: R character scalar -> Rust `char` -> R character scalar

conversions

Roundtrip test: R logical -> Rust `RLogical` -> R logical

conversions

Roundtrip test: R SEXP -> Rust `SEXP` -> R SEXP (passthrough)

conversions

Roundtrip test: R numeric -> Rust `isize` -> R numeric

conversions

Roundtrip test: R logical -> Rust `Rboolean` -> R logical

conversions

Roundtrip test: R numeric -> Rust `usize` -> R numeric

conversions

Roundtrip test: R character scalar -> Rust `String` -> R character scalar

conversions

Test `Vec<i8>` argument: returns the length of the vector

conversions

Test `Vec<u8>` argument: returns the length of the vector

conversions

Test `&i32` reference argument: dereferences and returns the value

Test `Vec<f32>` argument: returns the length of the vector

conversions

Test `Vec<f64>` argument: returns the length of the vector

conversions

Test `Vec<i16>` argument: returns the length of the vector

conversions

Test `Vec<i32>` argument: returns the length of the vector

conversions

Test `Vec<i64>` argument: returns the length of the vector

Test `Vec<u16>` argument: returns the length of the vector

conversions

Test `Vec<u32>` argument: returns the length of the vector

Test `Vec<u64>` argument: returns the length of the vector

conversions

conversions

conversions

conversions

Test `&[u8]` slice argument: returns the length of the slice

Test `Vec<bool>` argument: returns the length of the vector

conversions

Test `Vec<RLogical>` argument: returns the length of the vector

conversions

conversions

conversions

conversions

conversions

Test `&[f64]` slice argument: returns the length of the slice

Test `&[i32]` slice argument: returns the length of the slice

Test `Vec<isize>` argument: returns the length of the vector

Test `Vec<usize>` argument: returns the length of the vector

conversions

Test `HashSet<u8>` argument: returns the number of unique elements

conversions

Test `NamedList` O(1) lookup: retrieves "width" and "height" and returns their product

Test `NamedList` length queries: returns total length and named-element count

Test `Option<i8>` argument: returns 1 if Some, 0 if None (NA)

Test `Option<u8>` argument: returns 1 if Some, 0 if None (NA)

conversions

Test `Result<i32, ()>` argument: returns 1 if Ok, 0 if Err (NULL)

conversions

Test `&[RLogical]` slice argument: returns the length of the slice

Test `Vec<Option<i8>>` argument: returns the length of the vector

Test `Vec<Option<u8>>` argument: returns the length of the vector

Test `Vec<String>` argument: returns the length of the vector

conversions

conversions

Test `BTreeSet<u8>` argument: returns the number of unique sorted elements

conversions

Test `HashMap<String, f64>` argument: returns the number of entries

conversions

Test `HashMap<String, i32>` argument: returns the number of entries

conversions

conversions

Test `HashSet<i32>` argument: returns the number of unique elements

conversions

conversions

conversions

conversions

conversions

Test `Option<f32>` argument: returns 1 if Some, 0 if None (NA)

Test `Option<f64>` argument: returns 1 if Some, 0 if None (NA)

Test `Option<i16>` argument: returns 1 if Some, 0 if None (NA)

Test `Option<i32>` argument: returns 1 if Some, 0 if None (NA)

Test `Option<i64>` argument: returns 1 if Some, 0 if None (NA)

conversions

conversions

Test `Option<u16>` argument: returns 1 if Some, 0 if None (NA)

Test `Option<u32>` argument: returns 1 if Some, 0 if None (NA)

Test `Option<u64>` argument: returns 1 if Some, 0 if None (NA)

Test `Option<Vec<i32>>` argument: returns sum if Some, -999 if None (NULL)

Test `Vec<Option<f64>>` argument: returns the length of the vector

conversions

Test `Vec<Option<i32>>` argument: returns the length of the vector

conversions

Test `Vec<&i32>` argument: returns the length of the vector

Test `Vec<Vec<i32>>` argument: returns the sum across all inner vectors

conversions

Test `BTreeMap<String, f64>` argument: returns the number of entries

conversions

Test `BTreeMap<String, i32>` argument: returns the number of entries

conversions

conversions

Test `BTreeSet<i32>` argument: returns the number of unique sorted elements

conversions

conversions

Test `HashMap<String, RLogical>` argument: returns the number of entries

conversions

Test `HashSet<RLogical>` argument: returns the number of unique elements

conversions

Test `Option<bool>` argument: returns 1 if Some, 0 if None (NA)

Test `Option<RLogical>` argument: returns 1 if Some, 0 if None (NA)

conversions

Test `Vec<Option<bool>>` argument: returns the length of the vector

conversions

conversions

conversions

Test `BTreeMap<String, RLogical>` argument: returns the number of entries

conversions

Test `Option<isize>` argument: returns 1 if Some, 0 if None (NA)

Test `Option<Rboolean>` argument: returns 1 if Some, 0 if None (NA)

Test `Option<usize>` argument: returns 1 if Some, 0 if None (NA)

conversions

conversions

conversions

conversions

Test `HashMap<String, String>` argument: returns the number of entries

conversions

Test `HashSet<String>` argument: returns the number of unique elements

conversions

Note: `mut` tests macro handling of mutable parameters; becomes unused after expansion

Test `Option<String>` argument: returns 1 if Some, 0 if None (NA)

Test `Option<Vec<String>>` argument: returns length if Some, -999 if None (NULL)

Test `Result<Vec<i32>, ()>` argument: returns 1 if Ok, 0 if Err (NULL)

conversions

conversions

conversions

conversions

conversions

Test `Vec<Option<String>>` argument: returns the length of the vector

conversions

conversions

conversions

conversions

conversions

conversions

conversions

Test `BTreeMap<String, String>` argument: returns the number of entries

conversions

Test `BTreeSet<String>` argument: returns the number of unique sorted elements

conversions

Test `NamedList::contains`: checks existence of "a", "b", and "missing"

conversions

Test `Option<HashMap<String, i32>>` argument: returns sum if Some, -999 if None

Test `Option<HashSet<i32>>` argument: returns sum if Some, -999 if None

Test `Option<&i32>` argument: returns 1 if Some, 0 if None

Test `&mut i32` mutable reference: increments by one and returns

Test `Vec<HashMap<String, i32>>` argument: returns the sum across all maps

conversions

Test `NamedList` roundtrip: converts to `List` and back to R

Test `Option<BTreeMap<String, i32>>` argument: returns sum if Some, -999 if None

conversions

conversions

conversions

conversions

Test `&mut [u8]` mutable slice: wrapping-adds one to each element

Test `Vec<BTreeMap<String, i32>>` argument: returns the sum across all maps

conversions

conversions

conversions

conversions

conversions

Test `&mut [i32]` mutable slice: increments each element and returns length

conversions

conversions

conversions

conversions

conversions

conversions

conversions

Test `Vec<&[i32]>` argument: returns the total length across all slices

conversions

conversions

conversions

conversions

conversions

Roundtrip test: R numeric vector with NA -> Rust `Vec<Option<i64>>` -> R

conversions

conversions

conversions

conversions

Test `Option<&mut [i32]>` argument: returns 1 if Some, 0 if None

Test `Vec<&mut [i32]>` argument: increments all elements and returns total length

conversions

conversions

conversions

conversions

conversions
}
