% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{CounterTraitEnv}
\alias{CounterTraitEnv}
\alias{CounterTraitEnv$new}
\alias{CounterTraitEnv$get_value}
\alias{$.CounterTraitEnv}
\alias{[[.CounterTraitEnv}
\alias{CounterTraitEnv$MatrixCounter$custom_get}
\alias{CounterTraitEnv$MatrixCounter$custom_add}
\alias{CounterTraitEnv$MatrixCounter$default_value}
\alias{CounterTraitEnv$StaticXParam$from_value}
\title{CounterTraitEnv  Class}
\usage{
CounterTraitEnv

\method{$}{CounterTraitEnv}(self, name)

\method{[[}{CounterTraitEnv}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
CounterTraitEnv  Class
}
