% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{CounterTraitEnv}
\alias{CounterTraitEnv}
\alias{CounterTraitEnv$new}
\alias{CounterTraitEnv$get_value}
\alias{$.CounterTraitEnv}
\alias{[[.CounterTraitEnv}
\alias{CounterTraitEnv$MatrixCounter$custom_get}
\alias{CounterTraitEnv$MatrixCounter$custom_add}
\alias{CounterTraitEnv$MatrixCounter$default_value}
\title{CounterTraitEnv Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type \code{CounterTraitEnv}
}
\usage{
CounterTraitEnv

\method{$}{CounterTraitEnv}(self, name)

\method{[[}{CounterTraitEnv}(self, name)
}
\description{
CounterTraitEnv Class
}
\keyword{datasets}
