% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{CounterTraitR6}
\alias{CounterTraitR6}
\alias{r6_trait_MatrixCounter_custom_get}
\alias{r6_trait_MatrixCounter_custom_add}
\alias{CounterTraitR6$MatrixCounter$default_value}
\title{CounterTraitR6 R6 Class}
\source{
Generated by miniextendr from Rust type \code{CounterTraitR6}

Generated by miniextendr from \verb{impl MatrixCounter for CounterTraitR6}
R6 trait method \code{MatrixCounter::custom_get} for CounterTraitR6

Generated by miniextendr from \verb{impl MatrixCounter for CounterTraitR6}
R6 trait method \code{MatrixCounter::custom_add} for CounterTraitR6
}
\usage{
r6_trait_MatrixCounter_custom_get(x)

r6_trait_MatrixCounter_custom_add(x, n)
}
\arguments{
\item{x}{A \code{CounterTraitR6} object}

\item{n}{Amount to add.}
}
\description{
CounterTraitR6 R6 Class

CounterTraitR6 R6 Class

Static trait method MatrixCounter::default_value()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CounterTraitR6-new}{\code{CounterTraitR6$new()}}
\item \href{#method-CounterTraitR6-get_value}{\code{CounterTraitR6$get_value()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CounterTraitR6-new"></a>}}
\if{latex}{\out{\hypertarget{method-CounterTraitR6-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CounterTraitR6$new(v, .ptr = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.ptr}}{Internal pointer (used by static methods, not for direct use).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CounterTraitR6-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-CounterTraitR6-get_value}{}}}
\subsection{Method \code{get_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CounterTraitR6$get_value()}\if{html}{\out{</div>}}
}

}
}
