% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{CounterTraitR6}
\alias{CounterTraitR6}
\alias{CounterTraitR6$new}
\alias{CounterTraitR6$get_value}
\alias{$.CounterTraitR6}
\alias{[[.CounterTraitR6}
\alias{r6_trait_MatrixCounter_custom_get}
\alias{r6_trait_MatrixCounter_custom_add}
\alias{CounterTraitR6$MatrixCounter$default_value}
\title{CounterTraitR6  Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type \code{CounterTraitR6}

Generated by miniextendr from \code{CounterTraitR6::new}

Generated by miniextendr from \code{CounterTraitR6::get_value}

Generated by miniextendr from \verb{impl MatrixCounter for CounterTraitR6}
R6 trait method \code{MatrixCounter::custom_get} for CounterTraitR6

Generated by miniextendr from \verb{impl MatrixCounter for CounterTraitR6}
R6 trait method \code{MatrixCounter::custom_add} for CounterTraitR6
}
\usage{
CounterTraitR6

\method{$}{CounterTraitR6}(self, name)

\method{[[}{CounterTraitR6}(self, name)

r6_trait_MatrixCounter_custom_get(x)

r6_trait_MatrixCounter_custom_add(x, n)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}

\item{x}{A \code{CounterTraitR6} object}

\item{n}{Amount to add.}
}
\description{
CounterTraitR6  Class

Static trait method MatrixCounter::default_value()
}
\keyword{datasets}
