% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{CounterTraitS3}
\alias{CounterTraitS3}
\alias{CounterTraitS3$new}
\alias{CounterTraitS3$get_value}
\alias{$.CounterTraitS3}
\alias{[[.CounterTraitS3}
\alias{custom_get.CounterTraitS3}
\alias{custom_add.CounterTraitS3}
\alias{CounterTraitS3$MatrixCounter$default_value}
\title{CounterTraitS3 Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type \code{CounterTraitS3}
}
\usage{
CounterTraitS3

\method{$}{CounterTraitS3}(self, name)

\method{[[}{CounterTraitS3}(self, name)

\method{custom_get}{CounterTraitS3}(x, ...)

\method{custom_add}{CounterTraitS3}(x, n, ...)
}
\description{
CounterTraitS3 Class

Static trait method MatrixCounter::default_value()
}
\keyword{datasets}
