% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{CounterTraitS3}
\alias{CounterTraitS3}
\alias{CounterTraitS3$new}
\alias{CounterTraitS3$get_value}
\alias{$.CounterTraitS3}
\alias{[[.CounterTraitS3}
\alias{custom_get}
\alias{custom_get.CounterTraitS3}
\alias{custom_add}
\alias{custom_add.CounterTraitS3}
\alias{CounterTraitS3$MatrixCounter$default_value}
\title{CounterTraitS3  Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type \code{CounterTraitS3}

Generated by miniextendr from \code{CounterTraitS3::new}

Generated by miniextendr from \code{CounterTraitS3::get_value}

Generated by miniextendr from \verb{impl MatrixCounter for CounterTraitS3}
S3 generic for \code{custom_get}

Generated by miniextendr from \verb{impl MatrixCounter for CounterTraitS3}
S3 generic for \code{custom_add}
}
\usage{
CounterTraitS3

\method{$}{CounterTraitS3}(self, name)

\method{[[}{CounterTraitS3}(self, name)

\method{custom_get}{CounterTraitS3}(x, ...)

\method{custom_add}{CounterTraitS3}(x, n, ...)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}

\item{x}{An object}

\item{...}{Additional arguments passed to methods}

\item{n}{Amount to add.}
}
\description{
CounterTraitS3  Class

S3 generic for \code{custom_get}

S3 generic for \code{custom_add}

Static trait method MatrixCounter::default_value()
}
\keyword{datasets}
