% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{CounterTraitS4}
\alias{CounterTraitS4}
\alias{CounterTraitS4$new}
\alias{CounterTraitS4$get_value}
\alias{$.CounterTraitS4}
\alias{[[.CounterTraitS4}
\alias{s4_trait_MatrixCounter_custom_get}
\alias{s4_trait_MatrixCounter_custom_add}
\alias{CounterTraitS4_MatrixCounter_default_value}
\title{CounterTraitS4 Class}
\format{
An object of class \code{environment} of length 2.
}
\source{
Generated by miniextendr from Rust type \code{CounterTraitS4}

Generated by miniextendr from \verb{impl MatrixCounter for CounterTraitS4}
S4 generic for trait method \code{MatrixCounter::custom_get}

Generated by miniextendr from \verb{impl MatrixCounter for CounterTraitS4}
S4 generic for trait method \code{MatrixCounter::custom_add}
}
\usage{
CounterTraitS4

\method{$}{CounterTraitS4}(self, name)

\method{[[}{CounterTraitS4}(self, name)

CounterTraitS4_MatrixCounter_default_value()
}
\description{
CounterTraitS4 Class

Static trait method MatrixCounter::default_value() for CounterTraitS4
}
\keyword{datasets}
