% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{CounterTraitS7}
\alias{CounterTraitS7}
\alias{CounterTraitS7$new}
\alias{CounterTraitS7$get_value}
\alias{$.CounterTraitS7}
\alias{[[.CounterTraitS7}
\alias{s7_trait_MatrixCounter_custom_get}
\alias{s7_trait_MatrixCounter_custom_add}
\alias{CounterTraitS7$MatrixCounter$default_value}
\title{CounterTraitS7  Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type \code{CounterTraitS7}

Generated by miniextendr from \code{CounterTraitS7::new}

Generated by miniextendr from \code{CounterTraitS7::get_value}

Generated by miniextendr from \verb{impl MatrixCounter for CounterTraitS7}
S7 generic for trait method \code{MatrixCounter::custom_get}

Generated by miniextendr from \verb{impl MatrixCounter for CounterTraitS7}
S7 generic for trait method \code{MatrixCounter::custom_add}
}
\usage{
CounterTraitS7

\method{$}{CounterTraitS7}(self, name)

\method{[[}{CounterTraitS7}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
CounterTraitS7  Class

Static trait method MatrixCounter::default_value()
}
\keyword{datasets}
