% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{create_workspace_dirs}
\alias{create_workspace_dirs}
\title{Create directory structure for workspace}
\usage{
create_workspace_dirs(
  workspace_id,
  token = twist_token(),
  workspace_dir = "twist-workspace"
)
}
\arguments{
\item{workspace_id}{ID of the workspace}

\item{token}{Authentication token}

\item{workspace_dir}{Directory to create the workspace structure in}
}
\value{
Vector of created directory paths
}
\description{
Creates a directory for each channel in the workspace and returns the
created paths.
}
