% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{custom_get}
\alias{custom_get}
\title{S3 generic for \code{custom_get}}
\source{
Generated by miniextendr from \verb{impl MatrixCounter for Counter_TraitS3}
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments passed to methods}
}
\description{
S3 generic for \code{custom_get}
}
