% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D4_Traverse.R
\name{D4TraverseB}
\alias{D4TraverseB}
\title{Priority flow processing of D4 Stream Networks}
\usage{
D4TraverseB(dem, queue, marked, mask, step, direction, basins, d4 = c(1,
  2, 3, 4), printstep = F, nchunk = 100, epsilon = 0)
}
\arguments{
\item{dem}{Elevation matrix}

\item{queue}{a priority queue to start from three columns, x, y, elevation}

\item{d4}{directional numbering system: the numbers you want to assigne to down, left, top,right (defaults to 1,2,3,4)}

\item{marked:}{a matrix of which cells have been marked already}

\item{printstep:}{if true it will print out the step number and the size of the queue}

\item{epsilon:}{amount to add to filled areas to avoid creating flats, defaults to zero}

\item{mask:}{Mask with ones for cells to be processed and zeros for everything else - defaults to a mask of all 1's}

\item{step:}{a matrix of the step number for cells that have been processed - defaults to all zeros}

\item{direction:}{a matrix of the flow directions for cells that have been processed - defaults to all zeros}

\item{basins:}{a matrix of basin numbers that can be created by the initilizaiton script. If you input this every cell will be assigned the same basin as the cell that adds it}
}
\description{
'D4TraverB'  will process stream all network cells walking upstream on d4 neigbors
in a river mask. Where no D4 neigbhors exist it looks for d8 neigbors
and created d4 bridges to these diagonal cells
}
