% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{standalone_dataframe_roundtrip}
\alias{standalone_dataframe_roundtrip}
\title{dataframe derive alignment tests}
\source{
Generated by miniextendr from Rust fn `standalone_dataframe_roundtrip`
}
\usage{
standalone_dataframe_roundtrip()
}
\description{
dataframe derive alignment tests
}
