% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R, R/Connection.R, R/Result.R
\docType{class}
\name{DBI}
\alias{DBI}
\alias{graindb_driver-class}
\alias{show,graindb_driver-method}
\alias{dbConnect,graindb_driver-method}
\alias{dbDataType,graindb_driver-method}
\alias{dbIsValid,graindb_driver-method}
\alias{dbGetInfo,graindb_driver-method}
\alias{graindb_connection-class}
\alias{show,graindb_connection-method}
\alias{dbIsValid,graindb_connection-method}
\alias{dbDisconnect,graindb_connection-method}
\alias{dbSendQuery,graindb_connection,character-method}
\alias{dbSendStatement,graindb_connection,character-method}
\alias{dbDataType,graindb_connection-method}
\alias{dbWriteTable,graindb_connection,character,data.frame-method}
\alias{dbListTables,graindb_connection-method}
\alias{dbExistsTable,graindb_connection,character-method}
\alias{dbListFields,graindb_connection,character-method}
\alias{dbRemoveTable,graindb_connection,character-method}
\alias{dbGetInfo,graindb_connection-method}
\alias{dbBegin,graindb_connection-method}
\alias{dbCommit,graindb_connection-method}
\alias{dbRollback,graindb_connection-method}
\alias{graindb_result-class}
\alias{show,graindb_result-method}
\alias{dbClearResult,graindb_result-method}
\alias{dbFetch,graindb_result-method}
\alias{dbHasCompleted,graindb_result-method}
\alias{dbGetInfo,graindb_result-method}
\alias{dbIsValid,graindb_result-method}
\alias{dbGetStatement,graindb_result-method}
\alias{dbColumnInfo,graindb_result-method}
\alias{dbGetRowCount,graindb_result-method}
\alias{dbGetRowsAffected,graindb_result-method}
\alias{dbBind,graindb_result-method}
\title{DBI methods}
\usage{
\S4method{show}{graindb_driver}(object)

\S4method{dbConnect}{graindb_driver}(
  drv,
  dbdir = ":memory:",
  ...,
  debug = getOption("graindb.debug", FALSE),
  dbname = NA
)

\S4method{dbDataType}{graindb_driver}(dbObj, obj, ...)

\S4method{dbIsValid}{graindb_driver}(dbObj, ...)

\S4method{dbGetInfo}{graindb_driver}(dbObj, ...)

\S4method{show}{graindb_connection}(object)

\S4method{dbIsValid}{graindb_connection}(dbObj, ...)

\S4method{dbDisconnect}{graindb_connection}(conn, ...)

\S4method{dbSendQuery}{graindb_connection,character}(conn, statement, ...)

\S4method{dbSendStatement}{graindb_connection,character}(conn, statement, ...)

\S4method{dbDataType}{graindb_connection}(dbObj, obj, ...)

\S4method{dbWriteTable}{graindb_connection,character,data.frame}(
  conn,
  name,
  value,
  row.names = FALSE,
  overwrite = FALSE,
  append = FALSE,
  field.types = NULL,
  temporary = FALSE,
  ...
)

\S4method{dbListTables}{graindb_connection}(conn, ...)

\S4method{dbExistsTable}{graindb_connection,character}(conn, name, ...)

\S4method{dbListFields}{graindb_connection,character}(conn, name, ...)

\S4method{dbRemoveTable}{graindb_connection,character}(conn, name, ...)

\S4method{dbGetInfo}{graindb_connection}(dbObj, ...)

\S4method{dbBegin}{graindb_connection}(conn, ...)

\S4method{dbCommit}{graindb_connection}(conn, ...)

\S4method{dbRollback}{graindb_connection}(conn, ...)

\S4method{show}{graindb_result}(object)

\S4method{dbClearResult}{graindb_result}(res, ...)

\S4method{dbFetch}{graindb_result}(res, n = -1, ...)

\S4method{dbHasCompleted}{graindb_result}(res, ...)

\S4method{dbGetInfo}{graindb_result}(dbObj, ...)

\S4method{dbIsValid}{graindb_result}(dbObj, ...)

\S4method{dbGetStatement}{graindb_result}(res, ...)

\S4method{dbColumnInfo}{graindb_result}(res, ...)

\S4method{dbGetRowCount}{graindb_result}(res, ...)

\S4method{dbGetRowsAffected}{graindb_result}(res, ...)

\S4method{dbBind}{graindb_result}(res, params, ...)
}
\arguments{
\item{object}{Any R object}

\item{drv}{an object that inherits from \linkS4class{DBIDriver},
or an existing \linkS4class{DBIConnection}
object (in order to clone an existing connection).}

\item{...}{authentication arguments needed by the DBMS instance; these
typically include \code{user}, \code{password}, \code{host}, \code{port}, \code{dbname}, etc.
For details see the appropriate \code{DBIDriver}.}

\item{dbObj}{A object inheriting from \linkS4class{DBIDriver}
or \linkS4class{DBIConnection}}

\item{obj}{An R object whose SQL type we want to determine.}

\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[=dbConnect]{dbConnect()}}.}

\item{statement}{a character string containing SQL.}

\item{name}{A character string specifying the unquoted DBMS table name,
or the result of a call to \code{\link[=dbQuoteIdentifier]{dbQuoteIdentifier()}}.}

\item{value}{a \link[DBI]{data.frame} (or coercible to data.frame).}

\item{overwrite}{Allow overwriting the destination table. Cannot be
`TRUE` if `append` is also `TRUE`.}

\item{append}{Allow appending to the destination table. Cannot be
`TRUE` if `overwrite` is also `TRUE`.}

\item{res}{An object inheriting from \linkS4class{DBIResult}.}

\item{n}{maximum number of records to retrieve per fetch. Use \code{n = -1}
or \code{n = Inf}
to retrieve all pending records.  Some implementations may recognize other
special values.}

\item{params}{A list of bindings, named or unnamed.}
}
\description{
Implementations of pure virtual functions defined in the `DBI` package.
}
