% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{decimal_add}
\alias{decimal_add}
\alias{decimal_mul}
\alias{decimal_round}
\alias{decimal_scale}
\alias{decimal_is_zero}
\alias{decimal_roundtrip}
\title{Test Decimal addition of two string-encoded values}
\source{
Generated by miniextendr from Rust fn `decimal_add`

Generated by miniextendr from Rust fn `decimal_mul`

Generated by miniextendr from Rust fn `decimal_round`

Generated by miniextendr from Rust fn `decimal_scale`

Generated by miniextendr from Rust fn `decimal_is_zero`

Generated by miniextendr from Rust fn `decimal_roundtrip`
}
\usage{
decimal_add(a, b)

decimal_mul(a, b)

decimal_round(s, dp)

decimal_scale(s)

decimal_is_zero(s)

decimal_roundtrip(s)
}
\arguments{
\item{a}{First decimal as string.}

\item{b}{Second decimal as string.}

\item{s}{Decimal value as string.}

\item{dp}{Number of decimal places.}
}
\description{
Test Decimal addition of two string-encoded values

Test Decimal multiplication of two string-encoded values

Test rounding a Decimal to a given number of decimal places

Test extracting the scale (number of decimal digits) of a Decimal

Test whether a Decimal value parsed from string is zero

Test Decimal string roundtrip via parsing and formatting
}
