% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{DeepNest}
\alias{DeepNest}
\alias{DeepNest$new}
\alias{DeepNest$to_r}
\alias{DeepNest$from_r}
\alias{$.DeepNest}
\alias{[[.DeepNest}
\title{DeepNest Class}
\usage{
DeepNest

\method{$}{DeepNest}(self, name)

\method{[[}{DeepNest}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
DeepNest Class
}
