% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{DeepNest}
\alias{DeepNest}
\alias{DeepNest$new}
\alias{DeepNest$to_r}
\alias{DeepNest$from_r}
\alias{$.DeepNest}
\alias{[[.DeepNest}
\title{DeepNest  Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type `DeepNest`

Generated by miniextendr from `DeepNest::new`

Generated by miniextendr from `DeepNest::to_r`

Generated by miniextendr from `DeepNest::from_r`
}
\usage{
DeepNest

\method{$}{DeepNest}(self, name)

\method{[[}{DeepNest}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
DeepNest  Class
}
\keyword{datasets}
