% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{read_csv_duckdb}
\alias{read_csv_duckdb}
\title{Deprecated functions}
\usage{
read_csv_duckdb(conn, files, tablename, ...)
}
\description{
\code{read_csv_duckdb()} has been superseded by \code{duckdb_read_csv()}.
The order of the arguments has changed.
}
\keyword{internal}
