% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{detect_bubbles_i_j_igraph}
\alias{detect_bubbles_i_j_igraph}
\title{bubble detection between v_start and v_end}
\usage{
detect_bubbles_i_j_igraph(v_start, v_end, g, sgnodetypes)
}
\description{
bubble detection between v_start and v_end
}
