% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrsa.r
\docType{data}
\name{dff4}
\alias{dff4}
\title{Daily Fama French 4 Factor Returns}
\format{An \code{xts} object with 24,795 observations and 5 columns. 
The data run from January, 1927 through December, 2020.
As in the upstream source, the data are given in \emph{percents}, meaning a
value of 1.00 corresponds to a 1\% movement.
Note also that returns presumably are \sQuote{simple} returns, not log
returns, though this is not clarified by the upstream source.
The columns are defined as follows:
\describe{
 \item{\code{Mkt}}{The Market daily return. Note that the risk free rate has been added back
 to the excess returns published by the upstream source.}
 \item{\code{SMB}}{The cap factor daily return.}
 \item{\code{HML}}{The growth factor daily return.}
 \item{\code{UMD}}{The momentum factor daily return.}
 \item{\code{RF}}{The risk-free rate, presumably as an daily rate, though note that no corrections have been
 made for weekend effects when adding the risk-free rate back to the market
 rate.}
}}
\source{
Kenneth French data library.
See \url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html},
data description at \url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/Data_Library/f-f_factors.html}.
}
\usage{
dff4
}
\description{
The daily returns of the 4 Fama French Factors: Market, the cap factor SMB, 
the growth factor HML, and the momentum factor UMD.
}
\examples{
\dontrun{
data(dff4)
str(dff4)
}
}
\author{
Steven E. Pav \email{steven@sharperat.io}
}
\keyword{datasets}
