% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartition_analysis.R
\name{diff_exons_between_tx}
\alias{diff_exons_between_tx}
\title{Compute diff exons from graph and splicing structure for specific transcripts}
\usage{
diff_exons_between_tx(gene, g, sg, tx_ids, outdir)
}
\description{
Compute diff exons from graph and splicing structure for specific transcripts
}
