% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartition_analysis.R
\name{diff_exons_gene_ovr}
\alias{diff_exons_gene_ovr}
\title{Compute diff exons with one-vs-rest approach for each partition}
\usage{
diff_exons_gene_ovr(gene, g, outdir)
}
\description{
Compute diff exons with one-vs-rest approach for each partition
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
result <- grase::diff_exons_gene_ovr(
  gene = "ENSG00000000003.14", g = g, outdir = tempdir()
)
}
}
