% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focalexons.R
\name{diff_exons_gene_ovr}
\alias{diff_exons_gene_ovr}
\title{Compute diff exons from graph and splicing structure}
\usage{
diff_exons_gene_ovr(gene, g, sg, outdir)
}
\description{
Compute diff exons from graph and splicing structure
}
