% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{dim}
\alias{dim}
\alias{dim,fm-method}
\title{Dimensions of an FlashR Object}
\usage{
\S4method{dim}{fm}(x)
}
\arguments{
\item{x}{A FlashR matrix.}
}
\description{
Retrieve the dimension of an FlashR object.
}

