% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redivis-connection.R
\name{.fetch_biblio_table}
\alias{.fetch_biblio_table}
\title{Fetch Bibliography Metadata Table}
\usage{
.fetch_biblio_table()
}
\value{
A tibble containing filtered biblio information, with only the tables that exist in the IRW database.
}
\description{
Retrieves the metadata table from Redivis user("bdomingu")$dataset("irw_meta")$table("biblio").
Only fetches new data if the table version tag has changed.
This version filters out any tables that do not exist in the IRW database.
}
\keyword{internal}
