% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{.fetch_biblio_table}
\alias{.fetch_biblio_table}
\title{Fetch Bibliography Metadata Table}
\usage{
.fetch_biblio_table()
}
\value{
A cached or newly fetched tibble containing biblio information.
}
\description{
Retrieves the metadata table from Redivis user("bdomingu")$dataset("irw_meta")$table("biblio").
Only fetches new data if the table version tag has changed.
}
\keyword{internal}
