% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redivis-connection.R, R/redivis-metadata.R
\name{.fetch_tags_table}
\alias{.fetch_tags_table}
\title{Fetch Tags Metadata Table}
\usage{
.fetch_tags_table()

.fetch_tags_table()
}
\value{
A tibble containing filtered tags information.

A tibble containing filtered tags information.
}
\description{
Retrieves the tags metadata table from Redivis user("bdomingu")$dataset("irw_meta")$table("tags").
Only fetches new data if the table version tag has changed.
Filters out any tags referring to tables that do not exist in the IRW database.

Retrieves the tags metadata table from Redivis user("bdomingu")$dataset("irw_meta")$table("tags").
Only fetches new data if the table version tag has changed.
Filters out any tags referring to tables that do not exist in the IRW database.
}
\keyword{internal}
