% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redivis-datasets.R
\name{.initialize_datasource}
\alias{.initialize_datasource}
\title{Initialize Redivis Datasource(s)}
\usage{
.initialize_datasource(source = "core", sim = FALSE, comp = FALSE, nom = FALSE)
}
\arguments{
\item{source}{Character. One of \code{"core"}, \code{"nom"}, \code{"sim"}, \code{"comp"}.
Default is \code{"core"}.}

\item{sim}{Deprecated. Use \code{source = "sim"} instead.}

\item{comp}{Deprecated. Use \code{source = "comp"} instead.}

\item{nom}{Deprecated. Use \code{source = "nom"} instead.}
}
\value{
A list of one or more Redivis dataset objects.
}
\description{
Returns a list of Redivis dataset objects based on the selected source:
\itemize{
\item If \code{source = "sim"}, returns the IRW simulation dataset (\code{irw_simsyn:0btg})
\item If \code{source = "comp"}, returns the IRW competition dataset (\code{irw_competitions:cmd7})
\item If \code{source = "nom"}, returns the IRW nominal dataset (\code{irw_nominal:614n})
\item If \code{source = "core"} (default), returns the main IRW production datasets
}
}
\keyword{internal}
