% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{.initialize_datasource}
\alias{.initialize_datasource}
\title{Initialize Datasource (With Retry Logic)}
\usage{
.initialize_datasource()
}
\value{
The initialized datasource connection.
}
\description{
Ensures a stable connection to Redivis. If it fails, retries automatically.
}
\keyword{internal}
