% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redivis-connection.R
\name{.initialize_datasource}
\alias{.initialize_datasource}
\title{Initialize Datasource}
\usage{
.initialize_datasource()
}
\value{
A Redivis dataset object representing the connected datasource.
}
\description{
Establishes a connection to the Redivis datasource. Ensures \code{get()} is called
to properly retrieve the dataset.
}
\keyword{internal}
