% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{download_data}
\alias{download_data}
\title{Download a Dataset}
\usage{
download_data(table, path = NULL, overwrite = FALSE)
}
\arguments{
\item{table}{A table object from which data should be downloaded.}

\item{path}{A string specifying the file path where the data should be saved.
If NULL, the default file name will be used in the current working directory.}

\item{overwrite}{A logical indicating whether to overwrite an existing file at the specified path.
Default is FALSE.}
}
\value{
A message confirming the file download location.
}
\description{
This function downloads a specified dataset from the IRW datasource.
}
