% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{download_data}
\alias{download_data}
\title{Download a Dataset}
\usage{
download_data(table_name, path = NULL, overwrite = FALSE)
}
\arguments{
\item{table_name}{A character string specifying the name of the Redivis table
from which data should be downloaded. This should be the exact name
of the table as listed in the IRW Redivis database.}

\item{path}{A string specifying the file path where the data should be saved.
If \code{NULL}, the dataset will be saved in the current working directory with the table's name as the file name.}

\item{overwrite}{A logical indicating whether to overwrite an existing file at the specified path.
Default is \code{FALSE}.}
}
\value{
A message confirming the file download location.
}
\description{
This function downloads a specified dataset from the IRW datasource.
}
