% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\docType{package}
\name{duckdb-package}
\alias{duckdb-package}
\title{DuckDB client package for R}
\description{
R client package for DuckDB: an embeddable SQL OLAP Database Management System.
}
\seealso{
\code{\link[=duckdb]{duckdb()}} for connection instructions.

\url{https://duckdb.org/} for the project website.
}
