% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R
\name{duckdb}
\alias{duckdb}
\title{DuckDB main workhorse driver}
\usage{
duckdb(dbdir = DBDIR_MEMORY, read_only = FALSE)
}
\arguments{
\item{dbdir}{(`character`)\cr
FIXME}

\item{read_only}{(`logical`)\cr
FIXME}
}
\description{
TBD.
}
\examples{
\dontrun{
#' library(DBI)
duckdb::duckdb()
}

}
