% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R
\name{dbConnect,guinsoodb_driver-method}
\alias{dbConnect,guinsoodb_driver-method}
\alias{dbDisconnect,guinsoodb_connection-method}
\alias{guinsoodb}
\alias{guinsoodb_shutdown}
\title{Connect to a GuinsooDB database instance}
\usage{
\S4method{dbConnect}{guinsoodb_driver}(
  drv,
  dbdir = DBDIR_MEMORY,
  ...,
  debug = getOption("guinsoodb.debug", FALSE),
  read_only = FALSE,
  timezone_out = "UTC",
  tz_out_convert = c("with", "force")
)

\S4method{dbDisconnect}{guinsoodb_connection}(conn, ..., shutdown = FALSE)

guinsoodb(dbdir = DBDIR_MEMORY, read_only = FALSE)

guinsoodb_shutdown(drv)
}
\arguments{
\item{drv}{Object returned by \code{guinsoodb()}}

\item{dbdir}{Location for database files. Should be a path to an existing
directory in the file system. With the default, all
data is kept in RAM}

\item{...}{Ignored}

\item{debug}{Print additional debug information such as queries}

\item{read_only}{Set to \code{TRUE} for read-only operation}

\item{timezone_out}{The time zone returned to R, defaults to \code{"UTC"}, which
is currently the only timezone supported by guinsoodb.
If you want to display datetime values in the local timezone,
set to \code{\link[=Sys.timezone]{Sys.timezone()}} or \code{""}.}

\item{tz_out_convert}{How to convert timestamp columns to the timezone specified
in \code{timezone_out}. There are two options: \code{"with"}, and \code{"force"}. If \code{"with"}
is chosen, the timestamp will be returned as it would appear in the specified time zone.
If \code{"force"} is chosen, the timestamp will have the same clock
time as the timestamp in the database, but with the new time zone.}

\item{conn}{A \code{guinsoodb_connection} object}

\item{shutdown}{Set to \code{TRUE} to shut down the GuinsooDB database instance that this connection refers to.}
}
\value{
\code{dbConnect()} returns an object of class
\linkS4class{guinsoodb_connection}.

\code{guinsoodb()} returns an object of class \linkS4class{guinsoodb_driver}.

\code{dbDisconnect()} and \code{guinsoodb_shutdown()} are called for their
side effect.
}
\description{
\code{dbConnect()} connects to a database instance.

\code{dbDisconnect()} closes a GuinsooDB database connection, optionally shutting down
the associated instance.

\code{guinsoodb()} creates or reuses a database instance.

\code{guinsoodb_shutdown()} shuts down a database instance.
}
\examples{
drv <- guinsoodb()
con <- dbConnect(drv)

dbGetQuery(con, "SELECT 'Hello, world!'")

dbDisconnect(con)
guinsoodb_shutdown(drv)

# Shorter:
con <- dbConnect(guinsoodb())
dbGetQuery(con, "SELECT 'Hello, world!'")
dbDisconnect(con, shutdown = TRUE)
}
