% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R, R/Driver.R
\docType{class}
\name{duckdb_driver-class}
\alias{duckdb_driver-class}
\alias{duckdb_driver}
\alias{show,duckdb_driver-method}
\alias{dbIsValid,duckdb_driver-method}
\alias{dbGetInfo,duckdb_driver-method}
\title{DuckDB driver class}
\usage{
\S4method{show}{duckdb_driver}(object)

\S4method{dbIsValid}{duckdb_driver}(dbObj, ...)

\S4method{dbGetInfo}{duckdb_driver}(dbObj, ...)
}
\arguments{
\item{object}{Any R object}

\item{dbObj}{An object inheriting from \linkS4class{DBIObject},
i.e. \linkS4class{DBIDriver}, \linkS4class{DBIConnection},
or a \linkS4class{DBIResult}}

\item{...}{Other arguments to methods.}
}
\description{
Implements \linkS4class{DBIDriver}.
}
\keyword{internal}
