% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{EitherRoutesEnv}
\alias{EitherRoutesEnv}
\alias{EitherRoutesEnv$new}
\alias{EitherRoutesEnv$plan}
\alias{$.EitherRoutesEnv}
\alias{[[.EitherRoutesEnv}
\title{EitherRoutesEnv Class}
\usage{
EitherRoutesEnv

\method{$}{EitherRoutesEnv}(self, name)

\method{[[}{EitherRoutesEnv}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Route names or a data frame of doses, and omittable tier names or a number.
}
