% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{assert_utf8_locale_now}
\alias{assert_utf8_locale_now}
\alias{encoding_info_available}
\title{encoding tests}
\source{
Generated by miniextendr from Rust fn `assert_utf8_locale_now`

Generated by miniextendr from Rust fn `encoding_info_available`
}
\usage{
assert_utf8_locale_now()

encoding_info_available()
}
\description{
encoding tests

encoding tests
}
