% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{encoding_info_available}
\alias{encoding_info_available}
\title{encoding tests}
\source{
Generated by miniextendr from Rust fn \code{encoding_info_available}
}
\usage{
encoding_info_available()
}
\description{
encoding tests
}
