% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{EnvMatchArgCounter}
\alias{EnvMatchArgCounter}
\alias{EnvMatchArgCounter$new}
\alias{EnvMatchArgCounter$count}
\alias{EnvMatchArgCounter$reset}
\alias{$.EnvMatchArgCounter}
\alias{[[.EnvMatchArgCounter}
\title{EnvMatchArgCounter  Class}
\usage{
EnvMatchArgCounter

\method{$}{EnvMatchArgCounter}(self, name)

\method{[[}{EnvMatchArgCounter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
EnvMatchArgCounter  Class
}
