% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{EnvMatchArgCounter}
\alias{EnvMatchArgCounter}
\alias{EnvMatchArgCounter$new}
\alias{EnvMatchArgCounter$count}
\alias{EnvMatchArgCounter$reset}
\alias{$.EnvMatchArgCounter}
\alias{[[.EnvMatchArgCounter}
\title{EnvMatchArgCounter  Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type `EnvMatchArgCounter`

Generated by miniextendr from `EnvMatchArgCounter::new`

Generated by miniextendr from `EnvMatchArgCounter::count`

Generated by miniextendr from `EnvMatchArgCounter::reset`
}
\usage{
EnvMatchArgCounter

\method{$}{EnvMatchArgCounter}(self, name)

\method{[[}{EnvMatchArgCounter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
EnvMatchArgCounter  Class
}
\keyword{datasets}
