% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{EnvRaiser}
\alias{EnvRaiser}
\alias{EnvRaiser$new}
\alias{EnvRaiser$env_id}
\alias{EnvRaiser$env_raise_error}
\alias{EnvRaiser$env_raise_error_classed}
\alias{EnvRaiser$env_raise_warning}
\alias{EnvRaiser$env_raise_warning_classed}
\alias{EnvRaiser$env_raise_message}
\alias{EnvRaiser$env_raise_condition}
\alias{EnvRaiser$env_raise_condition_classed}
\alias{$.EnvRaiser}
\alias{[[.EnvRaiser}
\title{EnvRaiser  Class}
\usage{
EnvRaiser

\method{$}{EnvRaiser}(self, name)

\method{[[}{EnvRaiser}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
EnvRaiser  Class
}
