% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{EnvRaiser}
\alias{EnvRaiser}
\alias{EnvRaiser$new}
\alias{EnvRaiser$env_id}
\alias{EnvRaiser$env_raise_error}
\alias{EnvRaiser$env_raise_error_classed}
\alias{EnvRaiser$env_raise_warning}
\alias{EnvRaiser$env_raise_warning_classed}
\alias{EnvRaiser$env_raise_message}
\alias{EnvRaiser$env_raise_condition}
\alias{EnvRaiser$env_raise_condition_classed}
\alias{$.EnvRaiser}
\alias{[[.EnvRaiser}
\title{EnvRaiser  Class}
\format{
An object of class \code{environment} of length 9.
}
\source{
Generated by miniextendr from Rust type `EnvRaiser`

Generated by miniextendr from `EnvRaiser::new`

Generated by miniextendr from `EnvRaiser::env_id`

Generated by miniextendr from `EnvRaiser::env_raise_error`

Generated by miniextendr from `EnvRaiser::env_raise_error_classed`

Generated by miniextendr from `EnvRaiser::env_raise_warning`

Generated by miniextendr from `EnvRaiser::env_raise_warning_classed`

Generated by miniextendr from `EnvRaiser::env_raise_message`

Generated by miniextendr from `EnvRaiser::env_raise_condition`

Generated by miniextendr from `EnvRaiser::env_raise_condition_classed`
}
\usage{
EnvRaiser

\method{$}{EnvRaiser}(self, name)

\method{[[}{EnvRaiser}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
EnvRaiser  Class
}
\keyword{datasets}
