% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{error_in_r_i32_err}
\alias{error_in_r_i32_err}
\title{Function returning Result<i32, String> with Err -- becomes error value}
\source{
Generated by miniextendr from Rust fn \code{error_in_r_i32_err}
}
\usage{
error_in_r_i32_err()
}
\description{
Function returning Result<i32, String> with Err -- becomes error value
}
