% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{error_in_r_option_none}
\alias{error_in_r_option_none}
\title{Standalone function that returns Option::None (unit) -- becomes error value}
\source{
Generated by miniextendr from Rust fn \code{error_in_r_option_none}
}
\usage{
error_in_r_option_none()
}
\description{
Standalone function that returns Option::None (unit) -- becomes error value
}
