% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ErrorInRR6Widget}
\alias{ErrorInRR6Widget}
\title{ErrorInRR6Widget R6 Class}
\source{
Generated by miniextendr from Rust type \code{ErrorInRR6Widget}
}
\description{
ErrorInRR6Widget R6 Class

ErrorInRR6Widget R6 Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ErrorInRR6Widget-new}{\code{ErrorInRR6Widget$new()}}
\item \href{#method-ErrorInRR6Widget-get_name}{\code{ErrorInRR6Widget$get_name()}}
\item \href{#method-ErrorInRR6Widget-panic_method}{\code{ErrorInRR6Widget$panic_method()}}
\item \href{#method-ErrorInRR6Widget-failing_result}{\code{ErrorInRR6Widget$failing_result()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ErrorInRR6Widget-new"></a>}}
\if{latex}{\out{\hypertarget{method-ErrorInRR6Widget-new}{}}}
\subsection{Method \code{new()}}{
Create a new widget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErrorInRR6Widget$new(name, .ptr = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name for the widget.}

\item{\code{.ptr}}{Internal pointer (used by static methods, not for direct use).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ErrorInRR6Widget-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-ErrorInRR6Widget-get_name}{}}}
\subsection{Method \code{get_name()}}{
Get the name -- should always succeed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErrorInRR6Widget$get_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ErrorInRR6Widget-panic_method"></a>}}
\if{latex}{\out{\hypertarget{method-ErrorInRR6Widget-panic_method}{}}}
\subsection{Method \code{panic_method()}}{
Deliberately panic.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErrorInRR6Widget$panic_method()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ErrorInRR6Widget-failing_result"></a>}}
\if{latex}{\out{\hypertarget{method-ErrorInRR6Widget-failing_result}{}}}
\subsection{Method \code{failing_result()}}{
Return Result::Err.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErrorInRR6Widget$failing_result()}\if{html}{\out{</div>}}
}

}
}
