% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{export_control_normal}
\alias{export_control_normal}
\alias{export_control_internal}
\title{export_control_normal}
\source{
Generated by miniextendr from Rust fn \code{export_control_normal}

Generated by miniextendr from Rust fn \code{export_control_internal}
}
\usage{
export_control_normal()

export_control_internal()
}
\description{
A normal exported function for comparison.

An internal function -- callable but not exported, with @keywords internal.
}
\keyword{internal}
