% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{export_control_normal}
\alias{export_control_normal}
\alias{export_control_internal}
\alias{export_control_noexport}
\title{A normal exported function for comparison}
\source{
Generated by miniextendr from Rust fn `export_control_normal`

Generated by miniextendr from Rust fn `export_control_internal`

Generated by miniextendr from Rust fn `export_control_noexport`
}
\usage{
export_control_normal()

export_control_internal()

export_control_noexport()
}
\description{
A normal exported function for comparison

export control tests

export control tests
}
\keyword{internal}
