% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ExportControlTraitPoint}
\alias{ExportControlTraitPoint}
\alias{ExportControlTraitPoint$new}
\alias{ExportControlTraitPoint$RDebug$debug_str}
\alias{ExportControlTraitPoint$RDebug$debug_str_pretty}
\title{ExportControlTraitPoint  Class}
\usage{
ExportControlTraitPoint
}
\description{
ExportControlTraitPoint  Class
}
\keyword{internal}
