% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{ExportControlTraitPoint}
\alias{ExportControlTraitPoint}
\alias{ExportControlTraitPoint$new}
\alias{ExportControlTraitPoint$RDebug$debug_str}
\alias{ExportControlTraitPoint$RDebug$debug_str_pretty}
\alias{ExportControlTraitPoint$RDisplay$as_r_string}
\title{ExportControlTraitPoint  Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type \code{ExportControlTraitPoint}

Generated by miniextendr from \code{ExportControlTraitPoint::new}
}
\usage{
ExportControlTraitPoint
}
\description{
ExportControlTraitPoint  Class
}
\keyword{internal}
