% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational.R
\name{expr_reference}
\alias{expr_reference}
\title{Create a column reference expression}
\usage{
expr_reference(name, table = "")
}
\arguments{
\item{name}{the column name to be referenced}

\item{table}{the optional table name or a relation object to be referenced}
}
\value{
a column reference expression
}
\description{
Create a column reference expression
}
\examples{
col_ref_expr <- expr_reference("some_column_name")
col_ref_expr2 <- expr_reference("some_column_name", "some_table_name")
}
