% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{external_slice_len}
\alias{external_slice_len}
\title{Create an ExternalSlice from a Vec, return its length}
\source{
Generated by miniextendr from Rust fn \code{external_slice_len}
}
\usage{
external_slice_len(data)
}
\arguments{
\item{data}{(no documentation available)}
}
\description{
Create an ExternalSlice from a Vec, return its length
}
