% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rpkg_externalptr_identity}
\alias{rpkg_externalptr_identity}
\alias{ptr_identity}
\alias{ptr_pick_larger}
\title{Returns the ExternalPtr unchanged -- should preserve R identity}
\source{
Generated by miniextendr from Rust fn \code{ptr_identity}

Generated by miniextendr from Rust fn \code{ptr_pick_larger}
}
\usage{
ptr_identity(x)

ptr_pick_larger(a, b)
}
\arguments{
\item{x}{An ExternalPtr wrapping a PtrIdentityTest.}

\item{a}{An ExternalPtr wrapping a PtrIdentityTest.}

\item{b}{An ExternalPtr wrapping a PtrIdentityTest.}
}
\description{
Returns the ExternalPtr unchanged -- should preserve R identity

Takes two ExternalPtrs, returns the one with the larger value
}
\examples{
a <- PtrIdentityTest$new(10L)
b <- ptr_identity(a)
identical(a, b) # TRUE
}
