% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{extptr_point_new}
\alias{extptr_point_new}
\title{extptr_point_new}
\usage{
extptr_point_new(x, y)
}
\arguments{
\item{x}{Numeric x-coordinate.}

\item{y}{Numeric y-coordinate.}
}
\description{
Create a new Point ExternalPtr with the given coordinates.
}
