% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rpkg_externalptr}
\alias{rpkg_externalptr}
\alias{extptr_counter_new}
\alias{extptr_point_new}
\title{extptr_counter_new}
\source{
Generated by miniextendr from Rust fn \code{extptr_counter_new}

Generated by miniextendr from Rust fn \code{extptr_point_new}
}
\usage{
extptr_counter_new(initial)

extptr_point_new(x, y)
}
\arguments{
\item{initial}{Initial value for the counter.}

\item{x}{Numeric x-coordinate.}

\item{y}{Numeric y-coordinate.}
}
\description{
Create a new Counter ExternalPtr with the given initial value.

Create a new Point ExternalPtr with the given coordinates.
}
\examples{
ptr <- extptr_counter_new(1L)
p <- extptr_point_new(0.1, 0.2)
}
