% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{external_slice_len}
\alias{external_slice_len}
\alias{external_slice_sum}
\title{Create an ExternalSlice from a Vec, return its length}
\source{
Generated by miniextendr from Rust fn `external_slice_len`

Generated by miniextendr from Rust fn `external_slice_sum`
}
\usage{
external_slice_len(data)

external_slice_sum(data)
}
\arguments{
\item{data}{(no documentation available)}
}
\description{
Create an ExternalSlice from a Vec, return its length

Create an ExternalSlice, access elements via as_slice
}
