% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{extptr_any_erased_downcast}
\alias{extptr_any_erased_downcast}
\title{Test erased downcast_ref returns correct value}
\source{
Generated by miniextendr from Rust fn \code{extptr_any_erased_downcast}
}
\usage{
extptr_any_erased_downcast(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
Test erased downcast_ref returns correct value
}
