% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{Extract}
\alias{Extract}
\alias{[,fm,ANY,ANY,ANY-method}
\alias{[,fm,ANY,missing,ANY-method}
\alias{[,fm,missing,ANY,ANY-method}
\alias{[,fmV,ANY,ANY,ANY-method}
\alias{[<-,fm,ANY,missing,ANY-method}
\alias{[<-,fm,missing,ANY,ANY-method}
\title{Extract Parts of an object}
\usage{
\S4method{[}{fm,ANY,missing,ANY}(x, i, j, drop = TRUE)

\S4method{[}{fm,missing,ANY,ANY}(x, i, j, drop = TRUE)

\S4method{[}{fm,ANY,ANY,ANY}(x, i, j, drop = TRUE)

\S4method{[}{fmV,ANY,ANY,ANY}(x, i)

\S4method{[}{fm,ANY,missing,ANY}(x, i, j) <- value

\S4method{[}{fm,missing,ANY,ANY}(x, i, j) <- value
}
\arguments{
\item{x}{object from which to extract elements.}

\item{i, j}{indices specifying elements to extract. Indices are integer
or numeric vectors.}

\item{drop}{for matrices. If \code{TRUE}, the result is coerced to a vector.}

\item{value}{a FlashR vector or matrix that contains new data for the rows}
}
\description{
Operators acting on FlashR vectors and matrices to extract parts of
the objects.
}
\details{
FlashR supports extracting data from an object as well as replacing data
of a matrix. To assign data in a matrix, it only supports setting
the entire rows or columns.
}
\examples{
mat <- fm.runif.matrix(100, 10)
mat[1:5,]
mat[,1:5]
mat[1:5,1:5]
mat[,1] <- fm.runif(nrow(mat))
}

