% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{FallibleImpl}
\alias{FallibleImpl}
\alias{FallibleImpl$new}
\alias{FallibleImpl$inherent_value}
\alias{$.FallibleImpl}
\alias{[[.FallibleImpl}
\alias{FallibleImpl$Fallible$get_value}
\alias{FallibleImpl$Fallible$will_panic}
\title{FallibleImpl  Class}
\source{
Generated by miniextendr from Rust type \code{FallibleImpl}

Generated by miniextendr from \code{FallibleImpl::new}

Generated by miniextendr from \code{FallibleImpl::inherent_value}
}
\usage{
FallibleImpl

\method{$}{FallibleImpl}(self, name)

\method{[[}{FallibleImpl}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Create a new FallibleImpl.

Get value via inherent method.
}
