% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_query.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch Data from a Specified Table}
\usage{
fetch_data(name)
}
\arguments{
\item{name}{A character string specifying the name of the table to fetch.}
}
\value{
A data frame containing the data from the specified table.
}
\description{
Retrieves a specified table's data from the datasource, converting it to a data frame.
}
\details{
This function initializes the connection to the datasource (if not already initialized),
retrieves the specified table, and converts it to a data frame. The function requires
the user to have authenticated with Redivis and initialized the datasource.
}
\examples{
\dontrun{
  df <- fetch_data("abortion")
  head(df)
}
}
