% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redivis-connection.R
\name{fetch_table}
\alias{fetch_table}
\title{Fetch Table Data}
\usage{
fetch_table(name)
}
\arguments{
\item{name}{A character string specifying the name of the table to retrieve.}
}
\value{
A Redivis table object for the specified table.
}
\description{
This function initializes the datasource (if not already initialized) and retrieves a specified table by name.
It is primarily used by other functions to access a table from the datasource.
}
\examples{
\dontrun{
fetch_table("example_table_name")
}
}
