% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ffi_guard_fallback_ok}
\alias{ffi_guard_fallback_ok}
\alias{ffi_guard_fallback_panic}
\alias{ffi_guard_catch_unwind_ok}
\title{ffi guard tests}
\source{
Generated by miniextendr from Rust fn `ffi_guard_fallback_ok`

Generated by miniextendr from Rust fn `ffi_guard_fallback_panic`

Generated by miniextendr from Rust fn `ffi_guard_catch_unwind_ok`
}
\usage{
ffi_guard_fallback_ok()

ffi_guard_fallback_panic()

ffi_guard_catch_unwind_ok()
}
\description{
ffi guard tests

ffi guard tests

ffi guard tests
}
