% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ffi_guard_catch_unwind_ok}
\alias{ffi_guard_catch_unwind_ok}
\alias{ffi_guard_fallback_ok}
\alias{ffi_guard_fallback_panic}
\title{ffi_guard_catch_unwind_ok}
\source{
Generated by miniextendr from Rust fn \code{ffi_guard_catch_unwind_ok}

Generated by miniextendr from Rust fn \code{ffi_guard_fallback_ok}

Generated by miniextendr from Rust fn \code{ffi_guard_fallback_panic}
}
\usage{
ffi_guard_catch_unwind_ok()

ffi_guard_fallback_ok()

ffi_guard_fallback_panic()
}
\description{
Test guarded_ffi_call with CatchUnwind on a non-panicking closure.

Test guarded_ffi_call_with_fallback on a non-panicking closure.

Test guarded_ffi_call_with_fallback on a panicking closure — returns fallback.
}
