% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashgraph.R
\name{fg.exist.graph}
\alias{fg.exist.graph}
\title{Indicate whether a graph has been loaded to FlashR}
\usage{
fg.exist.graph(graph)
}
\arguments{
\item{graph}{A graph name.}
}
\value{
A boolean value: true if the graph has been loaded to FlashR.
}
\description{
This function indicates whether a graph has been loaded to FlashR.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

