% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashgraph.R
\name{fg.kmeans}
\alias{fg.kmeans}
\title{Perform k-means clustering on a data matrix.}
\usage{
fg.kmeans(mat, k, max.iters = 10, max.threads = 256, init = c("random",
  "forgy", "kmeanspp"), tol = -1)
}
\arguments{
\item{mat}{A numeric matrix of data.}

\item{k}{The number of clusters.}

\item{max.iters}{The maximum number of iterations allowed.}

\item{max.threads}{The maximum number of threads allowed (Default is 1 per core).}

\item{init}{The form of initialization to use when the algorithm begins.
The default is "random". For a desciption of each see:
http://en.wikipedia.org/wiki/K-means_clustering#Initialization_methods.}

\item{tol}{The tolerance for convergence. Between 0 and 1 and is the minimum fraction
                of cluster changes necessary to cause non-convergence. Default is -1 which
                represents no cluster changes.}
}
\value{
A named list with the following members:
        iters: The number of (outer) iterations performed.
        centers: A matrix of cluster centers.
        cluster: A vector of integers (from 1:k) indicating the cluster to which each point is allocated.
        sizes: The number of points in each cluster.
}
\description{
Assign each row of a matrix to a cluster denoted by a numeric. The clusters
are based on the euclidean distance of each row to one another. The assigned
cluster will have the smallest distance from the cluster center mean.
}
\examples{
num.clusts <- 3
mat <- replicate(5, rnorm(10))
kms <- fg.kmeans(mat, num.clusts)
}
\author{
Disa Mhembere <disa@jhu.edu>
}

