% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashgraph.R
\name{fg.list.graph}
\alias{fg.list.graph}
\alias{fg.list.graphs}
\title{List graphs loaded to FlashGraphR}
\usage{
fg.list.graphs()
}
\value{
A list of graphs in a data frame. The first column of the data
frame is the graph name. The second column indicates whether a graph
is stored in memory or on disks.
}
\description{
This function lists all graphs that have been loaded to FlashGraphR.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

