% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashgraph.R
\name{fg.multiply}
\alias{fg.multiply}
\alias{fg.multiply.matrix}
\title{Sparse matrix multiplication}
\usage{
fg.multiply(graph, vec, transpose = FALSE)

fg.multiply.matrix(graph, m, transpose = FALSE)
}
\arguments{
\item{graph}{The FlashGraphR object}

\item{vec}{A numueric vector}

\item{transpose}{Indicate whether or not to multiply with the transpose of
the matrix. It is ignored by an undirected graph.}

\item{m}{A numeric dense matrix.}
}
\value{
`fg.multiply' returns a numeric vector and `fg.multiply.matrix'
returns a numeric dense matrix.
}
\description{
Multiply a sparse matrix with a dense vector or a dense matrix.
}
\details{
`fg.multiply' multiplies a sparse matrix with a dense vector.

`fg.multiply.matrix' multiplies a sparse matrix with a dense matrix.

Note that the sparse matrix is represented by a graph. A symmetric matrix
is represented by an undirected graph and an asymmetric matrix is
represented by a directed graph. For an asymmetric matrix, we can multiply
the vector with the transpose of the matrix without really transposing
the matrix. Right now it only supports multiplication on the adjacency
matrix of the graph.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

