% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashgraph.R
\name{fg.spectral}
\alias{fg.spectral}
\alias{fg.spectral.clusters}
\title{Perform spectral clustering}
\usage{
fg.spectral.clusters(fg, k, ase)
}
\arguments{
\item{fg}{The FlashGraph object of the undirected graph.}

\item{k}{The number of clusters.}

\item{ase}{The function computes spectral embedding and generates
eigenvectors.}
}
\value{
A vector of integers (from '1:k') indicating the cluster to which
        each vertex belongs to.
}
\description{
Spectral clustering partitions an undirected graph into k groups based on
an embedding of the graph, which computes eigenvectors from a variant
of the adjacency matrix of the graph. Spectral clustering runs
KMeans clustering on eigenvectors.
}
\examples{
fg <- fg.load.graph("edge_list.txt")
res <- fg.spectral.clusters(fg, 10)
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

