% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fg.subgraph}
\alias{fg.fetch.subgraph}
\alias{fg.subgraph}
\title{Generate an induced subgraph}
\usage{
fg.fetch.subgraph(graph, vertices)
}
\arguments{
\item{graph}{The FlashGraphR object}

\item{vertices}{A numeric vector that contains the ids of vertices in
the induced subgraph.}
}
\value{
An iGraph object that represents the subgraph
}
\description{
Generate an induced subgraph that contains the specified vertices from
the input given graph.
}
\details{
The subgraph generated by this function is returned to the user as an
iGraph object.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

