% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_choose_2_analysis.R
\name{find_diff_and_ref_exparts_pairwise}
\alias{find_diff_and_ref_exparts_pairwise}
\title{Find differential and reference exonic parts for pairwise combinations}
\usage{
find_diff_and_ref_exparts_pairwise(
  g,
  source,
  sink,
  path_pairs,
  parsed_partitions,
  parsed_paths,
  tx_ex_parts,
  pairwise_list,
  gene = gene
)
}
\description{
Find differential and reference exonic parts for pairwise combinations
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_edge_names(g)
result <- grase::n_choose_2_paths(
  gene = "ENSG00000000003.14", g = g, sg = NULL, outdir = tempdir()
)
}
}
