% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartition_analysis.R
\name{find_reference_exonic_part}
\alias{find_reference_exonic_part}
\title{Find reference exonic part near bubble region (legacy version)}
\usage{
find_reference_exonic_part(
  source,
  sink,
  ex_part1_set,
  ex_part2_set,
  tx_ex_part1_set,
  tx_ex_part2_set,
  g
)
}
\arguments{
\item{source}{Character string. Name of the source vertex of the bubble.}

\item{sink}{Character string. Name of the sink vertex of the bubble.}

\item{ex_part1_set}{Integer vector. Edge indices of exonic parts covered by partition 1.}

\item{ex_part2_set}{Integer vector. Edge indices of exonic parts covered by partition 2.}

\item{tx_ex_part1_set}{Integer vector. Edge indices of exonic parts covered by representative transcript of partition 1.}

\item{tx_ex_part2_set}{Integer vector. Edge indices of exonic parts covered by representative transcript of partition 2.}

\item{g}{An \code{igraph} directed acyclic graph representing a gene splicing graph.}
}
\description{
Find reference exonic part near bubble region (legacy version)
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_edge_names(g)
ref <- grase::find_reference_exonic_part(
  source = "3", sink = "7",
  ex_part1_set = integer(0), ex_part2_set = integer(0),
  tx_ex_part1_set = integer(0), tx_ex_part2_set = integer(0),
  g = g
)
}
}
