% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focalexons.R
\name{find_reference_exonic_part}
\alias{find_reference_exonic_part}
\title{Find reference exonic part near bubble region}
\usage{
find_reference_exonic_part(
  source,
  sink,
  ex_part1_set,
  ex_part2_set,
  tx_ex_part1_set,
  tx_ex_part2_set,
  g
)
}
\description{
Find reference exonic part near bubble region
}
