% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_analysis.R
\name{find_reference_exonic_part_general}
\alias{find_reference_exonic_part_general}
\title{Find reference exonic parts for multiple partitions (generalization of simple binary case)}
\usage{
find_reference_exonic_part_general(
  source,
  sink,
  ex_part_sets,
  tx_ex_part_sets,
  g_expart
)
}
\description{
Find reference exonic parts for multiple partitions (generalization of simple binary case)
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_edge_names(g)
ex_or_in_vec <- igraph::edge_attr(g, "ex_or_in")
g_expart <- igraph::delete_edges(g, igraph::E(g)[ex_or_in_vec == "ex"])
ref <- grase::find_reference_exonic_part_general(
  source = "3", sink = "7",
  ex_part_sets = list(character(0), character(0)),
  tx_ex_part_sets = list(character(0), character(0)),
  g_expart = g_expart
)
}
}
