% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartition_analysis.R
\name{find_reference_exonic_part_simple}
\alias{find_reference_exonic_part_simple}
\title{Find reference exonic part near bubble region between two partitions}
\usage{
find_reference_exonic_part_simple(
  source,
  sink,
  ex_part1_set,
  ex_part2_set,
  tx_ex_part1_set,
  tx_ex_part2_set,
  g_expart
)
}
\arguments{
\item{source}{Character string. Name of the source vertex of the bubble.}

\item{sink}{Character string. Name of the sink vertex of the bubble.}

\item{ex_part1_set}{Character vector. Exonic part edge names covered by partition 1.}

\item{ex_part2_set}{Character vector. Exonic part edge names covered by partition 2.}

\item{tx_ex_part1_set}{Character vector. Exonic part edge names covered by representative transcript of partition 1.}

\item{tx_ex_part2_set}{Character vector. Exonic part edge names covered by representative transcript of partition 2.}

\item{g_expart}{An \code{igraph} graph of exonic parts.}
}
\description{
Find reference exonic part near bubble region between two partitions
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_edge_names(g)
ex_or_in_vec <- igraph::edge_attr(g, "ex_or_in")
g_expart <- igraph::delete_edges(g, igraph::E(g)[ex_or_in_vec == "ex"])
ref <- grase::find_reference_exonic_part_simple(
  source = "3", sink = "7",
  ex_part1_set = character(0), ex_part2_set = character(0),
  tx_ex_part1_set = character(0), tx_ex_part2_set = character(0),
  g_expart = g_expart
)
}
}
