% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{find_tx_with_epath}
\alias{find_tx_with_epath}
\title{Find transcripts that use given edge paths}
\usage{
find_tx_with_epath(g, trans, edges_list)
}
\description{
Find transcripts that use given edge paths
}
