% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_models}
\alias{fit_models}
\title{fit_model}
\usage{
fit_models(formula, data)
}
\arguments{
\item{formula}{formula}

\item{data}{data}
}
\value{
model
}
\description{
fit_model
}
\details{
DETAILS
}
\seealso{
\code{\link[lme4]{glmer}}
\code{\link[sjPlot]{tab_model}}
}
