% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R, R/oxy-fit_models.R
\docType{data}
\name{tj}
\alias{tj}
\alias{fit_models}
\title{FUNCTION_TITLE}
\format{An object of class \code{NULL} of length 0.}
\usage{
tj

fit_models(formula, data)
}
\arguments{
\item{formula}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}

\item{formula}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[lme4]{glmer}}
 \code{\link[sjPlot]{tab_model}}

\code{\link[lme4]{glmer}}
 \code{\link[sjPlot]{tab_model}}
}
\keyword{datasets}
