% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{\preformatted{    fit_models fits a model
}}
\usage{
fit_models(formula, data)
}
\arguments{
\item{formula}{formula to be used in fitting of model}

\item{data}{data.frame to be used in fitting of models}
}
\value{
\preformatted{   model diagnostics
}
}
\description{
fit_models is a function for fitting a poission fixed model
using the epilepsy data set.
}
\details{
\preformatted{  see the glmer function for a description of details
}
}
\examples{
fit_models(formula = y ~ trt*post + (1|subject), data = df_epil)
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[lme4]{glmer}}
}
