% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-fit_models.R
\name{fit_models}
\alias{fit_models}
\title{FUNCTION_TITLE}
\usage{
fit_models(formula, data)
}
\arguments{
\item{formula}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[lme4]{glmer}}
 \code{\link[sjPlot]{tab_model}}
}
