% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Border_Direction_FIx.R
\name{FixBorderDir}
\alias{FixBorderDir}
\title{Oringinally all border cells are initialized to point out of the domain. This script checks the slopes and areas to determine if they need to swap}
\usage{
FixBorderDir(direction, dem, d4 = c(1, 2, 3, 4))
}
\description{
Oringinally all border cells are initialized to point out of the domain. This script checks the slopes and areas to determine if they need to swap
}
