% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatterSIDb.R
\name{flatterSIDb}
\alias{flatterSIDb}
\title{Flattens SIDb database object for easier querying}
\usage{
flatterSIDb(database)
}
\arguments{
\item{database}{A database with the SIDb structure}
}
\value{
R list with two elements, linked by shared column "ID":\cr
1) \emph{timeseries}, a list of unique timeseries, each a 2-column dataframe (time, response)\cr
2) \emph{vars}, a list of all site, incubation and variables fields flattened into a dataframe for each entry
}
\description{
Flattens SIDb database object for easier querying
}
\examples{
sidb.flatter <- flatterSIDb(sidb)
}
