% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashmatrix.R
\name{fm.agg}
\alias{fm.agg}
\title{Aggregation on a FlashMatrix object.}
\usage{
fm.agg(fm, op)
}
\arguments{
\item{fm}{a FlashMatrix object}

\item{op}{the reference or the name of a predefined basic operator or
the reference to an aggregation operator returned by `fm.create.agg.op'.}
}
\value{
a scalar
}
\description{
This function accepts a basic operator and perform aggregation on
the FlashMatrix object with the basic operator.
}

