% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashmatrix.R
\name{fm.agg}
\alias{fm.agg}
\title{Aggregation on a FlashMatrixR object.}
\usage{
fm.agg(bop, fm)
}
\arguments{
\item{bop}{a basic operator}

\item{fm}{a FlashMatrixR object}
}
\value{
a scalar
}
\description{
This function accepts a basic operator and perform aggregation on
the FlashMatrixR object with the basic operator.
}

