% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashmatrix.R
\name{fm.agg}
\alias{fm.agg}
\title{Aggregation on a FlashMatrix object.}
\usage{
fm.agg(fm, op)
}
\arguments{
\item{fm}{a FlashMatrix object}

\item{op}{a basic operator or an aggregation operator}
}
\value{
a scalar
}
\description{
This function accepts a basic operator and perform aggregation on
the FlashMatrix object with the basic operator.
}

