% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.as.factor}
\alias{fm.as.factor}
\title{FlashR factor vector.}
\usage{
fm.as.factor(fm, num.levels = -1)
}
\arguments{
\item{fm}{a FlashR vector.}

\item{num.levels}{The number of levels in the factor vector.}
}
\value{
a FlashR factor vector.
}
\description{
\code{fm.as.factor} converts a FlashR vector to a FlashR factor vector.
}
\details{
Currently, this function only works for integer input vectors. If
the input vector isn't integers, it converts the vector to an integer
vector. By default, it uses the maximal value as the number of levels
for the factor.
}
\examples{
vec <- fm.as.factor(as.integer(fm.runif(100, min=0, max=100)))
}

