% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashmatrix.R
\name{fm.as.factor}
\alias{fm.as.factor}
\title{Convert a FlashMatrixR vector to a FlashMatrixR factor vector.}
\usage{
fm.as.factor(fm, num.levels = -1)
}
\arguments{
\item{fm}{a FlashMatrixR vector.}
}
\value{
a FlashMatrixR factor vector.
}
\description{
Convert a FlashMatrixR vector to a FlashMatrixR factor vector.
}

