% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.as.vector}
\alias{fm.as.vector}
\title{Convert a FlashMatrix matrix to a FlashMatrix vector.}
\usage{
fm.as.vector(obj)
}
\arguments{
\item{fm}{a FlashMatrix matrix}
}
\value{
a FlashMatrix vector
}
\description{
The matrix must have only one row or one column. Otherwise, the function
returns an error.
}

