% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashmatrix.R
\name{fm.as.vector}
\alias{fm.as.vector}
\title{Convert a FlashMatrixR matrix to a FlashMatrixR vector.}
\usage{
fm.as.vector(obj)
}
\arguments{
\item{fm}{a FlashMatrixR matrix}
}
\value{
a FlashMatrixR vector
}
\description{
The matrix must have only one row or one column. Otherwise, the function
returns an error.
}

