% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ASE.R
\name{fm.ase}
\alias{fm.ase}
\alias{fm.spectral.embedding}
\title{Spectral embedding}
\usage{
fm.spectral.embedding(fm, nev, which = c("A, Aug, L, nL"), c = 1/nrow(fm),
  ncv = 2 * nev, tol = 1e-12)
}
\arguments{
\item{fm}{The FlashMatrixR object}

\item{nev}{The number of eigenvalues/vectors required.}

\item{which}{The type of the embedding.}

\item{c}{The constant used in the Aug variant of embedding.}

\item{ncv}{The number of vectors in the vector subspace.}

\item{tol}{Numeric scalar. Stopping criterion: the relative accuracy
of the Ritz value is considered acceptable if its error is less than
tol times its estimated value. If this is set to zero then machine
precision is used.}
}
\value{
A named list with the following members:
\item{values}{Numeric vector, the desired eigenvalues.}
\item{vectors}{Numeric matrix, the desired eigenvectors as columns. If complex=TRUE
  (the default for non-symmetric problems), then the matrix is complex.}
\item{options}{A named list with the supplied options and some information
  about the performed calculation, including an ARPACK exit code.}
}
\description{
This computes spectral embedding of a given graph.
It supports four variants:
}
\details{
Adjacency matrix (A),

Augmented matrix (Aug), A - c * D, where c is a constant provided by a user.

Laplacian matrix (L), L = D - A,

Normalized Laplacian matrix (nL), nL = I - D^(-1/2) * A * D^(-1/2)

For the Laplacian matrix and the normalized Laplacian matrix, this function
only accepts a symmetric sparse matrix.

For the adjacency matrix (A) and augmented matrix (Aug), it computes
the largest eigenvalues and the corresponding eigenvectors.
For the Laplacian matrix and the normalized Laplacian matrix, it computes
the eigenvectors for the smallest eigenvalues.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

