% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashmatrix.R
\name{fm.basic.op}
\alias{fm.basic.op}
\alias{fm.get.basic.op}
\alias{fm.get.basic.uop}
\alias{fm.init.basic.op}
\title{The basic operators supported by FlashMatrix.}
\usage{
fm.get.basic.op(name)

fm.get.basic.uop(name)

fm.init.basic.op()
}
\arguments{
\item{name}{the name of the basic operator.}
}
\value{
a reference to the specified basic operator.
}
\description{
The basic operators are mainly used by the FlashMatrix functions that
accept operators as arguments. Such a function includes `fm.mapply',
`fm.inner.prod', etc.
}
\details{
`fm.get.basic.op' gets the predefined basic binary operator specified by a user.

`fm.get.basic.uop' gets the predefined basic unary operator specified by a user.

`fm.init.basic.op' defines the basic operators.

`fm.bo.add' is the predifined basic binary operator for addition.
`fm.bo.sub' is the predifined basic binary operator for subtraction.
`fm.bo.mul' is the predifined basic binary operator for multiplication.
`fm.bo.div' is the predifined basic binary operator for division.
`fm.bo.min' is the predifined basic binary operator for computing minimum.
`fm.bo.max' is the predifined basic binary operator for computing maximum.
`fm.bo.pow' is the predifined basic binary operator for computing exponential.
`fm.bo.eq', `fm.bo.gt' and `fm.bo.ge' are the predefined basic
logical operators to compare two elements: ==, >, >=.

`fm.buo.neg' is the predefined basic unary operator for negate.
`fm.buo.sqrt' is the predefined basic unary operator for square root.
`fm.buo.abs' is the predefined basic unary operator for absolute value.
`fm.buo.not' is the predefined logical NOT Operator.
`fm.buo.ceil' is the predefined basic unary Operator of computing a ceiling
of a number.
`fm.buo.floor' is the predefined basic unary Operator of computing a floor
of a number.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

