% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\docType{methods}
\name{fm.bind}
\alias{fm.bind}
\alias{fm.cbind.list}
\alias{fm.rbind.list}
\title{Combine FlashR Vectors/Matrices by Rows or Columns}
\usage{
fm.rbind.list(objs)

fm.cbind.list(objs)

\S4method{rbind}{fm}(..., deparse.level = 1)

\S4method{rbind}{fmV}(..., deparse.level = 1)

\S4method{cbind}{fm}(..., deparse.level = 1)

\S4method{cbind}{fmV}(..., deparse.level = 1)
}
\arguments{
\item{objs}{A list of FlashR matrices.

All arguments for \code{rbind} and \code{cbind} have to be either all
vectors or all matrices.}

\item{...}{A list of FlashR matrices.}
}
\value{
A FlashR matrix.
}
\description{
Take a list of FlashR vectors/matrices and combine them by Columns or rows
respectively.
}
\examples{
mat1 <- fm.runif.matrix(100, 10)
mat2 <- fm.runif.matrix(100, 10)
mat <- rbind(mat1, mat2)
mat <- cbind(mat1, mat2)
mat.list <- list(mat1, mat2)
mat <- fm.rbind.list(mat.list)
mat <- fm.cbind.list(mat.list)
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

