% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashmatrix.R
\name{fm.bind}
\alias{fm.bind}
\alias{fm.cbind}
\alias{fm.rbind}
\title{Combine FlashR matrices by rows or columns.}
\usage{
fm.rbind(...)

fm.cbind(...)
}
\arguments{
\item{...}{A list of FlashR matrices.

Currently, `fm.rbind' only supports combining a list of wide matrices.
`fm.cbind' can only combine a list of tall matrices.}
}
\value{
A FlashR matrix.
}
\description{
Take a list of FlashR matrices and combine them by Columns or rows
respectively.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

