% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.bind}
\alias{fm.bind}
\alias{fm.cbind}
\alias{fm.cbind.list}
\alias{fm.rbind}
\alias{fm.rbind.list}
\title{Combine FlashR matrices by rows or columns.}
\usage{
fm.rbind(...)

fm.rbind.list(objs)

fm.cbind(...)

fm.cbind.list(objs)
}
\arguments{
\item{...}{A list of FlashR matrices.}

\item{objs}{A list of FlashR matrices.

These functions are similar to the R counterparts: `rbind' and `cbind'.
Currently, `fm.rbind' and `fm.rbind.list' only supports combining
a list of wide matrices. `fm.cbind' and `fm.cbind.list' can only combine
a list of tall matrices.}
}
\value{
A FlashR matrix.
}
\description{
Take a list of FlashR matrices and combine them by Columns or rows
respectively.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

