% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fm.conv.FM2R}
\alias{fm.conv.FM2R}
\title{Convert a FlashMatrixR object to a regular R object}
\usage{
fm.conv.FM2R(obj)
}
\arguments{
\item{obj}{a FlashMatrixR object}
}
\value{
a regular R object.
}
\description{
Convert a FlashMatrixR object to a regular R object
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

