% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fm.conv.R2FM}
\alias{fm.conv.R2FM}
\title{Convert a regular R object to a FlashMatrixR object.}
\usage{
fm.conv.R2FM(obj, byrow = FALSE)
}
\arguments{
\item{obj}{a regular R object}

\item{byrow}{a logical value to determine the data layout of a FlashMatrixR
matrix.}
}
\value{
a FlashMatrixR object.
}
\description{
If the R object is a matrix, `byrow' determines how data in the generated
FlashMatrixR object is organized in memory.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

