% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.conv.store}
\alias{fm.conv.store}
\title{Convert the Storage of an Object.}
\usage{
fm.conv.store(fm, in.mem, name = "")
}
\arguments{
\item{fm}{a FlashMatrix object.}

\item{in.mem}{a logical value indicating whether to store the FlashMatrix
object in memory.}

\item{name}{a string to indicate the name of the new FlashMatrix object.}
}
\value{
a new FlashMatrix object with data stored in specified storage.
}
\description{
This function converts the storage of a FlashMatrix vector/matrix.
The storage can be memory or disks.
}
\details{
If a user provides \code{name} and \code{in.mem} is \code{TRUE},
the vector/matrix will be kept on disks persistently. That is, even if a user
exits from R, the vector/matrix will still be kept on disks.
}

