% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fm.eigen}
\alias{fm.eigen}
\title{Eigensolver}
\usage{
fm.eigen(func, extra = NULL, sym = FALSE, options = fm.eigen.default,
  env = parent.frame())
}
\arguments{
\item{func}{The function to perform the matrix-vector multiplication.}

\item{extra}{Extra argument to supply to `func'.}

\item{sym}{Logical scalar, whether the input matrix is symmetric.}

\item{options}{Options to Anasazi.

The `options' argument specifies what kind of computation to perform.
It is a list with the following members, which correspond directly to
Anasazi parameters:

which Specify which eigenvalues/vectors to compute, character
             constant with exactly two characters.
Possible values for symmetric input matrices:
"LA' Compute 'nev' largest (algebraic) eigenvalues.
"SA" Compute "nev" smallest (algebraic) eigenvalues.
"LM" Compute `nev' largest (in magnitude) eigenvalues.
"SM" Compute `nev' smallest (in magnitude) eigenvalues.
nev Numeric scalar. The number of eigenvalues to be computed.
ncv Number of Lanczos vectors to be generated.}
}
\value{
A named list with the following members:
        values: Numeric vector, the desired eigenvalues.
        vectors: Numeric matrix, the desired eigenvectors as columns.
}
\description{
Compute eigenvalues/vectors of the adjacency matrix of an undirected graph.
}
\details{
This eigensolver is powered by Anasazi package of Trilinos.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

