% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashmatrix.R
\name{fm.get.eles}
\alias{fm.get.cols}
\alias{fm.get.eles}
\alias{fm.get.rows}
\title{Get a submatrix from a FlashMatrixR matrix}
\usage{
fm.get.cols(fm, idxs)

fm.get.rows(fm, idxs)
}
\arguments{
\item{fm}{A FlashMatrixR matrix}

\item{idxs}{an array of column indices in fm.}
}
\value{
a submatrix
}
\description{
`fm.get.rows' gets specified rows in a FM matrix.
`fm.get.cols' gets specified columns in a FM matrix.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

