% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.get.matrix}
\alias{fm.get.dense.matrix}
\alias{fm.get.matrix}
\alias{fm.load.dense.matrix}
\alias{fm.load.dense.matrix.bin}
\alias{fm.load.sparse.matrix}
\title{Load a sparse matrix to FlashR.}
\usage{
fm.get.dense.matrix(name)

fm.load.dense.matrix(name, in.mem, ele.type = "D", delim = ",",
  ncol = .Machine$integer.max, mat.name = "")

fm.load.dense.matrix.bin(name, in.mem, nrow, ncol, byrow, ele.type,
  mat.name = "")

fm.load.sparse.matrix(spm, spm.idx, t.spm = NULL, t.spm.idx = NULL,
  in.mem = TRUE)
}
\arguments{
\item{name}{a string indicating the name of the dense matrix in the system.}

\item{in.mem}{Determine the loaded matrix is stored in memory or on SAFS.}

\item{ele.type}{The element type in a matrix.}

\item{delim}{The delimiter of separating elements in the text format.}

\item{ncol}{the number of columns in the binary dense matrix.}

\item{mat.name}{a string indicating the name of the matrix to be loaded.}

\item{nrow}{the number of rows in the binary dense matrix.}

\item{byrow}{a logical value indicating if the data in the binary matrix
is stored by rows.}

\item{spm}{The file that stores the sparse matrix.}

\item{spm.idx}{The file that stores the index of the sparse matrix.}

\item{t.spm}{The file that stores the transpose of the sparse matrix.}

\item{t.spm.idx}{The file that stores the index of the transpose of the sparse matrix.}
}
\value{
a FlashMatrix matrix.
}
\description{
Load a sparse matrix to FlashR from difference sources.
}
\details{
\code{fm.get.dense.matrix} gets a FlashMatrix dense matrix from SAFS.

\code{fm.load.dense.matrix} loads a dense matrix in the text format from
the Linux filesystem.

\code{fm.load.dense.matrix.bin} loads a dense matrix in the binary format
from the Linux filesystem.

\code{fm.load.sparse.matrix} loads a FlashMatrix sparse matrix from files.
The matrix in the file is in the FlashMatrix format.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

