% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.get.matrix}
\alias{fm.get.dense.matrix}
\alias{fm.get.matrix}
\alias{fm.get.sparse.matrix}
\alias{fm.load.sparse.matrix}
\title{Load a sparse matrix to FlashR.}
\usage{
fm.get.sparse.matrix(fg)

fm.get.dense.matrix(name)

fm.load.sparse.matrix(mat, index, t.mat = NULL, t.index = NULL,
  in.mem = TRUE)
}
\arguments{
\item{fg}{A FlashGraph object.}

\item{mat.file}{The file that stores the sparse matrix.}

\item{index.file}{The file that stores the index of the sparse matrix.}
}
\value{
a FlashMatrix matrix.
}
\description{
Load a sparse matrix to FlashR from difference sources.
}
\details{
`fm.get.sparse.matrix' gets a FlashMatrix sparse matrix that references
a matrix represented by a FlashGraph object.

`fm.get.dense.matrix' gets a FlashMatrix dense matrix from the underlying

`fm.load.sparse.matrix' loads a FlashMatrix sparse matrix from files.
The matrix in the file is in the FlashMatrix format.
}
\examples{
fg <- fg.load.graph("graph.adj", "graph.index")
fm <- fm.get.sparse.matrix(fg)
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

