% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fm.get.matrix}
\alias{fm.get.matrix}
\alias{fm.load.matrix}
\title{Load a sparse matrix to FlashMatrixR.}
\usage{
fm.get.matrix(fg)

fm.load.matrix(mat.file, index.file)
}
\arguments{
\item{fg}{A FlashGraphR object.}

\item{mat.file}{The file that stores the sparse matrix.}

\item{index.file}{The file that stores the index of the sparse matrix.}
}
\value{
a FlashMatrixR matrix.
}
\description{
Load a sparse matrix to FlashMatrixR from difference sources.
}
\details{
`fm.get.matrix' gets a FlashMatrixR matrix that references a matrix
represented by a FlashGraphR object.

`fm.load.matrix' loads a FlashMatrixR matrix from files.
The matrix in the file is in the FlashMatrix format.
}
\examples{
fg <- fg.load.graph("graph.adj", "graph.index")
fm <- fm.get.matrix(fg)
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

