% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashmatrix.R
\name{fm.groupby}
\alias{fm.groupby}
\alias{fm.sgroupby}
\title{Groupby on a FlashMatrixR vector.}
\usage{
fm.sgroupby(obj, FUN)
}
\arguments{
\item{obj}{a FlashMatrixR vector or matrix}

\item{FUN}{an aggregation operator returned by `fm.get.basic.op'.}

\item{margin}{the subscript which the function will be applied over.
E.g., for a matrix, `1' indicates rows, `2' indicates columns.}

\item{labels}{a FlashMatrixR vector that indicates the labels of
each element in `obj'.}
}
\value{
`fm.sgroupby' returns a data frame, where the column `val' stores
all of the unique values in the original data container, and the column
`agg' stores the aggregate result of the corresponding value.
}
\description{
`fm.sgroupby' groups elements in a vector based on corresponding `labels'
and applies `FUN' to the elements in each group. `FUN' is an aggregation
operator.
}
\details{
`fm.groupby' groups rows/columns of a matrix based on corresponding `labels'
and applies `FUN' to the rows/columns in each group. `FUN' is an aggregation
operator.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

