% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.info}
\alias{fm.in.mem}
\alias{fm.info}
\alias{fm.is.object}
\alias{fm.is.sink}
\alias{fm.is.sparse}
\alias{fm.is.sym}
\alias{fm.matrix.layout}
\title{The information of a FlashR object}
\usage{
fm.is.sym(fm)

fm.matrix.layout(fm)

fm.is.sparse(fm)

fm.is.sink(fm)

fm.in.mem(fm)

fm.is.object(fm)
}
\arguments{
\item{fm}{The FlashR object}
}
\value{
\code{fm.is.sym} and \code{fm.is.sparse} returns boolean constants.
}
\description{
These functions provide the basic information of a FlashR object.
}
\details{
\code{fm.is.sym} indicates whether a matrix is symmetric.

\code{fm.matrix.layout} indicates how data in a matrix is organized.

\code{fm.is.sparse} indicates whether a matrix is sparse.

\code{fm.is.sink} indicates whether a FlashR object is a sink matrix.

\code{fm.in.mem} indicates whether a FlashR object is stored in memory.

\code{fm.is.object} indicates whether this is a FlashR object.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

