% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashmatrix.R
\name{fm.inner.prod}
\alias{fm.inner.prod}
\title{Matrix inner product}
\usage{
fm.inner.prod(fm, mat, Fun1, Fun2)
}
\arguments{
\item{fm}{A FlashMatrix matrix}

\item{mat}{A FlashMatrix dense matrix.}
}
\value{
a FlashMatrix vector if the second argument is a vector;
a FlashMatrix matrix if the second argument is a matrix.
}
\description{
Inner product of a dense matrix with a dense vector/matrix.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

