% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMeans.R
\name{fm.kmeans}
\alias{fm.kmeans}
\title{KMeans clustering}
\usage{
fm.kmeans(data, centers, max.iters = 10, debug = FALSE, use.blas = FALSE)
}
\arguments{
\item{data}{the input data matrix where each row is a data point.}

\item{centers}{either the number of clusters, say k, or a set of
initial (distinct) cluster centers. If a number, a random set
of (distinct) rows in `x' is chosen as the initial centers.}

\item{max.iters}{the maximal number of iterations.}

\item{debug}{This indicates whether to print debug info.}

\item{use.blas}{a logical value indicating whether to use BLAS to
compute Euclidean distance.}
}
\value{
a vector that contains cluster Ids for each data point.
}
\description{
Perform k-means clustering on a data matrix.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

