% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fm.mapply2}
\alias{fm.mapply2}
\title{Apply a Function to two FlashMatrixR vectors/matrices.}
\usage{
fm.mapply2(FUN, o1, o2)
}
\arguments{
\item{FUN}{the user-defined operators.}

\item{o1,}{o2 a FlashMatrixR vector/matrix.}
}
\value{
a FlashMatrixR vector/matrix.
}
\description{
`mapply' applies `FUN' to the first elements of each vector, the second
elements, the third elements, and so on. Two vectors/matrices should have
the same shape. Currently, `mapply' only accepts predefined basic operators
returned by `fm.get.basic.op'.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

