% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashmatrix.R
\name{fm.mapply2}
\alias{fm.mapply2}
\title{Apply a Function to two FlashMatrix vectors/matrices.}
\usage{
fm.mapply2(o1, o2, FUN)
}
\arguments{
\item{o1,}{o2 a FlashMatrix vector/matrix.}

\item{FUN}{the reference or the name of one of the predefined basic binary
operators.}
}
\value{
a FlashMatrix vector/matrix.
}
\description{
`mapply' applies `FUN' to the first elements of each vector, the second
elements, the third elements, and so on. Two vectors/matrices should have
the same shape. Currently, `mapply' only accepts predefined basic operators
returned by `fm.get.basic.op'.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

