% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.multiply}
\alias{fm.multiply}
\title{Matrix multiplication}
\usage{
fm.multiply(fm, mat)
}
\arguments{
\item{fm}{A FlashR matrix}

\item{mat}{A FlashR dense matrix.}
}
\value{
a FlashR vector if the second argument is a vector;
a FlashR matrix if the second argument is a matrix.
}
\description{
Multiply a sparse/dense matrix with a dense vector/matrix.
}
\examples{
mat1 <- fm.runif.matrix(1000, 100)
mat2 <- fm.runif.matrix(100, 10)
mat <- fm.multiply(mat1, mat2)
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

