% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.multiply}
\alias{fm.multiply}
\title{Matrix multiplication}
\usage{
fm.multiply(fm, mat, lazy.wide = FALSE)
}
\arguments{
\item{fm}{A FlashMatrix matrix}

\item{mat}{A FlashMatrix dense matrix.}

\item{lazy.wide}{whether to lazily evaluate multiplication of a wide matrix
with a tall matrix.}
}
\value{
a FlashMatrix vector if the second argument is a vector;
a FlashMatrix matrix if the second argument is a matrix.
}
\description{
Multiply a sparse/dense matrix with a dense vector/matrix.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

