% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fm.multiply}
\alias{fm.multiply}
\title{Matrix multiplication}
\usage{
fm.multiply(fm, mat)
}
\arguments{
\item{fm}{A FlashMatrixR matrix}

\item{mat}{A FlashMatrixR dense matrix.}
}
\value{
a FlashMatrixR vector if the second argument is a vector;
a FlashMatrixR matrix if the second argument is a matrix.
}
\description{
Multiply a sparse/dense matrix with a dense vector/matrix.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

