% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fm.read.obj}
\alias{fm.read.obj}
\title{Read a FlashMatrixR object (vector/matrix) from a file.}
\usage{
fm.read.obj(file)
}
\arguments{
\item{file}{a file in the local filesystem.}
}
\value{
a FlashMatrixR object (vector/matrix)
}
\description{
Read a FlashMatrixR object (vector/matrix) from a file.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

