% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.read.obj}
\alias{fm.read.obj}
\title{Read a FlashMatrix object (vector/matrix) from a file.}
\usage{
fm.read.obj(file)
}
\arguments{
\item{file}{a file in the local filesystem.}
}
\value{
a FlashMatrix object (vector/matrix)
}
\description{
Read a FlashMatrix object (vector/matrix) from a file.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

