% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashmatrix.R
\name{fm.rep.int}
\alias{fm.rep.int}
\title{Create a FlashMatrixR vector with replicated elements.}
\usage{
fm.rep.int(x, times)
}
\arguments{
\item{x}{A constant initial value}

\item{times}{The length of the vector to be generated.}
}
\value{
A FlashMatrixR vector
}
\description{
Create a FlashMatrixR vector with replicated elements.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

