% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.rnorm}
\alias{fm.rnorm}
\alias{fm.rnorm.matrix}
\title{The Normal Distribution}
\usage{
fm.rnorm(n, mean = 0, sd = 1, in.mem = TRUE, name = "")

fm.rnorm.matrix(nrow, ncol, mean = 0, sd = 1, in.mem = TRUE, name = "")
}
\arguments{
\item{n}{the number of random numbers to be generated.}

\item{mean}{the mean of the distribution.}

\item{sd}{the standard deviation of the distribution.}

\item{in.mem}{whether the vector is stored in memory.}

\item{name}{the name of the matrix. It's stored on disks, it's used as
the file name.}

\item{nrow}{the number of rows in the generated matrix.}

\item{ncol}{the number of columns in the generated matrix.}
}
\description{
\code{fm.rnorm} creates a FlashR vector with random numbers from
normal distribution.
\code{fm.rnorm.matrix} creates a FlashR matrix with random numbers from
normal distribution.
}
\details{
If a user provides \code{name} and \code{in.mem} is \code{TRUE}, the created
vector/matrix will be kept on disks persistently. That is, even if a user
exits from R, the vector/matrix will still be kept on disks.
}
\examples{
vec <- fm.rnorm(10) # a vector of 10 elements filled with random numbers under normal distribution.
mat <- fm.rnorm.matrix(10, 2) # a 10x2 matrix.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

