% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.runif}
\alias{fm.runif}
\alias{fm.runif.matrix}
\title{The Uniform Distribution}
\usage{
fm.runif(n, min = 0, max = 1, in.mem = TRUE, name = "")

fm.runif.matrix(nrow, ncol, min = 0, max = 1, in.mem = TRUE, name = "")
}
\arguments{
\item{n}{the number of random numbers to be generated.}

\item{min}{lower limits of the distribution.}

\item{max}{upper limits of the distribution.}

\item{in.mem}{whether the vector is stored in memory.}

\item{name}{the name of the vector. It's stored on disks, it's used as
the file name.}

\item{nrow}{the number of rows in the generated matrix.}

\item{ncol}{the number of columns in the generated matrix.}
}
\description{
\code{fm.runif} creates a FlashR vector with uniformly random numbers.
\code{fm.runif.matrix} creates a FlashR matrix with uniformly random
numbers.
}
\details{
If a user provides \code{name} and \code{in.mem} is \code{TRUE}, the created
vector/matrix will be kept on disks persistently. That is, even if a user
exits from R, the vector/matrix will still be kept on disks.
}
\examples{
vec <- fm.runif(10) # a vector of 10 elements filled with uniform random numbers.
mat <- fm.runif.matrix(10, 2) a 10x2 matrix filled with uniform random numbers.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

