% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.runif}
\alias{fm.runif}
\title{Create a FlashMatrix vector with uniformly random numbers.}
\usage{
fm.runif(n, min = 0, max = 1, in.mem = TRUE, name = "")
}
\arguments{
\item{n}{the number of random numbers to be generated.}

\item{min}{lower limits of the distribution.}

\item{max}{upper limits of the distribution.}

\item{in.mem}{whether the vector is stored in memory.}

\item{name}{the name of the vector on SAFS if the vector is stored on SAFS
persistently.}
}
\description{
Create a FlashMatrix vector with uniformly random numbers.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

