% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashmatrix.R
\name{fm.runif}
\alias{fm.runif}
\title{Create a FlashMatrix vector with uniformly random numbers.}
\usage{
fm.runif(n, min = 0, max = 1)
}
\arguments{
\item{n}{the number of random numbers to be generated.}

\item{min}{lower limits of the distribution.}

\item{max}{upper limits of the distribution.}
}
\description{
Create a FlashMatrix vector with uniformly random numbers.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

