% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fm.runif}
\alias{fm.runif}
\title{Create a FlashMatrixR vector with uniformly random numbers.}
\usage{
fm.runif(n, min = 0, max = 1)
}
\arguments{
\item{n}{the number of random numbers to be generated.}

\item{min}{lower limits of the distribution.}

\item{max}{upper limits of the distribution.}
}
\description{
Create a FlashMatrixR vector with uniformly random numbers.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

