% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\docType{methods}
\name{fm.sapply}
\alias{fm.sapply}
\alias{fm.sapply,fm-method}
\alias{fm.sapply,fmV-method}
\title{Apply a Function to a FlashR vector/matrix.}
\usage{
fm.sapply(o, FUN, set.na = TRUE)

\S4method{fm.sapply}{fm}(o, FUN, set.na = TRUE)

\S4method{fm.sapply}{fmV}(o, FUN, set.na = TRUE)
}
\arguments{
\item{o}{a FlashR vector/matrix.}

\item{FUN}{the reference or the name of a predefined uniary operator.}

\item{set.na}{a logical value indicating whether to set the value in
the output vector or matrix to NA if the value in
the corresponding location of the input vector or matrix
is NA.}
}
\value{
a FlashR vector/matrix.
}
\description{
\code{sapply} applies \code{FUN} to every element of a vector/matrix.
Currently, \code{sapply} only accepts predefined basic operators
returned by \code{fm.get.basic.uop}.
}
\examples{
mat <- fm.runif.matrix(100, 10)
res <- fm.sapply(mat, "abs")
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

