% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.sapply}
\alias{fm.sapply}
\title{Apply a Function to a FlashMatrix vector/matrix.}
\usage{
fm.sapply(o, FUN, set.na)
}
\arguments{
\item{o}{a FlashMatrix vector/matrix.}

\item{FUN}{the reference or the name of a predefined uniary operator.}
}
\value{
a FlashMatrix vector/matrix.
}
\description{
`sapply' applies `FUN' to every element of a vector/matrix.
Currently, `sapply' only accepts predefined basic operators
returned by `fm.get.basic.uop'.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

