% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashmatrix.R
\name{fm.sapply}
\alias{fm.sapply}
\title{Apply a Function to a FlashMatrixR vector/matrix.}
\usage{
fm.sapply(o, FUN)
}
\arguments{
\item{o}{a FlashMatrixR vector/matrix.}

\item{FUN}{the user-defined operators.}
}
\value{
a FlashMatrixR vector/matrix.
}
\description{
`sapply' applies `FUN' to every element of a vector/matrix.
Currently, `sapply' only accepts predefined basic operators
returned by `fm.get.basic.uop'.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

