% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashmatrix.R
\name{fm.seq.int}
\alias{fm.seq.int}
\title{Create a FlashMatrix vector with a sequence of numbers.}
\usage{
fm.seq.int(from, to, by)
}
\arguments{
\item{from}{the starting value of the sequence.}

\item{to}{the end value of the sequence.}

\item{by}{number: increment of the sequence.}
}
\value{
a sequence of numbers that belongs to [from, to]
}
\description{
Create a FlashMatrix vector with a sequence of numbers.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

