% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.seq}
\alias{fm.seq}
\alias{fm.seq.int}
\alias{fm.seq.matrix}
\title{Sequence Generation}
\usage{
fm.seq.int(from, to, by)

fm.seq.matrix(from, to, nrow, ncol, byrow = FALSE)
}
\arguments{
\item{from}{the starting value of the sequence.}

\item{to}{the end value of the sequence.}

\item{by}{number. increment of the sequence.}

\item{nrow}{the number of rows in the generated matrix.}

\item{ncol}{the number of columns in the generated matrix.}

\item{byrow}{logical. If \code{FALSE} (the default) the matrix is filled by
columns, otherwise the matrix is filled by rows.}
}
\value{
a FlashR vector or matrix filled with a sequence of numbers
        that belongs to [from, to]
}
\description{
\code{fm.seq.int} creates a FlashR vector with a sequence of numbers.
\code{fm.seq.matrix} creates a FlashR matrix with a sequence of numbers.
}
\examples{
vec <- fm.seq.int(1, 10, 1) # create a FlashR vector of 10 elements whose values are from 1 to 10.
mat <- fm.seq.matrix(1, 20, 10, 2) # create a 10x2 FlashR matrix whose rows are sequence numbers from 1 to 10.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

