% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.set.cols}
\alias{fm.set.cols}
\title{Set the specified column of a FlashMatrix matrix.}
\usage{
fm.set.cols(fm, idxs, m2)
}
\arguments{
\item{fm}{A flashMatrixR matrix}

\item{idxs}{an array of column indices in fm.}

\item{m2}{A flashMatrixR matrix or vector.}
}
\value{
a logical value that indicates whether the operation is successful.
}
\description{
Set the specified column of a FlashMatrix matrix.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

