% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.set.conf}
\alias{fm.set.conf}
\title{Reconfigure FlashMatrix}
\usage{
fm.set.conf(conf.file)
}
\arguments{
\item{conf.file}{The configuration file.}
}
\description{
This reconfigures FlashMatrix with the settings in the configuration file.
The configuration file contains a list of key-value pairs. Each line in
the file is a key-value pair in the form of "key_name=value".
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

