% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR_base.R
\name{fm.summary}
\alias{fm.summary}
\title{FlashR Summaries}
\usage{
fm.summary(x)
}
\arguments{
\item{x}{a FlashR vector or matrix.}
}
\value{
A list containing the following named components:
\itemize{
\item{min}{The minimum value}
\item{max}{The maximum value}
\item{mean}{The average}
\item{normL1}{The L1 norm}
\item{normL2}{The L2 norm}
\item{numNonzeros}{The number of non-zero values}
}
}
\description{
\code{fm.summary} produces summaries of a FlashR vector or matrix.
}
\details{
It computes the min, max, sum, mean, L1, L2, number of non-zero values if
the argument is a vector, or these statistics for each column if the argument
is a matrix.
}

