% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashmatrix.R
\name{fm.t}
\alias{fm.t}
\title{Transpose a FlashMatrixR matrix.}
\usage{
fm.t(m)
}
\arguments{
\item{m}{a FlashMatrixR matrix}
}
\value{
a FlashMatrixR matrix
}
\description{
Transpose a FlashMatrixR matrix.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

