% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlashR.R
\name{fm.t}
\alias{fm.t}
\title{Transpose a FlashMatrix matrix.}
\usage{
fm.t(m)
}
\arguments{
\item{m}{a FlashMatrix matrix}
}
\value{
a FlashMatrix matrix
}
\description{
Transpose a FlashMatrix matrix.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

