% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{fm.table}
\alias{as.data.frame,fm.table-method}
\alias{as.vector,fm.table-method}
\alias{fm.table}
\title{Count the number of elements}
\usage{
fm.table(x)

\S4method{as.vector}{fm.table}(x)

\S4method{as.data.frame}{fm.table}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{a FlashR vector or a "table" object for \code{as.vector} and
\code{as.data.frame}.}
}
\value{
a table object.
\itemize{
\item{val}{The unique values in the vector.}
\item{Freq}{The number of occurences of each unique value.}
}
}
\description{
\code{fm.table} counts the number of occurences of each unique value
in a FlashR vector.
}
\examples{
vec <- as.integer(fm.runif(1000, min=0, max=10))
tbl <- fm.table(vec)
cnts <- as.vector(tbl)
df <- as.data.frame(tbl)
}

