% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashmatrix.R
\name{fm.write.obj}
\alias{fm.write.obj}
\title{Write a FlashMatrix object (vector/matrix) to a file}
\usage{
fm.write.obj(fm, file)
}
\arguments{
\item{fm}{a FlashMatrix object.}

\item{file}{a file in the local filesystem.}
}
\value{
a logical value. True if the object is written to a file
successfully. Otherwise, FALSE.
}
\description{
Write a FlashMatrix object (vector/matrix) to a file
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

