% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{format.percent}
\alias{format.percent}
\title{Print method for percent vectors}
\source{
Generated by miniextendr from Rust fn \code{format_percent}
}
\usage{
\method{format}{percent}(x, ...)
}
\arguments{
\item{x}{(no documentation available)}

\item{dots}{(no documentation available)}
}
\description{
Print method for percent vectors
}
