% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{from_epath_to_expart_path_simple}
\alias{from_epath_to_expart_path_simple}
\title{convert vertex path to exon path}
\usage{
from_epath_to_expart_path_simple(g, epath)
}
\description{
convert vertex path to exon path
}
