% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{from_epath_to_expart_path_simple}
\alias{from_epath_to_expart_path_simple}
\title{convert edge path (by name) to exonic part path}
\usage{
from_epath_to_expart_path_simple(g, epath)
}
\arguments{
\item{g}{An \code{igraph} directed acyclic graph representing a gene splicing graph.}

\item{epath}{Character vector. Sequence of edge names (as set by \code{set_edge_names}) forming
an exon-level path through the splicing graph.}
}
\description{
convert edge path (by name) to exonic part path
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_edge_names(g)
epath <- c("R-3:R", "3-5:ex", "5-7:ex", "7-L:L")
expart_path <- grase::from_epath_to_expart_path_simple(g, epath)
}
}
