% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{from_exon_path_to_exonic_part_path}
\alias{from_exon_path_to_exonic_part_path}
\title{Convert exon path to exonic part path}
\usage{
from_exon_path_to_exonic_part_path(g, exonpath)
}
\description{
Convert exon path to exonic part path
}
