% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{from_vpath_to_exon_path}
\alias{from_vpath_to_exon_path}
\title{Convert vertex path to exon path}
\usage{
from_vpath_to_exon_path(g, bubblepath, node2pos)
}
\description{
Convert vertex path to exon path
}
