% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future14C.R
\docType{data}
\name{future14C}
\alias{future14C}
\title{Future atmospheric 14C dataset for delta delta calculation}
\format{dataframe}
\usage{
future14C
}
\description{
Data from: Sierra, C. "Forecasting atmospheric radiocarbon decline to pre-bomb values", Radiocarbon, Vol 60, Nr 4, 2018, p 1055.1066 DOI:10.1017/RDC.2018.33
}
\keyword{datasets}
