% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaudinski_2001.R
\docType{data}
\name{Gaudinski_2001}
\alias{Gaudinski_2001}
\title{Gaudinski Harvard Forest example dataset}
\format{list}
\usage{
Gaudinski_2001
}
\description{
Data from Gaudinski, J., 2001, Belowground carbon cycling in three temperate forests of the eastern United States, University of California Irvine, Ph.D. thesis
}
\keyword{datasets}
