% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_list_set_elt}
\alias{test_list_set_elt}
\alias{test_strvec_set_str}
\alias{test_list_builder_set}
\alias{test_list_set_elt_with}
\alias{test_strvec_builder_set}
\alias{test_list_builder_length}
\alias{test_reprotect_slot_count}
\alias{test_strvec_builder_length}
\alias{test_reprotect_slot_no_growth}
\alias{test_reprotect_slot_accumulate}
\alias{test_list_from_pairs_strings_gctorture}
\alias{test_list_from_values_strings_gctorture}
\title{gc protect tests}
\source{
Generated by miniextendr from Rust fn \code{test_list_set_elt}

Generated by miniextendr from Rust fn \code{test_strvec_set_str}

Generated by miniextendr from Rust fn \code{test_list_builder_set}

Generated by miniextendr from Rust fn \code{test_list_set_elt_with}

Generated by miniextendr from Rust fn \code{test_strvec_builder_set}

Generated by miniextendr from Rust fn \code{test_list_builder_length}

Generated by miniextendr from Rust fn \code{test_reprotect_slot_count}

Generated by miniextendr from Rust fn \code{test_strvec_builder_length}

Generated by miniextendr from Rust fn \code{test_reprotect_slot_no_growth}

Generated by miniextendr from Rust fn \code{test_reprotect_slot_accumulate}

Generated by miniextendr from Rust fn \code{test_list_from_pairs_strings_gctorture}

Generated by miniextendr from Rust fn \code{test_list_from_values_strings_gctorture}
}
\usage{
test_list_set_elt()

test_strvec_set_str()

test_list_builder_set()

test_list_set_elt_with()

test_strvec_builder_set()

test_list_builder_length(n)

test_reprotect_slot_count()

test_strvec_builder_length(n)

test_reprotect_slot_no_growth(iterations)

test_reprotect_slot_accumulate(n)

test_list_from_pairs_strings_gctorture()

test_list_from_values_strings_gctorture()
}
\arguments{
\item{n}{Final expected vector length.}

\item{iterations}{Number of set() calls to perform.}
}
\description{
gc protect tests

gc protect tests

gc protect tests

gc protect tests

gc protect tests

Test that ListBuilder reports the correct length

gc protect tests

Test that StrVecBuilder reports the correct length

Test that ReprotectSlot::set does not grow the protect stack over many iterations

Test ReprotectSlot by repeatedly replacing with larger vectors up to length n

gc protect tests

gc protect tests
}
