% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{into_sexp_altrep}
\alias{into_sexp_altrep}
\alias{gc_stress_vec_option_borrowed}
\alias{gc_stress_vec_option_collection}
\title{Convert an R vector to an ALTREP-backed vector by materializing then re-wrapping}
\source{
Generated by miniextendr from Rust fn \code{into_sexp_altrep}

Generated by miniextendr from Rust fn \code{gc_stress_vec_option_borrowed}

Generated by miniextendr from Rust fn \code{gc_stress_vec_option_collection}
}
\usage{
into_sexp_altrep(x)

gc_stress_vec_option_borrowed()

gc_stress_vec_option_collection()
}
\arguments{
\item{x}{An integer, numeric, or character vector to convert.}
}
\description{
Convert an R vector to an ALTREP-backed vector by materializing then re-wrapping

gc stress fixtures

gc stress fixtures
}
