% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_bubble_topo_interval}
\alias{get_bubble_topo_interval}
\title{Get topological interval for bubble}
\usage{
get_bubble_topo_interval(bubble, topo_idx)
}
\description{
Get topological interval for bubble
}
