% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl_get_sample_annotation.R
\name{get_sample_annotation}
\alias{get_sample_annotation}
\title{Retrieve the sample annotation}
\usage{
get_sample_annotation(organism = c("mouse", "fly", "fish"),
  dataset = c("provided", "generated"), for_kallisto = FALSE,
  kallisto_parent_dir = NULL)
}
\arguments{
\item{organism}{either "mouse", "fly", or "fish".}

\item{dataset}{either "provided" or "generated"}
}
\value{
a data.frame of sample annotations
}
\description{
Returns the sample annotations for a particular dataset.
}
