% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{get_sgnodetypes}
\alias{get_sgnodetypes}
\title{identifies nodes that are exon starts and ends and marks them with 1 and 2 respectively.
copied from SplicingGraphs}
\usage{
get_sgnodetypes(txpathmat, check.matrix = FALSE)
}
\description{
identifies nodes that are exon starts and ends and marks them with 1 and 2 respectively.
copied from SplicingGraphs
}
