% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl0learn-pkg.R
\docType{package}
\name{gL0Learn-package}
\alias{gL0Learn}
\alias{gL0Learn-package}
\title{gL0Learn: Efficient Algorithms for Graphical Learning Problems}
\description{
Efficient Algorithms for Graphical Learning Problems
}
\details{
gL0learn is a package ....
a C++ class under S4.
}
\author{
\strong{Maintainer}: Tim Nonet \email{tim.nonet@gmail.com}

Authors:
\itemize{
  \item Wenyu Chen \email{wenyu@mit.edu}
  \item Rahul Mazumder \email{rahulmaz@mit.edu}
}

}
