% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_constant.R
\name{gL0Learn.generate_constant}
\alias{gL0Learn.generate_constant}
\title{Generate a constantly correlated data set for gL0Learn}
\usage{
gL0Learn.generate_constant(n, p, rho, normalize, seed = 1)
}
\arguments{
\item{n}{See `gL0Learn.generate_synthetic` for details}

\item{p}{See `gL0Learn.generate_synthetic` for details}

\item{rho}{[TODO: Add documentation]}

\item{normalize}{See `gL0Learn.generate_synthetic` for details}

\item{seed}{See `gL0Learn.generate_synthetic` for details}
}
\description{
Computes the ...
}
