% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_independent.R
\name{gL0Learn.generate_independent}
\alias{gL0Learn.generate_independent}
\title{Generate a independent correlated data set for gL0Learn}
\usage{
gL0Learn.generate_independent(n, p, normalize, seed = 1)
}
\arguments{
\item{n}{The number of observations to generated.
This will create a data matrix of shape (n, p)}

\item{p}{The number of features to generated.
This will create a data matrix of shape (n, p).}

\item{normalize}{The method for normalizing data
Currently only "covariance" is supported}

\item{seed}{A seed to a random number generated}
}
\description{
Computes the ...
}
