% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_regression.R
\name{gL0Learn.generate_regression}
\alias{gL0Learn.generate_regression}
\title{Generate a regression data set for gL0Learn}
\usage{
gL0Learn.generate_regression(n, p, k, val, normalize, seed = 1)
}
\arguments{
\item{n}{See `gL0Learn.generate_synthetic` for details}

\item{p}{See `gL0Learn.generate_synthetic` for details}

\item{k}{[TODO: Add documentation]}

\item{val}{[TODO: Add documentation]}

\item{normalize}{See `gL0Learn.generate_synthetic` for details}

\item{seed}{See `gL0Learn.generate_synthetic` for details}
}
\description{
Computes the ...
}
