% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_synthetic.R
\name{gL0Learn.generate_synthetic}
\alias{gL0Learn.generate_synthetic}
\title{Generate a synthetic data set for gL0Learn}
\usage{
gL0Learn.generate_synthetic(n, p, model, normalize, seed = 1, ...)
}
\arguments{
\item{n}{The number of observations to generated.
This will create a data matrix of shape (n, p)}

\item{p}{The number of features to generated.
This will create a data matrix of shape (n, p)}

\item{model}{The method for generating datasets.
Currently only AR1 is supported}

\item{normalize}{The method for normailzing data
Currently only "covariance" is supported}

\item{seed}{A seed to a random number generated}
}
\description{
Computes the ...
}
