% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graven_2017.R
\docType{data}
\name{Graven_2017}
\alias{Graven_2017}
\title{Graven dataset for delta-delta calculation}
\format{
dataframe
}
\usage{
data(Graven_2017)
}
\description{
Atmospheric 14C and 13C data for the period 1850-2016.
}
\details{
Data from Graven et al., 2017. Annual timestep. Three zones for 14C: 1) northern hemisphere, "NHc14" > 30 deg N, 2) southern hemisphere, "SHc14" > 30 deg S, 3) tropics, "Tropicsc14" < 30 deg N & < 30 deg S. 13C data are the global annual averages.
}
\references{
Graven, H., Allison, C. E., Etheridge, D. M., Hammer, S., Keeling, R. F., Levin, I., Meijer, H. A. J., Rubino, M., Tans, P. P., Trudinger, C. M., Vaughn, B. H., and White, J. W. C.: Compiled records of carbon isotopes in atmospheric CO2 for historical simulations in CMIP6, Geosci. Model Dev., 10, 4405–4417, https://doi.org/10.5194/gmd-10-4405-2017, 2017.
}
\keyword{datasets}
