% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{GrowableVec}
\alias{GrowableVec}
\alias{GrowableVec$new}
\alias{GrowableVec$to_vec}
\alias{GrowableVec$clear}
\alias{GrowableVec$from_vec}
\alias{$.GrowableVec}
\alias{[[.GrowableVec}
\alias{GrowableVec$RExtend$extend_from_vec}
\alias{GrowableVec$RExtend$len}
\alias{GrowableVec$RExtend$is_empty}
\title{GrowableVec Class}
\usage{
GrowableVec

\method{$}{GrowableVec}(self, name)

\method{[[}{GrowableVec}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
GrowableVec inherent methods: constructors, conversion, and clearing.
}
