% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{GrowableVec}
\alias{GrowableVec}
\alias{GrowableVec$new}
\alias{GrowableVec$extend}
\alias{GrowableVec$len}
\alias{GrowableVec$is_empty}
\alias{GrowableVec$to_vec}
\alias{GrowableVec$clear}
\alias{GrowableVec$from_vec}
\alias{$.GrowableVec}
\alias{[[.GrowableVec}
\title{Growable vector for testing RExtend}
\format{
An object of class \code{environment} of length 7.
}
\source{
Generated by miniextendr from Rust type \code{GrowableVec}

Generated by miniextendr from \code{GrowableVec::new}

Generated by miniextendr from \code{GrowableVec::extend}

Generated by miniextendr from \code{GrowableVec::len}

Generated by miniextendr from \code{GrowableVec::is_empty}

Generated by miniextendr from \code{GrowableVec::to_vec}

Generated by miniextendr from \code{GrowableVec::clear}

Generated by miniextendr from \code{GrowableVec::from_vec}
}
\usage{
GrowableVec

\method{$}{GrowableVec}(self, name)

\method{[[}{GrowableVec}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\value{
An ExternalPtr to a GrowableVec.
}
\description{
A vector that can be extended from R, demonstrating RExtend trait.
}
\examples{
v <- GrowableVec$new()
v$extend(c(1L, 2L, 3L))
v$len()
}
\keyword{datasets}
