% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{GrowableVec}
\alias{GrowableVec}
\alias{GrowableVec$new}
\alias{GrowableVec$to_vec}
\alias{GrowableVec$clear}
\alias{GrowableVec$from_vec}
\alias{$.GrowableVec}
\alias{[[.GrowableVec}
\alias{GrowableVec$RExtend$extend_from_vec}
\alias{GrowableVec$RExtend$len}
\alias{GrowableVec$RExtend$is_empty}
\title{GrowableVec inherent methods: constructors, conversion, and clearing}
\format{
An object of class \code{environment} of length 5.
}
\source{
Generated by miniextendr from Rust type \code{GrowableVec}

Generated by miniextendr from \code{GrowableVec::new}

Generated by miniextendr from \code{GrowableVec::to_vec}

Generated by miniextendr from \code{GrowableVec::clear}

Generated by miniextendr from \code{GrowableVec::from_vec}
}
\usage{
GrowableVec

\method{$}{GrowableVec}(self, name)

\method{[[}{GrowableVec}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}

\item{data}{Integer vector of initial elements.}
}
\description{
GrowableVec inherent methods: constructors, conversion, and clearing
}
\keyword{datasets}
