% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{unsafe_C_test_sexp_equality}
\alias{unsafe_C_test_sexp_equality}
\title{Test SEXP pointer equality vs R_compute_identical semantic equality}
\source{
Generated by miniextendr from Rust fn `C_test_sexp_equality`
}
\usage{
unsafe_C_test_sexp_equality(x, y)
}
\arguments{
\item{x}{First SEXP to compare.}

\item{y}{Second SEXP to compare.}
}
\description{
Test SEXP pointer equality vs R_compute_identical semantic equality
}
