% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{impl_return_vec}
\alias{impl_return_vec}
\alias{impl_return_string}
\title{impl trait tests}
\source{
Generated by miniextendr from Rust fn \code{impl_return_vec}

Generated by miniextendr from Rust fn \code{impl_return_string}
}
\usage{
impl_return_vec()

impl_return_string()
}
\description{
impl trait tests

impl trait tests
}
