% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ImplDotsS3}
\alias{ImplDotsS3}
\alias{new_impldotss3}
\alias{impl_dots_s3_ctor_dots.ImplDotsS3}
\alias{impl_dots_s3_add_with_dots.ImplDotsS3}
\title{ImplDotsS3 S3 Class}
\source{
Generated by miniextendr from Rust type \code{ImplDotsS3}

Generated by miniextendr from \code{ImplDotsS3::impl_dots_s3_ctor_dots}

Generated by miniextendr from \code{ImplDotsS3::impl_dots_s3_ctor_dots}

Generated by miniextendr from \code{ImplDotsS3::impl_dots_s3_add_with_dots}

Generated by miniextendr from \code{ImplDotsS3::impl_dots_s3_add_with_dots}
}
\usage{
new_impldotss3(seed, ...)

\method{impl_dots_s3_ctor_dots}{ImplDotsS3}(x, ...)

\method{impl_dots_s3_add_with_dots}{ImplDotsS3}(x, value, ...)

ImplDotsS3
}
\arguments{
\item{seed}{Integer base value.}

\item{...}{Additional arguments.}

\item{x}{An object.}

\item{value}{Integer value to add.}
}
\description{
S3 class exercising dots in both the constructor and an instance method.

Create an S3 dots fixture.

S3 generic for \code{impl_dots_s3_ctor_dots}

Return the constructor dots count.

S3 generic for \code{impl_dots_s3_add_with_dots}

Add a value and the number of method dots.
}
