% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{indexmap_len}
\alias{indexmap_len}
\alias{indexmap_keys}
\alias{indexmap_empty}
\alias{indexmap_single}
\alias{indexmap_duplicate_key}
\alias{indexmap_roundtrip_dbl}
\alias{indexmap_roundtrip_int}
\alias{indexmap_roundtrip_str}
\alias{indexmap_order_preserved}
\title{Test getting the number of entries in an IndexMap}
\source{
Generated by miniextendr from Rust fn `indexmap_len`

Generated by miniextendr from Rust fn `indexmap_keys`

Generated by miniextendr from Rust fn `indexmap_empty`

Generated by miniextendr from Rust fn `indexmap_single`

Generated by miniextendr from Rust fn `indexmap_duplicate_key`

Generated by miniextendr from Rust fn `indexmap_roundtrip_dbl`

Generated by miniextendr from Rust fn `indexmap_roundtrip_int`

Generated by miniextendr from Rust fn `indexmap_roundtrip_str`

Generated by miniextendr from Rust fn `indexmap_order_preserved`
}
\usage{
indexmap_len(map)

indexmap_keys(map)

indexmap_empty()

indexmap_single()

indexmap_duplicate_key()

indexmap_roundtrip_dbl(map)

indexmap_roundtrip_int(map)

indexmap_roundtrip_str(map)

indexmap_order_preserved()
}
\arguments{
\item{map}{Named character list from R.}
}
\description{
Test getting the number of entries in an IndexMap

Test extracting keys from an IndexMap in insertion order

indexmap adapter tests

indexmap adapter tests

indexmap adapter tests

Test IndexMap<String, f64> roundtrip through R named list

Test IndexMap<String, i32> roundtrip through R named list

Test IndexMap<String, String> roundtrip through R named list

indexmap adapter tests
}
