% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartition_analysis.R
\name{infer_diff_exons_from_tx_groups}
\alias{infer_diff_exons_from_tx_groups}
\title{Infer differential exonic parts given two groups of transcripts (e.g. changed vs constant)}
\usage{
infer_diff_exons_from_tx_groups(gene, g, tx_changed, tx_constant, outdir)
}
\arguments{
\item{gene}{Character string. Gene identifier.}

\item{g}{An \code{igraph} directed acyclic graph representing a gene splicing graph.}

\item{tx_changed}{Character vector. Transcript IDs in the "changed" group.}

\item{tx_constant}{Character vector. Transcript IDs in the "constant" group.}

\item{outdir}{Character string. Path to output directory.}
}
\description{
Infer differential exonic parts given two groups of transcripts (e.g. changed vs constant)
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
result <- grase::infer_diff_exons_from_tx_groups(
  gene       = "ENSG00000000003.14",
  g          = g,
  tx_changed  = c("ENST00000373020.8"),
  tx_constant = c("ENST00000494424.1", "ENST00000379263.3"),
  outdir     = tempdir()
)
}
}
