% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{initialize_datasource}
\alias{initialize_datasource}
\title{Initialize Datasource}
\usage{
initialize_datasource()
}
\description{
This function initializes the Redivis datasource connection if it is not already set.
}
