% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{initialize_datasource}
\alias{initialize_datasource}
\title{Initialize Datasource}
\usage{
initialize_datasource()
}
\value{
The initialized datasource connection.
}
\description{
This function initializes the Redivis datasource connection if it is not already set.
If the global variable \code{datasource} doesn't exist or is NULL, it sets up a new connection.
}
