% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{unsafe_C_check_interupt_after}
\alias{unsafe_C_check_interupt_after}
\alias{unsafe_C_check_interupt_unwind}
\title{interrupt tests}
\source{
Generated by miniextendr from Rust fn \code{C_check_interupt_after}

Generated by miniextendr from Rust fn \code{C_check_interupt_unwind}
}
\usage{
unsafe_C_check_interupt_after()

unsafe_C_check_interupt_unwind()
}
\description{
interrupt tests

interrupt tests
}
