% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{into_r_as_i64_to_i32}
\alias{into_r_as_i64_to_i32}
\title{Convert a \verb{Vec<i64>} to R integer vector via \verb{IntoRAs<i32>}}
\source{
Generated by miniextendr from Rust fn \code{into_r_as_i64_to_i32}
}
\usage{
into_r_as_i64_to_i32(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
Convert a \verb{Vec<i64>} to R integer vector via \verb{IntoRAs<i32>}
}
