% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{into_r_error_inner}
\alias{into_r_error_inner}
\alias{into_r_error_length_overflow}
\alias{into_r_error_string_too_long}
\title{into_r_error_inner}
\source{
Generated by miniextendr from Rust fn \code{into_r_error_inner}

Generated by miniextendr from Rust fn \code{into_r_error_length_overflow}

Generated by miniextendr from Rust fn \code{into_r_error_string_too_long}
}
\usage{
into_r_error_inner()

into_r_error_length_overflow()

into_r_error_string_too_long()
}
\description{
Trigger an Inner error and return its message.

Trigger a LengthOverflow error and return its message.

Trigger a StringTooLong error and return its message.
}
