% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{into_r_error_inner}
\alias{into_r_error_inner}
\alias{into_r_error_length_overflow}
\alias{into_r_error_string_too_long}
\title{into r error tests}
\source{
Generated by miniextendr from Rust fn \code{into_r_error_inner}

Generated by miniextendr from Rust fn \code{into_r_error_length_overflow}

Generated by miniextendr from Rust fn \code{into_r_error_string_too_long}
}
\usage{
into_r_error_inner()

into_r_error_length_overflow()

into_r_error_string_too_long()
}
\description{
into r error tests

into r error tests

into r error tests
}
