% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{IntSet}
\alias{IntSet}
\alias{IntSet$contains}
\alias{$.IntSet}
\alias{[[.IntSet}
\alias{IntSet$RToVec$to_vec}
\alias{IntSet$RToVec$len}
\alias{IntSet$RToVec$is_empty}
\alias{IntSet$RFromIter$from_vec}
\title{IntSet inherent methods: membership test}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type \code{IntSet}

Generated by miniextendr from \code{IntSet::contains}
}
\usage{
IntSet

\method{$}{IntSet}(self, name)

\method{[[}{IntSet}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}

\item{value}{Integer value to check for membership.}
}
\description{
IntSet inherent methods: membership test
}
\keyword{datasets}
