% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{IntSet}
\alias{IntSet}
\alias{IntSet$contains}
\alias{$.IntSet}
\alias{[[.IntSet}
\alias{IntSet$RToVec$to_vec}
\alias{IntSet$RToVec$len}
\alias{IntSet$RToVec$is_empty}
\alias{IntSet$RFromIter$from_vec}
\title{IntSet  Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type `IntSet`

Generated by miniextendr from `IntSet::contains`
}
\usage{
IntSet

\method{$}{IntSet}(self, name)

\method{[[}{IntSet}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
IntSet  Class

\describe{
  \item{\code{value}}{Integer value to check for membership.}
}
}
\keyword{datasets}
