% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{IntVecIter}
\alias{IntVecIter}
\alias{IntVecIter$new}
\alias{IntVecIter$next_item}
\alias{IntVecIter$size_hint}
\alias{IntVecIter$count}
\alias{IntVecIter$collect_n}
\alias{IntVecIter$skip}
\alias{IntVecIter$nth}
\alias{$.IntVecIter}
\alias{[[.IntVecIter}
\title{Iterator wrapper for testing RIterator}
\format{
An object of class \code{environment} of length 7.
}
\source{
Generated by miniextendr from Rust type \code{IntVecIter}

Generated by miniextendr from \code{IntVecIter::new}

Generated by miniextendr from \code{IntVecIter::next_item}

Generated by miniextendr from \code{IntVecIter::size_hint}

Generated by miniextendr from \code{IntVecIter::count}

Generated by miniextendr from \code{IntVecIter::collect_n}

Generated by miniextendr from \code{IntVecIter::skip}

Generated by miniextendr from \code{IntVecIter::nth}
}
\usage{
IntVecIter

\method{$}{IntVecIter}(self, name)

\method{[[}{IntVecIter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\value{
An ExternalPtr to an IntVecIter.
}
\description{
An iterator over integers demonstrating RIterator trait.
}
\examples{
it <- IntVecIter$new(c(1L, 2L, 3L))
it$next_item()
it$size_hint()
}
\keyword{datasets}
