% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{IntVecIter}
\alias{IntVecIter}
\alias{IntVecIter$new}
\alias{$.IntVecIter}
\alias{[[.IntVecIter}
\alias{IntVecIter$RIterator$next_item}
\alias{IntVecIter$RIterator$count}
\alias{IntVecIter$RIterator$collect_n}
\alias{IntVecIter$RIterator$skip}
\alias{IntVecIter$RIterator$nth}
\title{IntVecIter  Class}
\format{
An object of class \code{environment} of length 2.
}
\source{
Generated by miniextendr from Rust type \code{IntVecIter}

Generated by miniextendr from \code{IntVecIter::new}
}
\usage{
IntVecIter

\method{$}{IntVecIter}(self, name)

\method{[[}{IntVecIter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
IntVecIter  Class

\describe{
\item{\code{data}}{Integer vector of elements to iterate over.}
}
}
\keyword{datasets}
