% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite.r
\name{irw_citation}
\alias{irw_citation}
\title{Generate Citation and BibTeX for a Table}
\usage{
irw_citation(table_name)
}
\arguments{
\item{table_name}{A character string specifying the name of the table.}
}
\value{
A single character string containing the citation and BibTeX entry for the table.
}
\description{
Generates a well-formatted string containing both the citation and BibTeX entry for a table.
}
