% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{irw_db_metadata}
\alias{irw_db_metadata}
\title{Retrieve and Print IRW Database Metadata}
\usage{
irw_db_metadata()
}
\description{
Fetches and displays comprehensive metadata for the IRW database, including the database version,
number of tables, total data size, and relevant URLs.
}
\examples{
\dontrun{
  irw_db_metadata()
}
}
