% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{irw_download}
\alias{irw_download}
\title{Download a Dataset}
\usage{
irw_download(
  table_name,
  path = NULL,
  overwrite = FALSE,
  sim = FALSE,
  comp = FALSE
)
}
\arguments{
\item{table_name}{A character string specifying the name of the Redivis table
to download.}

\item{path}{A string specifying the file path where the data should be saved.
If \code{NULL}, the dataset will be saved in the current working directory
with the table's name as the file name.}

\item{overwrite}{Logical. Whether to overwrite an existing file.}

\item{sim}{Logical. If TRUE, download from the simulation dataset.}

\item{comp}{Logical. If TRUE, download from the competition dataset.}
}
\value{
A message confirming the file download location.
}
\description{
Downloads a specified dataset table from the IRW datasource.
}
