% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{irw_fetch}
\alias{irw_fetch}
\title{Fetch Tables(s) from the Item Response Warehouse}
\usage{
irw_fetch(name, sim = FALSE, dedup = FALSE)
}
\arguments{
\item{name}{Character vector of one or more table names (IRW table IDs).}

\item{sim}{Logical, optional. If TRUE, fetches from the IRW simulation table. Defaults to FALSE.}

\item{dedup}{Logical, optional. If TRUE, deduplicates responses based on timing variables. Defaults to FALSE.
\itemize{
\item If a 'date' column is present, no deduplication is performed.
\item If only a 'wave' column is present, one random response is retained per (id, item) pair within each wave.
\item If neither 'date' nor 'wave' is present, one random response is retained per (id, item) pair.
}}
}
\value{
If a single name is provided, returns a tibble. If multiple, returns a named list
of tibbles (or error messages, if retrieval failed).
}
\description{
Retrieves one or more tables from IRW and returns them as tibbles.
If the table includes a character \code{resp} column, the function attempts to
coerce it into numeric. Strings like \code{"NA"}, \code{""}, and \code{NA} are treated as missing values.
A warning is issued only if other non-numeric values are encountered.
}
\examples{
\dontrun{
irw_fetch("environment_ltm")
irw_fetch("gilbert_meta_3", sim = TRUE)
irw_fetch("pks_probability", dedup = TRUE)
}
}
