% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{irw_info}
\alias{irw_info}
\title{Retrieve and Print IRW Database or Table Information}
\usage{
irw_info(table_name = NULL)
}
\arguments{
\item{table_name}{Optional. A character string specifying the name of the table to retrieve information for.}
}
\description{
Fetches and displays information for the IRW database or a specified table.
If no table name is provided, it returns general database information.
If a table name is provided, it returns detailed information about that table.
}
\examples{
\dontrun{
irw_info() # Prints database information
irw_info("abortion") # Prints table-specific information
}
}
