% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{irw_license_options}
\alias{irw_license_options}
\title{View Unique License Options with Frequencies}
\usage{
irw_license_options()
}
\value{
A data.frame with 'license' and 'count' columns.
}
\description{
Returns a data frame showing the number of datasets associated with each license.
}
