% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemtext.R
\name{irw_list_itemtext_tables}
\alias{irw_list_itemtext_tables}
\title{List IRW tables with available item text metadata}
\usage{
irw_list_itemtext_tables()
}
\value{
A sorted character vector of IRW table names with item text metadata.
}
\description{
Returns a character vector of IRW table names for which item text is available.
}
\examples{
\dontrun{
  irw_list_itemtext_tables()
}

}
