% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{irw_list_tables}
\alias{irw_list_tables}
\title{List Available Tables}
\usage{
irw_list_tables(sim = FALSE, comp = FALSE, nom = FALSE)
}
\arguments{
\item{sim}{Logical. If TRUE, lists tables from the IRW simulation dataset (\code{irw_simsyn}).}

\item{comp}{Logical. If TRUE, lists tables from the IRW competition dataset (\code{irw_comp}).}

\item{nom}{Logical. If TRUE, lists tables from the IRW nominal dataset (\code{irw_nominal}).
If all of \code{sim}, \code{comp}, and \code{nom} are FALSE (default), lists tables from the main IRW database.
Only one of \code{sim}, \code{comp}, or \code{nom} should be TRUE at a time.}
}
\value{
A data frame with the following columns:
\item{name}{The name of the table, sorted alphabetically.}
\item{numRows}{The number of rows in the table.}
\item{variableCount}{The number of variables in the table.}
}
\description{
Retrieves a summary of available tables in the IRW database, including their name,
number of rows, and variable count, sorted alphabetically by table name.
}
\examples{
\dontrun{
irw_list_tables()             # Main IRW database
irw_list_tables(sim = TRUE)   # Simulation dataset
irw_list_tables(comp = TRUE)  # Competition dataset
}
}
