% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{irw_list_tables}
\alias{irw_list_tables}
\title{List Available Tables}
\usage{
irw_list_tables(sim = FALSE)
}
\arguments{
\item{sim}{Logical. If TRUE, lists tables from the IRW simulation dataset (\code{irw_simsyn}).}
}
\value{
A data frame with the following columns:
\item{name}{The name of the table, sorted alphabetically.}
\item{numRows}{The number of rows in the table.}
\item{variableCount}{The number of variables in the table.}
}
\description{
Retrieves a summary of available tables in the IRW database, including their name,
number of rows, and variable count, sorted alphabetically by table name.
}
\examples{
\dontrun{
irw_list_tables()            # IRW database
irw_list_tables(sim = TRUE) # Simulated datasets
}
}
