% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-filter.R
\name{irw_metadata}
\alias{irw_metadata}
\title{Retrieve IRW Metadata Table}
\usage{
irw_metadata()
}
\value{
A tibble containing metadata information.
}
\description{
Fetches the metadata table from Redivis and returns it as a tibble.
Automatically checks for updates and refreshes only when needed.
}
