% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{is_bubble}
\alias{is_bubble}
\title{copied from SplicingGraphs}
\usage{
is_bubble(txpathmat, i, j)
}
\description{
copied from SplicingGraphs
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_txpath_to_vertex_attr(g)
txmat <- grase::make_matrix_from_txpath_igraph(g)
grase::is_bubble(txmat, 2, 5)
}
}
