% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isCumulative.R
\name{isCumulative}
\alias{isCumulative}
\title{Test whether a time series of CO2 release is cumulative}
\usage{
isCumulative(x)
}
\arguments{
\item{x}{A data.frame with only two columns, first column for time and second column for CO2 release. Test is performed on the second column.}
}
\value{
logical
}
\description{
Test whether a time series of CO2 release is cumulative
}
\examples{
series=data.frame(time=seq(1,10), CO2=cumsum(rep(0.1, 10)))
isCumulative(x=series)
}
