% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.fill_coords.R
\name{ISRaD.extra.fill_coords}
\alias{ISRaD.extra.fill_coords}
\title{ISRaD.extra.fill_coords}
\usage{
ISRaD.extra.fill_coords(database)
}
\arguments{
\item{database}{ISRaD dataset object.}
}
\value{
ISRaD_data object with filled profile coordinates.
}
\description{
Fills profile coordinates from site coordinates if profile coordinates not reported.
}
\examples{
# Load example dataset Gaudinski_2001
database <- ISRaD::Gaudinski_2001
# Fill profile coordinates
database.x <- ISRaD.extra.fill_coords(database)
}
\author{
J. Beem-Miller
}
