% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.norm14c_year.R
\name{ISRaD.extra.norm14c_year}
\alias{ISRaD.extra.norm14c_year}
\title{ISRaD.extra.norm14c_year}
\usage{
ISRaD.extra.norm14c_year(
  df,
  obs_d14c,
  obs_year,
  atm_zone,
  norm_year,
  verbose = FALSE
)
}
\arguments{
\item{df}{data frame with columns for observed d14c (obs_d14c), observation year (obs_year), and atmospheric zone (atm_zone)}

\item{obs_d14c}{observed delta 14c value to be normalized}

\item{obs_year}{year in which obs_d14c was observed (sample collection year)}

\item{atm_zone}{atmospheric zone for obs_d14c. Must be one of c("NHc14", "SHc14", "Tropicsc14"). "NHc14" = > 30 degrees latitude; "SHc14" = < -30 latitude; "Tropicsc14" = < 30 & > -30 degrees latitude.}

\item{verbose}{Show progress bar? TRUE/FALSE (default = FALSE)}
}
\value{
data frame with new column of normalized 14c values.
}
\description{
Normalizes delta 14c values to a given year (norm_year)
}
\details{
Creates new column for normalized d14c in provided data frame. This function works by creating a one pool steady-state model using atmospheric 14c over the period 1850 to 2021. Turnover time is determined by fitting the model to the observed delta 14c (obs_d14c) in the observation year (obs_year), and the normalized 14c value is calculated by runnning the model forwards or backwards to the desired normalization year (norm_year). Note: there is no guarantee that the normalized 14c value will be meaningful as the model assumes a well-mixed homogenous system which is rarely the case in soils. See the example for information on how to run the function when the "df" argument corresponds to a table from an ISRaD object, e.g. "flux", "layer", etc. Can be very slow for large datasets!
}
\examples{
# Load example dataset Gaudinski_2001
database <- ISRaD::Gaudinski_2001
# Fill profile coordinates
database.x <- ISRaD.extra.fill_coords(database)
# Fill dates
database.x <- ISRaD.extra.fill_dates(database.x)
# Fill delta 14C from fraction modern
database.x <- ISRaD.extra.fill_rc(database.x)
# Fill atmospheric 14c
database.x <- ISRaD.extra.calc_atm14c(database.x)
# Run normalization function for the year 2010 with layer data
database.x$layer <- ISRaD.extra.norm14c_year(df = database.x$layer, obs_d14c = "lyr_14c", obs_year = "lyr_obs_date_y", atm_zone = "pro_graven_zone", norm_year = 2010, verbose = TRUE)
}
\author{
J. Beem-Miller and J. Randerson
}
