% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.rep.count.all.R
\name{ISRaD.rep.count.all}
\alias{ISRaD.rep.count.all}
\title{ISRaD.rep.count.all}
\usage{
ISRaD.rep.count.all(database)
}
\arguments{
\item{database}{ISRaD data object}
}
\value{
A tibble of observation counts, one column for each database table.
}
\description{
Generates a report of counts of observations at each level of the database
}
\examples{
# Load example dataset Gaudinski_2001
database <- ISRaD::Gaudinski_2001
ISRaD.rep.count.all(database)
}
