% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{IterableVec}
\alias{IterableVec}
\alias{IterableVec$new}
\alias{IterableVec$len}
\alias{IterableVec$to_vec}
\alias{IterableVec$make_iter}
\alias{$.IterableVec}
\alias{[[.IterableVec}
\title{Iterable vector for testing RMakeIter}
\format{
An object of class \code{environment} of length 4.
}
\source{
Generated by miniextendr from Rust type \code{IterableVec}

Generated by miniextendr from \code{IterableVec::new}

Generated by miniextendr from \code{IterableVec::len}

Generated by miniextendr from \code{IterableVec::to_vec}

Generated by miniextendr from \code{IterableVec::make_iter}
}
\usage{
IterableVec

\method{$}{IterableVec}(self, name)

\method{[[}{IterableVec}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\value{
An ExternalPtr to an IterableVec.
}
\description{
A vector that can create independent iterators.
}
\examples{
v <- IterableVec$new(c(1L, 2L, 3L))
it <- v$make_iter()
it$`next`()
}
\keyword{datasets}
