% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{IterableVecIter}
\alias{IterableVecIter}
\alias{IterableVecIter$collect_all}
\alias{$.IterableVecIter}
\alias{[[.IterableVecIter}
\alias{IterableVecIter$RIterator$next_item}
\alias{IterableVecIter$RIterator$count}
\alias{IterableVecIter$RIterator$collect_n}
\alias{IterableVecIter$RIterator$skip}
\alias{IterableVecIter$RIterator$nth}
\title{IterableVecIter  Class}
\format{
An object of class \code{environment} of length 2.
}
\source{
Generated by miniextendr from Rust type \code{IterableVecIter}

Generated by miniextendr from \code{IterableVecIter::collect_all}
}
\usage{
IterableVecIter

\method{$}{IterableVecIter}(self, name)

\method{[[}{IterableVecIter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
IterableVecIter  Class
}
\keyword{datasets}
