% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laboratoryTreatment.R
\name{laboratoryTreatment}
\alias{laboratoryTreatment}
\title{Creates an aggregated vector with a specific laboratory treatment}
\usage{
laboratoryTreatment(database, treatment)
}
\arguments{
\item{database}{A list with the sidb}

\item{treatment}{A character vector with the treatment to be extracted}
}
\value{
A vector with all values of the treatment from the database
}
\description{
Creates an aggregated vector with a specific laboratory treatment
}
\examples{
sidb=loadEntries(path="~/sidb/data/")
temperature=laboratoryTreatment(database=sidb, treatment="temperature")
}
