% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_lazy_vec_f64}
\alias{test_lazy_vec_f64}
\alias{test_lazy_vec_i32}
\alias{test_lazy_arrow_f64}
\alias{test_lazy_arrow_i32}
\alias{test_lazy_arrow_bool}
\alias{test_lazy_ndarray_f64}
\alias{test_lazy_ndarray_i32}
\alias{test_lazy_arrow_string}
\alias{test_lazy_nalgebra_f64}
\alias{test_lazy_nalgebra_i32}
\alias{test_lazy_arrow_f64_with_nulls}
\alias{test_lazy_arrow_string_with_nulls}
\title{test lazy vec f64}
\source{
Generated by miniextendr from Rust fn `test_lazy_vec_f64`

Generated by miniextendr from Rust fn `test_lazy_vec_i32`

Generated by miniextendr from Rust fn `test_lazy_arrow_f64`

Generated by miniextendr from Rust fn `test_lazy_arrow_i32`

Generated by miniextendr from Rust fn `test_lazy_arrow_bool`

Generated by miniextendr from Rust fn `test_lazy_ndarray_f64`

Generated by miniextendr from Rust fn `test_lazy_ndarray_i32`

Generated by miniextendr from Rust fn `test_lazy_arrow_string`

Generated by miniextendr from Rust fn `test_lazy_nalgebra_f64`

Generated by miniextendr from Rust fn `test_lazy_nalgebra_i32`

Generated by miniextendr from Rust fn `test_lazy_arrow_f64_with_nulls`

Generated by miniextendr from Rust fn `test_lazy_arrow_string_with_nulls`
}
\usage{
test_lazy_vec_f64(n)

test_lazy_vec_i32(n)

test_lazy_arrow_f64(n)

test_lazy_arrow_i32(n)

test_lazy_arrow_bool(n)

test_lazy_ndarray_f64(n)

test_lazy_ndarray_i32(n)

test_lazy_arrow_string()

test_lazy_nalgebra_f64(n)

test_lazy_nalgebra_i32(n)

test_lazy_arrow_f64_with_nulls()

test_lazy_arrow_string_with_nulls()
}
\arguments{
\item{n}{(no documentation available)}
}
\description{
test lazy vec f64

test lazy vec i32

test lazy arrow f64

test lazy arrow i32

test lazy arrow bool

test lazy ndarray f64

test lazy ndarray i32

test lazy arrow string

test lazy nalgebra f64

test lazy nalgebra i32

test lazy arrow f64 with nulls

test lazy arrow string with nulls
}
