% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{defunct_fn}
\alias{defunct_fn}
\alias{superseded_fn}
\alias{also_deprecated}
\alias{fully_deprecated}
\alias{old_deprecated_fn}
\alias{soft_deprecated_fn}
\alias{experimental_feature}
\title{A defunct function that should error}
\source{
Generated by miniextendr from Rust fn \code{defunct_fn}

Generated by miniextendr from Rust fn \code{superseded_fn}

Generated by miniextendr from Rust fn \code{also_deprecated}

Generated by miniextendr from Rust fn \code{fully_deprecated}

Generated by miniextendr from Rust fn \code{old_deprecated_fn}

Generated by miniextendr from Rust fn \code{soft_deprecated_fn}

Generated by miniextendr from Rust fn \code{experimental_feature}
}
\usage{
defunct_fn(x)

superseded_fn(x)

also_deprecated()

fully_deprecated(x)

old_deprecated_fn(x)

soft_deprecated_fn(x)

experimental_feature(x)
}
\arguments{
\item{x}{Numeric input value.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
