% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{defunct_fn}
\alias{defunct_fn}
\alias{superseded_fn}
\alias{also_deprecated}
\alias{fully_deprecated}
\alias{old_deprecated_fn}
\alias{soft_deprecated_fn}
\alias{experimental_feature}
\title{A defunct function that should error}
\source{
Generated by miniextendr from Rust fn `defunct_fn`

Generated by miniextendr from Rust fn `superseded_fn`

Generated by miniextendr from Rust fn `also_deprecated`

Generated by miniextendr from Rust fn `fully_deprecated`

Generated by miniextendr from Rust fn `old_deprecated_fn`

Generated by miniextendr from Rust fn `soft_deprecated_fn`

Generated by miniextendr from Rust fn `experimental_feature`
}
\usage{
defunct_fn(x)

superseded_fn(x)

also_deprecated()

fully_deprecated(x)

old_deprecated_fn(x)

soft_deprecated_fn(x)

experimental_feature(x)
}
\arguments{
\item{x}{Numeric input value.}
}
\description{
`r lifecycle::badge("deprecated")`

`r lifecycle::badge("superseded")`

`r lifecycle::badge("deprecated")`

`r lifecycle::badge("deprecated")`

`r lifecycle::badge("deprecated")`

`r lifecycle::badge("deprecated")`

`r lifecycle::badge("experimental")`
}
\keyword{internal}
