% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{old_deprecated_fn}
\alias{old_deprecated_fn}
\alias{also_deprecated}
\alias{experimental_feature}
\alias{superseded_fn}
\alias{soft_deprecated_fn}
\alias{fully_deprecated}
\alias{defunct_fn}
\title{old_deprecated_fn}
\source{
Generated by miniextendr from Rust fn \code{old_deprecated_fn}

Generated by miniextendr from Rust fn \code{also_deprecated}

Generated by miniextendr from Rust fn \code{experimental_feature}

Generated by miniextendr from Rust fn \code{superseded_fn}

Generated by miniextendr from Rust fn \code{soft_deprecated_fn}

Generated by miniextendr from Rust fn \code{fully_deprecated}

Generated by miniextendr from Rust fn \code{defunct_fn}
}
\usage{
old_deprecated_fn(x)

also_deprecated()

experimental_feature(x)

superseded_fn(x)

soft_deprecated_fn(x)

fully_deprecated(x)

defunct_fn(x)
}
\arguments{
\item{x}{Integer input value (unused, function is defunct).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A function marked with Rust's \verb{#[deprecated]} attribute. This should generate lifecycle::deprecate_warn in the R wrapper.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A deprecated function without replacement info.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} An experimental function that may change.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}} A superseded function with a better alternative.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A soft-deprecated function (warning on first use only).

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A deprecated function with full lifecycle spec.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A defunct function that should error.
}
\keyword{internal}
