% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{LifecycleDemo}
\alias{LifecycleDemo}
\title{R6 class demonstrating lifecycle attributes on methods}
\source{
Generated by miniextendr from Rust type `LifecycleDemo`
}
\description{
R6 class demonstrating lifecycle attributes on methods

R6 class demonstrating lifecycle attributes on methods
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LifecycleDemo-new}{\code{LifecycleDemo$new()}}
\item \href{#method-LifecycleDemo-get_value}{\code{LifecycleDemo$get_value()}}
\item \href{#method-LifecycleDemo-old_value}{\code{LifecycleDemo$old_value()}}
\item \href{#method-LifecycleDemo-experimental_method}{\code{LifecycleDemo$experimental_method()}}
\item \href{#method-LifecycleDemo-legacy_value}{\code{LifecycleDemo$legacy_value()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LifecycleDemo-new"></a>}}
\if{latex}{\out{\hypertarget{method-LifecycleDemo-new}{}}}
\subsection{Method \code{new()}}{
Creates a new LifecycleDemo.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LifecycleDemo$new(value, .ptr = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{(no documentation available)}

\item{\code{.ptr}}{Internal pointer (used by static methods, not for direct use).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LifecycleDemo-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-LifecycleDemo-get_value}{}}}
\subsection{Method \code{get_value()}}{
Returns the value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LifecycleDemo$get_value()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LifecycleDemo-old_value"></a>}}
\if{latex}{\out{\hypertarget{method-LifecycleDemo-old_value}{}}}
\subsection{Method \code{old_value()}}{
`r lifecycle::badge("deprecated")` A deprecated method using lifecycle attribute.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LifecycleDemo$old_value()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LifecycleDemo-experimental_method"></a>}}
\if{latex}{\out{\hypertarget{method-LifecycleDemo-experimental_method}{}}}
\subsection{Method \code{experimental_method()}}{
`r lifecycle::badge("experimental")` An experimental method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LifecycleDemo$experimental_method()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LifecycleDemo-legacy_value"></a>}}
\if{latex}{\out{\hypertarget{method-LifecycleDemo-legacy_value}{}}}
\subsection{Method \code{legacy_value()}}{
`r lifecycle::badge("deprecated")` A method marked with Rust's `#[deprecated]` attribute.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LifecycleDemo$legacy_value()}\if{html}{\out{</div>}}
}

}
}
