% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadEntries.R
\name{loadEntries}
\alias{loadEntries}
\title{Load entries of the sidb}
\usage{
loadEntries(path = "~/sidb/data/")
}
\arguments{
\item{path}{character string with the path where isdb is stored}
}
\value{
R list with all entries
}
\description{
Load entries of the sidb
}
\examples{
sidb=loadEntries()
}
