% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadEntries.R
\name{loadEntries}
\alias{loadEntries}
\title{Load SIDb entries}
\usage{
loadEntries(path)
}
\arguments{
\item{path}{character string with the path where SIDb entries are stored}
}
\value{
R list with all entries with the SIDb structure
}
\description{
Load SIDb entries
}
\examples{
# Two entries are included as examples in the sidb R package
# The path should be changed to a local directory where you store sidb entries
sidbEx <- loadEntries(path = paste0(system.file("extdata", "sidbEntries",
 package = "sidb"), "/"))
}
