% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadEntries.R
\name{loadEntries}
\alias{loadEntries}
\title{Load SIDb entries}
\usage{
loadEntries(path)
}
\arguments{
\item{path}{character string with the path where SIDb entries are stored}
}
\value{
R list with all entries with the SIDb structure
}
\description{
Load SIDb entries
}
\examples{
\dontrun{
sidb=loadEntries(path="~/sidb/data/")  # Assuming you downloaded
                    # the sidb repository in your home directory
}
}
