% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{make_matrix_from_txpath_igraph}
\alias{make_matrix_from_txpath_igraph}
\title{Build transcript-path logical matrix from index lists}
\usage{
make_matrix_from_txpath_igraph(g, txpath_vertex_list)
}
\description{
Build transcript-path logical matrix from index lists
}
