% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMATS.R
\name{map_rMATS}
\alias{map_rMATS}
\title{map rMATS events to grase graph}
\usage{
map_rMATS(
  g,
  gene,
  gff,
  fromGTF_A3SS,
  fromGTF_A5SS,
  fromGTF_SE,
  fromGTF_RI,
  bipartitions,
  grase_output_dir
)
}
\description{
map rMATS events to grase graph
}
