% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMATS.R
\name{map_rMATS}
\alias{map_rMATS}
\title{Map rMATS events to graph}
\usage{
map_rMATS(
  g,
  gene,
  gff,
  fromGTF_A3SS,
  fromGTF_A5SS,
  fromGTF_SE,
  fromGTF_RI,
  splits,
  analysis_type = NULL
)
}
\arguments{
\item{g}{Graph object}

\item{gene}{Gene identifier}

\item{gff}{GFF file path}

\item{fromGTF_A3SS}{Path to rMATS A3SS fromGTF file}

\item{fromGTF_A5SS}{Path to rMATS A5SS fromGTF file}

\item{fromGTF_SE}{Path to rMATS SE fromGTF file}

\item{fromGTF_RI}{Path to rMATS RI fromGTF file}

\item{splits}{Bipartition or multinomial splits data}

\item{analysis_type}{Optional string specifying analysis type ("bipartition", "multinomial", "n_choose_2").}
}
\value{
Modified graph object with rMATS event annotations
}
\description{
Map rMATS events to graph
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
splits <- read.table("ENSG00000000003.bipartition.txt", header = TRUE, sep = "\t")
fromGTF_SE <- read.table("fromGTF.SE.txt", header = TRUE, sep = "\t")
results <- grase::map_rMATS(
  g = g, gene = "ENSG00000000003.14",
  gff = "ENSG00000000003.dexseq.gff",
  fromGTF_A3SS = data.frame(), fromGTF_A5SS = data.frame(),
  fromGTF_SE = fromGTF_SE, fromGTF_RI = data.frame(),
  splits = splits, analysis_type = "bipartition"
)
}
}
