% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMATS.R
\name{map_rMATS_multinomial}
\alias{map_rMATS_multinomial}
\title{Map rMATS events using multinomial analysis (k-way partitions)}
\usage{
map_rMATS_multinomial(
  g,
  gene,
  gff,
  fromGTF_A3SS,
  fromGTF_A5SS,
  fromGTF_SE,
  fromGTF_RI,
  bipartitions,
  grase_output_dir,
  results_dir
)
}
\description{
Map rMATS events using multinomial analysis (k-way partitions)
}
