% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMATS.R
\name{map_rMATS_multinomial}
\alias{map_rMATS_multinomial}
\title{Map rMATS events using multinomial analysis (k-way partitions)}
\usage{
map_rMATS_multinomial(
  g,
  gene,
  gff,
  fromGTF_A3SS,
  fromGTF_A5SS,
  fromGTF_SE,
  fromGTF_RI,
  splits
)
}
\arguments{
\item{g}{Graph object}

\item{gene}{Gene identifier}

\item{gff}{GFF file path}

\item{fromGTF_A3SS}{Path to rMATS A3SS fromGTF file}

\item{fromGTF_A5SS}{Path to rMATS A5SS fromGTF file}

\item{fromGTF_SE}{Path to rMATS SE fromGTF file}

\item{fromGTF_RI}{Path to rMATS RI fromGTF file}

\item{splits}{Multinomial data frame with setdiff1, setdiff2, etc. columns}
}
\value{
Modified graph object with rMATS event annotations
}
\description{
Map rMATS events using multinomial analysis (k-way partitions)
}
