% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{Maps}
\alias{Maps}
\alias{Maps$new}
\alias{Maps$to_r}
\alias{Maps$from_r}
\alias{$.Maps}
\alias{[[.Maps}
\title{Maps  Class}
\source{
Generated by miniextendr from Rust type \code{Maps}

Generated by miniextendr from \code{Maps::new}

Generated by miniextendr from \code{Maps::to_r}

Generated by miniextendr from \code{Maps::from_r}
}
\usage{
Maps

\method{$}{Maps}(self, name)

\method{[[}{Maps}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Maps  Class
}
