% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_return_mode}
\alias{match_arg_return_mode}
\title{Test returning a MatchArg enum (uses IntoR -> character scalar)}
\source{
Generated by miniextendr from Rust fn \code{match_arg_return_mode}
}
\usage{
match_arg_return_mode(mode = NULL)
}
\arguments{
\item{mode}{A Mode enum value.}
}
\description{
Test returning a MatchArg enum (uses IntoR -> character scalar)
}
