% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_set_mode}
\alias{match_arg_set_mode}
\title{Set mode using match_arg}
\source{
Generated by miniextendr from Rust fn \code{match_arg_set_mode}
}
\usage{
match_arg_set_mode(mode = NULL)
}
\arguments{
\item{mode}{A Mode enum value.}
}
\description{
Set mode using match_arg
}
