% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{match_arg_set_status}
\alias{match_arg_set_status}
\title{Set status using match_arg}
\source{
Generated by miniextendr from Rust fn \code{match_arg_set_status}
}
\usage{
match_arg_set_status(status = NULL)
}
\arguments{
\item{status}{A BuildStatus enum value.}
}
\description{
Set status using match_arg
}
