% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR.R
\name{materialize}
\alias{fm.materialize}
\alias{fm.materialize.list}
\alias{fm.set.cached}
\alias{materialize}
\title{Materialize virtual FlashR objects.}
\usage{
fm.materialize.list(args)

fm.materialize(...)

fm.set.cached(fm, cached, in.mem = fm.in.mem(fm))
}
\arguments{
\item{args}{a list of virtual FlashR objects.}

\item{...}{a list of virtual FlashR objects.}

\item{fm}{a FlashR object.}

\item{in.mem}{a logical value, indicating whether to save the computation
results in memory.}
}
\value{
a list of materialized compuation results.
}
\description{
FlashR lazily evaluates many operations and outputs virtual objects
that represent computation results. \code{fm.materialize.list} and
\code{fm.materialize} explicitly materialize the virtualized computation
and save the computation results to memory or disks. Materialization of
these virtualized computations triggers materialization of other virtualized
computation. By default, FlashR only saves the computation results
specified by the arguments of \code{fm.materialize.list} and
\code{fm.materialize}.  \code{fm.set.cached} changes the default behavior and
notifies FlashR to save the materialized computation results of a virtual
matrix in memory or on disks.
}
\examples{
mat <- fm.mapply2(fm.runif.matrix(100, 10), fm.runif.matrix(100, 10), "+")
mat <- fm.materialize(mat)
mat2 <- fm.sapply(fm.runif.matrix(100, 10), "sqrt")
mat.list <- list(mat, mat2)
mat.list <- fm.materialize.list(mat.list)
mat <- fm.mapply2(fm.runif.matrix(100, 10), fm.runif.matrix(100, 10), "+")
fm.set.cached(mat, TRUE)
res <- fm.agg(mat, "+")
}

