% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{matmult}
\alias{\%*\%,ANY,fm-method}
\alias{\%*\%,fm,ANY-method}
\alias{\%*\%,fm,fm-method}
\alias{\%*\%,fm,fmV-method}
\alias{\%*\%,fmV,fm-method}
\alias{matmult}
\title{Matrix multiplication}
\usage{
\S4method{\%*\%}{fm,fm}(x, y)

\S4method{\%*\%}{fm,fmV}(x, y)

\S4method{\%*\%}{fmV,fm}(x, y)

\S4method{\%*\%}{fm,ANY}(x, y)

\S4method{\%*\%}{ANY,fm}(x, y)
}
\arguments{
\item{x}{a FlashR matrix.}

\item{y}{can be a FlashR vector or matrix, an R vector or matrix.}
}
\value{
a FlashR matrix.
}
\description{
Multiplies two matrices, if they are conformable. If one argument is
a vector, it will be promoted to either a row or column matrix to make
the two arguments conformable.
}
\examples{
mat1 <- fm.runif.matrix(100, 10)
mat2 <- fm.runif.matrix(10, 10)
mat <- mat1 \%*\% mat2
}

