% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl_plots.R
\name{mbl_plot_expression}
\alias{mbl_plot_expression}
\title{Plot expression of a gene/feature}
\usage{
mbl_plot_expression(y, gene, group, ...)
}
\description{
This function returns a raw ggplot object that you can decorate
further
}
