% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl_plot_expression.R
\name{mbl_plot_expression}
\alias{mbl_plot_expression}
\title{Plot expression of a gene/feature}
\usage{
mbl_plot_expression(y, gene, group, color_by = NULL, ...)
}
\arguments{
\item{y}{the DGEList or voomed object your data is in}

\item{gene}{the name of the gene identifier or symbol you want to plot.}

\item{group}{what column in the $samples (or $targets, for vm) pheno table
to use to plot split expression across the x axis}

\item{color_by}{name of column (like `group` param) to color your points by}
}
\value{
a ggplot object
}
\description{
This is a conveniencde function that takes your expression object
(`DGEList`, or "voomed" object) and plots the expression of one gene (`gene`)
across the entire dataset grouped by the sample covariate specified by
`group`. You can optionally color each point by a second sample covariate
specified by the `color_by` parameter.

Use the [mbl_tidy()] function combined with ggplot in order to make
more complex vizualizations of your expression data.
}
\examples{
y <- mbl_load_rnaseq("mouse", dataset = "mbl")
mbl_plot_expression(y, gene = "Fxyd6", group = "group",
                    color_by = "source")
mbl_plot_expression(y, gene = "Fxyd6", group = "genotype",
                    color_by = "source")
}
\seealso{
[mbl_tidy()]
}
