% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl_tidy.R
\name{mbl_tidy}
\alias{mbl_tidy}
\alias{mbl_tidy.DGEList}
\alias{mbl_tidy.EList}
\title{Converts your expression data into a (huge) "tidy" data.frame}
\usage{
mbl_tidy(x, ...)

\method{mbl_tidy}{DGEList}(x, normalized.lib.sizes = TRUE, prior.count = 3,
  ...)

\method{mbl_tidy}{EList}(x, ...)
}
\arguments{
\item{x}{The expression object to tidy}
}
\value{
a (huge) data.frame with your expression data. Each row holds the
  expression of one gene in one sample. The columns include all of the
  gene- and sample-level metadata for the obseration.
}
\description{
Use this function to extract all of your expression data into a data.frame
that can be used with dplyr and ggplot to make arbitrariy informative
plots.
}
\examples{
# Make a boxplot with points of Fxyd6 in the cheek wildtype/knockout
y <- mbl_load_rnaseq("mouse", dataset = "mbl")
ydat <- mbl_tidy(y) # all of the rnaseq data
gdat <- filter(ydat, source == "cheek", symbol == "Fxyd6")
ggplot(gdat, aes(x = genotype, y = cpm)) +
  geom_boxplot() +
  geom_point()
}
\seealso{
[mbl_plot_expression()]
}
