% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{md5_hex}
\alias{md5_hex}
\alias{md5_str}
\alias{md5_bytes}
\title{MD5 hash of a raw vector, as 32-char hex}
\source{
Generated by miniextendr from Rust fn \code{md5_hex}

Generated by miniextendr from Rust fn \code{md5_str}

Generated by miniextendr from Rust fn \code{md5_bytes}
}
\usage{
md5_hex(data)

md5_str(input)

md5_bytes(data)
}
\arguments{
\item{data}{Raw vector to hash.}

\item{input}{String to hash.}
}
\description{
MD5 hash of a raw vector, as 32-char hex

MD5 hash of a string, as 32-char hex

MD5 hash of a raw vector, as a 16-byte raw digest
}
