% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-documentation.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Metadata Summary}
\description{
Data representing metadata from the Redivis database. This dataset provides
information about tables, their number of rows, and their column names.
}
\details{
This data is generated during package development from the Redivis
database and stored in the \verb{data/} folder. It is used for internal functions
in the package to provide metadata information.
}
\keyword{datasets}
