% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flashmatrix_ext.R
\name{misc.math}
\alias{fm.abs}
\alias{fm.sqrt}
\alias{misc.math}
\title{Miscellaneous Mathematical Functions}
\usage{
fm.abs(m)

fm.sqrt(m)
}
\arguments{
\item{m}{a numeric vector or matrix.}
}
\value{
`abs' returns a vector/matrix of the same type as the input
vector/matrix; `sqrt' returns a numeric vector/matrix.
}
\description{
`abs(x)' computes the absolute value of x, `sqrt(x)' computes the
square root of x.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}

