% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{do_nothing}
\alias{do_nothing}
\title{do_nothing}
\source{
Generated by miniextendr from Rust fn \code{do_nothing}
}
\usage{
do_nothing()
}
\description{
Test returning a scalar integer SEXP directly.
}
