% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{missing_test_f64}
\alias{missing_test_f64}
\title{Test Missing\if{html}{\out{<f64>}} -- returns "absent" or the value as string}
\source{
Generated by miniextendr from Rust fn \code{missing_test_f64}
}
\usage{
missing_test_f64(x)
}
\arguments{
\item{x}{An optional numeric value.}
}
\description{
Test Missing\if{html}{\out{<f64>}} -- returns "absent" or the value as string
}
