% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{missing_test_option}
\alias{missing_test_option}
\title{Test Missing<Option\if{html}{\out{<f64>}}> -- distinguishes missing, NULL, and present}
\source{
Generated by miniextendr from Rust fn \code{missing_test_option}
}
\usage{
missing_test_option(x)
}
\arguments{
\item{x}{An optional nullable numeric.}
}
\description{
Test Missing<Option\if{html}{\out{<f64>}}> -- distinguishes missing, NULL, and present
}
