% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{missing_test_present}
\alias{missing_test_present}
\title{Test Missing\if{html}{\out{<i32>}} with is_present / is_absent}
\source{
Generated by miniextendr from Rust fn \code{missing_test_present}
}
\usage{
missing_test_present(x)
}
\arguments{
\item{x}{An optional integer value.}
}
\description{
Test Missing\if{html}{\out{<i32>}} with is_present / is_absent
}
