% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{missing_test_f64}
\alias{missing_test_f64}
\alias{missing_test_option}
\alias{missing_test_string}
\alias{missing_test_present}
\title{Test Missing\if{html}{\out{<f64>}} -- returns "absent" or the value as string}
\source{
Generated by miniextendr from Rust fn \code{missing_test_f64}

Generated by miniextendr from Rust fn \code{missing_test_option}

Generated by miniextendr from Rust fn \code{missing_test_string}

Generated by miniextendr from Rust fn \code{missing_test_present}
}
\usage{
missing_test_f64(x)

missing_test_option(x)

missing_test_string(x)

missing_test_present(x)
}
\arguments{
\item{x}{An optional integer value.}
}
\description{
Test Missing\if{html}{\out{<f64>}} -- returns "absent" or the value as string

Test Missing<Option\if{html}{\out{<f64>}}> -- distinguishes missing, NULL, and present

Test Missing\if{html}{\out{<String>}} with the helper methods

Test Missing\if{html}{\out{<i32>}} with is_present / is_absent
}
