% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingFields.R
\name{missingFields}
\alias{missingFields}
\title{Reports which fields are missing from an entry in relation to a template}
\usage{
missingFields(entry, template)
}
\arguments{
\item{entry}{An entry from the SIDb database}

\item{template}{A SIDb template in yaml format}
}
\description{
Reports which fields are missing from an entry in relation to a template
}
\details{
Prints one or more messages to the console if required data are missing
}
\examples{
\dontrun{
template=yaml::yaml.load_file(input="~/sidb/data/template_metadata.yaml")
missingFields(entry=sidb[[1]], template = template)
}
}
