% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_analysis.R
\name{multinomial_paths}
\alias{multinomial_paths}
\title{Compute diff exons from graph using multinomial partitions (each path as its own group)}
\usage{
multinomial_paths(
  gene,
  g,
  sg,
  outdir,
  max_path = 20,
  max_span = Inf,
  collapse_bubbles = FALSE
)
}
\description{
Compute diff exons from graph using multinomial partitions (each path as its own group)
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
result <- grase::multinomial_paths(
  gene = "ENSG00000000003.14", g = g, sg = NULL, outdir = tempdir(), max_path = 20
)
head(result[, c("gene", "source", "sink", "setdiff1", "setdiff2", "ref_ex_part")])
}
}
