% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_analysis.R
\name{multinomial_paths}
\alias{multinomial_paths}
\title{Compute diff exons from graph using multinomial partitions (each path as its own group)}
\usage{
multinomial_paths(
  gene,
  g,
  sg,
  outdir,
  max_path = 20,
  max_span = Inf,
  collapse_bubbles = FALSE
)
}
\description{
Compute diff exons from graph using multinomial partitions (each path as its own group)
}
