% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{MxBagHandle}
\alias{MxBagHandle}
\alias{new_mxbaghandle}
\alias{[[.MxBagHandle}
\alias{size.MxBagHandle}
\title{MxBagHandle S3 Class}
\usage{
new_mxbaghandle(values)

\method{[[}{MxBagHandle}(x, i, ...)

\method{size}{MxBagHandle}(x, ...)

MxBagHandle
}
\arguments{
\item{values}{Numeric values held by the handle.}

\item{x}{An object.}

\item{i}{(undocumented)}

\item{...}{Additional arguments.}
}
\description{
Element at 1-based position \code{i}.

S3 generic for \code{size}

Number of values.
}
