% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{MyFloat}
\alias{MyFloat}
\alias{MyFloat$new}
\alias{MyFloat$value}
\alias{MyFloat$nan}
\alias{$.MyFloat}
\alias{[[.MyFloat}
\alias{MyFloat$RPartialOrd$partial_cmp}
\title{MyFloat inherent methods: constructor and accessor}
\format{
An object of class \code{environment} of length 4.
}
\source{
Generated by miniextendr from Rust type \code{MyFloat}

Generated by miniextendr from \code{MyFloat::new}

Generated by miniextendr from \code{MyFloat::value}

Generated by miniextendr from \code{MyFloat::nan}
}
\usage{
MyFloat

\method{$}{MyFloat}(self, name)

\method{[[}{MyFloat}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}

\item{value}{Numeric value for the float wrapper.}
}
\description{
MyFloat inherent methods: constructor and accessor
}
\keyword{datasets}
