% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ndarray_roundtrip_vec}
\alias{ndarray_roundtrip_vec}
\alias{ndarray_roundtrip_array}
\alias{ndarray_roundtrip_matrix}
\alias{ndarray_roundtrip_int_vec}
\alias{ndarray_roundtrip_int_matrix}
\title{Test roundtripping a 1D numeric array through R and back}
\source{
Generated by miniextendr from Rust fn `ndarray_roundtrip_vec`

Generated by miniextendr from Rust fn `ndarray_roundtrip_array`

Generated by miniextendr from Rust fn `ndarray_roundtrip_matrix`

Generated by miniextendr from Rust fn `ndarray_roundtrip_int_vec`

Generated by miniextendr from Rust fn `ndarray_roundtrip_int_matrix`
}
\usage{
ndarray_roundtrip_vec(data)

ndarray_roundtrip_array(data)

ndarray_roundtrip_matrix(data)

ndarray_roundtrip_int_vec(data)

ndarray_roundtrip_int_matrix(data)
}
\arguments{
\item{data}{An integer matrix to roundtrip.}
}
\description{
Test roundtripping a 1D numeric array through R and back

Test roundtripping an N-dimensional numeric array through R and back

Test roundtripping a 2D numeric matrix through R and back

Test roundtripping a 1D integer array through R and back

Test roundtripping a 2D integer matrix through R and back
}
