% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{NdArrayDyn}
\alias{NdArrayDyn}
\alias{NdArrayDyn$new}
\alias{NdArrayDyn$len}
\alias{NdArrayDyn$is_empty}
\alias{NdArrayDyn$sum}
\alias{NdArrayDyn$mean}
\alias{NdArrayDyn$min}
\alias{NdArrayDyn$max}
\alias{NdArrayDyn$product}
\alias{NdArrayDyn$var}
\alias{NdArrayDyn$std}
\alias{NdArrayDyn$get_nd}
\alias{NdArrayDyn$slice_nd}
\alias{NdArrayDyn$shape_nd}
\alias{NdArrayDyn$ndim}
\alias{NdArrayDyn$len_nd}
\alias{NdArrayDyn$flatten}
\alias{NdArrayDyn$flatten_c}
\alias{NdArrayDyn$is_valid_nd}
\alias{NdArrayDyn$axis_slice}
\alias{NdArrayDyn$reshape}
\alias{NdArrayDyn$to_r}
\alias{NdArrayDyn$zeros}
\alias{NdArrayDyn$ones}
\alias{$.NdArrayDyn}
\alias{[[.NdArrayDyn}
\title{NdArrayDyn methods: N-dimensional array operations, indexing, and reshaping}
\format{
An object of class \code{environment} of length 23.
}
\source{
Generated by miniextendr from Rust type `NdArrayDyn`

Generated by miniextendr from `NdArrayDyn::new`

Generated by miniextendr from `NdArrayDyn::len`

Generated by miniextendr from `NdArrayDyn::is_empty`

Generated by miniextendr from `NdArrayDyn::sum`

Generated by miniextendr from `NdArrayDyn::mean`

Generated by miniextendr from `NdArrayDyn::min`

Generated by miniextendr from `NdArrayDyn::max`

Generated by miniextendr from `NdArrayDyn::product`

Generated by miniextendr from `NdArrayDyn::var`

Generated by miniextendr from `NdArrayDyn::std`

Generated by miniextendr from `NdArrayDyn::get_nd`

Generated by miniextendr from `NdArrayDyn::slice_nd`

Generated by miniextendr from `NdArrayDyn::shape_nd`

Generated by miniextendr from `NdArrayDyn::ndim`

Generated by miniextendr from `NdArrayDyn::len_nd`

Generated by miniextendr from `NdArrayDyn::flatten`

Generated by miniextendr from `NdArrayDyn::flatten_c`

Generated by miniextendr from `NdArrayDyn::is_valid_nd`

Generated by miniextendr from `NdArrayDyn::axis_slice`

Generated by miniextendr from `NdArrayDyn::reshape`

Generated by miniextendr from `NdArrayDyn::to_r`

Generated by miniextendr from `NdArrayDyn::zeros`

Generated by miniextendr from `NdArrayDyn::ones`
}
\usage{
NdArrayDyn

\method{$}{NdArrayDyn}(self, name)

\method{[[}{NdArrayDyn}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
NdArrayDyn methods: N-dimensional array operations, indexing, and reshaping
}
\keyword{datasets}
