% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{NdIntVec}
\alias{NdIntVec}
\alias{NdIntVec$new}
\alias{NdIntVec$len}
\alias{NdIntVec$is_empty}
\alias{NdIntVec$ndim}
\alias{NdIntVec$shape}
\alias{NdIntVec$sum}
\alias{NdIntVec$mean}
\alias{NdIntVec$min}
\alias{NdIntVec$max}
\alias{NdIntVec$product}
\alias{NdIntVec$var}
\alias{NdIntVec$std}
\alias{NdIntVec$get}
\alias{NdIntVec$first}
\alias{NdIntVec$last}
\alias{NdIntVec$slice_1d}
\alias{NdIntVec$to_r}
\alias{$.NdIntVec}
\alias{[[.NdIntVec}
\title{NdIntVec methods: 1D integer array operations, slicing, and conversion}
\format{
An object of class \code{environment} of length 17.
}
\source{
Generated by miniextendr from Rust type `NdIntVec`

Generated by miniextendr from `NdIntVec::new`

Generated by miniextendr from `NdIntVec::len`

Generated by miniextendr from `NdIntVec::is_empty`

Generated by miniextendr from `NdIntVec::ndim`

Generated by miniextendr from `NdIntVec::shape`

Generated by miniextendr from `NdIntVec::sum`

Generated by miniextendr from `NdIntVec::mean`

Generated by miniextendr from `NdIntVec::min`

Generated by miniextendr from `NdIntVec::max`

Generated by miniextendr from `NdIntVec::product`

Generated by miniextendr from `NdIntVec::var`

Generated by miniextendr from `NdIntVec::std`

Generated by miniextendr from `NdIntVec::get`

Generated by miniextendr from `NdIntVec::first`

Generated by miniextendr from `NdIntVec::last`

Generated by miniextendr from `NdIntVec::slice_1d`

Generated by miniextendr from `NdIntVec::to_r`
}
\usage{
NdIntVec

\method{$}{NdIntVec}(self, name)

\method{[[}{NdIntVec}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
NdIntVec methods: 1D integer array operations, slicing, and conversion
}
\keyword{datasets}
