% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{NdVec}
\alias{NdVec}
\alias{NdVec$new}
\alias{NdVec$len}
\alias{NdVec$is_empty}
\alias{NdVec$ndim}
\alias{NdVec$shape}
\alias{NdVec$sum}
\alias{NdVec$mean}
\alias{NdVec$min}
\alias{NdVec$max}
\alias{NdVec$product}
\alias{NdVec$var}
\alias{NdVec$std}
\alias{NdVec$get}
\alias{NdVec$first}
\alias{NdVec$last}
\alias{NdVec$slice_1d}
\alias{NdVec$get_many}
\alias{NdVec$is_valid_index}
\alias{NdVec$to_r}
\alias{NdVec$view_to_r}
\alias{NdVec$from_range}
\alias{$.NdVec}
\alias{[[.NdVec}
\title{1D ndarray wrapper for testing}
\format{
An object of class \code{environment} of length 21.
}
\source{
Generated by miniextendr from Rust type \code{NdVec}

Generated by miniextendr from \code{NdVec::new}

Generated by miniextendr from \code{NdVec::len}

Generated by miniextendr from \code{NdVec::is_empty}

Generated by miniextendr from \code{NdVec::ndim}

Generated by miniextendr from \code{NdVec::shape}

Generated by miniextendr from \code{NdVec::sum}

Generated by miniextendr from \code{NdVec::mean}

Generated by miniextendr from \code{NdVec::min}

Generated by miniextendr from \code{NdVec::max}

Generated by miniextendr from \code{NdVec::product}

Generated by miniextendr from \code{NdVec::var}

Generated by miniextendr from \code{NdVec::std}

Generated by miniextendr from \code{NdVec::get}

Generated by miniextendr from \code{NdVec::first}

Generated by miniextendr from \code{NdVec::last}

Generated by miniextendr from \code{NdVec::slice_1d}

Generated by miniextendr from \code{NdVec::get_many}

Generated by miniextendr from \code{NdVec::is_valid_index}

Generated by miniextendr from \code{NdVec::to_r}

Generated by miniextendr from \code{NdVec::view_to_r}

Generated by miniextendr from \code{NdVec::from_range}
}
\usage{
NdVec

\method{$}{NdVec}(self, name)

\method{[[}{NdVec}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\value{
An ExternalPtr to an NdVec.
}
\description{
Wraps a 1D ndarray and exposes adapter trait methods.
}
\examples{
v <- NdVec$new(c(1, 2, 3, 4, 5))
v$len()
v$sum()
v$get(2L)
}
\keyword{datasets}
