% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{NdVec}
\alias{NdVec}
\alias{NdVec$new}
\alias{NdVec$len}
\alias{NdVec$is_empty}
\alias{NdVec$ndim}
\alias{NdVec$shape}
\alias{NdVec$sum}
\alias{NdVec$mean}
\alias{NdVec$min}
\alias{NdVec$max}
\alias{NdVec$product}
\alias{NdVec$var}
\alias{NdVec$std}
\alias{NdVec$get}
\alias{NdVec$first}
\alias{NdVec$last}
\alias{NdVec$slice_1d}
\alias{NdVec$get_many}
\alias{NdVec$is_valid_index}
\alias{NdVec$to_r}
\alias{NdVec$view_to_r}
\alias{NdVec$from_range}
\alias{$.NdVec}
\alias{[[.NdVec}
\title{NdVec methods: 1D numeric array operations, slicing, and conversion}
\format{
An object of class \code{environment} of length 21.
}
\source{
Generated by miniextendr from Rust type `NdVec`

Generated by miniextendr from `NdVec::new`

Generated by miniextendr from `NdVec::len`

Generated by miniextendr from `NdVec::is_empty`

Generated by miniextendr from `NdVec::ndim`

Generated by miniextendr from `NdVec::shape`

Generated by miniextendr from `NdVec::sum`

Generated by miniextendr from `NdVec::mean`

Generated by miniextendr from `NdVec::min`

Generated by miniextendr from `NdVec::max`

Generated by miniextendr from `NdVec::product`

Generated by miniextendr from `NdVec::var`

Generated by miniextendr from `NdVec::std`

Generated by miniextendr from `NdVec::get`

Generated by miniextendr from `NdVec::first`

Generated by miniextendr from `NdVec::last`

Generated by miniextendr from `NdVec::slice_1d`

Generated by miniextendr from `NdVec::get_many`

Generated by miniextendr from `NdVec::is_valid_index`

Generated by miniextendr from `NdVec::to_r`

Generated by miniextendr from `NdVec::view_to_r`

Generated by miniextendr from `NdVec::from_range`
}
\usage{
NdVec

\method{$}{NdVec}(self, name)

\method{[[}{NdVec}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
NdVec methods: 1D numeric array operations, slicing, and conversion
}
\keyword{datasets}
