% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{new_derived_int_lists}
\alias{new_derived_int_lists}
\title{Create a new derived_int_lists vector (list_of\if{html}{\out{<integer>}})}
\source{
Generated by miniextendr from Rust fn \code{new_derived_int_lists}
}
\usage{
new_derived_int_lists(x)
}
\arguments{
\item{x}{A list of integer vectors.}
}
\value{
A derived_int_lists list_of vector.
}
\description{
Create a new derived_int_lists vector (list_of\if{html}{\out{<integer>}})
}
