% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{new_derived_temp}
\alias{new_derived_temp}
\title{Create a new derived_temp vector with arithmetic/math support}
\source{
Generated by miniextendr from Rust fn \code{new_derived_temp}
}
\usage{
new_derived_temp(x)
}
\arguments{
\item{x}{Temperature values.}
}
\value{
A derived_temp vector.
}
\description{
Create a new derived_temp vector with arithmetic/math support
}
