% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrogen.R
\name{nitrogen}
\alias{nitrogen}
\title{Creates a data frame with the nitrogen content in soil samples}
\usage{
nitrogen(database)
}
\arguments{
\item{database}{A list with the sidb}
}
\value{
A data.frame with the nitrogen content and the units of the soil samples from the database
}
\description{
Creates a data frame with the nitrogen content in soil samples
}
\examples{
sidb=loadEntries(path="~/sidb/data/")
N=nitrogen(database=sidb)
}
