% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{class}
\name{OmitPickS4}
\alias{OmitPickS4}
\alias{OmitPickS4-s4_omit_pick}
\alias{s4_omit_pick,OmitPickS4-method}
\title{OmitPickS4 S4 Class}
\usage{
OmitPickS4()
}
\description{
Omittable choices on an S4 method: reports what reached Rust for each of \code{mode}, \code{modes} and \code{level}.
}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to Rust \code{OmitPickS4} struct}
}}

