% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onepFit.R
\name{onepFit}
\alias{onepFit}
\title{Fits a one pool model to a time-series}
\usage{
onepFit(timeSeries, initialCarbon)
}
\arguments{
\item{timeSeries}{A time series of respiration values over time}

\item{initialCarbon}{The initial amount of carbon in units that correspond to the time series data}
}
\value{
R list with an FME model object, a SoilR model object, and the AIC value
}
\description{
Fits a one pool model to a time-series
}
\examples{
a=onepFit(timeSeries = sidb[[20]]$timeSeries[,1:2],
initialCarbon=sidb[[20]]$initConditions[1,"carbonMean"]*10)
}
