% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{OptsTarget}
\alias{OptsTarget}
\alias{OptsTarget$new}
\alias{$.OptsTarget}
\alias{[[.OptsTarget}
\alias{OptsTarget$OptionsDemo$basic_value}
\alias{OptsTarget$OptionsDemo$with_entry}
\alias{OptsTarget$OptionsDemo$with_exit}
\alias{OptsTarget$OptionsDemo$with_checks}
\alias{OptsTarget$OptionsDemo$deprecated_method}
\title{OptsTarget  Class}
\source{
Generated by miniextendr from Rust type \code{OptsTarget}

Generated by miniextendr from \code{OptsTarget::new}
}
\usage{
OptsTarget

\method{$}{OptsTarget}(self, name)

\method{[[}{OptsTarget}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
OptsTarget  Class
}
