% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{telemetry_get_count}
\alias{telemetry_get_count}
\alias{telemetry_clear_hook}
\alias{telemetry_install_counter}
\title{panic telemetry tests}
\source{
Generated by miniextendr from Rust fn \code{telemetry_get_count}

Generated by miniextendr from Rust fn \code{telemetry_clear_hook}

Generated by miniextendr from Rust fn \code{telemetry_install_counter}
}
\usage{
telemetry_get_count()

telemetry_clear_hook()

telemetry_install_counter()
}
\description{
panic telemetry tests

panic telemetry tests

panic telemetry tests
}
