% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{PanickyCounter}
\alias{PanickyCounter}
\alias{PanickyCounter$get_value}
\alias{PanickyCounter$try_add}
\alias{PanickyCounter$new_panicky}
\alias{$.PanickyCounter}
\alias{[[.PanickyCounter}
\alias{PanickyCounter$Counter$value}
\alias{PanickyCounter$Counter$increment}
\alias{PanickyCounter$Counter$add}
\title{PanickyCounter Class}
\format{
An object of class \code{environment} of length 4.
}
\source{
Generated by miniextendr from Rust type \code{PanickyCounter}
}
\usage{
PanickyCounter

\method{$}{PanickyCounter}(self, name)

\method{[[}{PanickyCounter}(self, name)
}
\description{
Get value.

Add (may panic).

Create a new panicky counter.
}
\keyword{datasets}
