% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ParamCheckHolder}
\alias{ParamCheckHolder}
\alias{ParamCheckHolder$new}
\alias{ParamCheckHolder$add}
\alias{$.ParamCheckHolder}
\alias{[[.ParamCheckHolder}
\title{ParamCheckHolder  Class}
\usage{
ParamCheckHolder

\method{$}{ParamCheckHolder}(self, name)

\method{[[}{ParamCheckHolder}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Env class whose \code{add()} method checks its arguments with \code{inherits(...)} / \code{no_na(...)}.

Add \code{y} if \code{x} is an \code{mx_obj}.

\describe{
\item{\code{x}}{An object of class \code{mx_obj} or \code{mx_other}.}
\item{\code{y}}{A non-NA double.}
}
}
