% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartitions.R
\name{path_subset_relation}
\alias{path_subset_relation}
\title{find a containment DAG that represents the hierarchical relationship between paths}
\usage{
path_subset_relation(sets)
}
\description{
find a containment DAG that represents the hierarchical relationship between paths
}
