% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{PipeCounter}
\alias{PipeCounter}
\alias{new_pipecounter}
\alias{bump.PipeCounter}
\alias{twice.PipeCounter}
\alias{peek.PipeCounter}
\title{Counter with pipe-friendly mutators returning \verb{&mut Self}}
\source{
Generated by miniextendr from Rust type \code{PipeCounter}

Generated by miniextendr from \code{PipeCounter::bump}

Generated by miniextendr from \code{PipeCounter::bump}

Generated by miniextendr from \code{PipeCounter::twice}

Generated by miniextendr from \code{PipeCounter::twice}

Generated by miniextendr from \code{PipeCounter::peek}

Generated by miniextendr from \code{PipeCounter::peek}
}
\usage{
new_pipecounter(initial)

\method{bump}{PipeCounter}(x, amount, ...)

\method{twice}{PipeCounter}(x, ...)

\method{peek}{PipeCounter}(x, ...)

PipeCounter
}
\arguments{
\item{initial}{Initial counter value.}

\item{x}{An object.}

\item{amount}{Amount to add.}

\item{...}{Additional arguments.}
}
\description{
S3 generic for \code{bump}

S3 generic for \code{twice}

S3 generic for \code{peek}
}
