% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/info.R
\name{pkg_info}
\alias{pkg_info}
\title{Information about a package}
\usage{
pkg_info(pkg, global = FALSE)
}
\arguments{
\item{pkg}{Package name and optionally version after a dash.}

\item{global}{Whether to check for installed packages globally
or locally.}
}
\value{
An rpkg package, invisibly.
}
\description{
Print the description of a CRAN package. It does not have
to be installed.
}

