% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/tree.R
\name{pkg_tree}
\alias{pkg_tree}
\title{Dependency tree of a package}
\usage{
pkg_tree(pkg, recursive = c("Imports*", "Depends*", "LinkingTo*"))
}
\arguments{
\item{pkg}{Package name, optionally with version number after a dash.}

\item{recursive}{What kind of dependencies to collect. If a type ends
  with a star, that type of dependency is collected recursively.}
}
\description{
Dependency tree of a package
}

