% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{Point}
\alias{Point}
\alias{Point$new}
\alias{Point$x}
\alias{Point$y}
\alias{$.Point}
\alias{[[.Point}
\alias{Point$ROrd$cmp}
\alias{Point$RCopy$copy}
\alias{Point$RCopy$is_copy}
\alias{Point$RHash$hash}
\alias{Point$RClone$clone}
\alias{Point$RDebug$debug_str}
\alias{Point$RDebug$debug_str_pretty}
\alias{Point$RDefault$default}
\alias{Point$RDisplay$as_r_string}
\alias{Point$RFromStr$from_str}
\title{Point inherent methods: constructor and field accessors}
\format{
An object of class \code{environment} of length 11.
}
\source{
Generated by miniextendr from Rust type `Point`

Generated by miniextendr from `Point::new`

Generated by miniextendr from `Point::x`

Generated by miniextendr from `Point::y`
}
\usage{
Point

\method{$}{Point}(self, name)

\method{[[}{Point}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Point inherent methods: constructor and field accessors
}
\keyword{datasets}
