% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{Point}
\alias{Point}
\alias{Point$new}
\alias{Point$x}
\alias{Point$y}
\alias{Point$debug_str}
\alias{Point$debug_str_pretty}
\alias{Point$to_r_string}
\alias{Point$hash}
\alias{Point$cmp_to}
\alias{Point$is_less_than}
\alias{Point$is_equal_to}
\alias{Point$is_greater_than}
\alias{Point$clone_point}
\alias{Point$default_point}
\alias{Point$from_str}
\alias{$.Point}
\alias{[[.Point}
\title{Point struct for testing adapter traits}
\format{
An object of class \code{environment} of length 14.
}
\source{
Generated by miniextendr from Rust type \code{Point}

Generated by miniextendr from \code{Point::new}

Generated by miniextendr from \code{Point::x}

Generated by miniextendr from \code{Point::y}

Generated by miniextendr from \code{Point::debug_str}

Generated by miniextendr from \code{Point::debug_str_pretty}

Generated by miniextendr from \code{Point::to_r_string}

Generated by miniextendr from \code{Point::hash}

Generated by miniextendr from \code{Point::cmp_to}

Generated by miniextendr from \code{Point::is_less_than}

Generated by miniextendr from \code{Point::is_equal_to}

Generated by miniextendr from \code{Point::is_greater_than}

Generated by miniextendr from \code{Point::clone_point}

Generated by miniextendr from \code{Point::default_point}

Generated by miniextendr from \code{Point::from_str}
}
\usage{
Point

\method{$}{Point}(self, name)

\method{[[}{Point}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\value{
An ExternalPtr to a Point.
}
\description{
A 2D point that demonstrates adapter trait functionality.
}
\examples{
p <- Point$new(3L, 4L)
p$debug_str()
p$to_r_string()
p$hash()
}
\keyword{datasets}
