% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,fm-method}
\alias{print,fm.bo-method}
\alias{print,fmV-method}
\alias{print,fmVFactor-method}
\title{Print FlashR objects.}
\usage{
\S4method{print}{fm}(x)

\S4method{print}{fmV}(x)

\S4method{print}{fmVFactor}(x)

\S4method{print}{fm.bo}(x)
}
\arguments{
\item{x}{a FlashR object. It can be a vector, a matrix or a FlashR
binary operator.}
}
\description{
\code{print} prints the information of a FlashR object.
}

