% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorityFlow.R
\docType{package}
\name{PriorityFlow-package}
\alias{PriorityFlow}
\alias{PriorityFlow-package}
\title{PriorityFlow: What the package does (short line)}
\description{
More about what it does (maybe more than one line)
}
\details{
The only function you're likely to need from roxygen2 is [roxygenize()]. 
Otherwise refer to the vignettes to see how to format the documentation.
}
\keyword{internal}
