\name{PriorityFlow-package}
\alias{PriorityFlow-package}
\alias{PriorityFlow}
\docType{package}
\title{
\packageTitle{PriorityFlow}
}
\description{
\packageDescription{PriorityFlow}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PriorityFlow}
\packageIndices{PriorityFlow}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PriorityFlow}

Maintainer: \packageMaintainer{PriorityFlow}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
