% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{protect_pool_multi}
\alias{protect_pool_multi}
\alias{protect_pool_roundtrip}
\title{protect pool tests}
\source{
Generated by miniextendr from Rust fn \code{protect_pool_multi}

Generated by miniextendr from Rust fn \code{protect_pool_roundtrip}
}
\usage{
protect_pool_multi()

protect_pool_roundtrip()
}
\description{
protect pool tests

protect pool tests
}
