% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{PtrIdentityTest}
\alias{PtrIdentityTest}
\alias{PtrIdentityTest$new}
\alias{PtrIdentityTest$value}
\alias{$.PtrIdentityTest}
\alias{[[.PtrIdentityTest}
\title{PtrIdentityTest Class}
\usage{
PtrIdentityTest

\method{$}{PtrIdentityTest}(self, name)

\method{[[}{PtrIdentityTest}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
PtrIdentityTest Class
}
