% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{PtrIdentityTest}
\alias{PtrIdentityTest}
\alias{PtrIdentityTest$new}
\alias{PtrIdentityTest$value}
\alias{$.PtrIdentityTest}
\alias{[[.PtrIdentityTest}
\title{PtrIdentityTest  Class}
\format{
An object of class \code{environment} of length 2.
}
\source{
Generated by miniextendr from Rust type `PtrIdentityTest`

Generated by miniextendr from `PtrIdentityTest::new`

Generated by miniextendr from `PtrIdentityTest::value`
}
\usage{
PtrIdentityTest

\method{$}{PtrIdentityTest}(self, name)

\method{[[}{PtrIdentityTest}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
PtrIdentityTest  Class
}
\keyword{datasets}
