% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{R6Animal}
\alias{R6Animal}
\title{R6Animal R6 Class}
\source{
Generated by miniextendr from Rust type \code{R6Animal}
}
\description{
R6 base animal class for inheritance testing.
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-R6Animal-initialize}{\code{R6Animal$new()}}
    \item \href{#method-R6Animal-name}{\code{R6Animal$name()}}
    \item \href{#method-R6Animal-speak}{\code{R6Animal$speak()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Animal-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-R6Animal-initialize}{}}}
\subsection{\code{R6Animal$new()}}{
  Create a new \code{R6Animal}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{R6Animal$new(name, sound, .ptr = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{Character name of the animal.}
      \item{\code{sound}}{Character sound the animal makes.}
      \item{\code{.ptr}}{Internal pointer (used by static methods, not for direct use).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Animal-name"></a>}}
\if{latex}{\out{\hypertarget{method-R6Animal-name}{}}}
\subsection{\code{R6Animal$name()}}{
  Method \code{name}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{R6Animal$name()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Animal-speak"></a>}}
\if{latex}{\out{\hypertarget{method-R6Animal-speak}{}}}
\subsection{\code{R6Animal$speak()}}{
  Speak the given number of times.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{R6Animal$speak(times)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{times}}{Number of times to speak.}
    }
    \if{html}{\out{</div>}}
  }
}

}
