% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{R6Counter}
\alias{R6Counter}
\title{R6Counter R6 Class}
\source{
Generated by miniextendr from Rust type \code{R6Counter}
}
\description{
R6 counter class that stores a single integer value.
}
\examples{
c <- R6Counter$new(1L)
c$value()
c$inc()
c$add(10L)
R6Counter$default_counter()$value()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.ptr}}{(private) External pointer to Rust \code{R6Counter} struct}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.ptr}}{(private) External pointer to Rust \code{R6Counter} struct}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-R6Counter-new}{\code{R6Counter$new()}}
\item \href{#method-R6Counter-value}{\code{R6Counter$value()}}
\item \href{#method-R6Counter-inc}{\code{R6Counter$inc()}}
\item \href{#method-R6Counter-add}{\code{R6Counter$add()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Counter-new"></a>}}
\if{latex}{\out{\hypertarget{method-R6Counter-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Counter$new(initial, .ptr = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Counter-value"></a>}}
\if{latex}{\out{\hypertarget{method-R6Counter-value}{}}}
\subsection{Method \code{value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Counter$value()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Counter-inc"></a>}}
\if{latex}{\out{\hypertarget{method-R6Counter-inc}{}}}
\subsection{Method \code{inc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Counter$inc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Counter-add"></a>}}
\if{latex}{\out{\hypertarget{method-R6Counter-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Counter$add(amount)}\if{html}{\out{</div>}}
}

}
}
