% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{R6Dog}
\alias{R6Dog}
\title{R6 dog class inheriting from R6Animal}
\source{
Generated by miniextendr from Rust type \code{R6Dog}
}
\description{
R6 dog class inheriting from R6Animal

R6 dog class inheriting from R6Animal
}
\section{Super class}{
\code{\link[miniextendr:R6Animal]{miniextendr::R6Animal}} -> \code{R6Dog}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-R6Dog-new}{\code{R6Dog$new()}}
\item \href{#method-R6Dog-breed}{\code{R6Dog$breed()}}
\item \href{#method-R6Dog-fetch}{\code{R6Dog$fetch()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="miniextendr" data-topic="R6Animal" data-id="name"><a href='../../miniextendr/html/R6Animal.html#method-R6Animal-name'><code>miniextendr::R6Animal$name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="miniextendr" data-topic="R6Animal" data-id="speak"><a href='../../miniextendr/html/R6Animal.html#method-R6Animal-speak'><code>miniextendr::R6Animal$speak()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Dog-new"></a>}}
\if{latex}{\out{\hypertarget{method-R6Dog-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{R6Dog}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Dog$new(breed, .ptr = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{breed}}{(no documentation available)}

\item{\code{.ptr}}{Internal pointer (used by static methods, not for direct use).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Dog-breed"></a>}}
\if{latex}{\out{\hypertarget{method-R6Dog-breed}{}}}
\subsection{Method \code{breed()}}{
Method \code{breed}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Dog$breed()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Dog-fetch"></a>}}
\if{latex}{\out{\hypertarget{method-R6Dog-fetch}{}}}
\subsection{Method \code{fetch()}}{
Method \code{fetch}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Dog$fetch()}\if{html}{\out{</div>}}
}

}
}
