% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{R6GoldenRetriever}
\alias{R6GoldenRetriever}
\title{R6 golden retriever class for 3-level inheritance chain testing}
\source{
Generated by miniextendr from Rust type \code{R6GoldenRetriever}
}
\description{
R6 golden retriever class for 3-level inheritance chain testing

R6 golden retriever class for 3-level inheritance chain testing
}
\section{Super classes}{
\code{miniextendr::R6Animal} -> \code{miniextendr::R6Dog} -> \code{R6GoldenRetriever}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-R6GoldenRetriever-new}{\code{R6GoldenRetriever$new()}}
\item \href{#method-R6GoldenRetriever-owner}{\code{R6GoldenRetriever$owner()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="miniextendr" data-topic="R6Animal" data-id="name"><a href='../../miniextendr/html/R6Animal.html#method-R6Animal-name'><code>miniextendr::R6Animal$name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="miniextendr" data-topic="R6Animal" data-id="speak"><a href='../../miniextendr/html/R6Animal.html#method-R6Animal-speak'><code>miniextendr::R6Animal$speak()</code></a></span></li>
<li><span class="pkg-link" data-pkg="miniextendr" data-topic="R6Dog" data-id="breed"><a href='../../miniextendr/html/R6Dog.html#method-R6Dog-breed'><code>miniextendr::R6Dog$breed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="miniextendr" data-topic="R6Dog" data-id="fetch"><a href='../../miniextendr/html/R6Dog.html#method-R6Dog-fetch'><code>miniextendr::R6Dog$fetch()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6GoldenRetriever-new"></a>}}
\if{latex}{\out{\hypertarget{method-R6GoldenRetriever-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6GoldenRetriever$new(owner, .ptr = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{owner}}{Character name of the owner.}

\item{\code{.ptr}}{Internal pointer (used by static methods, not for direct use).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6GoldenRetriever-owner"></a>}}
\if{latex}{\out{\hypertarget{method-R6GoldenRetriever-owner}{}}}
\subsection{Method \code{owner()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6GoldenRetriever$owner()}\if{html}{\out{</div>}}
}

}
}
