% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{R6Raiser}
\alias{R6Raiser}
\title{R6Raiser R6 Class}
\source{
Generated by miniextendr from Rust type \code{R6Raiser}
}
\description{
R6Raiser R6 Class

R6Raiser R6 Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-R6Raiser-new}{\code{R6Raiser$new()}}
\item \href{#method-R6Raiser-id}{\code{R6Raiser$id()}}
\item \href{#method-R6Raiser-raise_error}{\code{R6Raiser$raise_error()}}
\item \href{#method-R6Raiser-raise_error_classed}{\code{R6Raiser$raise_error_classed()}}
\item \href{#method-R6Raiser-raise_warning}{\code{R6Raiser$raise_warning()}}
\item \href{#method-R6Raiser-raise_warning_classed}{\code{R6Raiser$raise_warning_classed()}}
\item \href{#method-R6Raiser-raise_message}{\code{R6Raiser$raise_message()}}
\item \href{#method-R6Raiser-raise_condition}{\code{R6Raiser$raise_condition()}}
\item \href{#method-R6Raiser-raise_condition_classed}{\code{R6Raiser$raise_condition_classed()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Raiser-new"></a>}}
\if{latex}{\out{\hypertarget{method-R6Raiser-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{R6Raiser}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Raiser$new(id, .ptr = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(no documentation available)}

\item{\code{.ptr}}{Internal pointer (used by static methods, not for direct use).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Raiser-id"></a>}}
\if{latex}{\out{\hypertarget{method-R6Raiser-id}{}}}
\subsection{Method \code{id()}}{
Method \code{id}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Raiser$id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Raiser-raise_error"></a>}}
\if{latex}{\out{\hypertarget{method-R6Raiser-raise_error}{}}}
\subsection{Method \code{raise_error()}}{
Method \code{raise_error}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Raiser$raise_error(msg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{(no documentation available)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Raiser-raise_error_classed"></a>}}
\if{latex}{\out{\hypertarget{method-R6Raiser-raise_error_classed}{}}}
\subsection{Method \code{raise_error_classed()}}{
Method \code{raise_error_classed}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Raiser$raise_error_classed(class, msg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class}}{(no documentation available)}

\item{\code{msg}}{(no documentation available)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Raiser-raise_warning"></a>}}
\if{latex}{\out{\hypertarget{method-R6Raiser-raise_warning}{}}}
\subsection{Method \code{raise_warning()}}{
Method \code{raise_warning}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Raiser$raise_warning(msg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{(no documentation available)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Raiser-raise_warning_classed"></a>}}
\if{latex}{\out{\hypertarget{method-R6Raiser-raise_warning_classed}{}}}
\subsection{Method \code{raise_warning_classed()}}{
Method \code{raise_warning_classed}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Raiser$raise_warning_classed(class, msg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class}}{(no documentation available)}

\item{\code{msg}}{(no documentation available)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Raiser-raise_message"></a>}}
\if{latex}{\out{\hypertarget{method-R6Raiser-raise_message}{}}}
\subsection{Method \code{raise_message()}}{
Method \code{raise_message}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Raiser$raise_message(msg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{(no documentation available)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Raiser-raise_condition"></a>}}
\if{latex}{\out{\hypertarget{method-R6Raiser-raise_condition}{}}}
\subsection{Method \code{raise_condition()}}{
Method \code{raise_condition}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Raiser$raise_condition(msg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{(no documentation available)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Raiser-raise_condition_classed"></a>}}
\if{latex}{\out{\hypertarget{method-R6Raiser-raise_condition_classed}{}}}
\subsection{Method \code{raise_condition_classed()}}{
Method \code{raise_condition_classed}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Raiser$raise_condition_classed(class, msg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class}}{(no documentation available)}

\item{\code{msg}}{(no documentation available)}
}
\if{html}{\out{</div>}}
}
}
}
