% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{R6TraitCounter}
\alias{R6TraitCounter}
\alias{R6TraitCounter$new_r6trait}
\alias{r6_trait_Counter_value}
\alias{r6_trait_Counter_increment}
\alias{r6_trait_Counter_checked_add}
\alias{R6TraitCounter$Counter$default_initial}
\alias{R6TraitCounter$Counter$MAX_VALUE}
\title{R6TraitCounter R6 Class}
\source{
Generated by miniextendr from Rust type \code{R6TraitCounter}

Generated by miniextendr from \code{R6TraitCounter::new_r6trait}

Generated by miniextendr from \verb{impl Counter for R6TraitCounter}
R6 trait method \code{Counter::value} for R6TraitCounter

Generated by miniextendr from \verb{impl Counter for R6TraitCounter}
R6 trait method \code{Counter::increment} for R6TraitCounter

Generated by miniextendr from \verb{impl Counter for R6TraitCounter}
R6 trait method \code{Counter::checked_add} for R6TraitCounter
}
\usage{
r6_trait_Counter_value(x)

r6_trait_Counter_increment(x)

r6_trait_Counter_checked_add(x, n)
}
\arguments{
\item{x}{A \code{R6TraitCounter} object}

\item{n}{Amount to add.}
}
\description{
Create a new R6 trait counter.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-R6TraitCounter-new}{\code{R6TraitCounter$new()}}
\item \href{#method-R6TraitCounter-get_value}{\code{R6TraitCounter$get_value()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6TraitCounter-new"></a>}}
\if{latex}{\out{\hypertarget{method-R6TraitCounter-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6TraitCounter$new(.ptr = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.ptr}}{Internal pointer (used by static methods, not for direct use).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6TraitCounter-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-R6TraitCounter-get_value}{}}}
\subsection{Method \code{get_value()}}{
Get value via inherent method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6TraitCounter$get_value()}\if{html}{\out{</div>}}
}

}
}
