% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{R6TraitCounter}
\alias{R6TraitCounter}
\alias{R6TraitCounter$get_value}
\alias{R6TraitCounter$new_r6trait}
\alias{$.R6TraitCounter}
\alias{[[.R6TraitCounter}
\alias{r6_trait_Counter_value}
\alias{r6_trait_Counter_increment}
\alias{r6_trait_Counter_checked_add}
\alias{R6TraitCounter$Counter$default_initial}
\alias{R6TraitCounter$Counter$MAX_VALUE}
\title{R6TraitCounter Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type \code{R6TraitCounter}

Generated by miniextendr from \verb{impl Counter for R6TraitCounter}

Generated by miniextendr from \verb{impl Counter for R6TraitCounter}

Generated by miniextendr from \verb{impl Counter for R6TraitCounter}
}
\usage{
R6TraitCounter

\method{$}{R6TraitCounter}(self, name)

\method{[[}{R6TraitCounter}(self, name)

r6_trait_Counter_value(x)

r6_trait_Counter_increment(x)

r6_trait_Counter_checked_add(x, n)
}
\arguments{
\item{x}{A \code{R6TraitCounter} object}
}
\description{
Get value via inherent method.

Create a new R6 trait counter.
}
\keyword{datasets}
