% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{r6_trait_Counter_value}
\alias{r6_trait_Counter_value}
\alias{r6_trait_Counter_increment}
\alias{r6_trait_Counter_checked_add}
\alias{R6TraitCounter$Counter$default_initial}
\alias{R6TraitCounter$Counter$MAX_VALUE}
\title{Static trait method Counter::default_initial()}
\source{
Generated by miniextendr from \verb{impl Counter for R6TraitCounter}
R6 trait method \code{Counter::value} for R6TraitCounter

Generated by miniextendr from \verb{impl Counter for R6TraitCounter}
R6 trait method \code{Counter::increment} for R6TraitCounter

Generated by miniextendr from \verb{impl Counter for R6TraitCounter}
R6 trait method \code{Counter::checked_add} for R6TraitCounter
}
\usage{
r6_trait_Counter_value(x)

r6_trait_Counter_increment(x)

r6_trait_Counter_checked_add(x, n)
}
\arguments{
\item{x}{A \code{R6TraitCounter} object}

\item{n}{Amount to add.}
}
\description{
Static trait method Counter::default_initial()
}
