% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rarray_matrix_column}
\alias{rarray_matrix_column}
\title{Get a specific column from a numeric matrix as a Vec}
\source{
Generated by miniextendr from Rust fn \code{rarray_matrix_column}
}
\usage{
rarray_matrix_column(x, col)
}
\arguments{
\item{x}{(no documentation available)}

\item{col}{(no documentation available)}
}
\description{
Get a specific column from a numeric matrix as a Vec
}
