% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rarray_matrix_dims}
\alias{rarray_matrix_dims}
\title{Get dimensions of a matrix as integer vector}
\source{
Generated by miniextendr from Rust fn \code{rarray_matrix_dims}
}
\usage{
rarray_matrix_dims(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
Get dimensions of a matrix as integer vector
}
