% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{RCoerceErrorTest}
\alias{RCoerceErrorTest}
\alias{new_rcoerceerrortest}
\alias{as_data_frame.RCoerceErrorTest}
\alias{as_list.RCoerceErrorTest}
\alias{as_character.RCoerceErrorTest}
\alias{as.data.frame.RCoerceErrorTest}
\alias{as.list.RCoerceErrorTest}
\alias{as.character.RCoerceErrorTest}
\title{as.\if{html}{\out{<class>}}() Error Test Type}
\source{
Generated by miniextendr from Rust type \code{RCoerceErrorTest}

Generated by miniextendr from \code{RCoerceErrorTest::as_data_frame}

Generated by miniextendr from \code{RCoerceErrorTest::as_data_frame}

Generated by miniextendr from \code{RCoerceErrorTest::as_list}

Generated by miniextendr from \code{RCoerceErrorTest::as_list}

Generated by miniextendr from \code{RCoerceErrorTest::as_character}

Generated by miniextendr from \code{RCoerceErrorTest::as_character}

Generated by miniextendr from \code{RCoerceErrorTest::as_data_frame}

Generated by miniextendr from \code{RCoerceErrorTest::as_list}

Generated by miniextendr from \code{RCoerceErrorTest::as_character}
}
\usage{
new_rcoerceerrortest(is_empty)

\method{as_data_frame}{RCoerceErrorTest}(x, ...)

\method{as_list}{RCoerceErrorTest}(x, ...)

\method{as_character}{RCoerceErrorTest}(x, ...)

RCoerceErrorTest

\method{as.data.frame}{RCoerceErrorTest}(x, ...)

\method{as.list}{RCoerceErrorTest}(x, ...)

\method{as.character}{RCoerceErrorTest}(x, ...)
}
\arguments{
\item{is_empty}{Logical; if TRUE, triggers error on as.data.frame.}

\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
S3 generic for \code{as_data_frame}

S3 generic for \code{as_list}

S3 generic for \code{as_character}
}
