% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEntry.R
\name{readEntry}
\alias{readEntry}
\title{Load a single SIDb entry from entry files}
\usage{
readEntry(path, entryName)
}
\arguments{
\item{path}{character string with the path where SIDb entry files are stored}

\item{entryName}{character string with the name of the entry in the database}
}
\value{
R list of the entry with the SIDb structure
}
\description{
Load a single SIDb entry from entry files
}
\examples{
\dontrun{
Andrews2000SBB=readEntry(path="~/sidb/data/", entryName="Andrews2000SBB")
}
}
