% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEntry.R
\name{readEntry}
\alias{readEntry}
\title{Read single entry of the sidb}
\usage{
readEntry(path, entryName)
}
\arguments{
\item{path}{character string with the path where isdb is stored}

\item{entryName}{character string with the name of the entry in the database}
}
\value{
R list with the entry
}
\description{
Read single entry of the sidb
}
\examples{
\dontrun{
Andrews2000SBB=readEntry(path="~/sidb/data/", entryName="Andrews2000SBB")
}
}
