% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{ReceiverCounter}
\alias{ReceiverCounter}
\alias{ReceiverCounter$new}
\alias{ReceiverCounter$value}
\alias{ReceiverCounter$inc}
\alias{ReceiverCounter$add}
\alias{ReceiverCounter$default_counter}
\alias{$.ReceiverCounter}
\alias{[[.ReceiverCounter}
\title{ReceiverCounter  Class}
\format{
An object of class \code{environment} of length 5.
}
\source{
Generated by miniextendr from Rust type \code{ReceiverCounter}

Generated by miniextendr from \code{ReceiverCounter::new}

Generated by miniextendr from \code{ReceiverCounter::value}

Generated by miniextendr from \code{ReceiverCounter::inc}

Generated by miniextendr from \code{ReceiverCounter::add}

Generated by miniextendr from \code{ReceiverCounter::default_counter}
}
\usage{
ReceiverCounter

\method{$}{ReceiverCounter}(self, name)

\method{[[}{ReceiverCounter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Returns the current value.

Increments the counter by 1 and returns the new value.

Adds the given amount to the counter and returns the new value.

A static method that returns a default counter (value = 0).
}
\keyword{datasets}
