% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{Rectangle}
\alias{Rectangle}
\alias{Rectangle$new}
\alias{Rectangle$to_r}
\alias{Rectangle$with_color}
\alias{Rectangle$from_r}
\alias{$.Rectangle}
\alias{[[.Rectangle}
\title{Rectangle Class}
\usage{
Rectangle

\method{$}{Rectangle}(self, name)

\method{[[}{Rectangle}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Rectangle Class
}
