% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{Rectangle}
\alias{Rectangle}
\alias{Rectangle$new}
\alias{Rectangle$to_r}
\alias{Rectangle$with_color}
\alias{Rectangle$from_r}
\alias{$.Rectangle}
\alias{[[.Rectangle}
\title{Rectangle  Class}
\source{
Generated by miniextendr from Rust type \code{Rectangle}

Generated by miniextendr from \code{Rectangle::new}

Generated by miniextendr from \code{Rectangle::to_r}

Generated by miniextendr from \code{Rectangle::with_color}

Generated by miniextendr from \code{Rectangle::from_r}
}
\usage{
Rectangle

\method{$}{Rectangle}(self, name)

\method{[[}{Rectangle}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Rectangle  Class
}
