% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam-package.R
\docType{package}
\name{redatam-package}
\alias{redatam}
\alias{redatam-package}
\title{redatam: Import 'REDATAM' Files}
\description{
Import 'REDATAM' formats into R via the 'Open REDATAM' C++ library. The full context of this project and details about the implementation are available in \doi{10.1017/dap.2025.4} (Open Access).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/litalbarkai/open-redatam}
  \item Report bugs at \url{https://github.com/litalbarkai/open-redatam/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Authors:
\itemize{
  \item Lital Barkai
}

Other contributors:
\itemize{
  \item Ivan Krylov [contributor]
  \item Arseny Kapoulkine ('pugixml' C++ library) [contributor]
  \item  Oriental Republic of Uruguay (Aggregated census data from 2011) [data contributor]
}

}
\keyword{internal}
