% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam-package.R
\docType{package}
\name{redatam-package}
\alias{redatam}
\alias{redatam-package}
\title{redatam: Import 'REDATAM' Files}
\description{
Import 'REDATAM' files into R. 'REDATAM' is a software for the manipulation of microdata from censuses and surveys. This package provides functions to import 'REDATAM' files into R. It is based on De Grande (2016) \url{https://www.jstor.org/stable/24890658}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/litalbarkai/open-redatam}
  \item Report bugs at \url{https://github.com/litalbarkai/open-redatam/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Authors:
\itemize{
  \item Lital Barkai
}

Other contributors:
\itemize{
  \item Arseny Kapoulkine ('pugixml' C++ library) [contributor]
  \item  Republic of Ecuador (Galapagos census data) [data contributor]
}

}
\keyword{internal}
