% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{refcount_arena_roundtrip}
\alias{refcount_arena_roundtrip}
\title{refcount protect tests}
\source{
Generated by miniextendr from Rust fn \code{refcount_arena_roundtrip}
}
\usage{
refcount_arena_roundtrip()
}
\description{
refcount protect tests
}
