% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reformat.R
\name{reformat}
\alias{reformat}
\title{returns irw dataframes in the format required by the package}
\usage{
reformat(
  data,
  package = "mirt",
  id = "id",
  item = "item",
  resp = "resp",
  groups = NULL,
  timedate = NULL,
  covariates = NULL,
  levels = NULL,
  rt = NULL,
  qmatrix = NULL,
  item_groups = NULL,
  group_covariates = NULL,
  raters = NULL,
  rater_covariates = NULL,
  keep_all = F,
  facts2dummies = NULL,
  as_args_list = F,
  drop_na_vals = F,
  item_prefix = "item_",
  sep = "_",
  return_obj = "tibble",
  return_options = NULL
)
}
\arguments{
\item{data}{data.frame, tibble, or matrix: The data to be reformatted. The data should have names compatible with IRW nomenclature (see details).}

\item{package}{character: The name of the package for which the data should be reformatted. Currently supported packages are: mirt, lavaan, psych, ltm, mokken, and lme4. (see details)}

\item{id}{character: The name or names of the variables/column(s) in the data to be used in the role of subject identifier during reformatting. Default is "id".}

\item{item}{character: The name or names of the variables/column(s) in the data to be used in the role of item identifier during reformatting. Default is "item".}

\item{resp}{character: The name of the variable/column in the data to be used in the role of response variable. Default is "resp". Currently only one response variable is supported.}

\item{groups}{character or boolean: The name(s) of the variables/columns in the data to be used in the role of grouping subject-level observations OR a boolean where \code{TRUE} asks the function to attempt to automatically identify grouping variables. Default is NULL. (see details)}

\item{timedate}{character or boolean: The name(s) of the variables/columns in the data to be used in the role of longitudinal, session, or date-time OR a boolean where \code{TRUE} asks the function to attempt to automatically identify longitudinal, session, or date-time variables. Default is NULL. (see details)}

\item{covariates}{character or boolean: The name(s) of the variables/columns in the data to be used in the role of covariates for the individual/subject OR a boolean where \code{TRUE} asks the function to attempt to automatically identify covariates. Default is NULL. (see details)}

\item{levels}{character or boolean: The name(s) of the variables/columns in the data to be used in the role of level of the grouping variable for hierarchical models OR a boolean where \code{TRUE} asks the function to attempt to automatically identify levels. Default is NULL. (see details)}

\item{rt}{character or boolean: The name(s) of the variables/columns in the data to be used in the role of response time variable OR a boolean where \code{TRUE} asks the function to attempt to automatically identify response time variables. Default is NULL. (see details)}

\item{qmatrix}{character or boolean: The name(s) of the variables/columns in the data to be used in the role of Q-matrix for item response theory models OR a boolean where \code{TRUE} asks the function to attempt to automatically identify Q-matrix variables. Default is NULL. (see details)}

\item{item_groups}{character or boolean: The name(s) of the variables/columns in the data to be used in the role of grouping variable for items OR a boolean where \code{TRUE} asks the function to attempt to automatically identify item group variables. Default is NULL. (see details)}

\item{group_covariates}{character or boolean: The name(s) of the variables/columns in the data to be used in the role of covariates for the group OR a boolean where \code{TRUE} asks the function to attempt to automatically identify group covariate variables. Default is NULL. (see details)}

\item{raters}{character or boolean: The name(s) of the variables/columns in the data to be used in the role of rater variable OR a boolean where \code{TRUE} asks the function to attempt to automatically identify rater variables. Default is NULL. (see details)}

\item{rater_covariates}{character or boolean: The name(s) of the variables/columns in the data to be used in the role of covariates for the rater OR a boolean where \code{TRUE} asks the function to attempt to automatically identify rater covariate variables. Default is NULL. (see details)}

\item{keep_all}{boolean: If \code{TRUE}, the function will attempt to keep all columns in the data and identify the variable roles, regardless of whether they are used in the reformatting process. Default is \code{FALSE}.(see details)}

\item{facts2dummies}{character: The name(s) of the variables/columns in the data to be converted to dummy variables. Default is NULL. (see details)}

\item{as_args_list}{boolean: Does nothing. Currently not supported. Default is \code{FALSE}.}

\item{drop_na_vals}{boolean: If \code{TRUE}, the function will drop rows with missing values before returning. Default is \code{FALSE}.}

\item{item_prefix}{character: The prefix to be added to the item names when pivoting. Default is "item_".}

\item{sep}{character: The separator to be used when combining variables in the data. Default is "_".}

\item{return_obj}{character: The format in which the data should be returned. Options are "tibble", "data.frame", or "matrix". Default is "tibble".}

\item{return_options}{list: Currently does nothing. A list of additional options to be passed to the return object. Default is NULL. (see details)}
}
\value{
An object of class \code{irw_format} inheriting from either a data frame, tibble, or matrix in the format required by the specified package.
}
\description{
Overview ---------------------------------------------------------------------
The \code{reformat} function is designed to take a data frame and reformat it into the
format required by various R packages for analysis. The function is designed to
work with the following packages: mirt, lavaan, sem, psych, ltm, mokken, and lme4.
The function can handle a variety of data formats including wide and long formats,
as well as data with covariates, groups, item groups, raters, and more. The function
will automatically identify and convert factor columns to dummy variables if needed.
}
\details{
Currently supported packages:
mirt, lavaan, sem, psych, ltm, mokken, lme4
Note: not all functions within each package are supported. See details.
Manuals ---------------------------------------------------------------------
mirt: https://cran.r-project.org/web/packages/mirt/mirt.pdf
lavaan: https://cran.r-project.org/web/packages/lavaan/lavaan.pdf
psych: https://cran.r-project.org/web/packages/psych/psych.pdf
ltm: https://cran.r-project.org/web/packages/ltm/ltm.pdf
mokken: https://cran.r-project.org/web/packages/mokken/mokken.pdf
lme4: https://cran.r-project.org/web/packages/lme4/lme4.pdf
}
\examples{
df = data.frame(
id = rep(rep(1:3,3),2),
item = rep(rep(letters[1:3],each=3),2),
resp = sample(0:1,9*2,replace=TRUE),
cov_1 = rep(rnorm(9),2),
group = rep(c('G1','G2','G2'),each=3),
wave = rep(c(1,2),each=9)
)

}
