% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational.R
\name{rel_from_df}
\alias{rel_from_df}
\title{Convert a R data.frame to a DuckDB relation object}
\usage{
rel_from_df(con, df)
}
\arguments{
\item{con}{a DuckDB DBI connection object}

\item{df}{the data.frame}
}
\value{
the \code{duckdb_relation} object wrapping the data.frame
}
\description{
Convert a R data.frame to a DuckDB relation object
}
\examples{
con <- DBI::dbConnect(duckdb::duckdb())
rel <- rel_from_df(con, mtcars)
}
