% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{remove_symmetric_splits}
\alias{remove_symmetric_splits}
\title{Remove symmetric splits (where group1/group2 are swapped)}
\usage{
remove_symmetric_splits(splits)
}
\description{
Remove symmetric splits (where group1/group2 are swapped)
}
