% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rng_int}
\alias{rng_int}
\alias{rng_bool}
\alias{rng_range}
\alias{rng_normal}
\alias{rng_uniform}
\alias{rng_guard_test}
\alias{rng_exponential}
\alias{rng_chi_sq_approx}
\alias{rng_with_rng_test}
\alias{rng_with_interrupt}
\alias{rng_worker_uniform}
\title{Test generating n random integers in [0, max) using R's RNG}
\source{
Generated by miniextendr from Rust fn \code{rng_int}

Generated by miniextendr from Rust fn \code{rng_bool}

Generated by miniextendr from Rust fn \code{rng_range}

Generated by miniextendr from Rust fn \code{rng_normal}

Generated by miniextendr from Rust fn \code{rng_uniform}

Generated by miniextendr from Rust fn \code{rng_guard_test}

Generated by miniextendr from Rust fn \code{rng_exponential}

Generated by miniextendr from Rust fn \code{rng_chi_sq_approx}

Generated by miniextendr from Rust fn \code{rng_with_rng_test}

Generated by miniextendr from Rust fn \code{rng_with_interrupt}

Generated by miniextendr from Rust fn \code{rng_worker_uniform}
}
\usage{
rng_int(n, max)

rng_bool()

rng_range(min, max)

rng_normal(n)

rng_uniform(n)

rng_guard_test(n)

rng_exponential(n)

rng_chi_sq_approx(n, df)

rng_with_rng_test(n)

rng_with_interrupt(n)

rng_worker_uniform(n)
}
\arguments{
\item{n}{Integer number of uniform values to generate.}

\item{max}{Upper bound.}

\item{min}{Lower bound.}

\item{df}{Degrees of freedom (integer, used as count of squared normals).}
}
\description{
Test generating n random integers in [0, max) using R's RNG

rng tests

Generate a random double in [min, max) using R's RNG

Test generating n standard normal random numbers using R's RNG

Test generating n uniform random numbers using R's RNG

Test manual RNG management using RngGuard RAII pattern

Test generating n exponential random numbers using R's RNG

Generate n chi-squared-like values using the sum-of-normals method

Test manual RNG management using the with_rng closure API

Test RNG with interrupt checking enabled

Test RNG on worker thread (requires worker-thread feature)
}
