% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{RngSampler}
\alias{RngSampler}
\alias{RngSampler$new}
\alias{RngSampler$seed_hint}
\alias{RngSampler$sample_uniform}
\alias{RngSampler$sample_normal}
\alias{RngSampler$static_sample}
\alias{$.RngSampler}
\alias{[[.RngSampler}
\title{RngSampler  Class}
\format{
An object of class \code{environment} of length 5.
}
\source{
Generated by miniextendr from Rust type `RngSampler`

Generated by miniextendr from `RngSampler::new`

Generated by miniextendr from `RngSampler::seed_hint`

Generated by miniextendr from `RngSampler::sample_uniform`

Generated by miniextendr from `RngSampler::sample_normal`

Generated by miniextendr from `RngSampler::static_sample`
}
\usage{
RngSampler

\method{$}{RngSampler}(self, name)

\method{[[}{RngSampler}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Create a new RngSampler.

Get the seed hint (for testing).

Sample n uniform values using the rng attribute on a method.

Sample n normal values.

Static method with rng.
}
\keyword{datasets}
