% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpkg-package.R
\docType{package}
\name{rpkg-package}
\alias{rpkg}
\alias{rpkg-package}
\title{rpkg: My Wonderful First Package}
\description{
This my wonderful first package https://www.google.fr.
    This my wonderful first package. This my wonderful first package. 
    This my wonderful first package.
}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr}

Authors:
\itemize{
  \item Nicolas Rode \email{nicolas.rode@gmail.com}
}

}
\keyword{internal}
