% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpkg-package.R
\docType{package}
\name{rpkg-package}
\alias{rpkg}
\alias{rpkg-package}
\title{rpkg: An Example Package 0804}
\description{
A simple example package to demonstrate package creation_0804.
}
\author{
\strong{Maintainer}: Lucy van Pelt \email{lucy@example.com} (\href{https://orcid.org/LUCY-ORCID-ID}{ORCID})

}
\keyword{internal}
