% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpkg-package.R
\docType{package}
\name{rpkg-package}
\alias{rpkg}
\alias{rpkg-package}
\title{rpkg: What the Package Does (One Line, Title Case)}
\description{
gets lit.
}
\author{
\strong{Maintainer}: 

Other contributors:
\itemize{
  \item Jesse Islam \email{jesse.islam@mail.mcgill.ca} (itssa vibe)
}

}
\keyword{internal}
