% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPkg-package.R
\docType{package}
\name{RPkg-package}
\alias{RPkg}
\alias{RPkg-package}
\title{RPkg: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

}
\keyword{internal}
