% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{RPkg-package}
\alias{RPkg}
\alias{RPkg-package}
\title{RPkg: Generic Helpers}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Generic Helpers for use by R developers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig/RPkg}
  \item Report bugs at \url{https://github.com/jimbrig/RPkg/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

}
\keyword{internal}
