% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpkg-package.R
\docType{package}
\name{rpkg-package}
\alias{rpkg}
\alias{rpkg-package}
\title{rpkg: Play package for learning how to make R packages}
\description{
This package does not do anything
}
\author{
\strong{Maintainer}: Jonathan Werner \email{werner@cshl.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
