% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpkg-package.R
\docType{package}
\name{rpkg-package}
\alias{rpkg}
\alias{rpkg-package}
\title{rpkg: An implementation of the Grammar of Graphics}
\description{
An implementation of the grammar of graphics
      in R. It combines the advantages of both base and 
      lattice graphics: conditioning and shared axes
      are handled automatically, and you can still build
      up a plot step by step from multiple sources.
}
\author{
\strong{Maintainer}: Mark Leeds \email{markleeds2@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
