\name{Rpkg-package} 
\alias{Rpkg-package} 
\title{Rpkg package - package template for R packages} 
\description{The Rpkg package can be used as a template to create new packages from scratch.} 
\details{Some more details: 
The following list of objects and/or functions are available: 
\describe{ 
\item{\link[Rpkg:add]{add(x,y)}}{an illustrative add function} 
} 
} 
\examples{ 
library(Rpkg) 
add(1,2) 
} 
\author{Detlef Groth <dgroth@uni-potsdam.de>} 
\references{ 
\itemize{ 
   \item XYZ  
} 
} 
