% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpkg-package.R
\docType{package}
\name{rpkg-package}
\alias{rpkg}
\alias{rpkg-package}
\title{rpkg: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
