% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpkg-package.R
\docType{package}
\name{rpkg-package}
\alias{rpkg}
\alias{rpkg-package}
\title{rpkg: What the Package Does (One Line, Title Case)}
\description{
Absolutely nothing.
}
\author{
\strong{Maintainer}: Sara Zapata-Marin \email{sara.zapata-marin@mail.mcgill.ca} (YOUR-ORCID-ID)

}
\keyword{internal}
