% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{rpkg_altrep_helpers}
\alias{rpkg_altrep_helpers}
\alias{rpkg_doc_altrep_helpers}
\alias{altrep_compact_int}
\alias{altrep_from_doubles}
\alias{altrep_from_strings}
\alias{altrep_from_logicals}
\alias{altrep_from_raw}
\alias{altrep_from_list}
\alias{altrep_constant_int}
\alias{altrep_lazy_int_seq_is_materialized}
\title{ALTREP Helpers}
\source{
Generated by miniextendr from Rust fn \code{rpkg_doc_altrep_helpers}
}
\usage{
rpkg_doc_altrep_helpers()
}
\description{
ALTREP convenience wrappers (internal)
}
\examples{
x <- altrep_compact_int(5L, 1L, 2L)
y <- altrep_from_doubles(c(1, 2, 3))
z <- altrep_from_strings(c("a", "b"))
altrep_lazy_int_seq_is_materialized(lazy_int_seq(1L, 5L, 1L))
}
\keyword{internal}
