% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{rpkg_coercion_tests}
\alias{rpkg_coercion_tests}
\alias{test_coerce_identity}
\alias{test_coerce_widen}
\alias{test_coerce_bool_to_int}
\alias{test_coerce_via_helper}
\title{Coercion Tests}
\source{
Generated by miniextendr from Rust fn \code{test_coerce_identity}
}
\usage{
test_coerce_identity(x)
}
\description{
Coercion and RNativeType tests
}
\examples{
test_coerce_identity(1L)
test_coerce_widen(1L)
test_try_coerce_f64_to_i32(1.2)
test_coerce_attr_u16(10L)
test_per_arg_coerce_first(10L, 5L)
}
\keyword{internal}
