% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{rpkg_externalptr}
\alias{rpkg_externalptr}
\alias{extptr_counter_new}
\alias{extptr_point_new}
\alias{unsafe_C_extptr_counter_get}
\title{External Pointer Tests}
\source{
Generated by miniextendr from Rust fn \code{extptr_counter_new}
}
\usage{
extptr_counter_new(initial)
}
\arguments{
\item{initial}{Initial value for the counter.}
}
\description{
External Pointer Tests
}
\examples{
ptr <- extptr_counter_new(1L)
unsafe_C_extptr_counter_get(ptr)
unsafe_C_extptr_counter_increment(ptr)
p <- extptr_point_new(0.1, 0.2)
unsafe_C_extptr_point_get_x(p)
test_extptr_on_main_thread()
}
