% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{rpkg_externalptr}
\alias{rpkg_externalptr}
\alias{extptr_counter_new}
\alias{extptr_point_new}
\alias{unsafe_C_extptr_counter_get}
\alias{unsafe_C_extptr_counter_increment}
\alias{unsafe_C_extptr_point_get_x}
\alias{unsafe_C_extptr_point_get_y}
\alias{unsafe_C_extptr_type_mismatch_test}
\alias{unsafe_C_extptr_null_test}
\alias{unsafe_C_extptr_is_counter}
\alias{unsafe_C_extptr_is_point}
\alias{test_extptr_on_main_thread}
\title{External Pointer Tests}
\source{
Generated by miniextendr from Rust fn \code{extptr_counter_new}
}
\usage{
extptr_counter_new(initial)
}
\description{
External pointer helpers
}
\examples{
ptr <- extptr_counter_new(1L)
unsafe_C_extptr_counter_get(ptr)
unsafe_C_extptr_counter_increment(ptr)
p <- extptr_point_new(0.1, 0.2)
unsafe_C_extptr_point_get_x(p)
test_extptr_on_main_thread()
}
