% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rpkg_externalptr}
\alias{rpkg_externalptr}
\alias{extptr_counter_new}
\title{extptr_counter_new}
\source{
Generated by miniextendr from Rust fn \code{extptr_counter_new}
}
\usage{
extptr_counter_new(initial)
}
\arguments{
\item{initial}{Initial value for the counter.}
}
\description{
Create a new Counter ExternalPtr with the given initial value.
}
\examples{
ptr <- extptr_counter_new(1L)
p <- extptr_point_new(0.1, 0.2)
}
