% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rpkg_externalptr_identity}
\alias{rpkg_externalptr_identity}
\alias{ptr_identity}
\alias{ptr_pick_larger}
\title{ptr_identity}
\source{
Generated by miniextendr from Rust fn \code{ptr_identity}
}
\usage{
ptr_identity(x)
}
\arguments{
\item{x}{An ExternalPtr wrapping a PtrIdentityTest.}
}
\description{
Returns the ExternalPtr unchanged -- should preserve R identity.
}
\examples{
a <- PtrIdentityTest$new(10L)
b <- ptr_identity(a)
identical(a, b) # TRUE
}
