% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{rpkg_misc}
\alias{rpkg_misc}
\alias{underscore_it_all}
\alias{r6_standalone_add}
\title{Miscellaneous Tests}
\source{
Generated by miniextendr from Rust fn \code{underscore_it_all}
}
\usage{
underscore_it_all(private__unused0, private__unused1)
}
\description{
Miscellaneous test helpers
}
\examples{
underscore_it_all(1L, 2)
r6_standalone_add(1L, 2L)
}
\keyword{internal}
