% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rpkg_misc}
\alias{rpkg_misc}
\alias{underscore_it_all}
\title{Miscellaneous Tests}
\source{
Generated by miniextendr from Rust fn \code{underscore_it_all}
}
\usage{
underscore_it_all(unused0, unused1)
}
\arguments{
\item{unused0}{(no documentation available)}

\item{unused1}{(no documentation available)}
}
\description{
Miscellaneous test helpers
}
\examples{
underscore_it_all(1L, 2)
do_nothing()
}
