% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{rpkg_panic_tests}
\alias{rpkg_panic_tests}
\alias{add_panic}
\alias{nested_panic}
\alias{add_panic_heap}
\alias{add_r_error}
\alias{add_r_error_heap}
\title{Panic and Error Handling Tests}
\source{
Generated by miniextendr from Rust fn \code{add_panic}
}
\usage{
add_panic(left, right)
}
\description{
Panic and error handling tests (unsafe)
}
\examples{
try(add_panic(1L, 2L))
try(add_r_error(1L, 2L))
\dontrun{
unsafe_C_r_error_in_thread()
}
}
\keyword{internal}
