% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{rpkg_visibility_interrupts}
\alias{rpkg_visibility_interrupts}
\alias{invisibly_return_no_arrow}
\title{Visibility and Interrupt Tests}
\source{
Generated by miniextendr from Rust fn \code{invisibly_return_no_arrow}
}
\usage{
invisibly_return_no_arrow()
}
\description{
Visibility and interrupt checks
}
\examples{
invisibly_return_no_arrow()
force_invisible_i32()
with_interrupt_check(2L)
try(invisibly_option_return_none())
\dontrun{
unsafe_C_check_interupt_after()
}
}
