% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{S3Counter}
\alias{S3Counter}
\alias{new_s3counter}
\alias{s3_value.S3Counter}
\alias{s3_inc.S3Counter}
\alias{s3_add.S3Counter}
\alias{s3counter_default_counter}
\title{S3 counter class with constructor, accessor, and mutation methods}
\format{
An object of class \code{environment} of length 1.
}
\source{
Generated by miniextendr from Rust type `S3Counter`

Generated by miniextendr from `S3Counter::s3_value`

Generated by miniextendr from `S3Counter::s3_value`

Generated by miniextendr from `S3Counter::s3_inc`

Generated by miniextendr from `S3Counter::s3_inc`

Generated by miniextendr from `S3Counter::s3_add`

Generated by miniextendr from `S3Counter::s3_add`

Generated by miniextendr from `S3Counter::default_counter`
}
\usage{
new_s3counter(initial)

\method{s3_value}{S3Counter}(x, ...)

\method{s3_inc}{S3Counter}(x, ...)

\method{s3_add}{S3Counter}(x, amount, ...)

s3counter_default_counter()

S3Counter
}
\arguments{
\item{initial}{Initial value for the counter.}

\item{x}{An object.}

\item{...}{Additional arguments.}

\item{amount}{The amount to add to the counter.}
}
\description{
S3 generic for `s3_value`

Returns the current value (S3-specific method name to avoid conflicts).

S3 generic for `s3_inc`

Increments the counter by 1 and returns the new value.

S3 generic for `s3_add`

A static method that returns a default counter (value = 0).
}
\keyword{datasets}
