% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{S3MatchArgPoint}
\alias{S3MatchArgPoint}
\alias{new_s3matchargpoint}
\alias{label.S3MatchArgPoint}
\alias{relabel.S3MatchArgPoint}
\title{S3MatchArgPoint S3 Class}
\format{
An object of class \code{environment} of length 1.
}
\source{
Generated by miniextendr from Rust type \code{S3MatchArgPoint}

Generated by miniextendr from \code{S3MatchArgPoint::label}

Generated by miniextendr from \code{S3MatchArgPoint::label}

Generated by miniextendr from \code{S3MatchArgPoint::relabel}

Generated by miniextendr from \code{S3MatchArgPoint::relabel}
}
\usage{
new_s3matchargpoint(label = c("alpha", "beta", "gamma"))

\method{label}{S3MatchArgPoint}(x, ...)

\method{relabel}{S3MatchArgPoint}(x, mode = c("Fast", "Safe", "Debug"), ...)

S3MatchArgPoint
}
\arguments{
\item{label}{One of "alpha", "beta", "gamma".}

\item{x}{An object.}

\item{...}{Additional arguments.}

\item{mode}{One of "Fast", "Safe", "Debug".}
}
\description{
S3 generic for \code{label}

S3 generic for \code{relabel}
}
\keyword{datasets}
