% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{S3Raiser}
\alias{S3Raiser}
\alias{new_s3raiser}
\alias{s3_id.S3Raiser}
\alias{s3_raise_error.S3Raiser}
\alias{s3_raise_error_classed.S3Raiser}
\alias{s3_raise_warning.S3Raiser}
\alias{s3_raise_warning_classed.S3Raiser}
\alias{s3_raise_message.S3Raiser}
\alias{s3_raise_condition.S3Raiser}
\alias{s3_raise_condition_classed.S3Raiser}
\title{S3Raiser S3 Class}
\source{
Generated by miniextendr from Rust type \code{S3Raiser}

Generated by miniextendr from \code{S3Raiser::s3_id}

Generated by miniextendr from \code{S3Raiser::s3_id}

Generated by miniextendr from \code{S3Raiser::s3_raise_error}

Generated by miniextendr from \code{S3Raiser::s3_raise_error}

Generated by miniextendr from \code{S3Raiser::s3_raise_error_classed}

Generated by miniextendr from \code{S3Raiser::s3_raise_error_classed}

Generated by miniextendr from \code{S3Raiser::s3_raise_warning}

Generated by miniextendr from \code{S3Raiser::s3_raise_warning}

Generated by miniextendr from \code{S3Raiser::s3_raise_warning_classed}

Generated by miniextendr from \code{S3Raiser::s3_raise_warning_classed}

Generated by miniextendr from \code{S3Raiser::s3_raise_message}

Generated by miniextendr from \code{S3Raiser::s3_raise_message}

Generated by miniextendr from \code{S3Raiser::s3_raise_condition}

Generated by miniextendr from \code{S3Raiser::s3_raise_condition}

Generated by miniextendr from \code{S3Raiser::s3_raise_condition_classed}

Generated by miniextendr from \code{S3Raiser::s3_raise_condition_classed}
}
\usage{
new_s3raiser(id)

\method{s3_id}{S3Raiser}(x, ...)

\method{s3_raise_error}{S3Raiser}(x, msg, ...)

\method{s3_raise_error_classed}{S3Raiser}(x, class, msg, ...)

\method{s3_raise_warning}{S3Raiser}(x, msg, ...)

\method{s3_raise_warning_classed}{S3Raiser}(x, class, msg, ...)

\method{s3_raise_message}{S3Raiser}(x, msg, ...)

\method{s3_raise_condition}{S3Raiser}(x, msg, ...)

\method{s3_raise_condition_classed}{S3Raiser}(x, class, msg, ...)

S3Raiser
}
\arguments{
\item{id}{Numeric identifier for this raiser instance.}

\item{x}{An object.}

\item{...}{Additional arguments.}

\item{msg}{(undocumented)}

\item{class}{(undocumented)}
}
\description{
S3 generic for \code{s3_id}

S3 generic for \code{s3_raise_error}

S3 generic for \code{s3_raise_error_classed}

S3 generic for \code{s3_raise_warning}

S3 generic for \code{s3_raise_warning_classed}

S3 generic for \code{s3_raise_message}

S3 generic for \code{s3_raise_condition}

S3 generic for \code{s3_raise_condition_classed}
}
\keyword{internal}
