% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{S3TraitCounter}
\alias{S3TraitCounter}
\alias{S3TraitCounter$get_value}
\alias{S3TraitCounter$new_s3trait}
\alias{$.S3TraitCounter}
\alias{[[.S3TraitCounter}
\alias{value.S3TraitCounter}
\alias{increment.S3TraitCounter}
\alias{checked_add.S3TraitCounter}
\alias{S3TraitCounter$Counter$default_initial}
\alias{S3TraitCounter$Counter$MAX_VALUE}
\title{S3TraitCounter Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type \code{S3TraitCounter}
}
\usage{
S3TraitCounter

\method{$}{S3TraitCounter}(self, name)

\method{[[}{S3TraitCounter}(self, name)

\method{value}{S3TraitCounter}(x, ...)

\method{increment}{S3TraitCounter}(x, ...)

\method{checked_add}{S3TraitCounter}(x, n, ...)
}
\description{
Get value via inherent method.

Create a new S3 trait counter.
}
\keyword{datasets}
