% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{S3TraitCounter}
\alias{S3TraitCounter}
\alias{S3TraitCounter$get_value}
\alias{S3TraitCounter$new_s3trait}
\alias{$.S3TraitCounter}
\alias{[[.S3TraitCounter}
\title{S3TraitCounter  Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type \code{S3TraitCounter}

Generated by miniextendr from \code{S3TraitCounter::get_value}

Generated by miniextendr from \code{S3TraitCounter::new_s3trait}
}
\usage{
S3TraitCounter

\method{$}{S3TraitCounter}(self, name)

\method{[[}{S3TraitCounter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Get value via inherent method.

Create a new S3 trait counter.
}
\keyword{datasets}
