% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{get_value}
\alias{get_value}
\alias{get_value.S3TraitCounter}
\alias{s3traitcounter_new_s3trait}
\alias{S3TraitCounter}
\title{S3 generic for \code{get_value}
S3 generic for \code{get_value}}
\format{
An object of class \code{environment} of length 1.
}
\source{
Generated by miniextendr from \code{S3TraitCounter::get_value}

Generated by miniextendr from \code{S3TraitCounter::get_value}

Generated by miniextendr from \code{S3TraitCounter::new_s3trait}
}
\usage{
\method{get_value}{S3TraitCounter}(x, ...)

s3traitcounter_new_s3trait(initial)

S3TraitCounter
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Get value via inherent method.

Create a new S3 trait counter.
}
\keyword{datasets}
