% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{get_value.S3TraitCounter}
\alias{get_value.S3TraitCounter}
\alias{s3traitcounter_new_s3trait}
\alias{S3TraitCounter}
\alias{value.S3TraitCounter}
\alias{increment.S3TraitCounter}
\alias{checked_add.S3TraitCounter}
\alias{S3TraitCounter$Counter$default_initial}
\alias{S3TraitCounter$Counter$MAX_VALUE}
\title{S3 generic for \code{get_value}}
\format{
An object of class \code{environment} of length 1.
}
\source{
Generated by miniextendr from \code{S3TraitCounter::get_value}

Generated by miniextendr from \code{S3TraitCounter::get_value}

Generated by miniextendr from \code{S3TraitCounter::new_s3trait}

Generated by miniextendr from \verb{impl Counter for S3TraitCounter}
S3 generic for \code{value}

Generated by miniextendr from \verb{impl Counter for S3TraitCounter}
S3 generic for \code{increment}

Generated by miniextendr from \verb{impl Counter for S3TraitCounter}
S3 generic for \code{checked_add}
}
\usage{
\method{get_value}{S3TraitCounter}(x, ...)

s3traitcounter_new_s3trait(initial)

S3TraitCounter

\method{value}{S3TraitCounter}(x, ...)

\method{increment}{S3TraitCounter}(x, ...)

\method{checked_add}{S3TraitCounter}(x, n, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments passed to methods}

\item{initial}{(undocumented)}

\item{n}{Amount to add.}
}
\description{
S3 generic for \code{get_value}

Get value via inherent method.

Create a new S3 trait counter.
}
\keyword{datasets}
