% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{s4_has_slot_test}
\alias{s4_has_slot_test}
\title{Check if an S4 object has a named slot}
\source{
Generated by miniextendr from Rust fn \code{s4_has_slot_test}
}
\usage{
s4_has_slot_test(obj, slot_name)
}
\arguments{
\item{obj}{(no documentation available)}

\item{slot_name}{(no documentation available)}
}
\description{
Check if an S4 object has a named slot
}
