% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{class}
\name{S4Raiser}
\alias{S4Raiser}
\alias{s4_raise_error,S4Raiser-method}
\alias{s4_raise_error_classed,S4Raiser-method}
\alias{s4_raise_warning,S4Raiser-method}
\alias{s4_raise_warning_classed,S4Raiser-method}
\alias{s4_raise_message,S4Raiser-method}
\alias{s4_raise_condition,S4Raiser-method}
\alias{s4_raise_condition_classed,S4Raiser-method}
\alias{s4_id,S4Raiser-method}
\alias{S4Raiser-s4_id}
\alias{S4Raiser-s4_raise_error}
\alias{S4Raiser-s4_raise_error_classed}
\alias{S4Raiser-s4_raise_warning}
\alias{S4Raiser-s4_raise_warning_classed}
\alias{S4Raiser-s4_raise_message}
\alias{S4Raiser-s4_raise_condition}
\alias{S4Raiser-s4_raise_condition_classed}
\title{S4Raiser S4 Class}
\source{
Generated by miniextendr from Rust type `S4Raiser`

Generated by miniextendr from `S4Raiser::new`

Generated by miniextendr from `S4Raiser::s4_id`

Generated by miniextendr from `S4Raiser::s4_raise_error`

Generated by miniextendr from `S4Raiser::s4_raise_error_classed`

Generated by miniextendr from `S4Raiser::s4_raise_warning`

Generated by miniextendr from `S4Raiser::s4_raise_warning_classed`

Generated by miniextendr from `S4Raiser::s4_raise_message`

Generated by miniextendr from `S4Raiser::s4_raise_condition`

Generated by miniextendr from `S4Raiser::s4_raise_condition_classed`
}
\usage{
S4Raiser(id)
}
\arguments{
\item{id}{(undocumented)}
}
\description{
S4Raiser S4 Class
}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to Rust `S4Raiser` struct}
}}

