% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{S4TraitCounter}
\alias{S4TraitCounter}
\alias{S4TraitCounter$get_value}
\alias{S4TraitCounter$new_s4trait}
\alias{$.S4TraitCounter}
\alias{[[.S4TraitCounter}
\alias{s4_trait_Counter_value}
\alias{s4_trait_Counter_increment}
\alias{s4_trait_Counter_checked_add}
\alias{S4TraitCounter_Counter_default_initial}
\alias{S4TraitCounter_Counter_MAX_VALUE}
\title{S4TraitCounter Class}
\format{
An object of class \code{environment} of length 2.
}
\source{
Generated by miniextendr from Rust type \code{S4TraitCounter}

Generated by miniextendr from \verb{impl Counter for S4TraitCounter}
S4 generic for trait method \code{Counter::value}

Generated by miniextendr from \verb{impl Counter for S4TraitCounter}
S4 generic for trait method \code{Counter::increment}

Generated by miniextendr from \verb{impl Counter for S4TraitCounter}
S4 generic for trait method \code{Counter::checked_add}
}
\usage{
S4TraitCounter

\method{$}{S4TraitCounter}(self, name)

\method{[[}{S4TraitCounter}(self, name)

S4TraitCounter_Counter_default_initial()

S4TraitCounter_Counter_MAX_VALUE()
}
\description{
Get value via inherent method.

Create a new S4 trait counter.
}
\keyword{datasets}
