% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{s4_trait_Counter_value}
\alias{s4_trait_Counter_value}
\alias{s4_trait_Counter_value,S4TraitCounter-method}
\alias{s4_trait_Counter_increment}
\alias{s4_trait_Counter_increment,S4TraitCounter-method}
\alias{s4_trait_Counter_checked_add}
\alias{s4_trait_Counter_checked_add,S4TraitCounter-method}
\alias{S4TraitCounter_Counter_default_initial}
\alias{S4TraitCounter_Counter_MAX_VALUE}
\title{Static trait method Counter::default_initial() for S4TraitCounter}
\source{
Generated by miniextendr from \verb{impl Counter for S4TraitCounter}
S4 generic for trait method \code{Counter::value}

Generated by miniextendr from \verb{impl Counter for S4TraitCounter}
S4 generic for trait method \code{Counter::increment}

Generated by miniextendr from \verb{impl Counter for S4TraitCounter}
S4 generic for trait method \code{Counter::checked_add}
}
\usage{
\S4method{s4_trait_Counter_value}{S4TraitCounter}(x, ...)

\S4method{s4_trait_Counter_increment}{S4TraitCounter}(x, ...)

\S4method{s4_trait_Counter_checked_add}{S4TraitCounter}(x, n, ...)

S4TraitCounter_Counter_default_initial()

S4TraitCounter_Counter_MAX_VALUE()
}
\arguments{
\item{x}{A \code{S4TraitCounter} object}

\item{...}{Additional arguments passed to methods}

\item{n}{Amount to add.}
}
\description{
Static trait method Counter::default_initial() for S4TraitCounter
}
