% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{class}
\name{S4TraitCounter}
\alias{S4TraitCounter}
\alias{s4_get_value}
\alias{S4TraitCounter_new_s4trait}
\alias{s4_trait_Counter_value}
\alias{s4_trait_Counter_value,S4TraitCounter-method}
\alias{s4_trait_Counter_increment}
\alias{s4_trait_Counter_increment,S4TraitCounter-method}
\alias{s4_trait_Counter_checked_add}
\alias{s4_trait_Counter_checked_add,S4TraitCounter-method}
\alias{S4TraitCounter_Counter_default_initial}
\alias{S4TraitCounter_Counter_MAX_VALUE}
\title{S4TraitCounter S4 Class}
\source{
Generated by miniextendr from Rust type \code{S4TraitCounter}

Generated by miniextendr from \code{S4TraitCounter::s4_get_value}

Generated by miniextendr from \code{S4TraitCounter::new_s4trait}

Generated by miniextendr from \verb{impl Counter for S4TraitCounter}
S4 generic for trait method \code{Counter::value}

Generated by miniextendr from \verb{impl Counter for S4TraitCounter}
S4 generic for trait method \code{Counter::increment}

Generated by miniextendr from \verb{impl Counter for S4TraitCounter}
S4 generic for trait method \code{Counter::checked_add}
}
\usage{
S4TraitCounter_new_s4trait(initial)

\S4method{s4_trait_Counter_value}{S4TraitCounter}(x, ...)

\S4method{s4_trait_Counter_increment}{S4TraitCounter}(x, ...)

\S4method{s4_trait_Counter_checked_add}{S4TraitCounter}(x, n, ...)

S4TraitCounter_Counter_default_initial()

S4TraitCounter_Counter_MAX_VALUE()
}
\arguments{
\item{x}{A \code{S4TraitCounter} object}

\item{...}{Additional arguments passed to methods}

\item{n}{Amount to add.}
}
\description{
Get value via inherent method.

Create a new S4 trait counter.
}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to Rust \code{S4TraitCounter} struct}
}}

