% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{class}
\name{S4TraitCounter}
\alias{S4TraitCounter}
\alias{S4TraitCounter-s4_get_value}
\alias{s4_get_value,S4TraitCounter-method}
\alias{S4TraitCounter_new_s4trait}
\alias{s4_trait_Counter_value}
\alias{s4_trait_Counter_value,S4TraitCounter-method}
\alias{s4_trait_Counter_increment}
\alias{s4_trait_Counter_increment,S4TraitCounter-method}
\alias{s4_trait_Counter_checked_add}
\alias{s4_trait_Counter_checked_add,S4TraitCounter-method}
\alias{S4TraitCounter_Counter_default_initial}
\alias{S4TraitCounter_Counter_MAX_VALUE}
\title{S4TraitCounter S4 Class}
\source{
Generated by miniextendr from Rust type `S4TraitCounter`

Generated by miniextendr from `S4TraitCounter::s4_get_value`

Generated by miniextendr from `S4TraitCounter::new_s4trait`

Generated by miniextendr from `impl Counter for S4TraitCounter`
S4 generic for trait method `Counter::value`

Generated by miniextendr from `impl Counter for S4TraitCounter`
S4 generic for trait method `Counter::increment`

Generated by miniextendr from `impl Counter for S4TraitCounter`
S4 generic for trait method `Counter::checked_add`
}
\usage{
S4TraitCounter_new_s4trait(initial)

\S4method{s4_trait_Counter_value}{S4TraitCounter}(x, ...)

\S4method{s4_trait_Counter_increment}{S4TraitCounter}(x, ...)

\S4method{s4_trait_Counter_checked_add}{S4TraitCounter}(x, n, ...)

S4TraitCounter_Counter_default_initial()

S4TraitCounter_Counter_MAX_VALUE()
}
\arguments{
\item{initial}{(undocumented)}

\item{x}{A `S4TraitCounter` object}

\item{...}{Additional arguments passed to methods}

\item{n}{Amount to add.}
}
\description{
Get value via inherent method.

Create a new S4 trait counter.
}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to Rust `S4TraitCounter` struct}
}}

\keyword{internal}
