% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{S7Dog}
\alias{S7Dog}
\alias{S7Dog-dog_breed}
\alias{S7Dog$dog_breed}
\alias{S7Dog-bark}
\alias{S7Dog$bark}
\title{S7Dog S7 Class}
\source{
Generated by miniextendr from Rust type \code{S7Dog}

Generated by miniextendr from \code{S7Dog::dog_breed}

Generated by miniextendr from \code{S7Dog::bark}
}
\usage{
S7Dog(breed, .ptr = NULL)
}
\arguments{
\item{breed}{Character breed of the dog.}

\item{.ptr}{Internal pointer (used by static methods, not for direct use).}
}
\description{
Returns the dog breed.

All dogs bark.
}
