% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{S7Fahrenheit}
\alias{S7Fahrenheit}
\alias{S7Fahrenheit-value}
\alias{S7Fahrenheit$value}
\alias{S7Fahrenheit-to_celsius}
\alias{S7Fahrenheit$to_celsius}
\alias{S7Fahrenheit_from_celsius}
\alias{convert-S7Celsius-to-S7Fahrenheit}
\alias{convert}
\alias{convert-S7Fahrenheit-to-S7Celsius}
\title{S7Fahrenheit S7 Class}
\source{
Generated by miniextendr from Rust type `S7Fahrenheit`

Generated by miniextendr from `S7Fahrenheit::value`

Generated by miniextendr from `S7Fahrenheit::to_celsius`

Generated by miniextendr from `S7Fahrenheit::from_celsius`

Generated by miniextendr from `S7Fahrenheit::from_celsius`

Generated by miniextendr from `S7Fahrenheit::to_celsius`
}
\usage{
S7Fahrenheit(value, .ptr = NULL)

S7Fahrenheit_from_celsius(c)
}
\arguments{
\item{value}{Numeric temperature in Fahrenheit.}

\item{.ptr}{Internal pointer (used by static methods, not for direct use).}

\item{c}{(undocumented)}

\item{...}{Additional arguments passed to the S7 convert generic.}
}
\description{
Get the temperature value.

Convert Fahrenheit to Celsius. Uses S7::convert() for type coercion: S7::convert(fahrenheit_obj, S7Celsius)

Convert Fahrenheit from Celsius. Uses S7::convert() for type coercion: S7::convert(celsius_obj, S7Fahrenheit)
}
\keyword{internal}
