% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{S7MatchArgHolder}
\alias{S7MatchArgHolder}
\alias{S7MatchArgHolder-current}
\alias{S7MatchArgHolder$current}
\alias{S7MatchArgHolder-set}
\alias{S7MatchArgHolder$set}
\title{S7MatchArgHolder S7 Class}
\source{
Generated by miniextendr from Rust type \code{S7MatchArgHolder}

Generated by miniextendr from \code{S7MatchArgHolder::current}

Generated by miniextendr from \code{S7MatchArgHolder::set}
}
\usage{
S7MatchArgHolder(mode = c("Fast", "Safe", "Debug"), .ptr = NULL)
}
\arguments{
\item{mode}{(undocumented)}

\item{.ptr}{Internal pointer (used by static methods, not for direct use).}
}
\description{
S7MatchArgHolder S7 Class
}
