% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{S7PropInner}
\alias{S7PropInner}
\alias{S7PropInner-label}
\alias{S7PropInner$label}
\title{S7PropInner S7 Class}
\source{
Generated by miniextendr from Rust type \code{S7PropInner}

Generated by miniextendr from \code{S7PropInner::label}
}
\usage{
S7PropInner(label, .ptr = NULL)
}
\arguments{
\item{label}{Character label of the inner class.}

\item{.ptr}{Internal pointer (used by static methods, not for direct use).}
}
\description{
Returns the inner class label.
}
\keyword{internal}
