% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{S7PropOuter}
\alias{S7PropOuter}
\title{Outer S7 class with a getter returning another S7 class}
\source{
Generated by miniextendr from Rust type \code{S7PropOuter}
}
\usage{
S7PropOuter(inner_label, .ptr = NULL)
}
\arguments{
\item{inner_label}{Label used to reconstruct the inner class.}

\item{.ptr}{Internal pointer (used by static methods, not for direct use).}
}
\description{
Outer S7 class with a getter returning another S7 class
}
\keyword{internal}
\section{Additional properties}{

\describe{
\item{\code{@inner}}{Getter whose return type is another S7 class}
}}

