% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{S7PropOuter}
\alias{S7PropOuter}
\title{S7PropOuter S7 Class}
\source{
Generated by miniextendr from Rust type `S7PropOuter`
}
\usage{
S7PropOuter(inner_label, .ptr = NULL)
}
\arguments{
\item{inner_label}{Label used to reconstruct the inner class.}

\item{.ptr}{Internal pointer (used by static methods, not for direct use).}
}
\description{
S7PropOuter S7 Class
}
\keyword{internal}
