% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{S7Raiser}
\alias{S7Raiser}
\alias{S7Raiser-s7_id}
\alias{S7Raiser$s7_id}
\alias{S7Raiser-s7_raise_error}
\alias{S7Raiser$s7_raise_error}
\alias{S7Raiser-s7_raise_error_classed}
\alias{S7Raiser$s7_raise_error_classed}
\alias{S7Raiser-s7_raise_warning}
\alias{S7Raiser$s7_raise_warning}
\alias{S7Raiser-s7_raise_warning_classed}
\alias{S7Raiser$s7_raise_warning_classed}
\alias{S7Raiser-s7_raise_message}
\alias{S7Raiser$s7_raise_message}
\alias{S7Raiser-s7_raise_condition}
\alias{S7Raiser$s7_raise_condition}
\alias{S7Raiser-s7_raise_condition_classed}
\alias{S7Raiser$s7_raise_condition_classed}
\title{S7Raiser S7 Class}
\source{
Generated by miniextendr from Rust type \code{S7Raiser}

Generated by miniextendr from \code{S7Raiser::s7_id}

Generated by miniextendr from \code{S7Raiser::s7_raise_error}

Generated by miniextendr from \code{S7Raiser::s7_raise_error_classed}

Generated by miniextendr from \code{S7Raiser::s7_raise_warning}

Generated by miniextendr from \code{S7Raiser::s7_raise_warning_classed}

Generated by miniextendr from \code{S7Raiser::s7_raise_message}

Generated by miniextendr from \code{S7Raiser::s7_raise_condition}

Generated by miniextendr from \code{S7Raiser::s7_raise_condition_classed}
}
\usage{
S7Raiser(id, .ptr = NULL)
}
\arguments{
\item{id}{(undocumented)}

\item{.ptr}{Internal pointer (used by static methods, not for direct use).}
}
\description{
S7Raiser S7 Class
}
\keyword{internal}
