% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{S7Shape}
\alias{S7Shape}
\alias{shape_name}
\title{S7 abstract base shape class for inheritance testing}
\source{
Generated by miniextendr from Rust type \code{S7Shape}

Generated by miniextendr from \code{S7Shape::shape_name}
}
\usage{
S7Shape(name, .ptr = NULL)
}
\arguments{
\item{name}{(undocumented)}

\item{.ptr}{Internal pointer (used by static methods, not for direct use).}
}
\description{
Returns the shape name.
}
