% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{S7Shape}
\alias{S7Shape}
\alias{S7Shape-shape_name}
\alias{S7Shape$shape_name}
\title{S7Shape S7 Class}
\source{
Generated by miniextendr from Rust type \code{S7Shape}

Generated by miniextendr from \code{S7Shape::shape_name}
}
\usage{
S7Shape(name, .ptr = NULL)
}
\arguments{
\item{name}{Character name of the shape.}

\item{.ptr}{Internal pointer (used by static methods, not for direct use).}
}
\description{
Returns the shape name.
}
\keyword{internal}
