% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{S7TraitCounter}
\alias{S7TraitCounter}
\alias{S7TraitCounter$get_value}
\alias{S7TraitCounter$new_s7trait}
\alias{$.S7TraitCounter}
\alias{[[.S7TraitCounter}
\alias{s7_trait_Counter_value}
\alias{s7_trait_Counter_increment}
\alias{s7_trait_Counter_checked_add}
\alias{S7TraitCounter$Counter$default_initial}
\alias{S7TraitCounter$Counter$MAX_VALUE}
\title{S7TraitCounter  Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type \code{S7TraitCounter}

Generated by miniextendr from \code{S7TraitCounter::get_value}

Generated by miniextendr from \code{S7TraitCounter::new_s7trait}

Generated by miniextendr from \verb{impl Counter for S7TraitCounter}
S7 generic for trait method \code{Counter::value}

Generated by miniextendr from \verb{impl Counter for S7TraitCounter}
S7 generic for trait method \code{Counter::increment}

Generated by miniextendr from \verb{impl Counter for S7TraitCounter}
S7 generic for trait method \code{Counter::checked_add}
}
\usage{
S7TraitCounter

\method{$}{S7TraitCounter}(self, name)

\method{[[}{S7TraitCounter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Get value via inherent method.

Create a new S7 trait counter.
}
\keyword{datasets}
