% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{s7_trait_Counter_value}
\alias{s7_trait_Counter_value}
\alias{s7_trait_Counter_increment}
\alias{s7_trait_Counter_checked_add}
\alias{S7TraitCounter$Counter$default_initial}
\alias{S7TraitCounter$Counter$MAX_VALUE}
\title{Static trait method Counter::default_initial()}
\source{
Generated by miniextendr from \verb{impl Counter for S7TraitCounter}
S7 generic for trait method \code{Counter::value}

Generated by miniextendr from \verb{impl Counter for S7TraitCounter}
S7 generic for trait method \code{Counter::increment}

Generated by miniextendr from \verb{impl Counter for S7TraitCounter}
S7 generic for trait method \code{Counter::checked_add}
}
\description{
Static trait method Counter::default_initial()
}
