% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{S7TraitCounter}
\alias{S7TraitCounter}
\alias{get_value}
\alias{S7TraitCounter_new_s7trait}
\alias{s7_trait_Counter_value}
\alias{s7_trait_Counter_increment}
\alias{s7_trait_Counter_checked_add}
\alias{S7TraitCounter$Counter$default_initial}
\alias{S7TraitCounter$Counter$MAX_VALUE}
\title{S7TraitCounter S7 Class}
\source{
Generated by miniextendr from Rust type \code{S7TraitCounter}

Generated by miniextendr from \code{S7TraitCounter::get_value}

Generated by miniextendr from \code{S7TraitCounter::new_s7trait}

Generated by miniextendr from \verb{impl Counter for S7TraitCounter}
S7 generic for trait method \code{Counter::value}

Generated by miniextendr from \verb{impl Counter for S7TraitCounter}
S7 generic for trait method \code{Counter::increment}

Generated by miniextendr from \verb{impl Counter for S7TraitCounter}
S7 generic for trait method \code{Counter::checked_add}
}
\usage{
S7TraitCounter(.ptr = NULL)

S7TraitCounter_new_s7trait(initial)
}
\arguments{
\item{.ptr}{Internal pointer (used by static methods, not for direct use).}

\item{initial}{(undocumented)}
}
\description{
Get value via inherent method.

Create a new S7 trait counter.
}
