% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{ScalerR6}
\alias{ScalerR6}
\alias{r6_trait_Scaler_scale}
\alias{r6_trait_Scaler_bump}
\alias{r6_trait_Scaler_set_mode}
\title{ScalerR6 R6 Class}
\source{
Generated by miniextendr from Rust type \code{ScalerR6}

Generated by miniextendr from \verb{impl Scaler for ScalerR6}
R6 trait method \code{Scaler::scale} for ScalerR6

Generated by miniextendr from \verb{impl Scaler for ScalerR6}
R6 trait method \code{Scaler::bump} for ScalerR6

Generated by miniextendr from \verb{impl Scaler for ScalerR6}
R6 trait method \code{Scaler::set_mode} for ScalerR6
}
\usage{
r6_trait_Scaler_scale(x, x_factor)

r6_trait_Scaler_bump(x, amount)

r6_trait_Scaler_set_mode(x, mode = c("fast", "slow"))
}
\arguments{
\item{x}{A \code{ScalerR6} object}

\item{x_factor}{(undocumented)}

\item{amount}{(undocumented)}

\item{mode}{(undocumented)}
}
\description{
ScalerR6 R6 Class
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-ScalerR6-initialize}{\code{ScalerR6$new()}}
    \item \href{#method-ScalerR6-value}{\code{ScalerR6$value()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScalerR6-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-ScalerR6-initialize}{}}}
\subsection{\code{ScalerR6$new()}}{
  Create a new \code{ScalerR6}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScalerR6$new(v, .ptr = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{v}}{(no documentation available)}
      \item{\code{.ptr}}{Internal pointer (used by static methods, not for direct use).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScalerR6-value"></a>}}
\if{latex}{\out{\hypertarget{method-ScalerR6-value}{}}}
\subsection{\code{ScalerR6$value()}}{
  Method \code{value}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScalerR6$value()}
    \if{html}{\out{</div>}}
  }
}

}
