% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{SeededRng}
\alias{SeededRng}
\alias{SeededRng$new}
\alias{$.SeededRng}
\alias{[[.SeededRng}
\alias{SeededRng$RRngOps$random_f64}
\alias{SeededRng$RRngOps$random_i32}
\alias{SeededRng$RRngOps$random_bool}
\alias{SeededRng$RRngOps$gen_range_f64}
\alias{SeededRng$RRngOps$gen_range_i32}
\alias{SeededRng$RRngOps$gen_bool}
\alias{SeededRng$RRngOps$random_f64_vec}
\alias{SeededRng$RRngOps$gen_range_f64_vec}
\alias{SeededRng$RRngOps$gen_range_i32_vec}
\alias{SeededRng$RRngOps$gen_bool_vec}
\title{SeededRng  Class}
\usage{
SeededRng

\method{$}{SeededRng}(self, name)

\method{[[}{SeededRng}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
SeededRng inherent methods: constructor.
}
