% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{json_is_object}
\alias{json_is_object}
\alias{json_roundtrip}
\alias{json_to_pretty}
\alias{json_type_name}
\alias{json_object_keys}
\alias{json_serialize_point}
\title{Test whether a JSON value is an object}
\source{
Generated by miniextendr from Rust fn \code{json_is_object}

Generated by miniextendr from Rust fn \code{json_roundtrip}

Generated by miniextendr from Rust fn \code{json_to_pretty}

Generated by miniextendr from Rust fn \code{json_type_name}

Generated by miniextendr from Rust fn \code{json_object_keys}

Generated by miniextendr from Rust fn \code{json_serialize_point}
}
\usage{
json_is_object(value)

json_roundtrip(value)

json_to_pretty(value)

json_type_name(value)

json_object_keys(value)

json_serialize_point(x, y)
}
\arguments{
\item{value}{JSON object from R.}

\item{x}{X coordinate.}

\item{y}{Y coordinate.}
}
\description{
Test whether a JSON value is an object

Test JSON value roundtrip through R

Test pretty-printing a JSON value

Test getting the JSON type name of a value

Test extracting keys from a JSON object

Test serializing a Point struct to JSON string
}
