% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{SerdeRPoint}
\alias{SerdeRPoint}
\alias{SerdeRPoint$new}
\alias{SerdeRPoint$to_r}
\alias{SerdeRPoint$from_r}
\alias{$.SerdeRPoint}
\alias{[[.SerdeRPoint}
\title{SerdeRPoint  Class}
\source{
Generated by miniextendr from Rust type \code{SerdeRPoint}

Generated by miniextendr from \code{SerdeRPoint::new}

Generated by miniextendr from \code{SerdeRPoint::to_r}

Generated by miniextendr from \code{SerdeRPoint::from_r}
}
\usage{
SerdeRPoint

\method{$}{SerdeRPoint}(self, name)

\method{[[}{SerdeRPoint}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Serialize this Point to an R list.

Deserialize a Point from an R list.
}
