% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{SerdeRPoint3D}
\alias{SerdeRPoint3D}
\alias{SerdeRPoint3D$new}
\alias{SerdeRPoint3D$to_r}
\alias{SerdeRPoint3D$from_r}
\alias{$.SerdeRPoint3D}
\alias{[[.SerdeRPoint3D}
\title{SerdeRPoint3D  Class}
\source{
Generated by miniextendr from Rust type \code{SerdeRPoint3D}

Generated by miniextendr from \code{SerdeRPoint3D::new}

Generated by miniextendr from \code{SerdeRPoint3D::to_r}

Generated by miniextendr from \code{SerdeRPoint3D::from_r}
}
\usage{
SerdeRPoint3D

\method{$}{SerdeRPoint3D}(self, name)

\method{[[}{SerdeRPoint3D}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
SerdeRPoint3D  Class
}
