% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{Shape}
\alias{Shape}
\alias{Shape-shape_kind}
\alias{Shape$shape_kind}
\title{Shape S7 Class}
\source{
Generated by miniextendr from Rust type `S7OverrideShape`

Generated by miniextendr from `S7OverrideShape::shape_kind`
}
\usage{
Shape(kind, .ptr = NULL)
}
\arguments{
\item{kind}{Character kind of the shape.}

\item{.ptr}{Internal pointer (used by static methods, not for direct use).}
}
\description{
Returns the kind of shape.
}
\keyword{internal}
