% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{SharedData}
\alias{SharedData}
\title{SharedData R6 Class}
\source{
Generated by miniextendr from Rust type `SharedData`
}
\description{
SharedData R6 Class

SharedData R6 Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SharedData-new}{\code{SharedData$new()}}
\item \href{#method-SharedData-get_x}{\code{SharedData$get_x()}}
\item \href{#method-SharedData-get_y}{\code{SharedData$get_y()}}
\item \href{#method-SharedData-get_label}{\code{SharedData$get_label()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharedData-new"></a>}}
\if{latex}{\out{\hypertarget{method-SharedData-new}{}}}
\subsection{Method \code{new()}}{
Create a new `SharedData`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SharedData$new(x, y, label, .ptr = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(no documentation available)}

\item{\code{y}}{(no documentation available)}

\item{\code{label}}{(no documentation available)}

\item{\code{.ptr}}{Internal pointer (used by static methods, not for direct use).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharedData-get_x"></a>}}
\if{latex}{\out{\hypertarget{method-SharedData-get_x}{}}}
\subsection{Method \code{get_x()}}{
Method `get_x`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SharedData$get_x()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharedData-get_y"></a>}}
\if{latex}{\out{\hypertarget{method-SharedData-get_y}{}}}
\subsection{Method \code{get_y()}}{
Method `get_y`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SharedData$get_y()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharedData-get_label"></a>}}
\if{latex}{\out{\hypertarget{method-SharedData-get_label}{}}}
\subsection{Method \code{get_label()}}{
Method `get_label`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SharedData$get_label()}\if{html}{\out{</div>}}
}

}
}
