% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{SharedSimpleCounter}
\alias{SharedSimpleCounter}
\alias{SharedSimpleCounter$new}
\alias{SharedSimpleCounter$get_value}
\alias{$.SharedSimpleCounter}
\alias{[[.SharedSimpleCounter}
\alias{SharedSimpleCounter$Counter$value}
\alias{SharedSimpleCounter$Counter$increment}
\alias{SharedSimpleCounter$Counter$add}
\alias{SharedSimpleCounter$Counter$reset}
\title{SharedSimpleCounter Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type \code{SharedSimpleCounter}
}
\usage{
SharedSimpleCounter

\method{$}{SharedSimpleCounter}(self, name)

\method{[[}{SharedSimpleCounter}(self, name)
}
\description{
SharedSimpleCounter Class
}
\keyword{datasets}
