% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{SidecarEnv}
\alias{SidecarEnv}
\alias{$.SidecarEnv}
\alias{[[.SidecarEnv}
\title{SidecarEnv  Class}
\format{
An object of class \code{environment} of length 0.
}
\source{
Generated by miniextendr from Rust type \code{SidecarEnv}
}
\usage{
SidecarEnv

\method{$}{SidecarEnv}(self, name)

\method{[[}{SidecarEnv}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
SidecarEnv  Class
}
\keyword{datasets}
