% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{SidecarEnv}
\alias{SidecarEnv}
\alias{SidecarEnv_get_count}
\alias{SidecarEnv_set_count}
\alias{SidecarEnv_get_score}
\alias{SidecarEnv_set_score}
\alias{SidecarEnv_get_flag}
\alias{SidecarEnv_set_flag}
\alias{SidecarEnv_get_name}
\alias{SidecarEnv_set_name}
\alias{SidecarEnv_get_raw_slot}
\alias{SidecarEnv_set_raw_slot}
\alias{$.SidecarEnv}
\alias{[[.SidecarEnv}
\title{SidecarEnv Sidecar Accessors}
\source{
Generated by miniextendr from \verb{#[derive(ExternalPtr)]} on \code{SidecarEnv}

Generated by miniextendr from Rust type \code{SidecarEnv}
}
\usage{
SidecarEnv_get_count(x)

SidecarEnv_set_count(x, value)

SidecarEnv_get_score(x)

SidecarEnv_set_score(x, value)

SidecarEnv_get_flag(x)

SidecarEnv_set_flag(x, value)

SidecarEnv_get_name(x)

SidecarEnv_set_name(x, value)

SidecarEnv_get_raw_slot(x)

SidecarEnv_set_raw_slot(x, value)

SidecarEnv

\method{$}{SidecarEnv}(self, name)

\method{[[}{SidecarEnv}(self, name)
}
\arguments{
\item{x}{The SidecarEnv external pointer}

\item{value}{The new value to set}

\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\value{
The value of the \code{count} field

The SidecarEnv pointer (invisibly)

The value of the \code{score} field

The SidecarEnv pointer (invisibly)

The value of the \code{flag} field

The SidecarEnv pointer (invisibly)

The value of the \code{name} field

The SidecarEnv pointer (invisibly)

The value of the \code{raw_slot} field

The SidecarEnv pointer (invisibly)
}
\description{
Getter and setter functions for \verb{#[r_data]} fields on \code{SidecarEnv}.
}
