% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{SidecarEnv_get_count}
\alias{SidecarEnv_get_count}
\alias{SidecarEnv_set_count}
\alias{SidecarEnv_get_score}
\alias{SidecarEnv_set_score}
\alias{SidecarEnv_get_flag}
\alias{SidecarEnv_set_flag}
\alias{SidecarEnv_get_name}
\alias{SidecarEnv_set_name}
\alias{SidecarEnv_get_raw_slot}
\alias{SidecarEnv_set_raw_slot}
\title{Get \code{count} field from SidecarEnv}
\usage{
SidecarEnv_get_count(x)

SidecarEnv_set_count(x, value)

SidecarEnv_get_score(x)

SidecarEnv_set_score(x, value)

SidecarEnv_get_flag(x)

SidecarEnv_set_flag(x, value)

SidecarEnv_get_name(x)

SidecarEnv_set_name(x, value)

SidecarEnv_get_raw_slot(x)

SidecarEnv_set_raw_slot(x, value)
}
\arguments{
\item{x}{The SidecarEnv external pointer}

\item{value}{The new value to set}
}
\value{
The value of the \code{count} field

The SidecarEnv pointer (invisibly)

The value of the \code{score} field

The SidecarEnv pointer (invisibly)

The value of the \code{flag} field

The SidecarEnv pointer (invisibly)

The value of the \code{name} field

The SidecarEnv pointer (invisibly)

The value of the \code{raw_slot} field

The SidecarEnv pointer (invisibly)
}
\description{
Get \code{count} field from SidecarEnv

Set \code{count} field on SidecarEnv

Get \code{score} field from SidecarEnv

Set \code{score} field on SidecarEnv

Get \code{flag} field from SidecarEnv

Set \code{flag} field on SidecarEnv

Get \code{name} field from SidecarEnv

Set \code{name} field on SidecarEnv

Get \code{raw_slot} field from SidecarEnv

Set \code{raw_slot} field on SidecarEnv
}
