% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{SidecarEnv}
\alias{SidecarEnv}
\alias{SidecarEnv_get_count}
\alias{SidecarEnv_set_count}
\alias{SidecarEnv_get_score}
\alias{SidecarEnv_set_score}
\alias{SidecarEnv_get_flag}
\alias{SidecarEnv_set_flag}
\alias{SidecarEnv_get_name}
\alias{SidecarEnv_set_name}
\alias{SidecarEnv_get_raw_slot}
\alias{SidecarEnv_set_raw_slot}
\alias{$.SidecarEnv}
\alias{[[.SidecarEnv}
\title{SidecarEnv Sidecar Accessors}
\format{
An object of class \code{environment} of length 0.
}
\source{
Generated by miniextendr from `#[derive(ExternalPtr)]` on `SidecarEnv`

Generated by miniextendr from Rust type `SidecarEnv`
}
\usage{
SidecarEnv_get_count(x)

SidecarEnv_set_count(x, value)

SidecarEnv_get_score(x)

SidecarEnv_set_score(x, value)

SidecarEnv_get_flag(x)

SidecarEnv_set_flag(x, value)

SidecarEnv_get_name(x)

SidecarEnv_set_name(x, value)

SidecarEnv_get_raw_slot(x)

SidecarEnv_set_raw_slot(x, value)

SidecarEnv

\method{$}{SidecarEnv}(self, name)

\method{[[}{SidecarEnv}(self, name)
}
\arguments{
\item{x}{The SidecarEnv external pointer}

\item{value}{The new value to set}

\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\value{
The value of the `count` field

The SidecarEnv pointer (invisibly)

The value of the `score` field

The SidecarEnv pointer (invisibly)

The value of the `flag` field

The SidecarEnv pointer (invisibly)

The value of the `name` field

The SidecarEnv pointer (invisibly)

The value of the `raw_slot` field

The SidecarEnv pointer (invisibly)
}
\description{
Getter and setter functions for `#[r_data]` fields on `SidecarEnv`.
}
\keyword{datasets}
