% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{SidecarPanicky}
\alias{SidecarPanicky}
\alias{SidecarPanicky_get_boom}
\alias{SidecarPanicky_set_boom}
\alias{$.SidecarPanicky}
\alias{[[.SidecarPanicky}
\title{SidecarPanicky Sidecar Accessors}
\source{
Generated by miniextendr from \verb{#[derive(ExternalPtr)]} on \code{SidecarPanicky}

Generated by miniextendr from Rust type \code{SidecarPanicky}
}
\usage{
SidecarPanicky_get_boom(x)

SidecarPanicky_set_boom(x, value)

SidecarPanicky

\method{$}{SidecarPanicky}(self, name)

\method{[[}{SidecarPanicky}(self, name)
}
\arguments{
\item{x}{The SidecarPanicky external pointer}

\item{value}{The new value to set}

\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\value{
The value of the \code{boom} field

The SidecarPanicky pointer (invisibly)
}
\description{
Getter and setter functions for \verb{#[r_data]} fields on \code{SidecarPanicky}.

Env class registration for SidecarPanicky (panic-transport testing).
}
