% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{SidecarR6}
\alias{SidecarR6}
\alias{SidecarR6_get_value}
\alias{SidecarR6_set_value}
\alias{SidecarR6_get_label}
\alias{SidecarR6_set_label}
\title{SidecarR6 Sidecar Accessors}
\source{
Generated by miniextendr from \verb{#[derive(ExternalPtr)]} on \code{SidecarR6}
}
\usage{
SidecarR6_get_value(x)

SidecarR6_set_value(x, value)

SidecarR6_get_label(x)

SidecarR6_set_label(x, value)
}
\arguments{
\item{x}{The SidecarR6 external pointer}

\item{value}{The new value to set}
}
\value{
The value of the \code{value} field

The SidecarR6 pointer (invisibly)

The value of the \code{label} field

The SidecarR6 pointer (invisibly)
}
\description{
Getter and setter functions for \verb{#[r_data]} fields on \code{SidecarR6}.
}
