% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{SidecarRaw}
\alias{SidecarRaw}
\alias{$.SidecarRaw}
\alias{[[.SidecarRaw}
\alias{SidecarRaw_get_byte_val}
\alias{SidecarRaw_set_byte_val}
\title{SidecarRaw  Class}
\format{
An object of class \code{environment} of length 0.
}
\source{
Generated by miniextendr from Rust type \code{SidecarRaw}
}
\usage{
SidecarRaw

\method{$}{SidecarRaw}(self, name)

\method{[[}{SidecarRaw}(self, name)

SidecarRaw_get_byte_val(x)

SidecarRaw_set_byte_val(x, value)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}

\item{x}{The SidecarRaw external pointer}

\item{value}{The new value to set}
}
\value{
The value of the \code{byte_val} field

The SidecarRaw pointer (invisibly)
}
\description{
SidecarRaw  Class

Get \code{byte_val} field from SidecarRaw

Set \code{byte_val} field on SidecarRaw
}
\keyword{datasets}
