% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{SidecarRaw}
\alias{SidecarRaw}
\alias{SidecarRaw_get_byte_val}
\alias{SidecarRaw_set_byte_val}
\title{SidecarRaw Sidecar Accessors}
\source{
Generated by miniextendr from \verb{#[derive(ExternalPtr)]} on \code{SidecarRaw}
}
\usage{
SidecarRaw_get_byte_val(x)

SidecarRaw_set_byte_val(x, value)
}
\arguments{
\item{x}{The SidecarRaw external pointer}

\item{value}{The new value to set}
}
\value{
The value of the \code{byte_val} field

The SidecarRaw pointer (invisibly)
}
\description{
Getter and setter functions for \verb{#[r_data]} fields on \code{SidecarRaw}.
}
