% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{SidecarRaw_get_byte_val}
\alias{SidecarRaw_get_byte_val}
\alias{SidecarRaw_set_byte_val}
\title{Get \code{byte_val} field from SidecarRaw}
\usage{
SidecarRaw_get_byte_val(x)

SidecarRaw_set_byte_val(x, value)
}
\arguments{
\item{x}{The SidecarRaw external pointer}

\item{value}{The new value to set}
}
\value{
The value of the \code{byte_val} field

The SidecarRaw pointer (invisibly)
}
\description{
Get \code{byte_val} field from SidecarRaw

Set \code{byte_val} field on SidecarRaw
}
