% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{SidecarRaw}
\alias{SidecarRaw}
\alias{SidecarRaw_get_byte_val}
\alias{SidecarRaw_set_byte_val}
\alias{$.SidecarRaw}
\alias{[[.SidecarRaw}
\title{SidecarRaw Sidecar Accessors}
\format{
An object of class \code{environment} of length 0.
}
\source{
Generated by miniextendr from `#[derive(ExternalPtr)]` on `SidecarRaw`

Generated by miniextendr from Rust type `SidecarRaw`
}
\usage{
SidecarRaw_get_byte_val(x)

SidecarRaw_set_byte_val(x, value)

SidecarRaw

\method{$}{SidecarRaw}(self, name)

\method{[[}{SidecarRaw}(self, name)
}
\arguments{
\item{x}{The SidecarRaw external pointer}

\item{value}{The new value to set}

\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\value{
The value of the `byte_val` field

The SidecarRaw pointer (invisibly)
}
\description{
Getter and setter functions for `#[r_data]` fields on `SidecarRaw`.
}
\keyword{datasets}
