% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{SidecarRawSexp}
\alias{SidecarRawSexp}
\alias{SidecarRawSexp_get_int_vec}
\alias{SidecarRawSexp_set_int_vec}
\alias{SidecarRawSexp_get_real_vec}
\alias{SidecarRawSexp_set_real_vec}
\alias{SidecarRawSexp_get_char_vec}
\alias{SidecarRawSexp_set_char_vec}
\alias{SidecarRawSexp_get_list_val}
\alias{SidecarRawSexp_set_list_val}
\alias{SidecarRawSexp_get_func_val}
\alias{SidecarRawSexp_set_func_val}
\alias{SidecarRawSexp_get_env_val}
\alias{SidecarRawSexp_set_env_val}
\alias{$.SidecarRawSexp}
\alias{[[.SidecarRawSexp}
\title{SidecarRawSexp Sidecar Accessors}
\format{
An object of class \code{environment} of length 0.
}
\source{
Generated by miniextendr from `#[derive(ExternalPtr)]` on `SidecarRawSexp`

Generated by miniextendr from Rust type `SidecarRawSexp`
}
\usage{
SidecarRawSexp_get_int_vec(x)

SidecarRawSexp_set_int_vec(x, value)

SidecarRawSexp_get_real_vec(x)

SidecarRawSexp_set_real_vec(x, value)

SidecarRawSexp_get_char_vec(x)

SidecarRawSexp_set_char_vec(x, value)

SidecarRawSexp_get_list_val(x)

SidecarRawSexp_set_list_val(x, value)

SidecarRawSexp_get_func_val(x)

SidecarRawSexp_set_func_val(x, value)

SidecarRawSexp_get_env_val(x)

SidecarRawSexp_set_env_val(x, value)

SidecarRawSexp

\method{$}{SidecarRawSexp}(self, name)

\method{[[}{SidecarRawSexp}(self, name)
}
\arguments{
\item{x}{The SidecarRawSexp external pointer}

\item{value}{The new value to set}

\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\value{
The value of the `int_vec` field

The SidecarRawSexp pointer (invisibly)

The value of the `real_vec` field

The SidecarRawSexp pointer (invisibly)

The value of the `char_vec` field

The SidecarRawSexp pointer (invisibly)

The value of the `list_val` field

The SidecarRawSexp pointer (invisibly)

The value of the `func_val` field

The SidecarRawSexp pointer (invisibly)

The value of the `env_val` field

The SidecarRawSexp pointer (invisibly)
}
\description{
Getter and setter functions for `#[r_data]` fields on `SidecarRawSexp`.
}
\keyword{datasets}
