% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{SidecarRawSexp}
\alias{SidecarRawSexp}
\alias{$.SidecarRawSexp}
\alias{[[.SidecarRawSexp}
\alias{SidecarRawSexp_get_int_vec}
\alias{SidecarRawSexp_set_int_vec}
\alias{SidecarRawSexp_get_real_vec}
\alias{SidecarRawSexp_set_real_vec}
\alias{SidecarRawSexp_get_char_vec}
\alias{SidecarRawSexp_set_char_vec}
\alias{SidecarRawSexp_get_list_val}
\alias{SidecarRawSexp_set_list_val}
\alias{SidecarRawSexp_get_func_val}
\alias{SidecarRawSexp_set_func_val}
\alias{SidecarRawSexp_get_env_val}
\alias{SidecarRawSexp_set_env_val}
\title{SidecarRawSexp  Class}
\format{
An object of class \code{environment} of length 0.
}
\source{
Generated by miniextendr from Rust type \code{SidecarRawSexp}
}
\usage{
SidecarRawSexp

\method{$}{SidecarRawSexp}(self, name)

\method{[[}{SidecarRawSexp}(self, name)

SidecarRawSexp_get_int_vec(x)

SidecarRawSexp_set_int_vec(x, value)

SidecarRawSexp_get_real_vec(x)

SidecarRawSexp_set_real_vec(x, value)

SidecarRawSexp_get_char_vec(x)

SidecarRawSexp_set_char_vec(x, value)

SidecarRawSexp_get_list_val(x)

SidecarRawSexp_set_list_val(x, value)

SidecarRawSexp_get_func_val(x)

SidecarRawSexp_set_func_val(x, value)

SidecarRawSexp_get_env_val(x)

SidecarRawSexp_set_env_val(x, value)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}

\item{x}{The SidecarRawSexp external pointer}

\item{value}{The new value to set}
}
\value{
The value of the \code{int_vec} field

The SidecarRawSexp pointer (invisibly)

The value of the \code{real_vec} field

The SidecarRawSexp pointer (invisibly)

The value of the \code{char_vec} field

The SidecarRawSexp pointer (invisibly)

The value of the \code{list_val} field

The SidecarRawSexp pointer (invisibly)

The value of the \code{func_val} field

The SidecarRawSexp pointer (invisibly)

The value of the \code{env_val} field

The SidecarRawSexp pointer (invisibly)
}
\description{
SidecarRawSexp  Class

Get \code{int_vec} field from SidecarRawSexp

Set \code{int_vec} field on SidecarRawSexp

Get \code{real_vec} field from SidecarRawSexp

Set \code{real_vec} field on SidecarRawSexp

Get \code{char_vec} field from SidecarRawSexp

Set \code{char_vec} field on SidecarRawSexp

Get \code{list_val} field from SidecarRawSexp

Set \code{list_val} field on SidecarRawSexp

Get \code{func_val} field from SidecarRawSexp

Set \code{func_val} field on SidecarRawSexp

Get \code{env_val} field from SidecarRawSexp

Set \code{env_val} field on SidecarRawSexp
}
\keyword{datasets}
