% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{SidecarS3}
\alias{SidecarS3}
\alias{SidecarS3_get_data}
\alias{SidecarS3_set_data}
\title{SidecarS3 Sidecar Accessors}
\usage{
SidecarS3_get_data(x)

SidecarS3_set_data(x, value)

SidecarS3
}
\arguments{
\item{x}{The SidecarS3 external pointer}

\item{value}{The new value to set}
}
\value{
The value of the \code{data} field

The SidecarS3 pointer (invisibly)
}
\description{
Getter and setter functions for \verb{#[r_data]} fields on \code{SidecarS3}.
}
