% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{SidecarS3}
\alias{SidecarS3}
\alias{SidecarS3_get_data}
\alias{SidecarS3_set_data}
\title{Get \code{data} field from SidecarS3 (for S3)}
\format{
An object of class \code{environment} of length 0.
}
\usage{
SidecarS3

SidecarS3_get_data(x)

SidecarS3_set_data(x, value)
}
\arguments{
\item{x}{The SidecarS3 external pointer}

\item{value}{The new value to set}
}
\value{
The value of the \code{data} field

The SidecarS3 pointer (invisibly)
}
\description{
Get \code{data} field from SidecarS3 (for S3)

Set \code{data} field on SidecarS3 (for S3)
}
\keyword{datasets}
