% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{SidecarS3}
\alias{SidecarS3}
\alias{SidecarS3_get_data}
\alias{SidecarS3_set_data}
\title{SidecarS3 Sidecar Accessors}
\format{
An object of class \code{environment} of length 0.
}
\source{
Generated by miniextendr from `#[derive(ExternalPtr)]` on `SidecarS3`
}
\usage{
SidecarS3_get_data(x)

SidecarS3_set_data(x, value)

SidecarS3
}
\arguments{
\item{x}{The SidecarS3 external pointer}

\item{value}{The new value to set}
}
\value{
The value of the `data` field

The SidecarS3 pointer (invisibly)
}
\description{
Getter and setter functions for `#[r_data]` fields on `SidecarS3`.
}
\keyword{datasets}
