% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{SidecarS4_get_slot_int}
\alias{SidecarS4_get_slot_int}
\alias{SidecarS4_set_slot_int}
\alias{SidecarS4_get_slot_real}
\alias{SidecarS4_set_slot_real}
\alias{SidecarS4_get_slot_str}
\alias{SidecarS4_set_slot_str}
\title{Get \code{slot_int} field from SidecarS4 (for S4)}
\usage{
SidecarS4_get_slot_int(x)

SidecarS4_set_slot_int(x, value)

SidecarS4_get_slot_real(x)

SidecarS4_set_slot_real(x, value)

SidecarS4_get_slot_str(x)

SidecarS4_set_slot_str(x, value)
}
\arguments{
\item{x}{The SidecarS4 external pointer}

\item{value}{The new value to set}
}
\value{
The value of the \code{slot_int} field

The SidecarS4 pointer (invisibly)

The value of the \code{slot_real} field

The SidecarS4 pointer (invisibly)

The value of the \code{slot_str} field

The SidecarS4 pointer (invisibly)
}
\description{
Get \code{slot_int} field from SidecarS4 (for S4)

Set \code{slot_int} field on SidecarS4 (for S4)

Get \code{slot_real} field from SidecarS4 (for S4)

Set \code{slot_real} field on SidecarS4 (for S4)

Get \code{slot_str} field from SidecarS4 (for S4)

Set \code{slot_str} field on SidecarS4 (for S4)
}
