% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{class}
\name{SidecarS4}
\alias{SidecarS4}
\alias{SidecarS4_get_slot_int}
\alias{SidecarS4_set_slot_int}
\alias{SidecarS4_get_slot_real}
\alias{SidecarS4_set_slot_real}
\alias{SidecarS4_get_slot_str}
\alias{SidecarS4_set_slot_str}
\title{SidecarS4 Sidecar Accessors}
\source{
Generated by miniextendr from `#[derive(ExternalPtr)]` on `SidecarS4`

Generated by miniextendr from Rust type `SidecarS4`
}
\usage{
SidecarS4_get_slot_int(x)

SidecarS4_set_slot_int(x, value)

SidecarS4_get_slot_real(x)

SidecarS4_set_slot_real(x, value)

SidecarS4_get_slot_str(x)

SidecarS4_set_slot_str(x, value)
}
\arguments{
\item{x}{The SidecarS4 external pointer}

\item{value}{The new value to set}
}
\value{
The value of the `slot_int` field

The SidecarS4 pointer (invisibly)

The value of the `slot_real` field

The SidecarS4 pointer (invisibly)

The value of the `slot_str` field

The SidecarS4 pointer (invisibly)
}
\description{
Getter and setter functions for `#[r_data]` fields on `SidecarS4`.
}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to Rust `SidecarS4` struct}
}}

