% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{SidecarS7}
\alias{SidecarS7}
\alias{SidecarS7_get_prop_int}
\alias{SidecarS7_set_prop_int}
\alias{SidecarS7_get_prop_flag}
\alias{SidecarS7_set_prop_flag}
\alias{SidecarS7_get_prop_name}
\alias{SidecarS7_set_prop_name}
\title{SidecarS7 Sidecar Accessors}
\source{
Generated by miniextendr from \verb{#[derive(ExternalPtr)]} on \code{SidecarS7}
}
\usage{
SidecarS7_get_prop_int(x)

SidecarS7_set_prop_int(x, value)

SidecarS7_get_prop_flag(x)

SidecarS7_set_prop_flag(x, value)

SidecarS7_get_prop_name(x)

SidecarS7_set_prop_name(x, value)
}
\arguments{
\item{x}{The SidecarS7 external pointer}

\item{value}{The new value to set}
}
\value{
The value of the \code{prop_int} field

The SidecarS7 pointer (invisibly)

The value of the \code{prop_flag} field

The SidecarS7 pointer (invisibly)

The value of the \code{prop_name} field

The SidecarS7 pointer (invisibly)
}
\description{
Getter and setter functions for \verb{#[r_data]} fields on \code{SidecarS7}.
}
