% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{SidecarS7}
\alias{SidecarS7}
\alias{SidecarS7_get_prop_int}
\alias{SidecarS7_set_prop_int}
\alias{SidecarS7_get_prop_flag}
\alias{SidecarS7_set_prop_flag}
\alias{SidecarS7_get_prop_name}
\alias{SidecarS7_set_prop_name}
\title{SidecarS7 S7 Class}
\source{
Generated by miniextendr from Rust type \code{SidecarS7}
}
\usage{
SidecarS7(.ptr = NULL)

SidecarS7_get_prop_int(x)

SidecarS7_set_prop_int(x, value)

SidecarS7_get_prop_flag(x)

SidecarS7_set_prop_flag(x, value)

SidecarS7_get_prop_name(x)

SidecarS7_set_prop_name(x, value)
}
\arguments{
\item{.ptr}{Internal pointer (used by static methods, not for direct use).}

\item{x}{The SidecarS7 external pointer}

\item{value}{The new value to set}
}
\value{
The value of the \code{prop_int} field

The SidecarS7 pointer (invisibly)

The value of the \code{prop_flag} field

The SidecarS7 pointer (invisibly)

The value of the \code{prop_name} field

The SidecarS7 pointer (invisibly)
}
\description{
SidecarS7 S7 Class

Get \code{prop_int} field from SidecarS7 (for S7)

Set \code{prop_int} field on SidecarS7 (for S7)

Get \code{prop_flag} field from SidecarS7 (for S7)

Set \code{prop_flag} field on SidecarS7 (for S7)

Get \code{prop_name} field from SidecarS7 (for S7)

Set \code{prop_name} field on SidecarS7 (for S7)
}
