% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{vec_ptype2.SidecarVctrs.SidecarVctrs}
\alias{vec_ptype2.SidecarVctrs.SidecarVctrs}
\alias{vec_cast.SidecarVctrs.SidecarVctrs}
\alias{SidecarVctrs_get_vec_data}
\alias{SidecarVctrs_set_vec_data}
\alias{SidecarVctrs_get_vec_label}
\alias{SidecarVctrs_set_vec_label}
\title{Get \code{vec_data} field from SidecarVctrs (for vctrs)}
\usage{
\method{vec_ptype2}{SidecarVctrs.SidecarVctrs}(x, y, ...)

\method{vec_cast}{SidecarVctrs.SidecarVctrs}(x, to, ...)

SidecarVctrs_get_vec_data(x)

SidecarVctrs_set_vec_data(x, value)

SidecarVctrs_get_vec_label(x)

SidecarVctrs_set_vec_label(x, value)
}
\arguments{
\item{x}{The SidecarVctrs external pointer}

\item{y}{A SidecarVctrs vector.}

\item{...}{Additional arguments (unused).}

\item{to}{A SidecarVctrs prototype.}

\item{value}{The new value to set}
}
\value{
The value of the \code{vec_data} field

The SidecarVctrs pointer (invisibly)

The value of the \code{vec_label} field

The SidecarVctrs pointer (invisibly)
}
\description{
Get \code{vec_data} field from SidecarVctrs (for vctrs)

Set \code{vec_data} field on SidecarVctrs (for vctrs)

Get \code{vec_label} field from SidecarVctrs (for vctrs)

Set \code{vec_label} field on SidecarVctrs (for vctrs)
}
