% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{SimpleCounter}
\alias{SimpleCounter}
\alias{SimpleCounter$get_value}
\alias{SimpleCounter$trait_add}
\alias{SimpleCounter$new_counter}
\alias{$.SimpleCounter}
\alias{[[.SimpleCounter}
\alias{SimpleCounter$Counter$value}
\alias{SimpleCounter$Counter$increment}
\alias{SimpleCounter$Counter$checked_add}
\alias{SimpleCounter$Counter$default_initial}
\alias{SimpleCounter$Counter$MAX_VALUE}
\title{SimpleCounter  Class}
\usage{
SimpleCounter

\method{$}{SimpleCounter}(self, name)

\method{[[}{SimpleCounter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Get the counter's value.

Add to the counter using the trait method.

Create a new counter with given initial value.
}
