% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{SimpleCounter}
\alias{SimpleCounter}
\alias{SimpleCounter$get_value}
\alias{SimpleCounter$trait_add}
\alias{SimpleCounter$new_counter}
\alias{$.SimpleCounter}
\alias{[[.SimpleCounter}
\title{SimpleCounter  Class}
\format{
An object of class \code{environment} of length 4.
}
\source{
Generated by miniextendr from Rust type \code{SimpleCounter}

Generated by miniextendr from \code{SimpleCounter::get_value}

Generated by miniextendr from \code{SimpleCounter::trait_add}

Generated by miniextendr from \code{SimpleCounter::new_counter}
}
\usage{
SimpleCounter

\method{$}{SimpleCounter}(self, name)

\method{[[}{SimpleCounter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Get the counter's value.

Add to the counter using the trait method.

Create a new counter with given initial value.
}
\keyword{datasets}
