% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{SimpleCounter}
\alias{SimpleCounter}
\alias{SimpleCounter$get_value}
\alias{SimpleCounter$trait_add}
\alias{SimpleCounter$new_counter}
\alias{$.SimpleCounter}
\alias{[[.SimpleCounter}
\alias{SimpleCounter$Counter$value}
\alias{SimpleCounter$Counter$increment}
\alias{SimpleCounter$Counter$add}
\title{SimpleCounter Class}
\format{
An object of class \code{environment} of length 4.
}
\source{
Generated by miniextendr from Rust type \code{SimpleCounter}
}
\usage{
SimpleCounter

\method{$}{SimpleCounter}(self, name)

\method{[[}{SimpleCounter}(self, name)
}
\description{
Get the counter's value.

Add to the counter using the trait method.

Create a new counter with given initial value.
}
\keyword{datasets}
