% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql}
\alias{sql}
\title{Run a SQL query}
\usage{
sql(sql, conn = default_connection())
}
\arguments{
\item{sql}{A SQL string}

\item{conn}{An optional connection, defaults to built-in default}
}
\value{
A data frame with the query result
}
\description{
\code{sql()} runs an arbitrary SQL query and returns a data.frame the query results
}
\examples{
print(duckdb::sql("SELECT 42"))
}
