% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartitions.R
\name{standardize_multinomial_columns}
\alias{standardize_multinomial_columns}
\title{Standardize multinomial results from list to data frame with consistent column structure}
\usage{
standardize_multinomial_columns(row_list)
}
\arguments{
\item{row_list}{List of row-lists with variable multinomial columns}

\item{max_paths}{Maximum number of paths to standardize to}
}
\description{
Standardize multinomial results from list to data frame with consistent column structure
}
