% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_choose_2_analysis.R
\name{standardize_pairwise_columns}
\alias{standardize_pairwise_columns}
\title{Convert pairwise analysis results to standardized data frame}
\usage{
standardize_pairwise_columns(pairwise_list)
}
\arguments{
\item{pairwise_list}{A list of row-lists produced by \code{find_diff_and_ref_exparts_pairwise},
each containing fields \code{gene}, \code{source}, \code{sink}, \code{path1_idx},
\code{path2_idx}, \code{ref_ex_part}, \code{setdiff1}, \code{setdiff2},
\code{transcripts1}, \code{transcripts2}, \code{path1}, and \code{path2}.}
}
\description{
Convert pairwise analysis results to standardized data frame
}
\examples{
pairwise_list <- list(
  list(gene="GENE1", source="3", sink="7",
       path1_idx=1L, path2_idx=2L,
       ref_ex_part="E001", setdiff1="E002", setdiff2="E003",
       transcripts1="TX1", transcripts2="TX2",
       path1="3-5-7", path2="3-6-7")
)
df <- grase::standardize_pairwise_columns(pairwise_list)
colnames(df)
}
