% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{streaming_int_range}
\alias{streaming_int_range}
\alias{streaming_real_squares}
\title{Create a streaming integer ALTREP `1..=n`}
\source{
Generated by miniextendr from Rust fn `streaming_int_range`

Generated by miniextendr from Rust fn `streaming_real_squares`
}
\usage{
streaming_int_range(n)

streaming_real_squares(n)
}
\arguments{
\item{n}{(no documentation available)}
}
\description{
Create a streaming integer ALTREP `1..=n`

Create a streaming real ALTREP `1^2, 2^2, ..., n^2`
}
