% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{streaming_int_range}
\alias{streaming_int_range}
\title{Create a streaming integer ALTREP \link{1..=n}}
\source{
Generated by miniextendr from Rust fn \code{streaming_int_range}
}
\usage{
streaming_int_range(n)
}
\arguments{
\item{n}{(no documentation available)}
}
\description{
Create a streaming integer ALTREP \link{1..=n}
}
