% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{streaming_real_squares}
\alias{streaming_real_squares}
\title{Create a streaming real ALTREP \link{1^2, 2^2, ..., n^2}}
\source{
Generated by miniextendr from Rust fn \code{streaming_real_squares}
}
\usage{
streaming_real_squares(n)
}
\arguments{
\item{n}{(no documentation available)}
}
\description{
Create a streaming real ALTREP \link{1^2, 2^2, ..., n^2}
}
