% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{StrictCounter}
\alias{StrictCounter}
\title{StrictCounter R6 Class}
\source{
Generated by miniextendr from Rust type \code{StrictCounter}
}
\description{
StrictCounter R6 Class

StrictCounter R6 Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StrictCounter-new}{\code{StrictCounter$new()}}
\item \href{#method-StrictCounter-get_value}{\code{StrictCounter$get_value()}}
\item \href{#method-StrictCounter-add}{\code{StrictCounter$add()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrictCounter-new"></a>}}
\if{latex}{\out{\hypertarget{method-StrictCounter-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{StrictCounter}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrictCounter$new(value, .ptr = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{(no documentation available)}

\item{\code{.ptr}}{Internal pointer (used by static methods, not for direct use).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrictCounter-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-StrictCounter-get_value}{}}}
\subsection{Method \code{get_value()}}{
Method \code{get_value}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrictCounter$get_value()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrictCounter-add"></a>}}
\if{latex}{\out{\hypertarget{method-StrictCounter-add}{}}}
\subsection{Method \code{add()}}{
Method \code{add}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrictCounter$add(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(no documentation available)}
}
\if{html}{\out{</div>}}
}
}
}
