% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{style_and_plot}
\alias{style_and_plot}
\title{Style and plot}
\usage{
style_and_plot(g, gene, outdir)
}
\arguments{
\item{gene}{Character string. Gene identifier.}

\item{outdir}{Character string. Path to output directory where the PDF plot
is written.}
}
\description{
Style and plot
}
