% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{sum}
\alias{fm.sum}
\alias{sum}
\alias{sum,fm-method}
\alias{sum,fmV-method}
\title{Sum of Vector Elements}
\usage{
fm.sum(x, lazy = FALSE)

\S4method{sum}{fm}(x, ..., na.rm = FALSE)

\S4method{sum}{fmV}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{a FlashR vector or matrix.}

\item{lazy}{indicates whether or not to evaluate it lazily.}

\item{...}{zero or more vectors or matrices.}

\item{na.rm}{logical. Should missing values (including \code{NaN}) be removed?}
}
\value{
The sum.
}
\description{
\code{sum} returns the sum of all the values in its arguments.
\code{fm.sum} returns the sum of all the values in the input object. It can
evaluate the summation lazily.
}
\examples{
mat <- fm.runif.matrix(100, 10)
sum(mat)
fm.sum(mat)
}

