% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sylcount-package.r
\docType{package}
\name{sylcount-package}
\alias{sylcount-package}
\title{sylcount: Syllable Counting and Readability Measurements}
\description{
TODO
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}

