% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sylcount-package.r
\docType{package}
\name{sylcount-package}
\alias{sylcount-package}
\title{sylcount: Syllable Counting and Readability Measurements}
\description{
An English language syllable counter, plus readability score
measure-er.  The package has been carefully optimized and should be very
efficient, both in terms of run time performance and memory consumption.
The main methods 'vectorized' by document, and scores for multiple documents
are computed in parallel via 'OpenMP'.
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}
