% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_collect_strings_numbered}
\alias{test_collect_strings_numbered}
\title{test collect strings numbered}
\source{
Generated by miniextendr from Rust fn \code{test_collect_strings_numbered}
}
\usage{
test_collect_strings_numbered(n)
}
\arguments{
\item{n}{(no documentation available)}
}
\description{
test collect strings numbered
}
