% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_df_sql_query}
\alias{test_df_sql_query}
\title{test df sql query}
\source{
Generated by miniextendr from Rust fn \code{test_df_sql_query}
}
\usage{
test_df_sql_query(df, sql)
}
\description{
test df sql query
}
