% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{test_fromstr_ip}
\alias{test_fromstr_ip}
\title{test fromstr ip}
\source{
Generated by miniextendr from Rust fn \code{test_fromstr_ip}
}
\usage{
test_fromstr_ip(addr)
}
\arguments{
\item{addr}{(no documentation available)}
}
\description{
test fromstr ip
}
