% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exontest_functions.R
\name{test_model_glmmTMB_EB}
\alias{test_model_glmmTMB_EB}
\title{Test differential exon usage with a beta-binomial glmmTMB model using empirical Bayes fixed dispersion.}
\usage{
test_model_glmmTMB_EB(dd)
}
\arguments{
\item{dd}{A data frame of exon count data.}
}
\description{
Test differential exon usage with a beta-binomial glmmTMB model using empirical Bayes fixed dispersion.
}
\examples{
\dontrun{
splitcnts <- read.table("bipartition.internal.exoncnt.combined.txt",
                        header = TRUE, row.names = NULL)
phi_df <- read.table("phi.glmmtmb.txt", header = TRUE, row.names = NULL)
phi_table <- grase::moderate_phi_log_scale(phi_df)
splitcnts_eb <- dplyr::left_join(splitcnts, phi_table, by = c("gene", "event"))
splitcnts_eb$groups <- factor(splitcnts_eb$groups)
grouped_eb <- grase::group_by_event(splitcnts_eb, "diff", "n")
result <- grase::test_model_glmmTMB_EB(grouped_eb[[1]])
result
}
}
