% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threepsFit.R
\name{threepsFit}
\alias{threepsFit}
\title{Fits a three pool model with series structure to a time-series}
\usage{
threepsFit(timeSeries, initialCarbon, inipars = c(1, 0.5, 0.5, 0.5, 0.5,
  0.5, 0.5))
}
\arguments{
\item{timeSeries}{A time series of respiration values over time}

\item{initialCarbon}{The initial amount of carbon in units that correspond to the time series data}

\item{inipars}{vector of parameter values for the initial search of the optimization algorithm}
}
\value{
R list with an FME model object, a SoilR model object, and the AIC value
}
\description{
Fits a three pool model with series structure to a time-series
}
\examples{
db=loadEntries(path="~/sidb/data/")
a=threepsFit(timeSeries = db[["Crow2019a"]]$timeSeries[,c(1,5)], initialCarbon=db[["Crow2019a"]]$initConditions[4,"carbonMean"]*10000)
}
