% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twopfFit.R
\name{twopfFit}
\alias{twopfFit}
\title{Fits a two pool model with feedback structure to a time-series}
\usage{
twopfFit(timeSeries, initialCarbon, inipars = c(1, 0.5, 0.5, 0.5, 0.3))
}
\arguments{
\item{timeSeries}{A time series of respiration values over time}

\item{initialCarbon}{The initial amount of carbon in units that correspond to the time series data}

\item{inipars}{vector of parameter values for the initial search of the optimization algorithm}
}
\value{
R list with an FME model object, a SoilR model object, and the AIC value
}
\description{
Fits a two pool model with feedback structure to a time-series
}
\examples{
db=loadEntries(path="~/sidb/data/")
incubation=db[["Crow2019a"]]
b=twopfFit(timeSeries = incubation$timeSeries[,c(1,79)],
initialCarbon=incubation$initConditions[78,"carbonMean"]*10000,
inipars=c(0.005, 0.00001, 0.1, 0.01, 0.01))
}
