% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twopsFit.R
\name{twopsFit}
\alias{twopsFit}
\title{Fits a two pool model with series structure to a time-series}
\usage{
twopsFit(timeSeries, initialCarbon, inipars = c(-1, -0.5, 0.5, 0.3))
}
\arguments{
\item{timeSeries}{A time series of respiration values over time}

\item{initialCarbon}{The initial amount of carbon in units that correspond to the time series data}

\item{inipars}{vector of parameter values for the initial search of the optimization algorithm}
}
\value{
R list with an FME model object, a SoilR model object, and the AIC value
}
\description{
Fits a two pool model with series structure to a time-series
}
\examples{
db=loadEntries(path="~/sidb/clean/")
b=twopsFit(timeSeries = db[[20]]$timeSeries[,1:2], initialCarbon=db[[20]]$initConditions[1,"carbonMean"]*10)
}
