% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{txpath_from_edgeattr}
\alias{txpath_from_edgeattr}
\title{Generate transcript-paths from igraph edge attributes}
\usage{
txpath_from_edgeattr(g, type = "exon")
}
\description{
Generate transcript-paths from igraph edge attributes
}
