% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{TypeA}
\alias{TypeA}
\alias{TypeA$new}
\alias{TypeA$get_val}
\alias{$.TypeA}
\alias{[[.TypeA}
\title{TypeA Class}
\usage{
TypeA

\method{$}{TypeA}(self, name)

\method{[[}{TypeA}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
TypeA Class
}
