% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{TypeA}
\alias{TypeA}
\alias{TypeA$new}
\alias{TypeA$get_val}
\alias{$.TypeA}
\alias{[[.TypeA}
\title{TypeA  Class}
\format{
An object of class \code{environment} of length 2.
}
\source{
Generated by miniextendr from Rust type `TypeA`

Generated by miniextendr from `TypeA::new`

Generated by miniextendr from `TypeA::get_val`
}
\usage{
TypeA

\method{$}{TypeA}(self, name)

\method{[[}{TypeA}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
TypeA  Class
}
\keyword{datasets}
