% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{TypeB}
\alias{TypeB}
\alias{TypeB$new}
\alias{TypeB$get_val}
\alias{$.TypeB}
\alias{[[.TypeB}
\title{TypeB  Class}
\format{
An object of class \code{environment} of length 2.
}
\source{
Generated by miniextendr from Rust type \code{TypeB}

Generated by miniextendr from \code{TypeB::new}

Generated by miniextendr from \code{TypeB::get_val}
}
\usage{
TypeB

\method{$}{TypeB}(self, name)

\method{[[}{TypeB}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
TypeB  Class
}
\keyword{datasets}
