% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashR_base.R
\docType{methods}
\name{typeof}
\alias{typeof}
\alias{typeof,fm-method}
\alias{typeof,fmV-method}
\title{The Type of an Object}
\usage{
\S4method{typeof}{fm}(x)

\S4method{typeof}{fmV}(x)
}
\arguments{
\item{x}{a FlashR object.}
}
\value{
A character string. Current values are "logical", "integer",
"double".
}
\description{
\code{typeof} determines the R type of an FlashR object.
}
\examples{
mat <- fm.runif.matrix(100, 10)
typeof(mat)
}

