% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{unsafe_C_altrep_sexp_check}
\alias{unsafe_C_altrep_sexp_check}
\title{Check if a SEXP is ALTREP and return info about it}
\source{
Generated by miniextendr from Rust fn \code{C_altrep_sexp_check}
}
\usage{
unsafe_C_altrep_sexp_check(x)
}
\arguments{
\item{x}{A SEXP to inspect.}
}
\description{
Check if a SEXP is ALTREP and return info about it
}
