% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{update_txpaths_after_bubble_collapse2}
\alias{update_txpaths_after_bubble_collapse2}
\title{Update transcript paths after collapsing a bubble}
\usage{
update_txpaths_after_bubble_collapse2(g, tx_list, epath)
}
\arguments{
\item{g}{An igraph splicing graph}

\item{tx_list}{List of transcript groups to update (from find_tx_with_epath)}

\item{epath}{Edge paths to keep (list of edge name vectors)}
}
\value{
The updated igraph graph
}
\description{
After removing edges from a collapsed bubble, reassign the affected
transcripts to the kept paths by setting edge attributes.
}
\examples{
\dontrun{
g <- igraph::read_graph("ENSG00000000003.dexseq.graphml", format = "graphml")
g <- grase::set_edge_names(g)
trans <- grase::transcripts_from_igraph(g)
epaths <- list(c("3-5:ex", "5-7:ex"))
tx_list <- grase::find_tx_with_epath(g, trans, epaths)
g <- grase::update_txpaths_after_bubble_collapse2(g, tx_list, epaths)
}
}
