% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focalexons.R
\name{valid_partitions_slow}
\alias{valid_partitions_slow}
\title{find valid partitions that represents the hierarchical relationship between paths}
\usage{
valid_partitions_slow(g, max_powerset)
}
\description{
find valid partitions that represents the hierarchical relationship between paths
}
