% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{validate_class_args}
\alias{validate_class_args}
\title{Validate with class checking}
\source{
Generated by miniextendr from Rust fn \code{validate_class_args}
}
\usage{
validate_class_args(...)
}
\arguments{
\item{...}{Named arguments: \code{data} (data.frame).}
}
\description{
Validate with class checking
}
