% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{validate_with_attribute}
\alias{validate_with_attribute}
\title{Test the \verb{#[miniextendr(dots = typed_list!(...))]} attribute syntax}
\source{
Generated by miniextendr from Rust fn \code{validate_with_attribute}
}
\usage{
validate_with_attribute(...)
}
\arguments{
\item{...}{Named arguments: \code{x} (numeric), \code{y} (numeric).}
}
\description{
Test the \verb{#[miniextendr(dots = typed_list!(...))]} attribute syntax
}
