% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{VctrsMatchArgScale}
\alias{VctrsMatchArgScale}
\alias{new_vctrsmatchargscale}
\alias{vec_ptype_abbr.VctrsMatchArgScale}
\alias{vec_ptype2.VctrsMatchArgScale.VctrsMatchArgScale}
\alias{vec_cast.VctrsMatchArgScale.VctrsMatchArgScale}
\title{VctrsMatchArgScale vctrs S3 Class}
\source{
Generated by miniextendr from Rust type \code{VctrsMatchArgScale}
}
\usage{
new_vctrsmatchargscale(mode = c("Fast", "Safe", "Debug"))

\method{vec_ptype_abbr}{VctrsMatchArgScale}(x, ...)

\method{vec_ptype2}{VctrsMatchArgScale.VctrsMatchArgScale}(x, y, ...)

\method{vec_cast}{VctrsMatchArgScale.VctrsMatchArgScale}(x, to, ...)
}
\arguments{
\item{mode}{One of "Fast", "Safe", "Debug".}

\item{x}{A VctrsMatchArgScale vector to cast.}

\item{...}{Additional arguments (unused).}

\item{y}{A VctrsMatchArgScale vector.}

\item{to}{A VctrsMatchArgScale prototype.}
}
\description{
VctrsMatchArgScale vctrs S3 Class
}
