% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{vec_cast.double.percent}
\alias{vec_cast.double.percent}
\title{Cast from percent to double}
\source{
Generated by miniextendr from Rust fn \code{vec_cast_double_percent}
}
\usage{
\method{vec_cast}{double.percent}(x, to, ...)
}
\arguments{
\item{x}{(no documentation available)}

\item{to}{(no documentation available)}

\item{dots}{(no documentation available)}
}
\description{
Cast from percent to double
}
