% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{vec_cast.percent.double}
\alias{vec_cast.percent.double}
\title{Cast from double to percent}
\source{
Generated by miniextendr from Rust fn \code{vec_cast_percent_double}
}
\usage{
\method{vec_cast}{percent.double}(x, to, ...)
}
\arguments{
\item{x}{(no documentation available)}

\item{to}{(no documentation available)}

\item{dots}{(no documentation available)}
}
\description{
Cast from double to percent
}
