% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{vec_cast.percent.percent}
\alias{vec_cast.percent.percent}
\title{Self-cast (percent -> percent is identity)}
\source{
Generated by miniextendr from Rust fn \code{vec_cast_percent_percent}
}
\usage{
\method{vec_cast}{percent.percent}(x, to, ...)
}
\arguments{
\item{x}{(no documentation available)}

\item{to}{(no documentation available)}

\item{dots}{(no documentation available)}
}
\description{
Self-cast (percent -> percent is identity)
}
