% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{vec_proxy.percent}
\alias{vec_proxy.percent}
\title{Get the proxy for subsetting operations}
\source{
Generated by miniextendr from Rust fn \code{vec_proxy_percent}
}
\usage{
\method{vec_proxy}{percent}(x, ...)
}
\arguments{
\item{x}{(no documentation available)}

\item{dots}{(no documentation available)}
}
\description{
Get the proxy for subsetting operations
}
