% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{vec_ptype2.double.percent}
\alias{vec_ptype2.double.percent}
\title{Coercion: double + percent = percent (symmetric)}
\source{
Generated by miniextendr from Rust fn \code{vec_ptype2_double_percent}
}
\usage{
\method{vec_ptype2}{double.percent}(x, y, ...)
}
\arguments{
\item{x}{(no documentation available)}

\item{y}{(no documentation available)}

\item{dots}{(no documentation available)}
}
\description{
Coercion: double + percent = percent (symmetric)
}
