% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{vec_ptype2.percent.double}
\alias{vec_ptype2.percent.double}
\title{Coercion: double + percent = percent}
\source{
Generated by miniextendr from Rust fn \code{vec_ptype2_percent_double}
}
\usage{
\method{vec_ptype2}{percent.double}(x, y, ...)
}
\arguments{
\item{x}{(no documentation available)}

\item{y}{(no documentation available)}

\item{dots}{(no documentation available)}
}
\description{
Coercion: double + percent = percent
}
