% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{vec_ptype2.percent.percent}
\alias{vec_ptype2.percent.percent}
\title{Self-coercion prototype (percent + percent = percent)}
\source{
Generated by miniextendr from Rust fn \code{vec_ptype2_percent_percent}
}
\usage{
\method{vec_ptype2}{percent.percent}(x, y, ...)
}
\arguments{
\item{x}{(no documentation available)}

\item{y}{(no documentation available)}

\item{dots}{(no documentation available)}
}
\description{
Self-coercion prototype (percent + percent = percent)
}
