% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{vec_ptype_abbr.percent}
\alias{vec_ptype_abbr.percent}
\title{Print abbreviation for percent vectors}
\source{
Generated by miniextendr from Rust fn \code{vec_ptype_abbr_percent}
}
\usage{
\method{vec_ptype_abbr}{percent}(x, ...)
}
\arguments{
\item{x}{(no documentation available)}

\item{dots}{(no documentation available)}
}
\description{
Print abbreviation for percent vectors
}
