% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{vec_restore.percent}
\alias{vec_restore.percent}
\title{Restore from proxy after subsetting}
\source{
Generated by miniextendr from Rust fn \code{vec_restore_percent}
}
\usage{
\method{vec_restore}{percent}(x, to, ...)
}
\arguments{
\item{x}{(no documentation available)}

\item{to}{(no documentation available)}

\item{dots}{(no documentation available)}
}
\description{
Restore from proxy after subsetting
}
