% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{force_visible_unit}
\alias{force_visible_unit}
\alias{result_null_on_err}
\alias{result_unwrap_in_r}
\alias{force_invisible_i32}
\alias{with_interrupt_check}
\alias{invisibly_return_arrow}
\alias{rpkg_visibility_interrupts}
\alias{invisibly_return_no_arrow}
\alias{invisibly_option_return_none}
\alias{invisibly_result_return_ok}
\alias{invisibly_option_return_some}
\title{visibility tests}
\source{
Generated by miniextendr from Rust fn \code{force_visible_unit}

Generated by miniextendr from Rust fn \code{result_null_on_err}

Generated by miniextendr from Rust fn \code{result_unwrap_in_r}

Generated by miniextendr from Rust fn \code{force_invisible_i32}

Generated by miniextendr from Rust fn \code{with_interrupt_check}

Generated by miniextendr from Rust fn \code{invisibly_return_arrow}

Generated by miniextendr from Rust fn \code{invisibly_return_no_arrow}

Generated by miniextendr from Rust fn \code{invisibly_result_return_ok}

Generated by miniextendr from Rust fn \code{invisibly_option_return_none}

Generated by miniextendr from Rust fn \code{invisibly_option_return_some}
}
\usage{
force_visible_unit()

result_null_on_err(x)

result_unwrap_in_r(x)

force_invisible_i32()

with_interrupt_check(x)

invisibly_return_arrow()

invisibly_return_no_arrow()

invisibly_result_return_ok()

invisibly_option_return_none()

invisibly_option_return_some()
}
\arguments{
\item{x}{Integer scalar input.}
}
\description{
Visibility and interrupt checks
}
\examples{
invisibly_return_no_arrow()
force_invisible_i32()
with_interrupt_check(2L)
try(invisibly_option_return_none())
\dontrun{
unsafe_C_check_interupt_after()
}
}
