% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{WithEnums}
\alias{WithEnums}
\alias{WithEnums$to_r}
\alias{WithEnums$new_circle}
\alias{WithEnums$new_rectangle}
\alias{WithEnums$from_r}
\alias{$.WithEnums}
\alias{[[.WithEnums}
\title{WithEnums  Class}
\usage{
WithEnums

\method{$}{WithEnums}(self, name)

\method{[[}{WithEnums}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
WithEnums  Class
}
