% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{WithEnums}
\alias{WithEnums}
\alias{WithEnums$to_r}
\alias{WithEnums$new_circle}
\alias{WithEnums$new_rectangle}
\alias{WithEnums$from_r}
\alias{$.WithEnums}
\alias{[[.WithEnums}
\title{WithEnums  Class}
\source{
Generated by miniextendr from Rust type \code{WithEnums}

Generated by miniextendr from \code{WithEnums::to_r}

Generated by miniextendr from \code{WithEnums::new_circle}

Generated by miniextendr from \code{WithEnums::new_rectangle}

Generated by miniextendr from \code{WithEnums::from_r}
}
\usage{
WithEnums

\method{$}{WithEnums}(self, name)

\method{[[}{WithEnums}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
WithEnums  Class
}
