% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\docType{data}
\name{WithOptionals}
\alias{WithOptionals}
\alias{WithOptionals$to_r}
\alias{WithOptionals$all_present}
\alias{WithOptionals$all_none}
\alias{WithOptionals$mixed}
\alias{WithOptionals$from_r}
\alias{$.WithOptionals}
\alias{[[.WithOptionals}
\title{WithOptionals  Class}
\format{
An object of class \code{environment} of length 5.
}
\source{
Generated by miniextendr from Rust type `WithOptionals`

Generated by miniextendr from `WithOptionals::to_r`

Generated by miniextendr from `WithOptionals::all_present`

Generated by miniextendr from `WithOptionals::all_none`

Generated by miniextendr from `WithOptionals::mixed`

Generated by miniextendr from `WithOptionals::from_r`
}
\usage{
WithOptionals

\method{$}{WithOptionals}(self, name)

\method{[[}{WithOptionals}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
WithOptionals  Class
}
\keyword{datasets}
