% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{unsafe_C_test_worker_simple}
\alias{unsafe_C_test_worker_simple}
\alias{unsafe_C_worker_drop_on_panic}
\alias{test_main_thread_r_api}
\alias{test_worker_return_f64}
\alias{rpkg_worker_tests}
\alias{test_worker_return_i32}
\alias{test_worker_return_string}
\alias{unsafe_C_test_nested_with_error}
\alias{unsafe_C_test_nested_with_panic}
\alias{unsafe_C_worker_drop_on_success}
\alias{test_main_thread_r_error}
\alias{unsafe_C_test_extptr_from_worker}
\alias{unsafe_C_test_wrong_thread_r_api}
\alias{unsafe_C_test_nested_worker_calls}
\alias{unsafe_C_test_worker_panic_simple}
\alias{unsafe_C_test_nested_with_r_thread}
\alias{unsafe_C_test_worker_with_r_thread}
\alias{unsafe_C_test_nested_multiple_helpers}
\alias{unsafe_C_test_worker_multiple_r_calls}
\alias{unsafe_C_test_worker_panic_with_drops}
\alias{unsafe_C_test_call_worker_fn_from_main}
\alias{unsafe_C_test_worker_panic_in_r_thread}
\alias{unsafe_C_test_nested_helper_from_worker}
\alias{unsafe_C_test_worker_r_calls_then_error}
\alias{unsafe_C_test_worker_r_calls_then_panic}
\alias{unsafe_C_test_worker_r_error_with_drops}
\alias{unsafe_C_test_worker_r_error_in_r_thread}
\alias{unsafe_C_test_deep_with_r_thread_sequence}
\alias{unsafe_C_test_multiple_extptrs_from_worker}
\alias{test_main_thread_r_error_with_drops}
\alias{unsafe_C_test_worker_panic_in_r_thread_with_drops}
\title{worker tests}
\source{
Generated by miniextendr from Rust fn `C_test_worker_simple`

Generated by miniextendr from Rust fn `C_worker_drop_on_panic`

Generated by miniextendr from Rust fn `test_main_thread_r_api`

Generated by miniextendr from Rust fn `test_worker_return_f64`

Generated by miniextendr from Rust fn `test_worker_return_i32`

Generated by miniextendr from Rust fn `C_test_nested_with_error`

Generated by miniextendr from Rust fn `C_test_nested_with_panic`

Generated by miniextendr from Rust fn `C_worker_drop_on_success`

Generated by miniextendr from Rust fn `test_main_thread_r_error`

Generated by miniextendr from Rust fn `C_test_extptr_from_worker`

Generated by miniextendr from Rust fn `C_test_wrong_thread_r_api`

Generated by miniextendr from Rust fn `test_worker_return_string`

Generated by miniextendr from Rust fn `C_test_nested_worker_calls`

Generated by miniextendr from Rust fn `C_test_worker_panic_simple`

Generated by miniextendr from Rust fn `C_test_nested_with_r_thread`

Generated by miniextendr from Rust fn `C_test_worker_with_r_thread`

Generated by miniextendr from Rust fn `C_test_nested_multiple_helpers`

Generated by miniextendr from Rust fn `C_test_worker_multiple_r_calls`

Generated by miniextendr from Rust fn `C_test_worker_panic_with_drops`

Generated by miniextendr from Rust fn `C_test_call_worker_fn_from_main`

Generated by miniextendr from Rust fn `C_test_worker_panic_in_r_thread`

Generated by miniextendr from Rust fn `C_test_nested_helper_from_worker`

Generated by miniextendr from Rust fn `C_test_worker_r_calls_then_error`

Generated by miniextendr from Rust fn `C_test_worker_r_calls_then_panic`

Generated by miniextendr from Rust fn `C_test_worker_r_error_with_drops`

Generated by miniextendr from Rust fn `C_test_worker_r_error_in_r_thread`

Generated by miniextendr from Rust fn `C_test_deep_with_r_thread_sequence`

Generated by miniextendr from Rust fn `C_test_multiple_extptrs_from_worker`

Generated by miniextendr from Rust fn `test_main_thread_r_error_with_drops`

Generated by miniextendr from Rust fn `C_test_worker_panic_in_r_thread_with_drops`
}
\usage{
unsafe_C_test_worker_simple()

unsafe_C_worker_drop_on_panic()

test_main_thread_r_api()

test_worker_return_f64()

test_worker_return_i32()

unsafe_C_test_nested_with_error()

unsafe_C_test_nested_with_panic()

unsafe_C_worker_drop_on_success()

test_main_thread_r_error()

unsafe_C_test_extptr_from_worker()

unsafe_C_test_wrong_thread_r_api()

test_worker_return_string()

unsafe_C_test_nested_worker_calls()

unsafe_C_test_worker_panic_simple()

unsafe_C_test_nested_with_r_thread()

unsafe_C_test_worker_with_r_thread()

unsafe_C_test_nested_multiple_helpers()

unsafe_C_test_worker_multiple_r_calls()

unsafe_C_test_worker_panic_with_drops()

unsafe_C_test_call_worker_fn_from_main()

unsafe_C_test_worker_panic_in_r_thread()

unsafe_C_test_nested_helper_from_worker()

unsafe_C_test_worker_r_calls_then_error()

unsafe_C_test_worker_r_calls_then_panic()

unsafe_C_test_worker_r_error_with_drops()

unsafe_C_test_worker_r_error_in_r_thread()

unsafe_C_test_deep_with_r_thread_sequence()

unsafe_C_test_multiple_extptrs_from_worker()

test_main_thread_r_error_with_drops()

unsafe_C_test_worker_panic_in_r_thread_with_drops()
}
\description{
worker tests

worker tests

worker tests

worker tests

Test that i32 return values propagate correctly from the worker thread

worker tests

worker tests

worker tests

worker tests

worker tests

worker tests

worker tests

worker tests

worker tests

worker tests

worker tests

worker tests

worker tests

worker tests

worker tests

worker tests

worker tests

worker tests

worker tests

worker tests

worker tests

worker tests

worker tests

worker tests

worker tests
}
\examples{
test_worker_return_i32()
test_worker_return_string()
test_worker_return_f64()
try(test_main_thread_r_error())
\dontrun{
unsafe_C_test_worker_simple()
}
}
