% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{WrapperDemo}
\alias{WrapperDemo}
\title{WrapperDemo R6 Class}
\source{
Generated by miniextendr from Rust type \code{WrapperDemo}
}
\description{
WrapperDemo R6 Class
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-WrapperDemo-initialize}{\code{WrapperDemo$new()}}
    \item \href{#method-WrapperDemo-add_one}{\code{WrapperDemo$add_one()}}
    \item \href{#method-WrapperDemo-add_by}{\code{WrapperDemo$add_by()}}
    \item \href{#method-WrapperDemo-get_value}{\code{WrapperDemo$get_value()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WrapperDemo-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-WrapperDemo-initialize}{}}}
\subsection{\code{WrapperDemo$new()}}{
  Create a new WrapperDemo
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WrapperDemo$new(value, .ptr = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{value}}{Initial integer value.}
      \item{\code{.ptr}}{Internal pointer (used by static methods, not for direct use).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WrapperDemo-add_one"></a>}}
\if{latex}{\out{\hypertarget{method-WrapperDemo-add_one}{}}}
\subsection{\code{WrapperDemo$add_one()}}{
  Increment the value by one.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WrapperDemo$add_one()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WrapperDemo-add_by"></a>}}
\if{latex}{\out{\hypertarget{method-WrapperDemo-add_by}{}}}
\subsection{\code{WrapperDemo$add_by()}}{
  Add a custom amount to the value
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WrapperDemo$add_by(by)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{by}}{Amount to add.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WrapperDemo-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-WrapperDemo-get_value}{}}}
\subsection{\code{WrapperDemo$get_value()}}{
  Get the current value.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WrapperDemo$get_value()}
    \if{html}{\out{</div>}}
  }
}

}
