% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{zero_copy_alloc_r_backed}
\alias{zero_copy_alloc_r_backed}
\title{Allocate an R-backed Arrow buffer, fill it, return as Float64Array}
\source{
Generated by miniextendr from Rust fn \code{zero_copy_alloc_r_backed}
}
\usage{
zero_copy_alloc_r_backed(n)
}
\arguments{
\item{n}{(no documentation available)}
}
\description{
Allocate an R-backed Arrow buffer, fill it, return as Float64Array
}
