% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{zero_copy_arrow_f64_sliced}
\alias{zero_copy_arrow_f64_sliced}
\title{Slice a Float64Array and return it — recovery should fail (different pointer)}
\source{
Generated by miniextendr from Rust fn \code{zero_copy_arrow_f64_sliced}
}
\usage{
zero_copy_arrow_f64_sliced(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
Slice a Float64Array and return it — recovery should fail (different pointer)
}
