% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{zero_copy_arrow_i32_altrep}
\alias{zero_copy_arrow_i32_altrep}
\title{Create a Rust-allocated Int32Array and return as ALTREP}
\source{
Generated by miniextendr from Rust fn \code{zero_copy_arrow_i32_altrep}
}
\usage{
zero_copy_arrow_i32_altrep(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
Create a Rust-allocated Int32Array and return as ALTREP
}
