% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{zero_copy_vec_cow_str_all_borrowed}
\alias{zero_copy_vec_cow_str_all_borrowed}
\title{Returns TRUE if all elements of Vec<Cow\if{html}{\out{<str>}}> are Cow::Borrowed}
\source{
Generated by miniextendr from Rust fn \code{zero_copy_vec_cow_str_all_borrowed}
}
\usage{
zero_copy_vec_cow_str_all_borrowed(x)
}
\arguments{
\item{x}{(no documentation available)}
}
\description{
Returns TRUE if all elements of Vec<Cow\if{html}{\out{<str>}}> are Cow::Borrowed
}
