% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr-wrappers.R
\name{zero_copy_vec_f64_altrep}
\alias{zero_copy_vec_f64_altrep}
\title{Create a Vec\if{html}{\out{<f64>}} ALTREP (not Arrow, just plain Rust vec)}
\source{
Generated by miniextendr from Rust fn \code{zero_copy_vec_f64_altrep}
}
\usage{
zero_copy_vec_f64_altrep(n)
}
\arguments{
\item{n}{(no documentation available)}
}
\description{
Create a Vec\if{html}{\out{<f64>}} ALTREP (not Arrow, just plain Rust vec)
}
